Main API Endpoint (Current)

Main API Endpoint (Current)

Usage

CurrentMain(alias)

Arguments

alias

<chr> endpoint alias

Value

An S7 <CurrentMain> object.

Examples

CurrentMain("enrollees")
#> <providertwo::CurrentMain>
#>  @ title      : chr "Medicare Fee-For-Service Public Provider Enrollment"
#>  @ description: chr "The Medicare Fee-For-Service Public Provider Enrollment dataset includes information on providers who are activ"| __truncated__
#>  @ contact    : chr "Provider Enrollment Data Requests - CPI (mailto:ProviderEnrollmentDataRequests@cms.hhs.gov)"
#>  @ modified   : Date[1:1], format: "2025-03-18"
#>  @ identifier : chr "https://data.cms.gov/data-api/v1/dataset/2457ea29-fc82-48b0-86ec-3b0755de7515/data-viewer"
#>  @ rows       : int 2945846
#>  @ pages      : int 590
#>  @ fields     : chr [1:10] "NPI" "PECOS_ASCT_CNTL_ID" "ENRLMT_ID" "PROVIDER_TYPE_CD" ...
#>  @ download   : chr "https://data.cms.gov/sites/default/files/2025-03/3a6ec647-ed73-4c4e-849b-2d631e6b496f/PPEF_Enrollment_Extract_2025.02.03.csv"
#>  @ temporal   : chr "2023-01-01 • 2024-12-31"
#>  @ periodicity: chr "Quarterly [R/P3M]"
#>  @ resources  : chr "https://data.cms.gov/data-api/v1/dataset-resources/2457ea29-fc82-48b0-86ec-3b0755de7515"
#>  @ dictionary : chr "https://data.cms.gov/resources/medicare-fee-for-service-public-provider-enrollment-data-dictionary"
#>  @ site       : chr "https://data.cms.gov/provider-characteristics/medicare-provider-supplier-enrollment/medicare-fee-for-service-pu"| __truncated__
#>  @ references : chr "https://data.cms.gov/resources/fee-for-service-public-provider-enrollment-methodology"
CurrentMain("opt_out")
#> <providertwo::CurrentMain>
#>  @ title      : chr "Opt Out Affidavits"
#>  @ description: chr "The Opt Out Affidavits dataset provides information on providers who have decided not to participate in Medicar"| __truncated__
#>  @ contact    : chr "Provider Enrollment Data Requests - CPI (mailto:ProviderEnrollmentDataRequests@cms.hhs.gov)"
#>  @ modified   : Date[1:1], format: "2025-02-27"
#>  @ identifier : chr "https://data.cms.gov/data-api/v1/dataset/9887a515-7552-4693-bf58-735c77af46d7/data-viewer"
#>  @ rows       : int 49316
#>  @ pages      : int 10
#>  @ fields     : chr [1:13] "First Name" "Last Name" "NPI" "Specialty" ...
#>  @ download   : chr "https://data.cms.gov/sites/default/files/2025-02/101001b7-73df-419e-b101-602baa5015c5/Opt_Out_List_January.csv"
#>  @ temporal   : chr "2023-04-01 • 2025-01-31"
#>  @ periodicity: chr "Monthly [R/P1M]"
#>  @ resources  : chr "https://data.cms.gov/data-api/v1/dataset-resources/9887a515-7552-4693-bf58-735c77af46d7"
#>  @ dictionary : chr "https://data.cms.gov/resources/opt-out-affidavits-data-dictionary"
#>  @ site       : chr "https://data.cms.gov/provider-characteristics/medicare-provider-supplier-enrollment/opt-out-affidavits"
#>  @ references : chr "https://data.cms.gov/resources/opt-out-affidavits-methodology"
CurrentMain("order_refer")
#> <providertwo::CurrentMain>
#>  @ title      : chr "Order and Referring"
#>  @ description: chr "The Order and Referring dataset provides information on all physicians and non-physician practitioners, by thei"| __truncated__
#>  @ contact    : chr "Provider Enrollment Data Requests - CPI (mailto:ProviderEnrollmentDataRequests@cms.hhs.gov)"
#>  @ modified   : Date[1:1], format: "2025-03-18"
#>  @ identifier : chr "https://data.cms.gov/data-api/v1/dataset/c99b5865-1119-4436-bb80-c5af2773ea1f/data-viewer"
#>  @ rows       : int 1936503
#>  @ pages      : int 388
#>  @ fields     : chr [1:8] "NPI" "LAST_NAME" "FIRST_NAME" "PARTB" "DME" "HHA" "PMD" ...
#>  @ download   : chr "https://data.cms.gov/sites/default/files/2025-03/66c7dae3-0119-4258-9d97-999f92837f0a/OrderReferring_20250317.csv"
#>  @ temporal   : chr "2023-05-21 • 2025-03-15"
#>  @ periodicity: chr "Twice a Week [R/P3.5D]"
#>  @ resources  : chr "https://data.cms.gov/data-api/v1/dataset-resources/c99b5865-1119-4436-bb80-c5af2773ea1f"
#>  @ dictionary : chr "https://data.cms.gov/resources/order-and-referring-data-dictionary"
#>  @ site       : chr "https://data.cms.gov/provider-characteristics/medicare-provider-supplier-enrollment/order-and-referring"
#>  @ references : chr "https://data.cms.gov/resources/order-and-referring-methodology"
CurrentMain("reassignments")
#> <providertwo::CurrentMain>
#>  @ title      : chr "Revalidation Reassignment List"
#>  @ description: chr "The Revalidation Reassignment List dataset provides information on reassignments of providers who are due for revalidation."
#>  @ contact    : chr "Provider Enrollment Data Requests - CPI (mailto:ProviderEnrollmentDataRequests@cms.hhs.gov)"
#>  @ modified   : Date[1:1], format: "2025-03-17"
#>  @ identifier : chr "https://data.cms.gov/data-api/v1/dataset/20f51cff-4137-4f3a-b6b7-bfc9ad57983b/data-viewer"
#>  @ rows       : int 3682027
#>  @ pages      : int 737
#>  @ fields     : chr [1:16] "Group PAC ID" "Group Enrollment ID" ...
#>  @ download   : chr "https://data.cms.gov/sites/default/files/2025-02/693a548a-d960-4ea7-b17e-3422e1e20a75/revalidation_group_reassi"| __truncated__
#>  @ temporal   : chr "2022-02-01 • 2025-03-31"
#>  @ periodicity: chr "Monthly [R/P1M]"
#>  @ resources  : chr "https://data.cms.gov/data-api/v1/dataset-resources/20f51cff-4137-4f3a-b6b7-bfc9ad57983b"
#>  @ dictionary : chr "https://data.cms.gov/resources/revalidation-reassignments-list-data-dictionary"
#>  @ site       : chr "https://data.cms.gov/provider-characteristics/medicare-provider-supplier-enrollment/revalidation-reassignment-list"
#>  @ references : chr "https://data.cms.gov/resources/additional-information-on-revalidation"
CurrentMain("hospitals")
#> <providertwo::CurrentMain>
#>  @ title      : chr "Hospital Enrollments"
#>  @ description: chr "The Hospital Enrollments dataset provides enrollment information of all Hospitals currently enrolled in Medicar"| __truncated__
#>  @ contact    : chr "Provider Enrollment Data Requests - CPI (mailto:ProviderEnrollmentDataRequests@cms.hhs.gov)"
#>  @ modified   : Date[1:1], format: "2025-01-15"
#>  @ identifier : chr "https://data.cms.gov/data-api/v1/dataset/f6f6505c-e8b0-4d57-b258-e2b94133aaf2/data-viewer"
#>  @ rows       : int 9272
#>  @ pages      : int 2
#>  @ fields     : chr [1:39] "ENROLLMENT ID" "ENROLLMENT STATE" "PROVIDER TYPE CODE" ...
#>  @ download   : chr "https://data.cms.gov/sites/default/files/2025-01/0ef8ef0d-bf93-4b00-a542-a4e2853947cc/Hospital_Enrollments_2025.01.02.csv"
#>  @ temporal   : chr "2022-11-01 • 2025-01-31"
#>  @ periodicity: chr "Monthly [R/P1M]"
#>  @ resources  : chr "https://data.cms.gov/data-api/v1/dataset-resources/f6f6505c-e8b0-4d57-b258-e2b94133aaf2"
#>  @ dictionary : chr "https://data.cms.gov/sites/default/files/2023-07/Hospital_Enrollments_Data_Dictionary.pdf"
#>  @ site       : chr "https://data.cms.gov/provider-characteristics/hospitals-and-other-facilities/hospital-enrollments"
#>  @ references : chr "https://data.cms.gov/sites/default/files/2024-10/Hospital_Data_Guidance.pdf"
CurrentMain("laboratories")
#> <providertwo::CurrentMain>
#>  @ title      : chr "Provider of Services File - Clinical Laboratories"
#>  @ description: chr "The Provider of Services (POS) Clinical Laboratories (CLIA) data provides information on CLIA demographics and "| __truncated__
#>  @ contact    : chr "QIESData - CCSQ (mailto:QIESData@cms.hhs.gov)"
#>  @ modified   : Date[1:1], format: "2025-01-16"
#>  @ identifier : chr "https://data.cms.gov/data-api/v1/dataset/d3eb38ac-d8e9-40d3-b7b7-6205d3d1dc16/data-viewer"
#>  @ rows       : int 653223
#>  @ pages      : int 131
#>  @ fields     : chr [1:103] "PRVDR_CTGRY_SBTYP_CD" "PRVDR_CTGRY_CD" "CHOW_CNT" ...
#>  @ download   : chr "https://data.cms.gov/sites/default/files/2025-01/f16b0884-32a4-427f-b679-b48d3ef3146f/clia_data_Q4%202024.csv"
#>  @ temporal   : chr "2011-10-01 • 2024-12-31"
#>  @ periodicity: chr "Quarterly [R/P3M]"
#>  @ resources  : chr "https://data.cms.gov/data-api/v1/dataset-resources/d3eb38ac-d8e9-40d3-b7b7-6205d3d1dc16"
#>  @ dictionary : chr "https://data.cms.gov/sites/default/files/2023-07/50eacc07-7930-4e62-97e3-e04a2dc1e2a3/P.QWB.POSQ.CLIA.LAYOUT.MAR23.pdf"
#>  @ site       : chr "https://data.cms.gov/provider-characteristics/hospitals-and-other-facilities/provider-of-services-file-clinical-laboratories"
#>  @ references : chr "https://data.cms.gov/resources/provider-of-services-file-methodology"
CurrentMain("crosswalk")
#> <providertwo::CurrentMain>
#>  @ title      : chr "Medicare Provider and Supplier Taxonomy Crosswalk"
#>  @ description: chr "The Medicare Provider and Supplier Taxonomy Crosswalk dataset lists the providers and suppliers eligible to enr"| __truncated__
#>  @ contact    : chr "Provider Enrollment Data Requests - CPI (mailto:ProviderEnrollmentDataRequests@cms.hhs.gov)"
#>  @ modified   : Date[1:1], format: "2025-01-17"
#>  @ identifier : chr "https://data.cms.gov/data-api/v1/dataset/113eb0bc-0c9a-4d91-9f93-3f6b28c0bf6b/data-viewer"
#>  @ rows       : int 561
#>  @ pages      : int 1
#>  @ fields     : chr [1:4] "MEDICARE SPECIALTY CODE" ...
#>  @ download   : chr "https://data.cms.gov/sites/default/files/2025-01/7b0ea92d-2f70-424e-bec4-3fd35dc9a597/Medicare_Provider_and_Sup"| __truncated__
#>  @ temporal   : chr "2021-07-01 • 2025-06-30"
#>  @ periodicity: chr "Biannually [R/P6M]"
#>  @ resources  : chr "https://data.cms.gov/data-api/v1/dataset-resources/113eb0bc-0c9a-4d91-9f93-3f6b28c0bf6b"
#>  @ dictionary : chr "https://data.cms.gov/resources/medicare-provider-and-supplier-taxonomy-crosswalk-data-dictionary"
#>  @ site       : chr "https://data.cms.gov/provider-characteristics/medicare-provider-supplier-enrollment/medicare-provider-and-suppl"| __truncated__
#>  @ references : chr "https://data.cms.gov/resources/medicare-provider-and-supplier-taxonomy-crosswalk-methodology"
CurrentMain("rbcs")
#> <providertwo::CurrentMain>
#>  @ title      : chr "Restructured BETOS Classification System"
#>  @ description: chr "The Restructured BETOS Classification System (RBCS) dataset is a taxonomy that allows researchers to group heal"| __truncated__
#>  @ contact    : chr "PDAG Data Products - OEDA (mailto:PDAG_Data_Products@cms.hhs.gov)"
#>  @ modified   : Date[1:1], format: "2024-10-22"
#>  @ identifier : chr "https://data.cms.gov/data-api/v1/dataset/e3db6e56-149f-49ce-b374-40aecda2357b/data-viewer"
#>  @ rows       : int 18077
#>  @ pages      : int 4
#>  @ fields     : chr [1:14] "HCPCS_Cd" "RBCS_ID" "RBCS_Cat" "RBCS_Cat_Desc" ...
#>  @ download   : chr "https://data.cms.gov/sites/default/files/2024-10/30b74146-c5ce-47e0-8ae2-ad1204c81906/2024%20RBCS%20Taxonomy_CSV.csv"
#>  @ temporal   : chr "2024-01-01 • 2024-12-31"
#>  @ periodicity: chr "Annually [R/P1Y]"
#>  @ resources  : chr "https://data.cms.gov/data-api/v1/dataset-resources/e3db6e56-149f-49ce-b374-40aecda2357b"
#>  @ dictionary : chr "https://data.cms.gov/resources/rbcs-data-dictionary-2024"
#>  @ site       : chr "https://data.cms.gov/provider-summary-by-type-of-service/provider-service-classifications/restructured-betos-cl"| __truncated__
#>  @ references : chr "https://data.cms.gov/sites/default/files/2024-10/2024%20RBCS%20Final%20Report.pdf"
CurrentMain("facilities")
#> <providertwo::CurrentMain>
#>  @ title      : chr "Provider of Services File - Hospital & Non-Hospital Facilities"
#>  @ description: chr "Please be advised that as of Q4 2023 there is a new Provider of Service file (POS) that contains the provider a"| __truncated__
#>  @ contact    : chr "QIESData - CCSQ (mailto:QIESData@cms.hhs.gov)"
#>  @ modified   : Date[1:1], format: "2025-01-31"
#>  @ identifier : chr "https://data.cms.gov/data-api/v1/dataset/8ba0f9b4-9493-4aa0-9f82-44ea9468d1b5/data-viewer"
#>  @ rows       : int 115647
#>  @ pages      : int 24
#>  @ fields     : chr [1:473] "PRVDR_CTGRY_SBTYP_CD" "PRVDR_CTGRY_CD" "CHOW_CNT" ...
#>  @ download   : chr "https://data.cms.gov/sites/default/files/2025-01/b5fb8b31-8e6e-41d0-8d73-8da1e2185ee2/PQWB.POSQ.OTHER.DATA.DEC24.csv"
#>  @ temporal   : chr "2011-10-01 • 2024-12-31"
#>  @ periodicity: chr "Quarterly [R/P3M]"
#>  @ resources  : chr "https://data.cms.gov/data-api/v1/dataset-resources/8ba0f9b4-9493-4aa0-9f82-44ea9468d1b5"
#>  @ dictionary : chr "https://data.cms.gov/sites/default/files/2023-07/0ca58d5d-7914-4532-b22d-41741d3e6151/P.QWB.POSQ.OTHER.LAYOUT.MAR23.pdf"
#>  @ site       : chr "https://data.cms.gov/provider-characteristics/hospitals-and-other-facilities/provider-of-services-file-hospital"| __truncated__
#>  @ references : chr "https://data.cms.gov/resources/provider-of-services-file-methodology"
CurrentMain("home_health")
#> <providertwo::CurrentMain>
#>  @ title      : chr "Home Health Agency Enrollments"
#>  @ description: chr "The Home Health Agency (HHA) Enrollments dataset provides enrollment information on all HHAs currently enrolled"| __truncated__
#>  @ contact    : chr "Provider Enrollment Data Requests - CPI (mailto:ProviderEnrollmentDataRequests@cms.hhs.gov)"
#>  @ modified   : Date[1:1], format: "2025-01-15"
#>  @ identifier : chr "https://data.cms.gov/data-api/v1/dataset/15f64ab4-3172-4a27-b589-ebd67a6d28aa/data-viewer"
#>  @ rows       : int 11708
#>  @ pages      : int 3
#>  @ fields     : chr [1:22] "ENROLLMENT ID" "ENROLLMENT STATE" "PROVIDER TYPE CODE" ...
#>  @ download   : chr "https://data.cms.gov/sites/default/files/2025-01/c133f82f-5635-47d3-b85f-364b379a322a/HHA_Enrollments_2025.01.02.csv"
#>  @ temporal   : chr "2023-04-01 • 2025-03-31"
#>  @ periodicity: chr "Quarterly [R/P3M]"
#>  @ resources  : chr "https://data.cms.gov/data-api/v1/dataset-resources/15f64ab4-3172-4a27-b589-ebd67a6d28aa"
#>  @ dictionary : chr "https://data.cms.gov/sites/default/files/2024-10/HHA_Enrollments_Data_Dictionary.pdf"
#>  @ site       : chr "https://data.cms.gov/provider-characteristics/hospitals-and-other-facilities/home-health-agency-enrollments"
#>  @ references : chr "https://data.cms.gov/sites/default/files/2024-10/HHA_Data_Guidance.pdf"
CurrentMain("hospice")
#> <providertwo::CurrentMain>
#>  @ title      : chr "Hospice Enrollments"
#>  @ description: chr "The Hospice Enrollments dataset provides enrollment information of all hospices currently enrolled in Medicare."| __truncated__
#>  @ contact    : chr "Provider Enrollment Data Requests - CPI (mailto:ProviderEnrollmentDataRequests@cms.hhs.gov)"
#>  @ modified   : Date[1:1], format: "2025-01-15"
#>  @ identifier : chr "https://data.cms.gov/data-api/v1/dataset/25704213-e833-4b8b-9dbc-58dd17149209/data-viewer"
#>  @ rows       : int 6603
#>  @ pages      : int 2
#>  @ fields     : chr [1:20] "ENROLLMENT ID" "ENROLLMENT STATE" "PROVIDER TYPE CODE" ...
#>  @ download   : chr "https://data.cms.gov/sites/default/files/2025-01/afd62e57-8a99-4c49-a0b5-358a25ce39dd/Hospice_Enrollments_2025.01.02.csv"
#>  @ temporal   : chr "2023-04-01 • 2025-03-31"
#>  @ periodicity: chr "Quarterly [R/P3M]"
#>  @ resources  : chr "https://data.cms.gov/data-api/v1/dataset-resources/25704213-e833-4b8b-9dbc-58dd17149209"
#>  @ dictionary : chr "https://data.cms.gov/sites/default/files/2023-03/f736ce50-3d6b-4ee2-802f-2e5424ae71ac/Hospice_Enrollments_Data_Dictionary.pdf"
#>  @ site       : chr "https://data.cms.gov/provider-characteristics/hospitals-and-other-facilities/hospice-enrollments"
#>  @ references : chr "https://data.cms.gov/sites/default/files/2023-03/f193e9ef-9fd7-4de3-9a27-0d5281cf80a9/Hospice_Data_Guidance.pdf"
CurrentMain("dialysis")
#> <providertwo::CurrentMain>
#>  @ title      : chr "Medicare Dialysis Facilities"
#>  @ description: chr "The Medicare Dialysis Facilities data provides information on clinical and patient measures for Medicare-certif"| __truncated__
#>  @ contact    : chr "ESRD Questions - CCSQ (mailto:ESRDQuestions@cms.hhs.gov)"
#>  @ modified   : Date[1:1], format: "2024-02-29"
#>  @ identifier : chr "https://data.cms.gov/data-api/v1/dataset/f8610e87-ba25-43a3-a49e-927dbc8701ae/data-viewer"
#>  @ rows       : int 11679243
#>  @ pages      : int 2336
#>  @ fields     : chr [1:14] "state" "CCN" "Provider_Name" "city" "Ownership_Type" ...
#>  @ download   : chr "https://data.cms.gov/sites/default/files/2024-02/b2c2658d-e19d-48db-8b3b-9d343a4d395f/dfr_facility_socrata_fy2024.csv"
#>  @ temporal   : chr "2009-01-01 • 2024-12-31"
#>  @ periodicity: chr "Annually [R/P1Y]"
#>  @ resources  : chr "https://data.cms.gov/data-api/v1/dataset-resources/f8610e87-ba25-43a3-a49e-927dbc8701ae"
#>  @ dictionary : chr "https://data.cms.gov/sites/default/files/2024-02/FY2024%20Data%20Dictionary.xlsx"
#>  @ site       : chr "https://data.cms.gov/quality-of-care/medicare-dialysis-facilities"
#>  @ references : chr "https://data.cms.gov/sites/default/files/2024-02/FY2024_DFR_Guide%20and%20Methodology.pdf"
CurrentMain("snf")
#> <providertwo::CurrentMain>
#>  @ title      : chr "Skilled Nursing Facility Enrollments"
#>  @ description: chr "The Skilled Nursing Facility (SNF) Enrollments dataset provides enrollment information of all SNF s currently e"| __truncated__
#>  @ contact    : chr "Provider Enrollment Data Requests - CPI (mailto:ProviderEnrollmentDataRequests@cms.hhs.gov)"
#>  @ modified   : Date[1:1], format: "2025-03-17"
#>  @ identifier : chr "https://data.cms.gov/data-api/v1/dataset/5f2c306f-3b1c-42cd-b037-187b2ce22126/data-viewer"
#>  @ rows       : int 14462
#>  @ pages      : int 3
#>  @ fields     : chr [1:23] "ENROLLMENT ID" "ENROLLMENT STATE" "PROVIDER TYPE CODE" ...
#>  @ download   : chr "https://data.cms.gov/sites/default/files/2025-03/9ef2f127-250d-4809-8f72-99762554a08d/SNF_Enrollments_2025.02.28.csv"
#>  @ temporal   : chr "2022-09-01 • 2025-03-31"
#>  @ periodicity: chr "Monthly [R/P1M]"
#>  @ resources  : chr "https://data.cms.gov/data-api/v1/dataset-resources/5f2c306f-3b1c-42cd-b037-187b2ce22126"
#>  @ dictionary : chr "https://data.cms.gov/sites/default/files/2024-10/SNF_Enrollments_Data_Dictionary.pdf"
#>  @ site       : chr "https://data.cms.gov/provider-characteristics/hospitals-and-other-facilities/skilled-nursing-facility-enrollments"
#>  @ references : chr "https://data.cms.gov/sites/default/files/2024-12/1ea55414-f441-4694-a13c-6400d4e28ef6/SNF_Data_Guidance.pdf"