CurrentOpen Payments API Endpoint

CurrentOpen Payments API Endpoint

Usage

CurrentOpen(alias)

Arguments

alias

<chr> endpoint alias

Value

An S7 <CurrentOpen> object.

Examples

CurrentOpen("prof_cov")
#> <providertwo::CurrentOpen>
#>  @ title      : chr "Covered Recipient Profile Supplement"
#>  @ description: chr "The Covered Recipient Profile Supplement file contains information about physicians and non-physician practitio"| __truncated__
#>  @ contact    : chr "Open Payments (mailto:openpayments@cms.hhs.gov)"
#>  @ modified   : Date[1:1], format: "2025-01-30"
#>  @ identifier : chr "https://openpaymentsdata.cms.gov/api/1/datastore/query/23160558-6742-54ff-8b9f-cac7d514ff4e/0"
#>  @ rows       : int 1530412
#>  @ pages      : int 3061
#>  @ fields     : chr [1:32] "covered_recipient_profile_type" ...
#>  @ download   : chr "https://download.cms.gov/openpayments/PHPRFL_P01302025_01212025/OP_CVRD_RCPNT_PRFL_SPLMTL_P01302025_01212025.csv"
#>  @ uuid       : chr "23160558-6742-54ff-8b9f-cac7d514ff4e"
CurrentOpen("prof_phys")
#> <providertwo::CurrentOpen>
#>  @ title      : chr "Physician (Distinct) Profile Information"
#>  @ description: chr "This dataset is pre-filtered based on the most frequent searches of Open Payments data."
#>  @ contact    : chr "Open Payments (mailto:openpayments@cms.hhs.gov)"
#>  @ modified   : Date[1:1], format: "2025-01-30"
#>  @ identifier : chr "https://openpaymentsdata.cms.gov/api/1/datastore/query/6ed6ae76-2999-49da-b0b2-d7df150ac754/0"
#>  @ rows       : int 1528054
#>  @ pages      : int 3057
#>  @ fields     : chr [1:33] "covered_recipient_profile_type" ...
#>  @ download   : chr "https://download.cms.gov/openpayments/SMRY_RPTS_P01302025_01212025/PBLCTN_PHYSN_NON_PHYSN_PRCTNR_PRFL_SRCH_P013"| __truncated__
#>  @ uuid       : chr "6ed6ae76-2999-49da-b0b2-d7df150ac754"
CurrentOpen("prof_info")
#> <providertwo::CurrentOpen>
#>  @ title      : chr "Profile Information"
#>  @ description: chr "This dataset provides profile information to display on payment search results."
#>  @ contact    : chr "Open Payments (mailto:openpayments@cms.hhs.gov)"
#>  @ modified   : Date[1:1], format: "2025-01-30"
#>  @ identifier : chr "https://openpaymentsdata.cms.gov/api/1/datastore/query/f2ed17cc-f045-4e25-9b28-75c72f0c7efe/0"
#>  @ rows       : int 1534695
#>  @ pages      : int 3070
#>  @ fields     : chr [1:20] "entity_id" "entity_provider_type" "entity_npi" ...
#>  @ download   : chr "https://download.cms.gov/openpayments/SMRY_RPTS_P01302025_01212025/all-entities-profile-search.csv"
#>  @ uuid       : chr "f2ed17cc-f045-4e25-9b28-75c72f0c7efe"
CurrentOpen("prof_map")
#> <providertwo::CurrentOpen>
#>  @ title      : chr "Provider Profile ID Mapping Table"
#>  @ description: chr "Maps secondary/duplicate provider profile IDs to the primary ID in the Open Payments system. For multiple reaso"| __truncated__
#>  @ contact    : chr "Open Payments (mailto:openpayments@cms.hhs.gov)"
#>  @ modified   : Date[1:1], format: "2025-01-30"
#>  @ identifier : chr "https://openpaymentsdata.cms.gov/api/1/datastore/query/d19e5be2-5b8c-562c-af22-c243503bcfac/0"
#>  @ rows       : int 2358
#>  @ pages      : int 5
#>  @ fields     : chr [1:2] "primary_provider_profile_id" ...
#>  @ download   : chr "https://download.cms.gov/openpayments/SMRY_RPTS_P01302025_01212025/PBLCTN_PRVDR_PRFL_MAPPING_P01302025_01212025.csv"
#>  @ uuid       : chr "d19e5be2-5b8c-562c-af22-c243503bcfac"
CurrentOpen("prof_entity")
#> <providertwo::CurrentOpen>
#>  @ title      : chr "Reporting Entity Profile Information"
#>  @ description: chr "This dataset is pre-filtered based on the most frequent searches of Open Payments data."
#>  @ contact    : chr "Open Payments (mailto:openpayments@cms.hhs.gov)"
#>  @ modified   : Date[1:1], format: "2025-01-30"
#>  @ identifier : chr "https://openpaymentsdata.cms.gov/api/1/datastore/query/1cf0c6c0-c377-466e-b78f-037b442559f8/0"
#>  @ rows       : int 2843
#>  @ pages      : int 6
#>  @ fields     : chr [1:9] "amgpo_making_payment_id" "amgpo_making_payment_name" ...
#>  @ download   : chr "https://download.cms.gov/openpayments/SMRY_RPTS_P01302025_01212025/PBLCTN_RPTG_ORG_PRFL_SRCH_P01302025_01212025.csv"
#>  @ uuid       : chr "1cf0c6c0-c377-466e-b78f-037b442559f8"
CurrentOpen("prof_teach")
#> <providertwo::CurrentOpen>
#>  @ title      : chr "Teaching Hospital Profile Information"
#>  @ description: chr "This dataset is pre-filtered based on the most frequent searches of Open Payments data."
#>  @ contact    : chr "Open Payments (mailto:openpayments@cms.hhs.gov)"
#>  @ modified   : Date[1:1], format: "2025-01-30"
#>  @ identifier : chr "https://openpaymentsdata.cms.gov/api/1/datastore/query/8219fcb3-fb3d-4b51-9f65-be8a048c0608/0"
#>  @ rows       : int 1440
#>  @ pages      : int 3
#>  @ fields     : chr [1:12] "teaching_hospital_ccn" "teaching_hospital_name" ...
#>  @ download   : chr "https://download.cms.gov/openpayments/SMRY_RPTS_P01302025_01212025/PBLCTN_TH_PRFL_SRCH_P01302025_01212025.csv"
#>  @ uuid       : chr "8219fcb3-fb3d-4b51-9f65-be8a048c0608"
CurrentOpen("dashboard")
#> <providertwo::CurrentOpen>
#>  @ title      : chr "Summary Dashboard - all Years"
#>  @ description: chr "Summary data."
#>  @ contact    : chr "Open Payments (mailto:openpayments@cms.hhs.gov)"
#>  @ modified   : Date[1:1], format: "2025-01-30"
#>  @ identifier : chr "https://openpaymentsdata.cms.gov/api/1/datastore/query/e0d225fc-8230-401d-8fad-e2262fb22b4c/0"
#>  @ rows       : int 90
#>  @ pages      : int 1
#>  @ fields     : chr [1:10] "dashboard_row_number" "data_metrics" "py_2017" "py_2018" ...
#>  @ download   : chr "https://download.cms.gov/openpayments/SMRY_RPTS_P01302025_01212025/PBLCTN_DSHBRD_P01302025_01212025.csv"
#>  @ uuid       : chr "e0d225fc-8230-401d-8fad-e2262fb22b4c"
CurrentOpen("pay_state_total")
#> <providertwo::CurrentOpen>
#>  @ title      : chr "State Level Payment Total and Averages for all Years"
#>  @ description: chr "This dataset is pre-filtered based on the most frequent searches of Open Payments data."
#>  @ contact    : chr "Open Payments (mailto:openpayments@cms.hhs.gov)"
#>  @ modified   : Date[1:1], format: "2025-01-30"
#>  @ identifier : chr "https://openpaymentsdata.cms.gov/api/1/datastore/query/02157882-0f43-473a-a777-c96eb5b1b623/0"
#>  @ rows       : int 2765
#>  @ pages      : int 6
#>  @ fields     : chr [1:28] "country_code" "country_name" "state_code" "state_name" ...
#>  @ download   : chr "https://download.cms.gov/openpayments/SMRY_RPTS_P01302025_01212025/PBLCTN_STATE_SMRY_P01302025_01212025-joined.csv"
#>  @ uuid       : chr "02157882-0f43-473a-a777-c96eb5b1b623"
CurrentOpen("pay_state_group")
#> <providertwo::CurrentOpen>
#>  @ title      : chr "State Payment Totals and Averages Grouped by Nature of Payment for all Years"
#>  @ description: chr "Groups payment data by state, recipient type, and nature of payment."
#>  @ contact    : chr "Open Payments (mailto:openpayments@cms.hhs.gov)"
#>  @ modified   : Date[1:1], format: "2025-01-30"
#>  @ identifier : chr "https://openpaymentsdata.cms.gov/api/1/datastore/query/7c5144d9-1b2a-5039-a91a-cccb8ed7805c/0"
#>  @ rows       : int 16758
#>  @ pages      : int 34
#>  @ fields     : chr [1:28] "country_code" "country_name" "state_code" "state_name" ...
#>  @ download   : chr "https://download.cms.gov/openpayments/SMRY_RPTS_P01302025_01212025/PBLCTN_NTR_OF_PYMT_BY_STATE_SMRY_PGYRall_P01"| __truncated__
#>  @ uuid       : chr "7c5144d9-1b2a-5039-a91a-cccb8ed7805c"
CurrentOpen("pay_nat_group")
#> <providertwo::CurrentOpen>
#>  @ title      : chr "National Level Payment Total and Averages by Provider Specialty for all Years"
#>  @ description: chr "This dataset is pre-filtered based on the most frequent searches of Open Payments data."
#>  @ contact    : chr "Open Payments (mailto:openpayments@cms.hhs.gov)"
#>  @ modified   : Date[1:1], format: "2025-01-30"
#>  @ identifier : chr "https://openpaymentsdata.cms.gov/api/1/datastore/query/04c3355f-955b-4620-b3a8-08020cc9c1f2/0"
#>  @ rows       : int 4188
#>  @ pages      : int 9
#>  @ fields     : chr [1:21] "payment_type" "recipient_type" "program_year" ...
#>  @ download   : chr "https://download.cms.gov/openpayments/SMRY_RPTS_P01302025_01212025/PBLCTN_SPLTY_SMRY_P01302025_01212025-joined.csv"
#>  @ uuid       : chr "04c3355f-955b-4620-b3a8-08020cc9c1f2"
CurrentOpen("pay_nat_total")
#> <providertwo::CurrentOpen>
#>  @ title      : chr "National Level Payment Total and Averages for all Years"
#>  @ description: chr "This dataset is pre-filtered based on the most frequent searches of Open Payments data."
#>  @ contact    : chr "Open Payments (mailto:openpayments@cms.hhs.gov)"
#>  @ modified   : Date[1:1], format: "2025-01-30"
#>  @ identifier : chr "https://openpaymentsdata.cms.gov/api/1/datastore/query/aaaddc12-a3a7-4cc5-9c12-4a726317aa1c/0"
#>  @ rows       : int 24
#>  @ pages      : int 1
#>  @ fields     : chr [1:18] "metric_level" "payment_type" "recipient_type" ...
#>  @ download   : chr "https://download.cms.gov/openpayments/SMRY_RPTS_P01302025_01212025/PBLCTN_NTNL_SMRY_P01302025_01212025-joined.csv"
#>  @ uuid       : chr "aaaddc12-a3a7-4cc5-9c12-4a726317aa1c"