Current Open Payments API Endpoint

Current Open Payments API Endpoint

Temporal Open Payments API Endpoint

Open Payments API Endpoint Temporal Group

Usage

openMain(alias)

openTemp(alias)

openTempGroup(
  title = character(0),
  description = character(0),
  contact = character(0),
  rows = integer(0),
  pages = integer(0),
  fields = character(0),
  years = integer(0),
  endpoints = list()
)

Arguments

alias

<chr> endpoint alias

title

<chr> title

description

<chr> description

contact

<chr> contact name and email address

rows

<int> number of rows

pages

<int> number of pages

fields

<chr> field names

years

<int> years available

endpoints

<data.frame> endpoints

Value

An S7 <openCurr> object.

An S7 <openTemp> object.

An S7 <openTempGroup> object.

Examples

openMain("prof_cov")
#> ── Covered Recipient Profile Supplement ──
#> 
#>   The Covered Recipient Profile Supplement
#>   file contains information about physicians
#>   and non-physician practitioners who have
#>   been indicated as recipients of payments,
#>   other transfers of value, or ownership and
#>   investment interest in payment records, as
#>   well as physicians and non-physician
#>   practitioners who have been identified as
#>   principal investigators associated with
#>   research payment records published by Open
#>   Payments. This file contains only those
#>   physicians that have at least one
#>   published payment record in this cycle of
#>   the publication as of January 30, 2025.
#>   The criteria used by the Centers for
#>   Medicare & Medicaid Services (CMS) to
#>   determine which payment records are
#>   eligible for publication is available in
#>   the Open Payments Methodology and Data
#>   Dictionary Document. This document can be
#>   found on the Resources page of the Open
#>   Payments website
#>   (https://www.cms.gov/OpenPayments/Resources).
#>   The Methodology and Data Dictionary
#>   Document also includes information on the
#>   data collection and reporting methodology,
#>   data fields included in the files, and any
#>   notes or special considerations that users
#>   should be aware of.
#> 
#> • Rows 1,530,412
#> • Pages 3,061
#> • Fields 32
#> 
#> • Modified ❯ 2025-01-30
openMain("prof_phys")
#> ── Physician (Distinct) Profile Information ──
#> 
#>   This dataset is pre-filtered based on the
#>   most frequent searches of Open Payments
#>   data.
#> 
#> • Rows 1,528,054
#> • Pages 3,057
#> • Fields 33
#> 
#> • Modified ❯ 2025-01-30
openMain("prof_info")
#> ── Profile Information ──
#> 
#>   This dataset provides profile information
#>   to display on payment search results.
#> 
#> • Rows 1,534,695
#> • Pages 3,070
#> • Fields 20
#> 
#> • Modified ❯ 2025-01-30
openMain("prof_map")
#> ── Provider Profile ID Mapping Table ──
#> 
#>   Maps secondary/duplicate provider profile
#>   IDs to the primary ID in the Open Payments
#>   system. For multiple reasons, a small
#>   number of providers have more than one ID
#>   in the payments records.
#> 
#> • Rows 2,358
#> • Pages 5
#> • Fields 2
#> 
#> • Modified ❯ 2025-01-30
openMain("prof_entity")
#> ── Reporting Entity Profile Information ──
#> 
#>   This dataset is pre-filtered based on the
#>   most frequent searches of Open Payments
#>   data.
#> 
#> • Rows 2,843
#> • Pages 6
#> • Fields 9
#> 
#> • Modified ❯ 2025-01-30
openMain("prof_teach")
#> ── Teaching Hospital Profile Information ──
#> 
#>   This dataset is pre-filtered based on the
#>   most frequent searches of Open Payments
#>   data.
#> 
#> • Rows 1,440
#> • Pages 3
#> • Fields 12
#> 
#> • Modified ❯ 2025-01-30
openMain("dashboard")
#> ── Summary Dashboard - all Years ──
#> 
#>   Summary data.
#> 
#> • Rows 90
#> • Pages 1
#> • Fields 10
#> 
#> • Modified ❯ 2025-01-30
openMain("pay_state_total")
#> ── State Level Payment Total and Averages for all Years ──
#> 
#>   This dataset is pre-filtered based on the
#>   most frequent searches of Open Payments
#>   data.
#> 
#> • Rows 2,765
#> • Pages 6
#> • Fields 28
#> 
#> • Modified ❯ 2025-01-30
openMain("pay_state_group")
#> ── State Payment Totals and Averages Grouped by Nature of Payment for all Years ──
#> 
#>   Groups payment data by state, recipient
#>   type, and nature of payment.
#> 
#> • Rows 16,758
#> • Pages 34
#> • Fields 28
#> 
#> • Modified ❯ 2025-01-30
openMain("pay_nat_group")
#> ── National Level Payment Total and Averages by Provider Specialty for all Years ──
#> 
#>   This dataset is pre-filtered based on the
#>   most frequent searches of Open Payments
#>   data.
#> 
#> • Rows 4,188
#> • Pages 9
#> • Fields 21
#> 
#> • Modified ❯ 2025-01-30
openMain("pay_nat_total")
#> ── National Level Payment Total and Averages for all Years ──
#> 
#>   This dataset is pre-filtered based on the
#>   most frequent searches of Open Payments
#>   data.
#> 
#> • Rows 24
#> • Pages 1
#> • Fields 18
#> 
#> • Modified ❯ 2025-01-30
openTemp("general")
#> <openTemp>
#>  @ title      : chr "General Payment Data"
#>  @ description: chr "All general (non-research, non-ownership related) payments from the 2023 program year [January 1 – December 31, 2023]"
#>  @ contact    : chr "Open Payments (mailto:openpayments@cms.hhs.gov)"
#>  @ rows       : int 14607336
#>  @ pages      : int 29215
#>  @ fields     : chr [1:91] "change_type" "covered_recipient_type" ...
#>  @ years      : int [1:7] 2023 2022 2021 2020 2019 2018 2017
#>  @ endpoints  : tibble [7 × 4] (S3: tbl_df/tbl/data.frame)
#>  $ year      : int [1:7] 2023 2022 2021 2020 2019 2018 2017
#>  $ modified  : Date[1:7], format: "2025-01-30" "2025-01-30" ...
#>  $ identifier: chr [1:7] "fb3a65aa-c901-4a38-a813-b04b00dfa2a9" "df01c2f8-dc1f-4e79-96cb-8208beaf143c" "0380bbeb-aea1-58b6-b708-829f92a48202" "a08c4b30-5cf3-4948-ad40-36f404619019" ...
#>  $ download  : chr [1:7] "https://download.cms.gov/openpayments/PGYR2023_P01302025_01212025/OP_DTL_GNRL_PGYR2023_P01302025_01212025.csv" "https://download.cms.gov/openpayments/PGYR2022_P01302025_01212025/OP_DTL_GNRL_PGYR2022_P01302025_01212025.csv" "https://download.cms.gov/openpayments/PGYR2021_P01302025_01212025/OP_DTL_GNRL_PGYR2021_P01302025_01212025.csv" "https://download.cms.gov/openpayments/PGYR2020_P01302025_01212025/OP_DTL_GNRL_PGYR2020_P01302025_01212025.csv" ...
openTemp("ownership")
#> <openTemp>
#>  @ title      : chr "Ownership Payment Data"
#>  @ description: chr "Ownership Payment Data - Detailed Dataset 2023 Reporting Year\nAll ownership and investment payments from the 2"| __truncated__
#>  @ contact    : chr "Open Payments (mailto:openpayments@cms.hhs.gov)"
#>  @ rows       : int 4013
#>  @ pages      : int 9
#>  @ fields     : chr [1:30] "change_type" "physician_profile_id" "physician_npi" ...
#>  @ years      : int [1:7] 2023 2022 2021 2020 2019 2018 2017
#>  @ endpoints  : tibble [7 × 4] (S3: tbl_df/tbl/data.frame)
#>  $ year      : int [1:7] 2023 2022 2021 2020 2019 2018 2017
#>  $ modified  : Date[1:7], format: "2025-01-30" "2025-01-30" ...
#>  $ identifier: chr [1:7] "ac0bc85c-02e3-45d9-89e8-2ff43da85df7" "37792388-800f-427a-9e02-b11601454eeb" "b0c03b8d-06df-58f2-8ce2-4daeffee147e" "a9a0bf48-6b96-4589-b4c2-3c5dcfbeaca2" ...
#>  $ download  : chr [1:7] "https://download.cms.gov/openpayments/PGYR2023_P01302025_01212025/OP_DTL_OWNRSHP_PGYR2023_P01302025_01212025.csv" "https://download.cms.gov/openpayments/PGYR2022_P01302025_01212025/OP_DTL_OWNRSHP_PGYR2022_P01302025_01212025.csv" "https://download.cms.gov/openpayments/PGYR2021_P01302025_01212025/OP_DTL_OWNRSHP_PGYR2021_P01302025_01212025.csv" "https://download.cms.gov/openpayments/PGYR2020_P01302025_01212025/OP_DTL_OWNRSHP_PGYR2020_P01302025_01212025.csv" ...
openTemp("research")
#> <openTemp>
#>  @ title      : chr "Research Payment Data"
#>  @ description: chr "Research Payment Data – Detailed Dataset 2023 Reporting Year\nAll research-related payments from the 2023 progr"| __truncated__
#>  @ contact    : chr "Open Payments (mailto:openpayments@cms.hhs.gov)"
#>  @ rows       : int 1027926
#>  @ pages      : int 2056
#>  @ fields     : chr [1:252] "change_type" "covered_recipient_type" ...
#>  @ years      : int [1:7] 2023 2022 2021 2020 2019 2018 2017
#>  @ endpoints  : tibble [7 × 4] (S3: tbl_df/tbl/data.frame)
#>  $ year      : int [1:7] 2023 2022 2021 2020 2019 2018 2017
#>  $ modified  : Date[1:7], format: "2025-01-30" "2025-01-30" ...
#>  $ identifier: chr [1:7] "ec9521bf-9d97-4603-814c-f4132d34bc4f" "fdc3c773-018a-412c-8a81-d7b8a13a037b" "ce1d28dd-0094-5060-a036-580329439600" "9c248e7e-7c7f-478b-ab84-ce0919d72c1c" ...
#>  $ download  : chr [1:7] "https://download.cms.gov/openpayments/PGYR2023_P01302025_01212025/OP_DTL_RSRCH_PGYR2023_P01302025_01212025.csv" "https://download.cms.gov/openpayments/PGYR2022_P01302025_01212025/OP_DTL_RSRCH_PGYR2022_P01302025_01212025.csv" "https://download.cms.gov/openpayments/PGYR2021_P01302025_01212025/OP_DTL_RSRCH_PGYR2021_P01302025_01212025.csv" "https://download.cms.gov/openpayments/PGYR2020_P01302025_01212025/OP_DTL_RSRCH_PGYR2020_P01302025_01212025.csv" ...
openTemp("recipient_nature")
#> <openTemp>
#>  @ title      : chr "Payments Grouped by Covered Recipient and Nature of Payments"
#>  @ description: chr "This dataset is pre-filtered based on the most frequent searches of Open Payments data."
#>  @ contact    : chr "Open Payments (mailto:openpayments@cms.hhs.gov)"
#>  @ rows       : int 1184051
#>  @ pages      : int 2369
#>  @ fields     : chr [1:10] "recipient_id" "recipient_type" "covered_recipient_npi" ...
#>  @ years      : int [1:7] 2023 2022 2021 2020 2019 2018 2017
#>  @ endpoints  : tibble [7 × 4] (S3: tbl_df/tbl/data.frame)
#>  $ year      : int [1:7] 2023 2022 2021 2020 2019 2018 2017
#>  $ modified  : Date[1:7], format: "2025-01-30" "2025-01-30" ...
#>  $ identifier: chr [1:7] "4017acfa-0eeb-4b8a-b68c-3b41e779c4a2" "686b765e-f4b7-4d35-a050-fb5af926dbbe" "d35243be-55cb-5087-931e-a294e0e0917f" "1fd2be47-cb72-435a-b2df-142f38e99a05" ...
#>  $ download  : chr [1:7] "https://download.cms.gov/openpayments/SMRY_RPTS_P01302025_01212025/PBLCTN_SMRY_BY_CR_BY_NTR_OF_PYMT_PGYR2023_P0"| __truncated__ "https://download.cms.gov/openpayments/SMRY_RPTS_P01302025_01212025/PBLCTN_SMRY_BY_CR_BY_NTR_OF_PYMT_PGYR2022_P0"| __truncated__ "https://download.cms.gov/openpayments/SMRY_RPTS_P01302025_01212025/PBLCTN_SMRY_BY_CR_BY_NTR_OF_PYMT_PGYR2021_P0"| __truncated__ "https://download.cms.gov/openpayments/SMRY_RPTS_P01302025_01212025/PBLCTN_SMRY_BY_CR_BY_NTR_OF_PYMT_PGYR2020_P0"| __truncated__ ...
openTemp("recipient_entity")
#> <openTemp>
#>  @ title      : chr "Payments Grouped by Covered Recipient and Reporting Entities"
#>  @ description: chr "This dataset is pre-filtered based on the most frequent searches of Open Payments data."
#>  @ contact    : chr "Open Payments (mailto:openpayments@cms.hhs.gov)"
#>  @ rows       : int 4416875
#>  @ pages      : int 8834
#>  @ fields     : chr [1:12] "recipient_id" "recipient_type" "covered_recipient_npi" ...
#>  @ years      : int [1:7] 2023 2022 2021 2020 2019 2018 2017
#>  @ endpoints  : tibble [7 × 4] (S3: tbl_df/tbl/data.frame)
#>  $ year      : int [1:7] 2023 2022 2021 2020 2019 2018 2017
#>  $ modified  : Date[1:7], format: "2025-01-30" "2025-01-30" ...
#>  $ identifier: chr [1:7] "4fbf0094-3c7b-425a-aabe-5778052559ad" "d2d87f39-b117-4ac8-9f36-a088fb22a25b" "7c9ebb83-8f88-5462-ab7d-1925b0113cfc" "78ad9c8d-d7ae-4066-bf2d-a66e3c9d6e3b" ...
#>  $ download  : chr [1:7] "https://download.cms.gov/openpayments/SMRY_RPTS_P01302025_01212025/PBLCTN_SMRY_BY_CR_BY_AMGPO_PGYR2023_P01302025_01212025.csv" "https://download.cms.gov/openpayments/SMRY_RPTS_P01302025_01212025/PBLCTN_SMRY_BY_CR_BY_AMGPO_PGYR2022_P01302025_01212025.csv" "https://download.cms.gov/openpayments/SMRY_RPTS_P01302025_01212025/PBLCTN_SMRY_BY_CR_BY_AMGPO_PGYR2021_P01302025_01212025.csv" "https://download.cms.gov/openpayments/SMRY_RPTS_P01302025_01212025/PBLCTN_SMRY_BY_CR_BY_AMGPO_PGYR2020_P01302025_01212025.csv" ...
openTemp("entity_nature")
#> <openTemp>
#>  @ title      : chr "Payments Grouped by Reporting Entities and Nature of Payments"
#>  @ description: chr "This dataset is pre-filtered based on the most frequent searches of Open Payments data."
#>  @ contact    : chr "Open Payments (mailto:openpayments@cms.hhs.gov)"
#>  @ rows       : int 6192
#>  @ pages      : int 13
#>  @ fields     : chr [1:5] "amgpo_id" "nature_of_payment_type_code" ...
#>  @ years      : int [1:7] 2023 2022 2021 2020 2019 2018 2017
#>  @ endpoints  : tibble [7 × 4] (S3: tbl_df/tbl/data.frame)
#>  $ year      : int [1:7] 2023 2022 2021 2020 2019 2018 2017
#>  $ modified  : Date[1:7], format: "2025-01-30" "2025-01-30" ...
#>  $ identifier: chr [1:7] "72008dab-0953-4226-a4cd-9f1872e8170d" "cedcd327-4e5d-43f9-8eb1-c11850fa7c66" "a2714020-f599-52b0-b12b-8f5b510e547a" "7dfe6254-2aee-454f-9ea5-df611a9a7b55" ...
#>  $ download  : chr [1:7] "https://download.cms.gov/openpayments/SMRY_RPTS_P01302025_01212025/PBLCTN_SMRY_BY_AMGPO_BY_NTR_OF_PYMT_PGYR2023"| __truncated__ "https://download.cms.gov/openpayments/SMRY_RPTS_P01302025_01212025/PBLCTN_SMRY_BY_AMGPO_BY_NTR_OF_PYMT_PGYR2022"| __truncated__ "https://download.cms.gov/openpayments/SMRY_RPTS_P01302025_01212025/PBLCTN_SMRY_BY_AMGPO_BY_NTR_OF_PYMT_PGYR2021"| __truncated__ "https://download.cms.gov/openpayments/SMRY_RPTS_P01302025_01212025/PBLCTN_SMRY_BY_AMGPO_BY_NTR_OF_PYMT_PGYR2020"| __truncated__ ...
openTemp("entity_recipient_nature")
#> <openTemp>
#>  @ title      : chr "Payments Grouped by Reporting Entities, Covered Recipient, and Nature of Payments"
#>  @ description: chr "This dataset is pre-filtered based on the most frequent searches of Open Payments data."
#>  @ contact    : chr "Open Payments (mailto:openpayments@cms.hhs.gov)"
#>  @ rows       : int 4588015
#>  @ pages      : int 9177
#>  @ fields     : chr [1:8] "amgpo_id" "recipient_id" "recipient_type" ...
#>  @ years      : int [1:7] 2023 2022 2021 2020 2019 2018 2017
#>  @ endpoints  : tibble [7 × 4] (S3: tbl_df/tbl/data.frame)
#>  $ year      : int [1:7] 2023 2022 2021 2020 2019 2018 2017
#>  $ modified  : Date[1:7], format: "2025-01-30" "2025-01-30" ...
#>  $ identifier: chr [1:7] "29c23b6a-dd1e-4d82-a606-8d792c7f229e" "171e369a-6f25-49eb-aa36-37f0a497e5ca" "f240c437-9d8a-5be7-95a5-8ce8e0d970bf" "532df191-473f-4bcc-9056-a6260b9437b5" ...
#>  $ download  : chr [1:7] "https://download.cms.gov/openpayments/SMRY_RPTS_P01302025_01212025/PBLCTN_SMRY_BY_AMGPO_BY_CR_BY_NTR_OF_PYMT_PG"| __truncated__ "https://download.cms.gov/openpayments/SMRY_RPTS_P01302025_01212025/PBLCTN_SMRY_BY_AMGPO_BY_CR_BY_NTR_OF_PYMT_PG"| __truncated__ "https://download.cms.gov/openpayments/SMRY_RPTS_P01302025_01212025/PBLCTN_SMRY_BY_AMGPO_BY_CR_BY_NTR_OF_PYMT_PG"| __truncated__ "https://download.cms.gov/openpayments/SMRY_RPTS_P01302025_01212025/PBLCTN_SMRY_BY_AMGPO_BY_CR_BY_NTR_OF_PYMT_PG"| __truncated__ ...
openTemp("state_nature")
#> <openTemp>
#>  @ title      : chr "State Payment Totals Grouped by Nature of Payment for all Years"
#>  @ description: chr "Groups payment data by state, recipient type, and nature of payment."
#>  @ contact    : chr "Open Payments (mailto:openpayments@cms.hhs.gov)"
#>  @ rows       : int 16758
#>  @ pages      : int 34
#>  @ fields     : chr [1:28] "country_code" "country_name" "state_code" "state_name" ...
#>  @ years      : int [1:7] 2023 2022 2021 2020 2019 2018 2017
#>  @ endpoints  : tibble [7 × 4] (S3: tbl_df/tbl/data.frame)
#>  $ year      : int [1:7] 2023 2022 2021 2020 2019 2018 2017
#>  $ modified  : Date[1:7], format: "2025-01-30" "2025-01-30" ...
#>  $ identifier: chr [1:7] "f83aed3a-7504-4fca-aabb-ac0338c79688" "c64c40af-23a5-4419-8b64-692684483664" "7fad5b56-2cfe-52ab-ac4d-3200780287d4" "7f3055b6-73d8-5aff-be15-55873552e2cd" ...
#>  $ download  : chr [1:7] "https://download.cms.gov/openpayments/SMRY_RPTS_P01302025_01212025/PBLCTN_NTR_OF_PYMT_BY_STATE_SMRY_PGYR2023_P0"| __truncated__ "https://download.cms.gov/openpayments/SMRY_RPTS_P01302025_01212025/PBLCTN_NTR_OF_PYMT_BY_STATE_SMRY_PGYR2022_P0"| __truncated__ "https://download.cms.gov/openpayments/SMRY_RPTS_P01302025_01212025/PBLCTN_NTR_OF_PYMT_BY_STATE_SMRY_PGYR2021_P0"| __truncated__ "https://download.cms.gov/openpayments/SMRY_RPTS_P01302025_01212025/PBLCTN_NTR_OF_PYMT_BY_STATE_SMRY_PGYR2020_P0"| __truncated__ ...