Main API Endpoint Temporal Group

Main API Endpoint Temporal Group

Usage

TemporalMainGroup(alias)

Arguments

alias

<chr> endpoint alias

Value

An S7 <TemporalMainGroup> object.

Examples

TemporalMainGroup("utilization")
#> <TemporalMainGroup>
#>  @ title      : chr "Medicare Physician & Other Practitioners"
#>  @ contact    : chr "Medicare Provider Data - OEDA (mailto:IPAG_Data_Products@cms.hhs.gov)"
#>  @ endpoints  :List of 3
#>  .. $ by_geography_and_service:List of 7
#>  ..  ..$ description: chr "The Medicare Physician & Other Practitioners by Geography and Service dataset contains information on use, paym"| __truncated__
#>  ..  ..$ dictionary : chr "https://data.cms.gov/resources/medicare-physician-other-practitioners-by-geography-and-service-data-dictionary"
#>  ..  ..$ site       : chr "https://data.cms.gov/provider-summary-by-type-of-service/medicare-physician-other-practitioners/medicare-physic"| __truncated__
#>  ..  ..$ rows       : int 270673
#>  ..  ..$ pages      : int 55
#>  ..  ..$ fields     : chr [1:15] "Rndrng_Prvdr_Geo_Lvl" "Rndrng_Prvdr_Geo_Cd" "Rndrng_Prvdr_Geo_Desc" "HCPCS_Cd" ...
#>  ..  ..$ endpoints  : tibble [10 × 3] (S3: tbl_df/tbl/data.frame)
#>  ..  .. ..$ year      : int [1:10] 2022 2021 2020 2019 2018 2017 2016 2015 2014 2013
#>  ..  .. ..$ identifier: chr [1:10] "https://data.cms.gov/data-api/v1/dataset/87304f15-9ed0-41dc-a141-6141a0327453/data" "https://data.cms.gov/data-api/v1/dataset/f00f462f-bb61-48b1-a321-1c50d78ce175/data" "https://data.cms.gov/data-api/v1/dataset/31eb3018-43e0-4259-a0d8-e7a1112ffd08/data" "https://data.cms.gov/data-api/v1/dataset/673030ae-ceed-4561-8fca-b1275395a86a/data" ...
#>  ..  .. ..$ resources : chr [1:10] "https://data.cms.gov/data-api/v1/dataset-resources/87304f15-9ed0-41dc-a141-6141a0327453" "https://data.cms.gov/data-api/v1/dataset-resources/f00f462f-bb61-48b1-a321-1c50d78ce175" "https://data.cms.gov/data-api/v1/dataset-resources/31eb3018-43e0-4259-a0d8-e7a1112ffd08" "https://data.cms.gov/data-api/v1/dataset-resources/673030ae-ceed-4561-8fca-b1275395a86a" ...
#>  .. $ by_provider             :List of 7
#>  ..  ..$ description: chr "The Medicare Physician & Other Practitioners by Provider dataset provides information on use, payments, submitt"| __truncated__
#>  ..  ..$ dictionary : chr "https://data.cms.gov/resources/medicare-physician-other-practitioners-by-provider-data-dictionary"
#>  ..  ..$ site       : chr "https://data.cms.gov/provider-summary-by-type-of-service/medicare-physician-other-practitioners/medicare-physic"| __truncated__
#>  ..  ..$ rows       : int 1230293
#>  ..  ..$ pages      : int 247
#>  ..  ..$ fields     : chr [1:82] "Rndrng_NPI" "Rndrng_Prvdr_Last_Org_Name" "Rndrng_Prvdr_First_Name" "Rndrng_Prvdr_MI" ...
#>  ..  ..$ endpoints  : tibble [10 × 3] (S3: tbl_df/tbl/data.frame)
#>  ..  .. ..$ year      : int [1:10] 2022 2021 2020 2019 2018 2017 2016 2015 2014 2013
#>  ..  .. ..$ identifier: chr [1:10] "https://data.cms.gov/data-api/v1/dataset/8568be52-dfc9-4fe0-b8d7-95838b3cd0c6/data" "https://data.cms.gov/data-api/v1/dataset/1a3a5f39-09ac-48e2-b9bf-b371ac01c07e/data" "https://data.cms.gov/data-api/v1/dataset/6c0feedd-12a5-401c-bd35-15fabcd6a334/data" "https://data.cms.gov/data-api/v1/dataset/6859b742-ef67-4930-8262-03fb341001e2/data" ...
#>  ..  .. ..$ resources : chr [1:10] "https://data.cms.gov/data-api/v1/dataset-resources/8568be52-dfc9-4fe0-b8d7-95838b3cd0c6" "https://data.cms.gov/data-api/v1/dataset-resources/1a3a5f39-09ac-48e2-b9bf-b371ac01c07e" "https://data.cms.gov/data-api/v1/dataset-resources/6c0feedd-12a5-401c-bd35-15fabcd6a334" "https://data.cms.gov/data-api/v1/dataset-resources/6859b742-ef67-4930-8262-03fb341001e2" ...
#>  .. $ by_provider_and_service :List of 7
#>  ..  ..$ description: chr "The Medicare Physician & Other Practitioners by Provider and Service dataset provides information on use, payme"| __truncated__
#>  ..  ..$ dictionary : chr "https://data.cms.gov/resources/medicare-physician-other-practitioners-by-provider-and-service-data-dictionary"
#>  ..  ..$ site       : chr "https://data.cms.gov/provider-summary-by-type-of-service/medicare-physician-other-practitioners/medicare-physic"| __truncated__
#>  ..  ..$ rows       : int 9755427
#>  ..  ..$ pages      : int 1952
#>  ..  ..$ fields     : chr [1:29] "Rndrng_NPI" "Rndrng_Prvdr_Last_Org_Name" "Rndrng_Prvdr_First_Name" "Rndrng_Prvdr_MI" ...
#>  ..  ..$ endpoints  : tibble [10 × 3] (S3: tbl_df/tbl/data.frame)
#>  ..  .. ..$ year      : int [1:10] 2022 2021 2020 2019 2018 2017 2016 2015 2014 2013
#>  ..  .. ..$ identifier: chr [1:10] "https://data.cms.gov/data-api/v1/dataset/4f307be4-6868-4a9e-ae92-acf3fd4b5543/data" "https://data.cms.gov/data-api/v1/dataset/5c67d835-3862-4f63-897d-85d3eac82d5b/data" "https://data.cms.gov/data-api/v1/dataset/862ed658-1f38-4b2f-b02b-0b359e12c78a/data" "https://data.cms.gov/data-api/v1/dataset/5fccd951-9538-48a7-9075-6f02b9867868/data" ...
#>  ..  .. ..$ resources : chr [1:10] "https://data.cms.gov/data-api/v1/dataset-resources/4f307be4-6868-4a9e-ae92-acf3fd4b5543" "https://data.cms.gov/data-api/v1/dataset-resources/5c67d835-3862-4f63-897d-85d3eac82d5b" "https://data.cms.gov/data-api/v1/dataset-resources/862ed658-1f38-4b2f-b02b-0b359e12c78a" "https://data.cms.gov/data-api/v1/dataset-resources/5fccd951-9538-48a7-9075-6f02b9867868" ...
#>  @ periodicity: chr "Annually [R/P1Y]"
#>  @ years      : int [1:10] 2022 2021 2020 2019 2018 2017 2016 2015 2014 2013
TemporalMainGroup("prescribers")
#> <TemporalMainGroup>
#>  @ title      : chr "Medicare Part D Prescribers"
#>  @ contact    : chr "Medicare Provider Data - OEDA (mailto:IPAG_Data_Products@cms.hhs.gov)"
#>  @ endpoints  :List of 3
#>  .. $ by_geography_and_drug:List of 7
#>  ..  ..$ description: chr "The Medicare Part D Prescribers by Geography and Drug dataset contains information on prescription drugs prescr"| __truncated__
#>  ..  ..$ dictionary : chr "https://data.cms.gov/resources/medicare-part-d-prescribers-by-geography-and-drug-data-dictionary"
#>  ..  ..$ site       : chr "https://data.cms.gov/provider-summary-by-type-of-service/medicare-part-d-prescribers/medicare-part-d-prescriber"| __truncated__
#>  ..  ..$ rows       : int 115396
#>  ..  ..$ pages      : int 24
#>  ..  ..$ fields     : chr [1:22] "Prscrbr_Geo_Lvl" "Prscrbr_Geo_Cd" "Prscrbr_Geo_Desc" "Brnd_Name" ...
#>  ..  ..$ endpoints  : tibble [10 × 3] (S3: tbl_df/tbl/data.frame)
#>  ..  .. ..$ year      : int [1:10] 2022 2021 2020 2019 2018 2017 2016 2015 2014 2013
#>  ..  .. ..$ identifier: chr [1:10] "https://data.cms.gov/data-api/v1/dataset/1fc57194-a51d-4864-aee6-de0889488151/data" "https://data.cms.gov/data-api/v1/dataset/7dda2a9d-034a-446a-b4b3-e1254e0127b2/data" "https://data.cms.gov/data-api/v1/dataset/83891e77-99cf-4865-b60a-97703b916e09/data" "https://data.cms.gov/data-api/v1/dataset/73a6335e-f16f-4c81-a84b-6b5a986e2bf8/data" ...
#>  ..  .. ..$ resources : chr [1:10] "https://data.cms.gov/data-api/v1/dataset-resources/1fc57194-a51d-4864-aee6-de0889488151" "https://data.cms.gov/data-api/v1/dataset-resources/7dda2a9d-034a-446a-b4b3-e1254e0127b2" "https://data.cms.gov/data-api/v1/dataset-resources/83891e77-99cf-4865-b60a-97703b916e09" "https://data.cms.gov/data-api/v1/dataset-resources/73a6335e-f16f-4c81-a84b-6b5a986e2bf8" ...
#>  .. $ by_provider          :List of 7
#>  ..  ..$ description: chr "The Medicare Part D Prescribers by Provider dataset contains information on prescription drugs prescribed by in"| __truncated__
#>  ..  ..$ dictionary : chr "https://data.cms.gov/resources/medicare-part-d-prescribers-by-provider-data-dictionary"
#>  ..  ..$ site       : chr "https://data.cms.gov/provider-summary-by-type-of-service/medicare-part-d-prescribers/medicare-part-d-prescribers-by-provider"
#>  ..  ..$ rows       : int 1332309
#>  ..  ..$ pages      : int 267
#>  ..  ..$ fields     : chr [1:85] "Prscrbr_NPI" "Prscrbr_Last_Org_Name" "Prscrbr_First_Name" "Prscrbr_MI" ...
#>  ..  ..$ endpoints  : tibble [10 × 3] (S3: tbl_df/tbl/data.frame)
#>  ..  .. ..$ year      : int [1:10] 2022 2021 2020 2019 2018 2017 2016 2015 2014 2013
#>  ..  .. ..$ identifier: chr [1:10] "https://data.cms.gov/data-api/v1/dataset/5b7298b5-4a3b-4517-86b0-2c9077c64a86/data" "https://data.cms.gov/data-api/v1/dataset/c55024e4-4b4f-4106-b570-0cc05f173357/data" "https://data.cms.gov/data-api/v1/dataset/eca92c9a-cc1b-4d6e-9742-7cf34ed6f1f4/data" "https://data.cms.gov/data-api/v1/dataset/b6c4cbfd-0f23-4a38-8bc0-41a26fc71826/data" ...
#>  ..  .. ..$ resources : chr [1:10] "https://data.cms.gov/data-api/v1/dataset-resources/5b7298b5-4a3b-4517-86b0-2c9077c64a86" "https://data.cms.gov/data-api/v1/dataset-resources/c55024e4-4b4f-4106-b570-0cc05f173357" "https://data.cms.gov/data-api/v1/dataset-resources/eca92c9a-cc1b-4d6e-9742-7cf34ed6f1f4" "https://data.cms.gov/data-api/v1/dataset-resources/b6c4cbfd-0f23-4a38-8bc0-41a26fc71826" ...
#>  .. $ by_provider_and_drug :List of 7
#>  ..  ..$ description: chr "The Medicare Part D Prescribers by Provider and Drug dataset provides information on prescription drugs prescri"| __truncated__
#>  ..  ..$ dictionary : chr "https://data.cms.gov/resources/medicare-part-d-prescribers-by-provider-and-drug-data-dictionary"
#>  ..  ..$ site       : chr "https://data.cms.gov/provider-summary-by-type-of-service/medicare-part-d-prescribers/medicare-part-d-prescriber"| __truncated__
#>  ..  ..$ rows       : int 25869521
#>  ..  ..$ pages      : int 5174
#>  ..  ..$ fields     : chr [1:22] "Prscrbr_NPI" "Prscrbr_Last_Org_Name" "Prscrbr_First_Name" "Prscrbr_City" ...
#>  ..  ..$ endpoints  : tibble [10 × 3] (S3: tbl_df/tbl/data.frame)
#>  ..  .. ..$ year      : int [1:10] 2022 2021 2020 2019 2018 2017 2016 2015 2014 2013
#>  ..  .. ..$ identifier: chr [1:10] "https://data.cms.gov/data-api/v1/dataset/b101b457-ffa4-49bb-8fd9-27c1266086e2/data" "https://data.cms.gov/data-api/v1/dataset/f68114ed-f854-4ffc-9c6e-ed78b5e2f8d0/data" "https://data.cms.gov/data-api/v1/dataset/7795fe20-e80e-435a-a9ed-d2d65e05feeb/data" "https://data.cms.gov/data-api/v1/dataset/2a6705e6-7a1e-460c-ba22-35249a531918/data" ...
#>  ..  .. ..$ resources : chr [1:10] "https://data.cms.gov/data-api/v1/dataset-resources/b101b457-ffa4-49bb-8fd9-27c1266086e2" "https://data.cms.gov/data-api/v1/dataset-resources/f68114ed-f854-4ffc-9c6e-ed78b5e2f8d0" "https://data.cms.gov/data-api/v1/dataset-resources/7795fe20-e80e-435a-a9ed-d2d65e05feeb" "https://data.cms.gov/data-api/v1/dataset-resources/2a6705e6-7a1e-460c-ba22-35249a531918" ...
#>  @ periodicity: chr "Annually [R/P1Y]"
#>  @ years      : int [1:10] 2022 2021 2020 2019 2018 2017 2016 2015 2014 2013
TemporalMainGroup("suppliers")
#> <TemporalMainGroup>
#>  @ title      : chr "Medicare Durable Medical Equipment, Devices & Supplies"
#>  @ contact    : chr "Medicare Provider Data - OEDA (mailto:IPAG_Data_Products@cms.hhs.gov)"
#>  @ endpoints  :List of 5
#>  .. $ by_geography_and_service         :List of 7
#>  ..  ..$ description: chr "The Medicare Durable Medical Equipment, Devices & Supplies by Geography and Service dataset contains informatio"| __truncated__
#>  ..  ..$ dictionary : chr "https://data.cms.gov/resources/medicare-durable-medical-equipment-devices-supplies-by-geography-and-service-dat"| __truncated__
#>  ..  ..$ site       : chr "https://data.cms.gov/provider-summary-by-type-of-service/medicare-durable-medical-equipment-devices-supplies/me"| __truncated__
#>  ..  ..$ rows       : int 39129
#>  ..  ..$ pages      : int 8
#>  ..  ..$ fields     : chr [1:18] "Rfrg_Prvdr_Geo_Lvl" "Rfrg_Prvdr_Geo_Cd" "Rfrg_Prvdr_Geo_Desc" "RBCS_Lvl" ...
#>  ..  ..$ endpoints  : tibble [9 × 3] (S3: tbl_df/tbl/data.frame)
#>  ..  .. ..$ year      : int [1:9] 2022 2021 2020 2019 2018 2017 2016 2015 2014
#>  ..  .. ..$ identifier: chr [1:9] "https://data.cms.gov/data-api/v1/dataset/a9d1461a-9c63-49e1-8d91-696082e20e47/data" "https://data.cms.gov/data-api/v1/dataset/daedccd6-5881-4997-93cb-ad6f3d1dd2fd/data" "https://data.cms.gov/data-api/v1/dataset/c5da241c-71d2-4b04-9f44-b324ff4b4c01/data" "https://data.cms.gov/data-api/v1/dataset/3931609c-3f1d-4ab4-95f9-0e31d43613f0/data" ...
#>  ..  .. ..$ resources : chr [1:9] "https://data.cms.gov/data-api/v1/dataset-resources/a9d1461a-9c63-49e1-8d91-696082e20e47" "https://data.cms.gov/data-api/v1/dataset-resources/daedccd6-5881-4997-93cb-ad6f3d1dd2fd" "https://data.cms.gov/data-api/v1/dataset-resources/c5da241c-71d2-4b04-9f44-b324ff4b4c01" "https://data.cms.gov/data-api/v1/dataset-resources/3931609c-3f1d-4ab4-95f9-0e31d43613f0" ...
#>  .. $ by_referring_provider            :List of 7
#>  ..  ..$ description: chr "The Medicare Durable Medical Equipment, Devices & Supplies by Referring Provider dataset contains information o"| __truncated__
#>  ..  ..$ dictionary : chr "https://data.cms.gov/resources/medicare-durable-medical-equipment-devices-supplies-by-referring-provider-data-dictionary"
#>  ..  ..$ site       : chr "https://data.cms.gov/provider-summary-by-type-of-service/medicare-durable-medical-equipment-devices-supplies/me"| __truncated__
#>  ..  ..$ rows       : int 385637
#>  ..  ..$ pages      : int 78
#>  ..  ..$ fields     : chr [1:98] "Rfrg_NPI" "Rfrg_Prvdr_Last_Name_Org" "Rfrg_Prvdr_First_Name" "Rfrg_Prvdr_MI" ...
#>  ..  ..$ endpoints  : tibble [9 × 3] (S3: tbl_df/tbl/data.frame)
#>  ..  .. ..$ year      : int [1:9] 2022 2021 2020 2019 2018 2017 2016 2015 2014
#>  ..  .. ..$ identifier: chr [1:9] "https://data.cms.gov/data-api/v1/dataset/e0eb5710-5287-431b-984e-a5629556bbf2/data" "https://data.cms.gov/data-api/v1/dataset/527aab51-3316-4cb1-8740-a5e85a6670f2/data" "https://data.cms.gov/data-api/v1/dataset/5fb6377b-41b9-4002-828f-d257bccff371/data" "https://data.cms.gov/data-api/v1/dataset/c8cd9ce3-c005-4823-ad4d-b98bcd371f2f/data" ...
#>  ..  .. ..$ resources : chr [1:9] "https://data.cms.gov/data-api/v1/dataset-resources/e0eb5710-5287-431b-984e-a5629556bbf2" "https://data.cms.gov/data-api/v1/dataset-resources/527aab51-3316-4cb1-8740-a5e85a6670f2" "https://data.cms.gov/data-api/v1/dataset-resources/5fb6377b-41b9-4002-828f-d257bccff371" "https://data.cms.gov/data-api/v1/dataset-resources/c8cd9ce3-c005-4823-ad4d-b98bcd371f2f" ...
#>  .. $ by_referring_provider_and_service:List of 7
#>  ..  ..$ description: chr "The Medicare Durable Medical Equipment, Devices & Supplies by Referring Provider and Service dataset contains i"| __truncated__
#>  ..  ..$ dictionary : chr "https://data.cms.gov/resources/medicare-durable-medical-equipment-devices-supplies-by-referring-provider-and-se"| __truncated__
#>  ..  ..$ site       : chr "https://data.cms.gov/provider-summary-by-type-of-service/medicare-durable-medical-equipment-devices-supplies/me"| __truncated__
#>  ..  ..$ rows       : int 1457378
#>  ..  ..$ pages      : int 292
#>  ..  ..$ fields     : chr [1:34] "Rfrg_NPI" "Rfrg_Prvdr_Last_Name_Org" "Rfrg_Prvdr_First_Name" "Rfrg_Prvdr_MI" ...
#>  ..  ..$ endpoints  : tibble [9 × 3] (S3: tbl_df/tbl/data.frame)
#>  ..  .. ..$ year      : int [1:9] 2022 2021 2020 2019 2018 2017 2016 2015 2014
#>  ..  .. ..$ identifier: chr [1:9] "https://data.cms.gov/data-api/v1/dataset/7f98d91f-f2a6-44cd-a57f-1ac02eaaf795/data" "https://data.cms.gov/data-api/v1/dataset/442050e6-3d96-47b9-bf0e-d894f74a137f/data" "https://data.cms.gov/data-api/v1/dataset/72ce1cd5-efa9-484e-bd74-90ab3a10d244/data" "https://data.cms.gov/data-api/v1/dataset/5d2c51d1-50cb-4b10-8930-3e40b7561060/data" ...
#>  ..  .. ..$ resources : chr [1:9] "https://data.cms.gov/data-api/v1/dataset-resources/7f98d91f-f2a6-44cd-a57f-1ac02eaaf795" "https://data.cms.gov/data-api/v1/dataset-resources/442050e6-3d96-47b9-bf0e-d894f74a137f" "https://data.cms.gov/data-api/v1/dataset-resources/72ce1cd5-efa9-484e-bd74-90ab3a10d244" "https://data.cms.gov/data-api/v1/dataset-resources/5d2c51d1-50cb-4b10-8930-3e40b7561060" ...
#>  .. $ by_supplier                      :List of 7
#>  ..  ..$ description: chr "The Medicare Durable Medical Equipment, Devices & Supplies by Supplier dataset contains information on usage, p"| __truncated__
#>  ..  ..$ dictionary : chr "https://data.cms.gov/resources/medicare-durable-medical-equipment-devices-supplies-by-supplier-data-dictionary"
#>  ..  ..$ site       : chr "https://data.cms.gov/provider-summary-by-type-of-service/medicare-durable-medical-equipment-devices-supplies/me"| __truncated__
#>  ..  ..$ rows       : int 66406
#>  ..  ..$ pages      : int 14
#>  ..  ..$ fields     : chr [1:94] "Suplr_NPI" "Suplr_Prvdr_Last_Name_Org" "Suplr_Prvdr_First_Name" "Suplr_Prvdr_MI" ...
#>  ..  ..$ endpoints  : tibble [9 × 3] (S3: tbl_df/tbl/data.frame)
#>  ..  .. ..$ year      : int [1:9] 2022 2021 2020 2019 2018 2017 2016 2015 2014
#>  ..  .. ..$ identifier: chr [1:9] "https://data.cms.gov/data-api/v1/dataset/70b8ae50-1919-4468-9247-022cffd5e2b0/data" "https://data.cms.gov/data-api/v1/dataset/107436f3-cc94-4469-87ea-9c7f3d196124/data" "https://data.cms.gov/data-api/v1/dataset/40c9de9d-18a3-4476-ba08-761056e04f8a/data" "https://data.cms.gov/data-api/v1/dataset/ce5ee6de-c08b-4cf4-928d-17f29bbc2998/data" ...
#>  ..  .. ..$ resources : chr [1:9] "https://data.cms.gov/data-api/v1/dataset-resources/70b8ae50-1919-4468-9247-022cffd5e2b0" "https://data.cms.gov/data-api/v1/dataset-resources/107436f3-cc94-4469-87ea-9c7f3d196124" "https://data.cms.gov/data-api/v1/dataset-resources/40c9de9d-18a3-4476-ba08-761056e04f8a" "https://data.cms.gov/data-api/v1/dataset-resources/ce5ee6de-c08b-4cf4-928d-17f29bbc2998" ...
#>  .. $ by_supplier_and_service          :List of 7
#>  ..  ..$ description: chr "The Medicare Durable Medical Equipment, Devices & Supplies by Supplier and Service dataset contains information"| __truncated__
#>  ..  ..$ dictionary : chr "https://data.cms.gov/resources/medicare-durable-medical-equipment-devices-supplies-by-supplier-and-service-data-dictionary"
#>  ..  ..$ site       : chr "https://data.cms.gov/provider-summary-by-type-of-service/medicare-durable-medical-equipment-devices-supplies/me"| __truncated__
#>  ..  ..$ rows       : int 482638
#>  ..  ..$ pages      : int 97
#>  ..  ..$ fields     : chr [1:33] "Suplr_NPI" "Suplr_Prvdr_Last_Name_Org" "Suplr_Prvdr_First_Name" "Suplr_Prvdr_MI" ...
#>  ..  ..$ endpoints  : tibble [9 × 3] (S3: tbl_df/tbl/data.frame)
#>  ..  .. ..$ year      : int [1:9] 2022 2021 2020 2019 2018 2017 2016 2015 2014
#>  ..  .. ..$ identifier: chr [1:9] "https://data.cms.gov/data-api/v1/dataset/ce9ec4d2-d404-4463-98ef-aa5512f201e2/data" "https://data.cms.gov/data-api/v1/dataset/1ca065e5-adfa-4a43-9bf3-4089538571c9/data" "https://data.cms.gov/data-api/v1/dataset/6160f96b-710a-4380-a3ff-11324c2d5450/data" "https://data.cms.gov/data-api/v1/dataset/ea4a5834-aeaa-4176-8feb-976ae7874890/data" ...
#>  ..  .. ..$ resources : chr [1:9] "https://data.cms.gov/data-api/v1/dataset-resources/ce9ec4d2-d404-4463-98ef-aa5512f201e2" "https://data.cms.gov/data-api/v1/dataset-resources/1ca065e5-adfa-4a43-9bf3-4089538571c9" "https://data.cms.gov/data-api/v1/dataset-resources/6160f96b-710a-4380-a3ff-11324c2d5450" "https://data.cms.gov/data-api/v1/dataset-resources/ea4a5834-aeaa-4176-8feb-976ae7874890" ...
#>  @ periodicity: chr "Annually [R/P1Y]"
#>  @ years      : int [1:9] 2022 2021 2020 2019 2018 2017 2016 2015 2014
TemporalMainGroup("outpatient")
#> <TemporalMainGroup>
#>  @ title      : chr "Medicare Outpatient Hospitals"
#>  @ contact    : chr "Medicare Provider Data - OEDA (mailto:IPAG_Data_Products@cms.hhs.gov)"
#>  @ endpoints  :List of 2
#>  .. $ by_geography_and_service:List of 7
#>  ..  ..$ description: chr "The Medicare Outpatient Hospitals by Geography and Service dataset provides information on services for Origina"| __truncated__
#>  ..  ..$ dictionary : chr "https://data.cms.gov/resources/medicare-outpatient-hospitals-by-geography-and-service-data-dictionary"
#>  ..  ..$ site       : chr "https://data.cms.gov/provider-summary-by-type-of-service/medicare-outpatient-hospitals/medicare-outpatient-hosp"| __truncated__
#>  ..  ..$ rows       : int 90591
#>  ..  ..$ pages      : int 19
#>  ..  ..$ fields     : chr [1:15] "Rndrng_Prvdr_Geo_Lvl" "Rndrng_Prvdr_Geo_Cd" "Rndrng_Prvdr_Geo_Desc" "Srvc_Lvl" ...
#>  ..  ..$ endpoints  : tibble [8 × 3] (S3: tbl_df/tbl/data.frame)
#>  ..  .. ..$ year      : int [1:8] 2022 2021 2020 2019 2018 2017 2016 2015
#>  ..  .. ..$ identifier: chr [1:8] "https://data.cms.gov/data-api/v1/dataset/6af002ba-b934-4491-96dd-d041989b843c/data" "https://data.cms.gov/data-api/v1/dataset/b3537e90-c3c5-4895-914d-71f54123a0a4/data" "https://data.cms.gov/data-api/v1/dataset/31db4293-f75c-4940-bfae-32d554cd164f/data" "https://data.cms.gov/data-api/v1/dataset/46ae348d-8ba4-43df-868e-74ff87eada39/data" ...
#>  ..  .. ..$ resources : chr [1:8] "https://data.cms.gov/data-api/v1/dataset-resources/6af002ba-b934-4491-96dd-d041989b843c" "https://data.cms.gov/data-api/v1/dataset-resources/b3537e90-c3c5-4895-914d-71f54123a0a4" "https://data.cms.gov/data-api/v1/dataset-resources/31db4293-f75c-4940-bfae-32d554cd164f" "https://data.cms.gov/data-api/v1/dataset-resources/46ae348d-8ba4-43df-868e-74ff87eada39" ...
#>  .. $ by_provider_and_service :List of 7
#>  ..  ..$ description: chr "The Medicare Outpatient Hospitals by Provider and Service dataset provides information on services for Original"| __truncated__
#>  ..  ..$ dictionary : chr "https://data.cms.gov/resources/medicare-outpatient-hospitals-by-provider-and-service-data-dictionary"
#>  ..  ..$ site       : chr "https://data.cms.gov/provider-summary-by-type-of-service/medicare-outpatient-hospitals/medicare-outpatient-hosp"| __truncated__
#>  ..  ..$ rows       : int 116459
#>  ..  ..$ pages      : int 24
#>  ..  ..$ fields     : chr [1:18] "Rndrng_Prvdr_CCN" "Rndrng_Prvdr_Org_Name" "Rndrng_Prvdr_St" "Rndrng_Prvdr_City" ...
#>  ..  ..$ endpoints  : tibble [8 × 3] (S3: tbl_df/tbl/data.frame)
#>  ..  .. ..$ year      : int [1:8] 2022 2021 2020 2019 2018 2017 2016 2015
#>  ..  .. ..$ identifier: chr [1:8] "https://data.cms.gov/data-api/v1/dataset/456dbe68-8878-49dd-889b-bb8bf432fcbe/data" "https://data.cms.gov/data-api/v1/dataset/a7617aef-046f-4557-8a2a-54f9c9a88b57/data" "https://data.cms.gov/data-api/v1/dataset/8281e9f7-1874-4bd9-a183-a74bec2debd3/data" "https://data.cms.gov/data-api/v1/dataset/f3fc4327-f66e-49e1-ba06-f2607a164406/data" ...
#>  ..  .. ..$ resources : chr [1:8] "https://data.cms.gov/data-api/v1/dataset-resources/456dbe68-8878-49dd-889b-bb8bf432fcbe" "https://data.cms.gov/data-api/v1/dataset-resources/a7617aef-046f-4557-8a2a-54f9c9a88b57" "https://data.cms.gov/data-api/v1/dataset-resources/8281e9f7-1874-4bd9-a183-a74bec2debd3" "https://data.cms.gov/data-api/v1/dataset-resources/f3fc4327-f66e-49e1-ba06-f2607a164406" ...
#>  @ periodicity: chr "Annually [R/P1Y]"
#>  @ years      : int [1:8] 2022 2021 2020 2019 2018 2017 2016 2015
TemporalMainGroup("inpatient")
#> <TemporalMainGroup>
#>  @ title      : chr "Medicare Inpatient Hospitals"
#>  @ contact    : chr "Medicare Provider Data - OEDA (mailto:IPAG_Data_Products@cms.hhs.gov)"
#>  @ endpoints  :List of 3
#>  .. $ by_geography_and_service:List of 7
#>  ..  ..$ description: chr "The Medicare Inpatient Hospitals by Geography and Service dataset provides information on hospital discharges f"| __truncated__
#>  ..  ..$ dictionary : chr "https://data.cms.gov/resources/medicare-inpatient-hospitals-by-geography-and-service-data-dictionary-0"
#>  ..  ..$ site       : chr "https://data.cms.gov/provider-summary-by-type-of-service/medicare-inpatient-hospitals/medicare-inpatient-hospit"| __truncated__
#>  ..  ..$ rows       : int 26260
#>  ..  ..$ pages      : int 6
#>  ..  ..$ fields     : chr [1:9] "Rndrng_Prvdr_Geo_Lvl" "Rndrng_Prvdr_Geo_Cd" "Rndrng_Prvdr_Geo_Desc" "DRG_Cd" ...
#>  ..  ..$ endpoints  : tibble [10 × 3] (S3: tbl_df/tbl/data.frame)
#>  ..  .. ..$ year      : int [1:10] 2022 2021 2020 2019 2018 2017 2016 2015 2014 2013
#>  ..  .. ..$ identifier: chr [1:10] "https://data.cms.gov/data-api/v1/dataset/5c8b315f-a2a3-4b2f-bba6-7be19f8d7edc/data" "https://data.cms.gov/data-api/v1/dataset/68739613-1a4f-40ae-9664-49da07abd81f/data" "https://data.cms.gov/data-api/v1/dataset/e583de38-33d7-4895-a115-0d8af74f0795/data" "https://data.cms.gov/data-api/v1/dataset/337617e5-c18b-4cb9-8152-34851747584b/data" ...
#>  ..  .. ..$ resources : chr [1:10] "https://data.cms.gov/data-api/v1/dataset-resources/5c8b315f-a2a3-4b2f-bba6-7be19f8d7edc" "https://data.cms.gov/data-api/v1/dataset-resources/68739613-1a4f-40ae-9664-49da07abd81f" "https://data.cms.gov/data-api/v1/dataset-resources/e583de38-33d7-4895-a115-0d8af74f0795" "https://data.cms.gov/data-api/v1/dataset-resources/337617e5-c18b-4cb9-8152-34851747584b" ...
#>  .. $ by_provider             :List of 7
#>  ..  ..$ description: chr "The Medicare Inpatient Hospitals by Provider dataset provides information on inpatient discharges for Original "| __truncated__
#>  ..  ..$ dictionary : chr "https://data.cms.gov/resources/medicare-inpatient-hospitals-by-provider-data-dictionary"
#>  ..  ..$ site       : chr "https://data.cms.gov/provider-summary-by-type-of-service/medicare-inpatient-hospitals/medicare-inpatient-hospitals-by-provider"
#>  ..  ..$ rows       : int 3120
#>  ..  ..$ pages      : int 1
#>  ..  ..$ fields     : chr [1:57] "Rndrng_Prvdr_CCN" "Rndrng_Prvdr_Org_Name" "Rndrng_Prvdr_St" "Rndrng_Prvdr_City" ...
#>  ..  ..$ endpoints  : tibble [10 × 3] (S3: tbl_df/tbl/data.frame)
#>  ..  .. ..$ year      : int [1:10] 2022 2021 2020 2019 2018 2017 2016 2015 2014 2013
#>  ..  .. ..$ identifier: chr [1:10] "https://data.cms.gov/data-api/v1/dataset/cc73e77b-7367-48c9-b766-fafefe3f8ffa/data" "https://data.cms.gov/data-api/v1/dataset/c7e1501c-a770-4e2a-beb6-32b3525491a7/data" "https://data.cms.gov/data-api/v1/dataset/d7b2810d-ac10-498f-8794-1c3d9f72ee96/data" "https://data.cms.gov/data-api/v1/dataset/c349419e-3c6a-419e-8c74-0641b64cca12/data" ...
#>  ..  .. ..$ resources : chr [1:10] "https://data.cms.gov/data-api/v1/dataset-resources/cc73e77b-7367-48c9-b766-fafefe3f8ffa" "https://data.cms.gov/data-api/v1/dataset-resources/c7e1501c-a770-4e2a-beb6-32b3525491a7" "https://data.cms.gov/data-api/v1/dataset-resources/d7b2810d-ac10-498f-8794-1c3d9f72ee96" "https://data.cms.gov/data-api/v1/dataset-resources/c349419e-3c6a-419e-8c74-0641b64cca12" ...
#>  .. $ by_provider_and_service :List of 7
#>  ..  ..$ description: chr "The Medicare Inpatient Hospitals by Provider and Service dataset provides information on inpatient discharges f"| __truncated__
#>  ..  ..$ dictionary : chr "https://data.cms.gov/resources/medicare-inpatient-hospitals-by-provider-and-service-data-dictionary-0"
#>  ..  ..$ site       : chr "https://data.cms.gov/provider-summary-by-type-of-service/medicare-inpatient-hospitals/medicare-inpatient-hospit"| __truncated__
#>  ..  ..$ rows       : int 145742
#>  ..  ..$ pages      : int 30
#>  ..  ..$ fields     : chr [1:15] "Rndrng_Prvdr_CCN" "Rndrng_Prvdr_Org_Name" "Rndrng_Prvdr_City" "Rndrng_Prvdr_St" ...
#>  ..  ..$ endpoints  : tibble [10 × 3] (S3: tbl_df/tbl/data.frame)
#>  ..  .. ..$ year      : int [1:10] 2022 2021 2020 2019 2018 2017 2016 2015 2014 2013
#>  ..  .. ..$ identifier: chr [1:10] "https://data.cms.gov/data-api/v1/dataset/46bf50f8-0983-4ca2-b8d5-f2afbbf2e589/data" "https://data.cms.gov/data-api/v1/dataset/117d93f2-ce81-40fe-a4d4-8c03203b95e1/data" "https://data.cms.gov/data-api/v1/dataset/ef5bdbe1-27b4-4296-b320-52bd5d2183d7/data" "https://data.cms.gov/data-api/v1/dataset/6f6d93e1-ecf8-4b93-9845-091faf20f274/data" ...
#>  ..  .. ..$ resources : chr [1:10] "https://data.cms.gov/data-api/v1/dataset-resources/46bf50f8-0983-4ca2-b8d5-f2afbbf2e589" "https://data.cms.gov/data-api/v1/dataset-resources/117d93f2-ce81-40fe-a4d4-8c03203b95e1" "https://data.cms.gov/data-api/v1/dataset-resources/ef5bdbe1-27b4-4296-b320-52bd5d2183d7" "https://data.cms.gov/data-api/v1/dataset-resources/6f6d93e1-ecf8-4b93-9845-091faf20f274" ...
#>  @ periodicity: chr "Annually [R/P1Y]"
#>  @ years      : int [1:10] 2022 2021 2020 2019 2018 2017 2016 2015 2014 2013