Provider Facility Affiliations

Facility affiliations data publicly reported in the Provider Data Catalog.

Usage

affiliations(
  npi = NULL,
  pac = NULL,
  last_name = NULL,
  first_name = NULL,
  middle_name = NULL,
  suffix = NULL,
  facility_type = NULL,
  ccn_facility = NULL,
  ccn_primary = NULL
)

Arguments

npi

<chr> 10-digit Individual NPI

pac

<chr> 10-digit PECOS Associate Control (PAC) ID

first_name, middle_name, last_name, suffix

<chr> Individual provider's name

facility_type

<chr> Type of facility, one of the following:

"Hospital"

"Long-term care hospital"

"Nursing home"

"Inpatient rehabilitation facility"

"Home health agency"

"Skilled nursing facility"

"Hospice"

"Dialysis facility"

ccn_facility

<chr> 6-digit CMS Certification Number (CCN) of facility or unit within hospital where an individual provider provides service.

ccn_primary

<int> 6-digit CMS Certification Number (CCN) of a sub-unit's primary hospital, should the provider provide services in said unit.

Periodicity

Monthly (R/P1M)

Examples

affiliations(facility_type = "Home health agency")
#> # A tibble: 2,000 × 9
#>    npi   pac   last_name first_name middle_name suff  facility_type ccn_facility
#>  * <chr> <chr> <chr>     <chr>      <chr>       <chr> <chr>         <chr>       
#>  1 1003… 7517… ENKESHAFI ARDALAN    NA          NA    Hospital      390268      
#>  2 1003… 7517… ENKESHAFI ARDALAN    NA          NA    Hospital      390048      
#>  3 1003… 7517… ENKESHAFI ARDALAN    NA          NA    Hospital      210001      
#>  4 1003… 9931… KHALIL    RASHID     NA          NA    Hospital      360112      
#>  5 1003… 9133… VELOTTA   JENNIFER   A           NA    Hospital      360098      
#>  6 1003… 0446… ROTHCHILD KEVIN      B           NA    Hospital      060024      
#>  7 1003… 2163… SEMONCHE  AMANDA     M           NA    Hospital      390035      
#>  8 1003… 2163… SEMONCHE  AMANDA     M           NA    Hospital      390049      
#>  9 1003… 4082… KIM       DAE        NA          NA    Hospital      370001      
#> 10 1003… 4082… KIM       DAE        NA          NA    Hospital      370202      
#> # ℹ 1,990 more rows
#> # ℹ 1 more variable: ccn_primary <chr>
# affiliations(last_name = "CURRY")
# affiliations(ccn_primary = "670055")
# affiliations(ccn_facility = "370781")
# affiliations(ccn_facility = "331302")
# affiliations(ccn_facility = "33Z302")
# affiliations(npi = "1043245657")