Current Main API Endpoint

Current Main API Endpoint

Current Provider API Endpoint

Current Open Payments API Endpoint

TemporalMain API Endpoint

Temporal Open Payments API Endpoint

Main API Endpoint Temporal Group

Open Payments API Endpoint Temporal Group

Usage

MainCurrent(alias)

ProviderCurrent(alias)

OpenCurrent(alias)

MainTemporal(alias)

OpenTemporal(alias)

MainTemporalGroup(alias)

OpenTemporalGroup(
  title = character(0),
  description = character(0),
  contact = character(0),
  rows = integer(0),
  pages = integer(0),
  fields = character(0),
  years = integer(0),
  endpoints = list()
)

Arguments

alias

<chr> endpoint alias

title

<chr> title

description

<chr> description

contact

<chr> contact name and email address

rows

<int> number of rows

pages

<int> number of pages

fields

<chr> field names

years

<int> years available

endpoints

<data.frame> endpoints

Value

An S7 <MainCurrent> object.

An S7 <ProviderCurrent> object.

An S7 <OpenCurrent> object.

An S7 <MainTemporal> object.

An S7 <OpenTemporal> object.

An S7 <MainTemporalGroup> object.

An S7 <TemporalOpenGroup> object.

Examples

MainCurrent("enrollees")
#> ══ Medicare Fee-For-Service Public Provider Enrollment ══════
#> 
#> ── CMS Main API 
#> 
#>   The Medicare Fee-For-Service Public Provider
#>   Enrollment dataset includes information on
#>   providers who are actively approved to bill
#>   Medicare or have completed the 855O at the time
#>   the data was pulled from the Provider
#>   Enrollment, Chain, and Ownership System
#>   (PECOS). The release of this provider
#>   enrollment data is not related to other
#>   provider information releases such as Physician
#>   Compare or Data Transparency.
#> 
#> • Rows 2,945,846
#> • Pages 590
#> • Fields 10
#> 
#> • Modified ❯ 2025-03-18
#> • Timespan ❯ 2023-01-01 • 2024-12-31
#> • Periodicity ❯ Quarterly [R/P3M]
MainCurrent("opt_out")
#> ══ Opt Out Affidavits ══════
#> 
#> ── CMS Main API 
#> 
#>   The Opt Out Affidavits dataset provides
#>   information on providers who have decided not
#>   to participate in Medicare. It contains
#>   providers NPI, specialty, address, and
#>   effective dates.
#> 
#> • Rows 49,923
#> • Pages 10
#> • Fields 13
#> 
#> • Modified ❯ 2025-03-24
#> • Timespan ❯ 2023-04-01 • 2025-02-28
#> • Periodicity ❯ Monthly [R/P1M]
MainCurrent("order_refer")
#> ══ Order and Referring ══════
#> 
#> ── CMS Main API 
#> 
#>   The Order and Referring dataset provides
#>   information on all physicians and non-physician
#>   practitioners, by their National Provider
#>   Identifier (NPI), who are of a type/specialty
#>   that is legally eligible to order and refer in
#>   the Medicare program and who have current
#>   enrollment records in Medicare.
#> 
#> • Rows 1,938,384
#> • Pages 388
#> • Fields 8
#> 
#> • Modified ❯ 2025-03-24
#> • Timespan ❯ 2023-05-21 • 2025-03-22
#> • Periodicity ❯ Twice a Week [R/P3.5D]
MainCurrent("reassignments")
#> ══ Revalidation Reassignment List ══════
#> 
#> ── CMS Main API 
#> 
#>   The Revalidation Reassignment List dataset
#>   provides information on reassignments of
#>   providers who are due for revalidation.
#> 
#> • Rows 3,682,027
#> • Pages 737
#> • Fields 16
#> 
#> • Modified ❯ 2025-03-17
#> • Timespan ❯ 2022-02-01 • 2025-03-31
#> • Periodicity ❯ Monthly [R/P1M]
MainCurrent("hospitals")
#> ══ Hospital Enrollments ══════
#> 
#> ── CMS Main API 
#> 
#>   The Hospital Enrollments dataset provides
#>   enrollment information of all Hospitals
#>   currently enrolled in Medicare. This data
#>   includes information on the Hospitals sub-group
#>   type, legal business name, doing business as
#>   name, organization type and address.
#> 
#> • Rows 9,272
#> • Pages 2
#> • Fields 39
#> 
#> • Modified ❯ 2025-01-15
#> • Timespan ❯ 2022-11-01 • 2025-01-31
#> • Periodicity ❯ Monthly [R/P1M]
MainCurrent("laboratories")
#> ══ Provider of Services File - Clinical Laboratories ══════
#> 
#> ── CMS Main API 
#> 
#>   The Provider of Services (POS) Clinical
#>   Laboratories (CLIA) data provides information
#>   on CLIA demographics and types of testing
#>   services the facility provides. In this file
#>   you will find provider number, name, address
#>   and characteristics of the participating
#>   institution providers.
#> 
#> • Rows 653,223
#> • Pages 131
#> • Fields 103
#> 
#> • Modified ❯ 2025-01-16
#> • Timespan ❯ 2011-10-01 • 2024-12-31
#> • Periodicity ❯ Quarterly [R/P3M]
MainCurrent("crosswalk")
#> ══ Medicare Provider and Supplier Taxonomy Crosswalk ══════
#> 
#> ── CMS Main API 
#> 
#>   The Medicare Provider and Supplier Taxonomy
#>   Crosswalk dataset lists the providers and
#>   suppliers eligible to enroll in Medicare
#>   programs with the proper healthcare provider
#>   taxonomy code. This data includes the Medicare
#>   speciality codes, if available,
#>   provider/supplier type description, taxonomy
#>   code, and the taxonomy description.
#> 
#> • Rows 561
#> • Pages 1
#> • Fields 4
#> 
#> • Modified ❯ 2025-01-17
#> • Timespan ❯ 2021-07-01 • 2025-06-30
#> • Periodicity ❯ Biannually [R/P6M]
MainCurrent("rbcs")
#> ══ Restructured BETOS Classification System ══════
#> 
#> ── CMS Main API 
#> 
#>   The Restructured BETOS Classification System
#>   (RBCS) dataset is a taxonomy that allows
#>   researchers to group healthcare service codes
#>   for Medicare Part B services (i.e., HCPCS
#>   codes) into clinically meaningful categories
#>   and subcategories. It is based on the original
#>   Berenson-Eggers Type of Service (BETOS)
#>   classification created in the 1980s, and
#>   includes notable updates such as Part B
#>   non-physician services. The RBCS will undergo
#>   annual updates by a technical expert panel of
#>   researchers and clinicians.
#> 
#> • Rows 18,077
#> • Pages 4
#> • Fields 14
#> 
#> • Modified ❯ 2024-10-22
#> • Timespan ❯ 2024-01-01 • 2024-12-31
#> • Periodicity ❯ Annually [R/P1Y]
MainCurrent("facilities")
#> ══ Provider of Services File - Hospital & Non-Hospital Facilities ══════
#> 
#> ── CMS Main API 
#> 
#>   Please be advised that as of Q4 2023 there is a
#>   new Provider of Service file (POS) that
#>   contains the provider and certification details
#>   for Home Health Agencies (HHAs), Hospices, and
#>   Ambulatory Surgical Centers (ASCs). Data
#>   contained in this file are extracted from the
#>   Internet Quality Improvement and Evaluation
#>   System (iQIES) environment and will be updated
#>   quarterly along with the other two POS files.
#> 
#>    
#> 
#>   The Provider of Services File - Hospital &
#>   Non-Hospital Facilities data provide critical
#>   resources for other federal regulator
#>   requirements as well as supports the ongoing
#>   quality & research efforts sponsored by CMS. In
#>   this file you will find provider certification,
#>   termination, accreditation, ownership, name,
#>   location and other characteristics organized by
#>   CMS Certification Number.
#> 
#> • Rows 115,647
#> • Pages 24
#> • Fields 473
#> 
#> • Modified ❯ 2025-01-31
#> • Timespan ❯ 2011-10-01 • 2024-12-31
#> • Periodicity ❯ Quarterly [R/P3M]
MainCurrent("home_health")
#> ══ Home Health Agency Enrollments ══════
#> 
#> ── CMS Main API 
#> 
#>   The Home Health Agency (HHA) Enrollments
#>   dataset provides enrollment information on all
#>   HHAs currently enrolled in Medicare. This data
#>   includes information on the HHAs legal business
#>   name, doing business as name, organization type
#>   and address.
#> 
#> • Rows 11,708
#> • Pages 3
#> • Fields 22
#> 
#> • Modified ❯ 2025-01-15
#> • Timespan ❯ 2023-04-01 • 2025-03-31
#> • Periodicity ❯ Quarterly [R/P3M]
MainCurrent("hospice")
#> ══ Hospice Enrollments ══════
#> 
#> ── CMS Main API 
#> 
#>   The Hospice Enrollments dataset provides
#>   enrollment information of all hospices
#>   currently enrolled in Medicare. This data
#>   includes information on the hospices legal
#>   business name, doing business as name,
#>   organization type and address.
#> 
#> • Rows 6,603
#> • Pages 2
#> • Fields 20
#> 
#> • Modified ❯ 2025-01-15
#> • Timespan ❯ 2023-04-01 • 2025-03-31
#> • Periodicity ❯ Quarterly [R/P3M]
MainCurrent("dialysis")
#> ══ Medicare Dialysis Facilities ══════
#> 
#> ── CMS Main API 
#> 
#>   The Medicare Dialysis Facilities data provides
#>   information on clinical and patient measures
#>   for Medicare-certified ESRD facilities, also
#>   known as dialysis facilities. It contains data
#>   on patient characteristics, treatment patterns,
#>   hospitalization, mortality, and transplantation
#>   patterns in Medicare-certified dialysis
#>   facilities.
#> 
#> • Rows 11,679,243
#> • Pages 2,336
#> • Fields 14
#> 
#> • Modified ❯ 2024-02-29
#> • Timespan ❯ 2009-01-01 • 2024-12-31
#> • Periodicity ❯ Annually [R/P1Y]
MainCurrent("snf")
#> ══ Skilled Nursing Facility Enrollments ══════
#> 
#> ── CMS Main API 
#> 
#>   The Skilled Nursing Facility (SNF) Enrollments
#>   dataset provides enrollment information of all
#>   SNF s currently enrolled in Medicare. This data
#>   includes information on the SNFs legal business
#>   name, doing business as name, organization type
#>   and address.
#> 
#> • Rows 14,462
#> • Pages 3
#> • Fields 23
#> 
#> • Modified ❯ 2025-03-17
#> • Timespan ❯ 2022-09-01 • 2025-03-31
#> • Periodicity ❯ Monthly [R/P1M]

ProviderCurrent("affiliations")
#> ══ Facility Affiliation Data ══════
#> 
#> ── CMS Provider API 
#> 
#>   This is the facility affiliations data publicly
#>   reported in the Provider Data Catalog.
#> 
#> • Rows 1,565,148
#> • Pages 783
#> • Fields 9
#> 
#> • Modified ❯ 2025-03-07
#> • Issued ❯ 2023-08-17
#> • Released ❯ 2025-03-20
ProviderCurrent("clinicians")
#> ══ National Downloadable File ══════
#> 
#> ── CMS Provider API 
#> 
#>   The Doctors and Clinicians national
#>   downloadable file is organized such that each
#>   line is unique at the clinician/enrollment
#>   record/group/address level. Clinicians with
#>   multiple Medicare enrollment records and/or
#>   single enrollments linking to multiple practice
#>   locations are listed on multiple lines.
#> 
#> • Rows 2,667,516
#> • Pages 1,334
#> • Fields 31
#> 
#> • Modified ❯ 2025-03-07
#> • Issued ❯ 2023-08-17
#> • Released ❯ 2025-03-20
ProviderCurrent("utilization")
#> ══ Utilization Data ══════
#> 
#> ── CMS Provider API 
#> 
#>   The Doctors and Clinicians utilization data
#>   file reports volume information for procedures
#>   of interest on clinician profile pages and in
#>   the provider data catalog (PDC) to inform
#>   patients and caregivers about clinicians'
#>   experience.
#> 
#> • Rows 298,997
#> • Pages 150
#> • Fields 10
#> 
#> • Modified ❯ 2025-03-17
#> • Issued ❯ 2024-01-04
#> • Released ❯ 2025-03-20

OpenCurrent("prof_cov")
#> ══ Covered Recipient Profile Supplement ══════
#> 
#> ── CMS Open Payments API 
#> 
#>   The Covered Recipient Profile Supplement file
#>   contains information about physicians and
#>   non-physician practitioners who have been
#>   indicated as recipients of payments, other
#>   transfers of value, or ownership and investment
#>   interest in payment records, as well as
#>   physicians and non-physician practitioners who
#>   have been identified as principal investigators
#>   associated with research payment records
#>   published by Open Payments. This file contains
#>   only those physicians that have at least one
#>   published payment record in this cycle of the
#>   publication as of January 30, 2025. The
#>   criteria used by the Centers for Medicare &
#>   Medicaid Services (CMS) to determine which
#>   payment records are eligible for publication is
#>   available in the Open Payments Methodology and
#>   Data Dictionary Document. This document can be
#>   found on the Resources page of the Open
#>   Payments website
#>   (https://www.cms.gov/OpenPayments/Resources).
#>   The Methodology and Data Dictionary Document
#>   also includes information on the data
#>   collection and reporting methodology, data
#>   fields included in the files, and any notes or
#>   special considerations that users should be
#>   aware of.
#> 
#> • Rows 1,530,412
#> • Pages 3,061
#> • Fields 32
#> 
#> • Modified ❯ 2025-01-30
OpenCurrent("prof_phys")
#> ══ Physician (Distinct) Profile Information ══════
#> 
#> ── CMS Open Payments API 
#> 
#>   This dataset is pre-filtered based on the most
#>   frequent searches of Open Payments data.
#> 
#> • Rows 1,528,054
#> • Pages 3,057
#> • Fields 33
#> 
#> • Modified ❯ 2025-01-30
OpenCurrent("prof_info")
#> ══ Profile Information ══════
#> 
#> ── CMS Open Payments API 
#> 
#>   This dataset provides profile information to
#>   display on payment search results.
#> 
#> • Rows 1,534,695
#> • Pages 3,070
#> • Fields 20
#> 
#> • Modified ❯ 2025-01-30
OpenCurrent("prof_map")
#> ══ Provider Profile ID Mapping Table ══════
#> 
#> ── CMS Open Payments API 
#> 
#>   Maps secondary/duplicate provider profile IDs
#>   to the primary ID in the Open Payments system.
#>   For multiple reasons, a small number of
#>   providers have more than one ID in the payments
#>   records.
#> 
#> • Rows 2,358
#> • Pages 5
#> • Fields 2
#> 
#> • Modified ❯ 2025-01-30
OpenCurrent("prof_entity")
#> ══ Reporting Entity Profile Information ══════
#> 
#> ── CMS Open Payments API 
#> 
#>   This dataset is pre-filtered based on the most
#>   frequent searches of Open Payments data.
#> 
#> • Rows 2,843
#> • Pages 6
#> • Fields 9
#> 
#> • Modified ❯ 2025-01-30
OpenCurrent("prof_teach")
#> ══ Teaching Hospital Profile Information ══════
#> 
#> ── CMS Open Payments API 
#> 
#>   This dataset is pre-filtered based on the most
#>   frequent searches of Open Payments data.
#> 
#> • Rows 1,440
#> • Pages 3
#> • Fields 12
#> 
#> • Modified ❯ 2025-01-30
OpenCurrent("dashboard")
#> ══ Summary Dashboard - all Years ══════
#> 
#> ── CMS Open Payments API 
#> 
#>   Summary data.
#> 
#> • Rows 90
#> • Pages 1
#> • Fields 10
#> 
#> • Modified ❯ 2025-01-30
OpenCurrent("pay_state_total")
#> ══ State Level Payment Total and Averages for all Years ══════
#> 
#> ── CMS Open Payments API 
#> 
#>   This dataset is pre-filtered based on the most
#>   frequent searches of Open Payments data.
#> 
#> • Rows 2,765
#> • Pages 6
#> • Fields 28
#> 
#> • Modified ❯ 2025-01-30
OpenCurrent("pay_state_group")
#> ══ State Payment Totals and Averages Grouped by Nature of Payment for all Years ══════
#> 
#> ── CMS Open Payments API 
#> 
#>   Groups payment data by state, recipient type,
#>   and nature of payment.
#> 
#> • Rows 16,758
#> • Pages 34
#> • Fields 28
#> 
#> • Modified ❯ 2025-01-30
OpenCurrent("pay_nat_group")
#> ══ National Level Payment Total and Averages by Provider Specialty for all Years ══════
#> 
#> ── CMS Open Payments API 
#> 
#>   This dataset is pre-filtered based on the most
#>   frequent searches of Open Payments data.
#> 
#> • Rows 4,188
#> • Pages 9
#> • Fields 21
#> 
#> • Modified ❯ 2025-01-30
OpenCurrent("pay_nat_total")
#> ══ National Level Payment Total and Averages for all Years ══════
#> 
#> ── CMS Open Payments API 
#> 
#>   This dataset is pre-filtered based on the most
#>   frequent searches of Open Payments data.
#> 
#> • Rows 24
#> • Pages 1
#> • Fields 18
#> 
#> • Modified ❯ 2025-01-30

MainTemporal("quality_payment")
#> <MainTemporal>
#>  @ title      : chr "Quality Payment Program Experience"
#>  @ description: chr "The Quality Payment Program (QPP) Experience dataset provides participation and performance information in the "| __truncated__
#>  @ contact    : chr "Quality Payment Program - CCSQ (mailto:qpp@cms.hhs.gov)"
#>  @ rows       : int 624200
#>  @ pages      : int 125
#>  @ fields     : chr [1:165] "provider key" "practice state or us territory" ...
#>  @ years      : int [1:6] 2022 2021 2020 2019 2018 2017
#>  @ endpoints  : tibble [6 × 6] (S3: tbl_df/tbl/data.frame)
#>  $ year      : int [1:6] 2022 2021 2020 2019 2018 2017
#>  $ modified  : Date[1:6], format: "2024-05-08" "2023-06-12" ...
#>  $ identifier: chr [1:6] "https://data.cms.gov/data-api/v1/dataset/b3438273-b4a6-44ca-8fb2-9e6026b74642/data" "https://data.cms.gov/data-api/v1/dataset/8c2ab3e4-597e-43c9-9d94-7223c66d7f25/data" "https://data.cms.gov/data-api/v1/dataset/48a484e6-ff02-4284-8666-6045b3cac17f/data" "https://data.cms.gov/data-api/v1/dataset/5776e77d-a844-49fd-b697-a2b01bdacdc7/data" ...
#>  $ download  : chr [1:6] "https://data.cms.gov/sites/default/files/2023-05/0bac88bc-a4c8-4dff-bdbf-660d24e6a689/2021_puf_QPPAR-18260.csv" "https://data.cms.gov/sites/default/files/2023-04/fb3758dd-6ace-4ddb-863b-1b895cef4d7e/2020_puf_qppar_QPPAR-17660.csv" "https://data.cms.gov/sites/default/files/2023-04/f9ef9f64-91d8-49a2-a1be-5a48bbdcc1e9/2019_puf_qppar_QPPAR-17661.csv" "https://data.cms.gov/sites/default/files/2023-04/d2a19b56-7a5f-49c7-b78d-4d0dd49b9391/2018_puf_qppar-QPPAR-17662.csv" ...
#>  $ filetype  : chr [1:6] "text/csv" "text/csv" "text/csv" "text/csv" ...
#>  $ resources : chr [1:6] "https://data.cms.gov/data-api/v1/dataset-resources/b3438273-b4a6-44ca-8fb2-9e6026b74642" "https://data.cms.gov/data-api/v1/dataset-resources/8c2ab3e4-597e-43c9-9d94-7223c66d7f25" "https://data.cms.gov/data-api/v1/dataset-resources/48a484e6-ff02-4284-8666-6045b3cac17f" "https://data.cms.gov/data-api/v1/dataset-resources/5776e77d-a844-49fd-b697-a2b01bdacdc7" ...
#>  @ periodicity: chr "Annually [R/P1Y]"
#>  @ dictionary : chr "https://data.cms.gov/sites/default/files/2024-04/2022%20QPP%20Experience%20Report%20Public%20Use%20Files%20Data"| __truncated__
#>  @ site       : chr "https://data.cms.gov/quality-of-care/quality-payment-program-experience"

OpenTemporal("general")
#> <OpenTemporal>
#>  @ title      : chr "General Payment Data"
#>  @ description: chr "All general (non-research, non-ownership related) payments from the 2023 program year [January 1 – December 31, 2023]"
#>  @ contact    : chr "Open Payments (mailto:openpayments@cms.hhs.gov)"
#>  @ rows       : int 14607336
#>  @ pages      : int 29215
#>  @ fields     : chr [1:91] "change_type" "covered_recipient_type" ...
#>  @ years      : int [1:7] 2023 2022 2021 2020 2019 2018 2017
#>  @ endpoints  : tibble [7 × 4] (S3: tbl_df/tbl/data.frame)
#>  $ year      : int [1:7] 2023 2022 2021 2020 2019 2018 2017
#>  $ modified  : Date[1:7], format: "2025-01-30" "2025-01-30" ...
#>  $ identifier: chr [1:7] "fb3a65aa-c901-4a38-a813-b04b00dfa2a9" "df01c2f8-dc1f-4e79-96cb-8208beaf143c" "0380bbeb-aea1-58b6-b708-829f92a48202" "a08c4b30-5cf3-4948-ad40-36f404619019" ...
#>  $ download  : chr [1:7] "https://download.cms.gov/openpayments/PGYR2023_P01302025_01212025/OP_DTL_GNRL_PGYR2023_P01302025_01212025.csv" "https://download.cms.gov/openpayments/PGYR2022_P01302025_01212025/OP_DTL_GNRL_PGYR2022_P01302025_01212025.csv" "https://download.cms.gov/openpayments/PGYR2021_P01302025_01212025/OP_DTL_GNRL_PGYR2021_P01302025_01212025.csv" "https://download.cms.gov/openpayments/PGYR2020_P01302025_01212025/OP_DTL_GNRL_PGYR2020_P01302025_01212025.csv" ...
OpenTemporal("ownership")
#> <OpenTemporal>
#>  @ title      : chr "Ownership Payment Data"
#>  @ description: chr "Ownership Payment Data - Detailed Dataset 2023 Reporting Year\nAll ownership and investment payments from the 2"| __truncated__
#>  @ contact    : chr "Open Payments (mailto:openpayments@cms.hhs.gov)"
#>  @ rows       : int 4013
#>  @ pages      : int 9
#>  @ fields     : chr [1:30] "change_type" "physician_profile_id" "physician_npi" ...
#>  @ years      : int [1:7] 2023 2022 2021 2020 2019 2018 2017
#>  @ endpoints  : tibble [7 × 4] (S3: tbl_df/tbl/data.frame)
#>  $ year      : int [1:7] 2023 2022 2021 2020 2019 2018 2017
#>  $ modified  : Date[1:7], format: "2025-01-30" "2025-01-30" ...
#>  $ identifier: chr [1:7] "ac0bc85c-02e3-45d9-89e8-2ff43da85df7" "37792388-800f-427a-9e02-b11601454eeb" "b0c03b8d-06df-58f2-8ce2-4daeffee147e" "a9a0bf48-6b96-4589-b4c2-3c5dcfbeaca2" ...
#>  $ download  : chr [1:7] "https://download.cms.gov/openpayments/PGYR2023_P01302025_01212025/OP_DTL_OWNRSHP_PGYR2023_P01302025_01212025.csv" "https://download.cms.gov/openpayments/PGYR2022_P01302025_01212025/OP_DTL_OWNRSHP_PGYR2022_P01302025_01212025.csv" "https://download.cms.gov/openpayments/PGYR2021_P01302025_01212025/OP_DTL_OWNRSHP_PGYR2021_P01302025_01212025.csv" "https://download.cms.gov/openpayments/PGYR2020_P01302025_01212025/OP_DTL_OWNRSHP_PGYR2020_P01302025_01212025.csv" ...
OpenTemporal("research")
#> <OpenTemporal>
#>  @ title      : chr "Research Payment Data"
#>  @ description: chr "Research Payment Data – Detailed Dataset 2023 Reporting Year\nAll research-related payments from the 2023 progr"| __truncated__
#>  @ contact    : chr "Open Payments (mailto:openpayments@cms.hhs.gov)"
#>  @ rows       : int 1027926
#>  @ pages      : int 2056
#>  @ fields     : chr [1:252] "change_type" "covered_recipient_type" ...
#>  @ years      : int [1:7] 2023 2022 2021 2020 2019 2018 2017
#>  @ endpoints  : tibble [7 × 4] (S3: tbl_df/tbl/data.frame)
#>  $ year      : int [1:7] 2023 2022 2021 2020 2019 2018 2017
#>  $ modified  : Date[1:7], format: "2025-01-30" "2025-01-30" ...
#>  $ identifier: chr [1:7] "ec9521bf-9d97-4603-814c-f4132d34bc4f" "fdc3c773-018a-412c-8a81-d7b8a13a037b" "ce1d28dd-0094-5060-a036-580329439600" "9c248e7e-7c7f-478b-ab84-ce0919d72c1c" ...
#>  $ download  : chr [1:7] "https://download.cms.gov/openpayments/PGYR2023_P01302025_01212025/OP_DTL_RSRCH_PGYR2023_P01302025_01212025.csv" "https://download.cms.gov/openpayments/PGYR2022_P01302025_01212025/OP_DTL_RSRCH_PGYR2022_P01302025_01212025.csv" "https://download.cms.gov/openpayments/PGYR2021_P01302025_01212025/OP_DTL_RSRCH_PGYR2021_P01302025_01212025.csv" "https://download.cms.gov/openpayments/PGYR2020_P01302025_01212025/OP_DTL_RSRCH_PGYR2020_P01302025_01212025.csv" ...
OpenTemporal("recipient_nature")
#> <OpenTemporal>
#>  @ title      : chr "Payments Grouped by Covered Recipient and Nature of Payments"
#>  @ description: chr "This dataset is pre-filtered based on the most frequent searches of Open Payments data."
#>  @ contact    : chr "Open Payments (mailto:openpayments@cms.hhs.gov)"
#>  @ rows       : int 1184051
#>  @ pages      : int 2369
#>  @ fields     : chr [1:10] "recipient_id" "recipient_type" "covered_recipient_npi" ...
#>  @ years      : int [1:7] 2023 2022 2021 2020 2019 2018 2017
#>  @ endpoints  : tibble [7 × 4] (S3: tbl_df/tbl/data.frame)
#>  $ year      : int [1:7] 2023 2022 2021 2020 2019 2018 2017
#>  $ modified  : Date[1:7], format: "2025-01-30" "2025-01-30" ...
#>  $ identifier: chr [1:7] "4017acfa-0eeb-4b8a-b68c-3b41e779c4a2" "686b765e-f4b7-4d35-a050-fb5af926dbbe" "d35243be-55cb-5087-931e-a294e0e0917f" "1fd2be47-cb72-435a-b2df-142f38e99a05" ...
#>  $ download  : chr [1:7] "https://download.cms.gov/openpayments/SMRY_RPTS_P01302025_01212025/PBLCTN_SMRY_BY_CR_BY_NTR_OF_PYMT_PGYR2023_P0"| __truncated__ "https://download.cms.gov/openpayments/SMRY_RPTS_P01302025_01212025/PBLCTN_SMRY_BY_CR_BY_NTR_OF_PYMT_PGYR2022_P0"| __truncated__ "https://download.cms.gov/openpayments/SMRY_RPTS_P01302025_01212025/PBLCTN_SMRY_BY_CR_BY_NTR_OF_PYMT_PGYR2021_P0"| __truncated__ "https://download.cms.gov/openpayments/SMRY_RPTS_P01302025_01212025/PBLCTN_SMRY_BY_CR_BY_NTR_OF_PYMT_PGYR2020_P0"| __truncated__ ...
OpenTemporal("recipient_entity")
#> <OpenTemporal>
#>  @ title      : chr "Payments Grouped by Covered Recipient and Reporting Entities"
#>  @ description: chr "This dataset is pre-filtered based on the most frequent searches of Open Payments data."
#>  @ contact    : chr "Open Payments (mailto:openpayments@cms.hhs.gov)"
#>  @ rows       : int 4416875
#>  @ pages      : int 8834
#>  @ fields     : chr [1:12] "recipient_id" "recipient_type" "covered_recipient_npi" ...
#>  @ years      : int [1:7] 2023 2022 2021 2020 2019 2018 2017
#>  @ endpoints  : tibble [7 × 4] (S3: tbl_df/tbl/data.frame)
#>  $ year      : int [1:7] 2023 2022 2021 2020 2019 2018 2017
#>  $ modified  : Date[1:7], format: "2025-01-30" "2025-01-30" ...
#>  $ identifier: chr [1:7] "4fbf0094-3c7b-425a-aabe-5778052559ad" "d2d87f39-b117-4ac8-9f36-a088fb22a25b" "7c9ebb83-8f88-5462-ab7d-1925b0113cfc" "78ad9c8d-d7ae-4066-bf2d-a66e3c9d6e3b" ...
#>  $ download  : chr [1:7] "https://download.cms.gov/openpayments/SMRY_RPTS_P01302025_01212025/PBLCTN_SMRY_BY_CR_BY_AMGPO_PGYR2023_P01302025_01212025.csv" "https://download.cms.gov/openpayments/SMRY_RPTS_P01302025_01212025/PBLCTN_SMRY_BY_CR_BY_AMGPO_PGYR2022_P01302025_01212025.csv" "https://download.cms.gov/openpayments/SMRY_RPTS_P01302025_01212025/PBLCTN_SMRY_BY_CR_BY_AMGPO_PGYR2021_P01302025_01212025.csv" "https://download.cms.gov/openpayments/SMRY_RPTS_P01302025_01212025/PBLCTN_SMRY_BY_CR_BY_AMGPO_PGYR2020_P01302025_01212025.csv" ...
OpenTemporal("entity_nature")
#> <OpenTemporal>
#>  @ title      : chr "Payments Grouped by Reporting Entities and Nature of Payments"
#>  @ description: chr "This dataset is pre-filtered based on the most frequent searches of Open Payments data."
#>  @ contact    : chr "Open Payments (mailto:openpayments@cms.hhs.gov)"
#>  @ rows       : int 6192
#>  @ pages      : int 13
#>  @ fields     : chr [1:5] "amgpo_id" "nature_of_payment_type_code" ...
#>  @ years      : int [1:7] 2023 2022 2021 2020 2019 2018 2017
#>  @ endpoints  : tibble [7 × 4] (S3: tbl_df/tbl/data.frame)
#>  $ year      : int [1:7] 2023 2022 2021 2020 2019 2018 2017
#>  $ modified  : Date[1:7], format: "2025-01-30" "2025-01-30" ...
#>  $ identifier: chr [1:7] "72008dab-0953-4226-a4cd-9f1872e8170d" "cedcd327-4e5d-43f9-8eb1-c11850fa7c66" "a2714020-f599-52b0-b12b-8f5b510e547a" "7dfe6254-2aee-454f-9ea5-df611a9a7b55" ...
#>  $ download  : chr [1:7] "https://download.cms.gov/openpayments/SMRY_RPTS_P01302025_01212025/PBLCTN_SMRY_BY_AMGPO_BY_NTR_OF_PYMT_PGYR2023"| __truncated__ "https://download.cms.gov/openpayments/SMRY_RPTS_P01302025_01212025/PBLCTN_SMRY_BY_AMGPO_BY_NTR_OF_PYMT_PGYR2022"| __truncated__ "https://download.cms.gov/openpayments/SMRY_RPTS_P01302025_01212025/PBLCTN_SMRY_BY_AMGPO_BY_NTR_OF_PYMT_PGYR2021"| __truncated__ "https://download.cms.gov/openpayments/SMRY_RPTS_P01302025_01212025/PBLCTN_SMRY_BY_AMGPO_BY_NTR_OF_PYMT_PGYR2020"| __truncated__ ...
OpenTemporal("entity_recipient_nature")
#> <OpenTemporal>
#>  @ title      : chr "Payments Grouped by Reporting Entities, Covered Recipient, and Nature of Payments"
#>  @ description: chr "This dataset is pre-filtered based on the most frequent searches of Open Payments data."
#>  @ contact    : chr "Open Payments (mailto:openpayments@cms.hhs.gov)"
#>  @ rows       : int 4588015
#>  @ pages      : int 9177
#>  @ fields     : chr [1:8] "amgpo_id" "recipient_id" "recipient_type" ...
#>  @ years      : int [1:7] 2023 2022 2021 2020 2019 2018 2017
#>  @ endpoints  : tibble [7 × 4] (S3: tbl_df/tbl/data.frame)
#>  $ year      : int [1:7] 2023 2022 2021 2020 2019 2018 2017
#>  $ modified  : Date[1:7], format: "2025-01-30" "2025-01-30" ...
#>  $ identifier: chr [1:7] "29c23b6a-dd1e-4d82-a606-8d792c7f229e" "171e369a-6f25-49eb-aa36-37f0a497e5ca" "f240c437-9d8a-5be7-95a5-8ce8e0d970bf" "532df191-473f-4bcc-9056-a6260b9437b5" ...
#>  $ download  : chr [1:7] "https://download.cms.gov/openpayments/SMRY_RPTS_P01302025_01212025/PBLCTN_SMRY_BY_AMGPO_BY_CR_BY_NTR_OF_PYMT_PG"| __truncated__ "https://download.cms.gov/openpayments/SMRY_RPTS_P01302025_01212025/PBLCTN_SMRY_BY_AMGPO_BY_CR_BY_NTR_OF_PYMT_PG"| __truncated__ "https://download.cms.gov/openpayments/SMRY_RPTS_P01302025_01212025/PBLCTN_SMRY_BY_AMGPO_BY_CR_BY_NTR_OF_PYMT_PG"| __truncated__ "https://download.cms.gov/openpayments/SMRY_RPTS_P01302025_01212025/PBLCTN_SMRY_BY_AMGPO_BY_CR_BY_NTR_OF_PYMT_PG"| __truncated__ ...
OpenTemporal("state_nature")
#> <OpenTemporal>
#>  @ title      : chr "State Payment Totals Grouped by Nature of Payment for all Years"
#>  @ description: chr "Groups payment data by state, recipient type, and nature of payment."
#>  @ contact    : chr "Open Payments (mailto:openpayments@cms.hhs.gov)"
#>  @ rows       : int 16758
#>  @ pages      : int 34
#>  @ fields     : chr [1:28] "country_code" "country_name" "state_code" "state_name" ...
#>  @ years      : int [1:7] 2023 2022 2021 2020 2019 2018 2017
#>  @ endpoints  : tibble [7 × 4] (S3: tbl_df/tbl/data.frame)
#>  $ year      : int [1:7] 2023 2022 2021 2020 2019 2018 2017
#>  $ modified  : Date[1:7], format: "2025-01-30" "2025-01-30" ...
#>  $ identifier: chr [1:7] "f83aed3a-7504-4fca-aabb-ac0338c79688" "c64c40af-23a5-4419-8b64-692684483664" "7fad5b56-2cfe-52ab-ac4d-3200780287d4" "7f3055b6-73d8-5aff-be15-55873552e2cd" ...
#>  $ download  : chr [1:7] "https://download.cms.gov/openpayments/SMRY_RPTS_P01302025_01212025/PBLCTN_NTR_OF_PYMT_BY_STATE_SMRY_PGYR2023_P0"| __truncated__ "https://download.cms.gov/openpayments/SMRY_RPTS_P01302025_01212025/PBLCTN_NTR_OF_PYMT_BY_STATE_SMRY_PGYR2022_P0"| __truncated__ "https://download.cms.gov/openpayments/SMRY_RPTS_P01302025_01212025/PBLCTN_NTR_OF_PYMT_BY_STATE_SMRY_PGYR2021_P0"| __truncated__ "https://download.cms.gov/openpayments/SMRY_RPTS_P01302025_01212025/PBLCTN_NTR_OF_PYMT_BY_STATE_SMRY_PGYR2020_P0"| __truncated__ ...
MainTemporalGroup("utilization")
#> <MainTemporalGroup>
#>  @ title      : chr "Medicare Physician & Other Practitioners"
#>  @ contact    : chr "Medicare Provider Data - OEDA (mailto:IPAG_Data_Products@cms.hhs.gov)"
#>  @ endpoints  :List of 3
#>  .. $ by_geography_and_service:List of 7
#>  ..  ..$ description: chr "The Medicare Physician & Other Practitioners by Geography and Service dataset contains information on use, paym"| __truncated__
#>  ..  ..$ dictionary : chr "https://data.cms.gov/resources/medicare-physician-other-practitioners-by-geography-and-service-data-dictionary"
#>  ..  ..$ site       : chr "https://data.cms.gov/provider-summary-by-type-of-service/medicare-physician-other-practitioners/medicare-physic"| __truncated__
#>  ..  ..$ rows       : int 270673
#>  ..  ..$ pages      : int 55
#>  ..  ..$ fields     : chr [1:15] "Rndrng_Prvdr_Geo_Lvl" "Rndrng_Prvdr_Geo_Cd" "Rndrng_Prvdr_Geo_Desc" "HCPCS_Cd" ...
#>  ..  ..$ endpoints  : tibble [10 × 3] (S3: tbl_df/tbl/data.frame)
#>  ..  .. ..$ year      : int [1:10] 2022 2021 2020 2019 2018 2017 2016 2015 2014 2013
#>  ..  .. ..$ identifier: chr [1:10] "https://data.cms.gov/data-api/v1/dataset/87304f15-9ed0-41dc-a141-6141a0327453/data" "https://data.cms.gov/data-api/v1/dataset/f00f462f-bb61-48b1-a321-1c50d78ce175/data" "https://data.cms.gov/data-api/v1/dataset/31eb3018-43e0-4259-a0d8-e7a1112ffd08/data" "https://data.cms.gov/data-api/v1/dataset/673030ae-ceed-4561-8fca-b1275395a86a/data" ...
#>  ..  .. ..$ resources : chr [1:10] "https://data.cms.gov/data-api/v1/dataset-resources/87304f15-9ed0-41dc-a141-6141a0327453" "https://data.cms.gov/data-api/v1/dataset-resources/f00f462f-bb61-48b1-a321-1c50d78ce175" "https://data.cms.gov/data-api/v1/dataset-resources/31eb3018-43e0-4259-a0d8-e7a1112ffd08" "https://data.cms.gov/data-api/v1/dataset-resources/673030ae-ceed-4561-8fca-b1275395a86a" ...
#>  .. $ by_provider             :List of 7
#>  ..  ..$ description: chr "The Medicare Physician & Other Practitioners by Provider dataset provides information on use, payments, submitt"| __truncated__
#>  ..  ..$ dictionary : chr "https://data.cms.gov/resources/medicare-physician-other-practitioners-by-provider-data-dictionary"
#>  ..  ..$ site       : chr "https://data.cms.gov/provider-summary-by-type-of-service/medicare-physician-other-practitioners/medicare-physic"| __truncated__
#>  ..  ..$ rows       : int 1230293
#>  ..  ..$ pages      : int 247
#>  ..  ..$ fields     : chr [1:82] "Rndrng_NPI" "Rndrng_Prvdr_Last_Org_Name" "Rndrng_Prvdr_First_Name" "Rndrng_Prvdr_MI" ...
#>  ..  ..$ endpoints  : tibble [10 × 3] (S3: tbl_df/tbl/data.frame)
#>  ..  .. ..$ year      : int [1:10] 2022 2021 2020 2019 2018 2017 2016 2015 2014 2013
#>  ..  .. ..$ identifier: chr [1:10] "https://data.cms.gov/data-api/v1/dataset/8568be52-dfc9-4fe0-b8d7-95838b3cd0c6/data" "https://data.cms.gov/data-api/v1/dataset/1a3a5f39-09ac-48e2-b9bf-b371ac01c07e/data" "https://data.cms.gov/data-api/v1/dataset/6c0feedd-12a5-401c-bd35-15fabcd6a334/data" "https://data.cms.gov/data-api/v1/dataset/6859b742-ef67-4930-8262-03fb341001e2/data" ...
#>  ..  .. ..$ resources : chr [1:10] "https://data.cms.gov/data-api/v1/dataset-resources/8568be52-dfc9-4fe0-b8d7-95838b3cd0c6" "https://data.cms.gov/data-api/v1/dataset-resources/1a3a5f39-09ac-48e2-b9bf-b371ac01c07e" "https://data.cms.gov/data-api/v1/dataset-resources/6c0feedd-12a5-401c-bd35-15fabcd6a334" "https://data.cms.gov/data-api/v1/dataset-resources/6859b742-ef67-4930-8262-03fb341001e2" ...
#>  .. $ by_provider_and_service :List of 7
#>  ..  ..$ description: chr "The Medicare Physician & Other Practitioners by Provider and Service dataset provides information on use, payme"| __truncated__
#>  ..  ..$ dictionary : chr "https://data.cms.gov/resources/medicare-physician-other-practitioners-by-provider-and-service-data-dictionary"
#>  ..  ..$ site       : chr "https://data.cms.gov/provider-summary-by-type-of-service/medicare-physician-other-practitioners/medicare-physic"| __truncated__
#>  ..  ..$ rows       : int 9755427
#>  ..  ..$ pages      : int 1952
#>  ..  ..$ fields     : chr [1:29] "Rndrng_NPI" "Rndrng_Prvdr_Last_Org_Name" "Rndrng_Prvdr_First_Name" "Rndrng_Prvdr_MI" ...
#>  ..  ..$ endpoints  : tibble [10 × 3] (S3: tbl_df/tbl/data.frame)
#>  ..  .. ..$ year      : int [1:10] 2022 2021 2020 2019 2018 2017 2016 2015 2014 2013
#>  ..  .. ..$ identifier: chr [1:10] "https://data.cms.gov/data-api/v1/dataset/4f307be4-6868-4a9e-ae92-acf3fd4b5543/data" "https://data.cms.gov/data-api/v1/dataset/5c67d835-3862-4f63-897d-85d3eac82d5b/data" "https://data.cms.gov/data-api/v1/dataset/862ed658-1f38-4b2f-b02b-0b359e12c78a/data" "https://data.cms.gov/data-api/v1/dataset/5fccd951-9538-48a7-9075-6f02b9867868/data" ...
#>  ..  .. ..$ resources : chr [1:10] "https://data.cms.gov/data-api/v1/dataset-resources/4f307be4-6868-4a9e-ae92-acf3fd4b5543" "https://data.cms.gov/data-api/v1/dataset-resources/5c67d835-3862-4f63-897d-85d3eac82d5b" "https://data.cms.gov/data-api/v1/dataset-resources/862ed658-1f38-4b2f-b02b-0b359e12c78a" "https://data.cms.gov/data-api/v1/dataset-resources/5fccd951-9538-48a7-9075-6f02b9867868" ...
#>  @ periodicity: chr "Annually [R/P1Y]"
#>  @ years      : int [1:10] 2022 2021 2020 2019 2018 2017 2016 2015 2014 2013
MainTemporalGroup("prescribers")
#> <MainTemporalGroup>
#>  @ title      : chr "Medicare Part D Prescribers"
#>  @ contact    : chr "Medicare Provider Data - OEDA (mailto:IPAG_Data_Products@cms.hhs.gov)"
#>  @ endpoints  :List of 3
#>  .. $ by_geography_and_drug:List of 7
#>  ..  ..$ description: chr "The Medicare Part D Prescribers by Geography and Drug dataset contains information on prescription drugs prescr"| __truncated__
#>  ..  ..$ dictionary : chr "https://data.cms.gov/resources/medicare-part-d-prescribers-by-geography-and-drug-data-dictionary"
#>  ..  ..$ site       : chr "https://data.cms.gov/provider-summary-by-type-of-service/medicare-part-d-prescribers/medicare-part-d-prescriber"| __truncated__
#>  ..  ..$ rows       : int 115396
#>  ..  ..$ pages      : int 24
#>  ..  ..$ fields     : chr [1:22] "Prscrbr_Geo_Lvl" "Prscrbr_Geo_Cd" "Prscrbr_Geo_Desc" "Brnd_Name" ...
#>  ..  ..$ endpoints  : tibble [10 × 3] (S3: tbl_df/tbl/data.frame)
#>  ..  .. ..$ year      : int [1:10] 2022 2021 2020 2019 2018 2017 2016 2015 2014 2013
#>  ..  .. ..$ identifier: chr [1:10] "https://data.cms.gov/data-api/v1/dataset/1fc57194-a51d-4864-aee6-de0889488151/data" "https://data.cms.gov/data-api/v1/dataset/7dda2a9d-034a-446a-b4b3-e1254e0127b2/data" "https://data.cms.gov/data-api/v1/dataset/83891e77-99cf-4865-b60a-97703b916e09/data" "https://data.cms.gov/data-api/v1/dataset/73a6335e-f16f-4c81-a84b-6b5a986e2bf8/data" ...
#>  ..  .. ..$ resources : chr [1:10] "https://data.cms.gov/data-api/v1/dataset-resources/1fc57194-a51d-4864-aee6-de0889488151" "https://data.cms.gov/data-api/v1/dataset-resources/7dda2a9d-034a-446a-b4b3-e1254e0127b2" "https://data.cms.gov/data-api/v1/dataset-resources/83891e77-99cf-4865-b60a-97703b916e09" "https://data.cms.gov/data-api/v1/dataset-resources/73a6335e-f16f-4c81-a84b-6b5a986e2bf8" ...
#>  .. $ by_provider          :List of 7
#>  ..  ..$ description: chr "The Medicare Part D Prescribers by Provider dataset contains information on prescription drugs prescribed by in"| __truncated__
#>  ..  ..$ dictionary : chr "https://data.cms.gov/resources/medicare-part-d-prescribers-by-provider-data-dictionary"
#>  ..  ..$ site       : chr "https://data.cms.gov/provider-summary-by-type-of-service/medicare-part-d-prescribers/medicare-part-d-prescribers-by-provider"
#>  ..  ..$ rows       : int 1332309
#>  ..  ..$ pages      : int 267
#>  ..  ..$ fields     : chr [1:85] "Prscrbr_NPI" "Prscrbr_Last_Org_Name" "Prscrbr_First_Name" "Prscrbr_MI" ...
#>  ..  ..$ endpoints  : tibble [10 × 3] (S3: tbl_df/tbl/data.frame)
#>  ..  .. ..$ year      : int [1:10] 2022 2021 2020 2019 2018 2017 2016 2015 2014 2013
#>  ..  .. ..$ identifier: chr [1:10] "https://data.cms.gov/data-api/v1/dataset/5b7298b5-4a3b-4517-86b0-2c9077c64a86/data" "https://data.cms.gov/data-api/v1/dataset/c55024e4-4b4f-4106-b570-0cc05f173357/data" "https://data.cms.gov/data-api/v1/dataset/eca92c9a-cc1b-4d6e-9742-7cf34ed6f1f4/data" "https://data.cms.gov/data-api/v1/dataset/b6c4cbfd-0f23-4a38-8bc0-41a26fc71826/data" ...
#>  ..  .. ..$ resources : chr [1:10] "https://data.cms.gov/data-api/v1/dataset-resources/5b7298b5-4a3b-4517-86b0-2c9077c64a86" "https://data.cms.gov/data-api/v1/dataset-resources/c55024e4-4b4f-4106-b570-0cc05f173357" "https://data.cms.gov/data-api/v1/dataset-resources/eca92c9a-cc1b-4d6e-9742-7cf34ed6f1f4" "https://data.cms.gov/data-api/v1/dataset-resources/b6c4cbfd-0f23-4a38-8bc0-41a26fc71826" ...
#>  .. $ by_provider_and_drug :List of 7
#>  ..  ..$ description: chr "The Medicare Part D Prescribers by Provider and Drug dataset provides information on prescription drugs prescri"| __truncated__
#>  ..  ..$ dictionary : chr "https://data.cms.gov/resources/medicare-part-d-prescribers-by-provider-and-drug-data-dictionary"
#>  ..  ..$ site       : chr "https://data.cms.gov/provider-summary-by-type-of-service/medicare-part-d-prescribers/medicare-part-d-prescriber"| __truncated__
#>  ..  ..$ rows       : int 25869521
#>  ..  ..$ pages      : int 5174
#>  ..  ..$ fields     : chr [1:22] "Prscrbr_NPI" "Prscrbr_Last_Org_Name" "Prscrbr_First_Name" "Prscrbr_City" ...
#>  ..  ..$ endpoints  : tibble [10 × 3] (S3: tbl_df/tbl/data.frame)
#>  ..  .. ..$ year      : int [1:10] 2022 2021 2020 2019 2018 2017 2016 2015 2014 2013
#>  ..  .. ..$ identifier: chr [1:10] "https://data.cms.gov/data-api/v1/dataset/b101b457-ffa4-49bb-8fd9-27c1266086e2/data" "https://data.cms.gov/data-api/v1/dataset/f68114ed-f854-4ffc-9c6e-ed78b5e2f8d0/data" "https://data.cms.gov/data-api/v1/dataset/7795fe20-e80e-435a-a9ed-d2d65e05feeb/data" "https://data.cms.gov/data-api/v1/dataset/2a6705e6-7a1e-460c-ba22-35249a531918/data" ...
#>  ..  .. ..$ resources : chr [1:10] "https://data.cms.gov/data-api/v1/dataset-resources/b101b457-ffa4-49bb-8fd9-27c1266086e2" "https://data.cms.gov/data-api/v1/dataset-resources/f68114ed-f854-4ffc-9c6e-ed78b5e2f8d0" "https://data.cms.gov/data-api/v1/dataset-resources/7795fe20-e80e-435a-a9ed-d2d65e05feeb" "https://data.cms.gov/data-api/v1/dataset-resources/2a6705e6-7a1e-460c-ba22-35249a531918" ...
#>  @ periodicity: chr "Annually [R/P1Y]"
#>  @ years      : int [1:10] 2022 2021 2020 2019 2018 2017 2016 2015 2014 2013
MainTemporalGroup("suppliers")
#> <MainTemporalGroup>
#>  @ title      : chr "Medicare Durable Medical Equipment, Devices & Supplies"
#>  @ contact    : chr "Medicare Provider Data - OEDA (mailto:IPAG_Data_Products@cms.hhs.gov)"
#>  @ endpoints  :List of 5
#>  .. $ by_geography_and_service         :List of 7
#>  ..  ..$ description: chr "The Medicare Durable Medical Equipment, Devices & Supplies by Geography and Service dataset contains informatio"| __truncated__
#>  ..  ..$ dictionary : chr "https://data.cms.gov/resources/medicare-durable-medical-equipment-devices-supplies-by-geography-and-service-dat"| __truncated__
#>  ..  ..$ site       : chr "https://data.cms.gov/provider-summary-by-type-of-service/medicare-durable-medical-equipment-devices-supplies/me"| __truncated__
#>  ..  ..$ rows       : int 39129
#>  ..  ..$ pages      : int 8
#>  ..  ..$ fields     : chr [1:18] "Rfrg_Prvdr_Geo_Lvl" "Rfrg_Prvdr_Geo_Cd" "Rfrg_Prvdr_Geo_Desc" "RBCS_Lvl" ...
#>  ..  ..$ endpoints  : tibble [9 × 3] (S3: tbl_df/tbl/data.frame)
#>  ..  .. ..$ year      : int [1:9] 2022 2021 2020 2019 2018 2017 2016 2015 2014
#>  ..  .. ..$ identifier: chr [1:9] "https://data.cms.gov/data-api/v1/dataset/a9d1461a-9c63-49e1-8d91-696082e20e47/data" "https://data.cms.gov/data-api/v1/dataset/daedccd6-5881-4997-93cb-ad6f3d1dd2fd/data" "https://data.cms.gov/data-api/v1/dataset/c5da241c-71d2-4b04-9f44-b324ff4b4c01/data" "https://data.cms.gov/data-api/v1/dataset/3931609c-3f1d-4ab4-95f9-0e31d43613f0/data" ...
#>  ..  .. ..$ resources : chr [1:9] "https://data.cms.gov/data-api/v1/dataset-resources/a9d1461a-9c63-49e1-8d91-696082e20e47" "https://data.cms.gov/data-api/v1/dataset-resources/daedccd6-5881-4997-93cb-ad6f3d1dd2fd" "https://data.cms.gov/data-api/v1/dataset-resources/c5da241c-71d2-4b04-9f44-b324ff4b4c01" "https://data.cms.gov/data-api/v1/dataset-resources/3931609c-3f1d-4ab4-95f9-0e31d43613f0" ...
#>  .. $ by_referring_provider            :List of 7
#>  ..  ..$ description: chr "The Medicare Durable Medical Equipment, Devices & Supplies by Referring Provider dataset contains information o"| __truncated__
#>  ..  ..$ dictionary : chr "https://data.cms.gov/resources/medicare-durable-medical-equipment-devices-supplies-by-referring-provider-data-dictionary"
#>  ..  ..$ site       : chr "https://data.cms.gov/provider-summary-by-type-of-service/medicare-durable-medical-equipment-devices-supplies/me"| __truncated__
#>  ..  ..$ rows       : int 385637
#>  ..  ..$ pages      : int 78
#>  ..  ..$ fields     : chr [1:98] "Rfrg_NPI" "Rfrg_Prvdr_Last_Name_Org" "Rfrg_Prvdr_First_Name" "Rfrg_Prvdr_MI" ...
#>  ..  ..$ endpoints  : tibble [9 × 3] (S3: tbl_df/tbl/data.frame)
#>  ..  .. ..$ year      : int [1:9] 2022 2021 2020 2019 2018 2017 2016 2015 2014
#>  ..  .. ..$ identifier: chr [1:9] "https://data.cms.gov/data-api/v1/dataset/e0eb5710-5287-431b-984e-a5629556bbf2/data" "https://data.cms.gov/data-api/v1/dataset/527aab51-3316-4cb1-8740-a5e85a6670f2/data" "https://data.cms.gov/data-api/v1/dataset/5fb6377b-41b9-4002-828f-d257bccff371/data" "https://data.cms.gov/data-api/v1/dataset/c8cd9ce3-c005-4823-ad4d-b98bcd371f2f/data" ...
#>  ..  .. ..$ resources : chr [1:9] "https://data.cms.gov/data-api/v1/dataset-resources/e0eb5710-5287-431b-984e-a5629556bbf2" "https://data.cms.gov/data-api/v1/dataset-resources/527aab51-3316-4cb1-8740-a5e85a6670f2" "https://data.cms.gov/data-api/v1/dataset-resources/5fb6377b-41b9-4002-828f-d257bccff371" "https://data.cms.gov/data-api/v1/dataset-resources/c8cd9ce3-c005-4823-ad4d-b98bcd371f2f" ...
#>  .. $ by_referring_provider_and_service:List of 7
#>  ..  ..$ description: chr "The Medicare Durable Medical Equipment, Devices & Supplies by Referring Provider and Service dataset contains i"| __truncated__
#>  ..  ..$ dictionary : chr "https://data.cms.gov/resources/medicare-durable-medical-equipment-devices-supplies-by-referring-provider-and-se"| __truncated__
#>  ..  ..$ site       : chr "https://data.cms.gov/provider-summary-by-type-of-service/medicare-durable-medical-equipment-devices-supplies/me"| __truncated__
#>  ..  ..$ rows       : int 1457378
#>  ..  ..$ pages      : int 292
#>  ..  ..$ fields     : chr [1:34] "Rfrg_NPI" "Rfrg_Prvdr_Last_Name_Org" "Rfrg_Prvdr_First_Name" "Rfrg_Prvdr_MI" ...
#>  ..  ..$ endpoints  : tibble [9 × 3] (S3: tbl_df/tbl/data.frame)
#>  ..  .. ..$ year      : int [1:9] 2022 2021 2020 2019 2018 2017 2016 2015 2014
#>  ..  .. ..$ identifier: chr [1:9] "https://data.cms.gov/data-api/v1/dataset/7f98d91f-f2a6-44cd-a57f-1ac02eaaf795/data" "https://data.cms.gov/data-api/v1/dataset/442050e6-3d96-47b9-bf0e-d894f74a137f/data" "https://data.cms.gov/data-api/v1/dataset/72ce1cd5-efa9-484e-bd74-90ab3a10d244/data" "https://data.cms.gov/data-api/v1/dataset/5d2c51d1-50cb-4b10-8930-3e40b7561060/data" ...
#>  ..  .. ..$ resources : chr [1:9] "https://data.cms.gov/data-api/v1/dataset-resources/7f98d91f-f2a6-44cd-a57f-1ac02eaaf795" "https://data.cms.gov/data-api/v1/dataset-resources/442050e6-3d96-47b9-bf0e-d894f74a137f" "https://data.cms.gov/data-api/v1/dataset-resources/72ce1cd5-efa9-484e-bd74-90ab3a10d244" "https://data.cms.gov/data-api/v1/dataset-resources/5d2c51d1-50cb-4b10-8930-3e40b7561060" ...
#>  .. $ by_supplier                      :List of 7
#>  ..  ..$ description: chr "The Medicare Durable Medical Equipment, Devices & Supplies by Supplier dataset contains information on usage, p"| __truncated__
#>  ..  ..$ dictionary : chr "https://data.cms.gov/resources/medicare-durable-medical-equipment-devices-supplies-by-supplier-data-dictionary"
#>  ..  ..$ site       : chr "https://data.cms.gov/provider-summary-by-type-of-service/medicare-durable-medical-equipment-devices-supplies/me"| __truncated__
#>  ..  ..$ rows       : int 66406
#>  ..  ..$ pages      : int 14
#>  ..  ..$ fields     : chr [1:93] "Suplr_NPI" "Suplr_Prvdr_Last_Name_Org" "Suplr_Prvdr_First_Name" "Suplr_Prvdr_MI" ...
#>  ..  ..$ endpoints  : tibble [9 × 3] (S3: tbl_df/tbl/data.frame)
#>  ..  .. ..$ year      : int [1:9] 2022 2021 2020 2019 2018 2017 2016 2015 2014
#>  ..  .. ..$ identifier: chr [1:9] "https://data.cms.gov/data-api/v1/dataset/471f9c0d-12c4-4601-a547-559dcfa2d0e2/data" "https://data.cms.gov/data-api/v1/dataset/887561f5-1676-4c93-84b6-d2e50eb609d3/data" "https://data.cms.gov/data-api/v1/dataset/04f11092-de71-4a40-93b1-da2b41e4e938/data" "https://data.cms.gov/data-api/v1/dataset/e8c8eb83-a1d8-49d3-ac99-dda79b147021/data" ...
#>  ..  .. ..$ resources : chr [1:9] "https://data.cms.gov/data-api/v1/dataset-resources/471f9c0d-12c4-4601-a547-559dcfa2d0e2" "https://data.cms.gov/data-api/v1/dataset-resources/887561f5-1676-4c93-84b6-d2e50eb609d3" "https://data.cms.gov/data-api/v1/dataset-resources/04f11092-de71-4a40-93b1-da2b41e4e938" "https://data.cms.gov/data-api/v1/dataset-resources/e8c8eb83-a1d8-49d3-ac99-dda79b147021" ...
#>  .. $ by_supplier_and_service          :List of 7
#>  ..  ..$ description: chr "The Medicare Durable Medical Equipment, Devices & Supplies by Supplier and Service dataset contains information"| __truncated__
#>  ..  ..$ dictionary : chr "https://data.cms.gov/resources/medicare-durable-medical-equipment-devices-supplies-by-supplier-and-service-data-dictionary"
#>  ..  ..$ site       : chr "https://data.cms.gov/provider-summary-by-type-of-service/medicare-durable-medical-equipment-devices-supplies/me"| __truncated__
#>  ..  ..$ rows       : int 482638
#>  ..  ..$ pages      : int 97
#>  ..  ..$ fields     : chr [1:33] "Suplr_NPI" "Suplr_Prvdr_Last_Name_Org" "Suplr_Prvdr_First_Name" "Suplr_Prvdr_MI" ...
#>  ..  ..$ endpoints  : tibble [9 × 3] (S3: tbl_df/tbl/data.frame)
#>  ..  .. ..$ year      : int [1:9] 2022 2021 2020 2019 2018 2017 2016 2015 2014
#>  ..  .. ..$ identifier: chr [1:9] "https://data.cms.gov/data-api/v1/dataset/ce9ec4d2-d404-4463-98ef-aa5512f201e2/data" "https://data.cms.gov/data-api/v1/dataset/1ca065e5-adfa-4a43-9bf3-4089538571c9/data" "https://data.cms.gov/data-api/v1/dataset/6160f96b-710a-4380-a3ff-11324c2d5450/data" "https://data.cms.gov/data-api/v1/dataset/ea4a5834-aeaa-4176-8feb-976ae7874890/data" ...
#>  ..  .. ..$ resources : chr [1:9] "https://data.cms.gov/data-api/v1/dataset-resources/ce9ec4d2-d404-4463-98ef-aa5512f201e2" "https://data.cms.gov/data-api/v1/dataset-resources/1ca065e5-adfa-4a43-9bf3-4089538571c9" "https://data.cms.gov/data-api/v1/dataset-resources/6160f96b-710a-4380-a3ff-11324c2d5450" "https://data.cms.gov/data-api/v1/dataset-resources/ea4a5834-aeaa-4176-8feb-976ae7874890" ...
#>  @ periodicity: chr "Annually [R/P1Y]"
#>  @ years      : int [1:9] 2022 2021 2020 2019 2018 2017 2016 2015 2014
MainTemporalGroup("outpatient")
#> <MainTemporalGroup>
#>  @ title      : chr "Medicare Outpatient Hospitals"
#>  @ contact    : chr "Medicare Provider Data - OEDA (mailto:IPAG_Data_Products@cms.hhs.gov)"
#>  @ endpoints  :List of 2
#>  .. $ by_geography_and_service:List of 7
#>  ..  ..$ description: chr "The Medicare Outpatient Hospitals by Geography and Service dataset provides information on services for Origina"| __truncated__
#>  ..  ..$ dictionary : chr "https://data.cms.gov/resources/medicare-outpatient-hospitals-by-geography-and-service-data-dictionary"
#>  ..  ..$ site       : chr "https://data.cms.gov/provider-summary-by-type-of-service/medicare-outpatient-hospitals/medicare-outpatient-hosp"| __truncated__
#>  ..  ..$ rows       : int 90591
#>  ..  ..$ pages      : int 19
#>  ..  ..$ fields     : chr [1:15] "Rndrng_Prvdr_Geo_Lvl" "Rndrng_Prvdr_Geo_Cd" "Rndrng_Prvdr_Geo_Desc" "Srvc_Lvl" ...
#>  ..  ..$ endpoints  : tibble [8 × 3] (S3: tbl_df/tbl/data.frame)
#>  ..  .. ..$ year      : int [1:8] 2022 2021 2020 2019 2018 2017 2016 2015
#>  ..  .. ..$ identifier: chr [1:8] "https://data.cms.gov/data-api/v1/dataset/6af002ba-b934-4491-96dd-d041989b843c/data" "https://data.cms.gov/data-api/v1/dataset/b3537e90-c3c5-4895-914d-71f54123a0a4/data" "https://data.cms.gov/data-api/v1/dataset/31db4293-f75c-4940-bfae-32d554cd164f/data" "https://data.cms.gov/data-api/v1/dataset/46ae348d-8ba4-43df-868e-74ff87eada39/data" ...
#>  ..  .. ..$ resources : chr [1:8] "https://data.cms.gov/data-api/v1/dataset-resources/6af002ba-b934-4491-96dd-d041989b843c" "https://data.cms.gov/data-api/v1/dataset-resources/b3537e90-c3c5-4895-914d-71f54123a0a4" "https://data.cms.gov/data-api/v1/dataset-resources/31db4293-f75c-4940-bfae-32d554cd164f" "https://data.cms.gov/data-api/v1/dataset-resources/46ae348d-8ba4-43df-868e-74ff87eada39" ...
#>  .. $ by_provider_and_service :List of 7
#>  ..  ..$ description: chr "The Medicare Outpatient Hospitals by Provider and Service dataset provides information on services for Original"| __truncated__
#>  ..  ..$ dictionary : chr "https://data.cms.gov/resources/medicare-outpatient-hospitals-by-provider-and-service-data-dictionary"
#>  ..  ..$ site       : chr "https://data.cms.gov/provider-summary-by-type-of-service/medicare-outpatient-hospitals/medicare-outpatient-hosp"| __truncated__
#>  ..  ..$ rows       : int 116459
#>  ..  ..$ pages      : int 24
#>  ..  ..$ fields     : chr [1:18] "Rndrng_Prvdr_CCN" "Rndrng_Prvdr_Org_Name" "Rndrng_Prvdr_St" "Rndrng_Prvdr_City" ...
#>  ..  ..$ endpoints  : tibble [8 × 3] (S3: tbl_df/tbl/data.frame)
#>  ..  .. ..$ year      : int [1:8] 2022 2021 2020 2019 2018 2017 2016 2015
#>  ..  .. ..$ identifier: chr [1:8] "https://data.cms.gov/data-api/v1/dataset/456dbe68-8878-49dd-889b-bb8bf432fcbe/data" "https://data.cms.gov/data-api/v1/dataset/a7617aef-046f-4557-8a2a-54f9c9a88b57/data" "https://data.cms.gov/data-api/v1/dataset/8281e9f7-1874-4bd9-a183-a74bec2debd3/data" "https://data.cms.gov/data-api/v1/dataset/f3fc4327-f66e-49e1-ba06-f2607a164406/data" ...
#>  ..  .. ..$ resources : chr [1:8] "https://data.cms.gov/data-api/v1/dataset-resources/456dbe68-8878-49dd-889b-bb8bf432fcbe" "https://data.cms.gov/data-api/v1/dataset-resources/a7617aef-046f-4557-8a2a-54f9c9a88b57" "https://data.cms.gov/data-api/v1/dataset-resources/8281e9f7-1874-4bd9-a183-a74bec2debd3" "https://data.cms.gov/data-api/v1/dataset-resources/f3fc4327-f66e-49e1-ba06-f2607a164406" ...
#>  @ periodicity: chr "Annually [R/P1Y]"
#>  @ years      : int [1:8] 2022 2021 2020 2019 2018 2017 2016 2015
MainTemporalGroup("inpatient")
#> <MainTemporalGroup>
#>  @ title      : chr "Medicare Inpatient Hospitals"
#>  @ contact    : chr "Medicare Provider Data - OEDA (mailto:IPAG_Data_Products@cms.hhs.gov)"
#>  @ endpoints  :List of 3
#>  .. $ by_geography_and_service:List of 7
#>  ..  ..$ description: chr "The Medicare Inpatient Hospitals by Geography and Service dataset provides information on hospital discharges f"| __truncated__
#>  ..  ..$ dictionary : chr "https://data.cms.gov/resources/medicare-inpatient-hospitals-by-geography-and-service-data-dictionary-0"
#>  ..  ..$ site       : chr "https://data.cms.gov/provider-summary-by-type-of-service/medicare-inpatient-hospitals/medicare-inpatient-hospit"| __truncated__
#>  ..  ..$ rows       : int 26260
#>  ..  ..$ pages      : int 6
#>  ..  ..$ fields     : chr [1:9] "Rndrng_Prvdr_Geo_Lvl" "Rndrng_Prvdr_Geo_Cd" "Rndrng_Prvdr_Geo_Desc" "DRG_Cd" ...
#>  ..  ..$ endpoints  : tibble [10 × 3] (S3: tbl_df/tbl/data.frame)
#>  ..  .. ..$ year      : int [1:10] 2022 2021 2020 2019 2018 2017 2016 2015 2014 2013
#>  ..  .. ..$ identifier: chr [1:10] "https://data.cms.gov/data-api/v1/dataset/5c8b315f-a2a3-4b2f-bba6-7be19f8d7edc/data" "https://data.cms.gov/data-api/v1/dataset/68739613-1a4f-40ae-9664-49da07abd81f/data" "https://data.cms.gov/data-api/v1/dataset/e583de38-33d7-4895-a115-0d8af74f0795/data" "https://data.cms.gov/data-api/v1/dataset/337617e5-c18b-4cb9-8152-34851747584b/data" ...
#>  ..  .. ..$ resources : chr [1:10] "https://data.cms.gov/data-api/v1/dataset-resources/5c8b315f-a2a3-4b2f-bba6-7be19f8d7edc" "https://data.cms.gov/data-api/v1/dataset-resources/68739613-1a4f-40ae-9664-49da07abd81f" "https://data.cms.gov/data-api/v1/dataset-resources/e583de38-33d7-4895-a115-0d8af74f0795" "https://data.cms.gov/data-api/v1/dataset-resources/337617e5-c18b-4cb9-8152-34851747584b" ...
#>  .. $ by_provider             :List of 7
#>  ..  ..$ description: chr "The Medicare Inpatient Hospitals by Provider dataset provides information on inpatient discharges for Original "| __truncated__
#>  ..  ..$ dictionary : chr "https://data.cms.gov/resources/medicare-inpatient-hospitals-by-provider-data-dictionary"
#>  ..  ..$ site       : chr "https://data.cms.gov/provider-summary-by-type-of-service/medicare-inpatient-hospitals/medicare-inpatient-hospitals-by-provider"
#>  ..  ..$ rows       : int 3120
#>  ..  ..$ pages      : int 1
#>  ..  ..$ fields     : chr [1:57] "Rndrng_Prvdr_CCN" "Rndrng_Prvdr_Org_Name" "Rndrng_Prvdr_St" "Rndrng_Prvdr_City" ...
#>  ..  ..$ endpoints  : tibble [10 × 3] (S3: tbl_df/tbl/data.frame)
#>  ..  .. ..$ year      : int [1:10] 2022 2021 2020 2019 2018 2017 2016 2015 2014 2013
#>  ..  .. ..$ identifier: chr [1:10] "https://data.cms.gov/data-api/v1/dataset/cc73e77b-7367-48c9-b766-fafefe3f8ffa/data" "https://data.cms.gov/data-api/v1/dataset/c7e1501c-a770-4e2a-beb6-32b3525491a7/data" "https://data.cms.gov/data-api/v1/dataset/d7b2810d-ac10-498f-8794-1c3d9f72ee96/data" "https://data.cms.gov/data-api/v1/dataset/c349419e-3c6a-419e-8c74-0641b64cca12/data" ...
#>  ..  .. ..$ resources : chr [1:10] "https://data.cms.gov/data-api/v1/dataset-resources/cc73e77b-7367-48c9-b766-fafefe3f8ffa" "https://data.cms.gov/data-api/v1/dataset-resources/c7e1501c-a770-4e2a-beb6-32b3525491a7" "https://data.cms.gov/data-api/v1/dataset-resources/d7b2810d-ac10-498f-8794-1c3d9f72ee96" "https://data.cms.gov/data-api/v1/dataset-resources/c349419e-3c6a-419e-8c74-0641b64cca12" ...
#>  .. $ by_provider_and_service :List of 7
#>  ..  ..$ description: chr "The Medicare Inpatient Hospitals by Provider and Service dataset provides information on inpatient discharges f"| __truncated__
#>  ..  ..$ dictionary : chr "https://data.cms.gov/resources/medicare-inpatient-hospitals-by-provider-and-service-data-dictionary-0"
#>  ..  ..$ site       : chr "https://data.cms.gov/provider-summary-by-type-of-service/medicare-inpatient-hospitals/medicare-inpatient-hospit"| __truncated__
#>  ..  ..$ rows       : int 145742
#>  ..  ..$ pages      : int 30
#>  ..  ..$ fields     : chr [1:15] "Rndrng_Prvdr_CCN" "Rndrng_Prvdr_Org_Name" "Rndrng_Prvdr_City" "Rndrng_Prvdr_St" ...
#>  ..  ..$ endpoints  : tibble [10 × 3] (S3: tbl_df/tbl/data.frame)
#>  ..  .. ..$ year      : int [1:10] 2022 2021 2020 2019 2018 2017 2016 2015 2014 2013
#>  ..  .. ..$ identifier: chr [1:10] "https://data.cms.gov/data-api/v1/dataset/46bf50f8-0983-4ca2-b8d5-f2afbbf2e589/data" "https://data.cms.gov/data-api/v1/dataset/117d93f2-ce81-40fe-a4d4-8c03203b95e1/data" "https://data.cms.gov/data-api/v1/dataset/ef5bdbe1-27b4-4296-b320-52bd5d2183d7/data" "https://data.cms.gov/data-api/v1/dataset/6f6d93e1-ecf8-4b93-9845-091faf20f274/data" ...
#>  ..  .. ..$ resources : chr [1:10] "https://data.cms.gov/data-api/v1/dataset-resources/46bf50f8-0983-4ca2-b8d5-f2afbbf2e589" "https://data.cms.gov/data-api/v1/dataset-resources/117d93f2-ce81-40fe-a4d4-8c03203b95e1" "https://data.cms.gov/data-api/v1/dataset-resources/ef5bdbe1-27b4-4296-b320-52bd5d2183d7" "https://data.cms.gov/data-api/v1/dataset-resources/6f6d93e1-ecf8-4b93-9845-091faf20f274" ...
#>  @ periodicity: chr "Annually [R/P1Y]"
#>  @ years      : int [1:10] 2022 2021 2020 2019 2018 2017 2016 2015 2014 2013