Build a Query for an Endpoint

Build a Query for an Endpoint

Usage

build(obj, qry)

Arguments

obj

An <endpoint>, <collection> or <group> object.

qry

A <query> object.

Value

A list of query parameters matched to an endpoint's fields.

Examples

qry <- new_query(
   first_name  = starts_with("An"),
   middle_name = ends_with("e"),
   last_name   = contains("J"),
   state       = any_of(c("CA", "GA", "NY")),
   year        = 2021:2025,
   practice_state_or_us_territory = any_of(c("CA", "GA", "NY")))

build(endpoint("enroll_prov"), qry)
#> <class_results>
#>  @ title  : chr "Medicare Fee-For-Service Public Provider Enrollment"
#>  @ params : chr [1:2] "FIRST_NAME" "LAST_NAME"
#>  @ base   : chr "https://data.cms.gov/data-api/v1/dataset/2457ea29-fc82-48b0-86ec-3b0755de7515/data-viewer?offset=0&size=1&filte"| __truncated__
#>  @ limit  : int 5000
#>  @ found  : int 2765
#>  @ total  : int 2894910
#>  @ pages  : int 1
#>  @ problem: logi FALSE

build(endpoint("dial_facility"), qry)
#> <class_results>
#>  @ title  : chr "Medicare Dialysis Facilities"
#>  @ params : chr "state"
#>  @ base   :List of 5
#>  .. $ 2025: chr "https://data.cms.gov/data-api/v1/dataset/7d4e7bd1-f08a-4fca-8085-6af24026a19a/data?offset=0&size=1&filter[1][co"| __truncated__
#>  .. $ 2024: chr "https://data.cms.gov/data-api/v1/dataset/62d125ff-c959-4c5c-8a46-8cacba50bfd4/data?offset=0&size=1&filter[1][co"| __truncated__
#>  .. $ 2023: chr "https://data.cms.gov/data-api/v1/dataset/200e7e7e-a5a6-4450-ba18-ab53d237cc45/data?offset=0&size=1&filter[1][co"| __truncated__
#>  .. $ 2022: chr "https://data.cms.gov/data-api/v1/dataset/ebc64ad1-164c-455a-80ca-cb0753e44a46/data?offset=0&size=1&filter[1][co"| __truncated__
#>  .. $ 2021: chr "https://data.cms.gov/data-api/v1/dataset/e66eaf03-398f-4a8c-9d1b-6aaf25fb6425/data?offset=0&size=1&filter[1][co"| __truncated__
#>  @ limit  : int 5000
#>  @ found  : int [1:5] 2352531 2256637 2252660 2067765 9701304
#>  @ total  : int [1:5] 11985275 11679243 11716732 10821385 9701304
#>  @ pages  : int [1:5] 471 452 451 414 1941
#>  @ problem: logi [1:5] FALSE FALSE FALSE FALSE TRUE

build(endpoint("quality_payment"), qry)
#> <class_results>
#>  @ title  : chr "Quality Payment Program Experience"
#>  @ params : chr "practice state or us territory"
#>  @ base   :List of 3
#>  .. $ 2023: chr "https://data.cms.gov/data-api/v1/dataset/3545dc8e-7037-41ee-8a1d-9f3198fe4927/data?offset=0&size=1&filter[1][co"| __truncated__
#>  .. $ 2022: chr "https://data.cms.gov/data-api/v1/dataset/b3438273-b4a6-44ca-8fb2-9e6026b74642/data?offset=0&size=1&filter[1][co"| __truncated__
#>  .. $ 2021: chr "https://data.cms.gov/data-api/v1/dataset/8c2ab3e4-597e-43c9-9d94-7223c66d7f25/data?offset=0&size=1&filter[1][co"| __truncated__
#>  @ limit  : int 5000
#>  @ found  : int [1:3] 524998 624200 698730
#>  @ total  : int [1:3] 524998 624200 698730
#>  @ pages  : int [1:3] 105 125 140
#>  @ problem: logi [1:3] TRUE TRUE TRUE

build(endpoint("pdc_clinician"), qry)
#> $title
#> [1] "National Downloadable File"
#> 
#> $dimensions
#> <class_dimensions>
#>  @ limit: int 1500
#>  @ rows : int 2750892
#>  @ pages: int 1834
#> 
#> $identifier
#> [1] "https://data.cms.gov/provider-data/api/1/datastore/query/mj5m-pzi6/0?count=true&results=true&offset=0&limit=1"
#> 
#> $params
#> $params$state
#> [1] "conditions[0][property]=state&conditions[0][operator]=in&conditions[0][value][1]=CA&conditions[0][value][2]=GA&conditions[0][value][3]=NY"
#> 
#> 

build(endpoint("grp_cover_nature"), qry)
#> ! Payments Grouped by Covered Recipient and Nature of Payments: 5 params ≠ 10 fields
#> $title
#> [1] "Payments Grouped by Covered Recipient and Nature of Payments"
#> 
#> $dimensions
#> <class_dimensions>
#>  @ limit: int 500
#>  @ rows : int 1235646
#>  @ pages: int 2472
#> 
#> $identifier
#>                                                                                                                                     2024 
#> "https://openpaymentsdata.cms.gov/api/1/datastore/query/88822473-9093-46fc-ad38-81330eb8de4b/0?count=true&results=true&offset=0&limit=1" 
#>                                                                                                                                     2023 
#> "https://openpaymentsdata.cms.gov/api/1/datastore/query/4017acfa-0eeb-4b8a-b68c-3b41e779c4a2/0?count=true&results=true&offset=0&limit=1" 
#>                                                                                                                                     2022 
#> "https://openpaymentsdata.cms.gov/api/1/datastore/query/686b765e-f4b7-4d35-a050-fb5af926dbbe/0?count=true&results=true&offset=0&limit=1" 
#>                                                                                                                                     2021 
#> "https://openpaymentsdata.cms.gov/api/1/datastore/query/d35243be-55cb-5087-931e-a294e0e0917f/0?count=true&results=true&offset=0&limit=1" 
#> 
#> $params
#> NULL
#> 

build(collection("hha"), qry)
#> ! Home Health Agency All Owners: 5 params ≠ 38 fields
#> ! Home Health Agency Cost Report: 5 params ≠ 201 fields
#> $hha_owner
#> <class_results>
#>  @ title  : chr "Home Health Agency All Owners"
#>  @ params : NULL
#>  @ base   : chr "https://data.cms.gov/data-api/v1/dataset/fc009b2d-7846-44b1-b4a1-692f0c143879/data-viewer?offset=0&size=1"
#>  @ limit  : int 5000
#>  @ found  : int 0
#>  @ total  : int 102259
#>  @ pages  : int 1
#>  @ problem: logi FALSE
#> 
#> $hha_enroll
#> <class_results>
#>  @ title  : chr "Home Health Agency Enrollments"
#>  @ params : chr "STATE"
#>  @ base   : chr "https://data.cms.gov/data-api/v1/dataset/15f64ab4-3172-4a27-b589-ebd67a6d28aa/data-viewer?offset=0&size=1&filte"| __truncated__
#>  @ limit  : int 5000
#>  @ found  : int 3042
#>  @ total  : int 11632
#>  @ pages  : int 1
#>  @ problem: logi FALSE
#> 
#> $hha_cost
#> <class_results>
#>  @ title  : chr "Home Health Agency Cost Report"
#>  @ params : NULL
#>  @ base   :List of 2
#>  .. $ 2022: chr "https://data.cms.gov/data-api/v1/dataset/1adf48d9-9ff2-4641-ac51-d827aabf791c/data?offset=0&size=1"
#>  .. $ 2021: chr "https://data.cms.gov/data-api/v1/dataset/a02a324e-a8ad-4933-8bcd-e6a10ab28c8e/data?offset=0&size=1"
#>  @ limit  : int 5000
#>  @ found  : int [1:2] 10564 10379
#>  @ total  : int [1:2] 10564 10379
#>  @ pages  : int [1:2] 3 3
#>  @ problem: logi [1:2] TRUE TRUE
#> 

build(collection("hhc"), qry)
#> ! Home Health Care - Measure Date Range: 5 params ≠ 2 fields
#> ! Home Health Care - National Data: 5 params ≠ 40 fields
#> $hhc_range
#> $hhc_range$title
#> [1] "Home Health Care - Measure Date Range"
#> 
#> $hhc_range$dimensions
#> <class_dimensions>
#>  @ limit: int 1500
#>  @ rows : int 18
#>  @ pages: int 1
#> 
#> $hhc_range$identifier
#> [1] "https://data.cms.gov/provider-data/api/1/datastore/query/c886-nwpj/0?count=true&results=true&offset=0&limit=1"
#> 
#> $hhc_range$params
#> NULL
#> 
#> 
#> $hhc_nation
#> $hhc_nation$title
#> [1] "Home Health Care - National Data"
#> 
#> $hhc_nation$dimensions
#> <class_dimensions>
#>  @ limit: int 1500
#>  @ rows : int 1
#>  @ pages: int 1
#> 
#> $hhc_nation$identifier
#> [1] "https://data.cms.gov/provider-data/api/1/datastore/query/97z8-de96/0?count=true&results=true&offset=0&limit=1"
#> 
#> $hhc_nation$params
#> NULL
#> 
#> 
#> $hhc_state
#> $hhc_state$title
#> [1] "Home Health Care - State by State Data"
#> 
#> $hhc_state$dimensions
#> <class_dimensions>
#>  @ limit: int 1500
#>  @ rows : int 55
#>  @ pages: int 1
#> 
#> $hhc_state$identifier
#> [1] "https://data.cms.gov/provider-data/api/1/datastore/query/tee5-ixt5/0?count=true&results=true&offset=0&limit=1"
#> 
#> $hhc_state$params
#> $hhc_state$params$state
#> [1] "conditions[0][property]=state&conditions[0][operator]=in&conditions[0][value][1]=CA&conditions[0][value][2]=GA&conditions[0][value][3]=NY"
#> 
#> 
#> 
#> $hhc_zip
#> $hhc_zip$title
#> [1] "Home Health Care - Zip Codes"
#> 
#> $hhc_zip$dimensions
#> <class_dimensions>
#>  @ limit: int 1500
#>  @ rows : int 528025
#>  @ pages: int 353
#> 
#> $hhc_zip$identifier
#> [1] "https://data.cms.gov/provider-data/api/1/datastore/query/m5eg-upu5/0?count=true&results=true&offset=0&limit=1"
#> 
#> $hhc_zip$params
#> $hhc_zip$params$state
#> [1] "conditions[0][property]=state&conditions[0][operator]=in&conditions[0][value][1]=CA&conditions[0][value][2]=GA&conditions[0][value][3]=NY"
#> 
#> 
#> 
#> $hhc_agency
#> $hhc_agency$title
#> [1] "Home Health Care Agencies"
#> 
#> $hhc_agency$dimensions
#> <class_dimensions>
#>  @ limit: int 1500
#>  @ rows : int 12112
#>  @ pages: int 9
#> 
#> $hhc_agency$identifier
#> [1] "https://data.cms.gov/provider-data/api/1/datastore/query/6jpm-sxkc/0?count=true&results=true&offset=0&limit=1"
#> 
#> $hhc_agency$params
#> $hhc_agency$params$state
#> [1] "conditions[0][property]=state&conditions[0][operator]=in&conditions[0][value][1]=CA&conditions[0][value][2]=GA&conditions[0][value][3]=NY"
#> 
#> 
#>