Medicaid Endpoint

Medicaid Endpoint

Medicaid Endpoint Group

Usage

caid_endpoint(alias)

caid_group(alias)

Arguments

alias

<chr> title alias

Value

An S7 <caid_endpoint> object.

An S7 <caid_group> object.

Examples

caid_endpoint("MLR")
#> <caid_endpoint>
#>  @ title     : chr "MLR Summary Reports"
#>  @ identifier: chr "https://data.medicaid.gov/api/1/datastore/query/743f9f04-4473-41e2-9da2-9a89db65ee55/0"
#>  @ metadata  : <caid_metadata>
#>  .. @ description: chr "The MLR Summary Reports provide plan-specific MLRs for Medicaid managed care programs. The MLR summary report d"| __truncated__
#>  .. @ modified   : Date[1:1], format: "2024-11-04"
#>  .. @ dictionary : chr NA
#>  .. @ download   : chr "https://download.medicaid.gov/data/mlr-public-use-file-07302024final.csv"
#>  @ dimensions: <caid_dimensions>
#>  .. @ limit : int 8000
#>  .. @ rows  : int 1719
#>  .. @ pages : int 1
#>  .. @ fields:List of 14
#>  .. .. $ state                                            : chr "state"
#>  .. .. $ program_name                                     : chr "program_name"
#>  .. .. $ program_type                                     : chr "program_type"
#>  .. .. $ eligibility_group                                : chr "eligibility_group"
#>  .. .. $ if_other_describe_eligibility_group              : chr "if_other_describe_eligibility_group"
#>  .. .. $ mco_pihp_or_pahp_name                            : chr "mco_pihp_or_pahp_name"
#>  .. .. $ mlr_reporting_period_start_date                  : chr "mlr_reporting_period_start_date"
#>  .. .. $ mlr_reporting_period_end_date                    : chr "mlr_reporting_period_end_date"
#>  .. .. $ report_year                                      : chr "report_year"
#>  .. .. $ mlr_numerator                                    : chr "mlr_numerator"
#>  .. .. $ mlr_denominator                                  : chr "mlr_denominator"
#>  .. .. $ statereported_member_months                      : chr "statereported_member_months"
#>  .. .. $ adjusted_mlr                                     : chr "adjusted_mlr"
#>  .. .. $ remittance_dollar_amount_for_mlr_reporting_period: chr "remittance_dollar_amount_for_mlr_reporting_period"
caid_endpoint("enterprise")
#> <caid_endpoint>
#>  @ title     : chr "Medicaid Enterprise System Datatable"
#>  @ identifier: chr "https://data.medicaid.gov/api/1/datastore/query/13a06cdb-6bbb-4f86-bba7-9d6f3db41090/0"
#>  @ metadata  : <caid_metadata>
#>  .. @ description: chr "Medicaid Enterprise System Datatable"
#>  .. @ modified   : Date[1:1], format: "2025-01-17"
#>  .. @ dictionary : chr NA
#>  .. @ download   : chr "https://download.medicaid.gov/data/data-analysis-table-for-medcaid-01172025.csv"
#>  @ dimensions: <caid_dimensions>
#>  .. @ limit : int 8000
#>  .. @ rows  : int 68
#>  .. @ pages : int 1
#>  .. @ fields:List of 12
#>  .. .. $ implementation_year              : chr "implementation_year"
#>  .. .. $ policyupdate_type                : chr "policyupdate_type"
#>  .. .. $ policyupdate_title               : chr "policyupdate_title"
#>  .. .. $ potential_mes_impact             : chr "potential_mes_impact"
#>  .. .. $ statute_or_regulation_citation   : chr "statute_or_regulation_citation"
#>  .. .. $ compliance_month                 : chr "compliance_month"
#>  .. .. $ compliance_day                   : chr "compliance_day"
#>  .. .. $ type_of_potential_system_impact  : chr "type_of_potential_system_impact"
#>  .. .. $ category_of_potential_impact     : chr "category_of_potential_impact"
#>  .. .. $ link_to__source_document         : chr "link_to__source_document"
#>  .. .. $ impact_to_medicaid_chip_andor_bhp: chr "impact_to_medicaid_chip_andor_bhp"
#>  .. .. $ tshirt_size                      : chr "tshirt_size"
caid_group("demographics")
#> <caid_group>
#>  @ group  : chr "Medicaid and CHIP Enrollee Demographics"
#>  @ members:List of 8
#>  .. $ DEMO_well     : <caid_endpoint>
#>  ..  ..@ title     : chr "Medicaid and CHIP enrollees who received a well-child visit"
#>  ..  ..@ identifier: chr "https://data.medicaid.gov/api/1/datastore/query/289182e5-44c8-4fcc-96c4-4af69be9863f/0"
#>  ..  ..@ metadata  : <caid_metadata>
#>  .. .. .. @ description: chr "This data set includes annual counts and percentages of Medicaid and Childrens Health Insurance Program (CHIP) "| __truncated__
#>  .. .. .. @ modified   : Date[1:1], format: "2025-01-17"
#>  .. .. .. @ dictionary : chr NA
#>  .. .. .. @ download   : chr "https://download.medicaid.gov/data/well-child-visits-2020-2022-01162025.csv"
#>  ..  ..@ dimensions: <caid_dimensions>
#>  .. .. .. @ limit : int 8000
#>  .. .. .. @ rows  : int 63
#>  .. .. .. @ pages : int 1
#>  .. .. .. @ fields:List of 9
#>  .. .. .. .. $ year                          : chr "year"
#>  .. .. .. .. $ geography                     : chr "geography"
#>  .. .. .. .. $ subpopulation_topic           : chr "subpopulation_topic"
#>  .. .. .. .. $ subpopulation                 : chr "subpopulation"
#>  .. .. .. .. $ category                      : chr "category"
#>  .. .. .. .. $ count_of_enrollees            : chr "count_of_enrollees"
#>  .. .. .. .. $ denominator_count_of_enrollees: chr "denominator_count_of_enrollees"
#>  .. .. .. .. $ percentage_of_enrollees       : chr "percentage_of_enrollees"
#>  .. .. .. .. $ data_version                  : chr "data_version"
#>  .. $ DEMO_sud      : <caid_endpoint>
#>  ..  ..@ title     : chr "Medicaid and CHIP enrollees who received mental health or SUD services"
#>  ..  ..@ identifier: chr "https://data.medicaid.gov/api/1/datastore/query/8062e2f4-4c0a-41c9-8217-979468a80986/0"
#>  ..  ..@ metadata  : <caid_metadata>
#>  .. .. .. @ description: chr "This data set includes annual counts and percentages of Medicaid and Childrens Health Insurance Program (CHIP) "| __truncated__
#>  .. .. .. @ modified   : Date[1:1], format: "2025-01-17"
#>  .. .. .. @ dictionary : chr NA
#>  .. .. .. @ download   : chr "https://download.medicaid.gov/data/mh-and-sud-services-2020-2022-01162025.csv"
#>  ..  ..@ dimensions: <caid_dimensions>
#>  .. .. .. @ limit : int 8000
#>  .. .. .. @ rows  : int 216
#>  .. .. .. @ pages : int 1
#>  .. .. .. @ fields:List of 9
#>  .. .. .. .. $ year                          : chr "year"
#>  .. .. .. .. $ geography                     : chr "geography"
#>  .. .. .. .. $ subpopulation_topic           : chr "subpopulation_topic"
#>  .. .. .. .. $ subpopulation                 : chr "subpopulation"
#>  .. .. .. .. $ category                      : chr "category"
#>  .. .. .. .. $ count_of_enrollees            : chr "count_of_enrollees"
#>  .. .. .. .. $ denominator_count_of_enrollees: chr "denominator_count_of_enrollees"
#>  .. .. .. .. $ percentage_of_enrollees       : chr "percentage_of_enrollees"
#>  .. .. .. .. $ data_version                  : chr "data_version"
#>  .. $ DEMO_disable  : <caid_endpoint>
#>  ..  ..@ title     : chr "Medicaid enrollees who qualify for benefits based on disability"
#>  ..  ..@ identifier: chr "https://data.medicaid.gov/api/1/datastore/query/b12e390f-1721-47c0-bb38-b0b1805e0446/0"
#>  ..  ..@ metadata  : <caid_metadata>
#>  .. .. .. @ description: chr "This data set includes annual counts and percentages of Medicaid enrollees who are eligible for benefits based "| __truncated__
#>  .. .. .. @ modified   : Date[1:1], format: "2025-01-17"
#>  .. .. .. @ dictionary : chr NA
#>  .. .. .. @ download   : chr "https://download.medicaid.gov/data/disability-2020-2022-01162025.csv"
#>  ..  ..@ dimensions: <caid_dimensions>
#>  .. .. .. @ limit : int 8000
#>  .. .. .. @ rows  : int 114
#>  .. .. .. @ pages : int 1
#>  .. .. .. @ fields:List of 9
#>  .. .. .. .. $ year                          : chr "year"
#>  .. .. .. .. $ geography                     : chr "geography"
#>  .. .. .. .. $ subpopulation_topic           : chr "subpopulation_topic"
#>  .. .. .. .. $ subpopulation                 : chr "subpopulation"
#>  .. .. .. .. $ category                      : chr "category"
#>  .. .. .. .. $ count_of_enrollees            : chr "count_of_enrollees"
#>  .. .. .. .. $ denominator_count_of_enrollees: chr "denominator_count_of_enrollees"
#>  .. .. .. .. $ percentage_of_enrollees       : chr "percentage_of_enrollees"
#>  .. .. .. .. $ data_version                  : chr "data_version"
#>  .. $ DEMO_premature: <caid_endpoint>
#>  ..  ..@ title     : chr "Prematurity and severe maternal morbidity among Medicaid- and CHIP-covered live births"
#>  ..  ..@ identifier: chr "https://data.medicaid.gov/api/1/datastore/query/ee3b9534-0d19-4c1b-bf74-43f898d5de7c/0"
#>  ..  ..@ metadata  : <caid_metadata>
#>  .. .. .. @ description: chr "This data set includes annual counts and rates of Medicaid- and Childrens Health Insurance Program (CHIP)-cover"| __truncated__
#>  .. .. .. @ modified   : Date[1:1], format: "2025-01-17"
#>  .. .. .. @ dictionary : chr NA
#>  .. .. .. @ download   : chr "https://download.medicaid.gov/data/preterm-birth-and-smm-2020-2022-01172025.csv"
#>  ..  ..@ dimensions: <caid_dimensions>
#>  .. .. .. @ limit : int 8000
#>  .. .. .. @ rows  : int 435
#>  .. .. .. @ pages : int 1
#>  .. .. .. @ fields:List of 9
#>  .. .. .. .. $ year                           : chr "year"
#>  .. .. .. .. $ geography                      : chr "geography"
#>  .. .. .. .. $ subpopulation_topic            : chr "subpopulation_topic"
#>  .. .. .. .. $ subpopulation                  : chr "subpopulation"
#>  .. .. .. .. $ category                       : chr "category"
#>  .. .. .. .. $ count_of_deliveries            : chr "count_of_deliveries"
#>  .. .. .. .. $ denominator_count_of_deliveries: chr "denominator_count_of_deliveries"
#>  .. .. .. .. $ rate_of_deliveries             : chr "rate_of_deliveries"
#>  .. .. .. .. $ data_version                   : chr "data_version"
#>  .. $ DEMO_language : <caid_endpoint>
#>  ..  ..@ title     : chr "Primary language spoken by the Medicaid and CHIP population"
#>  ..  ..@ identifier: chr "https://data.medicaid.gov/api/1/datastore/query/3df86994-c0e4-4106-a404-e95e87c824e6/0"
#>  ..  ..@ metadata  : <caid_metadata>
#>  .. .. .. @ description: chr "This data set includes annual counts and percentages of Medicaid and Childrens Health Insurance Program (CHIP) "| __truncated__
#>  .. .. .. @ modified   : Date[1:1], format: "2025-01-18"
#>  .. .. .. @ dictionary : chr NA
#>  .. .. .. @ download   : chr "https://download.medicaid.gov/data/primary-language-2020-2022-01162025.csv"
#>  ..  ..@ dimensions: <caid_dimensions>
#>  .. .. .. @ limit : int 8000
#>  .. .. .. @ rows  : int 675
#>  .. .. .. @ pages : int 1
#>  .. .. .. @ fields:List of 9
#>  .. .. .. .. $ year                          : chr "year"
#>  .. .. .. .. $ geography                     : chr "geography"
#>  .. .. .. .. $ subpopulation_topic           : chr "subpopulation_topic"
#>  .. .. .. .. $ subpopulation                 : chr "subpopulation"
#>  .. .. .. .. $ category                      : chr "category"
#>  .. .. .. .. $ count_of_enrollees            : chr "count_of_enrollees"
#>  .. .. .. .. $ denominator_count_of_enrollees: chr "denominator_count_of_enrollees"
#>  .. .. .. .. $ percentage_of_enrollees       : chr "percentage_of_enrollees"
#>  .. .. .. .. $ data_version                  : chr "data_version"
#>  .. $ DEMO_race     : <caid_endpoint>
#>  ..  ..@ title     : chr "Race and ethnicity of the national Medicaid and CHIP population"
#>  ..  ..@ identifier: chr "https://data.medicaid.gov/api/1/datastore/query/b9e18d80-c765-4fb9-95e7-81951e01a6a3/0"
#>  ..  ..@ metadata  : <caid_metadata>
#>  .. .. .. @ description: chr "This data set includes annual counts and percentages of Medicaid and Childrens Health Insurance Program (CHIP) "| __truncated__
#>  .. .. .. @ modified   : Date[1:1], format: "2025-01-16"
#>  .. .. .. @ dictionary : chr NA
#>  .. .. .. @ download   : chr "https://download.medicaid.gov/data/race-and-ethnicity-2020-2022-01162025.csv"
#>  ..  ..@ dimensions: <caid_dimensions>
#>  .. .. .. @ limit : int 8000
#>  .. .. .. @ rows  : int 216
#>  .. .. .. @ pages : int 1
#>  .. .. .. @ fields:List of 9
#>  .. .. .. .. $ year                          : chr "year"
#>  .. .. .. .. $ geography                     : chr "geography"
#>  .. .. .. .. $ subpopulation_topic           : chr "subpopulation_topic"
#>  .. .. .. .. $ subpopulation                 : chr "subpopulation"
#>  .. .. .. .. $ category                      : chr "category"
#>  .. .. .. .. $ count_of_enrollees            : chr "count_of_enrollees"
#>  .. .. .. .. $ denominator_count_of_enrollees: chr "denominator_count_of_enrollees"
#>  .. .. .. .. $ percentage_of_enrollees       : chr "percentage_of_enrollees"
#>  .. .. .. .. $ data_version                  : chr "data_version"
#>  .. $ DEMO_rural    : <caid_endpoint>
#>  ..  ..@ title     : chr "Rural Medicaid and CHIP enrollees"
#>  ..  ..@ identifier: chr "https://data.medicaid.gov/api/1/datastore/query/73fdbf7c-569e-427b-9bff-eeed34a51dcb/0"
#>  ..  ..@ metadata  : <caid_metadata>
#>  .. .. .. @ description: chr "This data set includes annual counts and percentages of Medicaid and Childrens Health Insurance Program (CHIP) "| __truncated__
#>  .. .. .. @ modified   : Date[1:1], format: "2025-01-17"
#>  .. .. .. @ dictionary : chr NA
#>  .. .. .. @ download   : chr "https://download.medicaid.gov/data/rural-residence-2020-2022-01162025.csv"
#>  ..  ..@ dimensions: <caid_dimensions>
#>  .. .. .. @ limit : int 8000
#>  .. .. .. @ rows  : int 603
#>  .. .. .. @ pages : int 1
#>  .. .. .. @ fields:List of 9
#>  .. .. .. .. $ year                          : chr "year"
#>  .. .. .. .. $ geography                     : chr "geography"
#>  .. .. .. .. $ subpopulation_topic           : chr "subpopulation_topic"
#>  .. .. .. .. $ subpopulation                 : chr "subpopulation"
#>  .. .. .. .. $ category                      : chr "category"
#>  .. .. .. .. $ count_of_enrollees            : chr "count_of_enrollees"
#>  .. .. .. .. $ denominator_count_of_enrollees: chr "denominator_count_of_enrollees"
#>  .. .. .. .. $ percentage_of_enrollees       : chr "percentage_of_enrollees"
#>  .. .. .. .. $ data_version                  : chr "data_version"
#>  .. $ DEMO_waive    : <caid_endpoint>
#>  ..  ..@ title     : chr "Section 1915(c) waiver program participants"
#>  ..  ..@ identifier: chr "https://data.medicaid.gov/api/1/datastore/query/4d4eaf55-33d3-4468-80b4-63553f4530ae/0"
#>  ..  ..@ metadata  : <caid_metadata>
#>  .. .. .. @ description: chr "This data set includes annual counts and percentages of Medicaid and Childrens Health Insurance Program (CHIP) "| __truncated__
#>  .. .. .. @ modified   : Date[1:1], format: "2025-01-17"
#>  .. .. .. @ dictionary : chr NA
#>  .. .. .. @ download   : chr "https://download.medicaid.gov/data/1915c-waiver-2020-2022-01162025.csv"
#>  ..  ..@ dimensions: <caid_dimensions>
#>  .. .. .. @ limit : int 8000
#>  .. .. .. @ rows  : int 450
#>  .. .. .. @ pages : int 1
#>  .. .. .. @ fields:List of 9
#>  .. .. .. .. $ year                          : chr "year"
#>  .. .. .. .. $ geography                     : chr "geography"
#>  .. .. .. .. $ subpopulation_topic           : chr "subpopulation_topic"
#>  .. .. .. .. $ subpopulation                 : chr "subpopulation"
#>  .. .. .. .. $ category                      : chr "category"
#>  .. .. .. .. $ count_of_enrollees            : chr "count_of_enrollees"
#>  .. .. .. .. $ denominator_count_of_enrollees: chr "denominator_count_of_enrollees"
#>  .. .. .. .. $ percentage_of_enrollees       : chr "percentage_of_enrollees"
#>  .. .. .. .. $ data_version                  : chr "data_version"