Medicare Endpoint Group

Medicare Endpoint Group

Usage

careGroup(alias)

Arguments

alias

<chr> title alias

Value

An S7 <careGroup> object.

Examples

careGroup("HHA")
#> <careGroup>
#>  @ group  : chr "Home Health Agencies"
#>  @ members:List of 3
#>  .. $ HHA_owners     : <careMain>
#>  ..  ..@ title     : chr "Home Health Agency All Owners"
#>  ..  ..@ metadata  : <careMeta>
#>  .. .. .. @ description: chr "The Home Health Agency (HHA) All Owners dataset provides ownership information on all HHAs currently enrolled i"| __truncated__
#>  .. .. .. @ temporal   : chr "2023-04-01 • 2025-03-31"
#>  .. .. .. @ periodicity: chr "Quarterly [R/P3M]"
#>  .. .. .. @ modified   : Date[1:1], format: "2025-01-15"
#>  .. .. .. @ dictionary : chr "https://data.cms.gov/sites/default/files/2024-10/HHA_All_Owners_Data_Dictionary.pdf"
#>  .. .. .. @ site       : chr "https://data.cms.gov/provider-characteristics/hospitals-and-other-facilities/home-health-agency-all-owners"
#>  .. .. .. @ references : chr "https://data.cms.gov/sites/default/files/2024-10/HHA_Data_Guidance.pdf"
#>  .. .. .. @ download   : chr "https://data.cms.gov/sites/default/files/2025-01/a0443d0f-247e-481e-8d62-cdd0621238d4/HHA_All_Owners_2025.01.02.csv"
#>  ..  ..@ dimensions: <careDim>
#>  .. .. .. @ rows  : int 102925
#>  .. .. .. @ pages : int 21
#>  .. .. .. @ fields: chr [1:34] "ENROLLMENT ID" "ASSOCIATE ID" "ORGANIZATION NAME" "ASSOCIATE ID - OWNER" ...
#>  ..  ..@ identifier: chr "https://data.cms.gov/data-api/v1/dataset/fc009b2d-7846-44b1-b4a1-692f0c143879/data-viewer"
#>  ..  ..@ resources : chr "https://data.cms.gov/data-api/v1/dataset-resources/fc009b2d-7846-44b1-b4a1-692f0c143879"
#>  .. $ HHA_cost_report: <careMain>
#>  ..  ..@ title     : chr "Home Health Agency Cost Report"
#>  ..  ..@ metadata  : <careMeta>
#>  .. .. .. @ description: chr "The Home Health Agency Provider Cost Report dataset provides select measures from the home health agency annual"| __truncated__
#>  .. .. .. @ temporal   : chr "2020-01-01 • 2022-12-31"
#>  .. .. .. @ periodicity: chr "Annually [R/P1Y]"
#>  .. .. .. @ modified   : Date[1:1], format: "2024-10-25"
#>  .. .. .. @ dictionary : chr "https://data.cms.gov/resources/home-health-agency-cost-report-data-dictionary"
#>  .. .. .. @ site       : chr "https://data.cms.gov/provider-compliance/cost-report/home-health-agency-cost-report"
#>  .. .. .. @ references : chr "https://data.cms.gov/resources/home-health-agency-cost-report-methodology"
#>  .. .. .. @ download   : chr "https://data.cms.gov/sites/default/files/2024-10/19db07e2-b626-480a-8554-8cb2ea64f5df/CostReporthha_Final_22.csv"
#>  ..  ..@ dimensions: <careDim>
#>  .. .. .. @ rows  : int 10564
#>  .. .. .. @ pages : int 3
#>  .. .. .. @ fields: chr [1:201] "rpt_rec_num" "Provider CCN" "HHA Name" "Street Address" ...
#>  ..  ..@ identifier: chr "https://data.cms.gov/data-api/v1/dataset/4999da74-1d8d-4a6f-934e-2d7ea470cc63/data-viewer"
#>  ..  ..@ resources : chr "https://data.cms.gov/data-api/v1/dataset-resources/4999da74-1d8d-4a6f-934e-2d7ea470cc63"
#>  .. $ HHA_enrollments: <careMain>
#>  ..  ..@ title     : chr "Home Health Agency Enrollments"
#>  ..  ..@ metadata  : <careMeta>
#>  .. .. .. @ description: chr "The Home Health Agency (HHA) Enrollments dataset provides enrollment information on all HHAs currently enrolled"| __truncated__
#>  .. .. .. @ temporal   : chr "2023-04-01 • 2025-03-31"
#>  .. .. .. @ periodicity: chr "Quarterly [R/P3M]"
#>  .. .. .. @ modified   : Date[1:1], format: "2025-01-15"
#>  .. .. .. @ dictionary : chr "https://data.cms.gov/sites/default/files/2024-10/HHA_Enrollments_Data_Dictionary.pdf"
#>  .. .. .. @ site       : chr "https://data.cms.gov/provider-characteristics/hospitals-and-other-facilities/home-health-agency-enrollments"
#>  .. .. .. @ references : chr "https://data.cms.gov/sites/default/files/2024-10/HHA_Data_Guidance.pdf"
#>  .. .. .. @ download   : chr "https://data.cms.gov/sites/default/files/2025-01/c133f82f-5635-47d3-b85f-364b379a322a/HHA_Enrollments_2025.01.02.csv"
#>  ..  ..@ dimensions: <careDim>
#>  .. .. .. @ rows  : int 11708
#>  .. .. .. @ pages : int 3
#>  .. .. .. @ fields: chr [1:22] "ENROLLMENT ID" "ENROLLMENT STATE" "PROVIDER TYPE CODE" "PROVIDER TYPE TEXT" ...
#>  ..  ..@ identifier: chr "https://data.cms.gov/data-api/v1/dataset/15f64ab4-3172-4a27-b589-ebd67a6d28aa/data-viewer"
#>  ..  ..@ resources : chr "https://data.cms.gov/data-api/v1/dataset-resources/15f64ab4-3172-4a27-b589-ebd67a6d28aa"
careGroup("hospice")
#> <careGroup>
#>  @ group  : chr "Hospices"
#>  @ members:List of 3
#>  .. $ hospice_owners     : <careMain>
#>  ..  ..@ title     : chr "Hospice All Owners"
#>  ..  ..@ metadata  : <careMeta>
#>  .. .. .. @ description: chr "The  Hospice All Owners dataset provides ownership information on all hospices currently enrolled in Medicare. "| __truncated__
#>  .. .. .. @ temporal   : chr "2023-04-01 • 2025-03-31"
#>  .. .. .. @ periodicity: chr "Quarterly [R/P3M]"
#>  .. .. .. @ modified   : Date[1:1], format: "2025-01-15"
#>  .. .. .. @ dictionary : chr "https://data.cms.gov/sites/default/files/2024-10/Hospice_All_Owners_Data_Dictionary.pdf"
#>  .. .. .. @ site       : chr "https://data.cms.gov/provider-characteristics/hospitals-and-other-facilities/hospice-all-owners"
#>  .. .. .. @ references : chr "https://data.cms.gov/sites/default/files/2023-03/f193e9ef-9fd7-4de3-9a27-0d5281cf80a9/Hospice_Data_Guidance.pdf"
#>  .. .. .. @ download   : chr "https://data.cms.gov/sites/default/files/2025-01/662c8886-9074-4669-bbfb-3807ba85b484/Hospice_All_Owners_2025.01.02.csv"
#>  ..  ..@ dimensions: <careDim>
#>  .. .. .. @ rows  : int 71905
#>  .. .. .. @ pages : int 15
#>  .. .. .. @ fields: chr [1:34] "ENROLLMENT ID" "ASSOCIATE ID" "ORGANIZATION NAME" "ASSOCIATE ID - OWNER" ...
#>  ..  ..@ identifier: chr "https://data.cms.gov/data-api/v1/dataset/e983965e-1603-4cb8-82b5-c40090e380d1/data-viewer"
#>  ..  ..@ resources : chr "https://data.cms.gov/data-api/v1/dataset-resources/e983965e-1603-4cb8-82b5-c40090e380d1"
#>  .. $ hospice_enrollments: <careMain>
#>  ..  ..@ title     : chr "Hospice Enrollments"
#>  ..  ..@ metadata  : <careMeta>
#>  .. .. .. @ description: chr "The Hospice Enrollments dataset provides enrollment information of all hospices currently enrolled in Medicare."| __truncated__
#>  .. .. .. @ temporal   : chr "2023-04-01 • 2025-03-31"
#>  .. .. .. @ periodicity: chr "Quarterly [R/P3M]"
#>  .. .. .. @ modified   : Date[1:1], format: "2025-01-15"
#>  .. .. .. @ dictionary : chr "https://data.cms.gov/sites/default/files/2023-03/f736ce50-3d6b-4ee2-802f-2e5424ae71ac/Hospice_Enrollments_Data_Dictionary.pdf"
#>  .. .. .. @ site       : chr "https://data.cms.gov/provider-characteristics/hospitals-and-other-facilities/hospice-enrollments"
#>  .. .. .. @ references : chr "https://data.cms.gov/sites/default/files/2023-03/f193e9ef-9fd7-4de3-9a27-0d5281cf80a9/Hospice_Data_Guidance.pdf"
#>  .. .. .. @ download   : chr "https://data.cms.gov/sites/default/files/2025-01/afd62e57-8a99-4c49-a0b5-358a25ce39dd/Hospice_Enrollments_2025.01.02.csv"
#>  ..  ..@ dimensions: <careDim>
#>  .. .. .. @ rows  : int 6603
#>  .. .. .. @ pages : int 2
#>  .. .. .. @ fields: chr [1:20] "ENROLLMENT ID" "ENROLLMENT STATE" "PROVIDER TYPE CODE" "PROVIDER TYPE TEXT" ...
#>  ..  ..@ identifier: chr "https://data.cms.gov/data-api/v1/dataset/25704213-e833-4b8b-9dbc-58dd17149209/data-viewer"
#>  ..  ..@ resources : chr "https://data.cms.gov/data-api/v1/dataset-resources/25704213-e833-4b8b-9dbc-58dd17149209"
#>  .. $ hospice_acute      : <careMain>
#>  ..  ..@ title     : chr "Medicare Post-Acute Care and Hospice - by Geography & Provider"
#>  ..  ..@ metadata  : <careMeta>
#>  .. .. .. @ description: chr "The Medicare Post-Acute Care and Hospice Provider Utilization and Payment Public Use Files (PAC PUF) contains i"| __truncated__
#>  .. .. .. @ temporal   : chr "2022-01-01 • 2022-12-31"
#>  .. .. .. @ periodicity: chr "Annually [R/P1Y]"
#>  .. .. .. @ modified   : Date[1:1], format: "2024-06-14"
#>  .. .. .. @ dictionary : chr "https://data.cms.gov/sites/default/files/2024-05/MUP_PAC_RY24_Data_Dictionary_2024_FINAL.xlsx"
#>  .. .. .. @ site       : chr "https://data.cms.gov/provider-summary-by-type-of-service/medicare-post-acute-care-hospice/medicare-post-acute-c"| __truncated__
#>  .. .. .. @ references : chr "https://data.cms.gov/sites/default/files/2024-05/MUP_PAC_RY24_Methodology_2024_508.pdf"
#>  .. .. .. @ download   : chr "https://data.cms.gov/sites/default/files/2024-06/e19ecccb-355b-479f-9fd1-2e1250822bd8/ry2024_pacpuf_main_2020_2022s.csv"
#>  ..  ..@ dimensions: <careDim>
#>  .. .. .. @ rows  : int 89392
#>  .. .. .. @ pages : int 18
#>  .. .. .. @ fields: chr [1:88] "YEAR" "YEAR_TYPE" "SMRY_CTGRY" "SRVC_CTGRY" ...
#>  ..  ..@ identifier: chr "https://data.cms.gov/data-api/v1/dataset/fa8caa65-d5c3-492e-be49-835fe42c9190/data-viewer"
#>  ..  ..@ resources : chr "https://data.cms.gov/data-api/v1/dataset-resources/fa8caa65-d5c3-492e-be49-835fe42c9190"
careGroup("hospital")
#> <careGroup>
#>  @ group  : chr "Hospitals"
#>  @ members:List of 4
#>  .. $ hospital_owners     : <careMain>
#>  ..  ..@ title     : chr "Hospital All Owners"
#>  ..  ..@ metadata  : <careMeta>
#>  .. .. .. @ description: chr "The Hospital All Owners Information dataset provides information on all owners of the hospitals. This data incl"| __truncated__
#>  .. .. .. @ temporal   : chr "2022-11-01 • 2025-01-31"
#>  .. .. .. @ periodicity: chr "Monthly [R/P1M]"
#>  .. .. .. @ modified   : Date[1:1], format: "2025-01-15"
#>  .. .. .. @ dictionary : chr "https://data.cms.gov/sites/default/files/2024-10/Hospital_All_Owners_Data_Dictionary.pdf"
#>  .. .. .. @ site       : chr "https://data.cms.gov/provider-characteristics/hospitals-and-other-facilities/hospital-all-owners"
#>  .. .. .. @ references : chr "https://data.cms.gov/sites/default/files/2024-10/Hospital_Data_Guidance.pdf"
#>  .. .. .. @ download   : chr "https://data.cms.gov/sites/default/files/2025-01/d20ffa23-37f1-40ca-8ecd-03413f021acb/Hospital_All_Owners_2025.01.02.csv"
#>  ..  ..@ dimensions: <careDim>
#>  .. .. .. @ rows  : int 148305
#>  .. .. .. @ pages : int 30
#>  .. .. .. @ fields: chr [1:34] "ENROLLMENT ID" "ASSOCIATE ID" "ORGANIZATION NAME" "ASSOCIATE ID - OWNER" ...
#>  ..  ..@ identifier: chr "https://data.cms.gov/data-api/v1/dataset/029c119f-f79c-49be-9100-344d31d10344/data-viewer"
#>  ..  ..@ resources : chr "https://data.cms.gov/data-api/v1/dataset-resources/029c119f-f79c-49be-9100-344d31d10344"
#>  .. $ hospital_chow       : <careMain>
#>  ..  ..@ title     : chr "Hospital Change of Ownership"
#>  ..  ..@ metadata  : <careMeta>
#>  .. .. .. @ description: chr "The Hospital Change of Ownership (CHOW) dataset provides information on the hospital ownership changes that occ"| __truncated__
#>  .. .. .. @ temporal   : chr "2022-01-01 • 2024-12-31"
#>  .. .. .. @ periodicity: chr "Quarterly [R/P3M]"
#>  .. .. .. @ modified   : Date[1:1], format: "2025-01-15"
#>  .. .. .. @ dictionary : chr "https://data.cms.gov/sites/default/files/2023-01/Hospital_CHOW_Data_Dictionary_2022.12.30.pdf"
#>  .. .. .. @ site       : chr "https://data.cms.gov/provider-characteristics/hospitals-and-other-facilities/hospital-change-of-ownership"
#>  .. .. .. @ references : chr "https://data.cms.gov/sites/default/files/2024-10/Hospital_CHOW_Data_Guidance.pdf"
#>  .. .. .. @ download   : chr "https://data.cms.gov/sites/default/files/2025-01/2e9ddac3-eb7e-4805-a3e0-9c52152af37a/Hospital_CHOW_2025.01.02.csv"
#>  ..  ..@ dimensions: <careDim>
#>  .. .. .. @ rows  : int 707
#>  .. .. .. @ pages : int 1
#>  .. .. .. @ fields: chr [1:23] "ENROLLMENT ID - BUYER" "ENROLLMENT STATE - BUYER" "PROVIDER TYPE CODE - BUYER" "PROVIDER TYPE TEXT - BUYER" ...
#>  ..  ..@ identifier: chr "https://data.cms.gov/data-api/v1/dataset/c04031db-54ce-461c-85d1-d2613d71f167/data-viewer"
#>  ..  ..@ resources : chr "https://data.cms.gov/data-api/v1/dataset-resources/c04031db-54ce-461c-85d1-d2613d71f167"
#>  .. $ hospital_chow_owner : <careMain>
#>  ..  ..@ title     : chr "Hospital Change of Ownership - Owner Information"
#>  ..  ..@ metadata  : <careMeta>
#>  .. .. .. @ description: chr "The Hospital Change of Ownership (CHOW) - Owner Information dataset provides information on individual and orga"| __truncated__
#>  .. .. .. @ temporal   : chr "2022-01-01 • 2024-12-31"
#>  .. .. .. @ periodicity: chr "Quarterly [R/P3M]"
#>  .. .. .. @ modified   : Date[1:1], format: "2025-01-15"
#>  .. .. .. @ dictionary : chr "https://data.cms.gov/sites/default/files/2024-10/Hospital_CHOW_Owners_Data_Dictionary.pdf"
#>  .. .. .. @ site       : chr "https://data.cms.gov/provider-characteristics/hospitals-and-other-facilities/hospital-change-of-ownership-owner-information"
#>  .. .. .. @ references : chr "https://data.cms.gov/sites/default/files/2024-10/Hospital_CHOW_Data_Guidance.pdf"
#>  .. .. .. @ download   : chr "https://data.cms.gov/sites/default/files/2025-01/e42d8d95-39eb-43a4-86ca-de774b00c7ca/Hospital_CHOW_Owners_2025.01.02.csv"
#>  ..  ..@ dimensions: <careDim>
#>  .. .. .. @ rows  : int 20909
#>  .. .. .. @ pages : int 5
#>  .. .. .. @ fields: chr [1:34] "ENROLLMENT ID" "ASSOCIATE ID" "ORGANIZATION NAME" "ASSOCIATE ID - OWNER" ...
#>  ..  ..@ identifier: chr "https://data.cms.gov/data-api/v1/dataset/60625dc8-b621-45f0-9423-077fd133b13e/data-viewer"
#>  ..  ..@ resources : chr "https://data.cms.gov/data-api/v1/dataset-resources/60625dc8-b621-45f0-9423-077fd133b13e"
#>  .. $ hospital_enrollments: <careMain>
#>  ..  ..@ title     : chr "Hospital Enrollments"
#>  ..  ..@ metadata  : <careMeta>
#>  .. .. .. @ description: chr "The Hospital Enrollments dataset provides enrollment information of all Hospitals currently enrolled in Medicar"| __truncated__
#>  .. .. .. @ temporal   : chr "2022-11-01 • 2025-01-31"
#>  .. .. .. @ periodicity: chr "Monthly [R/P1M]"
#>  .. .. .. @ modified   : Date[1:1], format: "2025-01-15"
#>  .. .. .. @ dictionary : chr "https://data.cms.gov/sites/default/files/2023-07/Hospital_Enrollments_Data_Dictionary.pdf"
#>  .. .. .. @ site       : chr "https://data.cms.gov/provider-characteristics/hospitals-and-other-facilities/hospital-enrollments"
#>  .. .. .. @ references : chr "https://data.cms.gov/sites/default/files/2024-10/Hospital_Data_Guidance.pdf"
#>  .. .. .. @ download   : chr "https://data.cms.gov/sites/default/files/2025-01/0ef8ef0d-bf93-4b00-a542-a4e2853947cc/Hospital_Enrollments_2025.01.02.csv"
#>  ..  ..@ dimensions: <careDim>
#>  .. .. .. @ rows  : int 9272
#>  .. .. .. @ pages : int 2
#>  .. .. .. @ fields: chr [1:39] "ENROLLMENT ID" "ENROLLMENT STATE" "PROVIDER TYPE CODE" "PROVIDER TYPE TEXT" ...
#>  ..  ..@ identifier: chr "https://data.cms.gov/data-api/v1/dataset/f6f6505c-e8b0-4d57-b258-e2b94133aaf2/data-viewer"
#>  ..  ..@ resources : chr "https://data.cms.gov/data-api/v1/dataset-resources/f6f6505c-e8b0-4d57-b258-e2b94133aaf2"
careGroup("RHC")
#> <careGroup>
#>  @ group  : chr "Rural Health Clinics"
#>  @ members:List of 2
#>  .. $ RHC_owners     : <careMain>
#>  ..  ..@ title     : chr "Rural Health Clinic All Owners"
#>  ..  ..@ metadata  : <careMeta>
#>  .. .. .. @ description: chr "The Rural Health Clinic (RHC) All Owners dataset provides ownership information on all  RHCs currently enrolled"| __truncated__
#>  .. .. .. @ temporal   : chr "2023-10-01 • 2025-03-31"
#>  .. .. .. @ periodicity: chr "Quarterly [R/P3M]"
#>  .. .. .. @ modified   : Date[1:1], format: "2025-01-15"
#>  .. .. .. @ dictionary : chr "https://data.cms.gov/sites/default/files/2023-09/RHC_All_Owners_Data_Dictionary.pdf"
#>  .. .. .. @ site       : chr "https://data.cms.gov/provider-characteristics/hospitals-and-other-facilities/rural-health-clinic-all-owners"
#>  .. .. .. @ references : chr "https://data.cms.gov/sites/default/files/2024-10/RHC_Data_Guidance.pdf"
#>  .. .. .. @ download   : chr "https://data.cms.gov/sites/default/files/2025-01/da02dffa-e26d-4c9c-85cb-3d8df1b98613/RHC_All_Owners_2025.01.02.csv"
#>  ..  ..@ dimensions: <careDim>
#>  .. .. .. @ rows  : int 62784
#>  .. .. .. @ pages : int 13
#>  .. .. .. @ fields: chr [1:34] "ENROLLMENT ID" "ASSOCIATE ID" "ORGANIZATION NAME" "ASSOCIATE ID - OWNER" ...
#>  ..  ..@ identifier: chr "https://data.cms.gov/data-api/v1/dataset/ab03c9bc-0c22-4ca4-b032-21dd3408210d/data-viewer"
#>  ..  ..@ resources : chr "https://data.cms.gov/data-api/v1/dataset-resources/ab03c9bc-0c22-4ca4-b032-21dd3408210d"
#>  .. $ RHC_enrollments: <careMain>
#>  ..  ..@ title     : chr "Rural Health Clinic Enrollments"
#>  ..  ..@ metadata  : <careMeta>
#>  .. .. .. @ description: chr "The Rural Health Clinic (RHC) Enrollments dataset provides enrollment information on all RHCs currently enrolle"| __truncated__
#>  .. .. .. @ temporal   : chr "2023-10-01 • 2025-03-31"
#>  .. .. .. @ periodicity: chr "Quarterly [R/P3M]"
#>  .. .. .. @ modified   : Date[1:1], format: "2025-01-15"
#>  .. .. .. @ dictionary : chr "https://data.cms.gov/sites/default/files/2023-09/RHC_Enrollments_Data_Dictionary.pdf"
#>  .. .. .. @ site       : chr "https://data.cms.gov/provider-characteristics/hospitals-and-other-facilities/rural-health-clinic-enrollments"
#>  .. .. .. @ references : chr "https://data.cms.gov/sites/default/files/2024-10/RHC_Data_Guidance.pdf"
#>  .. .. .. @ download   : chr "https://data.cms.gov/sites/default/files/2025-01/31a07f52-068f-4849-8731-aa0216554368/RHC_Enrollments_2025.01.02.csv"
#>  ..  ..@ dimensions: <careDim>
#>  .. .. .. @ rows  : int 5399
#>  .. .. .. @ pages : int 2
#>  .. .. .. @ fields: chr [1:20] "ENROLLMENT ID" "ENROLLMENT STATE" "PROVIDER TYPE CODE" "PROVIDER TYPE TEXT" ...
#>  ..  ..@ identifier: chr "https://data.cms.gov/data-api/v1/dataset/3b7e7659-067e-41ea-8e36-f9ee2036e1f6/data-viewer"
#>  ..  ..@ resources : chr "https://data.cms.gov/data-api/v1/dataset-resources/3b7e7659-067e-41ea-8e36-f9ee2036e1f6"
careGroup("FQHC")
#> <careGroup>
#>  @ group  : chr "Federally Qualified Health Centers"
#>  @ members:List of 2
#>  .. $ FQHC_owners     : <careMain>
#>  ..  ..@ title     : chr "Federally Qualified Health Center All Owners"
#>  ..  ..@ metadata  : <careMeta>
#>  .. .. .. @ description: chr "The Federally Qualified Health Center (FQHC) All Owners dataset provides ownership information on all FQHCs cur"| __truncated__
#>  .. .. .. @ temporal   : chr "2023-10-01 • 2025-03-31"
#>  .. .. .. @ periodicity: chr "Quarterly [R/P3M]"
#>  .. .. .. @ modified   : Date[1:1], format: "2025-01-15"
#>  .. .. .. @ dictionary : chr "https://data.cms.gov/sites/default/files/2024-10/FQHC_All_Owners_Data_Dictionary.pdf"
#>  .. .. .. @ site       : chr "https://data.cms.gov/provider-characteristics/hospitals-and-other-facilities/federally-qualified-health-center-all-owners"
#>  .. .. .. @ references : chr "https://data.cms.gov/sites/default/files/2023-09/FQHC_Data_Guidance.pdf"
#>  .. .. .. @ download   : chr "https://data.cms.gov/sites/default/files/2025-01/02441a9f-3380-4438-a1a0-be365d433ed9/FQHC_All_Owners_2025.01.02.csv"
#>  ..  ..@ dimensions: <careDim>
#>  .. .. .. @ rows  : int 142522
#>  .. .. .. @ pages : int 29
#>  .. .. .. @ fields: chr [1:34] "ENROLLMENT ID" "ASSOCIATE ID" "ORGANIZATION NAME" "ASSOCIATE ID - OWNER" ...
#>  ..  ..@ identifier: chr "https://data.cms.gov/data-api/v1/dataset/ed289c89-0bb8-4221-a20a-85776066381b/data-viewer"
#>  ..  ..@ resources : chr "https://data.cms.gov/data-api/v1/dataset-resources/ed289c89-0bb8-4221-a20a-85776066381b"
#>  .. $ FQHC_enrollments: <careMain>
#>  ..  ..@ title     : chr "Federally Qualified Health Center Enrollments"
#>  ..  ..@ metadata  : <careMeta>
#>  .. .. .. @ description: chr "The Federally Qualified Health Center (FQHC) Enrollments dataset provides enrollment information on all FQHCs c"| __truncated__
#>  .. .. .. @ temporal   : chr "2023-10-01 • 2025-03-31"
#>  .. .. .. @ periodicity: chr "Quarterly [R/P3M]"
#>  .. .. .. @ modified   : Date[1:1], format: "2025-01-15"
#>  .. .. .. @ dictionary : chr "https://data.cms.gov/sites/default/files/2023-09/FQHC_Enrollments_Data_Dictionary.pdf"
#>  .. .. .. @ site       : chr "https://data.cms.gov/provider-characteristics/hospitals-and-other-facilities/federally-qualified-health-center-enrollments"
#>  .. .. .. @ references : chr "https://data.cms.gov/sites/default/files/2023-09/FQHC_Data_Guidance.pdf"
#>  .. .. .. @ download   : chr "https://data.cms.gov/sites/default/files/2025-01/472db45d-c7ce-4b0b-8251-0470e625e521/FQHC_Enrollments_2025.01.02.csv"
#>  ..  ..@ dimensions: <careDim>
#>  .. .. .. @ rows  : int 10513
#>  .. .. .. @ pages : int 3
#>  .. .. .. @ fields: chr [1:20] "ENROLLMENT ID" "ENROLLMENT STATE" "PROVIDER TYPE CODE" "PROVIDER TYPE TEXT" ...
#>  ..  ..@ identifier: chr "https://data.cms.gov/data-api/v1/dataset/4bcae866-3411-439a-b762-90a6187c194b/data-viewer"
#>  ..  ..@ resources : chr "https://data.cms.gov/data-api/v1/dataset-resources/4bcae866-3411-439a-b762-90a6187c194b"
careGroup("pending")
#> <careGroup>
#>  @ group  : chr "Pending Initial Logging and Tracking"
#>  @ members:List of 2
#>  .. $ PILAT_non_physicians: <careMain>
#>  ..  ..@ title     : chr "Pending Initial Logging and Tracking Non Physicians"
#>  ..  ..@ metadata  : <careMeta>
#>  .. .. .. @ description: chr "The Pending Initial Logging and Tracking (L & T) Non Physicians dataset provides a list of pending applications"| __truncated__
#>  .. .. .. @ temporal   : chr "2023-05-21 • 2025-01-18"
#>  .. .. .. @ periodicity: chr "Twice a Week [R/P3.5D]"
#>  .. .. .. @ modified   : Date[1:1], format: "2025-04-10"
#>  .. .. .. @ dictionary : chr "https://data.cms.gov/resources/pending-initial-logging-and-tracking-physician-and-non-physicians-data-dictionary"
#>  .. .. .. @ site       : chr "https://data.cms.gov/provider-characteristics/medicare-provider-supplier-enrollment/pending-initial-logging-and"| __truncated__
#>  .. .. .. @ references : chr "https://data.cms.gov/resources/pending-initial-logging-and-tracking-physician-and-non-physicians-methodology"
#>  .. .. .. @ download   : chr "https://data.cms.gov/sites/default/files/2025-01/e394d10f-d2ae-4156-95b6-3d99c7a0f830/PendingInitialLandTsNonPh"| __truncated__
#>  ..  ..@ dimensions: <careDim>
#>  .. .. .. @ rows  : int 5847
#>  .. .. .. @ pages : int 2
#>  .. .. .. @ fields: chr [1:3] "NPI" "LAST_NAME" "FIRST_NAME"
#>  ..  ..@ identifier: chr "https://data.cms.gov/data-api/v1/dataset/261b83b6-b89f-43ad-ae7b-0d419a3bc24b/data-viewer"
#>  ..  ..@ resources : chr "https://data.cms.gov/data-api/v1/dataset-resources/261b83b6-b89f-43ad-ae7b-0d419a3bc24b"
#>  .. $ PILAT_physicians    : <careMain>
#>  ..  ..@ title     : chr "Pending Initial Logging and Tracking Physicians"
#>  ..  ..@ metadata  : <careMeta>
#>  .. .. .. @ description: chr "The Pending Initial Logging and Tracking (L & T) Physicians dataset provides a list of pending applications tha"| __truncated__
#>  .. .. .. @ temporal   : chr "2023-05-21 • 2025-01-18"
#>  .. .. .. @ periodicity: chr "Twice a Week [R/P3.5D]"
#>  .. .. .. @ modified   : Date[1:1], format: "2025-04-10"
#>  .. .. .. @ dictionary : chr "https://data.cms.gov/resources/pending-initial-logging-and-tracking-physician-and-non-physicians-data-dictionary"
#>  .. .. .. @ site       : chr "https://data.cms.gov/provider-characteristics/medicare-provider-supplier-enrollment/pending-initial-logging-and"| __truncated__
#>  .. .. .. @ references : chr "https://data.cms.gov/resources/pending-initial-logging-and-tracking-physician-and-non-physicians-methodology"
#>  .. .. .. @ download   : chr "https://data.cms.gov/sites/default/files/2025-01/fb10c137-60ec-4328-9af5-b6fc3b55b937/PendingInitialLandTsPhysi"| __truncated__
#>  ..  ..@ dimensions: <careDim>
#>  .. .. .. @ rows  : int 3350
#>  .. .. .. @ pages : int 1
#>  .. .. .. @ fields: chr [1:3] "NPI" "LAST_NAME" "FIRST_NAME"
#>  ..  ..@ identifier: chr "https://data.cms.gov/data-api/v1/dataset/6bd6b1dd-208c-4f9c-88b8-b15fec6db548/data-viewer"
#>  ..  ..@ resources : chr "https://data.cms.gov/data-api/v1/dataset-resources/6bd6b1dd-208c-4f9c-88b8-b15fec6db548"
careGroup("reassignment")
#> <careGroup>
#>  @ group  : chr "Revalidation Reassignment Lists"
#>  @ members:List of 3
#>  .. $ REVAL_group       : <careMain>
#>  ..  ..@ title     : chr "Revalidation Clinic Group Practice Reassignment"
#>  ..  ..@ metadata  : <careMeta>
#>  .. .. .. @ description: chr "The Revalidation Clinic Group Practice Reassignment dataset provides information between the physician and the "| __truncated__
#>  .. .. .. @ temporal   : chr "2022-01-01 • 2025-04-30"
#>  .. .. .. @ periodicity: chr "Monthly [R/P1M]"
#>  .. .. .. @ modified   : Date[1:1], format: "2025-04-01"
#>  .. .. .. @ dictionary : chr "https://data.cms.gov/resources/revalidation-clinic-group-practice-reassignment-data-dictionary"
#>  .. .. .. @ site       : chr "https://data.cms.gov/provider-characteristics/medicare-provider-supplier-enrollment/revalidation-clinic-group-p"| __truncated__
#>  .. .. .. @ references : chr "https://data.cms.gov/resources/additional-information-on-revalidation"
#>  .. .. .. @ download   : chr "https://data.cms.gov/sites/default/files/2025-03/f8e7bd82-2f79-4738-a927-58f3fd147ccd/revalidation_clinic_group"| __truncated__
#>  ..  ..@ dimensions: <careDim>
#>  .. .. .. @ rows  : int 3490044
#>  .. .. .. @ pages : int 699
#>  .. .. .. @ fields: chr [1:15] "Group PAC ID" "Group Enrollment ID" "Group Legal Business Name" "Group State Code" ...
#>  ..  ..@ identifier: chr "https://data.cms.gov/data-api/v1/dataset/e1f1fa9a-d6b4-417e-948a-c72dead8a41c/data-viewer"
#>  ..  ..@ resources : chr "https://data.cms.gov/data-api/v1/dataset-resources/e1f1fa9a-d6b4-417e-948a-c72dead8a41c"
#>  .. $ REVAL_due_date    : <careMain>
#>  ..  ..@ title     : chr "Revalidation Due Date List"
#>  ..  ..@ metadata  : <careMeta>
#>  .. .. .. @ description: chr "The Revalidation Due Date List dataset contains revalidation due dates for Medicare providers who are due to re"| __truncated__
#>  .. .. .. @ temporal   : chr "2022-02-01 • 2025-04-30"
#>  .. .. .. @ periodicity: chr "Monthly [R/P1M]"
#>  .. .. .. @ modified   : Date[1:1], format: "2025-04-01"
#>  .. .. .. @ dictionary : chr "https://data.cms.gov/resources/revalidation-due-date-list-data-dictionary"
#>  .. .. .. @ site       : chr "https://data.cms.gov/provider-characteristics/medicare-provider-supplier-enrollment/revalidation-due-date-list"
#>  .. .. .. @ references : chr "https://data.cms.gov/resources/additional-information-on-revalidation"
#>  .. .. .. @ download   : chr "https://data.cms.gov/sites/default/files/2025-03/e9c149be-aa7b-42cf-92a3-6b7675753153/revalidation_base.csv"
#>  ..  ..@ dimensions: <careDim>
#>  .. .. .. @ rows  : int 2820803
#>  .. .. .. @ pages : int 565
#>  .. .. .. @ fields: chr [1:13] "Enrollment ID" "National Provider Identifier" "First Name" "Last Name" ...
#>  ..  ..@ identifier: chr "https://data.cms.gov/data-api/v1/dataset/3746498e-874d-45d8-9c69-68603cafea60/data-viewer"
#>  ..  ..@ resources : chr "https://data.cms.gov/data-api/v1/dataset-resources/3746498e-874d-45d8-9c69-68603cafea60"
#>  .. $ REVAL_reassignment: <careMain>
#>  ..  ..@ title     : chr "Revalidation Reassignment List"
#>  ..  ..@ metadata  : <careMeta>
#>  .. .. .. @ description: chr "The Revalidation Reassignment List dataset provides information on reassignments of providers who are due for revalidation."
#>  .. .. .. @ temporal   : chr "2022-02-01 • 2025-04-30"
#>  .. .. .. @ periodicity: chr "Monthly [R/P1M]"
#>  .. .. .. @ modified   : Date[1:1], format: "2025-04-01"
#>  .. .. .. @ dictionary : chr "https://data.cms.gov/resources/revalidation-reassignments-list-data-dictionary"
#>  .. .. .. @ site       : chr "https://data.cms.gov/provider-characteristics/medicare-provider-supplier-enrollment/revalidation-reassignment-list"
#>  .. .. .. @ references : chr "https://data.cms.gov/resources/additional-information-on-revalidation"
#>  .. .. .. @ download   : chr "https://data.cms.gov/sites/default/files/2025-03/d9204531-f293-4b1e-be25-7ae88b6f0a77/revalidation_group_reassi"| __truncated__
#>  ..  ..@ dimensions: <careDim>
#>  .. .. .. @ rows  : int 3658408
#>  .. .. .. @ pages : int 732
#>  .. .. .. @ fields: chr [1:16] "Group PAC ID" "Group Enrollment ID" "Group Legal Business Name" "Group State Code" ...
#>  ..  ..@ identifier: chr "https://data.cms.gov/data-api/v1/dataset/20f51cff-4137-4f3a-b6b7-bfc9ad57983b/data-viewer"
#>  ..  ..@ resources : chr "https://data.cms.gov/data-api/v1/dataset-resources/20f51cff-4137-4f3a-b6b7-bfc9ad57983b"
careGroup("SNF")
#> <careGroup>
#>  @ group  : chr "Skilled Nursing Facilities"
#>  @ members:List of 5
#>  .. $ SNF_owners     : <careMain>
#>  ..  ..@ title     : chr "Skilled Nursing Facility All Owners"
#>  ..  ..@ metadata  : <careMeta>
#>  .. .. .. @ description: chr "The Skilled Nursing Facility (SNF) All Owners dataset provides information on all owners of SNFs currently enro"| __truncated__
#>  .. .. .. @ temporal   : chr "2022-09-01 • 2025-04-30"
#>  .. .. .. @ periodicity: chr "Monthly [R/P1M]"
#>  .. .. .. @ modified   : Date[1:1], format: "2025-04-15"
#>  .. .. .. @ dictionary : chr "https://data.cms.gov/sites/default/files/2024-12/7a976b38-70d9-4332-a913-446c89641840/SNF_All_Owners_Data_Dictionary.pdf"
#>  .. .. .. @ site       : chr "https://data.cms.gov/provider-characteristics/hospitals-and-other-facilities/skilled-nursing-facility-all-owners"
#>  .. .. .. @ references : chr "https://data.cms.gov/sites/default/files/2024-12/1ea55414-f441-4694-a13c-6400d4e28ef6/SNF_Data_Guidance.pdf"
#>  .. .. .. @ download   : chr "https://data.cms.gov/sites/default/files/2025-04/ba353275-5451-4d57-97c7-7943ee7ac09d/SNF_All_Owners_2025.04.01.csv"
#>  ..  ..@ dimensions: <careDim>
#>  .. .. .. @ rows  : int 157183
#>  .. .. .. @ pages : int 32
#>  .. .. .. @ fields: chr [1:40] "ENROLLMENT ID" "ASSOCIATE ID" "ORGANIZATION NAME" "ASSOCIATE ID - OWNER" ...
#>  ..  ..@ identifier: chr "https://data.cms.gov/data-api/v1/dataset/afe44b85-cc6d-40d7-b5df-00ae8910d1d2/data-viewer"
#>  ..  ..@ resources : chr "https://data.cms.gov/data-api/v1/dataset-resources/afe44b85-cc6d-40d7-b5df-00ae8910d1d2"
#>  .. $ SNF_chow       : <careMain>
#>  ..  ..@ title     : chr "Skilled Nursing Facility Change of Ownership"
#>  ..  ..@ metadata  : <careMeta>
#>  .. .. .. @ description: chr "The Skilled Nursing Facility (SNF) Change of Ownership (CHOW) dataset provides information on the SNF ownership"| __truncated__
#>  .. .. .. @ temporal   : chr "2022-01-01 • 2024-12-31"
#>  .. .. .. @ periodicity: chr "Quarterly [R/P3M]"
#>  .. .. .. @ modified   : Date[1:1], format: "2025-01-15"
#>  .. .. .. @ dictionary : chr "https://data.cms.gov/sites/default/files/2023-01/SNF_CHOW_Data_Dictionary_2022.12.30.pdf"
#>  .. .. .. @ site       : chr "https://data.cms.gov/provider-characteristics/hospitals-and-other-facilities/skilled-nursing-facility-change-of-ownership"
#>  .. .. .. @ references : chr "https://data.cms.gov/sites/default/files/2024-10/SNF_CHOW_Data_Guidance.pdf"
#>  .. .. .. @ download   : chr "https://data.cms.gov/sites/default/files/2025-01/7c721fd1-d17f-431c-a9d2-aab269919d21/SNF_CHOW_2025.01.02.csv"
#>  ..  ..@ dimensions: <careDim>
#>  .. .. .. @ rows  : int 4698
#>  .. .. .. @ pages : int 1
#>  .. .. .. @ fields: chr [1:23] "ENROLLMENT ID - BUYER" "ENROLLMENT STATE - BUYER" "PROVIDER TYPE CODE - BUYER" "PROVIDER TYPE TEXT - BUYER" ...
#>  ..  ..@ identifier: chr "https://data.cms.gov/data-api/v1/dataset/f557a6ed-95b3-4a22-8433-4175db2dec1c/data-viewer"
#>  ..  ..@ resources : chr "https://data.cms.gov/data-api/v1/dataset-resources/f557a6ed-95b3-4a22-8433-4175db2dec1c"
#>  .. $ SNF_chow_owner : <careMain>
#>  ..  ..@ title     : chr "Skilled Nursing Facility Change of Ownership - Owner Information"
#>  ..  ..@ metadata  : <careMeta>
#>  .. .. .. @ description: chr "The Skilled Nursing Facility (SNF) Change of Ownership (CHOW) - Owner Information dataset provides information "| __truncated__
#>  .. .. .. @ temporal   : chr "2022-01-01 • 2024-12-31"
#>  .. .. .. @ periodicity: chr "Quarterly [R/P3M]"
#>  .. .. .. @ modified   : Date[1:1], format: "2025-01-15"
#>  .. .. .. @ dictionary : chr "https://data.cms.gov/sites/default/files/2024-10/SNF_CHOW_Owners_Data_Dictionary.pdf"
#>  .. .. .. @ site       : chr "https://data.cms.gov/provider-characteristics/hospitals-and-other-facilities/skilled-nursing-facility-change-of"| __truncated__
#>  .. .. .. @ references : chr "https://data.cms.gov/sites/default/files/2024-10/SNF_CHOW_Data_Guidance.pdf"
#>  .. .. .. @ download   : chr "https://data.cms.gov/sites/default/files/2025-01/ba5a5dd5-f2b1-4047-a80a-9da9ec038312/SNF_CHOW_Owners_2025.01.02.csv"
#>  ..  ..@ dimensions: <careDim>
#>  .. .. .. @ rows  : int 101711
#>  .. .. .. @ pages : int 21
#>  .. .. .. @ fields: chr [1:34] "ENROLLMENT ID" "ASSOCIATE ID" "ORGANIZATION NAME" "ASSOCIATE ID - OWNER" ...
#>  ..  ..@ identifier: chr "https://data.cms.gov/data-api/v1/dataset/a4358712-e910-4eaf-8f24-5e90ba3cf8d0/data-viewer"
#>  ..  ..@ resources : chr "https://data.cms.gov/data-api/v1/dataset-resources/a4358712-e910-4eaf-8f24-5e90ba3cf8d0"
#>  .. $ SNF_cost_report: <careMain>
#>  ..  ..@ title     : chr "Skilled Nursing Facility Cost Report"
#>  ..  ..@ metadata  : <careMeta>
#>  .. .. .. @ description: chr "The Skilled Nursing Facility (SNF) Cost Report dataset is a public use file that provides select measures from "| __truncated__
#>  .. .. .. @ temporal   : chr "2011-01-01 • 2022-12-31"
#>  .. .. .. @ periodicity: chr "Annually [R/P1Y]"
#>  .. .. .. @ modified   : Date[1:1], format: "2024-11-25"
#>  .. .. .. @ dictionary : chr "https://data.cms.gov/sites/default/files/2023-12/Skilled%20Nursing%20Facility%20Cost%20Report%20Data%20Dictionary_508.pdf"
#>  .. .. .. @ site       : chr "https://data.cms.gov/provider-compliance/cost-report/skilled-nursing-facility-cost-report"
#>  .. .. .. @ references : chr "https://data.cms.gov/resources/skilled-nursing-facility-cost-report-methodology"
#>  .. .. .. @ download   : chr "https://data.cms.gov/sites/default/files/2024-10/94119bf8-a863-421d-acfa-ec4cc0b9e96b/CostReportsnf_Final_22.csv"
#>  ..  ..@ dimensions: <careDim>
#>  .. .. .. @ rows  : int 14960
#>  .. .. .. @ pages : int 3
#>  .. .. .. @ fields: chr [1:122] "rpt_rec_num" "Provider CCN" "Facility Name" "Street Address" ...
#>  ..  ..@ identifier: chr "https://data.cms.gov/data-api/v1/dataset/a69d3df7-3f66-4a0d-b5b8-0d66049bd565/data-viewer"
#>  ..  ..@ resources : chr "https://data.cms.gov/data-api/v1/dataset-resources/a69d3df7-3f66-4a0d-b5b8-0d66049bd565"
#>  .. $ SNF_enrollments: <careMain>
#>  ..  ..@ title     : chr "Skilled Nursing Facility Enrollments"
#>  ..  ..@ metadata  : <careMeta>
#>  .. .. .. @ description: chr "The Skilled Nursing Facility (SNF) Enrollments dataset provides enrollment information of all SNF s currently e"| __truncated__
#>  .. .. .. @ temporal   : chr "2022-09-01 • 2025-04-30"
#>  .. .. .. @ periodicity: chr "Monthly [R/P1M]"
#>  .. .. .. @ modified   : Date[1:1], format: "2025-04-15"
#>  .. .. .. @ dictionary : chr "https://data.cms.gov/sites/default/files/2024-10/SNF_Enrollments_Data_Dictionary.pdf"
#>  .. .. .. @ site       : chr "https://data.cms.gov/provider-characteristics/hospitals-and-other-facilities/skilled-nursing-facility-enrollments"
#>  .. .. .. @ references : chr "https://data.cms.gov/sites/default/files/2024-12/1ea55414-f441-4694-a13c-6400d4e28ef6/SNF_Data_Guidance.pdf"
#>  .. .. .. @ download   : chr "https://data.cms.gov/sites/default/files/2025-04/23e84217-1576-4fef-95e0-d1dd524c5971/SNF_Enrollments_2025.04.01.csv"
#>  ..  ..@ dimensions: <careDim>
#>  .. .. .. @ rows  : int 14462
#>  .. .. .. @ pages : int 3
#>  .. .. .. @ fields: chr [1:23] "ENROLLMENT ID" "ENROLLMENT STATE" "PROVIDER TYPE CODE" "PROVIDER TYPE TEXT" ...
#>  ..  ..@ identifier: chr "https://data.cms.gov/data-api/v1/dataset/5f2c306f-3b1c-42cd-b037-187b2ce22126/data-viewer"
#>  ..  ..@ resources : chr "https://data.cms.gov/data-api/v1/dataset-resources/5f2c306f-3b1c-42cd-b037-187b2ce22126"