Medicare Endpoint

Medicare Endpoint

Medicare Endpoint 2

Usage

careMain(alias)

careMain2(alias)

Arguments

alias

<chr> endpoint alias

Value

An S7 <careMain> object.

An S7 <careMain> object.

Examples

if (FALSE) { # interactive()
careMain("contact")
careMain("crosswalk")
careMain("dialysis")
careMain("enrollees")
careMain("facilities")
careMain("hospice_acute")
careMain("IQIES")
careMain("laboratories")
careMain("long_term")
careMain("opt_out")
careMain("order_refer")
careMain("rbcs")
careMain("transparency")
}
careMain2("contact")
#> <careMain2>
#>  @ title     : chr "Medicare Fee-For-Service Public Provider Enrollment"
#>  @ metadata  : <care_metadata>
#>  .. @ description: chr "The Medicare Fee-For-Service Public Provider Enrollment dataset includes information on providers who are activ"| __truncated__
#>  .. @ temporal   : chr "2023-01-01 • 2025-03-31"
#>  .. @ periodicity: chr "Quarterly [R/P3M]"
#>  .. @ dictionary : chr "https://data.cms.gov/resources/medicare-fee-for-service-public-provider-enrollment-data-dictionary"
#>  .. @ site       : chr "https://data.cms.gov/provider-characteristics/medicare-provider-supplier-enrollment/medicare-fee-for-service-pu"| __truncated__
#>  .. @ references : chr "https://data.cms.gov/resources/fee-for-service-public-provider-enrollment-methodology"
#>  .. @ modified   : Date[1:1], format: "2025-04-15"
#>  .. @ download   : chr "https://data.cms.gov/sites/default/files/2025-04/5d10bc83-29bd-498e-a1ff-3e6e2065a7e3/PPEF_Enrollment_Extract_2025.04.01.csv"
#>  @ dimensions: <care_dimensions>
#>  .. @ rows  : int 2922752
#>  .. @ pages : int 585
#>  .. @ fields: chr [1:10] "NPI" "PECOS_ASCT_CNTL_ID" "ENRLMT_ID" ...
#>  @ identifier: chr "https://data.cms.gov/data-api/v1/dataset/2457ea29-fc82-48b0-86ec-3b0755de7515/data-viewer"
#>  @ resources : chr "https://data.cms.gov/data-api/v1/dataset-resources/2457ea29-fc82-48b0-86ec-3b0755de7515"
careMain2("crosswalk")
#> <careMain2>
#>  @ title     : chr "Medicare Fee-For-Service Public Provider Enrollment"
#>  @ metadata  : <care_metadata>
#>  .. @ description: chr "The Medicare Fee-For-Service Public Provider Enrollment dataset includes information on providers who are activ"| __truncated__
#>  .. @ temporal   : chr "2023-01-01 • 2025-03-31"
#>  .. @ periodicity: chr "Quarterly [R/P3M]"
#>  .. @ dictionary : chr "https://data.cms.gov/resources/medicare-fee-for-service-public-provider-enrollment-data-dictionary"
#>  .. @ site       : chr "https://data.cms.gov/provider-characteristics/medicare-provider-supplier-enrollment/medicare-fee-for-service-pu"| __truncated__
#>  .. @ references : chr "https://data.cms.gov/resources/fee-for-service-public-provider-enrollment-methodology"
#>  .. @ modified   : Date[1:1], format: "2025-04-15"
#>  .. @ download   : chr "https://data.cms.gov/sites/default/files/2025-04/5d10bc83-29bd-498e-a1ff-3e6e2065a7e3/PPEF_Enrollment_Extract_2025.04.01.csv"
#>  @ dimensions: <care_dimensions>
#>  .. @ rows  : int 2922752
#>  .. @ pages : int 585
#>  .. @ fields: chr [1:10] "NPI" "PECOS_ASCT_CNTL_ID" "ENRLMT_ID" ...
#>  @ identifier: chr "https://data.cms.gov/data-api/v1/dataset/2457ea29-fc82-48b0-86ec-3b0755de7515/data-viewer"
#>  @ resources : chr "https://data.cms.gov/data-api/v1/dataset-resources/2457ea29-fc82-48b0-86ec-3b0755de7515"
careMain2("dialysis")
#> <careMain2>
#>  @ title     : chr "Medicare Fee-For-Service Public Provider Enrollment"
#>  @ metadata  : <care_metadata>
#>  .. @ description: chr "The Medicare Fee-For-Service Public Provider Enrollment dataset includes information on providers who are activ"| __truncated__
#>  .. @ temporal   : chr "2023-01-01 • 2025-03-31"
#>  .. @ periodicity: chr "Quarterly [R/P3M]"
#>  .. @ dictionary : chr "https://data.cms.gov/resources/medicare-fee-for-service-public-provider-enrollment-data-dictionary"
#>  .. @ site       : chr "https://data.cms.gov/provider-characteristics/medicare-provider-supplier-enrollment/medicare-fee-for-service-pu"| __truncated__
#>  .. @ references : chr "https://data.cms.gov/resources/fee-for-service-public-provider-enrollment-methodology"
#>  .. @ modified   : Date[1:1], format: "2025-04-15"
#>  .. @ download   : chr "https://data.cms.gov/sites/default/files/2025-04/5d10bc83-29bd-498e-a1ff-3e6e2065a7e3/PPEF_Enrollment_Extract_2025.04.01.csv"
#>  @ dimensions: <care_dimensions>
#>  .. @ rows  : int 2922752
#>  .. @ pages : int 585
#>  .. @ fields: chr [1:10] "NPI" "PECOS_ASCT_CNTL_ID" "ENRLMT_ID" ...
#>  @ identifier: chr "https://data.cms.gov/data-api/v1/dataset/2457ea29-fc82-48b0-86ec-3b0755de7515/data-viewer"
#>  @ resources : chr "https://data.cms.gov/data-api/v1/dataset-resources/2457ea29-fc82-48b0-86ec-3b0755de7515"
careMain2("enrollees")
#> <careMain2>
#>  @ title     : chr "Medicare Fee-For-Service Public Provider Enrollment"
#>  @ metadata  : <care_metadata>
#>  .. @ description: chr "The Medicare Fee-For-Service Public Provider Enrollment dataset includes information on providers who are activ"| __truncated__
#>  .. @ temporal   : chr "2023-01-01 • 2025-03-31"
#>  .. @ periodicity: chr "Quarterly [R/P3M]"
#>  .. @ dictionary : chr "https://data.cms.gov/resources/medicare-fee-for-service-public-provider-enrollment-data-dictionary"
#>  .. @ site       : chr "https://data.cms.gov/provider-characteristics/medicare-provider-supplier-enrollment/medicare-fee-for-service-pu"| __truncated__
#>  .. @ references : chr "https://data.cms.gov/resources/fee-for-service-public-provider-enrollment-methodology"
#>  .. @ modified   : Date[1:1], format: "2025-04-15"
#>  .. @ download   : chr "https://data.cms.gov/sites/default/files/2025-04/5d10bc83-29bd-498e-a1ff-3e6e2065a7e3/PPEF_Enrollment_Extract_2025.04.01.csv"
#>  @ dimensions: <care_dimensions>
#>  .. @ rows  : int 2922752
#>  .. @ pages : int 585
#>  .. @ fields: chr [1:10] "NPI" "PECOS_ASCT_CNTL_ID" "ENRLMT_ID" ...
#>  @ identifier: chr "https://data.cms.gov/data-api/v1/dataset/2457ea29-fc82-48b0-86ec-3b0755de7515/data-viewer"
#>  @ resources : chr "https://data.cms.gov/data-api/v1/dataset-resources/2457ea29-fc82-48b0-86ec-3b0755de7515"
careMain2("facilities")
#> <careMain2>
#>  @ title     : chr "Medicare Fee-For-Service Public Provider Enrollment"
#>  @ metadata  : <care_metadata>
#>  .. @ description: chr "The Medicare Fee-For-Service Public Provider Enrollment dataset includes information on providers who are activ"| __truncated__
#>  .. @ temporal   : chr "2023-01-01 • 2025-03-31"
#>  .. @ periodicity: chr "Quarterly [R/P3M]"
#>  .. @ dictionary : chr "https://data.cms.gov/resources/medicare-fee-for-service-public-provider-enrollment-data-dictionary"
#>  .. @ site       : chr "https://data.cms.gov/provider-characteristics/medicare-provider-supplier-enrollment/medicare-fee-for-service-pu"| __truncated__
#>  .. @ references : chr "https://data.cms.gov/resources/fee-for-service-public-provider-enrollment-methodology"
#>  .. @ modified   : Date[1:1], format: "2025-04-15"
#>  .. @ download   : chr "https://data.cms.gov/sites/default/files/2025-04/5d10bc83-29bd-498e-a1ff-3e6e2065a7e3/PPEF_Enrollment_Extract_2025.04.01.csv"
#>  @ dimensions: <care_dimensions>
#>  .. @ rows  : int 2922752
#>  .. @ pages : int 585
#>  .. @ fields: chr [1:10] "NPI" "PECOS_ASCT_CNTL_ID" "ENRLMT_ID" ...
#>  @ identifier: chr "https://data.cms.gov/data-api/v1/dataset/2457ea29-fc82-48b0-86ec-3b0755de7515/data-viewer"
#>  @ resources : chr "https://data.cms.gov/data-api/v1/dataset-resources/2457ea29-fc82-48b0-86ec-3b0755de7515"
careMain2("hospice_acute")
#> <careMain2>
#>  @ title     : chr "Medicare Fee-For-Service Public Provider Enrollment"
#>  @ metadata  : <care_metadata>
#>  .. @ description: chr "The Medicare Fee-For-Service Public Provider Enrollment dataset includes information on providers who are activ"| __truncated__
#>  .. @ temporal   : chr "2023-01-01 • 2025-03-31"
#>  .. @ periodicity: chr "Quarterly [R/P3M]"
#>  .. @ dictionary : chr "https://data.cms.gov/resources/medicare-fee-for-service-public-provider-enrollment-data-dictionary"
#>  .. @ site       : chr "https://data.cms.gov/provider-characteristics/medicare-provider-supplier-enrollment/medicare-fee-for-service-pu"| __truncated__
#>  .. @ references : chr "https://data.cms.gov/resources/fee-for-service-public-provider-enrollment-methodology"
#>  .. @ modified   : Date[1:1], format: "2025-04-15"
#>  .. @ download   : chr "https://data.cms.gov/sites/default/files/2025-04/5d10bc83-29bd-498e-a1ff-3e6e2065a7e3/PPEF_Enrollment_Extract_2025.04.01.csv"
#>  @ dimensions: <care_dimensions>
#>  .. @ rows  : int 2922752
#>  .. @ pages : int 585
#>  .. @ fields: chr [1:10] "NPI" "PECOS_ASCT_CNTL_ID" "ENRLMT_ID" ...
#>  @ identifier: chr "https://data.cms.gov/data-api/v1/dataset/2457ea29-fc82-48b0-86ec-3b0755de7515/data-viewer"
#>  @ resources : chr "https://data.cms.gov/data-api/v1/dataset-resources/2457ea29-fc82-48b0-86ec-3b0755de7515"
careMain2("IQIES")
#> <careMain2>
#>  @ title     : chr "Medicare Fee-For-Service Public Provider Enrollment"
#>  @ metadata  : <care_metadata>
#>  .. @ description: chr "The Medicare Fee-For-Service Public Provider Enrollment dataset includes information on providers who are activ"| __truncated__
#>  .. @ temporal   : chr "2023-01-01 • 2025-03-31"
#>  .. @ periodicity: chr "Quarterly [R/P3M]"
#>  .. @ dictionary : chr "https://data.cms.gov/resources/medicare-fee-for-service-public-provider-enrollment-data-dictionary"
#>  .. @ site       : chr "https://data.cms.gov/provider-characteristics/medicare-provider-supplier-enrollment/medicare-fee-for-service-pu"| __truncated__
#>  .. @ references : chr "https://data.cms.gov/resources/fee-for-service-public-provider-enrollment-methodology"
#>  .. @ modified   : Date[1:1], format: "2025-04-15"
#>  .. @ download   : chr "https://data.cms.gov/sites/default/files/2025-04/5d10bc83-29bd-498e-a1ff-3e6e2065a7e3/PPEF_Enrollment_Extract_2025.04.01.csv"
#>  @ dimensions: <care_dimensions>
#>  .. @ rows  : int 2922752
#>  .. @ pages : int 585
#>  .. @ fields: chr [1:10] "NPI" "PECOS_ASCT_CNTL_ID" "ENRLMT_ID" ...
#>  @ identifier: chr "https://data.cms.gov/data-api/v1/dataset/2457ea29-fc82-48b0-86ec-3b0755de7515/data-viewer"
#>  @ resources : chr "https://data.cms.gov/data-api/v1/dataset-resources/2457ea29-fc82-48b0-86ec-3b0755de7515"
careMain2("laboratories")
#> <careMain2>
#>  @ title     : chr "Medicare Fee-For-Service Public Provider Enrollment"
#>  @ metadata  : <care_metadata>
#>  .. @ description: chr "The Medicare Fee-For-Service Public Provider Enrollment dataset includes information on providers who are activ"| __truncated__
#>  .. @ temporal   : chr "2023-01-01 • 2025-03-31"
#>  .. @ periodicity: chr "Quarterly [R/P3M]"
#>  .. @ dictionary : chr "https://data.cms.gov/resources/medicare-fee-for-service-public-provider-enrollment-data-dictionary"
#>  .. @ site       : chr "https://data.cms.gov/provider-characteristics/medicare-provider-supplier-enrollment/medicare-fee-for-service-pu"| __truncated__
#>  .. @ references : chr "https://data.cms.gov/resources/fee-for-service-public-provider-enrollment-methodology"
#>  .. @ modified   : Date[1:1], format: "2025-04-15"
#>  .. @ download   : chr "https://data.cms.gov/sites/default/files/2025-04/5d10bc83-29bd-498e-a1ff-3e6e2065a7e3/PPEF_Enrollment_Extract_2025.04.01.csv"
#>  @ dimensions: <care_dimensions>
#>  .. @ rows  : int 2922752
#>  .. @ pages : int 585
#>  .. @ fields: chr [1:10] "NPI" "PECOS_ASCT_CNTL_ID" "ENRLMT_ID" ...
#>  @ identifier: chr "https://data.cms.gov/data-api/v1/dataset/2457ea29-fc82-48b0-86ec-3b0755de7515/data-viewer"
#>  @ resources : chr "https://data.cms.gov/data-api/v1/dataset-resources/2457ea29-fc82-48b0-86ec-3b0755de7515"
careMain2("long_term")
#> <careMain2>
#>  @ title     : chr "Medicare Fee-For-Service Public Provider Enrollment"
#>  @ metadata  : <care_metadata>
#>  .. @ description: chr "The Medicare Fee-For-Service Public Provider Enrollment dataset includes information on providers who are activ"| __truncated__
#>  .. @ temporal   : chr "2023-01-01 • 2025-03-31"
#>  .. @ periodicity: chr "Quarterly [R/P3M]"
#>  .. @ dictionary : chr "https://data.cms.gov/resources/medicare-fee-for-service-public-provider-enrollment-data-dictionary"
#>  .. @ site       : chr "https://data.cms.gov/provider-characteristics/medicare-provider-supplier-enrollment/medicare-fee-for-service-pu"| __truncated__
#>  .. @ references : chr "https://data.cms.gov/resources/fee-for-service-public-provider-enrollment-methodology"
#>  .. @ modified   : Date[1:1], format: "2025-04-15"
#>  .. @ download   : chr "https://data.cms.gov/sites/default/files/2025-04/5d10bc83-29bd-498e-a1ff-3e6e2065a7e3/PPEF_Enrollment_Extract_2025.04.01.csv"
#>  @ dimensions: <care_dimensions>
#>  .. @ rows  : int 2922752
#>  .. @ pages : int 585
#>  .. @ fields: chr [1:10] "NPI" "PECOS_ASCT_CNTL_ID" "ENRLMT_ID" ...
#>  @ identifier: chr "https://data.cms.gov/data-api/v1/dataset/2457ea29-fc82-48b0-86ec-3b0755de7515/data-viewer"
#>  @ resources : chr "https://data.cms.gov/data-api/v1/dataset-resources/2457ea29-fc82-48b0-86ec-3b0755de7515"
careMain2("opt_out")
#> <careMain2>
#>  @ title     : chr "Medicare Fee-For-Service Public Provider Enrollment"
#>  @ metadata  : <care_metadata>
#>  .. @ description: chr "The Medicare Fee-For-Service Public Provider Enrollment dataset includes information on providers who are activ"| __truncated__
#>  .. @ temporal   : chr "2023-01-01 • 2025-03-31"
#>  .. @ periodicity: chr "Quarterly [R/P3M]"
#>  .. @ dictionary : chr "https://data.cms.gov/resources/medicare-fee-for-service-public-provider-enrollment-data-dictionary"
#>  .. @ site       : chr "https://data.cms.gov/provider-characteristics/medicare-provider-supplier-enrollment/medicare-fee-for-service-pu"| __truncated__
#>  .. @ references : chr "https://data.cms.gov/resources/fee-for-service-public-provider-enrollment-methodology"
#>  .. @ modified   : Date[1:1], format: "2025-04-15"
#>  .. @ download   : chr "https://data.cms.gov/sites/default/files/2025-04/5d10bc83-29bd-498e-a1ff-3e6e2065a7e3/PPEF_Enrollment_Extract_2025.04.01.csv"
#>  @ dimensions: <care_dimensions>
#>  .. @ rows  : int 2922752
#>  .. @ pages : int 585
#>  .. @ fields: chr [1:10] "NPI" "PECOS_ASCT_CNTL_ID" "ENRLMT_ID" ...
#>  @ identifier: chr "https://data.cms.gov/data-api/v1/dataset/2457ea29-fc82-48b0-86ec-3b0755de7515/data-viewer"
#>  @ resources : chr "https://data.cms.gov/data-api/v1/dataset-resources/2457ea29-fc82-48b0-86ec-3b0755de7515"
careMain2("order_refer")
#> <careMain2>
#>  @ title     : chr "Medicare Fee-For-Service Public Provider Enrollment"
#>  @ metadata  : <care_metadata>
#>  .. @ description: chr "The Medicare Fee-For-Service Public Provider Enrollment dataset includes information on providers who are activ"| __truncated__
#>  .. @ temporal   : chr "2023-01-01 • 2025-03-31"
#>  .. @ periodicity: chr "Quarterly [R/P3M]"
#>  .. @ dictionary : chr "https://data.cms.gov/resources/medicare-fee-for-service-public-provider-enrollment-data-dictionary"
#>  .. @ site       : chr "https://data.cms.gov/provider-characteristics/medicare-provider-supplier-enrollment/medicare-fee-for-service-pu"| __truncated__
#>  .. @ references : chr "https://data.cms.gov/resources/fee-for-service-public-provider-enrollment-methodology"
#>  .. @ modified   : Date[1:1], format: "2025-04-15"
#>  .. @ download   : chr "https://data.cms.gov/sites/default/files/2025-04/5d10bc83-29bd-498e-a1ff-3e6e2065a7e3/PPEF_Enrollment_Extract_2025.04.01.csv"
#>  @ dimensions: <care_dimensions>
#>  .. @ rows  : int 2922752
#>  .. @ pages : int 585
#>  .. @ fields: chr [1:10] "NPI" "PECOS_ASCT_CNTL_ID" "ENRLMT_ID" ...
#>  @ identifier: chr "https://data.cms.gov/data-api/v1/dataset/2457ea29-fc82-48b0-86ec-3b0755de7515/data-viewer"
#>  @ resources : chr "https://data.cms.gov/data-api/v1/dataset-resources/2457ea29-fc82-48b0-86ec-3b0755de7515"
careMain2("rbcs")
#> <careMain2>
#>  @ title     : chr "Medicare Fee-For-Service Public Provider Enrollment"
#>  @ metadata  : <care_metadata>
#>  .. @ description: chr "The Medicare Fee-For-Service Public Provider Enrollment dataset includes information on providers who are activ"| __truncated__
#>  .. @ temporal   : chr "2023-01-01 • 2025-03-31"
#>  .. @ periodicity: chr "Quarterly [R/P3M]"
#>  .. @ dictionary : chr "https://data.cms.gov/resources/medicare-fee-for-service-public-provider-enrollment-data-dictionary"
#>  .. @ site       : chr "https://data.cms.gov/provider-characteristics/medicare-provider-supplier-enrollment/medicare-fee-for-service-pu"| __truncated__
#>  .. @ references : chr "https://data.cms.gov/resources/fee-for-service-public-provider-enrollment-methodology"
#>  .. @ modified   : Date[1:1], format: "2025-04-15"
#>  .. @ download   : chr "https://data.cms.gov/sites/default/files/2025-04/5d10bc83-29bd-498e-a1ff-3e6e2065a7e3/PPEF_Enrollment_Extract_2025.04.01.csv"
#>  @ dimensions: <care_dimensions>
#>  .. @ rows  : int 2922752
#>  .. @ pages : int 585
#>  .. @ fields: chr [1:10] "NPI" "PECOS_ASCT_CNTL_ID" "ENRLMT_ID" ...
#>  @ identifier: chr "https://data.cms.gov/data-api/v1/dataset/2457ea29-fc82-48b0-86ec-3b0755de7515/data-viewer"
#>  @ resources : chr "https://data.cms.gov/data-api/v1/dataset-resources/2457ea29-fc82-48b0-86ec-3b0755de7515"
careMain2("transparency")
#> <careMain2>
#>  @ title     : chr "Medicare Fee-For-Service Public Provider Enrollment"
#>  @ metadata  : <care_metadata>
#>  .. @ description: chr "The Medicare Fee-For-Service Public Provider Enrollment dataset includes information on providers who are activ"| __truncated__
#>  .. @ temporal   : chr "2023-01-01 • 2025-03-31"
#>  .. @ periodicity: chr "Quarterly [R/P3M]"
#>  .. @ dictionary : chr "https://data.cms.gov/resources/medicare-fee-for-service-public-provider-enrollment-data-dictionary"
#>  .. @ site       : chr "https://data.cms.gov/provider-characteristics/medicare-provider-supplier-enrollment/medicare-fee-for-service-pu"| __truncated__
#>  .. @ references : chr "https://data.cms.gov/resources/fee-for-service-public-provider-enrollment-methodology"
#>  .. @ modified   : Date[1:1], format: "2025-04-15"
#>  .. @ download   : chr "https://data.cms.gov/sites/default/files/2025-04/5d10bc83-29bd-498e-a1ff-3e6e2065a7e3/PPEF_Enrollment_Extract_2025.04.01.csv"
#>  @ dimensions: <care_dimensions>
#>  .. @ rows  : int 2922752
#>  .. @ pages : int 585
#>  .. @ fields: chr [1:10] "NPI" "PECOS_ASCT_CNTL_ID" "ENRLMT_ID" ...
#>  @ identifier: chr "https://data.cms.gov/data-api/v1/dataset/2457ea29-fc82-48b0-86ec-3b0755de7515/data-viewer"
#>  @ resources : chr "https://data.cms.gov/data-api/v1/dataset-resources/2457ea29-fc82-48b0-86ec-3b0755de7515"