Group of Medicare Temporal Endpoints

Group of Medicare Temporal Endpoints

Usage

careTempGroup(alias)

Arguments

alias

<chr> title alias

Value

An S7 <careTempGroup> object.

Examples

careTempGroup("inpatient")
#> <careTempGroup>
#>  @ group  : chr "Medicare Inpatient Hospitals"
#>  @ members:List of 3
#>  .. $ IN_geography_and_service: <careTemp>
#>  ..  ..@ title     : chr "Medicare Inpatient Hospitals - by Geography and Service"
#>  ..  ..@ metadata  : <careMeta>
#>  .. .. .. @ description: chr "The Medicare Inpatient Hospitals by Geography and Service dataset provides information on hospital discharges f"| __truncated__
#>  .. .. .. @ periodicity: chr "Annually [R/P1Y]"
#>  .. .. .. @ dictionary : chr "https://data.cms.gov/resources/medicare-inpatient-hospitals-by-geography-and-service-data-dictionary-0"
#>  .. .. .. @ site       : chr "https://data.cms.gov/provider-summary-by-type-of-service/medicare-inpatient-hospitals/medicare-inpatient-hospit"| __truncated__
#>  ..  ..@ dimensions: <careTempDim>
#>  .. .. .. @ rows  : int 26260
#>  .. .. .. @ pages : int 6
#>  .. .. .. @ fields: chr [1:9] "Rndrng_Prvdr_Geo_Lvl" "Rndrng_Prvdr_Geo_Cd" "Rndrng_Prvdr_Geo_Desc" "DRG_Cd" ...
#>  ..  ..@ endpoints : tibble [10 × 5] (S3: tbl_df/tbl/data.frame)
#>  $ year      : int [1:10] 2022 2021 2020 2019 2018 2017 2016 2015 2014 2013
#>  $ modified  : Date[1:10], format: "2024-06-04" "2023-05-10" ...
#>  $ identifier: chr [1:10] "https://data.cms.gov/data-api/v1/dataset/5c8b315f-a2a3-4b2f-bba6-7be19f8d7edc/data" "https://data.cms.gov/data-api/v1/dataset/68739613-1a4f-40ae-9664-49da07abd81f/data" "https://data.cms.gov/data-api/v1/dataset/e583de38-33d7-4895-a115-0d8af74f0795/data" "https://data.cms.gov/data-api/v1/dataset/337617e5-c18b-4cb9-8152-34851747584b/data" ...
#>  $ download  : chr [1:10] "https://data.cms.gov/sites/default/files/2023-05/ee35f629-d386-426e-b51e-70451cf896d2/MUP_IHP_RY23_P03_V10_DY21_GEO.CSV" "https://data.cms.gov/sites/default/files/2023-05/ef9ecf4c-ad7c-4112-881e-ccb9a83acb1c/MUP_IHP_RY23_P03_V10_DY20_GEO.CSV" "https://data.cms.gov/sites/default/files/2023-05/f07177de-e932-4b1a-a754-eac7bdfc4b28/MUP_IHP_RY23_P03_V10_DY19_GEO.CSV" "https://data.cms.gov/sites/default/files/2023-05/9c685ed4-a63c-4d22-9027-f225ce631c56/MUP_IHP_RY23_P03_V10_DY18_GEO.CSV" ...
#>  $ resources : chr [1:10] "https://data.cms.gov/data-api/v1/dataset-resources/5c8b315f-a2a3-4b2f-bba6-7be19f8d7edc" "https://data.cms.gov/data-api/v1/dataset-resources/68739613-1a4f-40ae-9664-49da07abd81f" "https://data.cms.gov/data-api/v1/dataset-resources/e583de38-33d7-4895-a115-0d8af74f0795" "https://data.cms.gov/data-api/v1/dataset-resources/337617e5-c18b-4cb9-8152-34851747584b" ...
#>  .. $ IN_provider             : <careTemp>
#>  ..  ..@ title     : chr "Medicare Inpatient Hospitals - by Provider"
#>  ..  ..@ metadata  : <careMeta>
#>  .. .. .. @ description: chr "The Medicare Inpatient Hospitals by Provider dataset provides information on inpatient discharges for Original "| __truncated__
#>  .. .. .. @ periodicity: chr "Annually [R/P1Y]"
#>  .. .. .. @ dictionary : chr "https://data.cms.gov/resources/medicare-inpatient-hospitals-by-provider-data-dictionary"
#>  .. .. .. @ site       : chr "https://data.cms.gov/provider-summary-by-type-of-service/medicare-inpatient-hospitals/medicare-inpatient-hospitals-by-provider"
#>  ..  ..@ dimensions: <careTempDim>
#>  .. .. .. @ rows  : int 3120
#>  .. .. .. @ pages : int 1
#>  .. .. .. @ fields: chr [1:57] "Rndrng_Prvdr_CCN" "Rndrng_Prvdr_Org_Name" "Rndrng_Prvdr_St" "Rndrng_Prvdr_City" ...
#>  ..  ..@ endpoints : tibble [10 × 5] (S3: tbl_df/tbl/data.frame)
#>  $ year      : int [1:10] 2022 2021 2020 2019 2018 2017 2016 2015 2014 2013
#>  $ modified  : Date[1:10], format: "2024-06-04" "2024-06-04" ...
#>  $ identifier: chr [1:10] "https://data.cms.gov/data-api/v1/dataset/cc73e77b-7367-48c9-b766-fafefe3f8ffa/data" "https://data.cms.gov/data-api/v1/dataset/c7e1501c-a770-4e2a-beb6-32b3525491a7/data" "https://data.cms.gov/data-api/v1/dataset/d7b2810d-ac10-498f-8794-1c3d9f72ee96/data" "https://data.cms.gov/data-api/v1/dataset/c349419e-3c6a-419e-8c74-0641b64cca12/data" ...
#>  $ download  : chr [1:10] "https://data.cms.gov/sites/default/files/2024-05/ef80e842-4949-46e8-ab66-0c4c4310d400/MUP_INP_RY24_P04_V10_DY21_Prv.CSV" "https://data.cms.gov/sites/default/files/2024-05/c2f4c3df-c7ec-4402-8064-7f7d2586ba56/MUP_INP_RY24_P04_V10_DY20_Prv.CSV" "https://data.cms.gov/sites/default/files/2024-05/1c1f6004-9fe2-43d0-907c-203322cd40e7/MUP_INP_RY24_P04_V10_DY19_Prv.CSV" "https://data.cms.gov/sites/default/files/2024-05/7bb03a80-a0ed-4fd9-a7c8-c851c70d484b/MUP_INP_RY24_P04_V10_DY18_Prv.CSV" ...
#>  $ resources : chr [1:10] "https://data.cms.gov/data-api/v1/dataset-resources/cc73e77b-7367-48c9-b766-fafefe3f8ffa" "https://data.cms.gov/data-api/v1/dataset-resources/c7e1501c-a770-4e2a-beb6-32b3525491a7" "https://data.cms.gov/data-api/v1/dataset-resources/d7b2810d-ac10-498f-8794-1c3d9f72ee96" "https://data.cms.gov/data-api/v1/dataset-resources/c349419e-3c6a-419e-8c74-0641b64cca12" ...
#>  .. $ IN_provider_and_service : <careTemp>
#>  ..  ..@ title     : chr "Medicare Inpatient Hospitals - by Provider and Service"
#>  ..  ..@ metadata  : <careMeta>
#>  .. .. .. @ description: chr "The Medicare Inpatient Hospitals by Provider and Service dataset provides information on inpatient discharges f"| __truncated__
#>  .. .. .. @ periodicity: chr "Annually [R/P1Y]"
#>  .. .. .. @ dictionary : chr "https://data.cms.gov/resources/medicare-inpatient-hospitals-by-provider-and-service-data-dictionary-0"
#>  .. .. .. @ site       : chr "https://data.cms.gov/provider-summary-by-type-of-service/medicare-inpatient-hospitals/medicare-inpatient-hospit"| __truncated__
#>  ..  ..@ dimensions: <careTempDim>
#>  .. .. .. @ rows  : int 145742
#>  .. .. .. @ pages : int 30
#>  .. .. .. @ fields: chr [1:15] "Rndrng_Prvdr_CCN" "Rndrng_Prvdr_Org_Name" "Rndrng_Prvdr_City" "Rndrng_Prvdr_St" ...
#>  ..  ..@ endpoints : tibble [10 × 5] (S3: tbl_df/tbl/data.frame)
#>  $ year      : int [1:10] 2022 2021 2020 2019 2018 2017 2016 2015 2014 2013
#>  $ modified  : Date[1:10], format: "2024-06-04" "2023-09-05" ...
#>  $ identifier: chr [1:10] "https://data.cms.gov/data-api/v1/dataset/46bf50f8-0983-4ca2-b8d5-f2afbbf2e589/data" "https://data.cms.gov/data-api/v1/dataset/117d93f2-ce81-40fe-a4d4-8c03203b95e1/data" "https://data.cms.gov/data-api/v1/dataset/ef5bdbe1-27b4-4296-b320-52bd5d2183d7/data" "https://data.cms.gov/data-api/v1/dataset/6f6d93e1-ecf8-4b93-9845-091faf20f274/data" ...
#>  $ download  : chr [1:10] "https://data.cms.gov/sites/default/files/2023-05/a754bf0b-0c51-4daf-876e-272f90a11c05/MUP_IHP_RY23_P03_V10_DY21_PRVSVC.CSV" "https://data.cms.gov/sites/default/files/2023-05/e57818f2-318c-4979-a612-c91eba44b011/MUP_IHP_RY23_P03_V10_DY20_PRVSVC.CSV" "https://data.cms.gov/sites/default/files/2023-05/6602e715-b301-4a38-954d-b8d5aec12b87/MUP_IHP_RY23_P03_V10_DY19_PRVSVC.CSV" "https://data.cms.gov/sites/default/files/2023-05/78d80cf4-fc5b-40db-8e88-ca44bc86102f/MUP_IHP_RY23_P03_V10_DY18_PRVSVC.CSV" ...
#>  $ resources : chr [1:10] "https://data.cms.gov/data-api/v1/dataset-resources/46bf50f8-0983-4ca2-b8d5-f2afbbf2e589" "https://data.cms.gov/data-api/v1/dataset-resources/117d93f2-ce81-40fe-a4d4-8c03203b95e1" "https://data.cms.gov/data-api/v1/dataset-resources/ef5bdbe1-27b4-4296-b320-52bd5d2183d7" "https://data.cms.gov/data-api/v1/dataset-resources/6f6d93e1-ecf8-4b93-9845-091faf20f274" ...
careTempGroup("outpatient")
#> <careTempGroup>
#>  @ group  : chr "Medicare Outpatient Hospitals"
#>  @ members:List of 2
#>  .. $ OUT_geography_and_service: <careTemp>
#>  ..  ..@ title     : chr "Medicare Outpatient Hospitals - by Geography and Service"
#>  ..  ..@ metadata  : <careMeta>
#>  .. .. .. @ description: chr "The Medicare Outpatient Hospitals by Geography and Service dataset provides information on services for Origina"| __truncated__
#>  .. .. .. @ periodicity: chr "Annually [R/P1Y]"
#>  .. .. .. @ dictionary : chr "https://data.cms.gov/resources/medicare-outpatient-hospitals-by-geography-and-service-data-dictionary"
#>  .. .. .. @ site       : chr "https://data.cms.gov/provider-summary-by-type-of-service/medicare-outpatient-hospitals/medicare-outpatient-hosp"| __truncated__
#>  ..  ..@ dimensions: <careTempDim>
#>  .. .. .. @ rows  : int 90591
#>  .. .. .. @ pages : int 19
#>  .. .. .. @ fields: chr [1:15] "Rndrng_Prvdr_Geo_Lvl" "Rndrng_Prvdr_Geo_Cd" "Rndrng_Prvdr_Geo_Desc" "Srvc_Lvl" ...
#>  ..  ..@ endpoints : tibble [8 × 5] (S3: tbl_df/tbl/data.frame)
#>  $ year      : int [1:8] 2022 2021 2020 2019 2018 2017 2016 2015
#>  $ modified  : Date[1:8], format: "2024-06-04" "2024-06-04" ...
#>  $ identifier: chr [1:8] "https://data.cms.gov/data-api/v1/dataset/6af002ba-b934-4491-96dd-d041989b843c/data" "https://data.cms.gov/data-api/v1/dataset/b3537e90-c3c5-4895-914d-71f54123a0a4/data" "https://data.cms.gov/data-api/v1/dataset/31db4293-f75c-4940-bfae-32d554cd164f/data" "https://data.cms.gov/data-api/v1/dataset/46ae348d-8ba4-43df-868e-74ff87eada39/data" ...
#>  $ download  : chr [1:8] "https://data.cms.gov/sites/default/files/2024-06/ced83326-3281-4132-bbf2-b33bef69a4a6/MUP_OUT_RY24_P04_V10_DY21_Geo.csv" "https://data.cms.gov/sites/default/files/2024-06/2b035ff9-f248-4fe5-a51e-1a756729e1de/MUP_OUT_RY24_P04_V10_DY20_Geo.csv" "https://data.cms.gov/sites/default/files/2021-08/MUP_OUT_RY21_P04_V10_DY19_Geo.csv" "https://data.cms.gov/sites/default/files/2020-11/MUP_OHP_R20_P04_V10_D18_Geo_0.csv" ...
#>  $ resources : chr [1:8] "https://data.cms.gov/data-api/v1/dataset-resources/6af002ba-b934-4491-96dd-d041989b843c" "https://data.cms.gov/data-api/v1/dataset-resources/b3537e90-c3c5-4895-914d-71f54123a0a4" "https://data.cms.gov/data-api/v1/dataset-resources/31db4293-f75c-4940-bfae-32d554cd164f" "https://data.cms.gov/data-api/v1/dataset-resources/46ae348d-8ba4-43df-868e-74ff87eada39" ...
#>  .. $ OUT_provider_and_service : <careTemp>
#>  ..  ..@ title     : chr "Medicare Outpatient Hospitals - by Provider and Service"
#>  ..  ..@ metadata  : <careMeta>
#>  .. .. .. @ description: chr "The Medicare Outpatient Hospitals by Provider and Service dataset provides information on services for Original"| __truncated__
#>  .. .. .. @ periodicity: chr "Annually [R/P1Y]"
#>  .. .. .. @ dictionary : chr "https://data.cms.gov/resources/medicare-outpatient-hospitals-by-provider-and-service-data-dictionary"
#>  .. .. .. @ site       : chr "https://data.cms.gov/provider-summary-by-type-of-service/medicare-outpatient-hospitals/medicare-outpatient-hosp"| __truncated__
#>  ..  ..@ dimensions: <careTempDim>
#>  .. .. .. @ rows  : int 116459
#>  .. .. .. @ pages : int 24
#>  .. .. .. @ fields: chr [1:18] "Rndrng_Prvdr_CCN" "Rndrng_Prvdr_Org_Name" "Rndrng_Prvdr_St" "Rndrng_Prvdr_City" ...
#>  ..  ..@ endpoints : tibble [8 × 5] (S3: tbl_df/tbl/data.frame)
#>  $ year      : int [1:8] 2022 2021 2020 2019 2018 2017 2016 2015
#>  $ modified  : Date[1:8], format: "2024-06-04" "2024-06-04" ...
#>  $ identifier: chr [1:8] "https://data.cms.gov/data-api/v1/dataset/456dbe68-8878-49dd-889b-bb8bf432fcbe/data" "https://data.cms.gov/data-api/v1/dataset/a7617aef-046f-4557-8a2a-54f9c9a88b57/data" "https://data.cms.gov/data-api/v1/dataset/8281e9f7-1874-4bd9-a183-a74bec2debd3/data" "https://data.cms.gov/data-api/v1/dataset/f3fc4327-f66e-49e1-ba06-f2607a164406/data" ...
#>  $ download  : chr [1:8] "https://data.cms.gov/sites/default/files/2024-06/657f3480-6e15-4978-8e4d-78bfa03b7a79/MUP_OUT_RY24_P04_V10_DY21_Prov_Svc.csv" "https://data.cms.gov/sites/default/files/2024-06/62b3e7e6-674f-4f74-8c52-c1e32d3905eb/MUP_OUT_RY24_P04_V10_DY20_Prov_Svc.csv" "https://data.cms.gov/sites/default/files/2021-08/MUP_OUT_RY21_P04_V10_DY19_Prov_Svc.csv" "https://data.cms.gov/sites/default/files/2020-11/MUP_OHP_R20_P04_V10_D18_Prov_Svc.csv" ...
#>  $ resources : chr [1:8] "https://data.cms.gov/data-api/v1/dataset-resources/456dbe68-8878-49dd-889b-bb8bf432fcbe" "https://data.cms.gov/data-api/v1/dataset-resources/a7617aef-046f-4557-8a2a-54f9c9a88b57" "https://data.cms.gov/data-api/v1/dataset-resources/8281e9f7-1874-4bd9-a183-a74bec2debd3" "https://data.cms.gov/data-api/v1/dataset-resources/f3fc4327-f66e-49e1-ba06-f2607a164406" ...
careTempGroup("utilization")
#> <careTempGroup>
#>  @ group  : chr "Medicare Physician & Other Practitioners"
#>  @ members:List of 3
#>  .. $ UTIL_geography_and_service: <careTemp>
#>  ..  ..@ title     : chr "Medicare Physician & Other Practitioners - by Geography and Service"
#>  ..  ..@ metadata  : <careMeta>
#>  .. .. .. @ description: chr "The Medicare Physician & Other Practitioners by Geography and Service dataset contains information on use, paym"| __truncated__
#>  .. .. .. @ periodicity: chr "Annually [R/P1Y]"
#>  .. .. .. @ dictionary : chr "https://data.cms.gov/resources/medicare-physician-other-practitioners-by-geography-and-service-data-dictionary"
#>  .. .. .. @ site       : chr "https://data.cms.gov/provider-summary-by-type-of-service/medicare-physician-other-practitioners/medicare-physic"| __truncated__
#>  ..  ..@ dimensions: <careTempDim>
#>  .. .. .. @ rows  : int 268634
#>  .. .. .. @ pages : int 54
#>  .. .. .. @ fields: chr [1:15] "Rndrng_Prvdr_Geo_Lvl" "Rndrng_Prvdr_Geo_Cd" "Rndrng_Prvdr_Geo_Desc" "HCPCS_Cd" ...
#>  ..  ..@ endpoints : tibble [11 × 5] (S3: tbl_df/tbl/data.frame)
#>  $ year      : int [1:11] 2023 2022 2021 2020 2019 2018 2017 2016 2015 2014 ...
#>  $ modified  : Date[1:11], format: "2025-04-24" "2024-06-04" ...
#>  $ identifier: chr [1:11] "https://data.cms.gov/data-api/v1/dataset/ddee9e22-7889-4bef-975a-7853e4cd0fbb/data" "https://data.cms.gov/data-api/v1/dataset/87304f15-9ed0-41dc-a141-6141a0327453/data" "https://data.cms.gov/data-api/v1/dataset/f00f462f-bb61-48b1-a321-1c50d78ce175/data" "https://data.cms.gov/data-api/v1/dataset/31eb3018-43e0-4259-a0d8-e7a1112ffd08/data" ...
#>  $ download  : chr [1:11] "https://data.cms.gov/sites/default/files/2024-05/3167b4d9-10c0-48f0-a680-1165f4eec064/MUP_PHY_R24_P05_V10_D22_Geo.csv" "https://data.cms.gov/sites/default/files/2023-05/0a47308e-812e-42cb-8ecf-0c1f457ab849/MUP_PHY_R23_P05_V10_D21_Geo.csv" "https://data.cms.gov/sites/default/files/2022-07/0d730fe9-deb8-453b-a2fd-ce778286d9f6/MUP_PHY_R22_P05_V10_D20_Geo.csv" "https://data.cms.gov/sites/default/files/2021-08/MUP_PHY_R21_P04_V10_D19_Geo.csv" ...
#>  $ resources : chr [1:11] "https://data.cms.gov/data-api/v1/dataset-resources/ddee9e22-7889-4bef-975a-7853e4cd0fbb" "https://data.cms.gov/data-api/v1/dataset-resources/87304f15-9ed0-41dc-a141-6141a0327453" "https://data.cms.gov/data-api/v1/dataset-resources/f00f462f-bb61-48b1-a321-1c50d78ce175" "https://data.cms.gov/data-api/v1/dataset-resources/31eb3018-43e0-4259-a0d8-e7a1112ffd08" ...
#>  .. $ UTIL_provider             : <careTemp>
#>  ..  ..@ title     : chr "Medicare Physician & Other Practitioners - by Provider"
#>  ..  ..@ metadata  : <careMeta>
#>  .. .. .. @ description: chr "The Medicare Physician & Other Practitioners by Provider dataset provides information on use, payments, submitt"| __truncated__
#>  .. .. .. @ periodicity: chr "Annually [R/P1Y]"
#>  .. .. .. @ dictionary : chr "https://data.cms.gov/resources/medicare-physician-other-practitioners-by-provider-data-dictionary"
#>  .. .. .. @ site       : chr "https://data.cms.gov/provider-summary-by-type-of-service/medicare-physician-other-practitioners/medicare-physic"| __truncated__
#>  ..  ..@ dimensions: <careTempDim>
#>  .. .. .. @ rows  : int 1259343
#>  .. .. .. @ pages : int 252
#>  .. .. .. @ fields: chr [1:81] "Rndrng_NPI" "Rndrng_Prvdr_Last_Org_Name" "Rndrng_Prvdr_First_Name" "Rndrng_Prvdr_MI" ...
#>  ..  ..@ endpoints : tibble [11 × 5] (S3: tbl_df/tbl/data.frame)
#>  $ year      : int [1:11] 2023 2022 2021 2020 2019 2018 2017 2016 2015 2014 ...
#>  $ modified  : Date[1:11], format: "2025-04-24" "2025-03-26" ...
#>  $ identifier: chr [1:11] "https://data.cms.gov/data-api/v1/dataset/8ba584c6-a43a-4b0b-a35a-eb9a59e3a571/data" "https://data.cms.gov/data-api/v1/dataset/21555c17-ec1b-4e74-b2c6-925c6cbb3147/data" "https://data.cms.gov/data-api/v1/dataset/44e0a489-666c-4ea4-a1a2-360b6cdc19db/data" "https://data.cms.gov/data-api/v1/dataset/29d799aa-c660-44fe-a51a-72c4b3e661ac/data" ...
#>  $ download  : chr [1:11] "https://data.cms.gov/sites/default/files/2025-03/88113a46-badb-41ac-855c-62d60ca8c55d/MUP_PHY_R25_P07_V10_D22_Prov.csv" "https://data.cms.gov/sites/default/files/2025-03/0cd67673-9f7c-42e9-bf0f-ca73e8dc4fc4/MUP_PHY_R25_P07_V10_D21_Prov.csv" "https://data.cms.gov/sites/default/files/2025-03/4b0fefe3-eaed-43bd-be63-d45364cb9fd7/MUP_PHY_R25_P07_V10_D20_Prov.csv" "https://data.cms.gov/sites/default/files/2025-03/6f29e406-eb9c-4a60-bbf3-29d2ead443aa/MUP_PHY_R25_P07_V10_D19_Prov.csv" ...
#>  $ resources : chr [1:11] "https://data.cms.gov/data-api/v1/dataset-resources/8ba584c6-a43a-4b0b-a35a-eb9a59e3a571" "https://data.cms.gov/data-api/v1/dataset-resources/21555c17-ec1b-4e74-b2c6-925c6cbb3147" "https://data.cms.gov/data-api/v1/dataset-resources/44e0a489-666c-4ea4-a1a2-360b6cdc19db" "https://data.cms.gov/data-api/v1/dataset-resources/29d799aa-c660-44fe-a51a-72c4b3e661ac" ...
#>  .. $ UTIL_provider_and_service : <careTemp>
#>  ..  ..@ title     : chr "Medicare Physician & Other Practitioners - by Provider and Service"
#>  ..  ..@ metadata  : <careMeta>
#>  .. .. .. @ description: chr "The Medicare Physician & Other Practitioners by Provider and Service dataset provides information on use, payme"| __truncated__
#>  .. .. .. @ periodicity: chr "Annually [R/P1Y]"
#>  .. .. .. @ dictionary : chr "https://data.cms.gov/resources/medicare-physician-other-practitioners-by-provider-and-service-data-dictionary"
#>  .. .. .. @ site       : chr "https://data.cms.gov/provider-summary-by-type-of-service/medicare-physician-other-practitioners/medicare-physic"| __truncated__
#>  ..  ..@ dimensions: <careTempDim>
#>  .. .. .. @ rows  : int 9660647
#>  .. .. .. @ pages : int 1933
#>  .. .. .. @ fields: chr [1:28] "Rndrng_NPI" "Rndrng_Prvdr_Last_Org_Name" "Rndrng_Prvdr_First_Name" "Rndrng_Prvdr_MI" ...
#>  ..  ..@ endpoints : tibble [11 × 5] (S3: tbl_df/tbl/data.frame)
#>  $ year      : int [1:11] 2023 2022 2021 2020 2019 2018 2017 2016 2015 2014 ...
#>  $ modified  : Date[1:11], format: "2025-04-24" "2025-03-26" ...
#>  $ identifier: chr [1:11] "https://data.cms.gov/data-api/v1/dataset/0e9f2f2b-7bf9-451a-912c-e02e654dd725/data" "https://data.cms.gov/data-api/v1/dataset/e650987d-01b7-4f09-b75e-b0b075afbf98/data" "https://data.cms.gov/data-api/v1/dataset/31dc2c47-f297-4948-bfb4-075e1bec3a02/data" "https://data.cms.gov/data-api/v1/dataset/c957b49e-1323-49e7-8678-c09da387551d/data" ...
#>  $ download  : chr [1:11] "https://data.cms.gov/sites/default/files/2025-03/e3333bd8-f6a7-4bbd-b44c-45798e180223/MUP_PHY_R25_P05_V10_D22_Prov_Svc.csv" "https://data.cms.gov/sites/default/files/2025-03/07e1fd37-fed3-4d19-8b8f-274130b18738/MUP_PHY_R25_P05_V10_D21_Prov_Svc.csv" "https://data.cms.gov/sites/default/files/2025-03/e9f5e673-1620-4ed4-9207-4cbccdb4c0ad/MUP_PHY_R25_P05_V10_D20_Prov_Svc.csv" "https://data.cms.gov/sites/default/files/2025-03/6717506c-304a-4d34-bc0b-4a7e466790e6/MUP_PHY_R25_P04_V10_D19_Prov_Svc.csv" ...
#>  $ resources : chr [1:11] "https://data.cms.gov/data-api/v1/dataset-resources/0e9f2f2b-7bf9-451a-912c-e02e654dd725" "https://data.cms.gov/data-api/v1/dataset-resources/e650987d-01b7-4f09-b75e-b0b075afbf98" "https://data.cms.gov/data-api/v1/dataset-resources/31dc2c47-f297-4948-bfb4-075e1bec3a02" "https://data.cms.gov/data-api/v1/dataset-resources/c957b49e-1323-49e7-8678-c09da387551d" ...
careTempGroup("suppliers")
#> <careTempGroup>
#>  @ group  : chr "Medicare Durable Medical Equipment, Devices & Supplies"
#>  @ members:List of 5
#>  .. $ DME_geography_and_service: <careTemp>
#>  ..  ..@ title     : chr "Medicare Durable Medical Equipment, Devices & Supplies - by Geography and Service"
#>  ..  ..@ metadata  : <careMeta>
#>  .. .. .. @ description: chr "The Medicare Durable Medical Equipment, Devices & Supplies by Geography and Service dataset contains informatio"| __truncated__
#>  .. .. .. @ periodicity: chr "Annually [R/P1Y]"
#>  .. .. .. @ dictionary : chr "https://data.cms.gov/resources/medicare-durable-medical-equipment-devices-supplies-by-geography-and-service-dat"| __truncated__
#>  .. .. .. @ site       : chr "https://data.cms.gov/provider-summary-by-type-of-service/medicare-durable-medical-equipment-devices-supplies/me"| __truncated__
#>  ..  ..@ dimensions: <careTempDim>
#>  .. .. .. @ rows  : int 39129
#>  .. .. .. @ pages : int 8
#>  .. .. .. @ fields: chr [1:18] "Rfrg_Prvdr_Geo_Lvl" "Rfrg_Prvdr_Geo_Cd" "Rfrg_Prvdr_Geo_Desc" "RBCS_Lvl" ...
#>  ..  ..@ endpoints : tibble [9 × 5] (S3: tbl_df/tbl/data.frame)
#>  $ year      : int [1:9] 2022 2021 2020 2019 2018 2017 2016 2015 2014
#>  $ modified  : Date[1:9], format: "2024-06-04" "2024-06-04" ...
#>  $ identifier: chr [1:9] "https://data.cms.gov/data-api/v1/dataset/a9d1461a-9c63-49e1-8d91-696082e20e47/data" "https://data.cms.gov/data-api/v1/dataset/daedccd6-5881-4997-93cb-ad6f3d1dd2fd/data" "https://data.cms.gov/data-api/v1/dataset/c5da241c-71d2-4b04-9f44-b324ff4b4c01/data" "https://data.cms.gov/data-api/v1/dataset/3931609c-3f1d-4ab4-95f9-0e31d43613f0/data" ...
#>  $ download  : chr [1:9] "https://data.cms.gov/sites/default/files/2024-05/0de02a58-f287-4757-83e1-b6eeac59bafe/mup_dme_ry24_p05_v10_dy21_geor.csv" "https://data.cms.gov/sites/default/files/2024-05/3de69379-e812-45be-8291-12ee3ae2fc1b/mup_dme_ry24_p05_v10_dy20_geor.csv" "https://data.cms.gov/sites/default/files/2024-05/eaec290e-2e73-48e7-8969-781716a9d154/mup_dme_ry24_p05_v10_dy19_geor.csv" "https://data.cms.gov/sites/default/files/2024-05/18ebb7be-a6d2-487b-a024-7f1132d1c774/mup_dme_ry24_p05_v10_dy18_geor.csv" ...
#>  $ resources : chr [1:9] "https://data.cms.gov/data-api/v1/dataset-resources/a9d1461a-9c63-49e1-8d91-696082e20e47" "https://data.cms.gov/data-api/v1/dataset-resources/daedccd6-5881-4997-93cb-ad6f3d1dd2fd" "https://data.cms.gov/data-api/v1/dataset-resources/c5da241c-71d2-4b04-9f44-b324ff4b4c01" "https://data.cms.gov/data-api/v1/dataset-resources/3931609c-3f1d-4ab4-95f9-0e31d43613f0" ...
#>  .. $ DME_provider             : <careTemp>
#>  ..  ..@ title     : chr "Medicare Durable Medical Equipment, Devices & Supplies - by Referring Provider"
#>  ..  ..@ metadata  : <careMeta>
#>  .. .. .. @ description: chr "The Medicare Durable Medical Equipment, Devices & Supplies by Referring Provider dataset contains information o"| __truncated__
#>  .. .. .. @ periodicity: chr "Annually [R/P1Y]"
#>  .. .. .. @ dictionary : chr "https://data.cms.gov/resources/medicare-durable-medical-equipment-devices-supplies-by-referring-provider-data-dictionary"
#>  .. .. .. @ site       : chr "https://data.cms.gov/provider-summary-by-type-of-service/medicare-durable-medical-equipment-devices-supplies/me"| __truncated__
#>  ..  ..@ dimensions: <careTempDim>
#>  .. .. .. @ rows  : int 385637
#>  .. .. .. @ pages : int 78
#>  .. .. .. @ fields: chr [1:97] "Rfrg_NPI" "Rfrg_Prvdr_Last_Name_Org" "Rfrg_Prvdr_First_Name" "Rfrg_Prvdr_MI" ...
#>  ..  ..@ endpoints : tibble [9 × 5] (S3: tbl_df/tbl/data.frame)
#>  $ year      : int [1:9] 2022 2021 2020 2019 2018 2017 2016 2015 2014
#>  $ modified  : Date[1:9], format: "2025-03-27" "2025-03-27" ...
#>  $ identifier: chr [1:9] "https://data.cms.gov/data-api/v1/dataset/3dccdda1-c3fb-4ac5-908f-18b387495cb3/data" "https://data.cms.gov/data-api/v1/dataset/ac919fb0-dc48-4290-8abf-a4edb8e09139/data" "https://data.cms.gov/data-api/v1/dataset/d3c866db-3caf-4f3c-acc6-086a1770fb6b/data" "https://data.cms.gov/data-api/v1/dataset/d3e0a332-082d-4479-bb3e-f2911eb86497/data" ...
#>  $ download  : chr [1:9] "https://data.cms.gov/sites/default/files/2025-03/461ab785-11fe-414b-866e-d4ceeb108348/mup_dme_ry25_p05_v10_dy21_rfrr.csv" "https://data.cms.gov/sites/default/files/2025-03/86480766-b460-4430-a1b6-2aa78844ac28/mup_dme_ry25_p05_v10_dy20_rfrr.csv" "https://data.cms.gov/sites/default/files/2025-03/de1b2d3b-9cb5-4535-a9fb-b74f2735ea38/mup_dme_ry25_p05_v10_dy19_rfrr.csv" "https://data.cms.gov/sites/default/files/2025-03/8a34ba44-78d3-490c-ade9-979c1c4326fa/mup_dme_ry25_p05_v10_dy18_rfrr.csv" ...
#>  $ resources : chr [1:9] "https://data.cms.gov/data-api/v1/dataset-resources/3dccdda1-c3fb-4ac5-908f-18b387495cb3" "https://data.cms.gov/data-api/v1/dataset-resources/ac919fb0-dc48-4290-8abf-a4edb8e09139" "https://data.cms.gov/data-api/v1/dataset-resources/d3c866db-3caf-4f3c-acc6-086a1770fb6b" "https://data.cms.gov/data-api/v1/dataset-resources/d3e0a332-082d-4479-bb3e-f2911eb86497" ...
#>  .. $ DME_provider_and_service : <careTemp>
#>  ..  ..@ title     : chr "Medicare Durable Medical Equipment, Devices & Supplies - by Referring Provider and Service"
#>  ..  ..@ metadata  : <careMeta>
#>  .. .. .. @ description: chr "The Medicare Durable Medical Equipment, Devices & Supplies by Referring Provider and Service dataset contains i"| __truncated__
#>  .. .. .. @ periodicity: chr "Annually [R/P1Y]"
#>  .. .. .. @ dictionary : chr "https://data.cms.gov/resources/medicare-durable-medical-equipment-devices-supplies-by-referring-provider-and-se"| __truncated__
#>  .. .. .. @ site       : chr "https://data.cms.gov/provider-summary-by-type-of-service/medicare-durable-medical-equipment-devices-supplies/me"| __truncated__
#>  ..  ..@ dimensions: <careTempDim>
#>  .. .. .. @ rows  : int 1457378
#>  .. .. .. @ pages : int 292
#>  .. .. .. @ fields: chr [1:33] "Rfrg_NPI" "Rfrg_Prvdr_Last_Name_Org" "Rfrg_Prvdr_First_Name" "Rfrg_Prvdr_MI" ...
#>  ..  ..@ endpoints : tibble [9 × 5] (S3: tbl_df/tbl/data.frame)
#>  $ year      : int [1:9] 2022 2021 2020 2019 2018 2017 2016 2015 2014
#>  $ modified  : Date[1:9], format: "2025-03-26" "2025-03-26" ...
#>  $ identifier: chr [1:9] "https://data.cms.gov/data-api/v1/dataset/0dd53b4b-67ba-48c7-b8fa-fecbdfc83b70/data" "https://data.cms.gov/data-api/v1/dataset/46ae675c-bc81-40ca-aa79-64da1c1ec9d9/data" "https://data.cms.gov/data-api/v1/dataset/323df359-ceac-4525-a350-e2cd9eb128fe/data" "https://data.cms.gov/data-api/v1/dataset/eb0019f6-791d-4065-ae4e-4761d2f6c9f2/data" ...
#>  $ download  : chr [1:9] "https://data.cms.gov/sites/default/files/2025-03/09669250-0962-4d47-abb2-0387cffefcee/mup_dme_ry25_p05_v10_dy21_rfrhpr.csv" "https://data.cms.gov/sites/default/files/2025-03/5f8059ab-56ad-4789-b298-b0d87d8530f6/mup_dme_ry25_p05_v10_dy20_rfrhpr.csv" "https://data.cms.gov/sites/default/files/2025-03/1dd01437-c808-4f5e-aaf8-af85ff5fce64/mup_dme_ry25_p05_v10_dy19_rfrhpr.csv" "https://data.cms.gov/sites/default/files/2025-03/3d42dca5-bafa-4077-8b58-cfa150a50602/mup_dme_ry25_p05_v10_dy18_rfrhpr.csv" ...
#>  $ resources : chr [1:9] "https://data.cms.gov/data-api/v1/dataset-resources/0dd53b4b-67ba-48c7-b8fa-fecbdfc83b70" "https://data.cms.gov/data-api/v1/dataset-resources/46ae675c-bc81-40ca-aa79-64da1c1ec9d9" "https://data.cms.gov/data-api/v1/dataset-resources/323df359-ceac-4525-a350-e2cd9eb128fe" "https://data.cms.gov/data-api/v1/dataset-resources/eb0019f6-791d-4065-ae4e-4761d2f6c9f2" ...
#>  .. $ DME_supplier             : <careTemp>
#>  ..  ..@ title     : chr "Medicare Durable Medical Equipment, Devices & Supplies - by Supplier"
#>  ..  ..@ metadata  : <careMeta>
#>  .. .. .. @ description: chr "The Medicare Durable Medical Equipment, Devices & Supplies by Supplier dataset contains information on usage, p"| __truncated__
#>  .. .. .. @ periodicity: chr "Annually [R/P1Y]"
#>  .. .. .. @ dictionary : chr "https://data.cms.gov/resources/medicare-durable-medical-equipment-devices-supplies-by-supplier-data-dictionary"
#>  .. .. .. @ site       : chr "https://data.cms.gov/provider-summary-by-type-of-service/medicare-durable-medical-equipment-devices-supplies/me"| __truncated__
#>  ..  ..@ dimensions: <careTempDim>
#>  .. .. .. @ rows  : int 66406
#>  .. .. .. @ pages : int 14
#>  .. .. .. @ fields: chr [1:93] "Suplr_NPI" "Suplr_Prvdr_Last_Name_Org" "Suplr_Prvdr_First_Name" "Suplr_Prvdr_MI" ...
#>  ..  ..@ endpoints : tibble [9 × 5] (S3: tbl_df/tbl/data.frame)
#>  $ year      : int [1:9] 2022 2021 2020 2019 2018 2017 2016 2015 2014
#>  $ modified  : Date[1:9], format: "2025-03-26" "2025-03-26" ...
#>  $ identifier: chr [1:9] "https://data.cms.gov/data-api/v1/dataset/471f9c0d-12c4-4601-a547-559dcfa2d0e2/data" "https://data.cms.gov/data-api/v1/dataset/887561f5-1676-4c93-84b6-d2e50eb609d3/data" "https://data.cms.gov/data-api/v1/dataset/04f11092-de71-4a40-93b1-da2b41e4e938/data" "https://data.cms.gov/data-api/v1/dataset/e8c8eb83-a1d8-49d3-ac99-dda79b147021/data" ...
#>  $ download  : chr [1:9] "https://data.cms.gov/sites/default/files/2025-03/eee83fd5-89a9-47ca-9dde-316c279ca886/mup_dme_ry25_p05_v10_dy21_supr.csv" "https://data.cms.gov/sites/default/files/2025-03/256f2e66-3bd6-4f89-aefb-d13d05c7b47f/mup_dme_ry25_p05_v10_dy20_supr.csv" "https://data.cms.gov/sites/default/files/2025-03/99a8b743-f7be-4968-bfae-c0cfd8770290/mup_dme_ry25_p05_v10_dy19_supr.csv" "https://data.cms.gov/sites/default/files/2025-03/7e2b13a4-9f55-4047-8bd5-df69e0314703/mup_dme_ry25_p05_v10_dy18_supr.csv" ...
#>  $ resources : chr [1:9] "https://data.cms.gov/data-api/v1/dataset-resources/471f9c0d-12c4-4601-a547-559dcfa2d0e2" "https://data.cms.gov/data-api/v1/dataset-resources/887561f5-1676-4c93-84b6-d2e50eb609d3" "https://data.cms.gov/data-api/v1/dataset-resources/04f11092-de71-4a40-93b1-da2b41e4e938" "https://data.cms.gov/data-api/v1/dataset-resources/e8c8eb83-a1d8-49d3-ac99-dda79b147021" ...
#>  .. $ DME_supplier_and_service : <careTemp>
#>  ..  ..@ title     : chr "Medicare Durable Medical Equipment, Devices & Supplies - by Supplier and Service"
#>  ..  ..@ metadata  : <careMeta>
#>  .. .. .. @ description: chr "The Medicare Durable Medical Equipment, Devices & Supplies by Supplier and Service dataset contains information"| __truncated__
#>  .. .. .. @ periodicity: chr "Annually [R/P1Y]"
#>  .. .. .. @ dictionary : chr "https://data.cms.gov/resources/medicare-durable-medical-equipment-devices-supplies-by-supplier-and-service-data-dictionary"
#>  .. .. .. @ site       : chr "https://data.cms.gov/provider-summary-by-type-of-service/medicare-durable-medical-equipment-devices-supplies/me"| __truncated__
#>  ..  ..@ dimensions: <careTempDim>
#>  .. .. .. @ rows  : int 482638
#>  .. .. .. @ pages : int 97
#>  .. .. .. @ fields: chr [1:32] "Suplr_NPI" "Suplr_Prvdr_Last_Name_Org" "Suplr_Prvdr_First_Name" "Suplr_Prvdr_MI" ...
#>  ..  ..@ endpoints : tibble [9 × 5] (S3: tbl_df/tbl/data.frame)
#>  $ year      : int [1:9] 2022 2021 2020 2019 2018 2017 2016 2015 2014
#>  $ modified  : Date[1:9], format: "2025-03-26" "2025-03-26" ...
#>  $ identifier: chr [1:9] "https://data.cms.gov/data-api/v1/dataset/dad703df-2e52-4ac6-a26f-51ecbf463d77/data" "https://data.cms.gov/data-api/v1/dataset/e531ffc9-57ad-48db-8d88-f686edb8b8e1/data" "https://data.cms.gov/data-api/v1/dataset/1fe3a8c3-ecfd-44b9-9418-fda1843f72a5/data" "https://data.cms.gov/data-api/v1/dataset/5165dbdd-eae3-4b94-82cd-da635e00e0fa/data" ...
#>  $ download  : chr [1:9] "https://data.cms.gov/sites/default/files/2025-03/d298ae41-63f8-4d67-b249-5a6f4e01a36e/mup_dme_ry25_p05_v10_dy21_suphpr.csv" "https://data.cms.gov/sites/default/files/2025-03/d5f7f823-76ba-4f22-a590-0b674d20c937/mup_dme_ry25_p05_v10_dy20_suphpr.csv" "https://data.cms.gov/sites/default/files/2025-03/d505e663-630d-4b9c-843e-cd56a3cab3fd/mup_dme_ry25_p05_v10_dy19_suphpr.csv" "https://data.cms.gov/sites/default/files/2025-03/9d0f8f71-292a-4303-a8a9-2e39b6924f9d/mup_dme_ry25_p05_v10_dy18_suphpr.csv" ...
#>  $ resources : chr [1:9] "https://data.cms.gov/data-api/v1/dataset-resources/dad703df-2e52-4ac6-a26f-51ecbf463d77" "https://data.cms.gov/data-api/v1/dataset-resources/e531ffc9-57ad-48db-8d88-f686edb8b8e1" "https://data.cms.gov/data-api/v1/dataset-resources/1fe3a8c3-ecfd-44b9-9418-fda1843f72a5" "https://data.cms.gov/data-api/v1/dataset-resources/5165dbdd-eae3-4b94-82cd-da635e00e0fa" ...
careTempGroup("prescribers")
#> <careTempGroup>
#>  @ group  : chr "Medicare Part D Prescribers"
#>  @ members:List of 3
#>  .. $ PRX_geography_and_drug: <careTemp>
#>  ..  ..@ title     : chr "Medicare Part D Prescribers - by Geography and Drug"
#>  ..  ..@ metadata  : <careMeta>
#>  .. .. .. @ description: chr "The Medicare Part D Prescribers by Geography and Drug dataset contains information on prescription drugs prescr"| __truncated__
#>  .. .. .. @ periodicity: chr "Annually [R/P1Y]"
#>  .. .. .. @ dictionary : chr "https://data.cms.gov/resources/medicare-part-d-prescribers-by-geography-and-drug-data-dictionary"
#>  .. .. .. @ site       : chr "https://data.cms.gov/provider-summary-by-type-of-service/medicare-part-d-prescribers/medicare-part-d-prescriber"| __truncated__
#>  ..  ..@ dimensions: <careTempDim>
#>  .. .. .. @ rows  : int 115936
#>  .. .. .. @ pages : int 24
#>  .. .. .. @ fields: chr [1:22] "Prscrbr_Geo_Lvl" "Prscrbr_Geo_Cd" "Prscrbr_Geo_Desc" "Brnd_Name" ...
#>  ..  ..@ endpoints : tibble [11 × 5] (S3: tbl_df/tbl/data.frame)
#>  $ year      : int [1:11] 2023 2022 2021 2020 2019 2018 2017 2016 2015 2014 ...
#>  $ modified  : Date[1:11], format: "2025-04-24" "2024-06-04" ...
#>  $ identifier: chr [1:11] "https://data.cms.gov/data-api/v1/dataset/3463648b-1971-478d-84ca-80cadc758153/data" "https://data.cms.gov/data-api/v1/dataset/1fc57194-a51d-4864-aee6-de0889488151/data" "https://data.cms.gov/data-api/v1/dataset/7dda2a9d-034a-446a-b4b3-e1254e0127b2/data" "https://data.cms.gov/data-api/v1/dataset/83891e77-99cf-4865-b60a-97703b916e09/data" ...
#>  $ download  : chr [1:11] "https://data.cms.gov/sites/default/files/2024-05/410ed206-d782-4dba-b442-6bcd45ae2016/MUP_DPR_RY24_P04_V10_DY22_Geo.csv" "https://data.cms.gov/sites/default/files/2023-04/3d3ebd5b-b4bf-45b4-876d-afa7916d1b72/MUP_DPR_RY23_P04_V10_DY21_Geo.csv" "https://data.cms.gov/sites/default/files/2022-07/ca71b7df-4d48-4c2d-aded-2ca22285739c/MUP_DPR_RY22_P04_V10_DY20_Geo.csv" "https://data.cms.gov/sites/default/files/2021-08/MUP_DPR_RY21_P04_V10_DY19_Geo.csv" ...
#>  $ resources : chr [1:11] "https://data.cms.gov/data-api/v1/dataset-resources/3463648b-1971-478d-84ca-80cadc758153" "https://data.cms.gov/data-api/v1/dataset-resources/1fc57194-a51d-4864-aee6-de0889488151" "https://data.cms.gov/data-api/v1/dataset-resources/7dda2a9d-034a-446a-b4b3-e1254e0127b2" "https://data.cms.gov/data-api/v1/dataset-resources/83891e77-99cf-4865-b60a-97703b916e09" ...
#>  .. $ PRX_provider          : <careTemp>
#>  ..  ..@ title     : chr "Medicare Part D Prescribers - by Provider"
#>  ..  ..@ metadata  : <careMeta>
#>  .. .. .. @ description: chr "The Medicare Part D Prescribers by Provider dataset contains information on prescription drugs prescribed by in"| __truncated__
#>  .. .. .. @ periodicity: chr "Annually [R/P1Y]"
#>  .. .. .. @ dictionary : chr "https://data.cms.gov/resources/medicare-part-d-prescribers-by-provider-data-dictionary"
#>  .. .. .. @ site       : chr "https://data.cms.gov/provider-summary-by-type-of-service/medicare-part-d-prescribers/medicare-part-d-prescribers-by-provider"
#>  ..  ..@ dimensions: <careTempDim>
#>  .. .. .. @ rows  : int 1380665
#>  .. .. .. @ pages : int 277
#>  .. .. .. @ fields: chr [1:84] "PRSCRBR_NPI" "Prscrbr_Last_Org_Name" "Prscrbr_First_Name" "Prscrbr_MI" ...
#>  ..  ..@ endpoints : tibble [11 × 5] (S3: tbl_df/tbl/data.frame)
#>  $ year      : int [1:11] 2023 2022 2021 2020 2019 2018 2017 2016 2015 2014 ...
#>  $ modified  : Date[1:11], format: "2025-04-24" "2025-03-27" ...
#>  $ identifier: chr [1:11] "https://data.cms.gov/data-api/v1/dataset/6428fad7-4fb1-454c-93a4-772934d73922/data" "https://data.cms.gov/data-api/v1/dataset/bed99012-c527-4d9d-92ea-67ec2510abea/data" "https://data.cms.gov/data-api/v1/dataset/3f7ab9ce-6fb6-4e6b-9af3-b681f2d3a95e/data" "https://data.cms.gov/data-api/v1/dataset/ed6548c1-c905-4fdf-84cc-1c897ca6210d/data" ...
#>  $ download  : chr [1:11] "https://data.cms.gov/sites/default/files/2025-03/dc2568de-65a2-490e-9f8c-cbcfc4ea2514/MUP_DPR_RY25_P04_V10_DY22_NPI.csv" "https://data.cms.gov/sites/default/files/2025-03/537a8fd7-f826-4eaf-b12e-d8ed6b631355/MUP_DPR_RY25_P04_V10_DY21_NPI.csv" "https://data.cms.gov/sites/default/files/2025-03/d8af7b1a-1261-454d-b7bd-2e3b1f671a0a/MUP_DPR_RY25_P04_V10_DY20_NPI.csv" "https://data.cms.gov/sites/default/files/2025-03/512bf0ea-628e-4ffd-a17f-387171ab5e71/MUP_DPR_RY25_P04_V10_DY19_NPI.csv" ...
#>  $ resources : chr [1:11] "https://data.cms.gov/data-api/v1/dataset-resources/6428fad7-4fb1-454c-93a4-772934d73922" "https://data.cms.gov/data-api/v1/dataset-resources/bed99012-c527-4d9d-92ea-67ec2510abea" "https://data.cms.gov/data-api/v1/dataset-resources/3f7ab9ce-6fb6-4e6b-9af3-b681f2d3a95e" "https://data.cms.gov/data-api/v1/dataset-resources/ed6548c1-c905-4fdf-84cc-1c897ca6210d" ...
#>  .. $ PRX_provider_and_drug : <careTemp>
#>  ..  ..@ title     : chr "Medicare Part D Prescribers - by Provider and Drug"
#>  ..  ..@ metadata  : <careMeta>
#>  .. .. .. @ description: chr "The Medicare Part D Prescribers by Provider and Drug dataset provides information on prescription drugs prescri"| __truncated__
#>  .. .. .. @ periodicity: chr "Annually [R/P1Y]"
#>  .. .. .. @ dictionary : chr "https://data.cms.gov/resources/medicare-part-d-prescribers-by-provider-and-drug-data-dictionary"
#>  .. .. .. @ site       : chr "https://data.cms.gov/provider-summary-by-type-of-service/medicare-part-d-prescribers/medicare-part-d-prescriber"| __truncated__
#>  ..  ..@ dimensions: <careTempDim>
#>  .. .. .. @ rows  : int 26794878
#>  .. .. .. @ pages : int 5359
#>  .. .. .. @ fields: chr [1:22] "Prscrbr_NPI" "Prscrbr_Last_Org_Name" "Prscrbr_First_Name" "Prscrbr_City" ...
#>  ..  ..@ endpoints : tibble [11 × 5] (S3: tbl_df/tbl/data.frame)
#>  $ year      : int [1:11] 2023 2022 2021 2020 2019 2018 2017 2016 2015 2014 ...
#>  $ modified  : Date[1:11], format: "2025-04-24" "2024-06-04" ...
#>  $ identifier: chr [1:11] "https://data.cms.gov/data-api/v1/dataset/e54db557-cd82-4e91-a0fe-61aad5865d69/data" "https://data.cms.gov/data-api/v1/dataset/b101b457-ffa4-49bb-8fd9-27c1266086e2/data" "https://data.cms.gov/data-api/v1/dataset/f68114ed-f854-4ffc-9c6e-ed78b5e2f8d0/data" "https://data.cms.gov/data-api/v1/dataset/7795fe20-e80e-435a-a9ed-d2d65e05feeb/data" ...
#>  $ download  : chr [1:11] "https://data.cms.gov/sites/default/files/2024-05/18f82097-61a6-4889-9941-9a0b6ad7523c/MUP_DPR_RY24_P04_V10_DY22_NPIBN.csv" "https://data.cms.gov/sites/default/files/2024-05/43359391-e7fa-40b9-9bd4-5dc295e18712/MUP_DPR_RY24_P04_V10_DY21_NPIBN.csv" "https://data.cms.gov/sites/default/files/2024-05/75fecc51-c9e8-4904-b570-9da9dc101721/MUP_DPR_RY24_P04_V10_DY20_NPIBN.csv" "https://data.cms.gov/sites/default/files/2024-05/129e7c21-d492-425b-be03-d2e59d933ab6/MUP_DPR_RY24_P04_V10_DY19_NPIBN.csv" ...
#>  $ resources : chr [1:11] "https://data.cms.gov/data-api/v1/dataset-resources/e54db557-cd82-4e91-a0fe-61aad5865d69" "https://data.cms.gov/data-api/v1/dataset-resources/b101b457-ffa4-49bb-8fd9-27c1266086e2" "https://data.cms.gov/data-api/v1/dataset-resources/f68114ed-f854-4ffc-9c6e-ed78b5e2f8d0" "https://data.cms.gov/data-api/v1/dataset-resources/7795fe20-e80e-435a-a9ed-d2d65e05feeb" ...
careTempGroup("staffing")
#> <careTempGroup>
#>  @ group  : chr "Nursing Home Payroll-Based Journal Staffing"
#>  @ members:List of 3
#>  .. $ STAFF_non_nurse: <careTemp>
#>  ..  ..@ title     : chr "Payroll Based Journal Daily Non-Nurse Staffing"
#>  ..  ..@ metadata  : <careMeta>
#>  .. .. .. @ description: chr "The Payroll Based Journal (PBJ) Nurse Staffing and Non-Nurse Staffing datasets provide information submitted by"| __truncated__
#>  .. .. .. @ periodicity: chr "Quarterly [R/P3M]"
#>  .. .. .. @ dictionary : chr "https://data.cms.gov/resources/payroll-based-journal-daily-non-nurse-staffing-data-dictionary"
#>  .. .. .. @ site       : chr "https://data.cms.gov/quality-of-care/payroll-based-journal-daily-non-nurse-staffing"
#>  ..  ..@ dimensions: <careTempDim>
#>  .. .. .. @ rows  : int 1338416
#>  .. .. .. @ pages : int 268
#>  .. .. .. @ fields: chr [1:82] "PROVNUM" "PROVNAME" "CITY" "STATE" ...
#>  ..  ..@ endpoints : tibble [8 × 5] (S3: tbl_df/tbl/data.frame)
#>  $ year      : int [1:8] 2024 2023 2022 2021 2020 2019 2018 2017
#>  $ modified  : Date[1:8], format: "2025-04-02" "2024-04-24" ...
#>  $ identifier: chr [1:8] "https://data.cms.gov/data-api/v1/dataset/293788f3-20a9-4ebd-ab4a-9989325ce3e7/data" "https://data.cms.gov/data-api/v1/dataset/80e14392-7640-4fef-b226-309929273962/data" "https://data.cms.gov/data-api/v1/dataset/161d21ed-2f94-46da-aeb3-6bdd58d57264/data" "https://data.cms.gov/data-api/v1/dataset/d805bf46-faa2-48d8-89af-1f40def16c36/data" ...
#>  $ download  : chr [1:8] "https://data.cms.gov/sites/default/files/2024-10/f9a724c1-1785-494f-b34d-1b66c1a53b8c/PBJ_dailyNonnursestaffing_CY2024Q2.csv" "https://data.cms.gov/sites/default/files/2024-01/da7cceca-fbef-4461-8286-b8a777323344/PBJ_dailyNonnurseStaffing_CY2023Q3.csv" "https://data.cms.gov/sites/default/files/2023-01/f51658e7-a4f8-4711-ae29-98de30cfe5aa/PBJ_dailyNonnursestaffing_CY2022Q3.csv" "https://data.cms.gov/sites/default/files/2022-01/6460c80d-40b1-4de2-98ba-a850d0a30ffa/PBJ_dailyNonnursestaffing_CY2021Q3.csv" ...
#>  $ resources : chr [1:8] "https://data.cms.gov/data-api/v1/dataset-resources/293788f3-20a9-4ebd-ab4a-9989325ce3e7" "https://data.cms.gov/data-api/v1/dataset-resources/80e14392-7640-4fef-b226-309929273962" "https://data.cms.gov/data-api/v1/dataset-resources/161d21ed-2f94-46da-aeb3-6bdd58d57264" "https://data.cms.gov/data-api/v1/dataset-resources/d805bf46-faa2-48d8-89af-1f40def16c36" ...
#>  .. $ STAFF_nurse    : <careTemp>
#>  ..  ..@ title     : chr "Payroll Based Journal Daily Nurse Staffing"
#>  ..  ..@ metadata  : <careMeta>
#>  .. .. .. @ description: chr "The Payroll Based Journal (PBJ) Nurse Staffing and Non-Nurse Staffing datasets provide information submitted by"| __truncated__
#>  .. .. .. @ periodicity: chr "Quarterly [R/P3M]"
#>  .. .. .. @ dictionary : chr "https://data.cms.gov/resources/payroll-based-journal-daily-nurse-staffing-data-dictionary"
#>  .. .. .. @ site       : chr "https://data.cms.gov/quality-of-care/payroll-based-journal-daily-nurse-staffing"
#>  ..  ..@ dimensions: <careTempDim>
#>  .. .. .. @ rows  : int 1338416
#>  .. .. .. @ pages : int 268
#>  .. .. .. @ fields: chr [1:33] "PROVNUM" "PROVNAME" "CITY" "STATE" ...
#>  ..  ..@ endpoints : tibble [8 × 5] (S3: tbl_df/tbl/data.frame)
#>  $ year      : int [1:8] 2024 2023 2022 2021 2020 2019 2018 2017
#>  $ modified  : Date[1:8], format: "2025-04-02" "2024-04-24" ...
#>  $ identifier: chr [1:8] "https://data.cms.gov/data-api/v1/dataset/989fbc78-1655-487d-9f24-d68e9a0ab3af/data" "https://data.cms.gov/data-api/v1/dataset/ef20ab44-833c-4611-9600-2862e460072c/data" "https://data.cms.gov/data-api/v1/dataset/7a678e92-0418-44e0-9d72-043fd23f3959/data" "https://data.cms.gov/data-api/v1/dataset/93881636-0cde-4118-a110-ba2fd8cc890a/data" ...
#>  $ download  : chr [1:8] "https://data.cms.gov/sites/default/files/2024-10/79f8602d-e17a-4ba3-9182-3287ccff9d3c/PBJ_dailynursestaffing_CY2024Q2.csv" "https://data.cms.gov/sites/default/files/2024-01/8e987728-d824-48a0-b041-2d3c7a0ee2e8/PBJ_dailynursestaffing_CY2023Q3.csv" "https://data.cms.gov/sites/default/files/2023-01/893d09c1-df09-41ef-b7b9-b47e9dbb6e05/PBJ_dailynursestaffing_CY2022Q3.csv" "https://data.cms.gov/sites/default/files/2022-01/416c11d0-1ef7-4178-94a3-6b4ebeef4677/PBJ_dailynursestaffing_CY2021Q3.csv" ...
#>  $ resources : chr [1:8] "https://data.cms.gov/data-api/v1/dataset-resources/989fbc78-1655-487d-9f24-d68e9a0ab3af" "https://data.cms.gov/data-api/v1/dataset-resources/ef20ab44-833c-4611-9600-2862e460072c" "https://data.cms.gov/data-api/v1/dataset-resources/7a678e92-0418-44e0-9d72-043fd23f3959" "https://data.cms.gov/data-api/v1/dataset-resources/93881636-0cde-4118-a110-ba2fd8cc890a" ...
#>  .. $ STAFF_employee : <careTemp>
#>  ..  ..@ title     : chr "Payroll Based Journal Employee Detail Nursing Home Staffing"
#>  ..  ..@ metadata  : <careMeta>
#>  .. .. .. @ description: chr "The Payroll Based Journal (PBJ) Employee Detail Nursing Home Staffing dataset provides information submitted by"| __truncated__
#>  .. .. .. @ periodicity: chr "Quarterly [R/P3M]"
#>  .. .. .. @ dictionary : chr "https://data.cms.gov/resources/payroll-based-journal-employee-detail-nursing-home-staffing-data-dictionary"
#>  .. .. .. @ site       : chr "https://data.cms.gov/quality-of-care/payroll-based-journal-employee-detail-nursing-home-staffing"
#>  ..  ..@ dimensions: <careTempDim>
#>  .. .. .. @ rows  : int 63260729
#>  .. .. .. @ pages : int 12653
#>  .. .. .. @ fields: chr [1:9] "PROVNUM" "STATE" "CY_Qtr" "WorkDate" ...
#>  ..  ..@ endpoints : tibble [5 × 5] (S3: tbl_df/tbl/data.frame)
#>  $ year      : int [1:5] 2024 2023 2022 2021 2020
#>  $ modified  : Date[1:5], format: "2025-04-02" "2024-04-24" ...
#>  $ identifier: chr [1:5] "https://data.cms.gov/data-api/v1/dataset/f513e55c-236c-437c-9a02-b53cc4acad72/data" "https://data.cms.gov/data-api/v1/dataset/86f6f9ad-2cb2-447c-b460-cf4f819fea7c/data" "https://data.cms.gov/data-api/v1/dataset/91c55823-f84a-4b68-b770-9f594f9e776f/data" "https://data.cms.gov/data-api/v1/dataset/24e7953e-ed2a-4931-a272-a4c0995c4865/data" ...
#>  $ download  : chr [1:5] "https://data.cms.gov/sites/default/files/2024-10/a284dec3-92da-4de4-a1a2-3c91e15c43e7/PBJ_employeedetail_CY2024Q2.csv" "https://data.cms.gov/sites/default/files/2024-01/009f82da-932f-4daf-82dc-25f33b655a45/PBJ_employeedetail_CY2023Q3.csv" "https://data.cms.gov/sites/default/files/2023-04/PBJ_employeedetail_CY2022Q3.csv" "https://data.cms.gov/sites/default/files/2023-05/5aff92dc-fa72-47dd-b438-a5b4791666db/PBJ_employeedetail_CY2021Q3.csv" ...
#>  $ resources : chr [1:5] "https://data.cms.gov/data-api/v1/dataset-resources/f513e55c-236c-437c-9a02-b53cc4acad72" "https://data.cms.gov/data-api/v1/dataset-resources/86f6f9ad-2cb2-447c-b460-cf4f819fea7c" "https://data.cms.gov/data-api/v1/dataset-resources/91c55823-f84a-4b68-b770-9f594f9e776f" "https://data.cms.gov/data-api/v1/dataset-resources/24e7953e-ed2a-4931-a272-a4c0995c4865" ...