Medicare Endpoint

Medicare Endpoint

Usage

care_endpoint(alias)

Arguments

alias

<chr> endpoint alias

Value

An S7 <care_endpoint> object.

Examples

care_endpoint("contact")
#> <care_endpoint>
#>  @ title     : chr "Public Reporting of Missing Digital Contact Information"
#>  @ identifier: chr "https://data.cms.gov/data-api/v1/dataset/63a83bb1-4c02-43b3-8ef4-e3d3c6cf62fa/data-viewer"
#>  @ dimensions: <care_dimensions>
#>  .. @ limit : int 5000
#>  .. @ rows  : int 3505696
#>  .. @ pages : int 702
#>  .. @ fields:List of 2
#>  .. .. $ NPI          : chr "NPI"
#>  .. .. $ Provider Name: chr "Provider Name"
#>  @ metadata  : <care_metadata>
#>  .. @ description: chr "In the May 2020 CMS Interoperability and Patient Access final rule, CMS finalized the policy to publicly report"| __truncated__
#>  .. @ temporal   : chr "2023-01-01 • 2024-09-30"
#>  .. @ periodicity: chr "Quarterly [R/P3M]"
#>  .. @ modified   : Date[1:1], format: "2024-10-25"
#>  .. @ dictionary : chr "https://data.cms.gov/resources/public-reporting-of-missing-digital-contact-information-data-dictionary"
#>  .. @ site       : chr "https://data.cms.gov/provider-compliance/public-reporting-of-missing-digital-contact-information"
#>  .. @ references : chr "https://data.cms.gov/sites/default/files/2021-12/8eb2b4bf-6e5f-4e05-bcdb-39c07ad8f77a/Missing_Digital_Contact_I"| __truncated__
#>  .. @ download   : chr "https://data.cms.gov/sites/default/files/2024-10/fb284594-a4ff-442f-8b66-f6838c4355ce/Providers%20without%20End"| __truncated__
#>  @ resources : chr "https://data.cms.gov/data-api/v1/dataset-resources/63a83bb1-4c02-43b3-8ef4-e3d3c6cf62fa"
care_endpoint("crosswalk")
#> <care_endpoint>
#>  @ title     : chr "Medicare Provider and Supplier Taxonomy Crosswalk"
#>  @ identifier: chr "https://data.cms.gov/data-api/v1/dataset/113eb0bc-0c9a-4d91-9f93-3f6b28c0bf6b/data-viewer"
#>  @ dimensions: <care_dimensions>
#>  .. @ limit : int 5000
#>  .. @ rows  : int 561
#>  .. @ pages : int 1
#>  .. @ fields:List of 4
#>  .. .. $ MEDICARE SPECIALTY CODE                                             : chr "MEDICARE SPECIALTY CODE"
#>  .. .. $ MEDICARE PROVIDER/SUPPLIER TYPE DESCRIPTION                         : chr "MEDICARE PROVIDER/SUPPLIER TYPE DESCRIPTION"
#>  .. .. $ PROVIDER TAXONOMY CODE                                              : chr "PROVIDER TAXONOMY CODE"
#>  .. .. $ PROVIDER TAXONOMY DESCRIPTION:  TYPE, CLASSIFICATION, SPECIALIZATION: chr "PROVIDER TAXONOMY DESCRIPTION:  TYPE, CLASSIFICATION, SPECIALIZATION"
#>  @ metadata  : <care_metadata>
#>  .. @ description: chr "The Medicare Provider and Supplier Taxonomy Crosswalk dataset lists the providers and suppliers eligible to enr"| __truncated__
#>  .. @ temporal   : chr "2021-07-01 • 2025-06-30"
#>  .. @ periodicity: chr "Biannually [R/P6M]"
#>  .. @ modified   : Date[1:1], format: "2025-04-22"
#>  .. @ dictionary : chr "https://data.cms.gov/resources/medicare-provider-and-supplier-taxonomy-crosswalk-data-dictionary"
#>  .. @ site       : chr "https://data.cms.gov/provider-characteristics/medicare-provider-supplier-enrollment/medicare-provider-and-suppl"| __truncated__
#>  .. @ references : chr "https://data.cms.gov/resources/medicare-provider-and-supplier-taxonomy-crosswalk-methodology"
#>  .. @ download   : chr "https://data.cms.gov/sites/default/files/2025-04/d7c24ae8-908c-4174-8470-c31c1c55e2a8/Medicare_Provider_and_Sup"| __truncated__
#>  @ resources : chr "https://data.cms.gov/data-api/v1/dataset-resources/113eb0bc-0c9a-4d91-9f93-3f6b28c0bf6b"
care_endpoint("CARE_dialysis")
#> <care_endpoint>
#>  @ title     : chr "Medicare Dialysis Facilities"
#>  @ identifier: chr "https://data.cms.gov/data-api/v1/dataset/f8610e87-ba25-43a3-a49e-927dbc8701ae/data-viewer"
#>  @ dimensions: <care_dimensions>
#>  .. @ limit : int 5000
#>  .. @ rows  : int 11679243
#>  .. @ pages : int 2336
#>  .. @ fields:List of 14
#>  .. .. $ state         : chr "state"
#>  .. .. $ CCN           : chr "CCN"
#>  .. .. $ Provider_Name : chr "Provider_Name"
#>  .. .. $ city          : chr "city"
#>  .. .. $ Ownership_Type: chr "Ownership_Type"
#>  .. .. $ ESRD_Network  : chr "ESRD_Network"
#>  .. .. $ NPI           : chr "NPI"
#>  .. .. $ Chain         : chr "Chain"
#>  .. .. $ Modality      : chr "Modality"
#>  .. .. $ Alternate_CCNs: chr "Alternate_CCNs"
#>  .. .. $ Measure       : chr "Measure"
#>  .. .. $ Measure_Score : chr "Measure_Score"
#>  .. .. $ year          : chr "year"
#>  .. .. $ Measure_ID    : chr "Measure_ID"
#>  @ metadata  : <care_metadata>
#>  .. @ description: chr "The Medicare Dialysis Facilities data provides information on clinical and patient measures for Medicare-certif"| __truncated__
#>  .. @ temporal   : chr "2009-01-01 • 2024-12-31"
#>  .. @ periodicity: chr "Annually [R/P1Y]"
#>  .. @ modified   : Date[1:1], format: "2024-02-29"
#>  .. @ dictionary : chr "https://data.cms.gov/sites/default/files/2024-02/FY2024%20Data%20Dictionary.xlsx"
#>  .. @ site       : chr "https://data.cms.gov/quality-of-care/medicare-dialysis-facilities"
#>  .. @ references : chr "https://data.cms.gov/sites/default/files/2024-02/FY2024_DFR_Guide%20and%20Methodology.pdf"
#>  .. @ download   : chr "https://data.cms.gov/sites/default/files/2024-02/b2c2658d-e19d-48db-8b3b-9d343a4d395f/dfr_facility_socrata_fy2024.csv"
#>  @ resources : chr "https://data.cms.gov/data-api/v1/dataset-resources/f8610e87-ba25-43a3-a49e-927dbc8701ae"
care_endpoint("enrollees")
#> <care_endpoint>
#>  @ title     : chr "Medicare Fee-For-Service Public Provider Enrollment"
#>  @ identifier: chr "https://data.cms.gov/data-api/v1/dataset/2457ea29-fc82-48b0-86ec-3b0755de7515/data-viewer"
#>  @ dimensions: <care_dimensions>
#>  .. @ limit : int 5000
#>  .. @ rows  : int 2922752
#>  .. @ pages : int 585
#>  .. @ fields:List of 10
#>  .. .. $ NPI               : chr "NPI"
#>  .. .. $ PECOS_ASCT_CNTL_ID: chr "PECOS_ASCT_CNTL_ID"
#>  .. .. $ ENRLMT_ID         : chr "ENRLMT_ID"
#>  .. .. $ PROVIDER_TYPE_CD  : chr "PROVIDER_TYPE_CD"
#>  .. .. $ PROVIDER_TYPE_DESC: chr "PROVIDER_TYPE_DESC"
#>  .. .. $ STATE_CD          : chr "STATE_CD"
#>  .. .. $ FIRST_NAME        : chr "FIRST_NAME"
#>  .. .. $ MDL_NAME          : chr "MDL_NAME"
#>  .. .. $ LAST_NAME         : chr "LAST_NAME"
#>  .. .. $ ORG_NAME          : chr "ORG_NAME"
#>  @ metadata  : <care_metadata>
#>  .. @ description: chr "The Medicare Fee-For-Service Public Provider Enrollment dataset includes information on providers who are activ"| __truncated__
#>  .. @ temporal   : chr "2023-01-01 • 2025-03-31"
#>  .. @ periodicity: chr "Quarterly [R/P3M]"
#>  .. @ modified   : Date[1:1], format: "2025-04-15"
#>  .. @ dictionary : chr "https://data.cms.gov/resources/medicare-fee-for-service-public-provider-enrollment-data-dictionary"
#>  .. @ site       : chr "https://data.cms.gov/provider-characteristics/medicare-provider-supplier-enrollment/medicare-fee-for-service-pu"| __truncated__
#>  .. @ references : chr "https://data.cms.gov/resources/fee-for-service-public-provider-enrollment-methodology"
#>  .. @ download   : chr "https://data.cms.gov/sites/default/files/2025-04/5d10bc83-29bd-498e-a1ff-3e6e2065a7e3/PPEF_Enrollment_Extract_2025.04.01.csv"
#>  @ resources : chr "https://data.cms.gov/data-api/v1/dataset-resources/2457ea29-fc82-48b0-86ec-3b0755de7515"
care_endpoint("facilities")
#> <care_endpoint>
#>  @ title     : chr "Provider of Services File - Hospital & Non-Hospital Facilities"
#>  @ identifier: chr "https://data.cms.gov/data-api/v1/dataset/8ba0f9b4-9493-4aa0-9f82-44ea9468d1b5/data-viewer"
#>  @ dimensions: <care_dimensions>
#>  .. @ limit : int 5000
#>  .. @ rows  : int 116047
#>  .. @ pages : int 24
#>  .. @ fields:List of 473
#>  .. .. $ PRVDR_CTGRY_SBTYP_CD          : chr "PRVDR_CTGRY_SBTYP_CD"
#>  .. .. $ PRVDR_CTGRY_CD                : chr "PRVDR_CTGRY_CD"
#>  .. .. $ CHOW_CNT                      : chr "CHOW_CNT"
#>  .. .. $ CHOW_DT                       : chr "CHOW_DT"
#>  .. .. $ CITY_NAME                     : chr "CITY_NAME"
#>  .. .. $ ACPTBL_POC_SW                 : chr "ACPTBL_POC_SW"
#>  .. .. $ CMPLNC_STUS_CD                : chr "CMPLNC_STUS_CD"
#>  .. .. $ SSA_CNTY_CD                   : chr "SSA_CNTY_CD"
#>  .. .. $ CROSS_REF_PROVIDER_NUMBER     : chr "CROSS_REF_PROVIDER_NUMBER"
#>  .. .. $ CRTFCTN_DT                    : chr "CRTFCTN_DT"
#>  .. .. $ ELGBLTY_SW                    : chr "ELGBLTY_SW"
#>  .. .. $ FAC_NAME                      : chr "FAC_NAME"
#>  .. .. $ INTRMDRY_CARR_CD              : chr "INTRMDRY_CARR_CD"
#>  .. .. $ MDCD_VNDR_NUM                 : chr "MDCD_VNDR_NUM"
#>  .. .. $ ORGNL_PRTCPTN_DT              : chr "ORGNL_PRTCPTN_DT"
#>  .. .. $ CHOW_PRIOR_DT                 : chr "CHOW_PRIOR_DT"
#>  .. .. $ INTRMDRY_CARR_PRIOR_CD        : chr "INTRMDRY_CARR_PRIOR_CD"
#>  .. .. $ PRVDR_NUM                     : chr "PRVDR_NUM"
#>  .. .. $ RGN_CD                        : chr "RGN_CD"
#>  .. .. $ SKLTN_REC_SW                  : chr "SKLTN_REC_SW"
#>  .. .. $ STATE_CD                      : chr "STATE_CD"
#>  .. .. $ SSA_STATE_CD                  : chr "SSA_STATE_CD"
#>  .. .. $ STATE_RGN_CD                  : chr "STATE_RGN_CD"
#>  .. .. $ ST_ADR                        : chr "ST_ADR"
#>  .. .. $ PHNE_NUM                      : chr "PHNE_NUM"
#>  .. .. $ PGM_TRMNTN_CD                 : chr "PGM_TRMNTN_CD"
#>  .. .. $ TRMNTN_EXPRTN_DT              : chr "TRMNTN_EXPRTN_DT"
#>  .. .. $ CRTFCTN_ACTN_TYPE_CD          : chr "CRTFCTN_ACTN_TYPE_CD"
#>  .. .. $ GNRL_CNTL_TYPE_CD             : chr "GNRL_CNTL_TYPE_CD"
#>  .. .. $ ZIP_CD                        : chr "ZIP_CD"
#>  .. .. $ FIPS_STATE_CD                 : chr "FIPS_STATE_CD"
#>  .. .. $ FIPS_CNTY_CD                  : chr "FIPS_CNTY_CD"
#>  .. .. $ CBSA_URBN_RRL_IND             : chr "CBSA_URBN_RRL_IND"
#>  .. .. $ CBSA_CD                       : chr "CBSA_CD"
#>  .. .. $ ACRDTN_EFCTV_DT               : chr "ACRDTN_EFCTV_DT"
#>  .. .. $ ACRDTN_EXPRTN_DT              : chr "ACRDTN_EXPRTN_DT"
#>  .. .. $ ACRDTN_TYPE_CD                : chr "ACRDTN_TYPE_CD"
#>  .. .. $ TOT_AFLTD_AMBLNC_SRVC_CNT     : chr "TOT_AFLTD_AMBLNC_SRVC_CNT"
#>  .. .. $ TOT_AFLTD_ASC_CNT             : chr "TOT_AFLTD_ASC_CNT"
#>  .. .. $ TOT_COLCTD_HOSP_CNT           : chr "TOT_COLCTD_HOSP_CNT"
#>  .. .. $ TOT_AFLTD_ESRD_CNT            : chr "TOT_AFLTD_ESRD_CNT"
#>  .. .. $ TOT_AFLTD_FQHC_CNT            : chr "TOT_AFLTD_FQHC_CNT"
#>  .. .. $ TOT_AFLTD_HHA_CNT             : chr "TOT_AFLTD_HHA_CNT"
#>  .. .. $ TOT_AFLTD_HOSPC_CNT           : chr "TOT_AFLTD_HOSPC_CNT"
#>  .. .. $ TOT_AFLTD_OPO_CNT             : chr "TOT_AFLTD_OPO_CNT"
#>  .. .. $ TOT_AFLTD_PRTF_CNT            : chr "TOT_AFLTD_PRTF_CNT"
#>  .. .. $ TOT_AFLTD_RHC_CNT             : chr "TOT_AFLTD_RHC_CNT"
#>  .. .. $ TOT_AFLTD_SNF_CNT             : chr "TOT_AFLTD_SNF_CNT"
#>  .. .. $ AFLTD_PRVDR_CNT               : chr "AFLTD_PRVDR_CNT"
#>  .. .. $ RSDNT_PGM_ALPTHC_SW           : chr "RSDNT_PGM_ALPTHC_SW"
#>  .. .. $ RSDNT_PGM_DNTL_SW             : chr "RSDNT_PGM_DNTL_SW"
#>  .. .. $ RSDNT_PGM_OSTPTHC_SW          : chr "RSDNT_PGM_OSTPTHC_SW"
#>  .. .. $ RSDNT_PGM_OTHR_SW             : chr "RSDNT_PGM_OTHR_SW"
#>  .. .. $ RSDNT_PGM_PDTRC_SW            : chr "RSDNT_PGM_PDTRC_SW"
#>  .. .. $ LAB_SRVC_CD                   : chr "LAB_SRVC_CD"
#>  .. .. $ PHRMCY_SRVC_CD                : chr "PHRMCY_SRVC_CD"
#>  .. .. $ RDLGY_SRVC_CD                 : chr "RDLGY_SRVC_CD"
#>  .. .. $ ASC_BGN_SRVC_DT               : chr "ASC_BGN_SRVC_DT"
#>  .. .. $ FREESTNDNG_ASC_SW             : chr "FREESTNDNG_ASC_SW"
#>  .. .. $ OVRRD_BED_CNT_SW              : chr "OVRRD_BED_CNT_SW"
#>  .. .. $ CRTFD_BED_CNT                 : chr "CRTFD_BED_CNT"
#>  .. .. $ ICFIID_BED_CNT                : chr "ICFIID_BED_CNT"
#>  .. .. $ MDCD_NF_BED_CNT               : chr "MDCD_NF_BED_CNT"
#>  .. .. $ MDCR_SNF_BED_CNT              : chr "MDCR_SNF_BED_CNT"
#>  .. .. $ MDCR_MDCD_SNF_BED_CNT         : chr "MDCR_MDCD_SNF_BED_CNT"
#>  .. .. $ AIDS_BED_CNT                  : chr "AIDS_BED_CNT"
#>  .. .. $ ALZHMR_BED_CNT                : chr "ALZHMR_BED_CNT"
#>  .. .. $ DLYS_BED_CNT                  : chr "DLYS_BED_CNT"
#>  .. .. $ DSBL_CHLDRN_BED_CNT           : chr "DSBL_CHLDRN_BED_CNT"
#>  .. .. $ HEAD_TRMA_BED_CNT             : chr "HEAD_TRMA_BED_CNT"
#>  .. .. $ HOSPC_BED_CNT                 : chr "HOSPC_BED_CNT"
#>  .. .. $ HNTGTN_DEASE_BED_CNT          : chr "HNTGTN_DEASE_BED_CNT"
#>  .. .. $ REHAB_BED_CNT                 : chr "REHAB_BED_CNT"
#>  .. .. $ VNTLTR_BED_CNT                : chr "VNTLTR_BED_CNT"
#>  .. .. $ BED_CNT                       : chr "BED_CNT"
#>  .. .. $ BRNCH_CNT                     : chr "BRNCH_CNT"
#>  .. .. $ BRNCH_OPRTN_SW                : chr "BRNCH_OPRTN_SW"
#>  .. .. $ CAH_PSYCH_DPU_SW              : chr "CAH_PSYCH_DPU_SW"
#>  .. .. $ CAH_REHAB_DPU_SW              : chr "CAH_REHAB_DPU_SW"
#>  .. .. $ CAH_SB_SW                     : chr "CAH_SB_SW"
#>  .. .. $ CRDC_CTHRTZTN_PRCDR_ROOMS_CNT : chr "CRDC_CTHRTZTN_PRCDR_ROOMS_CNT"
#>  .. .. $ GNRL_FAC_TYPE_CD              : chr "GNRL_FAC_TYPE_CD"
#>  .. .. $ CHOW_SW                       : chr "CHOW_SW"
#>  .. .. $ CLIA_ID_NUMBER_1              : chr "CLIA_ID_NUMBER_1"
#>  .. .. $ CLIA_ID_NUMBER_2              : chr "CLIA_ID_NUMBER_2"
#>  .. .. $ CLIA_ID_NUMBER_3              : chr "CLIA_ID_NUMBER_3"
#>  .. .. $ CLIA_ID_NUMBER_4              : chr "CLIA_ID_NUMBER_4"
#>  .. .. $ CLIA_ID_NUMBER_5              : chr "CLIA_ID_NUMBER_5"
#>  .. .. $ COLCTN_STUS_SW                : chr "COLCTN_STUS_SW"
#>  .. .. $ RN_24_HR_WVR_SW               : chr "RN_24_HR_WVR_SW"
#>  .. .. $ RN_7_DAY_WVR_SW               : chr "RN_7_DAY_WVR_SW"
#>  .. .. $ BED_PER_ROOM_WVR_SW           : chr "BED_PER_ROOM_WVR_SW"
#>  .. .. $ LSC_WVR_SW                    : chr "LSC_WVR_SW"
#>  .. .. $ ROOM_SIZE_WVR_SW              : chr "ROOM_SIZE_WVR_SW"
#>  .. .. $ ENDSCPY_PRCDR_ROOMS_CNT       : chr "ENDSCPY_PRCDR_ROOMS_CNT"
#>  .. .. $ ESRD_NTWRK_NUM                : chr "ESRD_NTWRK_NUM"
#>  .. .. $ EXPRMT_RSRCH_CNDCTD_SW        : chr "EXPRMT_RSRCH_CNDCTD_SW"
#>  .. .. $ FAX_PHNE_NUM                  : chr "FAX_PHNE_NUM"
#>  .. .. $ FY_END_MO_DAY_CD              : chr "FY_END_MO_DAY_CD"
#>  .. ..  [list output truncated]
#>  @ metadata  : <care_metadata>
#>  .. @ description: chr "Please be advised that as of Q4 2023 there is a new Provider of Service file (POS) that contains the provider a"| __truncated__
#>  .. @ temporal   : chr "2011-10-01 • 2025-03-31"
#>  .. @ periodicity: chr "Quarterly [R/P3M]"
#>  .. @ modified   : Date[1:1], format: "2025-04-25"
#>  .. @ dictionary : chr "https://data.cms.gov/sites/default/files/2023-07/0ca58d5d-7914-4532-b22d-41741d3e6151/P.QWB.POSQ.OTHER.LAYOUT.MAR23.pdf"
#>  .. @ site       : chr "https://data.cms.gov/provider-characteristics/hospitals-and-other-facilities/provider-of-services-file-hospital"| __truncated__
#>  .. @ references : chr "https://data.cms.gov/resources/provider-of-services-file-methodology"
#>  .. @ download   : chr "https://data.cms.gov/sites/default/files/2025-04/52ac594d-c42d-4377-9c03-309d6a783dc4/PQWB.POSQ.CLIA.DATA.MAR25.csv"
#>  @ resources : chr "https://data.cms.gov/data-api/v1/dataset-resources/8ba0f9b4-9493-4aa0-9f82-44ea9468d1b5"
care_endpoint("IQIES")
#> <care_endpoint>
#>  @ title     : chr "Provider of Services File - Internet Quality Improvement and Evaluation System - Home Health Agency, Ambulatory"| __truncated__
#>  @ identifier: chr "https://data.cms.gov/data-api/v1/dataset/086e48c4-87a6-4be1-8823-29e8da8f225b/data-viewer"
#>  @ dimensions: <care_dimensions>
#>  .. @ limit : int 5000
#>  .. @ rows  : int 39912
#>  .. @ pages : int 8
#>  .. @ fields:List of 103
#>  .. .. $ prvdr_num                    : chr "prvdr_num"
#>  .. .. $ fac_name                     : chr "fac_name"
#>  .. .. $ prvdr_sbtyp_id               : chr "prvdr_sbtyp_id"
#>  .. .. $ mdcd_vndr_num                : chr "mdcd_vndr_num"
#>  .. .. $ orgnl_prtcptn_dt             : chr "orgnl_prtcptn_dt"
#>  .. .. $ prvdr_type_id                : chr "prvdr_type_id"
#>  .. .. $ st_adr                       : chr "st_adr"
#>  .. .. $ city_name                    : chr "city_name"
#>  .. .. $ ssa_cnty_cd                  : chr "ssa_cnty_cd"
#>  .. .. $ zip_cd                       : chr "zip_cd"
#>  .. .. $ phne_num                     : chr "phne_num"
#>  .. .. $ crtfctn_dt                   : chr "crtfctn_dt"
#>  .. .. $ trmntn_exprtn_dt             : chr "trmntn_exprtn_dt"
#>  .. .. $ state_cd                     : chr "state_cd"
#>  .. .. $ ssa_state_cd                 : chr "ssa_state_cd"
#>  .. .. $ state_rgn_cd                 : chr "state_rgn_cd"
#>  .. .. $ rgn_cd                       : chr "rgn_cd"
#>  .. .. $ fips_state_cd                : chr "fips_state_cd"
#>  .. .. $ fips_cnty_cd                 : chr "fips_cnty_cd"
#>  .. .. $ cbsa_cd                      : chr "cbsa_cd"
#>  .. .. $ cbsa_urbn_rrl_ind            : chr "cbsa_urbn_rrl_ind"
#>  .. .. $ acrdtn_type_cd               : chr "acrdtn_type_cd"
#>  .. .. $ intrmdry_carr_cd             : chr "intrmdry_carr_cd"
#>  .. .. $ acptbl_poc_sw                : chr "acptbl_poc_sw"
#>  .. .. $ fy_end_mo_day_cd             : chr "fy_end_mo_day_cd"
#>  .. .. $ cmplnc_stus_cd               : chr "cmplnc_stus_cd"
#>  .. .. $ crtfctn_actn_type_cd         : chr "crtfctn_actn_type_cd"
#>  .. .. $ chow_dt                      : chr "chow_dt"
#>  .. .. $ phrmcy_srvc_cd               : chr "phrmcy_srvc_cd"
#>  .. .. $ hh_aide_srvc_cd              : chr "hh_aide_srvc_cd"
#>  .. .. $ lab_srvc_cd                  : chr "lab_srvc_cd"
#>  .. .. $ lpn_lvn_cnt                  : chr "lpn_lvn_cnt"
#>  .. .. $ mdcl_scl_srvc_cd             : chr "mdcl_scl_srvc_cd"
#>  .. .. $ othr_srvc_cd                 : chr "othr_srvc_cd"
#>  .. .. $ ot_srvc_cd                   : chr "ot_srvc_cd"
#>  .. .. $ nrsng_srvc_cd                : chr "nrsng_srvc_cd"
#>  .. .. $ prsnel_othr_cnt              : chr "prsnel_othr_cnt"
#>  .. .. $ pt_srvc_cd                   : chr "pt_srvc_cd"
#>  .. .. $ rn_cnt                       : chr "rn_cnt"
#>  .. .. $ asc_bgn_srvc_dt              : chr "asc_bgn_srvc_dt"
#>  .. .. $ dntl_srgry_sw                : chr "dntl_srgry_sw"
#>  .. .. $ endscpy_srgry_sw             : chr "endscpy_srgry_sw"
#>  .. .. $ ft_srgry_sw                  : chr "ft_srgry_sw"
#>  .. .. $ ob_gyn_srgry_sw              : chr "ob_gyn_srgry_sw"
#>  .. .. $ oprtg_room_cnt               : chr "oprtg_room_cnt"
#>  .. .. $ opthmlgy_srgry_sw            : chr "opthmlgy_srgry_sw"
#>  .. .. $ orthpdc_srgry_sw             : chr "orthpdc_srgry_sw"
#>  .. .. $ othr_srgry_sw                : chr "othr_srgry_sw"
#>  .. .. $ otlrynglgy_srgry_sw          : chr "otlrynglgy_srgry_sw"
#>  .. .. $ pain_srgry_sw                : chr "pain_srgry_sw"
#>  .. .. $ plstc_srgry_sw               : chr "plstc_srgry_sw"
#>  .. .. $ rdlgy_srvc_cd                : chr "rdlgy_srvc_cd"
#>  .. .. $ aplnc_equip_srvc_cd          : chr "aplnc_equip_srvc_cd"
#>  .. .. $ brnch_cnt                    : chr "brnch_cnt"
#>  .. .. $ brnch_oprtn_sw               : chr "brnch_oprtn_sw"
#>  .. .. $ chow_sw                      : chr "chow_sw"
#>  .. .. $ dietn_cnt                    : chr "dietn_cnt"
#>  .. .. $ hh_aide_cnt                  : chr "hh_aide_cnt"
#>  .. .. $ hh_aide_trng_pgm_cd          : chr "hh_aide_trng_pgm_cd"
#>  .. .. $ intrn_rsdnt_srvc_cd          : chr "intrn_rsdnt_srvc_cd"
#>  .. .. $ mdcr_hospc_sw                : chr "mdcr_hospc_sw"
#>  .. .. $ medicare_hospice_provider_num: chr "medicare_hospice_provider_num"
#>  .. .. $ ntrtnl_gdnc_srvc_cd          : chr "ntrtnl_gdnc_srvc_cd"
#>  .. .. $ ocptnl_thrpst_cnt            : chr "ocptnl_thrpst_cnt"
#>  .. .. $ pgm_prtcptn_cd               : chr "pgm_prtcptn_cd"
#>  .. .. $ phys_thrpst_stf_cnt          : chr "phys_thrpst_stf_cnt"
#>  .. .. $ reg_phrmcst_cnt              : chr "reg_phrmcst_cnt"
#>  .. .. $ sbunit_cnt                   : chr "sbunit_cnt"
#>  .. .. $ sbunit_oprtn_sw              : chr "sbunit_oprtn_sw"
#>  .. .. $ sbunit_sw                    : chr "sbunit_sw"
#>  .. .. $ scl_workr_cnt                : chr "scl_workr_cnt"
#>  .. .. $ spch_pthlgst_audlgst_cnt     : chr "spch_pthlgst_audlgst_cnt"
#>  .. .. $ spch_thrpy_srvc_cd           : chr "spch_thrpy_srvc_cd"
#>  .. .. $ vctnl_gdnc_srvc_cd           : chr "vctnl_gdnc_srvc_cd"
#>  .. .. $ hospice_type                 : chr "hospice_type"
#>  .. .. $ acute_resp_care_cd           : chr "acute_resp_care_cd"
#>  .. .. $ cnslng_srvc_cd               : chr "cnslng_srvc_cd"
#>  .. .. $ cnslr_emplee_cnt             : chr "cnslr_emplee_cnt"
#>  .. .. $ cnslr_vlntr_cnt              : chr "cnslr_vlntr_cnt"
#>  .. .. $ emplee_cnt                   : chr "emplee_cnt"
#>  .. .. $ hh_aide_emplee_cnt           : chr "hh_aide_emplee_cnt"
#>  .. .. $ hh_aide_vlntr_cnt            : chr "hh_aide_vlntr_cnt"
#>  .. .. $ hmmkr_emplee_cnt             : chr "hmmkr_emplee_cnt"
#>  .. .. $ hmmkr_srvc_cd                : chr "hmmkr_srvc_cd"
#>  .. .. $ hmmkr_vlntr_cnt              : chr "hmmkr_vlntr_cnt"
#>  .. .. $ lpn_lvn_vlntr_cnt            : chr "lpn_lvn_vlntr_cnt"
#>  .. .. $ mdcl_scl_workr_cnt           : chr "mdcl_scl_workr_cnt"
#>  .. .. $ mdcl_scl_workr_vlntr_cnt     : chr "mdcl_scl_workr_vlntr_cnt"
#>  .. .. $ mdcl_suply_srvc_cd           : chr "mdcl_suply_srvc_cd"
#>  .. .. $ physn_cnt                    : chr "physn_cnt"
#>  .. .. $ physn_srvc_cd                : chr "physn_srvc_cd"
#>  .. .. $ physn_vlntr_cnt              : chr "physn_vlntr_cnt"
#>  .. .. $ rn_vlntr_cnt                 : chr "rn_vlntr_cnt"
#>  .. .. $ shrt_term_ip_srvc_cd         : chr "shrt_term_ip_srvc_cd"
#>  .. .. $ spch_pthlgy_srvc_cd          : chr "spch_pthlgy_srvc_cd"
#>  .. .. $ vlntr_cnt                    : chr "vlntr_cnt"
#>  .. .. $ vlntr_othr_cnt               : chr "vlntr_othr_cnt"
#>  .. .. $ fed_crtfctn_stus_id          : chr "fed_crtfctn_stus_id"
#>  .. .. $ fed_crtfctn_stus_name        : chr "fed_crtfctn_stus_name"
#>  .. ..  [list output truncated]
#>  @ metadata  : <care_metadata>
#>  .. @ description: chr "The Provider of Services File (POS) - Internet Quality Improvement and Evaluation System (iQIES) - Home Health "| __truncated__
#>  .. @ temporal   : chr "2023-10-01 • 2025-03-31"
#>  .. @ periodicity: chr "Quarterly [R/P3M]"
#>  .. @ modified   : Date[1:1], format: "2025-04-15"
#>  .. @ dictionary : chr "https://data.cms.gov/sites/default/files/2025-01/47cff7a0-5575-4a62-be85-ae72dcedf677/iQIES%20POS%20Data%20Dictionary.zip"
#>  .. @ site       : chr "https://data.cms.gov/provider-characteristics/hospitals-and-other-facilities/provider-of-services-file-internet"| __truncated__
#>  .. @ references : chr "https://data.cms.gov/resources/pos-file-iqies-for-hha-asc-and-hospice-providers-methodology"
#>  .. @ download   : chr "https://data.cms.gov/sites/default/files/2025-04/6aeb9473-03a7-4db9-9b71-057d60cbcada/iQIES%20HHA_HOSPICE_ASC_DATA.MAR25.csv"
#>  @ resources : chr "https://data.cms.gov/data-api/v1/dataset-resources/086e48c4-87a6-4be1-8823-29e8da8f225b"
care_endpoint("laboratories")
#> <care_endpoint>
#>  @ title     : chr "Provider of Services File - Clinical Laboratories"
#>  @ identifier: chr "https://data.cms.gov/data-api/v1/dataset/d3eb38ac-d8e9-40d3-b7b7-6205d3d1dc16/data-viewer"
#>  @ dimensions: <care_dimensions>
#>  .. @ limit : int 5000
#>  .. @ rows  : int 657403
#>  .. @ pages : int 132
#>  .. @ fields:List of 103
#>  .. .. $ PRVDR_CTGRY_SBTYP_CD         : chr "PRVDR_CTGRY_SBTYP_CD"
#>  .. .. $ PRVDR_CTGRY_CD               : chr "PRVDR_CTGRY_CD"
#>  .. .. $ CHOW_CNT                     : chr "CHOW_CNT"
#>  .. .. $ CHOW_DT                      : chr "CHOW_DT"
#>  .. .. $ CITY_NAME                    : chr "CITY_NAME"
#>  .. .. $ ACPTBL_POC_SW                : chr "ACPTBL_POC_SW"
#>  .. .. $ CMPLNC_STUS_CD               : chr "CMPLNC_STUS_CD"
#>  .. .. $ SSA_CNTY_CD                  : chr "SSA_CNTY_CD"
#>  .. .. $ CROSS_REF_PROVIDER_NUMBER    : chr "CROSS_REF_PROVIDER_NUMBER"
#>  .. .. $ CRTFCTN_DT                   : chr "CRTFCTN_DT"
#>  .. .. $ ELGBLTY_SW                   : chr "ELGBLTY_SW"
#>  .. .. $ FAC_NAME                     : chr "FAC_NAME"
#>  .. .. $ INTRMDRY_CARR_CD             : chr "INTRMDRY_CARR_CD"
#>  .. .. $ MDCD_VNDR_NUM                : chr "MDCD_VNDR_NUM"
#>  .. .. $ ORGNL_PRTCPTN_DT             : chr "ORGNL_PRTCPTN_DT"
#>  .. .. $ CHOW_PRIOR_DT                : chr "CHOW_PRIOR_DT"
#>  .. .. $ INTRMDRY_CARR_PRIOR_CD       : chr "INTRMDRY_CARR_PRIOR_CD"
#>  .. .. $ PRVDR_NUM                    : chr "PRVDR_NUM"
#>  .. .. $ RGN_CD                       : chr "RGN_CD"
#>  .. .. $ SKLTN_REC_SW                 : chr "SKLTN_REC_SW"
#>  .. .. $ STATE_CD                     : chr "STATE_CD"
#>  .. .. $ SSA_STATE_CD                 : chr "SSA_STATE_CD"
#>  .. .. $ STATE_RGN_CD                 : chr "STATE_RGN_CD"
#>  .. .. $ ST_ADR                       : chr "ST_ADR"
#>  .. .. $ PHNE_NUM                     : chr "PHNE_NUM"
#>  .. .. $ PGM_TRMNTN_CD                : chr "PGM_TRMNTN_CD"
#>  .. .. $ TRMNTN_EXPRTN_DT             : chr "TRMNTN_EXPRTN_DT"
#>  .. .. $ CRTFCTN_ACTN_TYPE_CD         : chr "CRTFCTN_ACTN_TYPE_CD"
#>  .. .. $ GNRL_CNTL_TYPE_CD            : chr "GNRL_CNTL_TYPE_CD"
#>  .. .. $ ZIP_CD                       : chr "ZIP_CD"
#>  .. .. $ FIPS_STATE_CD                : chr "FIPS_STATE_CD"
#>  .. .. $ FIPS_CNTY_CD                 : chr "FIPS_CNTY_CD"
#>  .. .. $ CBSA_URBN_RRL_IND            : chr "CBSA_URBN_RRL_IND"
#>  .. .. $ CBSA_CD                      : chr "CBSA_CD"
#>  .. .. $ ADDTNL_ST_ADR                : chr "ADDTNL_ST_ADR"
#>  .. .. $ AFFILIATED_PROVIDER_NUMBER_1 : chr "AFFILIATED_PROVIDER_NUMBER_1"
#>  .. .. $ AFFILIATED_PROVIDER_NUMBER_2 : chr "AFFILIATED_PROVIDER_NUMBER_2"
#>  .. .. $ AFFILIATED_PROVIDER_NUMBER_3 : chr "AFFILIATED_PROVIDER_NUMBER_3"
#>  .. .. $ AFFILIATED_PROVIDER_NUMBER_4 : chr "AFFILIATED_PROVIDER_NUMBER_4"
#>  .. .. $ AFFILIATED_PROVIDER_NUMBER_5 : chr "AFFILIATED_PROVIDER_NUMBER_5"
#>  .. .. $ AFFILIATED_PROVIDER_NUMBER_6 : chr "AFFILIATED_PROVIDER_NUMBER_6"
#>  .. .. $ AFFILIATED_PROVIDER_NUMBER_7 : chr "AFFILIATED_PROVIDER_NUMBER_7"
#>  .. .. $ AFFILIATED_PROVIDER_NUMBER_8 : chr "AFFILIATED_PROVIDER_NUMBER_8"
#>  .. .. $ A2LA_ACRDTD_CD               : chr "A2LA_ACRDTD_CD"
#>  .. .. $ A2LA_ACRDTD_Y_MATCH_DT       : chr "A2LA_ACRDTD_Y_MATCH_DT"
#>  .. .. $ A2LA_ACRDTD_Y_MATCH_SW       : chr "A2LA_ACRDTD_Y_MATCH_SW"
#>  .. .. $ AABB_ACRDTD_CD               : chr "AABB_ACRDTD_CD"
#>  .. .. $ AABB_ACRDTD_Y_MATCH_DT       : chr "AABB_ACRDTD_Y_MATCH_DT"
#>  .. .. $ AABB_ACRDTD_Y_MATCH_SW       : chr "AABB_ACRDTD_Y_MATCH_SW"
#>  .. .. $ AOA_ACRDTD_CD                : chr "AOA_ACRDTD_CD"
#>  .. .. $ AOA_ACRDTD_Y_MATCH_DT        : chr "AOA_ACRDTD_Y_MATCH_DT"
#>  .. .. $ AOA_ACRDTD_Y_MATCH_SW        : chr "AOA_ACRDTD_Y_MATCH_SW"
#>  .. .. $ ASHI_ACRDTD_CD               : chr "ASHI_ACRDTD_CD"
#>  .. .. $ ASHI_ACRDTD_Y_MATCH_DT       : chr "ASHI_ACRDTD_Y_MATCH_DT"
#>  .. .. $ ASHI_ACRDTD_Y_MATCH_SW       : chr "ASHI_ACRDTD_Y_MATCH_SW"
#>  .. .. $ CAP_ACRDTD_CD                : chr "CAP_ACRDTD_CD"
#>  .. .. $ CAP_ACRDTD_Y_MATCH_DT        : chr "CAP_ACRDTD_Y_MATCH_DT"
#>  .. .. $ CAP_ACRDTD_Y_MATCH_SW        : chr "CAP_ACRDTD_Y_MATCH_SW"
#>  .. .. $ COLA_ACRDTD_CD               : chr "COLA_ACRDTD_CD"
#>  .. .. $ COLA_ACRDTD_Y_MATCH_DT       : chr "COLA_ACRDTD_Y_MATCH_DT"
#>  .. .. $ COLA_ACRDTD_Y_MATCH_SW       : chr "COLA_ACRDTD_Y_MATCH_SW"
#>  .. .. $ JCAHO_ACRDTD_CD              : chr "JCAHO_ACRDTD_CD"
#>  .. .. $ JCAHO_ACRDTD_Y_MATCH_DT      : chr "JCAHO_ACRDTD_Y_MATCH_DT"
#>  .. .. $ JCAHO_ACRDTD_Y_MATCH_SW      : chr "JCAHO_ACRDTD_Y_MATCH_SW"
#>  .. .. $ APLCTN_RCVD_DT               : chr "APLCTN_RCVD_DT"
#>  .. .. $ APLCTN_TYPE_CD               : chr "APLCTN_TYPE_CD"
#>  .. .. $ GNRL_FAC_TYPE_CD             : chr "GNRL_FAC_TYPE_CD"
#>  .. .. $ CRTFCT_EFCTV_DT              : chr "CRTFCT_EFCTV_DT"
#>  .. .. $ CRTFCT_MAIL_DT               : chr "CRTFCT_MAIL_DT"
#>  .. .. $ CRTFCT_TYPE_CD               : chr "CRTFCT_TYPE_CD"
#>  .. .. $ CURRENT_CLIA_LAB_CLSFCTN_CD  : chr "CURRENT_CLIA_LAB_CLSFCTN_CD"
#>  .. .. $ CLIA_LAB_CLASSIFICATION_CD_1 : chr "CLIA_LAB_CLASSIFICATION_CD_1"
#>  .. .. $ CLIA_LAB_CLASSIFICATION_CD_10: chr "CLIA_LAB_CLASSIFICATION_CD_10"
#>  .. .. $ CLIA_LAB_CLASSIFICATION_CD_2 : chr "CLIA_LAB_CLASSIFICATION_CD_2"
#>  .. .. $ CLIA_LAB_CLASSIFICATION_CD_3 : chr "CLIA_LAB_CLASSIFICATION_CD_3"
#>  .. .. $ CLIA_LAB_CLASSIFICATION_CD_4 : chr "CLIA_LAB_CLASSIFICATION_CD_4"
#>  .. .. $ CLIA_LAB_CLASSIFICATION_CD_5 : chr "CLIA_LAB_CLASSIFICATION_CD_5"
#>  .. .. $ CLIA_LAB_CLASSIFICATION_CD_6 : chr "CLIA_LAB_CLASSIFICATION_CD_6"
#>  .. .. $ CLIA_LAB_CLASSIFICATION_CD_7 : chr "CLIA_LAB_CLASSIFICATION_CD_7"
#>  .. .. $ CLIA_LAB_CLASSIFICATION_CD_8 : chr "CLIA_LAB_CLASSIFICATION_CD_8"
#>  .. .. $ CLIA_LAB_CLASSIFICATION_CD_9 : chr "CLIA_LAB_CLASSIFICATION_CD_9"
#>  .. .. $ CLIA_MDCR_NUM                : chr "CLIA_MDCR_NUM"
#>  .. .. $ CLIA_TRMNTN_CD               : chr "CLIA_TRMNTN_CD"
#>  .. .. $ ACRDTN_SCHDL_CD              : chr "ACRDTN_SCHDL_CD"
#>  .. .. $ FORM_116_ACRDTD_TEST_VOL_CNT : chr "FORM_116_ACRDTD_TEST_VOL_CNT"
#>  .. .. $ FORM_116_TEST_VOL_CNT        : chr "FORM_116_TEST_VOL_CNT"
#>  .. .. $ DRCTLY_AFLTD_LAB_CNT         : chr "DRCTLY_AFLTD_LAB_CNT"
#>  .. .. $ FAX_PHNE_NUM                 : chr "FAX_PHNE_NUM"
#>  .. .. $ FY_END_MO_DAY_CD             : chr "FY_END_MO_DAY_CD"
#>  .. .. $ MLT_SITE_EXCPTN_SW           : chr "MLT_SITE_EXCPTN_SW"
#>  .. .. $ HOSP_LAB_EXCPTN_SW           : chr "HOSP_LAB_EXCPTN_SW"
#>  .. .. $ NON_PRFT_EXCPTN_SW           : chr "NON_PRFT_EXCPTN_SW"
#>  .. .. $ LAB_TEMP_TSTG_SITE_SW        : chr "LAB_TEMP_TSTG_SITE_SW"
#>  .. .. $ LAB_SITE_CNT                 : chr "LAB_SITE_CNT"
#>  .. .. $ PPMP_TEST_VOL_CNT            : chr "PPMP_TEST_VOL_CNT"
#>  .. .. $ ADDTNL_FAC_NAME              : chr "ADDTNL_FAC_NAME"
#>  .. .. $ RELATED_PROVIDER_NUMBER      : chr "RELATED_PROVIDER_NUMBER"
#>  .. .. $ SHR_LAB_SW                   : chr "SHR_LAB_SW"
#>  .. .. $ SHARED_LAB_XREF_NUMBER       : chr "SHARED_LAB_XREF_NUMBER"
#>  .. ..  [list output truncated]
#>  @ metadata  : <care_metadata>
#>  .. @ description: chr "The Provider of Services (POS) Clinical Laboratories (CLIA) data provides information on CLIA demographics and "| __truncated__
#>  .. @ temporal   : chr "2011-10-01 • 2025-03-31"
#>  .. @ periodicity: chr "Quarterly [R/P3M]"
#>  .. @ modified   : Date[1:1], format: "2025-04-15"
#>  .. @ dictionary : chr "https://data.cms.gov/sites/default/files/2023-07/50eacc07-7930-4e62-97e3-e04a2dc1e2a3/P.QWB.POSQ.CLIA.LAYOUT.MAR23.pdf"
#>  .. @ site       : chr "https://data.cms.gov/provider-characteristics/hospitals-and-other-facilities/provider-of-services-file-clinical-laboratories"
#>  .. @ references : chr "https://data.cms.gov/resources/provider-of-services-file-methodology"
#>  .. @ download   : chr "https://data.cms.gov/sites/default/files/2025-04/d0fc9f7a-b8d2-4988-a6e9-eb4cff2d45c3/PQWB.POSQ.CLIA.DATA.MAR25.csv"
#>  @ resources : chr "https://data.cms.gov/data-api/v1/dataset-resources/d3eb38ac-d8e9-40d3-b7b7-6205d3d1dc16"
care_endpoint("long_term")
#> <care_endpoint>
#>  @ title     : chr "Long-Term Care Facility Characteristics"
#>  @ identifier: chr "https://data.cms.gov/data-api/v1/dataset/129a6503-c0f1-4132-b186-4c0232c2d894/data-viewer"
#>  @ dimensions: <care_dimensions>
#>  .. @ limit : int 5000
#>  .. @ rows  : int 14799
#>  .. @ pages : int 3
#>  .. @ fields:List of 33
#>  .. .. $ Provider Number                                      : chr "Provider Number"
#>  .. .. $ Facility Name                                        : chr "Facility Name"
#>  .. .. $ City                                                 : chr "City"
#>  .. .. $ State                                                : chr "State"
#>  .. .. $ Zip Code                                             : chr "Zip Code"
#>  .. .. $ Certification Date                                   : chr "Certification Date"
#>  .. .. $ Medicare Census                                      : chr "Medicare Census"
#>  .. .. $ Medicaid Census                                      : chr "Medicaid Census"
#>  .. .. $ Other Census                                         : chr "Other Census"
#>  .. .. $ Total Residents                                      : chr "Total Residents"
#>  .. .. $ Program Participation Code                           : chr "Program Participation Code"
#>  .. .. $ Hospital Based                                       : chr "Hospital Based"
#>  .. .. $ Ownership Type                                       : chr "Ownership Type"
#>  .. .. $ Multi-Facility Organization                          : chr "Multi-Facility Organization"
#>  .. .. $ Multi-Facility Organization Name                     : chr "Multi-Facility Organization Name"
#>  .. .. $ Number of AIDS Beds                                  : chr "Number of AIDS Beds"
#>  .. .. $ Number of Alzheimer's Disease Beds                   : chr "Number of Alzheimer's Disease Beds"
#>  .. .. $ Number of Dialysis Beds                              : chr "Number of Dialysis Beds"
#>  .. .. $ Number of Disabled Children/Young Adult Beds         : chr "Number of Disabled Children/Young Adult Beds"
#>  .. .. $ Number of Head Trauma Beds                           : chr "Number of Head Trauma Beds"
#>  .. .. $ Number of Hospice Beds                               : chr "Number of Hospice Beds"
#>  .. .. $ Number of Huntington's Disease Beds                  : chr "Number of Huntington's Disease Beds"
#>  .. .. $ Number of Ventilator Beds                            : chr "Number of Ventilator Beds"
#>  .. .. $ Number of Other Specialized Rehab Beds               : chr "Number of Other Specialized Rehab Beds"
#>  .. .. $ Organized Residents' Group                           : chr "Organized Residents' Group"
#>  .. .. $ Organized Family Member Group                        : chr "Organized Family Member Group"
#>  .. .. $ Facility Conducts Experimental Research              : chr "Facility Conducts Experimental Research"
#>  .. .. $ Continuing Care Retirement Community (CCRC)          : chr "Continuing Care Retirement Community (CCRC)"
#>  .. .. $ 7 Day RN Waiver Date                                 : chr "7 Day RN Waiver Date"
#>  .. .. $ 7 Day RN Hrs Waived Per Week                         : chr "7 Day RN Hrs Waived Per Week"
#>  .. .. $ 24 Hr Licensed Nursing Waiver Date                   : chr "24 Hr Licensed Nursing Waiver Date"
#>  .. .. $ 24 Hr Licensed Nursing Hrs Waived Per Week           : chr "24 Hr Licensed Nursing Hrs Waived Per Week"
#>  .. .. $ Nurse Aide Training and Competency Evaluation Program: chr "Nurse Aide Training and Competency Evaluation Program"
#>  @ metadata  : <care_metadata>
#>  .. @ description: chr "The Long-Term Care Facility Characteristics, CMS Form 671, dataset provides information submitted by nursing ho"| __truncated__
#>  .. @ temporal   : chr "2023-10-01 • 2024-12-31"
#>  .. @ periodicity: chr "Quarterly [R/P3M]"
#>  .. @ modified   : Date[1:1], format: "2025-04-02"
#>  .. @ dictionary : chr "https://data.cms.gov/resources/long-term-care-facility-characteristics-data-dictionary"
#>  .. @ site       : chr "https://data.cms.gov/quality-of-care/long-term-care-facility-characteristics"
#>  .. @ references : chr "https://data.cms.gov/resources/long-term-care-facility-characteristics-methodology"
#>  .. @ download   : chr "https://data.cms.gov/sites/default/files/2025-01/d51e2acf-28d6-49cf-9be0-a306c1b17837/Form_671_Q4_2024.csv"
#>  @ resources : chr "https://data.cms.gov/data-api/v1/dataset-resources/129a6503-c0f1-4132-b186-4c0232c2d894"
care_endpoint("opt_out")
#> <care_endpoint>
#>  @ title     : chr "Opt Out Affidavits"
#>  @ identifier: chr "https://data.cms.gov/data-api/v1/dataset/9887a515-7552-4693-bf58-735c77af46d7/data-viewer"
#>  @ dimensions: <care_dimensions>
#>  .. @ limit : int 5000
#>  .. @ rows  : int 50473
#>  .. @ pages : int 11
#>  .. @ fields:List of 13
#>  .. .. $ First Name                 : chr "First Name"
#>  .. .. $ Last Name                  : chr "Last Name"
#>  .. .. $ NPI                        : chr "NPI"
#>  .. .. $ Specialty                  : chr "Specialty"
#>  .. .. $ Optout Effective Date      : chr "Optout Effective Date"
#>  .. .. $ Optout End Date            : chr "Optout End Date"
#>  .. .. $ First Line Street Address  : chr "First Line Street Address"
#>  .. .. $ Second Line Street Address : chr "Second Line Street Address"
#>  .. .. $ City Name                  : chr "City Name"
#>  .. .. $ State Code                 : chr "State Code"
#>  .. .. $ Zip code                   : chr "Zip code"
#>  .. .. $ Eligible to Order and Refer: chr "Eligible to Order and Refer"
#>  .. .. $ Last updated               : chr "Last updated"
#>  @ metadata  : <care_metadata>
#>  .. @ description: chr "ATTENTION USERSSome Providers Opt-Out Status may end early due to COVID 19 waivers. Please contact your respect"| __truncated__
#>  .. @ temporal   : chr "2023-04-01 • 2025-03-31"
#>  .. @ periodicity: chr "Monthly [R/P1M]"
#>  .. @ modified   : Date[1:1], format: "2025-04-24"
#>  .. @ dictionary : chr "https://data.cms.gov/resources/opt-out-affidavits-data-dictionary"
#>  .. @ site       : chr "https://data.cms.gov/provider-characteristics/medicare-provider-supplier-enrollment/opt-out-affidavits"
#>  .. @ references : chr "https://data.cms.gov/resources/opt-out-affidavits-methodology"
#>  .. @ download   : chr "https://data.cms.gov/sites/default/files/2025-04/fcfeb052-e800-4934-844d-0bd2e0b48937/Opt_Out_List_March.csv"
#>  @ resources : chr "https://data.cms.gov/data-api/v1/dataset-resources/9887a515-7552-4693-bf58-735c77af46d7"
care_endpoint("order_refer")
#> <care_endpoint>
#>  @ title     : chr "Order and Referring"
#>  @ identifier: chr "https://data.cms.gov/data-api/v1/dataset/c99b5865-1119-4436-bb80-c5af2773ea1f/data-viewer"
#>  @ dimensions: <care_dimensions>
#>  .. @ limit : int 5000
#>  .. @ rows  : int 1945216
#>  .. @ pages : int 390
#>  .. @ fields:List of 8
#>  .. .. $ NPI       : chr "NPI"
#>  .. .. $ LAST_NAME : chr "LAST_NAME"
#>  .. .. $ FIRST_NAME: chr "FIRST_NAME"
#>  .. .. $ PARTB     : chr "PARTB"
#>  .. .. $ DME       : chr "DME"
#>  .. .. $ HHA       : chr "HHA"
#>  .. .. $ PMD       : chr "PMD"
#>  .. .. $ HOSPICE   : chr "HOSPICE"
#>  @ metadata  : <care_metadata>
#>  .. @ description: chr "The Order and Referring dataset provides information on all physicians and non-physician practitioners, by thei"| __truncated__
#>  .. @ temporal   : chr "2023-05-21 • 2025-04-26"
#>  .. @ periodicity: chr "Twice a Week [R/P3.5D]"
#>  .. @ modified   : Date[1:1], format: "2025-04-29"
#>  .. @ dictionary : chr "https://data.cms.gov/resources/order-and-referring-data-dictionary"
#>  .. @ site       : chr "https://data.cms.gov/provider-characteristics/medicare-provider-supplier-enrollment/order-and-referring"
#>  .. @ references : chr "https://data.cms.gov/resources/order-and-referring-methodology"
#>  .. @ download   : chr "https://data.cms.gov/sites/default/files/2025-04/d7415ae9-61a0-4d57-9cc7-6b5781415def/OrderReferring_20250428.csv"
#>  @ resources : chr "https://data.cms.gov/data-api/v1/dataset-resources/c99b5865-1119-4436-bb80-c5af2773ea1f"
care_endpoint("RBCS")
#> <care_endpoint>
#>  @ title     : chr "Restructured BETOS Classification System"
#>  @ identifier: chr "https://data.cms.gov/data-api/v1/dataset/e3db6e56-149f-49ce-b374-40aecda2357b/data-viewer"
#>  @ dimensions: <care_dimensions>
#>  .. @ limit : int 5000
#>  .. @ rows  : int 18077
#>  .. @ pages : int 4
#>  .. @ fields:List of 14
#>  .. .. $ HCPCS_Cd                      : chr "HCPCS_Cd"
#>  .. .. $ RBCS_ID                       : chr "RBCS_ID"
#>  .. .. $ RBCS_Cat                      : chr "RBCS_Cat"
#>  .. .. $ RBCS_Cat_Desc                 : chr "RBCS_Cat_Desc"
#>  .. .. $ RBCS_Cat_Subcat               : chr "RBCS_Cat_Subcat"
#>  .. .. $ RBCS_SubCat_Desc              : chr "RBCS_SubCat_Desc"
#>  .. .. $ RBCS_FamNumb                  : chr "RBCS_FamNumb"
#>  .. .. $ RBCS_Family_Desc              : chr "RBCS_Family_Desc"
#>  .. .. $ RBCS_Major_Ind                : chr "RBCS_Major_Ind"
#>  .. .. $ HCPCS_CD_ADD_DT               : chr "HCPCS_CD_ADD_DT"
#>  .. .. $ HCPCS_CD_END_DT               : chr "HCPCS_CD_END_DT"
#>  .. .. $ RBCS_Release_Year             : chr "RBCS_Release_Year"
#>  .. .. $ RBCS_Reassigned               : chr "RBCS_Reassigned"
#>  .. .. $ Final_CY_Prior_to_Reassignment: chr "Final_CY_Prior_to_Reassignment"
#>  @ metadata  : <care_metadata>
#>  .. @ description: chr "The Restructured BETOS Classification System (RBCS) dataset is a taxonomy that allows researchers to group heal"| __truncated__
#>  .. @ temporal   : chr "2024-01-01 • 2024-12-31"
#>  .. @ periodicity: chr "Annually [R/P1Y]"
#>  .. @ modified   : Date[1:1], format: "2024-10-22"
#>  .. @ dictionary : chr "https://data.cms.gov/resources/rbcs-data-dictionary-2024"
#>  .. @ site       : chr "https://data.cms.gov/provider-summary-by-type-of-service/provider-service-classifications/restructured-betos-cl"| __truncated__
#>  .. @ references : chr "https://data.cms.gov/sites/default/files/2024-10/2024%20RBCS%20Final%20Report.pdf"
#>  .. @ download   : chr "https://data.cms.gov/sites/default/files/2024-10/30b74146-c5ce-47e0-8ae2-ad1204c81906/2024%20RBCS%20Taxonomy_CSV.csv"
#>  @ resources : chr "https://data.cms.gov/data-api/v1/dataset-resources/e3db6e56-149f-49ce-b374-40aecda2357b"
care_endpoint("transparency")
#> <care_endpoint>
#>  @ title     : chr "Hospital Price Transparency Enforcement Activities and Outcomes"
#>  @ identifier: chr "https://data.cms.gov/data-api/v1/dataset/6a3aa708-3c9d-411a-a1a4-e046d3ade7ef/data-viewer"
#>  @ dimensions: <care_dimensions>
#>  .. @ limit : int 5000
#>  .. @ rows  : int 5290
#>  .. @ pages : int 2
#>  .. @ fields:List of 7
#>  .. .. $ Case_ID       : chr "Case_ID"
#>  .. .. $ Hosp_Name     : chr "Hosp_Name"
#>  .. .. $ Hosp_Address  : chr "Hosp_Address"
#>  .. .. $ City          : chr "City"
#>  .. .. $ State         : chr "State"
#>  .. .. $ Action        : chr "Action"
#>  .. .. $ Date_of_Action: chr "Date_of_Action"
#>  @ metadata  : <care_metadata>
#>  .. @ description: chr "The Hospital Price Transparency Enforcement Activities and Outcomes dataset contains information related to enf"| __truncated__
#>  .. @ temporal   : chr "2023-10-01 • 2024-12-31"
#>  .. @ periodicity: chr "Quarterly [R/P3M]"
#>  .. @ modified   : Date[1:1], format: "2025-02-06"
#>  .. @ dictionary : chr "https://data.cms.gov/resources/hospital-price-transparency-enforcement-activities-and-outcomes-data-dictionary"
#>  .. @ site       : chr "https://data.cms.gov/provider-characteristics/hospitals-and-other-facilities/hospital-price-transparency-enforc"| __truncated__
#>  .. @ references : chr "https://data.cms.gov/resources/hospital-price-transparency-enforcement-activities-and-outcomes-methodology"
#>  .. @ download   : chr "https://data.cms.gov/sites/default/files/2025-01/3febdccd-aae4-4d5b-be37-7fdcb10b67ea/Hospital_Price_Transparen"| __truncated__
#>  @ resources : chr "https://data.cms.gov/data-api/v1/dataset-resources/6a3aa708-3c9d-411a-a1a4-e046d3ade7ef"