API Catalogs

List of API catalogs:

care: CMS Medicare API

prov: CMS Provider API

open: CMS Open Payments API

caid: CMS Medicaid API

hgov: CMS HealthCare.gov API

Usage

catalogs()

Value

<list> of catalogs

Examples

catalogs()
#> ! The following functions will be optimised package-wide:
#> `sum`
#> `prod`
#> `mean`
#> `median`
#> `min`
#> `max`
#> `sd`
#> `var`
#> `dplyr::n`
#> `dplyr::first`
#> `dplyr::last`
#> `dplyr::n_distinct`
#> `dplyr::row_number`
#> `dplyr::lag`
#> `dplyr::lead`
#> `dplyr::cur_group`
#> `dplyr::cur_group_id`
#> `dplyr::cur_group_rows`
#> 
#> Optimised expressions are independent from each other and typical data-masking
#> rules may not apply
#> 
#> Run `options(fastplyr.optimise = FALSE)` to disable optimisations globally
#> 
#> Run `options(fastplyr.inform = FALSE)` to disable this message
#> This message is displayed once per session.
#> $care
#> $care$current
#> # A tibble: 147 × 12
#>    title        description identifier references modified   periodicity contact
#>  * <chr>        <chr>       <chr>      <chr>      <date>     <chr>       <chr>  
#>  1 ACO REACH A… The Accoun… https://d… https://d… 2025-05-21 Annually [… ACO RE…
#>  2 ACO REACH E… Accountabl… https://d… https://d… 2025-05-21 Annually [… ACO RE…
#>  3 ACO REACH F… The Accoun… https://d… https://d… 2025-09-10 Annually [… ACO RE…
#>  4 ACO REACH P… The Accoun… https://d… https://d… 2025-05-21 Annually [… ACO RE…
#>  5 Accountable… The Accoun… https://d… https://d… 2025-09-10 Annually [… Shared…
#>  6 Accountable… The Accoun… https://d… https://d… 2025-09-10 Annually [… Shared…
#>  7 Accountable… TheAccount… https://d… https://d… 2025-09-10 Annually [… Shared…
#>  8 Advance Inv… The Advanc… https://d… https://d… 2025-03-18 Annually [… Shared…
#>  9 Agency for … The Agency… https://d… https://d… 2020-12-08 Annually [… Defici…
#> 10 CMS Program… The CMS Pr… https://d… https://d… 2025-09-08 Annually [… CMS Pr…
#> # ℹ 137 more rows
#> # ℹ 5 more variables: temporal <chr>, dictionary <chr>, site <chr>,
#> #   download <chr>, resources <chr>
#> 
#> $care$temporal
#> # A tibble: 119 × 8
#>    title   description periodicity contact dictionary site  references endpoints
#>    <chr>   <chr>       <chr>       <chr>   <chr>      <chr> <chr>      <list<ti>
#>  1 ACO RE… The Accoun… Annually [… ACO RE… https://d… http… https://d…   [3 × 6]
#>  2 ACO RE… Accountabl… Annually [… ACO RE… https://d… http… https://d…   [3 × 6]
#>  3 ACO RE… The Accoun… Annually [… ACO RE… https://d… http… https://d…   [3 × 6]
#>  4 ACO RE… The Accoun… Annually [… ACO RE… https://d… http… https://d…   [3 × 6]
#>  5 Accoun… The Accoun… Annually [… Shared… https://d… http… https://d…  [13 × 6]
#>  6 Accoun… The Accoun… Annually [… Shared… https://d… http… https://d…  [10 × 6]
#>  7 Accoun… TheAccount… Annually [… Shared… https://d… http… https://d…  [13 × 6]
#>  8 Advanc… The Advanc… Annually [… Shared… https://d… http… https://d…   [2 × 6]
#>  9 Agency… The Agency… Annually [… Defici… https://d… http… https://d…   [2 × 6]
#> 10 CPC In… The Compre… Annually [… CMS In… https://d… http… https://d…   [1 × 6]
#> # ℹ 109 more rows
#> 
#> $care$download_only
#> # A tibble: 27 × 8
#>    title   description periodicity contact dictionary site  references endpoints
#>    <chr>   <chr>       <chr>       <chr>   <chr>      <chr> <chr>      <list<ti>
#>  1 CMS Pr… The CMS Pr… Annually [… CMS Pr… https://d… http… https://d…   [9 × 6]
#>  2 CMS Pr… The CMS Pr… Annually [… CMS Pr… https://d… http… https://d…   [6 × 6]
#>  3 CMS Pr… The CMS Pr… Annually [… CMS Pr… https://d… http… https://d…   [6 × 6]
#>  4 CMS Pr… The CMS Pr… Annually [… CMS Pr… https://d… http… https://d…   [6 × 6]
#>  5 CMS Pr… The CMS Pr… Annually [… CMS Pr… https://d… http… https://d…   [5 × 6]
#>  6 CMS Pr… The CMS Pr… Annually [… CMS Pr… https://d… http… https://d…   [9 × 6]
#>  7 CMS Pr… The Medica… Annually [… CMS Pr… https://d… http… https://d…   [9 × 6]
#>  8 CMS Pr… The CMS Pr… Annually [… CMS Pr… https://d… http… https://d…   [9 × 6]
#>  9 CMS Pr… The CMS Pr… Annually [… CMS Pr… https://d… http… https://d…   [9 × 6]
#> 10 CMS Pr… The CMS Pr… Annually [… CMS Pr… https://d… http… https://d…   [9 × 6]
#> # ℹ 17 more rows
#> 
#> 
#> $prov
#> $prov$current
#> # A tibble: 147 × 12
#>    title          group description issued     modified   released   next_update
#>    <chr>          <chr> <chr>       <date>     <date>     <date>     <date>     
#>  1 Dialysis Faci… Dial… A list of … 2020-03-14 2025-06-17 2025-07-16 2025-10-22 
#>  2 Dialysis Faci… Dial… National a… 2020-03-14 2025-06-17 2025-07-16 2025-10-22 
#>  3 Dialysis Faci… Dial… State aver… 2020-03-14 2025-06-17 2025-07-16 2025-10-22 
#>  4 ESRD QIP - Cl… Dial… This datas… 2021-01-20 2024-12-03 2025-01-15 NA         
#>  5 ESRD QIP - Co… Dial… This datas… 2021-01-20 2024-12-03 2025-01-15 NA         
#>  6 ESRD QIP - Co… Dial… This datas… 2025-01-15 2024-12-03 2025-01-15 NA         
#>  7 ESRD QIP - Di… Dial… This datas… 2021-01-20 2024-12-03 2025-01-15 NA         
#>  8 ESRD QIP - Fo… Dial… This table… 2021-01-20 2024-12-03 2025-01-15 NA         
#>  9 ESRD QIP - Hy… Dial… This datas… 2021-01-20 2024-12-03 2025-01-15 NA         
#> 10 ESRD QIP - In… Dial… This datas… 2021-01-20 2024-12-03 2025-01-15 NA         
#> # ℹ 137 more rows
#> # ℹ 5 more variables: identifier <chr>, contact <chr>, download <chr>,
#> #   site <chr>, dictionary <chr>
#> 
#> 
#> $open
#> $open$current
#> # A tibble: 18 × 6
#>    title                      description modified   identifier contact download
#>    <chr>                      <chr>       <date>     <chr>      <chr>   <chr>   
#>  1 Covered Recipient Profile… The Covere… 2025-06-30 https://o… Open P… https:/…
#>  2 National level payment to… This datas… 2025-06-30 https://o… Open P… https:/…
#>  3 National level payment to… This datas… 2025-06-30 https://o… Open P… https:/…
#>  4 Payments grouped by Teach… This datas… 2025-06-30 https://o… Open P… https:/…
#>  5 Payments grouped by cover… This datas… 2025-06-30 https://o… Open P… https:/…
#>  6 Payments grouped by cover… This datas… 2025-06-30 https://o… Open P… https:/…
#>  7 Payments grouped by physi… This datas… 2025-06-30 https://o… Open P… https:/…
#>  8 Payments grouped by repor… This datas… 2025-06-30 https://o… Open P… https:/…
#>  9 Payments grouped by repor… This datas… 2025-06-30 https://o… Open P… https:/…
#> 10 Payments grouped by repor… This datas… 2025-06-30 https://o… Open P… https:/…
#> 11 Physician (distinct) prof… This datas… 2025-06-30 https://o… Open P… https:/…
#> 12 Profile Information        This datas… 2025-06-30 https://o… Open P… https:/…
#> 13 Provider profile ID mappi… Maps secon… 2025-06-30 https://o… Open P… https:/…
#> 14 Reporting entity profile … This datas… 2025-06-30 https://o… Open P… https:/…
#> 15 State level payment total… This datas… 2025-06-30 https://o… Open P… https:/…
#> 16 State payment totals and … Groups pay… 2025-06-30 https://o… Open P… https:/…
#> 17 Summary Dashboard - all y… Summary da… 2025-06-30 https://o… Open P… https:/…
#> 18 Teaching Hospital profile… This datas… 2025-06-30 https://o… Open P… https:/…
#> 
#> $open$temporal
#> # A tibble: 8 × 2
#>   title                                                                endpoints
#>   <chr>                                                                <list<ti>
#> 1 General Payment Data                                                   [7 × 6]
#> 2 Ownership Payment Data                                                 [7 × 6]
#> 3 Research Payment Data                                                  [7 × 6]
#> 4 payments grouped by covered recipient and nature of payments           [7 × 6]
#> 5 payments grouped by covered recipient and reporting entities           [7 × 6]
#> 6 payments grouped by reporting entities and nature of payments          [7 × 6]
#> 7 payments grouped by reporting entities, covered recipient, and natu…   [7 × 6]
#> 8 state payment totals grouped by Nature of Payment for all years        [7 × 6]
#> 
#> 
#> $caid
#> $caid$current
#> # A tibble: 85 × 7
#>    title          identifier description periodicity modified   contact download
#>    <chr>          <chr>      <chr>       <chr>       <date>     <chr>   <chr>   
#>  1 ACA Federal U… https://d… Affordable… Decenniall… 2025-08-28 Medica… https:/…
#>  2 Acute Care Se… https://d… This data … Decenniall… 2024-01-05 Medica… https:/…
#>  3 Behavioral He… https://d… This data … Decenniall… 2024-01-05 Medica… NA      
#>  4 Beneficiaries… https://d… This table… Decenniall… 2024-01-05 Medica… https:/…
#>  5 Beneficiaries… https://d… This table… Decenniall… 2024-01-05 Medica… https:/…
#>  6 Beneficiaries… https://d… This table… Decenniall… 2024-01-05 Medica… https:/…
#>  7 Beneficiaries… https://d… This table… Decenniall… 2024-01-05 Medica… https:/…
#>  8 Benefit Packa… https://d… This data … Decenniall… 2024-01-05 Medica… https:/…
#>  9 Benefit Packa… https://d… This data … Decenniall… 2024-01-05 Medica… https:/…
#> 10 Blood Lead Sc… https://d… This data … Decenniall… 2024-01-05 Medica… https:/…
#> # ℹ 75 more rows
#> 
#> $caid$temporal
#> # A tibble: 6 × 4
#>   title                                        description periodicity endpoints
#>   <chr>                                        <chr>       <chr>       <list<ti>
#> 1 Child and Adult Health Care Quality Measures Performanc… Annually […  [10 × 4]
#> 2 Managed Care Programs by State               Provides p… Decenniall…   [6 × 4]
#> 3 NADAC                                        National A… Decenniall…  [13 × 4]
#> 4 Pricing Comparison for Blood Disorder Treat… The Refere… Decenniall…   [8 × 4]
#> 5 Product Data for Newly Reported Drugs in th… The data b… Decenniall… [100 × 4]
#> 6 State Drug Utilization Data                  Drug utili… Decenniall…  [34 × 4]
#> 
#> 
#> $hgov
#> $hgov$current
#> # A tibble: 28 × 9
#>    title        identifier description periodicity issued     modified   contact
#>    <chr>        <chr>      <chr>       <chr>       <date>     <date>     <chr>  
#>  1 Agent Broke… https://d… NA          Decenniall… 2018-07-31 2025-09-19 FFMPro…
#>  2 AB Registra… https://d… Access Dat… Decenniall… 2016-03-30 2025-09-19 FFMPro…
#>  3 AB Suspensi… https://d… Access Dat… NA          2016-05-03 2025-09-19 FFMPro…
#>  4 Issuer_Part… https://d… The datase… Decenniall… 2020-06-05 2025-01-13 Domini…
#>  5 Rolling Dra… https://d… NA          Decenniall… 2016-08-22 2024-08-09 CMS (m…
#>  6 Auto-popula… https://d… ECP Autofi… Decenniall… 2016-08-22 2024-08-09 CMS (m…
#>  7 Issuer Part… https://d… NA          NA          2019-10-09 2020-09-29 Hiko N…
#>  8 NIPR Valid … https://d… This datas… Decenniall… 2016-11-17 2020-08-10 diaz_j…
#>  9 Direct Enro… https://d… NA          Decenniall… 2018-09-27 2024-08-07 Hiko N…
#> 10 Issuer and … https://d… NA          NA          2019-10-09 2019-11-22 Hiko N…
#> # ℹ 18 more rows
#> # ℹ 2 more variables: download <chr>, resources <chr>
#> 
#> $hgov$temporal
#> # A tibble: 16 × 2
#>    title                                            endpoints
#>    <chr>                                   <list<tibble[,9]>>
#>  1 Benefits and Cost Sharing PUF                     [12 × 9]
#>  2 Business Rules PUF                                [12 × 9]
#>  3 MLR Dataset                                        [9 × 9]
#>  4 Machine Readable PUF                               [4 × 9]
#>  5 Network PUF                                       [12 × 9]
#>  6 Plan Attributes PUF                               [12 × 9]
#>  7 Plan ID Crosswalk PUF                             [11 × 9]
#>  8 QHP Landscape Individual Market Dental            [17 × 9]
#>  9 QHP Landscape Individual Market Medical           [17 × 9]
#> 10 QHP Landscape SHOP Market Dental                  [16 × 9]
#> 11 QHP Landscape SHOP Market Medical                 [38 × 9]
#> 12 QHP Landscape SHOP Medical                         [2 × 9]
#> 13 Quality PUF                                        [8 × 9]
#> 14 Rate PUF                                          [12 × 9]
#> 15 Service Area PUF                                  [12 × 9]
#> 16 Transparency in Coverage PUF                       [5 × 9]
#> 
#>