Clinicians National Downloadable File

The Doctors and Clinicians national downloadable file is organized such that each line is unique at the clinician/enrollment record/group/address level. Clinicians with multiple Medicare enrollment records and/or single enrollments linking to multiple practice locations are listed on multiple lines.

Usage

clinicians(
  npi = NULL,
  pac = NULL,
  enid = NULL,
  last_name = NULL,
  first_name = NULL,
  middle_name = NULL,
  suffix = NULL,
  gender = NULL,
  credential = NULL,
  med_school = NULL,
  grad_year = NULL,
  spec_prim = NULL,
  spec_sec = NULL,
  telehealth = NULL,
  pac_org = NULL,
  org_n_members = NULL,
  facility_name = NULL,
  add1 = NULL,
  add2 = NULL,
  city = NULL,
  state = NULL,
  zip = NULL,
  phone = NULL,
  ind_assign = NULL,
  grp_assign = NULL,
  add_id = NULL
)

Arguments

npi

<chr> 10-digit Individual NPI

pac

<chr> 10-digit Individual PECOS Associate Control (PAC) ID

enid

<chr> 15-digit Medicare Enrollment ID

first_name, middle_name, last_name, suffix

<chr> Individual provider's name

gender

Individual provider's gender

credential

Individual provider's credentials

med_school

Medical school attended

grad_year

Year of graduation from medical school

spec_prim

Primary specialty

spec_sec

Secondary specialty

telehealth

Telehealth services

pac_org

<chr> 10-digit Organization PECOS Associate Control (PAC) ID

org_n_members

Number of members in the organization

facility_name

Facility name

add1

Address line 1

add2

Address line 2

city

City

state

State

zip

Zip code

phone

Phone number

ind_assign

Individual assignment

grp_assign

Group assignment

add_id

Address ID

Value

<tibble> of search results

Periodicity

Monthly (R/P1M)

Examples

clinicians(last_name = "CURRY")
#> INFO [2025-05-08 21:09:09] Loading endpoint...
#> INFO [2025-05-08 21:09:10] Querying National Downloadable File...
#> INFO [2025-05-08 21:09:10] Returned 465 results
#> # A tibble: 465 × 24
#>    npi        pac      enid  last_name first_name middle_name suffix_name gender
#>  * <chr>      <chr>    <chr> <chr>     <chr>      <chr>       <chr>       <chr> 
#>  1 1013162536 2264585… I200… CURRY     CHOLADDA   V.          NA          F     
#>  2 1023591120 8820342… I201… CURRY     EVAN       NA          NA          M     
#>  3 1023591120 8820342… I201… CURRY     EVAN       NA          NA          M     
#>  4 1033205539 8325104… I200… CURRY     ROBERT     EUGENE      NA          M     
#>  5 1033406533 4082858… I201… CURRY     ROBIN      G           NA          F     
#>  6 1033406533 4082858… I201… CURRY     ROBIN      G           NA          F     
#>  7 1033406533 4082858… I201… CURRY     ROBIN      G           NA          F     
#>  8 1033406533 4082858… I201… CURRY     ROBIN      G           NA          F     
#>  9 1033655212 8628334… I201… CURRY     SONDRA     NA          NA          F     
#> 10 1033764634 6204266… I202… CURRY     HALIE      NA          NA          F     
#> # ℹ 455 more rows
#> # ℹ 16 more variables: credential <chr>, med_school <chr>, grad_year <int>,
#> #   spec_prim <chr>, spec_sec <chr>, telehealth <int>, facility_name <chr>,
#> #   org_pac <chr>, org_members <int>, city <chr>, state <chr>, zip <chr>,
#> #   phone <chr>, ind_assgn <int>, grp_assgn <int>, address <chr>
clinicians(npi = "1043245657")
#> INFO [2025-05-08 21:09:10] Loading endpoint...
#> INFO [2025-05-08 21:09:10] Querying National Downloadable File...
#> INFO [2025-05-08 21:09:10] Returned 1 results
#> # A tibble: 1 × 24
#>   npi        pac       enid  last_name first_name middle_name suffix_name gender
#> * <chr>      <chr>     <chr> <chr>     <chr>      <chr>       <chr>       <chr> 
#> 1 1043245657 78108910… I200… FUNG      MARK       K           NA          M     
#> # ℹ 16 more variables: credential <chr>, med_school <chr>, grad_year <int>,
#> #   spec_prim <chr>, spec_sec <chr>, telehealth <int>, facility_name <chr>,
#> #   org_pac <chr>, org_members <int>, city <chr>, state <chr>, zip <chr>,
#> #   phone <chr>, ind_assgn <int>, grp_assgn <int>, address <chr>
clinicians(state = "GA")
#> INFO [2025-05-08 21:09:10] Loading endpoint...
#> INFO [2025-05-08 21:09:10] Querying National Downloadable File...
#> INFO [2025-05-08 21:09:10] Returned 2000 results
#> # A tibble: 2,000 × 24
#>    npi        pac      enid  last_name first_name middle_name suffix_name gender
#>  * <chr>      <chr>    <chr> <chr>     <chr>      <chr>       <chr>       <chr> 
#>  1 1003001843 5597718… I200… LUECHING… INGE       C           NA          F     
#>  2 1003002734 9234354… I202… MCCOLLUM  JONATHAN   P           NA          M     
#>  3 1003006115 9335223… I200… DURHAM    BENJAMIN   W           NA          M     
#>  4 1003006115 9335223… I200… DURHAM    BENJAMIN   W           NA          M     
#>  5 1003006115 9335223… I201… DURHAM    BENJAMIN   W           NA          M     
#>  6 1003008111 6406027… I201… EASTERLIN MARIE      NA          NA          F     
#>  7 1003008111 6406027… I201… EASTERLIN MARIE      NA          NA          F     
#>  8 1003009309 9638239… I200… BREWER    DEBRA      L           NA          F     
#>  9 1003012915 9133275… I200… JOHN      JAYASHREE  S           NA          F     
#> 10 1003012949 3476625… I200… PATEL RA… PINECCA    NA          NA          F     
#> # ℹ 1,990 more rows
#> # ℹ 16 more variables: credential <chr>, med_school <chr>, grad_year <int>,
#> #   spec_prim <chr>, spec_sec <chr>, telehealth <int>, facility_name <chr>,
#> #   org_pac <chr>, org_members <int>, city <chr>, state <chr>, zip <chr>,
#> #   phone <chr>, ind_assgn <int>, grp_assgn <int>, address <chr>