Load a collection of endpoints by alias

Load a collection of endpoints by alias

Usage

collection(alias)

Arguments

alias

<chr> collection alias

Value

An S7 <class_collection> object.

Examples

collection("unwind")
#> <class_collection>
#>  @ name   : chr "Medicaid Unwinding Report"
#>  @ members:List of 4
#>  .. $ wind_mark    : <class_caid>
#>  ..  ..@ access: <class_current>
#>  .. .. .. @ identifier: chr "https://data.medicaid.gov/api/1/datastore/query/9a83ba5e-05f5-47f5-82de-f3a59233a912/0?count=true&results=true&"| __truncated__
#>  .. .. .. @ metadata  :List of 5
#>  .. .. .. .. $ title      : chr "HealthCare.gov Marketplace Medicaid Unwinding Report"
#>  .. .. .. .. $ description: chr "Metrics from individual Marketplaces during the current reporting period. The report includes data for the stat"| __truncated__
#>  .. .. .. .. $ modified   : Date[1:1], format: "2024-08-30"
#>  .. .. .. .. $ periodicity: chr "Decennially [R/P10Y]"
#>  .. .. .. .. $ download   : chr NA
#>  .. .. .. @ dimensions: <class_dimensions>
#>  .. .. .. .. @ limit : int 8000
#>  .. .. .. .. @ rows  : int 204
#>  .. .. .. .. @ pages : int 1
#>  .. .. .. .. @ fields:List of 30
#>  .. .. .. .. .. $ state                           : chr NA
#>  .. .. .. .. .. $ metric                          : chr NA
#>  .. .. .. .. .. $ april_2023_net_count            : chr NA
#>  .. .. .. .. .. $ april_2023_net_percentage       : chr NA
#>  .. .. .. .. .. $ may_2023_net_count              : chr NA
#>  .. .. .. .. .. $ may_2023_net_percentage         : chr NA
#>  .. .. .. .. .. $ june_2023_net_count             : chr NA
#>  .. .. .. .. .. $ june_2023_net_percentage        : chr NA
#>  .. .. .. .. .. $ july_2023_net_count             : chr NA
#>  .. .. .. .. .. $ july_2023_net_percentage        : chr NA
#>  .. .. .. .. .. $ august_2023_net_count           : chr NA
#>  .. .. .. .. .. $ august_2023_net_percentage      : chr NA
#>  .. .. .. .. .. $ september_2023_net_count        : chr NA
#>  .. .. .. .. .. $ september_2023_net_percentage   : chr NA
#>  .. .. .. .. .. $ october_2023_net_count          : chr NA
#>  .. .. .. .. .. $ october_2023_net_percentage     : chr NA
#>  .. .. .. .. .. $ november_2023_net_count         : chr NA
#>  .. .. .. .. .. $ november_2023_net_percentage    : chr NA
#>  .. .. .. .. .. $ december_2023_net_count         : chr NA
#>  .. .. .. .. .. $ december_2023_net_percentage    : chr NA
#>  .. .. .. .. .. $ january_2024_net_count          : chr NA
#>  .. .. .. .. .. $ january_2024_net_percentage     : chr NA
#>  .. .. .. .. .. $ february_2024_net_count         : chr NA
#>  .. .. .. .. .. $ february_2024_net_percentage    : chr NA
#>  .. .. .. .. .. $ march_2024_net_count            : chr NA
#>  .. .. .. .. .. $ march_2024_net_percentage       : chr NA
#>  .. .. .. .. .. $ april_2024_net_count            : chr NA
#>  .. .. .. .. .. $ april_2024_net_percentage       : chr NA
#>  .. .. .. .. .. $ april_2024_cumulative_count     : chr NA
#>  .. .. .. .. .. $ april_cumulative_2024_percentage: chr NA
#>  .. $ wind_trans   : <class_caid>
#>  ..  ..@ access: <class_current>
#>  .. .. .. @ identifier: chr "https://data.medicaid.gov/api/1/datastore/query/5636a78c-fe18-4229-aee1-e40fa910a8a0/0?count=true&results=true&"| __truncated__
#>  .. .. .. @ metadata  :List of 5
#>  .. .. .. .. $ title      : chr "HealthCare.gov Transitions Marketplace Medicaid Unwinding Report"
#>  .. .. .. .. $ description: chr "Metrics from individual Marketplaces during the current reporting period. The report includes data for the stat"| __truncated__
#>  .. .. .. .. $ modified   : Date[1:1], format: "2024-12-27"
#>  .. .. .. .. $ periodicity: chr "Decennially [R/P10Y]"
#>  .. .. .. .. $ download   : chr NA
#>  .. .. .. @ dimensions: <class_dimensions>
#>  .. .. .. .. @ limit : int 8000
#>  .. .. .. .. @ rows  : int 59527
#>  .. .. .. .. @ pages : int 8
#>  .. .. .. .. @ fields:List of 9
#>  .. .. .. .. .. $ state                                                           : chr NA
#>  .. .. .. .. .. $ data_record_note                                                : chr NA
#>  .. .. .. .. .. $ metric                                                          : chr NA
#>  .. .. .. .. .. $ time_period                                                     : chr NA
#>  .. .. .. .. .. $ release_through                                                 : chr NA
#>  .. .. .. .. .. $ count_of_individuals_whose_medicaid_or_chip_coverage_was_te_dd11: chr NA
#>  .. .. .. .. .. $ percentage_of_individuals_whose_medicaid_or_chip_coverage_w_da6e: chr NA
#>  .. .. .. .. .. $ cumulative_count_of_individuals_whose_medicaid_or_chip_cove_28d9: chr NA
#>  .. .. .. .. .. $ cumulative_percentage_of_individuals_whose_medicaid_or_chip_6ac2: chr NA
#>  .. $ wind_historic: <class_caid>
#>  ..  ..@ access: <class_current>
#>  .. .. .. @ identifier: chr "https://data.medicaid.gov/api/1/datastore/query/d30cfc7c-4b32-4df1-b2bf-e0a850befd77/0?count=true&results=true&"| __truncated__
#>  .. .. .. @ metadata  :List of 5
#>  .. .. .. .. $ title      : chr "Separate CHIP Enrollment by Month and State Historic CAA/Unwinding Period"
#>  .. .. .. .. $ description: chr "<p>This historic dataset with total enrollment in separate CHIP programs by month and state was created to fulf"| __truncated__
#>  .. .. .. .. $ modified   : Date[1:1], format: "2024-12-27"
#>  .. .. .. .. $ periodicity: chr "Monthly [R/P1M]"
#>  .. .. .. .. $ download   : chr NA
#>  .. .. .. @ dimensions: <class_dimensions>
#>  .. .. .. .. @ limit : int 8000
#>  .. .. .. .. @ rows  : int 780
#>  .. .. .. .. @ pages : int 1
#>  .. .. .. .. @ fields:List of 5
#>  .. .. .. .. .. $ state           : chr NA
#>  .. .. .. .. .. $ reporting_period: chr NA
#>  .. .. .. .. .. $ release_date    : chr NA
#>  .. .. .. .. .. $ value           : chr NA
#>  .. .. .. .. .. $ data_notes      : chr NA
#>  .. $ wind_sbm     : <class_caid>
#>  ..  ..@ access: <class_current>
#>  .. .. .. @ identifier: chr "https://data.medicaid.gov/api/1/datastore/query/5670e72c-e44e-4282-ab67-4ebebaba3cbd/0?count=true&results=true&"| __truncated__
#>  .. .. .. @ metadata  :List of 5
#>  .. .. .. .. $ title      : chr "State-based Marketplace (SBM) Medicaid Unwinding Report"
#>  .. .. .. .. $ description: chr "Metrics from individual Marketplaces during the current reporting period. The report includes data for the stat"| __truncated__
#>  .. .. .. .. $ modified   : Date[1:1], format: "2024-12-27"
#>  .. .. .. .. $ periodicity: chr "Decennially [R/P10Y]"
#>  .. .. .. .. $ download   : chr NA
#>  .. .. .. @ dimensions: <class_dimensions>
#>  .. .. .. .. @ limit : int 8000
#>  .. .. .. .. @ rows  : int 128
#>  .. .. .. .. @ pages : int 1
#>  .. .. .. .. @ fields:List of 59
#>  .. .. .. .. .. $ state                                      : chr NA
#>  .. .. .. .. .. $ metric                                     : chr NA
#>  .. .. .. .. .. $ april_2023_count                           : chr NA
#>  .. .. .. .. .. $ april_2023_percentage_applications         : chr NA
#>  .. .. .. .. .. $ april_2023_percentage_account_transfers    : chr NA
#>  .. .. .. .. .. $ may_2023_count                             : chr NA
#>  .. .. .. .. .. $ may_2023_percentage_applications           : chr NA
#>  .. .. .. .. .. $ may_2023_percentage_account_transfers      : chr NA
#>  .. .. .. .. .. $ june_2023_count                            : chr NA
#>  .. .. .. .. .. $ june_2023_percentage_applications          : chr NA
#>  .. .. .. .. .. $ june_2023_percentage_account_transfers     : chr NA
#>  .. .. .. .. .. $ july_2023_count                            : chr NA
#>  .. .. .. .. .. $ july_2023_percentage_applications          : chr NA
#>  .. .. .. .. .. $ july_2023_percentage_account_transfers     : chr NA
#>  .. .. .. .. .. $ august_2023_count                          : chr NA
#>  .. .. .. .. .. $ august_2023_percentage_applications        : chr NA
#>  .. .. .. .. .. $ august_2023_percentage_account_transfers   : chr NA
#>  .. .. .. .. .. $ september_2023_count                       : chr NA
#>  .. .. .. .. .. $ september_2023_percentage_applications     : chr NA
#>  .. .. .. .. .. $ september_2023_percentage_account_transfers: chr NA
#>  .. .. .. .. .. $ october_2023_count                         : chr NA
#>  .. .. .. .. .. $ october_2023_percentage_applications       : chr NA
#>  .. .. .. .. .. $ october_2023_percentage_account_transfers  : chr NA
#>  .. .. .. .. .. $ november_2023_count                        : chr NA
#>  .. .. .. .. .. $ november_2023_percentage_applications      : chr NA
#>  .. .. .. .. .. $ november_2023_percentage_account_transfers : chr NA
#>  .. .. .. .. .. $ december_2023_count                        : chr NA
#>  .. .. .. .. .. $ december_2023_percentage_applications      : chr NA
#>  .. .. .. .. .. $ december_2023_percentage_account_transfers : chr NA
#>  .. .. .. .. .. $ january_2024_count                         : chr NA
#>  .. .. .. .. .. $ january_2024_percentage_applications       : chr NA
#>  .. .. .. .. .. $ january_2024_percentage_account_transfers  : chr NA
#>  .. .. .. .. .. $ february_2024_count                        : chr NA
#>  .. .. .. .. .. $ february_2024_percentage_applications      : chr NA
#>  .. .. .. .. .. $ february_2024_percentage_account_transfers : chr NA
#>  .. .. .. .. .. $ march_2024_count                           : chr NA
#>  .. .. .. .. .. $ march_2024_percentage_applications         : chr NA
#>  .. .. .. .. .. $ march_2024_percentage_account_transfers    : chr NA
#>  .. .. .. .. .. $ april_2024_count                           : chr NA
#>  .. .. .. .. .. $ april_2024_percentage_applications         : chr NA
#>  .. .. .. .. .. $ april_2024_percentage_account_transfers    : chr NA
#>  .. .. .. .. .. $ may_2024_count                             : chr NA
#>  .. .. .. .. .. $ may_2024_percentage_applications           : chr NA
#>  .. .. .. .. .. $ may_2024_percentage_account_transfers      : chr NA
#>  .. .. .. .. .. $ june_2024_count                            : chr NA
#>  .. .. .. .. .. $ june_2024_percentage_applications          : chr NA
#>  .. .. .. .. .. $ june_2024_percentage_account_transfers     : chr NA
#>  .. .. .. .. .. $ july_2024_count                            : chr NA
#>  .. .. .. .. .. $ july_2024_percentage_applications          : chr NA
#>  .. .. .. .. .. $ july_2024_percentage_account_transfers     : chr NA
#>  .. .. .. .. .. $ august_2024_count                          : chr NA
#>  .. .. .. .. .. $ august_2024_percentage_applications        : chr NA
#>  .. .. .. .. .. $ august_2024_percentage_account_transfers   : chr NA
#>  .. .. .. .. .. $ september_2024_count                       : chr NA
#>  .. .. .. .. .. $ september_2024_percentage_applications     : chr NA
#>  .. .. .. .. .. $ september_2024_percentage_account_transfers: chr NA
#>  .. .. .. .. .. $ total_count                                : chr NA
#>  .. .. .. .. .. $ total_percentage_applications              : chr NA
#>  .. .. .. .. .. $ total_percentage_account_transfers         : chr NA
collection("managed")
#> <class_collection>
#>  @ name   : chr "Managed Care Enrollment"
#>  @ members:List of 12
#>  .. $ man_mltss     : <class_caid>
#>  ..  ..@ access: <class_current>
#>  .. .. .. @ identifier: chr "https://data.medicaid.gov/api/1/datastore/query/5394bcab-c748-5e4b-af07-b5bf77ed3aa3/0?count=true&results=true&"| __truncated__
#>  .. .. .. @ metadata  :List of 5
#>  .. .. .. .. $ title      : chr "Managed Long Term Services and Supports (MLTSS) Enrollees"
#>  .. .. .. .. $ description: chr "1. Enrollment includes both Medicaid-only and Medicare-Medicaid (dual) enrollees. For both types of enrollees, "| __truncated__
#>  .. .. .. .. $ modified   : Date[1:1], format: "2024-10-16"
#>  .. .. .. .. $ periodicity: chr "Decennially [R/P10Y]"
#>  .. .. .. .. $ download   : chr NA
#>  .. .. .. @ dimensions: <class_dimensions>
#>  .. .. .. .. @ limit : int 8000
#>  .. .. .. .. @ rows  : int 399
#>  .. .. .. .. @ pages : int 1
#>  .. .. .. .. @ fields:List of 8
#>  .. .. .. .. .. $ state                                    : chr NA
#>  .. .. .. .. .. $ notes                                    : chr NA
#>  .. .. .. .. .. $ total_any_managed_care_enrollees         : chr NA
#>  .. .. .. .. .. $ comprehensive_managed_care_ltss_enrollees: chr NA
#>  .. .. .. .. .. $ comprehensive_managed_care_ltss_percent  : chr NA
#>  .. .. .. .. .. $ managed_ltss_only_enrollees              : chr NA
#>  .. .. .. .. .. $ managed_ltss_only_percent                : chr NA
#>  .. .. .. .. .. $ year                                     : chr NA
#>  .. $ man_summary   : <class_caid>
#>  ..  ..@ access: <class_current>
#>  .. .. .. @ identifier: chr "https://data.medicaid.gov/api/1/datastore/query/52ed908b-0cb8-5dd2-846d-99d4af12b369/0?count=true&results=true&"| __truncated__
#>  .. .. .. @ metadata  :List of 5
#>  .. .. .. .. $ title      : chr "Managed Care Enrollment Summary"
#>  .. .. .. .. $ description: chr "The Medicaid Managed Care Enrollment Report profiles enrollment statistics on Medicaid managed care programs on"| __truncated__
#>  .. .. .. .. $ modified   : Date[1:1], format: "2024-10-16"
#>  .. .. .. .. $ periodicity: chr "Decennially [R/P10Y]"
#>  .. .. .. .. $ download   : chr NA
#>  .. .. .. @ dimensions: <class_dimensions>
#>  .. .. .. .. @ limit : int 8000
#>  .. .. .. .. @ rows  : int 399
#>  .. .. .. .. @ pages : int 1
#>  .. .. .. .. @ fields:List of 7
#>  .. .. .. .. .. $ state                                                        : chr NA
#>  .. .. .. .. .. $ notes                                                        : chr NA
#>  .. .. .. .. .. $ total_medicaid_enrollees                                     : chr NA
#>  .. .. .. .. .. $ total_medicaid_enrollment_in_any_type_of_managed_care        : chr NA
#>  .. .. .. .. .. $ medicaid_enrollment_in_comprehensive_managed_care            : chr NA
#>  .. .. .. .. .. $ medicaid_newly_eligible_adults_enrolled_in_comprehensive_mcos: chr NA
#>  .. .. .. .. .. $ year                                                         : chr NA
#>  .. $ man_state_end : <class_caid>
#>  ..  ..@ access: <class_current>
#>  .. .. .. @ identifier: chr "https://data.medicaid.gov/api/1/datastore/query/0e112ea8-8e8e-5dee-a7e2-7ed551c3baa4/0?count=true&results=true&"| __truncated__
#>  .. .. .. @ metadata  :List of 5
#>  .. .. .. .. $ title      : chr "Managed Care Programs by State"
#>  .. .. .. .. $ description: chr "Provides program names, program features, population enrolled, benefits covered, quality assurance and improvem"| __truncated__
#>  .. .. .. .. $ modified   : Date[1:1], format: "2018-08-22"
#>  .. .. .. .. $ periodicity: chr "Decennially [R/P10Y]"
#>  .. .. .. .. $ download   : chr NA
#>  .. .. .. @ dimensions: <class_dimensions>
#>  .. .. .. .. @ limit : int 8000
#>  .. .. .. .. @ rows  : int 195
#>  .. .. .. .. @ pages : int 1
#>  .. .. .. .. @ fields:List of 69
#>  .. .. .. .. .. $ state                                                           : chr NA
#>  .. .. .. .. .. $ program_name                                                    : chr NA
#>  .. .. .. .. .. $ program_type                                                    : chr NA
#>  .. .. .. .. .. $ statewide_or_regionspecific                                     : chr NA
#>  .. .. .. .. .. $ federal_operating_authority                                     : chr NA
#>  .. .. .. .. .. $ program_start_date                                              : chr NA
#>  .. .. .. .. .. $ waiver_expiration_date_if_applicable                            : chr NA
#>  .. .. .. .. .. $ if_the_program_ended_in_2014_indicate_the_end_date              : chr NA
#>  .. .. .. .. .. $ populations_enrolled                                            : chr NA
#>  .. .. .. .. .. $ lowincome_adults_not_eligible_under_aca_section_viii            : chr NA
#>  .. .. .. .. .. $ aged_blind_or_disabled_children_or_adults                       : chr NA
#>  .. .. .. .. .. $ nondisabled_children_excluding_children_in_foster_care_or_r_f40d: chr NA
#>  .. .. .. .. .. $ individuals_receiving_limited_benefits                          : chr NA
#>  .. .. .. .. .. $ lowincome_adults_eligible_under_aca_section_viii                : chr NA
#>  .. .. .. .. .. $ full_duals                                                      : chr NA
#>  .. .. .. .. .. $ partial_duals                                                   : chr NA
#>  .. .. .. .. .. $ children_with_special_health_care_needs                         : chr NA
#>  .. .. .. .. .. $ native_americanalaskan_natives                                  : chr NA
#>  .. .. .. .. .. $ foster_care_and_adoption_assistance_children                    : chr NA
#>  .. .. .. .. .. $ enrollment_choice_period                                        : chr NA
#>  .. .. .. .. .. $ enrollment_broker_name_if_applicable                            : chr NA
#>  .. .. .. .. .. $ notes_on_enrollment_choice_period                               : chr NA
#>  .. .. .. .. .. $ benefits_covered                                                : chr NA
#>  .. .. .. .. .. $ inpatient_hospital_physical_health                              : chr NA
#>  .. .. .. .. .. $ inpatient_hospital_behavioral_health_mh_andor_sud               : chr NA
#>  .. .. .. .. .. $ outpatient_hospital_physical_health                             : chr NA
#>  .. .. .. .. .. $ outpatient_hospital_behavioral_health_mh_andor_sud              : chr NA
#>  .. .. .. .. .. $ partial_hospitalization                                         : chr NA
#>  .. .. .. .. .. $ physician                                                       : chr NA
#>  .. .. .. .. .. $ nurse_practitioner                                              : chr NA
#>  .. .. .. .. .. $ rural_health_clinics_and_fqhcs                                  : chr NA
#>  .. .. .. .. .. $ clinic_services                                                 : chr NA
#>  .. .. .. .. .. $ lab_and_xray                                                    : chr NA
#>  .. .. .. .. .. $ prescription_drugs_and_prosthetic_devices                       : chr NA
#>  .. .. .. .. .. $ epsdt                                                           : chr NA
#>  .. .. .. .. .. $ case_management                                                 : chr NA
#>  .. .. .. .. .. $ health_home_ssa_1945                                            : chr NA
#>  .. .. .. .. .. $ family_planning                                                 : chr NA
#>  .. .. .. .. .. $ dental_services_medicalsurgical                                 : chr NA
#>  .. .. .. .. .. $ dental_preventative_or_corrective                               : chr NA
#>  .. .. .. .. .. $ home_health_agency_services                                     : chr NA
#>  .. .. .. .. .. $ personal_care_state_plan_option                                 : chr NA
#>  .. .. .. .. .. $ hcbs_waiver_services                                            : chr NA
#>  .. .. .. .. .. $ private_duty_nursing                                            : chr NA
#>  .. .. .. .. .. $ icfidd                                                          : chr NA
#>  .. .. .. .. .. $ nursing_facility_services                                       : chr NA
#>  .. .. .. .. .. $ hospice_care                                                    : chr NA
#>  .. .. .. .. .. $ nonemergency_medical_transportation                             : chr NA
#>  .. .. .. .. .. $ other_eg_nurse_midwife_services_freestanding_birth_centers__8f59: chr NA
#>  .. .. .. .. .. $ quality_assurance_and_improvement                               : chr NA
#>  .. .. .. .. .. $ hedis_data_required                                             : chr NA
#>  .. .. .. .. .. $ cahps_data_required                                             : chr NA
#>  .. .. .. .. .. $ accreditation_required                                          : chr NA
#>  .. .. .. .. .. $ accrediting_organization                                        : chr NA
#>  .. .. .. .. .. $ eqro_contractor_name_if_applicable                              : chr NA
#>  .. .. .. .. .. $ performance_incentives                                          : chr NA
#>  .. .. .. .. .. $ payment_bonusesdifferentials_to_reward_plans                    : chr NA
#>  .. .. .. .. .. $ preferential_autoenrollment_to_reward_plans                     : chr NA
#>  .. .. .. .. .. $ public_reports_comparing_mco_performance_on_key_metrics         : chr NA
#>  .. .. .. .. .. $ withholds_tied_to_performance_metrics                           : chr NA
#>  .. .. .. .. .. $ mcosphps_requiredencouraged_to_pay_providers_for_valuequali_ea98: chr NA
#>  .. .. .. .. .. $ provider_valuebased_purchasing                                  : chr NA
#>  .. .. .. .. .. $ state_pays_providerbased_entities_such_as_acos_or_pcmhs_dir_89da: chr NA
#>  .. .. .. .. .. $ participating_plans_and_regions_served                          : chr NA
#>  .. .. .. .. .. $ plans_in_program                                                : chr NA
#>  .. .. .. .. .. $ notes                                                           : chr NA
#>  .. .. .. .. .. $ program_notes                                                   : chr NA
#>  .. .. .. .. .. $ year                                                            : chr NA
#>  .. .. .. .. .. $ location                                                        : chr NA
#>  .. $ man_program   : <class_caid>
#>  ..  ..@ access: <class_current>
#>  .. .. .. @ identifier: chr "https://data.medicaid.gov/api/1/datastore/query/0bef7b8a-c663-5b14-9a46-0b5c2b86b0fe/0?count=true&results=true&"| __truncated__
#>  .. .. .. @ metadata  :List of 5
#>  .. .. .. .. $ title      : chr "Managed Care Enrollment by Program and Plan"
#>  .. .. .. .. $ description: chr "The Medicaid Managed Care Enrollment Report profiles enrollment statistics on Medicaid managed care programs on"| __truncated__
#>  .. .. .. .. $ modified   : Date[1:1], format: "2024-10-16"
#>  .. .. .. .. $ periodicity: chr "Decennially [R/P10Y]"
#>  .. .. .. .. $ download   : chr NA
#>  .. .. .. @ dimensions: <class_dimensions>
#>  .. .. .. .. @ limit : int 8000
#>  .. .. .. .. @ rows  : int 5974
#>  .. .. .. .. @ pages : int 1
#>  .. .. .. .. @ fields:List of 10
#>  .. .. .. .. .. $ state                  : chr NA
#>  .. .. .. .. .. $ notes                  : chr NA
#>  .. .. .. .. .. $ program_name           : chr NA
#>  .. .. .. .. .. $ plan_name              : chr NA
#>  .. .. .. .. .. $ geographic_region      : chr NA
#>  .. .. .. .. .. $ medicaidonly_enrollment: chr NA
#>  .. .. .. .. .. $ dual_enrollment        : chr NA
#>  .. .. .. .. .. $ total_enrollment       : chr NA
#>  .. .. .. .. .. $ year                   : chr NA
#>  .. .. .. .. .. $ parent_organization    : chr NA
#>  .. $ man_pop       : <class_caid>
#>  ..  ..@ access: <class_current>
#>  .. .. .. @ identifier: chr "https://data.medicaid.gov/api/1/datastore/query/e2ce0d2f-07c5-5213-947a-31e19bc649f6/0?count=true&results=true&"| __truncated__
#>  .. .. .. @ metadata  :List of 5
#>  .. .. .. .. $ title      : chr "Managed Care Enrollment by Program and Population (All)"
#>  .. .. .. .. $ description: chr "The Medicaid Managed Care Enrollment Report profiles enrollment statistics on Medicaid managed care programs on"| __truncated__
#>  .. .. .. .. $ modified   : Date[1:1], format: "2024-10-16"
#>  .. .. .. .. $ periodicity: chr "Decennially [R/P10Y]"
#>  .. .. .. .. $ download   : chr NA
#>  .. .. .. @ dimensions: <class_dimensions>
#>  .. .. .. .. @ limit : int 8000
#>  .. .. .. .. @ rows  : int 401
#>  .. .. .. .. @ pages : int 1
#>  .. .. .. .. @ fields:List of 13
#>  .. .. .. .. .. $ state                                  : chr NA
#>  .. .. .. .. .. $ notes                                  : chr NA
#>  .. .. .. .. .. $ total_medicaid_enrollees               : chr NA
#>  .. .. .. .. .. $ comprehensive_mco_with_or_without_mltss: chr NA
#>  .. .. .. .. .. $ pccm                                   : chr NA
#>  .. .. .. .. .. $ mltss_only                             : chr NA
#>  .. .. .. .. .. $ bho_pihp_andor_pahp                    : chr NA
#>  .. .. .. .. .. $ dental                                 : chr NA
#>  .. .. .. .. .. $ transportation                         : chr NA
#>  .. .. .. .. .. $ pace                                   : chr NA
#>  .. .. .. .. .. $ other                                  : chr NA
#>  .. .. .. .. .. $ year                                   : chr NA
#>  .. .. .. .. .. $ pccm_entity                            : chr NA
#>  .. $ man_dual      : <class_caid>
#>  ..  ..@ access: <class_current>
#>  .. .. .. @ identifier: chr "https://data.medicaid.gov/api/1/datastore/query/8e7be65b-97ba-5ecf-8394-ca8e6f63685a/0?count=true&results=true&"| __truncated__
#>  .. .. .. @ metadata  :List of 5
#>  .. .. .. .. $ title      : chr "Managed Care Enrollment by Program and Population (Duals)"
#>  .. .. .. .. $ description: chr "The Medicaid Managed Care Enrollment Report profiles enrollment statistics on Medicaid managed care programs on"| __truncated__
#>  .. .. .. .. $ modified   : Date[1:1], format: "2024-10-16"
#>  .. .. .. .. $ periodicity: chr "Decennially [R/P10Y]"
#>  .. .. .. .. $ download   : chr NA
#>  .. .. .. @ dimensions: <class_dimensions>
#>  .. .. .. .. @ limit : int 8000
#>  .. .. .. .. @ rows  : int 399
#>  .. .. .. .. @ pages : int 1
#>  .. .. .. .. @ fields:List of 13
#>  .. .. .. .. .. $ state                                  : chr NA
#>  .. .. .. .. .. $ notes                                  : chr NA
#>  .. .. .. .. .. $ total_beneficiaries                    : chr NA
#>  .. .. .. .. .. $ comprehensive_mco_with_or_without_mltss: chr NA
#>  .. .. .. .. .. $ pccm                                   : chr NA
#>  .. .. .. .. .. $ pccm_entity                            : chr NA
#>  .. .. .. .. .. $ mltss_only                             : chr NA
#>  .. .. .. .. .. $ bho_pihp_andor_pahp                    : chr NA
#>  .. .. .. .. .. $ dental                                 : chr NA
#>  .. .. .. .. .. $ transportation                         : chr NA
#>  .. .. .. .. .. $ pace                                   : chr NA
#>  .. .. .. .. .. $ other                                  : chr NA
#>  .. .. .. .. .. $ year                                   : chr NA
#>  .. $ man_feat_pop  : <class_caid>
#>  ..  ..@ access: <class_current>
#>  .. .. .. @ identifier: chr "https://data.medicaid.gov/api/1/datastore/query/927f4847-2c0a-50c1-8f50-9103de7d048b/0?count=true&results=true&"| __truncated__
#>  .. .. .. @ metadata  :List of 5
#>  .. .. .. .. $ title      : chr "Managed Care Features By Enrollment Population"
#>  .. .. .. .. $ description: chr "Number of Managed Care Programs Enrolling Certain Populations on a Mandatory or Voluntary Basis at any point in 2022"
#>  .. .. .. .. $ modified   : Date[1:1], format: "2024-10-16"
#>  .. .. .. .. $ periodicity: chr "Decennially [R/P10Y]"
#>  .. .. .. .. $ download   : chr NA
#>  .. .. .. @ dimensions: <class_dimensions>
#>  .. .. .. .. @ limit : int 8000
#>  .. .. .. .. @ rows  : int 74
#>  .. .. .. .. @ pages : int 1
#>  .. .. .. .. @ fields:List of 21
#>  .. .. .. .. .. $ features                                 : chr NA
#>  .. .. .. .. .. $ notes                                    : chr NA
#>  .. .. .. .. .. $ comprehensive_mco_with_or_without_mltss  : chr NA
#>  .. .. .. .. .. $ comprehensive_mco_with_or_without_mltss_2: chr NA
#>  .. .. .. .. .. $ pccm                                     : chr NA
#>  .. .. .. .. .. $ pccm_2                                   : chr NA
#>  .. .. .. .. .. $ pccm_entity                              : chr NA
#>  .. .. .. .. .. $ pccm_entity_2                            : chr NA
#>  .. .. .. .. .. $ mltss                                    : chr NA
#>  .. .. .. .. .. $ mltss_2                                  : chr NA
#>  .. .. .. .. .. $ bho_pihp_andor_pahp                      : chr NA
#>  .. .. .. .. .. $ bho_pihp_andor_pahp_2                    : chr NA
#>  .. .. .. .. .. $ dental                                   : chr NA
#>  .. .. .. .. .. $ dental_2                                 : chr NA
#>  .. .. .. .. .. $ transportation                           : chr NA
#>  .. .. .. .. .. $ transportation_2                         : chr NA
#>  .. .. .. .. .. $ other_php                                : chr NA
#>  .. .. .. .. .. $ other_php_2                              : chr NA
#>  .. .. .. .. .. $ pace                                     : chr NA
#>  .. .. .. .. .. $ pace_2                                   : chr NA
#>  .. .. .. .. .. $ year                                     : chr NA
#>  .. $ man_feat_qa   : <class_caid>
#>  ..  ..@ access: <class_current>
#>  .. .. .. @ identifier: chr "https://data.medicaid.gov/api/1/datastore/query/a6d2c261-210d-587c-8fa4-892444b7de42/0?count=true&results=true&"| __truncated__
#>  .. .. .. @ metadata  :List of 5
#>  .. .. .. .. $ title      : chr "Managed Care Features by QA and Performance Incentive"
#>  .. .. .. .. $ description: chr "Number of Managed Care Program Types, by Quality Assurance Requirements, Performance Incentives, and Provider V"| __truncated__
#>  .. .. .. .. $ modified   : Date[1:1], format: "2024-10-16"
#>  .. .. .. .. $ periodicity: chr "Decennially [R/P10Y]"
#>  .. .. .. .. $ download   : chr NA
#>  .. .. .. @ dimensions: <class_dimensions>
#>  .. .. .. .. @ limit : int 8000
#>  .. .. .. .. @ rows  : int 63
#>  .. .. .. .. @ pages : int 1
#>  .. .. .. .. @ fields:List of 11
#>  .. .. .. .. .. $ features                               : chr NA
#>  .. .. .. .. .. $ comprehensive_mco_with_or_without_mltss: chr NA
#>  .. .. .. .. .. $ pccm                                   : chr NA
#>  .. .. .. .. .. $ pccm_entity                            : chr NA
#>  .. .. .. .. .. $ mltss                                  : chr NA
#>  .. .. .. .. .. $ bho_pihp_andor_pahp                    : chr NA
#>  .. .. .. .. .. $ dental                                 : chr NA
#>  .. .. .. .. .. $ transportation                         : chr NA
#>  .. .. .. .. .. $ other_php                              : chr NA
#>  .. .. .. .. .. $ pace                                   : chr NA
#>  .. .. .. .. .. $ year                                   : chr NA
#>  .. $ man_bene_month: <class_caid>
#>  ..  ..@ access: <class_current>
#>  .. .. .. @ identifier: chr "https://data.medicaid.gov/api/1/datastore/query/89baf100-259b-4763-b9e2-337972f988c4/0?count=true&results=true&"| __truncated__
#>  .. .. .. @ metadata  :List of 5
#>  .. .. .. .. $ title      : chr "Managed Care Information for Medicaid and CHIP Beneficiaries by Month"
#>  .. .. .. .. $ description: chr "This data set includes monthly enrollment counts of Medicaid and CHIP beneficiaries by managed care participati"| __truncated__
#>  .. .. .. .. $ modified   : Date[1:1], format: "2024-01-05"
#>  .. .. .. .. $ periodicity: chr "Decennially [R/P10Y]"
#>  .. .. .. .. $ download   : chr "https://download.medicaid.gov/data/ManagedCare-montly.csv"
#>  .. .. .. @ dimensions: <class_dimensions>
#>  .. .. .. .. @ limit : int 8000
#>  .. .. .. .. @ rows  : int 31164
#>  .. .. .. .. @ pages : int 4
#>  .. .. .. .. @ fields:List of 5
#>  .. .. .. .. .. $ state                   : chr NA
#>  .. .. .. .. .. $ month                   : chr NA
#>  .. .. .. .. .. $ managedcareparticipation: chr NA
#>  .. .. .. .. .. $ countenrolled           : chr NA
#>  .. .. .. .. .. $ dqunusable              : chr NA
#>  .. $ man_bene_year : <class_caid>
#>  ..  ..@ access: <class_current>
#>  .. .. .. @ identifier: chr "https://data.medicaid.gov/api/1/datastore/query/c0bef49e-46ea-4b9d-9153-76b4078d58b7/0?count=true&results=true&"| __truncated__
#>  .. .. .. @ metadata  :List of 5
#>  .. .. .. .. $ title      : chr "Managed Care Information for Medicaid and CHIP Beneficiaries by Year"
#>  .. .. .. .. $ description: chr "This data set presents annual enrollment counts of Medicaid and CHIP beneficiaries by managed care participatio"| __truncated__
#>  .. .. .. .. $ modified   : Date[1:1], format: "2023-03-20"
#>  .. .. .. .. $ periodicity: chr "Decennially [R/P10Y]"
#>  .. .. .. .. $ download   : chr "https://download.medicaid.gov/data/ManagedCare-anul.csv"
#>  .. .. .. @ dimensions: <class_dimensions>
#>  .. .. .. .. @ limit : int 8000
#>  .. .. .. .. @ rows  : int 2597
#>  .. .. .. .. @ pages : int 1
#>  .. .. .. .. @ fields:List of 7
#>  .. .. .. .. .. $ state                    : chr NA
#>  .. .. .. .. .. $ year                     : chr NA
#>  .. .. .. .. .. $ managedcareparticipation : chr NA
#>  .. .. .. .. .. $ counteverenrolled        : chr NA
#>  .. .. .. .. .. $ countlastmonthenrollment : chr NA
#>  .. .. .. .. .. $ averageenrollmentpermonth: chr NA
#>  .. .. .. .. .. $ dqunusable               : chr NA
#>  .. $ man_share     : <class_caid>
#>  ..  ..@ access: <class_current>
#>  .. .. .. @ identifier: chr "https://data.medicaid.gov/api/1/datastore/query/79692ea5-21e1-56bf-8149-97d437120c4b/0?count=true&results=true&"| __truncated__
#>  .. .. .. @ metadata  :List of 5
#>  .. .. .. .. $ title      : chr "Share of Medicaid Enrollees in Managed Care"
#>  .. .. .. .. $ description: chr "The Share of Medicaid Enrollees in any Managed Care and in Comprehensive Managed CaAre profiles state-level enr"| __truncated__
#>  .. .. .. .. $ modified   : Date[1:1], format: "2024-10-16"
#>  .. .. .. .. $ periodicity: chr "Decennially [R/P10Y]"
#>  .. .. .. .. $ download   : chr NA
#>  .. .. .. @ dimensions: <class_dimensions>
#>  .. .. .. .. @ limit : int 8000
#>  .. .. .. .. @ rows  : int 399
#>  .. .. .. .. @ pages : int 1
#>  .. .. .. .. @ fields:List of 8
#>  .. .. .. .. .. $ state                                          : chr NA
#>  .. .. .. .. .. $ notes                                          : chr NA
#>  .. .. .. .. .. $ total_medicaid_enrollees                       : chr NA
#>  .. .. .. .. .. $ individuals_enrolled_any                       : chr NA
#>  .. .. .. .. .. $ percent_of_all_medicaid_enrollees_any          : chr NA
#>  .. .. .. .. .. $ individuals_enrolled_comprehensive             : chr NA
#>  .. .. .. .. .. $ percent_of_all_medicaid_enrollees_comprehensive: chr NA
#>  .. .. .. .. .. $ year                                           : chr NA
#>  .. $ man_tmp       : <class_caid>
#>  ..  ..@ access: <class_temporal>
#>  .. .. .. @ identifier: tibble [1 × 4] (S3: tbl_df/tbl/data.frame)
#>  $ year      : int 2017
#>  $ modified  : Date[1:1], format: "2019-05-03"
#>  $ identifier: chr "https://data.medicaid.gov/api/1/datastore/query/8f39b637-9bb1-5894-9062-2c4f2ad70fba/0?count=true&results=true&"| __truncated__
#>  $ download  : chr NA
#>  .. .. .. @ metadata  :List of 3
#>  .. .. .. .. $ title      : chr "Managed Care Programs by State"
#>  .. .. .. .. $ description: chr "Provides program names, program features, population enrolled, benefits covered, quality assurance and improvem"| __truncated__
#>  .. .. .. .. $ periodicity: chr "Decennially [R/P10Y]"
#>  .. .. .. @ dimensions: <class_dimensions>
#>  .. .. .. .. @ limit : int 8000
#>  .. .. .. .. @ rows  : int 175
#>  .. .. .. .. @ pages : int 1
#>  .. .. .. .. @ fields:List of 61
#>  .. .. .. .. .. $ state                                                           : chr NA
#>  .. .. .. .. .. $ program_name                                                    : chr NA
#>  .. .. .. .. .. $ program_type                                                    : chr NA
#>  .. .. .. .. .. $ statewide_or_regionspecific                                     : chr NA
#>  .. .. .. .. .. $ federal_operating_authority                                     : chr NA
#>  .. .. .. .. .. $ program_start_date                                              : chr NA
#>  .. .. .. .. .. $ waiver_expiration_date_if_applicable                            : chr NA
#>  .. .. .. .. .. $ if_the_program_ended_in_2017_indicate_the_end_date              : chr NA
#>  .. .. .. .. .. $ populations_enrolled_lowincome_adults_not_covered_under_aca_0778: chr NA
#>  .. .. .. .. .. $ populations_enrolled_lowincome_adults_covered_under__aca_se_9567: chr NA
#>  .. .. .. .. .. $ populations_enrolled_aged_blind_or_disabled_children_or_adults  : chr NA
#>  .. .. .. .. .. $ populations_enrolled_nondisabled_children_excludes_children_1dda: chr NA
#>  .. .. .. .. .. $ populations_enrolled_individuals_receiving_limited_benefits_facc: chr NA
#>  .. .. .. .. .. $ populations_enrolled_full_duals                                 : chr NA
#>  .. .. .. .. .. $ populations_enrolled_partial_duals                              : chr NA
#>  .. .. .. .. .. $ populations_enrolled_children_with_special_health_care_needs    : chr NA
#>  .. .. .. .. .. $ populations_enrolled_native_americanalaskan_natives             : chr NA
#>  .. .. .. .. .. $ populations_enrolled_foster_care_and_adoption_assistance_ch_b424: chr NA
#>  .. .. .. .. .. $ populations_enrolled_enrollment_choice_period                   : chr NA
#>  .. .. .. .. .. $ populations_enrolled_enrollment_broker_name_if_applicable       : chr NA
#>  .. .. .. .. .. $ populations_enrolled_notes_on_enrollment_choice_period          : chr NA
#>  .. .. .. .. .. $ benefits_covered_inpatient_hospital_physical_health             : chr NA
#>  .. .. .. .. .. $ benefits_covered_inpatient_hospital_behavioral_health_mh_an_4041: chr NA
#>  .. .. .. .. .. $ benefits_covered_outpatient_hospital_physical_health            : chr NA
#>  .. .. .. .. .. $ benefits_covered_outpatient_hospital_behavioral_health_mh_a_ceee: chr NA
#>  .. .. .. .. .. $ benefits_covered_partial_hospitalization                        : chr NA
#>  .. .. .. .. .. $ benefits_covered_physician                                      : chr NA
#>  .. .. .. .. .. $ benefits_covered_nurse_practitioner                             : chr NA
#>  .. .. .. .. .. $ benefits_covered_rural_health_clinics_and_fqhcs                 : chr NA
#>  .. .. .. .. .. $ benefits_covered_clinic_services                                : chr NA
#>  .. .. .. .. .. $ benefits_covered_lab_and_xray                                   : chr NA
#>  .. .. .. .. .. $ benefits_covered_prescription_drugs_and_prosthetic_devices      : chr NA
#>  .. .. .. .. .. $ benefits_covered_epsdt                                          : chr NA
#>  .. .. .. .. .. $ benefits_covered_case_management                                : chr NA
#>  .. .. .. .. .. $ benefits_covered_ssa_section_1945authorized_health_home         : chr NA
#>  .. .. .. .. .. $ benefits_covered_health_home_care_services_in_home              : chr NA
#>  .. .. .. .. .. $ benefits_covered_family_planning                                : chr NA
#>  .. .. .. .. .. $ benefits_covered_dental_services_medicalsurgical                : chr NA
#>  .. .. .. .. .. $ benefits_covered_dental_preventative_or_corrective              : chr NA
#>  .. .. .. .. .. $ benefits_covered_personal_care_state_plan_option                : chr NA
#>  .. .. .. .. .. $ benefits_covered_hcbs_waiver_services                           : chr NA
#>  .. .. .. .. .. $ benefits_covered_private_duty_nursing                           : chr NA
#>  .. .. .. .. .. $ benefits_covered_icfidd                                         : chr NA
#>  .. .. .. .. .. $ benefits_covered_nursing_facility_services                      : chr NA
#>  .. .. .. .. .. $ benefits_covered_hospice_care                                   : chr NA
#>  .. .. .. .. .. $ benefits_covered_nonemergency_medical_transportation            : chr NA
#>  .. .. .. .. .. $ benefits_covered_institution_for_mental_disease_inpatient_t_448a: chr NA
#>  .. .. .. .. .. $ benefits_covered_other_eg_nurse_midwife_services_freestandi_d504: chr NA
#>  .. .. .. .. .. $ quality_assurance_and_improvement_hedis_data_required           : chr NA
#>  .. .. .. .. .. $ quality_assurance_and_improvement_cahps_data_required           : chr NA
#>  .. .. .. .. .. $ quality_assurance_and_improvement_accreditation_required        : chr NA
#>  .. .. .. .. .. $ quality_assurance_and_improvement_accrediting_organization      : chr NA
#>  .. .. .. .. .. $ quality_assurance_and_improvement_eqro_contractor_name_if_a_f2dc: chr NA
#>  .. .. .. .. .. $ performance_incentives_payment_bonusesdifferentials_to_rewa_eece: chr NA
#>  .. .. .. .. .. $ performance_incentives_preferential_autoenrollment_to_rewar_ab96: chr NA
#>  .. .. .. .. .. $ performance_incentives_public_reports_comparing_plan_perfor_42d0: chr NA
#>  .. .. .. .. .. $ performance_incentives_withholds_tied_to_performance_metrics    : chr NA
#>  .. .. .. .. .. $ performance_incentives_mcosphps_required_or_encouraged_to_p_bf59: chr NA
#>  .. .. .. .. .. $ participating_plans_plans_in_program                            : chr NA
#>  .. .. .. .. .. $ notes_program_notes                                             : chr NA
#>  .. .. .. .. .. $ year                                                            : chr NA