Load Endpoints

Load a data endpoint or collection by its alias.

Usage

endpoint(alias, call = rlang::caller_env())

collection(alias, call = rlang::caller_env())

group(..., .title = NULL, call = rlang::caller_env())

Arguments

alias

<chr> endpoint or collection alias

call

The environment from which the function is called.

...

<dynamic-dots> endpoint aliases to be grouped

.title

<chr> Group title. Defaults to NULL, which will concatenate the aliases for the title.

Value

An S7 <endpoint/collection/group> object.

Details

The endpoint() function loads a single data endpoint, while the collection() function loads a collection of endpoints, and the group() function loads a group of endpoints. The alias parameter is used to specify the endpoint or collection alias. If the alias is not found, an error will be raised.

Examples

endpoint("dial_facility")
#> Error: ✖ "dial_facility" is not an <endpoint> alias.
endpoint("man_state")
#> Error: ✖ "man_state" is not an <endpoint> alias.
try(endpoint(c("asc_facility", "enterprise")))
#> Error in eval(expr, envir) : 
#>   ✖ Only one <endpoint> can be loaded at a time.
try(endpoint("unwind"))
#> Error in eval(expr, envir) : 
#>   ✖ "unwind" is not an <endpoint> alias.
try(endpoint("ex"))
#> Error in eval(expr, envir) : 
#>   ✖ "ex" is not an <endpoint> alias.
collection("unwind")
#> Error: ✖ "unwind" is not a <collection> alias.
try(collection(c("asc_facility", "enterprise")))
#> Error in eval(expr, envir) : 
#>   ✖ Only one <collection> can be loaded at a time.
try(collection("asc_facility"))
#> Error in eval(expr, envir) : 
#>   ✖ "asc_facility" is not a <collection> alias.
group("asc_facility", "enterprise")
#> Error in eval(mc, parent.frame()): ℹ In index: 1.
#> Caused by error in `map_()`:
#> ✖ "asc_facility" is not an <endpoint> alias.
try(group("asc_facility"))
#> Error in eval(expr, envir) : 
#>   ✖ A <group> must contain more than one <endpoint>.
#> → Run endpoint("asc_facility") to load this endpoint.
try(group("util"))
#> Error in eval(expr, envir) : 
#>   ✖ A <group> must contain more than one <endpoint>.
#> → Run endpoint("util") to load this endpoint.