Medicare Provider Enrollees

Information on a point in time snapshot of enrollment level data for providers actively enrolled in Medicare.

Usage

enrollees(
  npi = NULL,
  pac = NULL,
  enid = NULL,
  state = NULL,
  first_name = NULL,
  middle_name = NULL,
  last_name = NULL,
  org_name = NULL,
  specialty_code = NULL,
  specialty_description = NULL
)

Arguments

npi

<chr> 10-digit Individual NPI

pac

<chr> 10-digit PECOS Associate Control (PAC) ID

enid

<chr> 15-digit Medicare Enrollment ID

state

<chr> Enrollment state abbreviation

first_name, middle_name, last_name

<chr> Individual provider's name

org_name

<chr> Organizational provider's name

specialty_code

<chr> Enrollment specialty code

specialty_description

<chr> Enrollment specialty description

Value

<tibble> of search results

Details

Individual & organizational enrollment-level data on providers actively approved to bill Medicare.

Periodicity

Quarterly (R/P3M)

Links

Enrollment API

Enrollment Data Dictionary

Examples

enrollees(enid = "I20040309000221")
#> # A tibble: 1 × 10
#>   npi        pac     enid  specialty_code specialty_description state first_name
#>   <chr>      <chr>   <chr> <chr>          <chr>                 <chr> <chr>     
#> 1 1417918293 387048… I200… 14-41          PRACTITIONER - OPTOM… FL    STEVEN    
#> # ℹ 3 more variables: middle_name <chr>, last_name <chr>, org_name <chr>
enrollees(npi = "1417918293")
#> # A tibble: 1 × 10
#>   npi        pac     enid  specialty_code specialty_description state first_name
#>   <chr>      <chr>   <chr> <chr>          <chr>                 <chr> <chr>     
#> 1 1417918293 387048… I200… 14-41          PRACTITIONER - OPTOM… FL    STEVEN    
#> # ℹ 3 more variables: middle_name <chr>, last_name <chr>, org_name <chr>
enrollees(pac = "2860305554")
#> # A tibble: 1 × 10
#>   npi        pac     enid  specialty_code specialty_description state first_name
#>   <chr>      <chr>   <chr> <chr>          <chr>                 <chr> <chr>     
#> 1 1134122260 286030… I200… 14-41          PRACTITIONER - OPTOM… TX    ROBERT    
#> # ℹ 3 more variables: middle_name <chr>, last_name <chr>, org_name <chr>
enrollees(state = "GA")
#> # A tibble: 5,000 × 10
#>    npi        pac    enid  specialty_code specialty_description state first_name
#>    <chr>      <chr>  <chr> <chr>          <chr>                 <chr> <chr>     
#>  1 1245249333 99310… I200… 14-08          PRACTITIONER - FAMIL… GA    MICHELLE  
#>  2 1346396462 27693… I200… 14-97          PRACTITIONER - PHYSI… GA    ROBERT    
#>  3 1356345136 07491… I200… 14-93          PRACTITIONER - EMERG… GA    RONALD    
#>  4 1508866765 80229… I200… 14-48          PRACTITIONER - PODIA… GA    LISSETTE  
#>  5 1942384953 37794… I200… 14-08          PRACTITIONER - FAMIL… GA    STANLEY   
#>  6 1699718056 70128… I200… 14-01          PRACTITIONER - GENER… GA    JOHN      
#>  7 1881634483 67097… I200… 14-08          PRACTITIONER - FAMIL… GA    BRUCE     
#>  8 1609936566 47895… I200… 14-08          PRACTITIONER - FAMIL… GA    ADOLFO    
#>  9 1497796718 29613… I200… 14-93          PRACTITIONER - EMERG… GA    REGINALD  
#> 10 1679547061 04461… I200… 14-08          PRACTITIONER - FAMIL… GA    CHARLES   
#> # ℹ 4,990 more rows
#> # ℹ 3 more variables: middle_name <chr>, last_name <chr>, org_name <chr>