Load a group of endpoints by alias

Load a group of endpoints by alias

Usage

group(alias, description = NULL, call = caller_env())

Arguments

alias

<chr> Alias representing the CMS data endpoint.

description

<chr> Group description. Defaults to NULL, which will use the aliases as the description.

call

<env> Environment from which to call the function.

Value

S7 <class_group> object.

Examples

group(c("asc_facility", "enterprise"))
#> <class_group>
#>  @ name   : chr "[asc_facility, enterprise]"
#>  @ members:List of 2
#>  .. $ asc_facility: <class_prov>
#>  ..  ..@ access: <class_current>
#>  .. .. .. @ identifier: chr "https://data.cms.gov/provider-data/api/1/datastore/query/4jcv-atw7/0?count=true&results=true&offset=0&limit=1"
#>  .. .. .. @ metadata  :List of 9
#>  .. .. .. .. $ title      : chr "Ambulatory Surgical Center Quality Measures - Facility"
#>  .. .. .. .. $ description: chr "A list of ambulatory surgical centers participating in the Ambulatory Surgical Center Quality Reporting (ASCQR)"| __truncated__
#>  .. .. .. .. $ modified   : Date[1:1], format: "2025-04-08"
#>  .. .. .. .. $ group      : chr "Hospitals"
#>  .. .. .. .. $ issued     : Date[1:1], format: "2024-10-10"
#>  .. .. .. .. $ released   : Date[1:1], format: "2025-04-30"
#>  .. .. .. .. $ download   : chr "https://data.cms.gov/provider-data/sites/default/files/resources/dd03994fc93e296bb0297f1cd43cc987_1744668308/ASC_Facility.csv"
#>  .. .. .. .. $ dictionary : chr "https://data.cms.gov/provider-data/dataset/4jcv-atw7#data-dictionary"
#>  .. .. .. .. $ site       : chr "https://data.cms.gov/provider-data/dataset/4jcv-atw7"
#>  .. .. .. @ dimensions: <class_dimensions>
#>  .. .. .. .. @ limit : int 1500
#>  .. .. .. .. @ rows  : int 11288
#>  .. .. .. .. @ pages : int 8
#>  .. .. .. .. @ fields: chr [1:50] "facility_name" "facility_id" "npi" "citytown" ...
#>  .. $ enterprise  : <class_caid>
#>  ..  ..@ access: <class_current>
#>  .. .. .. @ identifier: chr "https://data.medicaid.gov/api/1/datastore/query/13a06cdb-6bbb-4f86-bba7-9d6f3db41090/0?count=true&results=true&"| __truncated__
#>  .. .. .. @ metadata  :List of 3
#>  .. .. .. .. $ title      : chr "Medicaid Enterprise System Datatable"
#>  .. .. .. .. $ description: chr "Medicaid Enterprise System Datatable"
#>  .. .. .. .. $ modified   : Date[1:1], format: "2025-05-21"
#>  .. .. .. @ dimensions: <class_dimensions>
#>  .. .. .. .. @ limit : int 8000
#>  .. .. .. .. @ rows  : int 68
#>  .. .. .. .. @ pages : int 1
#>  .. .. .. .. @ fields: chr [1:13] "implementation_year" "policyupdate_type" "policyupdate_title" "potential_mes_impact" ...
try(group("asc_facility"))
#> Error in eval(expr, envir) : 
#>   A <class_group> must have more than one `alias`.
#> ℹ Run `endpoint("asc_facility")` to load this endpoint.