Create Groups of Endpoints

Create Groups of Endpoints

Usage

new_group(
  member_names,
  group_name = NULL,
  quick_call = FALSE,
  offset = 0,
  limit = 10
)

pro_group(alias, ...)

Arguments

member_names

<chr> Alias representing the CMS data endpoint.

group_name

<chr> Name of the group.

quick_call

<lgl> call quick_() method on group if TRUE.

offset

<int> Offset for pagination.

limit

<int> Limit for pagination.

alias

<chr> endpoint alias

...

Additional arguments passed to the group constructor

Value

S7 class_group object.

Examples

new_group(c("qhp_bus_rule_variables", "managed_care_share"))
#> <class_group>
#>  @ group  : chr "qhp_bus_rule_variables | managed_care_share"
#>  @ members:List of 2
#>  .. $ qhp_bus_rule_variables: <hgov_endpoint>
#>  ..  ..@ identifier: chr "https://data.healthcare.gov/api/1/datastore/query/3yfc-yz8a/0"
#>  ..  ..@ metadata  :List of 5
#>  .. .. .. $ title      : chr "QHP Landscape Health Plan Business Rule Variables"
#>  .. .. .. $ description: chr "QHP Landscape Health Plan Business Rule Variables"
#>  .. .. .. $ modified   : Date[1:1], format: "2014-04-16"
#>  .. .. .. $ issued     : Date[1:1], format: "2013-11-26"
#>  .. .. .. $ download   : chr "https://data.healthcare.gov/datafile/qhp-landscape-health-plan-business-rule-variables.3yfc-yz8a.csv"
#>  ..  ..@ dimensions: <class_dimensions>
#>  .. .. .. @ limit : int 500
#>  .. .. .. @ rows  : int 5891
#>  .. .. .. @ pages : int 12
#>  .. .. .. @ fields:List of 19
#>  .. .. .. .. $ standard_component_id                                           : chr(0) 
#>  .. .. .. .. $ issuer_id                                                       : chr(0) 
#>  .. .. .. .. $ issuer_legal_name                                               : chr(0) 
#>  .. .. .. .. $ state                                                           : chr(0) 
#>  .. .. .. .. $ metal_level                                                     : chr(0) 
#>  .. .. .. .. $ market_coverage                                                 : chr(0) 
#>  .. .. .. .. $ on_or_off_exchange                                              : chr(0) 
#>  .. .. .. .. $ sadp_or_qhp                                                     : chr(0) 
#>  .. .. .. .. $ child_only_offering                                             : chr(0) 
#>  .. .. .. .. $ guaranteed_or_estimated_rate                                    : chr(0) 
#>  .. .. .. .. $ how_are_rates_for_contracts_covering_two_or_more_enrollees__e264: chr(0) 
#>  .. .. .. .. $ what_are_the_maximum_number_of_under_age_under_21_dependent_364a: chr(0) 
#>  .. .. .. .. $ what_are_the_maximum_number_of_under_age_under_21_dependent_e6fb: chr(0) 
#>  .. .. .. .. $ is_there_a_maximum_age_for_a_dependent                          : chr(0) 
#>  .. .. .. .. $ what_are_the_maximum_number_of_children_used_to_quote_a_chi_733a: chr(0) 
#>  .. .. .. .. $ are_domestic_partners_treated_the_same_as_secondary_subscribers : chr(0) 
#>  .. .. .. .. $ are_samesex_partners_treated_the_same_as_secondary_subscribers  : chr(0) 
#>  .. .. .. .. $ how_is_age_determined_for_rating_and_eligibility_purposes       : chr(0) 
#>  .. .. .. .. $ what_relationships_between_primary_and_dependent_are_allowe_e280: chr(0) 
#>  .. $ managed_care_share    : <caid_endpoint>
#>  ..  ..@ identifier: chr "https://data.medicaid.gov/api/1/datastore/query/79692ea5-21e1-56bf-8149-97d437120c4b/0"
#>  ..  ..@ metadata  :List of 6
#>  .. .. .. $ title      : chr "Share of Medicaid Enrollees in Managed Care"
#>  .. .. .. $ description: chr "The Share of Medicaid Enrollees in any Managed Care and in Comprehensive Managed CaAre profiles state-level enr"| __truncated__
#>  .. .. .. $ modified   : Date[1:1], format: "2024-10-16"
#>  .. .. .. $ periodicity: chr "Decennially [R/P10Y]"
#>  .. .. .. $ download   : chr "https://data.medicaid.gov/sites/default/files/uploaded_resources/share-of-medicaid-enrollees-in-managed-care.2022.csv"
#>  .. .. .. $ dictionary : chr "https://data.medicaid.gov/api/1/metastore/schemas/data-dictionary/items/71030135-eef9-46b2-9d13-e401008e0fc6"
#>  ..  ..@ dimensions: <class_dimensions>
#>  .. .. .. @ limit : int 8000
#>  .. .. .. @ rows  : int 399
#>  .. .. .. @ pages : int 1
#>  .. .. .. @ fields:List of 8
#>  .. .. .. .. $ state                                          : chr(0) 
#>  .. .. .. .. $ notes                                          : chr(0) 
#>  .. .. .. .. $ total_medicaid_enrollees                       : chr(0) 
#>  .. .. .. .. $ individuals_enrolled_any                       : chr(0) 
#>  .. .. .. .. $ percent_of_all_medicaid_enrollees_any          : chr(0) 
#>  .. .. .. .. $ individuals_enrolled_comprehensive             : chr(0) 
#>  .. .. .. .. $ percent_of_all_medicaid_enrollees_comprehensive: chr(0) 
#>  .. .. .. .. $ year                                           : chr(0) 
new_group(c("qhp_bus_rule_variables", "managed_care_share"), quick_call = TRUE)
#> $qhp_bus_rule_variables
#> # A tibble: 10 × 19
#>    standard_component_id issuer_id issuer_legal_name           state metal_level
#>    <chr>                 <chr>     <chr>                       <chr> <chr>      
#>  1 10191NJ0050001        10191     Freelancers CO-OP of New J… NJ    Silver     
#>  2 10191NJ0040002        10191     Freelancers CO-OP of New J… NJ    Silver     
#>  3 10191NJ0070004        10191     Freelancers CO-OP of New J… NJ    Silver     
#>  4 10191NJ0050003        10191     Freelancers CO-OP of New J… NJ    Platinum   
#>  5 10191NJ0120001        10191     Freelancers CO-OP of New J… NJ    Silver     
#>  6 10191NJ0140002        10191     Freelancers CO-OP of New J… NJ    Silver     
#>  7 10191NJ0050002        10191     Freelancers CO-OP of New J… NJ    Gold       
#>  8 10191NJ0070003        10191     Freelancers CO-OP of New J… NJ    Gold       
#>  9 10191NJ0070001        10191     Freelancers CO-OP of New J… NJ    Bronze     
#> 10 10191NJ0100002        10191     Freelancers CO-OP of New J… NJ    Silver     
#> # ℹ 14 more variables: market_coverage <chr>, on_or_off_exchange <chr>,
#> #   sadp_or_qhp <chr>, child_only_offering <chr>,
#> #   guaranteed_or_estimated_rate <chr>,
#> #   how_are_rates_for_contracts_covering_two_or_more_enrollees__e264 <chr>,
#> #   what_are_the_maximum_number_of_under_age_under_21_dependent_364a <chr>,
#> #   what_are_the_maximum_number_of_under_age_under_21_dependent_e6fb <chr>,
#> #   is_there_a_maximum_age_for_a_dependent <chr>, …
#> 
#> $managed_care_share
#> # A tibble: 10 × 8
#>    state                notes      total_medicaid_enrol…¹ individuals_enrolled…²
#>    <chr>                <chr>      <chr>                  <chr>                 
#>  1 TOTALS               NA         80,263,839             65,034,032            
#>  2 Arkansas5            Arkansas … 1,026,621              512,041               
#>  3 District of Columbia NA         251,791                180,942               
#>  4 Wyoming              NA         64,442                 391                   
#>  5 Pennsylvania         NA         2,753,618              2,521,421             
#>  6 Massachusetts        NA         1,889,306              1,268,120             
#>  7 North Dakota         NA         93,422                 48,621                
#>  8 Michigan             NA         4,448,582              4,370,138             
#>  9 Virginia             NA         1,111,999              761,019               
#> 10 Hawaii               NA         358,302                354,289               
#> # ℹ abbreviated names: ¹​total_medicaid_enrollees, ²​individuals_enrolled_any
#> # ℹ 4 more variables: percent_of_all_medicaid_enrollees_any <chr>,
#> #   individuals_enrolled_comprehensive <chr>,
#> #   percent_of_all_medicaid_enrollees_comprehensive <chr>, year <chr>
#> 
pro_group("pro_dialysis")
#> <class_group>
#>  @ group  : chr "Dialysis Facilities"
#>  @ members:List of 4
#>  .. $ dialysis_facility  : <care_temporal>
#>  ..  ..@ metadata  :List of 5
#>  .. .. .. $ title      : chr "Medicare Dialysis Facilities"
#>  .. .. .. $ description: chr "The Medicare Dialysis Facilities data provides information on clinical and patient measures for Medicare-certif"| __truncated__
#>  .. .. .. $ periodicity: chr "Annually [R/P1Y]"
#>  .. .. .. $ dictionary : chr "https://data.cms.gov/sites/default/files/2024-02/FY2024%20Data%20Dictionary.xlsx"
#>  .. .. .. $ site       : chr "https://data.cms.gov/quality-of-care/medicare-dialysis-facilities"
#>  ..  ..@ dimensions: <class_dimensions>
#>  .. .. .. @ limit : int 5000
#>  .. .. .. @ rows  : int 11679243
#>  .. .. .. @ pages : int 2336
#>  .. .. .. @ fields:List of 14
#>  .. .. .. .. $ state         : chr(0) 
#>  .. .. .. .. $ CCN           : chr(0) 
#>  .. .. .. .. $ Provider_Name : chr(0) 
#>  .. .. .. .. $ city          : chr(0) 
#>  .. .. .. .. $ Ownership_Type: chr(0) 
#>  .. .. .. .. $ ESRD_Network  : chr(0) 
#>  .. .. .. .. $ NPI           : chr(0) 
#>  .. .. .. .. $ Chain         : chr(0) 
#>  .. .. .. .. $ Modality      : chr(0) 
#>  .. .. .. .. $ Alternate_CCNs: chr(0) 
#>  .. .. .. .. $ Measure       : chr(0) 
#>  .. .. .. .. $ Measure_Score : chr(0) 
#>  .. .. .. .. $ year          : chr(0) 
#>  .. .. .. .. $ Measure_ID    : chr(0) 
#>  ..  ..@ endpoints : tibble [5 × 6] (S3: tbl_df/tbl/data.frame)
#>  $ year      : int [1:5] 2024 2023 2022 2021 2020
#>  $ modified  : Date[1:5], format: "2024-02-29" "2023-03-06" ...
#>  $ temporal  : chr [1:5] "2024-01-01 • 2024-12-31" "2023-01-01 • 2023-12-31" "2022-01-01 • 2022-12-31" "2021-01-01 • 2021-12-31" ...
#>  $ identifier: chr [1:5] "https://data.cms.gov/data-api/v1/dataset/62d125ff-c959-4c5c-8a46-8cacba50bfd4/data" "https://data.cms.gov/data-api/v1/dataset/200e7e7e-a5a6-4450-ba18-ab53d237cc45/data" "https://data.cms.gov/data-api/v1/dataset/ebc64ad1-164c-455a-80ca-cb0753e44a46/data" "https://data.cms.gov/data-api/v1/dataset/e66eaf03-398f-4a8c-9d1b-6aaf25fb6425/data" ...
#>  $ download  : chr [1:5] "https://data.cms.gov/sites/default/files/2023-02/2eeb2aa7-05cd-4b7a-a4b2-62e82da03b96/dfr_facility_socrata_fy2023.csv" "https://data.cms.gov/sites/default/files/2022-02/48802694-7450-47c3-a61f-dad7d392037c/dfr_facility_socrata_fy2022.csv" "https://data.cms.gov/sites/default/files/2021-01/FY_2021_Facility_Level_Dialysis_Facility_Reports.csv" "https://data.cms.gov/sites/default/files/2020-12/7q25-ht2t.csv" ...
#>  $ resources : chr [1:5] "https://data.cms.gov/data-api/v1/dataset-resources/62d125ff-c959-4c5c-8a46-8cacba50bfd4" "https://data.cms.gov/data-api/v1/dataset-resources/200e7e7e-a5a6-4450-ba18-ab53d237cc45" "https://data.cms.gov/data-api/v1/dataset-resources/ebc64ad1-164c-455a-80ca-cb0753e44a46" "https://data.cms.gov/data-api/v1/dataset-resources/e66eaf03-398f-4a8c-9d1b-6aaf25fb6425" ...
#>  .. $ dialysis_byfacility: <pro_endpoint>
#>  ..  ..@ identifier: chr "https://data.cms.gov/provider-data/api/1/datastore/query/23ew-n7w9/0"
#>  ..  ..@ metadata  :List of 9
#>  .. .. .. $ title      : chr "Dialysis Facility - Listing by Facility"
#>  .. .. .. $ description: chr "A list of all dialysis facilities registered with Medicare that includes addresses and phone numbers, as well a"| __truncated__
#>  .. .. .. $ modified   : Date[1:1], format: "2025-03-24"
#>  .. .. .. $ group      : chr "Dialysis facilities"
#>  .. .. .. $ issued     : Date[1:1], format: "2020-03-14"
#>  .. .. .. $ released   : Date[1:1], format: "2025-04-23"
#>  .. .. .. $ download   : chr "https://data.cms.gov/provider-data/sites/default/files/resources/c04d84bc5c641284494bee4f20f17f9c_1742846704/DFC_FACILITY.csv"
#>  .. .. .. $ dictionary : chr "https://data.cms.gov/provider-data/dataset/23ew-n7w9#data-dictionary"
#>  .. .. .. $ site       : chr "https://data.cms.gov/provider-data/dataset/23ew-n7w9"
#>  ..  ..@ dimensions: <class_dimensions>
#>  .. .. .. @ limit : int 2000
#>  .. .. .. @ rows  : int 7556
#>  .. .. .. @ pages : int 4
#>  .. .. .. @ fields:List of 142
#>  .. .. .. .. $ cms_certification_number_ccn                                    : chr(0) 
#>  .. .. .. .. $ network                                                         : chr(0) 
#>  .. .. .. .. $ facility_name                                                   : chr(0) 
#>  .. .. .. .. $ five_star_date                                                  : chr(0) 
#>  .. .. .. .. $ five_star                                                       : chr(0) 
#>  .. .. .. .. $ five_star_data_availability_code                                : chr(0) 
#>  .. .. .. .. $ address_line_1                                                  : chr(0) 
#>  .. .. .. .. $ address_line_2                                                  : chr(0) 
#>  .. .. .. .. $ citytown                                                        : chr(0) 
#>  .. .. .. .. $ state                                                           : chr(0) 
#>  .. .. .. .. $ zip_code                                                        : chr(0) 
#>  .. .. .. .. $ countyparish                                                    : chr(0) 
#>  .. .. .. .. $ telephone_number                                                : chr(0) 
#>  .. .. .. .. $ profit_or_nonprofit                                             : chr(0) 
#>  .. .. .. .. $ chain_owned                                                     : chr(0) 
#>  .. .. .. .. $ chain_organization                                              : chr(0) 
#>  .. .. .. .. $ late_shift                                                      : chr(0) 
#>  .. .. .. .. $ of_dialysis_stations                                            : chr(0) 
#>  .. .. .. .. $ offers_incenter_hemodialysis                                    : chr(0) 
#>  .. .. .. .. $ offers_peritoneal_dialysis                                      : chr(0) 
#>  .. .. .. .. $ offers_home_hemodialysis_training                               : chr(0) 
#>  .. .. .. .. $ certification_date                                              : chr(0) 
#>  .. .. .. .. $ claims_date                                                     : chr(0) 
#>  .. .. .. .. $ eqrs_date                                                       : chr(0) 
#>  .. .. .. .. $ smr_date                                                        : chr(0) 
#>  .. .. .. .. $ patient_survival_category_text                                  : chr(0) 
#>  .. .. .. .. $ patient_survival_data_availability_code                         : chr(0) 
#>  .. .. .. .. $ number_of_patients_included_in_survival_summary                 : chr(0) 
#>  .. .. .. .. $ mortality_rate_facility                                         : chr(0) 
#>  .. .. .. .. $ mortality_rate_upper_confidence_limit_975                       : chr(0) 
#>  .. .. .. .. $ mortality_rate_lower_confidence_limit_25                        : chr(0) 
#>  .. .. .. .. $ shr_date                                                        : chr(0) 
#>  .. .. .. .. $ patient_hospitalization_category_text                           : chr(0) 
#>  .. .. .. .. $ patient_hospitalization_data_availability_code                  : chr(0) 
#>  .. .. .. .. $ number_of_patients_included_in_hospitalization_summary          : chr(0) 
#>  .. .. .. .. $ hospitalization_rate_facility                                   : chr(0) 
#>  .. .. .. .. $ hospitalization_rate_upper_confidence_limit_975                 : chr(0) 
#>  .. .. .. .. $ hospitalization_rate_lower_confidence_limit_25                  : chr(0) 
#>  .. .. .. .. $ srr_date                                                        : chr(0) 
#>  .. .. .. .. $ patient_hospital_readmission_category                           : chr(0) 
#>  .. .. .. .. $ patient_hospital_readmission_data_availability_code             : chr(0) 
#>  .. .. .. .. $ number_of_hospitalizations_included_in_hospital_readmission_fc2b: chr(0) 
#>  .. .. .. .. $ readmission_rate_facility                                       : chr(0) 
#>  .. .. .. .. $ readmission_rate_upper_confidence_limit_975                     : chr(0) 
#>  .. .. .. .. $ readmission_rate_lower_confidence_limit_25                      : chr(0) 
#>  .. .. .. .. $ strr_date                                                       : chr(0) 
#>  .. .. .. .. $ patient_transfusion_category_text                               : chr(0) 
#>  .. .. .. .. $ patient_transfusion_data_availability_code                      : chr(0) 
#>  .. .. .. .. $ number_of_patients_included_in_the_transfusion_summary          : chr(0) 
#>  .. .. .. .. $ transfusion_rate_facility                                       : chr(0) 
#>  .. .. .. .. $ transfusion_rate_upper_confidence_limit_975                     : chr(0) 
#>  .. .. .. .. $ transfusion_rate_lower_confidence_limit_25                      : chr(0) 
#>  .. .. .. .. $ swr_date                                                        : chr(0) 
#>  .. .. .. .. $ swr_category_text                                               : chr(0) 
#>  .. .. .. .. $ patient_transplant_waitlist_data_availability_code              : chr(0) 
#>  .. .. .. .. $ number_of_patients_in_this_facility_for_swr                     : chr(0) 
#>  .. .. .. .. $ standardized_first_kidney_transplant_waitlist_ratio             : chr(0) 
#>  .. .. .. .. $ 95_ci_upper_limit_for_swr                                       : chr(0) 
#>  .. .. .. .. $ 95_ci_lower_limit_for_swr                                       : chr(0) 
#>  .. .. .. .. $ pppw_category_text                                              : chr(0) 
#>  .. .. .. .. $ patient_prevalent_transplant_waitlist_data_availability_code    : chr(0) 
#>  .. .. .. .. $ number_of_patients_for_pppw                                     : chr(0) 
#>  .. .. .. .. $ percentage_of_prevalent_patients_waitlisted                     : chr(0) 
#>  .. .. .. .. $ 95_ci_upper_limit_for_pppw                                      : chr(0) 
#>  .. .. .. .. $ 95_ci_lower_limit_for_pppw                                      : chr(0) 
#>  .. .. .. .. $ sedr_date                                                       : chr(0) 
#>  .. .. .. .. $ sedr_category_text                                              : chr(0) 
#>  .. .. .. .. $ emergency_department_encounter_data_availability_code           : chr(0) 
#>  .. .. .. .. $ number_of_patients_included_in_sedr_summary                     : chr(0) 
#>  .. .. .. .. $ standardized_ed_visits_ratio_facility                           : chr(0) 
#>  .. .. .. .. $ sedr_upper_confidence_limit_975                                 : chr(0) 
#>  .. .. .. .. $ sedr_lower_confidence_limit_25                                  : chr(0) 
#>  .. .. .. .. $ ed30_date                                                       : chr(0) 
#>  .. .. .. .. $ ed30_category_text                                              : chr(0) 
#>  .. .. .. .. $ emergency_department_encounter_ratio_occurring_within_30_da_f8e3: chr(0) 
#>  .. .. .. .. $ number_of_hospitalizations_included_in_ed30_summary             : chr(0) 
#>  .. .. .. .. $ standardized_ed_visits_within_30_days_of_hospital_discharge_6307: chr(0) 
#>  .. .. .. .. $ ed30_upper_confidence_limit_975                                 : chr(0) 
#>  .. .. .. .. $ ed30_lower_confidence_limit_25                                  : chr(0) 
#>  .. .. .. .. $ years_modality_switch_based_upon                                : chr(0) 
#>  .. .. .. .. $ smosr_classification_category_facility                          : chr(0) 
#>  .. .. .. .. $ modality_switch_data_availability_code                          : chr(0) 
#>  .. .. .. .. $ smosr_number_of_eligible_patients_facility                      : chr(0) 
#>  .. .. .. .. $ smosr_standardized_modality_switch_ratio_facility               : chr(0) 
#>  .. .. .. .. $ smosr_upper_confidence_limit_facility                           : chr(0) 
#>  .. .. .. .. $ smosr_lower_confidence_limit_facility                           : chr(0) 
#>  .. .. .. .. $ sir_date                                                        : chr(0) 
#>  .. .. .. .. $ patient_infection_category_text                                 : chr(0) 
#>  .. .. .. .. $ patient_infection_data_availability_code                        : chr(0) 
#>  .. .. .. .. $ standard_infection_ratio                                        : chr(0) 
#>  .. .. .. .. $ sir_upper_confidence_limit_975                                  : chr(0) 
#>  .. .. .. .. $ sir_lower_confidence_limit_25                                   : chr(0) 
#>  .. .. .. .. $ fistula_category_text                                           : chr(0) 
#>  .. .. .. .. $ fistula_data_availability_code                                  : chr(0) 
#>  .. .. .. .. $ number_of_patients_included_in_fistula_summary                  : chr(0) 
#>  .. .. .. .. $ fistula_rate_facility                                           : chr(0) 
#>  .. .. .. .. $ fistula_rate_upper_confidence_limit_975                         : chr(0) 
#>  .. .. .. .. $ fistula_rate_lower_confidence_limit_25                          : chr(0) 
#>  .. .. .. .. $ hcp_vaccination_data_collection_dates                           : chr(0) 
#>  .. .. .. ..  [list output truncated]
#>  .. $ dialysis_national  : <pro_endpoint>
#>  ..  ..@ identifier: chr "https://data.cms.gov/provider-data/api/1/datastore/query/2rkq-ygai/0"
#>  ..  ..@ metadata  :List of 9
#>  .. .. .. $ title      : chr "Dialysis Facility - National Averages"
#>  .. .. .. $ description: chr "National averages of common dialysis quality measures."
#>  .. .. .. $ modified   : Date[1:1], format: "2025-03-24"
#>  .. .. .. $ group      : chr "Dialysis facilities"
#>  .. .. .. $ issued     : Date[1:1], format: "2020-03-14"
#>  .. .. .. $ released   : Date[1:1], format: "2025-04-23"
#>  .. .. .. $ download   : chr "https://data.cms.gov/provider-data/sites/default/files/resources/ae019f56e76839ee1da1b51f502954cc_1742846706/DFC_NATIONAL.csv"
#>  .. .. .. $ dictionary : chr "https://data.cms.gov/provider-data/dataset/2rkq-ygai#data-dictionary"
#>  .. .. .. $ site       : chr "https://data.cms.gov/provider-data/dataset/2rkq-ygai"
#>  ..  ..@ dimensions: <class_dimensions>
#>  .. .. .. @ limit : int 2000
#>  .. .. .. @ rows  : int 1
#>  .. .. .. @ pages : int 1
#>  .. .. .. @ fields:List of 59
#>  .. .. .. .. $ country                                                         : chr(0) 
#>  .. .. .. .. $ transfusion_rate_us                                             : chr(0) 
#>  .. .. .. .. $ transfusions__better_than_expected_us                           : chr(0) 
#>  .. .. .. .. $ transfusions__as_expected_us                                    : chr(0) 
#>  .. .. .. .. $ transfusions__worse_than_expected_us                            : chr(0) 
#>  .. .. .. .. $ percentage_of_adult_hd_patients_with_ktv12                      : chr(0) 
#>  .. .. .. .. $ percentage_of_adult_pd_patients_with_ktv17                      : chr(0) 
#>  .. .. .. .. $ percentage_of_pediatric_hd_patients_with_ktv12                  : chr(0) 
#>  .. .. .. .. $ percentage_of_pediatric_pd_patients_with_ktv18                  : chr(0) 
#>  .. .. .. .. $ percentage_of_pediatric_hd_patients_with_npcr                   : chr(0) 
#>  .. .. .. .. $ fistula_rate_us                                                 : chr(0) 
#>  .. .. .. .. $ fistula_rate__better_than_expected_us                           : chr(0) 
#>  .. .. .. .. $ fistula_rate__as_expected_us                                    : chr(0) 
#>  .. .. .. .. $ fistula_rate__worse_than_expected_us                            : chr(0) 
#>  .. .. .. .. $ percentage_of_adult_patients_with_long_term_catheter_in_use     : chr(0) 
#>  .. .. .. .. $ percentage_of_adult_patients_with_hypercalcemia_serum_calci_044d: chr(0) 
#>  .. .. .. .. $ percentage_of_adult_patients_with_serum_phosphorus_less_tha_c222: chr(0) 
#>  .. .. .. .. $ percentage_of_adult_patients_with_serum_phosphorus_between__85e8: chr(0) 
#>  .. .. .. .. $ percentage_of_adult_patients_with_serum_phosphorus_between__fad7: chr(0) 
#>  .. .. .. .. $ percentage_of_adult_patients_with_serum_phosphorus_between__ff32: chr(0) 
#>  .. .. .. .. $ percentage_of_adult_patients_with_serum_phosphorus_greater__d8e3: chr(0) 
#>  .. .. .. .. $ hospitalization_rate_us                                         : chr(0) 
#>  .. .. .. .. $ hospitalizations__better_than_expected_us                       : chr(0) 
#>  .. .. .. .. $ hospitalizations__as_expected_us                                : chr(0) 
#>  .. .. .. .. $ hospitalizations__worse_than_expected_us                        : chr(0) 
#>  .. .. .. .. $ readmission_rate_us                                             : chr(0) 
#>  .. .. .. .. $ hospital_readmission__better_than_expected_us                   : chr(0) 
#>  .. .. .. .. $ hospital_readmission__as_expected_us                            : chr(0) 
#>  .. .. .. .. $ hospital__readmission__worse_than_expected_us                   : chr(0) 
#>  .. .. .. .. $ mortality_rate_us                                               : chr(0) 
#>  .. .. .. .. $ survival__better_than_expected_us                               : chr(0) 
#>  .. .. .. .. $ survival__as_expected_us                                        : chr(0) 
#>  .. .. .. .. $ survival__worse_than_expected_us                                : chr(0) 
#>  .. .. .. .. $ standardized_first_kidney_transplant_waitlist_ratio_us          : chr(0) 
#>  .. .. .. .. $ incident_patients_transplant_waitlisting__better_than_expec_a24d: chr(0) 
#>  .. .. .. .. $ incident_patients_transplant_waitlisting__as_expected_us        : chr(0) 
#>  .. .. .. .. $ incident_patients_transplant_waitlisting__worse_than_expected_us: chr(0) 
#>  .. .. .. .. $ percentage_of_prevalent_patients_waitlisted_us                  : chr(0) 
#>  .. .. .. .. $ prevalent_patients_transplant_waitlisting_better_than_expec_c0cd: chr(0) 
#>  .. .. .. .. $ prevalent_patients_transplant_waitlisting__as_expected_us       : chr(0) 
#>  .. .. .. .. $ prevalent_patients_transplant_waitlisting__worse_than_expec_e6e9: chr(0) 
#>  .. .. .. .. $ healthcare_worker_covid19_vaccination_adherence_percentage_us   : chr(0) 
#>  .. .. .. .. $ percentage_of_patients_with_hgb10_gdl_us                        : chr(0) 
#>  .. .. .. .. $ percentage_of_patients_with_hgb12_gdl_us                        : chr(0) 
#>  .. .. .. .. $ infection__better_than_expected_us                              : chr(0) 
#>  .. .. .. .. $ infection__as_expected_us                                       : chr(0) 
#>  .. .. .. .. $ infection__worse_than_expected_us                               : chr(0) 
#>  .. .. .. .. $ standardized_ed_visits_ratio_us                                 : chr(0) 
#>  .. .. .. .. $ sedr__better_than_expected_us                                   : chr(0) 
#>  .. .. .. .. $ sedr__as_expected_us                                            : chr(0) 
#>  .. .. .. .. $ sedr__worse_than_expected_us                                    : chr(0) 
#>  .. .. .. .. $ standardized_ed_visits_within_30_days_of_hospital_discharge_d0c6: chr(0) 
#>  .. .. .. .. $ ed30__better_than_expected_us                                   : chr(0) 
#>  .. .. .. .. $ ed30__as_expected_us                                            : chr(0) 
#>  .. .. .. .. $ ed30__worse_than_expected_us                                    : chr(0) 
#>  .. .. .. .. $ standardized_modality_switch_ratio                              : chr(0) 
#>  .. .. .. .. $ n_for_smsr_category_1_smsrbetter_than_expected_us               : chr(0) 
#>  .. .. .. .. $ n_for_smsr_category_2_smsras_expected_us                        : chr(0) 
#>  .. .. .. .. $ n_for_smsr_category_3_smsrworse_than_expected_us                : chr(0) 
#>  .. $ dialysis_state     : <pro_endpoint>
#>  ..  ..@ identifier: chr "https://data.cms.gov/provider-data/api/1/datastore/query/2fpu-cgbb/0"
#>  ..  ..@ metadata  :List of 9
#>  .. .. .. $ title      : chr "Dialysis Facility - State Averages"
#>  .. .. .. $ description: chr "State averages of common dialysis quality measures."
#>  .. .. .. $ modified   : Date[1:1], format: "2025-03-24"
#>  .. .. .. $ group      : chr "Dialysis facilities"
#>  .. .. .. $ issued     : Date[1:1], format: "2020-03-14"
#>  .. .. .. $ released   : Date[1:1], format: "2025-04-23"
#>  .. .. .. $ download   : chr "https://data.cms.gov/provider-data/sites/default/files/resources/3b5f0449f531ee07c467e49abed89954_1742846705/DFC_STATE.csv"
#>  .. .. .. $ dictionary : chr "https://data.cms.gov/provider-data/dataset/2fpu-cgbb#data-dictionary"
#>  .. .. .. $ site       : chr "https://data.cms.gov/provider-data/dataset/2fpu-cgbb"
#>  ..  ..@ dimensions: <class_dimensions>
#>  .. .. .. @ limit : int 2000
#>  .. .. .. @ rows  : int 56
#>  .. .. .. @ pages : int 1
#>  .. .. .. @ fields:List of 49
#>  .. .. .. .. $ state                                                           : chr(0) 
#>  .. .. .. .. $ survival_better_than_expected_state                             : chr(0) 
#>  .. .. .. .. $ survival_as_expected_state                                      : chr(0) 
#>  .. .. .. .. $ survival_worse_than_expected_state                              : chr(0) 
#>  .. .. .. .. $ hospitalizations_better_than_expected_state                     : chr(0) 
#>  .. .. .. .. $ hospitalizations_as_expected_state                              : chr(0) 
#>  .. .. .. .. $ hospitalizations_worse_than_expected_state                      : chr(0) 
#>  .. .. .. .. $ hospital_readmission__better_than_expected_state                : chr(0) 
#>  .. .. .. .. $ hospital_readmission__as_expected_state                         : chr(0) 
#>  .. .. .. .. $ hospital__readmission__worse_than_expected_state                : chr(0) 
#>  .. .. .. .. $ transfusions_better_than_expected_state                         : chr(0) 
#>  .. .. .. .. $ transfusions_as_expected_state                                  : chr(0) 
#>  .. .. .. .. $ transfusions_worse_than_expected_state                          : chr(0) 
#>  .. .. .. .. $ infection_better_than_expected_state                            : chr(0) 
#>  .. .. .. .. $ infection_as_expected_state                                     : chr(0) 
#>  .. .. .. .. $ infection_worse_than_expected_state                             : chr(0) 
#>  .. .. .. .. $ percentage_of_adult_hd_patients_with_ktv12                      : chr(0) 
#>  .. .. .. .. $ percentage_of_adult_pd_patients_with_ktv17                      : chr(0) 
#>  .. .. .. .. $ percentage_of_pediatric_hd_patients_with_ktv12                  : chr(0) 
#>  .. .. .. .. $ percentage_of_pediatric_pd_patients_with_ktv18                  : chr(0) 
#>  .. .. .. .. $ percentage_of_pediatric_hd_patients_with_npcr_in_use            : chr(0) 
#>  .. .. .. .. $ fistula_rate__better_than_expected_state                        : chr(0) 
#>  .. .. .. .. $ fistula_rate__as_expected_state                                 : chr(0) 
#>  .. .. .. .. $ fistula_rate__worse_than_expected_state                         : chr(0) 
#>  .. .. .. .. $ percentage_of_adult_patients_with_long_term_catheter_in_use     : chr(0) 
#>  .. .. .. .. $ percentage_of_adult_patients_with_hypercalcemia_serum_calci_044d: chr(0) 
#>  .. .. .. .. $ percentage_of_adult_patients_with_serum_phosphorus_less_tha_c222: chr(0) 
#>  .. .. .. .. $ percentage_of_adult_patients_with_serum_phosphorus_between__85e8: chr(0) 
#>  .. .. .. .. $ percentage_of_adult_patients_with_serum_phosphorus_between__fad7: chr(0) 
#>  .. .. .. .. $ percentage_of_adult_patients_with_serum_phosphorus_between__ff32: chr(0) 
#>  .. .. .. .. $ percentage_of_adult_patients_with_serum_phosphorus_greater__d8e3: chr(0) 
#>  .. .. .. .. $ incident_patients_transplant_waitlisting_better_than_expect_49e4: chr(0) 
#>  .. .. .. .. $ incident_patients_transplant_waitlisting__as_expected_state     : chr(0) 
#>  .. .. .. .. $ incident_patients_transplant_waitlisting__worse_than_expect_9b13: chr(0) 
#>  .. .. .. .. $ prevalent_patients_transplant_waitlisting_better_than_expec_c046: chr(0) 
#>  .. .. .. .. $ prevalent_patients_transplant_waitlisting__as_expected_state    : chr(0) 
#>  .. .. .. .. $ prevalent_patients_transplant_waitlisting__worse_than_expec_eeb0: chr(0) 
#>  .. .. .. .. $ healthcare_worker_covid19_vaccination_adherence_percentage_state: chr(0) 
#>  .. .. .. .. $ percentage_of_patients_with_hgb10_gdl_state                     : chr(0) 
#>  .. .. .. .. $ percentage_of_patients_with_hgb12_gdl_state                     : chr(0) 
#>  .. .. .. .. $ sedr__better_than_expected_state                                : chr(0) 
#>  .. .. .. .. $ sedr__as_expected_state                                         : chr(0) 
#>  .. .. .. .. $ sedr__worse_than_expected_state                                 : chr(0) 
#>  .. .. .. .. $ ed30__better_than_expected_state                                : chr(0) 
#>  .. .. .. .. $ ed30__as_expected_state                                         : chr(0) 
#>  .. .. .. .. $ ed30__worse_than_expected_state                                 : chr(0) 
#>  .. .. .. .. $ n_for_smsr_category_1_smsrbetter_than_expected_state            : chr(0) 
#>  .. .. .. .. $ n_for_smsr_category_2_smsras_expected_state                     : chr(0) 
#>  .. .. .. .. $ n_for_smsr_category_3_smsrworse_than_expected_state             : chr(0)