Healthcare.Gov API Endpoints

Healthcare.Gov API Endpoints

Usage

hgov_endpoint(alias, call = caller_env())

hgov_temporal(alias, call = caller_env())

Arguments

alias

<chr> endpoint or group alias

call

<env> environment to use for error reporting

Value

An S7 <class_endpoint> or <class_temporal> object

Examples

hgov_endpoint("hgov_ab_reg_comp")
#> <class_endpoint>
#>  @ metadata  :List of 8
#>  .. $ clog       : chr "hgov"
#>  .. $ api        : chr "end"
#>  .. $ title      : chr "AB Registration Completion List"
#>  .. $ description: chr "Access Data Dictionary Here - https://www.cms.gov/CCIIO/Programs-and-Initiatives/Health-Insurance-Marketplaces/"| __truncated__
#>  .. $ modified   : Date[1:1], format: "2025-06-11"
#>  .. $ issued     : Date[1:1], format: "2016-03-30"
#>  .. $ periodicity: chr "Decennially [R/P10Y]"
#>  .. $ download   : chr "https://data.healthcare.gov/sites/default/files/uploaded_resources/TBL_RCL_157.csv"
#>  @ identifier: chr "https://data.healthcare.gov/api/1/datastore/query/wb6u-x2ny/0"
#>  @ dimensions: <class_dimensions>
#>  .. @ limit : int 500
#>  .. @ rows  : int 1022897
#>  .. @ pages : int 2046
#>  .. @ fields:List of 7
#>  .. .. $ npn                                    : chr(0) 
#>  .. .. $ applicable_plan_year                   : chr(0) 
#>  .. .. $ individual_registration_completion_date: chr(0) 
#>  .. .. $ individual_marketplace_end_date        : chr(0) 
#>  .. .. $ shop_registration_completion_date      : chr(0) 
#>  .. .. $ shop_end_date                          : chr(0) 
#>  .. .. $ npn_valid_current_plan_year_only       : chr(0) 
hgov_temporal("hgov_puf_rate")
#> <class_temporal>
#>  @ metadata  :List of 3
#>  .. $ clog : chr "hgov"
#>  .. $ api  : chr "tmp"
#>  .. $ title: chr "Rate PUF"
#>  @ endpoints : tibble [12 × 9] (S3: tbl_df/tbl/data.frame)
#>  $ year       : chr [1:12] "2025" "2024" "2023" "2022" ...
#>  $ identifier : chr [1:12] "https://data.healthcare.gov/api/1/datastore/query/672d5f6a-b8a7-4ebe-87f6-67db641e192d/0" "https://data.healthcare.gov/api/1/datastore/query/2638f347-5311-4655-8d2b-1466848adeae/0" "https://data.healthcare.gov/api/1/datastore/query/e6deb20e-dd27-495b-8eb1-c67db7bbd788/0" "https://data.healthcare.gov/api/1/datastore/query/3edfe68f-b783-4e98-9120-20014dd0fe49/0" ...
#>  $ description: chr [1:12] "The Rate PUF (Rate-PUF) is one of the files that comprise the Health Insurance Exchange Public Use Files. The R"| __truncated__ "The Rate PUF (Rate-PUF) is one of the files that comprise the Health Insurance Exchange Public Use Files. The R"| __truncated__ "The Rate PUF (Rate-PUF) is one of the files that comprise the Health Insurance Exchange Public Use Files. The R"| __truncated__ "The Rate PUF (Rate-PUF) is one of the files that comprise the Health Insurance Exchange Public Use Files. The R"| __truncated__ ...
#>  $ periodicity: chr [1:12] "Annually [R/P1Y]" "Annually [R/P1Y]" "Annually [R/P1Y]" "Annually [R/P1Y]" ...
#>  $ issued     : Date[1:12], format: "2025-04-29" "2024-08-06" ...
#>  $ modified   : Date[1:12], format: "2025-05-06" "2024-08-14" ...
#>  $ contact    : chr [1:12] "Center for Consumer Information and Insurance Oversight (mailto:CMS_FEPS@cms.hhs.gov)" "Center for Consumer Information and Insurance Oversight (mailto:CMS_FEPS@cms.hhs.gov)" "Center for Consumer Information and Insurance Oversight (CMS_FEPS@cms.hhs.gov)" "Center for Consumer Information and Insurance Oversight (CMS_FEPS@cms.hhs.gov)" ...
#>  $ download   : chr [1:12] "https://data.healthcare.gov/datafile/py2025/Rate_PUF.csv" "https://data.healthcare.gov/datafile/py2024/Rate_PUF.csv" "https://data.healthcare.gov/datafile/py2023/rate_puf.csv" "https://data.healthcare.gov/datafile/py2022/rate_puf.csv" ...
#>  $ resources  : chr [1:12] NA NA NA NA ...
#>  @ dimensions: <class_dimensions>
#>  .. @ limit : int 500
#>  .. @ rows  : int 2418441
#>  .. @ pages : int 4837
#>  .. @ fields:List of 20
#>  .. .. $ businessyear                             : chr(0) 
#>  .. .. $ statecode                                : chr(0) 
#>  .. .. $ issuerid                                 : chr(0) 
#>  .. .. $ sourcename                               : chr(0) 
#>  .. .. $ importdate                               : chr(0) 
#>  .. .. $ rateeffectivedate                        : chr(0) 
#>  .. .. $ rateexpirationdate                       : chr(0) 
#>  .. .. $ planid                                   : chr(0) 
#>  .. .. $ ratingareaid                             : chr(0) 
#>  .. .. $ tobacco                                  : chr(0) 
#>  .. .. $ age                                      : chr(0) 
#>  .. .. $ individualrate                           : chr(0) 
#>  .. .. $ individualtobaccorate                    : chr(0) 
#>  .. .. $ couple                                   : chr(0) 
#>  .. .. $ primarysubscriberandonedependent         : chr(0) 
#>  .. .. $ primarysubscriberandtwodependents        : chr(0) 
#>  .. .. $ primarysubscriberandthreeormoredependents: chr(0) 
#>  .. .. $ coupleandonedependent                    : chr(0) 
#>  .. .. $ coupleandtwodependents                   : chr(0) 
#>  .. .. $ coupleandthreeormoredependents           : chr(0)