Load Main API Dataset

Load Main API Dataset

Usage

main_current(alias)

Arguments

alias

<chr> endpoint alias

Value

<Dataset> object

Examples

main_current("enrollees")
#> $title
#> [1] "Medicare Fee-For-Service  Public Provider Enrollment"
#> 
#> $description
#> [1] "The Medicare Fee-For-Service Public Provider Enrollment dataset includes information on providers who are actively approved to bill Medicare or have completed the 855O at the time the data was pulled from the Provider Enrollment, Chain, and Ownership System (PECOS). The release of this provider enrollment data is not related to other provider information releases such as Physician Compare or Data Transparency.. .  . . Note: This full dataset contains more records than most spreadsheet programs can handle, which will result in an incomplete load of data. Use of a database or statistical software is required."
#> 
#> $periodicity
#> [1] "Quarterly [R/P3M]"
#> 
#> $contact
#> [1] "Provider Enrollment Data Requests - CPI (mailto:ProviderEnrollmentDataRequests@cms.hhs.gov)"
#> 
#> $dictionary
#> [1] "https://data.cms.gov/resources/medicare-fee-for-service-public-provider-enrollment-data-dictionary"
#> 
#> $identifier
#> [1] "https://data.cms.gov/data-api/v1/dataset/2457ea29-fc82-48b0-86ec-3b0755de7515/data-viewer"
#> 
#> $modified
#> [1] "2025-01-15"
#> 
#> $site
#> [1] "https://data.cms.gov/provider-characteristics/medicare-provider-supplier-enrollment/medicare-fee-for-service-public-provider-enrollment"
#> 
#> $temporal
#> [1] "2023-01-01/2024-12-31"
#> 
#> $references
#> [1] "https://data.cms.gov/resources/fee-for-service-public-provider-enrollment-methodology"
#> 
main_current("opt_out")
#> $title
#> [1] "Opt Out Affidavits"
#> 
#> $description
#> [1] "ATTENTION USERS. . . Some Providers' Opt-Out Status may end early due to COVID 19 waivers. Please contact your respective MAC for further information.. .  . . For more information on the opt-out process, see Manage Your Enrollment or view the FAQ section below.. .  . . The Opt Out Affidavits dataset provides information on providers who have decided not to participate in Medicare. It contains provider's NPI, specialty, address, and effective dates."
#> 
#> $periodicity
#> [1] "Monthly [R/P1M]"
#> 
#> $contact
#> [1] "Provider Enrollment Data Requests - CPI (mailto:ProviderEnrollmentDataRequests@cms.hhs.gov)"
#> 
#> $dictionary
#> [1] "https://data.cms.gov/resources/opt-out-affidavits-data-dictionary"
#> 
#> $identifier
#> [1] "https://data.cms.gov/data-api/v1/dataset/9887a515-7552-4693-bf58-735c77af46d7/data-viewer"
#> 
#> $modified
#> [1] "2025-02-27"
#> 
#> $site
#> [1] "https://data.cms.gov/provider-characteristics/medicare-provider-supplier-enrollment/opt-out-affidavits"
#> 
#> $temporal
#> [1] "2023-04-01/2025-01-31"
#> 
#> $references
#> [1] "https://data.cms.gov/resources/opt-out-affidavits-methodology"
#> 
main_current("order_refer")
#> $title
#> [1] "Order and Referring"
#> 
#> $description
#> [1] "The Order and Referring dataset provides information on all physicians and non-physician practitioners, by their National Provider Identifier (NPI), who are of a type/specialty that is legally eligible to order and refer in the Medicare program and who have current enrollment records in Medicare. . .  . . Note: This full dataset contains more records than most spreadsheet programs can handle, which will result in an incomplete load of data. Use of a database or statistical software is required."
#> 
#> $periodicity
#> [1] "Twice a Week [R/P3.5D]"
#> 
#> $contact
#> [1] "Provider Enrollment Data Requests - CPI (mailto:ProviderEnrollmentDataRequests@cms.hhs.gov)"
#> 
#> $dictionary
#> [1] "https://data.cms.gov/resources/order-and-referring-data-dictionary"
#> 
#> $identifier
#> [1] "https://data.cms.gov/data-api/v1/dataset/c99b5865-1119-4436-bb80-c5af2773ea1f/data-viewer"
#> 
#> $modified
#> [1] "2025-03-07"
#> 
#> $site
#> [1] "https://data.cms.gov/provider-characteristics/medicare-provider-supplier-enrollment/order-and-referring"
#> 
#> $temporal
#> [1] "2023-05-21/2025-03-01"
#> 
#> $references
#> [1] "https://data.cms.gov/resources/order-and-referring-methodology"
#> 
main_current("reassignments")
#> $title
#> [1] "Revalidation Reassignment List"
#> 
#> $description
#> [1] "The Revalidation Reassignment List dataset provides information on reassignments of providers who are due for revalidation.. .  . . Note: This full dataset contains more records than most spreadsheet programs can handle, which will result in an incomplete load of data. Use of a database or statistical software is required."
#> 
#> $periodicity
#> [1] "Monthly [R/P1M]"
#> 
#> $contact
#> [1] "Provider Enrollment Data Requests - CPI (mailto:ProviderEnrollmentDataRequests@cms.hhs.gov)"
#> 
#> $dictionary
#> [1] "https://data.cms.gov/resources/revalidation-reassignments-list-data-dictionary"
#> 
#> $identifier
#> [1] "https://data.cms.gov/data-api/v1/dataset/20f51cff-4137-4f3a-b6b7-bfc9ad57983b/data-viewer"
#> 
#> $modified
#> [1] "2025-01-02"
#> 
#> $site
#> [1] "https://data.cms.gov/provider-characteristics/medicare-provider-supplier-enrollment/revalidation-reassignment-list"
#> 
#> $temporal
#> [1] "2022-02-01/2025-01-31"
#> 
#> $references
#> [1] "https://data.cms.gov/resources/additional-information-on-revalidation"
#> 
main_current("hospitals")
#> $title
#> [1] "Hospital Enrollments"
#> 
#> $description
#> [1] "The Hospital Enrollments dataset provides enrollment information of all Hospitals currently enrolled in Medicare. This data includes information on the Hospital's sub-group type, legal business name, doing business as name, organization type and address. "
#> 
#> $periodicity
#> [1] "Monthly [R/P1M]"
#> 
#> $contact
#> [1] "Provider Enrollment Data Requests - CPI (mailto:ProviderEnrollmentDataRequests@cms.hhs.gov)"
#> 
#> $dictionary
#> [1] "https://data.cms.gov/sites/default/files/2023-07/Hospital_Enrollments_Data_Dictionary.pdf"
#> 
#> $identifier
#> [1] "https://data.cms.gov/data-api/v1/dataset/f6f6505c-e8b0-4d57-b258-e2b94133aaf2/data-viewer"
#> 
#> $modified
#> [1] "2025-01-15"
#> 
#> $site
#> [1] "https://data.cms.gov/provider-characteristics/hospitals-and-other-facilities/hospital-enrollments"
#> 
#> $temporal
#> [1] "2022-11-01/2025-01-31"
#> 
#> $references
#> [1] "https://data.cms.gov/sites/default/files/2024-10/Hospital_Data_Guidance.pdf"
#> 
main_current("laboratories")
#> $title
#> [1] "Provider of Services File - Clinical Laboratories"
#> 
#> $description
#> [1] "The Provider of Services (POS) Clinical Laboratories (CLIA) data provides information on CLIA demographics and types of testing services the facility provides. In this file you will find provider number, name, address and characteristics of the participating institution providers."
#> 
#> $periodicity
#> [1] "Quarterly [R/P3M]"
#> 
#> $contact
#> [1] "QIESData - CCSQ (mailto:QIESData@cms.hhs.gov)"
#> 
#> $dictionary
#> [1] "https://data.cms.gov/sites/default/files/2023-07/50eacc07-7930-4e62-97e3-e04a2dc1e2a3/P.QWB.POSQ.CLIA.LAYOUT.MAR23.pdf"
#> 
#> $identifier
#> [1] "https://data.cms.gov/data-api/v1/dataset/d3eb38ac-d8e9-40d3-b7b7-6205d3d1dc16/data-viewer"
#> 
#> $modified
#> [1] "2025-01-16"
#> 
#> $site
#> [1] "https://data.cms.gov/provider-characteristics/hospitals-and-other-facilities/provider-of-services-file-clinical-laboratories"
#> 
#> $temporal
#> [1] "2011-10-01/2024-12-31"
#> 
#> $references
#> [1] "https://data.cms.gov/resources/provider-of-services-file-methodology"
#> 
main_current("crosswalk")
#> $title
#> [1] "Medicare Provider and Supplier Taxonomy Crosswalk"
#> 
#> $description
#> [1] "The Medicare Provider and Supplier Taxonomy Crosswalk dataset lists the providers and suppliers eligible to enroll in Medicare programs with the proper healthcare provider taxonomy code. This data includes the Medicare speciality codes, if available, provider/supplier type description, taxonomy code, and the taxonomy description."
#> 
#> $periodicity
#> [1] "Biannually [R/P6M]"
#> 
#> $contact
#> [1] "Provider Enrollment Data Requests - CPI (mailto:ProviderEnrollmentDataRequests@cms.hhs.gov)"
#> 
#> $dictionary
#> [1] "https://data.cms.gov/resources/medicare-provider-and-supplier-taxonomy-crosswalk-data-dictionary"
#> 
#> $identifier
#> [1] "https://data.cms.gov/data-api/v1/dataset/113eb0bc-0c9a-4d91-9f93-3f6b28c0bf6b/data-viewer"
#> 
#> $modified
#> [1] "2025-01-17"
#> 
#> $site
#> [1] "https://data.cms.gov/provider-characteristics/medicare-provider-supplier-enrollment/medicare-provider-and-supplier-taxonomy-crosswalk"
#> 
#> $temporal
#> [1] "2021-07-01/2025-06-30"
#> 
#> $references
#> [1] "https://data.cms.gov/resources/medicare-provider-and-supplier-taxonomy-crosswalk-methodology"
#> 
main_current("rbcs")
#> $title
#> [1] "Restructured BETOS Classification System"
#> 
#> $description
#> [1] "The Restructured BETOS Classification System (RBCS) dataset is a taxonomy that allows researchers to group healthcare service codes for Medicare Part B services (i.e., HCPCS codes) into clinically meaningful categories and subcategories. It is based on the original Berenson-Eggers Type of Service (BETOS) classification created in the 1980s, and includes notable updates such as Part B non-physician services. The RBCS will undergo annual updates by a technical expert panel of researchers and clinicians."
#> 
#> $periodicity
#> [1] "Annually [R/P1Y]"
#> 
#> $contact
#> [1] "PDAG Data Products - OEDA (mailto:PDAG_Data_Products@cms.hhs.gov)"
#> 
#> $dictionary
#> [1] "https://data.cms.gov/resources/rbcs-data-dictionary-2024"
#> 
#> $identifier
#> [1] "https://data.cms.gov/data-api/v1/dataset/e3db6e56-149f-49ce-b374-40aecda2357b/data-viewer"
#> 
#> $modified
#> [1] "2024-10-22"
#> 
#> $site
#> [1] "https://data.cms.gov/provider-summary-by-type-of-service/provider-service-classifications/restructured-betos-classification-system"
#> 
#> $temporal
#> [1] "2024-01-01/2024-12-31"
#> 
#> $references
#> [1] "https://data.cms.gov/sites/default/files/2024-10/2024%20RBCS%20Final%20Report.pdf"
#> 
main_current("rhc")
#> $title
#> [1] "Rural Health Clinic Enrollments"
#> 
#> $description
#> [1] "The Rural Health Clinic (RHC) Enrollments dataset provides enrollment information on all RHCs currently enrolled in Medicare. This data includes information on the RHC's legal business name, doing business as name, organization type and address."
#> 
#> $periodicity
#> [1] "Quarterly [R/P3M]"
#> 
#> $contact
#> [1] "Provider Enrollment Data Requests - CPI (mailto:ProviderEnrollmentDataRequests@cms.hhs.gov)"
#> 
#> $dictionary
#> [1] "https://data.cms.gov/sites/default/files/2023-09/RHC_Enrollments_Data_Dictionary.pdf"
#> 
#> $identifier
#> [1] "https://data.cms.gov/data-api/v1/dataset/3b7e7659-067e-41ea-8e36-f9ee2036e1f6/data-viewer"
#> 
#> $modified
#> [1] "2025-01-15"
#> 
#> $site
#> [1] "https://data.cms.gov/provider-characteristics/hospitals-and-other-facilities/rural-health-clinic-enrollments"
#> 
#> $temporal
#> [1] "2023-10-01/2025-03-31"
#> 
#> $references
#> [1] "https://data.cms.gov/sites/default/files/2024-10/RHC_Data_Guidance.pdf"
#> 
main_current("fqhc")
#> $title
#> [1] "Federally Qualified Health Center Enrollments"
#> 
#> $description
#> [1] "The Federally Qualified Health Center (FQHC) Enrollments dataset provides enrollment information on all FQHCs currently enrolled in Medicare. This data includes information on the FQHC's legal business name, doing business as name, organization type, and address."
#> 
#> $periodicity
#> [1] "Quarterly [R/P3M]"
#> 
#> $contact
#> [1] "Provider Enrollment Data Requests - CPI (mailto:ProviderEnrollmentDataRequests@cms.hhs.gov)"
#> 
#> $dictionary
#> [1] "https://data.cms.gov/sites/default/files/2023-09/FQHC_Enrollments_Data_Dictionary.pdf"
#> 
#> $identifier
#> [1] "https://data.cms.gov/data-api/v1/dataset/4bcae866-3411-439a-b762-90a6187c194b/data-viewer"
#> 
#> $modified
#> [1] "2025-01-15"
#> 
#> $site
#> [1] "https://data.cms.gov/provider-characteristics/hospitals-and-other-facilities/federally-qualified-health-center-enrollments"
#> 
#> $temporal
#> [1] "2023-10-01/2025-03-31"
#> 
#> $references
#> [1] "https://data.cms.gov/sites/default/files/2023-09/FQHC_Data_Guidance.pdf"
#> 
main_current("home_health")
#> $title
#> [1] "Home Health Agency Enrollments"
#> 
#> $description
#> [1] "The Home Health Agency (HHA) Enrollments dataset provides enrollment information on all HHAs currently enrolled in Medicare. This data includes information on the HHA's legal business name, doing business as name, organization type and address. "
#> 
#> $periodicity
#> [1] "Quarterly [R/P3M]"
#> 
#> $contact
#> [1] "Provider Enrollment Data Requests - CPI (mailto:ProviderEnrollmentDataRequests@cms.hhs.gov)"
#> 
#> $dictionary
#> [1] "https://data.cms.gov/sites/default/files/2024-10/HHA_Enrollments_Data_Dictionary.pdf"
#> 
#> $identifier
#> [1] "https://data.cms.gov/data-api/v1/dataset/15f64ab4-3172-4a27-b589-ebd67a6d28aa/data-viewer"
#> 
#> $modified
#> [1] "2025-01-15"
#> 
#> $site
#> [1] "https://data.cms.gov/provider-characteristics/hospitals-and-other-facilities/home-health-agency-enrollments"
#> 
#> $temporal
#> [1] "2023-04-01/2025-03-31"
#> 
#> $references
#> [1] "https://data.cms.gov/sites/default/files/2024-10/HHA_Data_Guidance.pdf"
#> 
main_current("hospice")
#> $title
#> [1] "Hospice Enrollments"
#> 
#> $description
#> [1] "The Hospice Enrollments dataset provides enrollment information of all hospices currently enrolled in Medicare. This data includes information on the hospice's legal business name, doing business as name, organization type and address. "
#> 
#> $periodicity
#> [1] "Quarterly [R/P3M]"
#> 
#> $contact
#> [1] "Provider Enrollment Data Requests - CPI (mailto:ProviderEnrollmentDataRequests@cms.hhs.gov)"
#> 
#> $dictionary
#> [1] "https://data.cms.gov/sites/default/files/2023-03/f736ce50-3d6b-4ee2-802f-2e5424ae71ac/Hospice_Enrollments_Data_Dictionary.pdf"
#> 
#> $identifier
#> [1] "https://data.cms.gov/data-api/v1/dataset/25704213-e833-4b8b-9dbc-58dd17149209/data-viewer"
#> 
#> $modified
#> [1] "2025-01-15"
#> 
#> $site
#> [1] "https://data.cms.gov/provider-characteristics/hospitals-and-other-facilities/hospice-enrollments"
#> 
#> $temporal
#> [1] "2023-04-01/2025-03-31"
#> 
#> $references
#> [1] "https://data.cms.gov/sites/default/files/2023-03/f193e9ef-9fd7-4de3-9a27-0d5281cf80a9/Hospice_Data_Guidance.pdf"
#> 
main_current("snf")
#> $title
#> [1] "Skilled Nursing Facility Enrollments"
#> 
#> $description
#> [1] "The Skilled Nursing Facility (SNF) Enrollments dataset provides enrollment information of all SNF 's currently enrolled in Medicare. This data includes information on the SNF's legal business name, doing business as name, organization type and address.. .  . . On November 17, 2023, CMS published in the Federal Register a final rule titled, “Medicare and Medicaid Programs; Disclosures of Ownership and Additional Disclosable Parties Information for Skilled Nursing Facilities and Nursing Facilities; Medicare Providers’ and Suppliers’ Disclosure of Private Equity Companies and Real Estate Investment Trusts” (88 FR 80141). This final rule implements parts of section 1124(c) of the Act which requires SNFs to disclose detailed information about their ownership and management as well as additional data regarding: (1) other parties with which the SNF is associated; and (2) the ownership structures of these other parties.  Refer to Medicare Enrollment for Providers & Suppliers for more information on the Skilled Nursing Facility disclosure requirements. . . Section 6101(b) of the Affordable Care Act states that no later than 1 year after final regulations promulgated under section 1124(c) of the Act are published in the Federal Register, the Secretary shall make the information reported available to the public.. .  On November 21, 2024 CMS updated this dataset to include this reported information."
#> 
#> $periodicity
#> [1] "Monthly [R/P1M]"
#> 
#> $contact
#> [1] "Provider Enrollment Data Requests - CPI (mailto:ProviderEnrollmentDataRequests@cms.hhs.gov)"
#> 
#> $dictionary
#> [1] "https://data.cms.gov/sites/default/files/2024-10/SNF_Enrollments_Data_Dictionary.pdf"
#> 
#> $identifier
#> [1] "https://data.cms.gov/data-api/v1/dataset/5f2c306f-3b1c-42cd-b037-187b2ce22126/data-viewer"
#> 
#> $modified
#> [1] "2025-01-15"
#> 
#> $site
#> [1] "https://data.cms.gov/provider-characteristics/hospitals-and-other-facilities/skilled-nursing-facility-enrollments"
#> 
#> $temporal
#> [1] "2022-09-01/2025-01-31"
#> 
#> $references
#> [1] "https://data.cms.gov/sites/default/files/2024-12/1ea55414-f441-4694-a13c-6400d4e28ef6/SNF_Data_Guidance.pdf"
#> 
main_current("pending_nonphysicians")
#> $title
#> [1] "Pending Initial Logging and Tracking Non Physicians"
#> 
#> $description
#> [1] "The Pending Initial Logging and Tracking (L & T) Non Physicians dataset provides a list of pending applications that have not been processed by CMS contractors for Non Physicians."
#> 
#> $periodicity
#> [1] "Twice a Week [R/P3.5D]"
#> 
#> $contact
#> [1] "Provider Enrollment Data Requests - CPI (mailto:ProviderEnrollmentDataRequests@cms.hhs.gov)"
#> 
#> $dictionary
#> [1] "https://data.cms.gov/resources/pending-initial-logging-and-tracking-physician-and-non-physicians-data-dictionary"
#> 
#> $identifier
#> [1] "https://data.cms.gov/data-api/v1/dataset/261b83b6-b89f-43ad-ae7b-0d419a3bc24b/data-viewer"
#> 
#> $modified
#> [1] "2025-01-21"
#> 
#> $site
#> [1] "https://data.cms.gov/provider-characteristics/medicare-provider-supplier-enrollment/pending-initial-logging-and-tracking-non-physicians"
#> 
#> $temporal
#> [1] "2023-05-21/2025-01-18"
#> 
#> $references
#> [1] "https://data.cms.gov/resources/pending-initial-logging-and-tracking-physician-and-non-physicians-methodology"
#> 
main_current("pending_physicians")
#> $title
#> [1] "Pending Initial Logging and Tracking Physicians"
#> 
#> $description
#> [1] "The Pending Initial Logging and Tracking (L & T) Physicians dataset provides a list of pending applications that have not been processed by CMS contractors."
#> 
#> $periodicity
#> [1] "Twice a Week [R/P3.5D]"
#> 
#> $contact
#> [1] "Provider Enrollment Data Requests - CPI (mailto:ProviderEnrollmentDataRequests@cms.hhs.gov)"
#> 
#> $dictionary
#> [1] "https://data.cms.gov/resources/pending-initial-logging-and-tracking-physician-and-non-physicians-data-dictionary"
#> 
#> $identifier
#> [1] "https://data.cms.gov/data-api/v1/dataset/6bd6b1dd-208c-4f9c-88b8-b15fec6db548/data-viewer"
#> 
#> $modified
#> [1] "2025-01-21"
#> 
#> $site
#> [1] "https://data.cms.gov/provider-characteristics/medicare-provider-supplier-enrollment/pending-initial-logging-and-tracking-physicians"
#> 
#> $temporal
#> [1] "2023-05-21/2025-01-18"
#> 
#> $references
#> [1] "https://data.cms.gov/resources/pending-initial-logging-and-tracking-physician-and-non-physicians-methodology"
#>