Load <TemporalMainGroup> API Endpoints

Load <TemporalMainGroup> API Endpoints

Usage

main_temporal_group(alias)

Arguments

alias

<chr> dataset title

Value

<TemporalMainGroup> object

Examples

main_temporal_group("utilization")
#> $title
#> [1] "Medicare Physician & Other Practitioners"
#> 
#> $periodicity
#> [1] "Annually [R/P1Y]"
#> 
#> $contact
#> [1] "Medicare Provider Data - OEDA (mailto:IPAG_Data_Products@cms.hhs.gov)"
#> 
#> $description
#>                                                                                                                                                                                                                                                                               by_geography_and_service 
#> "The Medicare Physician & Other Practitioners by Geography and Service dataset contains information on use, payments, submitted charges, and beneficiary demographic and health characteristics organized by geography, Healthcare Common Procedure Coding System (HCPCS) code, and place of service." 
#>                                                                                                                                                                                                                                                                                            by_provider 
#>                                                                    "The Medicare Physician & Other Practitioners by Provider dataset provides information on use, payments, submitted charges and beneficiary demographic and health characteristics organized by National Provider Identifier (NPI)." 
#>                                                                                                                                                                                                                                                                                by_provider_and_service 
#>                             "The Medicare Physician & Other Practitioners by Provider and Service dataset provides information on use, payments, and submitted charges organized by National Provider Identifier (NPI), Healthcare Common Procedure Coding System (HCPCS) code, and place of service." 
#> 
#> $dictionary
#>                                                                                         by_geography_and_service 
#> "https://data.cms.gov/resources/medicare-physician-other-practitioners-by-geography-and-service-data-dictionary" 
#>                                                                                                      by_provider 
#>              "https://data.cms.gov/resources/medicare-physician-other-practitioners-by-provider-data-dictionary" 
#>                                                                                          by_provider_and_service 
#>  "https://data.cms.gov/resources/medicare-physician-other-practitioners-by-provider-and-service-data-dictionary" 
#> 
#> $site
#>                                                                                                                                          by_geography_and_service 
#> "https://data.cms.gov/provider-summary-by-type-of-service/medicare-physician-other-practitioners/medicare-physician-other-practitioners-by-geography-and-service" 
#>                                                                                                                                                       by_provider 
#>              "https://data.cms.gov/provider-summary-by-type-of-service/medicare-physician-other-practitioners/medicare-physician-other-practitioners-by-provider" 
#>                                                                                                                                           by_provider_and_service 
#>  "https://data.cms.gov/provider-summary-by-type-of-service/medicare-physician-other-practitioners/medicare-physician-other-practitioners-by-provider-and-service" 
#> 
#> $by_geography_and_service
#> # A tibble: 10 × 7
#>     year modified   temporal              identifier download filetype resources
#>    <int> <date>     <chr>                 <chr>      <chr>    <chr>    <chr>    
#>  1  2022 2024-06-04 2022-01-01 • 2022-12… https://d… https:/… text/csv https://…
#>  2  2021 2024-03-08 2021-01-01 • 2021-12… https://d… https:/… text/csv https://…
#>  3  2020 2023-05-10 2020-01-01 • 2020-12… https://d… https:/… text/csv https://…
#>  4  2019 2023-05-10 2019-01-01 • 2019-12… https://d… https:/… text/csv https://…
#>  5  2018 2023-05-10 2018-01-01 • 2018-12… https://d… https:/… text/csv https://…
#>  6  2017 2023-05-10 2017-01-01 • 2017-12… https://d… https:/… text/csv https://…
#>  7  2016 2023-05-10 2016-01-01 • 2016-12… https://d… https:/… text/csv https://…
#>  8  2015 2023-05-10 2015-01-01 • 2015-12… https://d… https:/… text/csv https://…
#>  9  2014 2023-05-10 2014-01-01 • 2014-12… https://d… https:/… text/csv https://…
#> 10  2013 2023-05-10 2013-01-01 • 2013-12… https://d… NA       NA       https://…
#> 
#> $by_provider
#> # A tibble: 10 × 7
#>     year modified   temporal              identifier download filetype resources
#>    <int> <date>     <chr>                 <chr>      <chr>    <chr>    <chr>    
#>  1  2022 2024-06-04 2022-01-01 • 2022-12… https://d… https:/… text/csv https://…
#>  2  2021 2024-06-04 2021-01-01 • 2021-12… https://d… https:/… text/csv https://…
#>  3  2020 2024-06-04 2020-01-01 • 2020-12… https://d… https:/… text/csv https://…
#>  4  2019 2024-06-07 2019-01-01 • 2019-12… https://d… https:/… text/csv https://…
#>  5  2018 2024-06-07 2018-01-01 • 2018-12… https://d… https:/… text/csv https://…
#>  6  2017 2024-06-04 2017-01-01 • 2017-12… https://d… https:/… text/csv https://…
#>  7  2016 2024-06-04 2016-01-01 • 2016-12… https://d… https:/… text/csv https://…
#>  8  2015 2024-06-04 2015-01-01 • 2015-12… https://d… https:/… text/csv https://…
#>  9  2014 2024-06-04 2014-01-01 • 2014-12… https://d… https:/… text/csv https://…
#> 10  2013 2024-06-04 2013-01-01 • 2013-12… https://d… NA       NA       https://…
#> 
#> $by_provider_and_service
#> # A tibble: 10 × 7
#>     year modified   temporal              identifier download filetype resources
#>    <int> <date>     <chr>                 <chr>      <chr>    <chr>    <chr>    
#>  1  2022 2024-06-04 2022-01-01 • 2022-12… https://d… https:/… text/csv https://…
#>  2  2021 2023-05-10 2021-01-01 • 2021-12… https://d… https:/… text/csv https://…
#>  3  2020 2023-05-10 2020-01-01 • 2020-12… https://d… https:/… text/csv https://…
#>  4  2019 2023-05-10 2019-01-01 • 2019-12… https://d… https:/… text/csv https://…
#>  5  2018 2023-05-10 2018-01-01 • 2018-12… https://d… https:/… text/csv https://…
#>  6  2017 2023-05-10 2017-01-01 • 2017-12… https://d… https:/… text/csv https://…
#>  7  2016 2023-05-10 2016-01-01 • 2016-12… https://d… https:/… text/csv https://…
#>  8  2015 2023-05-10 2015-01-01 • 2015-12… https://d… https:/… text/csv https://…
#>  9  2014 2023-05-10 2014-01-01 • 2014-12… https://d… https:/… text/csv https://…
#> 10  2013 2023-05-10 2013-01-01 • 2013-12… https://d… NA       NA       https://…
#> 
main_temporal_group("prescribers")
#> $title
#> [1] "Medicare Part D Prescribers"
#> 
#> $periodicity
#> [1] "Annually [R/P1Y]"
#> 
#> $contact
#> [1] "Medicare Provider Data - OEDA (mailto:IPAG_Data_Products@cms.hhs.gov)"
#> 
#> $description
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                               by_geography_and_drug 
#> "The Medicare Part D Prescribers by Geography and Drug dataset contains information on prescription drugs prescribed by individual physicians and other health care providers and paid for under the Medicare Part D Prescription Drug Program. For each drug, the dataset includes the total number of prescriptions that were dispensed, which include original prescriptions and any refills, and the total drug cost. The total drug cost includes the ingredient cost of the medication, dispensing fees, sales tax, and any applicable administration fees and is based on the amount paid by the Part D plan, Medicare beneficiary, government subsidies, and any other third-party payers." 
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                         by_provider 
#>                                                                                                                                                                                                        "The Medicare Part D Prescribers by Provider dataset contains information on prescription drugs prescribed by individual physicians and other health care providers and paid for under the Medicare Part D Prescription Drug Program. The dataset identifies providers by their National Provider Identifier (NPI) and summarizes for each prescriber the total number of prescriptions that were dispensed, which include original prescriptions and any refills, and the total drug cost." 
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                by_provider_and_drug 
#>                                                                                                                                                                                                                                                  "The Medicare Part D Prescribers by Provider and Drug dataset provides information on prescription drugs prescribed to Medicare beneficiaries enrolled in Part D by physicians and other health care providers. This dataset contains the total number of prescription fills that were dispensed and the total drug cost paid organized by prescribing National Provider Identifier (NPI), drug brand name (if applicable) and drug generic name." 
#> 
#> $dictionary
#>                                                                              by_geography_and_drug 
#> "https://data.cms.gov/resources/medicare-part-d-prescribers-by-geography-and-drug-data-dictionary" 
#>                                                                                        by_provider 
#>           "https://data.cms.gov/resources/medicare-part-d-prescribers-by-provider-data-dictionary" 
#>                                                                               by_provider_and_drug 
#>  "https://data.cms.gov/resources/medicare-part-d-prescribers-by-provider-and-drug-data-dictionary" 
#> 
#> $site
#>                                                                                                                    by_geography_and_drug 
#> "https://data.cms.gov/provider-summary-by-type-of-service/medicare-part-d-prescribers/medicare-part-d-prescribers-by-geography-and-drug" 
#>                                                                                                                              by_provider 
#>           "https://data.cms.gov/provider-summary-by-type-of-service/medicare-part-d-prescribers/medicare-part-d-prescribers-by-provider" 
#>                                                                                                                     by_provider_and_drug 
#>  "https://data.cms.gov/provider-summary-by-type-of-service/medicare-part-d-prescribers/medicare-part-d-prescribers-by-provider-and-drug" 
#> 
#> $by_geography_and_drug
#> # A tibble: 10 × 7
#>     year modified   temporal              identifier download filetype resources
#>    <int> <date>     <chr>                 <chr>      <chr>    <chr>    <chr>    
#>  1  2022 2024-06-04 2022-01-01 • 2022-12… https://d… https:/… text/csv https://…
#>  2  2021 2023-10-25 2021-01-01 • 2021-12… https://d… https:/… text/csv https://…
#>  3  2020 2024-05-21 2020-01-01 • 2020-12… https://d… https:/… text/csv https://…
#>  4  2019 2023-05-10 2019-01-01 • 2019-12… https://d… https:/… text/csv https://…
#>  5  2018 2023-05-10 2018-01-01 • 2018-12… https://d… https:/… text/csv https://…
#>  6  2017 2023-05-10 2017-01-01 • 2017-12… https://d… https:/… text/csv https://…
#>  7  2016 2024-01-30 2016-01-01 • 2016-12… https://d… https:/… text/csv https://…
#>  8  2015 2023-05-10 2015-01-01 • 2015-12… https://d… https:/… text/csv https://…
#>  9  2014 2023-05-10 2014-01-01 • 2014-12… https://d… https:/… text/csv https://…
#> 10  2013 2023-05-10 2013-01-01 • 2013-12… https://d… https:/… applica… https://…
#> 
#> $by_provider
#> # A tibble: 10 × 7
#>     year modified   temporal              identifier download filetype resources
#>    <int> <date>     <chr>                 <chr>      <chr>    <chr>    <chr>    
#>  1  2022 2024-06-04 2022-01-01 • 2022-12… https://d… https:/… text/csv https://…
#>  2  2021 2024-06-04 2021-01-01 • 2021-12… https://d… https:/… text/csv https://…
#>  3  2020 2024-06-04 2020-01-01 • 2020-12… https://d… https:/… text/csv https://…
#>  4  2019 2024-06-04 2019-01-01 • 2019-12… https://d… https:/… text/csv https://…
#>  5  2018 2024-06-04 2018-01-01 • 2018-12… https://d… https:/… text/csv https://…
#>  6  2017 2024-06-04 2017-01-01 • 2017-12… https://d… https:/… text/csv https://…
#>  7  2016 2024-06-04 2016-01-01 • 2016-12… https://d… https:/… text/csv https://…
#>  8  2015 2024-06-04 2015-01-01 • 2015-12… https://d… https:/… text/csv https://…
#>  9  2014 2024-06-04 2014-01-01 • 2014-12… https://d… https:/… text/csv https://…
#> 10  2013 2024-06-04 2013-01-01 • 2013-12… https://d… NA       NA       https://…
#> 
#> $by_provider_and_drug
#> # A tibble: 10 × 7
#>     year modified   temporal              identifier download filetype resources
#>    <int> <date>     <chr>                 <chr>      <chr>    <chr>    <chr>    
#>  1  2022 2024-06-04 2022-01-01 • 2022-12… https://d… https:/… text/csv https://…
#>  2  2021 2024-06-04 2021-01-01 • 2021-12… https://d… https:/… text/csv https://…
#>  3  2020 2024-06-04 2020-01-01 • 2020-12… https://d… https:/… text/csv https://…
#>  4  2019 2024-06-04 2019-01-01 • 2019-12… https://d… https:/… text/csv https://…
#>  5  2018 2024-06-04 2018-01-01 • 2018-12… https://d… https:/… text/csv https://…
#>  6  2017 2024-06-04 2017-01-01 • 2017-12… https://d… https:/… text/csv https://…
#>  7  2016 2024-06-04 2016-01-01 • 2016-12… https://d… https:/… text/csv https://…
#>  8  2015 2024-06-04 2015-01-01 • 2015-12… https://d… https:/… text/csv https://…
#>  9  2014 2024-06-04 2014-01-01 • 2014-12… https://d… https:/… text/csv https://…
#> 10  2013 2024-06-04 2013-01-01 • 2013-12… https://d… NA       NA       https://…
#> 
main_temporal_group("suppliers")
#> $title
#> [1] "Medicare Durable Medical Equipment, Devices & Supplies"
#> 
#> $periodicity
#> [1] "Annually [R/P1Y]"
#> 
#> $contact
#> [1] "Medicare Provider Data - OEDA (mailto:IPAG_Data_Products@cms.hhs.gov)"
#> 
#> $description
#>                                                                                                                                                                                                                                                                                     by_geography_and_service 
#>                                  "The Medicare Durable Medical Equipment, Devices & Supplies by Geography and Service dataset contains information on usage, payments, and submitted charges organized by geography, Healthcare Common Procedure Coding System (HCPCS) code, and supplier rental indicator." 
#>                                                                                                                                                                                                                                                                                        by_referring_provider 
#>                                                "The Medicare Durable Medical Equipment, Devices & Supplies by Referring Provider dataset contains information on usage, payments, submitted charges and beneficiary demographic and health characteristics organized by National Provider Identifier (NPI)." 
#>                                                                                                                                                                                                                                                                            by_referring_provider_and_service 
#> "The Medicare Durable Medical Equipment, Devices & Supplies by Referring Provider and Service dataset contains information on usage, payments and submitted charges organized by National Provider Identifier (NPI), Healthcare Common Procedure Coding System (HCPCS) code, and supplier rental indicator." 
#>                                                                                                                                                                                                                                                                                                  by_supplier 
#>                                                 "The Medicare Durable Medical Equipment, Devices & Supplies by Supplier dataset contains information on usage, payments, submitted charges and beneficiary demographic and health characteristics organized by supplier National Provider Identifier (NPI)." 
#>                                                                                                                                                                                                                                                                                      by_supplier_and_service 
#>  "The Medicare Durable Medical Equipment, Devices & Supplies by Supplier and Service dataset contains information on usage, payments and submitted charges organized by supplier National Provider Identifier (NPI), Healthcare Common Procedure Coding System (HCPCS) code, and supplier rental indicator." 
#> 
#> $dictionary
#>                                                                                                               by_geography_and_service 
#>     "https://data.cms.gov/resources/medicare-durable-medical-equipment-devices-supplies-by-geography-and-service-data-dictionary-2022" 
#>                                                                                                                  by_referring_provider 
#>             "https://data.cms.gov/resources/medicare-durable-medical-equipment-devices-supplies-by-referring-provider-data-dictionary" 
#>                                                                                                      by_referring_provider_and_service 
#> "https://data.cms.gov/resources/medicare-durable-medical-equipment-devices-supplies-by-referring-provider-and-service-data-dictionary" 
#>                                                                                                                            by_supplier 
#>                       "https://data.cms.gov/resources/medicare-durable-medical-equipment-devices-supplies-by-supplier-data-dictionary" 
#>                                                                                                                by_supplier_and_service 
#>           "https://data.cms.gov/resources/medicare-durable-medical-equipment-devices-supplies-by-supplier-and-service-data-dictionary" 
#> 
#> $site
#>                                                                                                                                                                             by_geography_and_service 
#>          "https://data.cms.gov/provider-summary-by-type-of-service/medicare-durable-medical-equipment-devices-supplies/medicare-durable-medical-equipment-devices-supplies-by-geography-and-service" 
#>                                                                                                                                                                                by_referring_provider 
#>             "https://data.cms.gov/provider-summary-by-type-of-service/medicare-durable-medical-equipment-devices-supplies/medicare-durable-medical-equipment-devices-supplies-by-referring-provider" 
#>                                                                                                                                                                    by_referring_provider_and_service 
#> "https://data.cms.gov/provider-summary-by-type-of-service/medicare-durable-medical-equipment-devices-supplies/medicare-durable-medical-equipment-devices-supplies-by-referring-provider-and-service" 
#>                                                                                                                                                                                          by_supplier 
#>                       "https://data.cms.gov/provider-summary-by-type-of-service/medicare-durable-medical-equipment-devices-supplies/medicare-durable-medical-equipment-devices-supplies-by-supplier" 
#>                                                                                                                                                                              by_supplier_and_service 
#>           "https://data.cms.gov/provider-summary-by-type-of-service/medicare-durable-medical-equipment-devices-supplies/medicare-durable-medical-equipment-devices-supplies-by-supplier-and-service" 
#> 
#> $by_geography_and_service
#> # A tibble: 9 × 7
#>    year modified   temporal               identifier download filetype resources
#>   <int> <date>     <chr>                  <chr>      <chr>    <chr>    <chr>    
#> 1  2022 2024-06-04 2022-01-01 • 2022-12-… https://d… https:/… text/csv https://…
#> 2  2021 2024-06-04 2021-01-01 • 2021-12-… https://d… https:/… text/csv https://…
#> 3  2020 2024-06-04 2020-01-01 • 2020-12-… https://d… https:/… text/csv https://…
#> 4  2019 2024-06-04 2019-01-01 • 2019-12-… https://d… https:/… text/csv https://…
#> 5  2018 2024-06-04 2018-01-01 • 2018-12-… https://d… https:/… text/csv https://…
#> 6  2017 2024-06-04 2017-01-01 • 2017-12-… https://d… https:/… text/csv https://…
#> 7  2016 2024-06-04 2016-01-01 • 2016-12-… https://d… https:/… text/csv https://…
#> 8  2015 2024-06-04 2015-01-01 • 2015-12-… https://d… https:/… text/csv https://…
#> 9  2014 2024-06-04 2014-01-01 • 2014-12-… https://d… NA       NA       https://…
#> 
#> $by_referring_provider
#> # A tibble: 9 × 7
#>    year modified   temporal               identifier download filetype resources
#>   <int> <date>     <chr>                  <chr>      <chr>    <chr>    <chr>    
#> 1  2022 2024-06-04 2022-01-01 • 2022-12-… https://d… https:/… text/csv https://…
#> 2  2021 2024-06-05 2021-01-01 • 2021-12-… https://d… https:/… text/csv https://…
#> 3  2020 2024-06-05 2020-01-01 • 2020-12-… https://d… https:/… text/csv https://…
#> 4  2019 2024-06-05 2019-01-01 • 2019-12-… https://d… https:/… text/csv https://…
#> 5  2018 2024-06-05 2018-01-01 • 2018-12-… https://d… https:/… text/csv https://…
#> 6  2017 2024-06-05 2017-01-01 • 2017-12-… https://d… https:/… text/csv https://…
#> 7  2016 2024-06-05 2016-01-01 • 2016-12-… https://d… https:/… text/csv https://…
#> 8  2015 2024-06-05 2015-01-01 • 2015-12-… https://d… https:/… text/csv https://…
#> 9  2014 2024-06-05 2014-01-01 • 2014-12-… https://d… NA       NA       https://…
#> 
#> $by_referring_provider_and_service
#> # A tibble: 9 × 7
#>    year modified   temporal               identifier download filetype resources
#>   <int> <date>     <chr>                  <chr>      <chr>    <chr>    <chr>    
#> 1  2022 2024-06-04 2022-01-01 • 2022-12-… https://d… https:/… text/csv https://…
#> 2  2021 2024-06-04 2021-01-01 • 2021-12-… https://d… https:/… text/csv https://…
#> 3  2020 2024-06-04 2020-01-01 • 2020-12-… https://d… https:/… text/csv https://…
#> 4  2019 2024-06-04 2019-01-01 • 2019-12-… https://d… https:/… text/csv https://…
#> 5  2018 2024-06-04 2018-01-01 • 2018-12-… https://d… https:/… text/csv https://…
#> 6  2017 2024-06-04 2017-01-01 • 2017-12-… https://d… https:/… text/csv https://…
#> 7  2016 2024-06-04 2016-01-01 • 2016-12-… https://d… https:/… text/csv https://…
#> 8  2015 2024-06-04 2015-01-01 • 2015-12-… https://d… https:/… text/csv https://…
#> 9  2014 2024-06-04 2014-01-01 • 2014-12-… https://d… NA       NA       https://…
#> 
#> $by_supplier
#> # A tibble: 9 × 7
#>    year modified   temporal               identifier download filetype resources
#>   <int> <date>     <chr>                  <chr>      <chr>    <chr>    <chr>    
#> 1  2022 2024-06-04 2022-01-01 • 2022-12-… https://d… https:/… text/csv https://…
#> 2  2021 2024-06-05 2021-01-01 • 2021-12-… https://d… https:/… text/csv https://…
#> 3  2020 2024-06-05 2020-01-01 • 2020-12-… https://d… https:/… text/csv https://…
#> 4  2019 2024-06-05 2019-01-01 • 2019-12-… https://d… https:/… text/csv https://…
#> 5  2018 2024-06-05 2018-01-01 • 2018-12-… https://d… https:/… text/csv https://…
#> 6  2017 2024-06-05 2017-01-01 • 2017-12-… https://d… https:/… text/csv https://…
#> 7  2016 2024-06-05 2016-01-01 • 2016-12-… https://d… https:/… text/csv https://…
#> 8  2015 2024-06-05 2015-01-01 • 2015-12-… https://d… https:/… text/csv https://…
#> 9  2014 2024-06-05 2014-01-01 • 2014-12-… https://d… NA       NA       https://…
#> 
#> $by_supplier_and_service
#> # A tibble: 9 × 7
#>    year modified   temporal               identifier download filetype resources
#>   <int> <date>     <chr>                  <chr>      <chr>    <chr>    <chr>    
#> 1  2022 2024-06-04 2022-01-01 • 2022-12-… https://d… https:/… text/csv https://…
#> 2  2021 2024-06-04 2021-01-01 • 2021-12-… https://d… https:/… text/csv https://…
#> 3  2020 2024-06-04 2020-01-01 • 2020-12-… https://d… https:/… text/csv https://…
#> 4  2019 2024-06-04 2019-01-01 • 2019-12-… https://d… https:/… text/csv https://…
#> 5  2018 2024-06-04 2018-01-01 • 2018-12-… https://d… https:/… text/csv https://…
#> 6  2017 2024-06-04 2017-01-01 • 2017-12-… https://d… https:/… text/csv https://…
#> 7  2016 2024-06-04 2016-01-01 • 2016-12-… https://d… https:/… text/csv https://…
#> 8  2015 2024-06-04 2015-01-01 • 2015-12-… https://d… https:/… applica… https://…
#> 9  2014 2024-06-04 2014-01-01 • 2014-12-… https://d… NA       NA       https://…
#> 
main_temporal_group("outpatient")
#> $title
#> [1] "Medicare Outpatient Hospitals"
#> 
#> $periodicity
#> [1] "Annually [R/P1Y]"
#> 
#> $contact
#> [1] "Medicare Provider Data - OEDA (mailto:IPAG_Data_Products@cms.hhs.gov)"
#> 
#> $description
#>                                                                                                                                                                                                                                                                                                                                                     by_geography_and_service 
#>                          "The Medicare Outpatient Hospitals by Geography and Service dataset provides information on services for Original Medicare Part B beneficiaries by OPPS hospitals. These datasets contain information on the number of services, payments, and submitted charges organized by geography and comprehensive Ambulatory Payment Classification (APC)." 
#>                                                                                                                                                                                                                                                                                                                                                      by_provider_and_service 
#> "The Medicare Outpatient Hospitals by Provider and Service dataset provides information on services for Original Medicare Part B beneficiaries by OPPS hospitals. These datasets contain information on the number of services, payments, and submitted charges organized by provider CMS Certified Number (CCN) and comprehensive Ambulatory Payment Classification (APC)." 
#> 
#> $dictionary
#>                                                                                by_geography_and_service 
#> "https://data.cms.gov/resources/medicare-outpatient-hospitals-by-geography-and-service-data-dictionary" 
#>                                                                                 by_provider_and_service 
#>  "https://data.cms.gov/resources/medicare-outpatient-hospitals-by-provider-and-service-data-dictionary" 
#> 
#> $site
#>                                                                                                                        by_geography_and_service 
#> "https://data.cms.gov/provider-summary-by-type-of-service/medicare-outpatient-hospitals/medicare-outpatient-hospitals-by-geography-and-service" 
#>                                                                                                                         by_provider_and_service 
#>  "https://data.cms.gov/provider-summary-by-type-of-service/medicare-outpatient-hospitals/medicare-outpatient-hospitals-by-provider-and-service" 
#> 
#> $by_geography_and_service
#> # A tibble: 8 × 7
#>    year modified   temporal               identifier download filetype resources
#>   <int> <date>     <chr>                  <chr>      <chr>    <chr>    <chr>    
#> 1  2022 2024-06-04 2022-01-01 • 2022-12-… https://d… https:/… text/csv https://…
#> 2  2021 2024-06-04 2021-01-01 • 2021-12-… https://d… https:/… text/csv https://…
#> 3  2020 2024-06-04 2020-01-01 • 2020-12-… https://d… https:/… text/csv https://…
#> 4  2019 2021-08-31 2019-01-01 • 2019-12-… https://d… https:/… text/csv https://…
#> 5  2018 2021-06-07 2018-01-01 • 2018-12-… https://d… https:/… text/csv https://…
#> 6  2017 2020-10-15 2017-01-01 • 2017-12-… https://d… https:/… text/csv https://…
#> 7  2016 2020-07-10 2016-01-01 • 2016-12-… https://d… https:/… text/csv https://…
#> 8  2015 2021-01-11 2015-01-01 • 2015-12-… https://d… NA       NA       https://…
#> 
#> $by_provider_and_service
#> # A tibble: 8 × 7
#>    year modified   temporal               identifier download filetype resources
#>   <int> <date>     <chr>                  <chr>      <chr>    <chr>    <chr>    
#> 1  2022 2024-06-04 2022-01-01 • 2022-12-… https://d… https:/… text/csv https://…
#> 2  2021 2024-06-04 2021-01-01 • 2021-12-… https://d… https:/… text/csv https://…
#> 3  2020 2024-06-04 2020-01-01 • 2020-12-… https://d… https:/… text/csv https://…
#> 4  2019 2021-08-31 2019-01-01 • 2019-12-… https://d… https:/… text/csv https://…
#> 5  2018 2021-06-07 2018-01-01 • 2018-12-… https://d… https:/… text/csv https://…
#> 6  2017 2020-10-28 2017-01-01 • 2017-12-… https://d… https:/… text/csv https://…
#> 7  2016 2020-07-10 2016-01-01 • 2016-12-… https://d… https:/… text/csv https://…
#> 8  2015 2021-01-11 2015-01-01 • 2015-12-… https://d… NA       NA       https://…
#> 
main_temporal_group("inpatient")
#> $title
#> [1] "Medicare Inpatient Hospitals"
#> 
#> $periodicity
#> [1] "Annually [R/P1Y]"
#> 
#> $contact
#> [1] "Medicare Provider Data - OEDA (mailto:IPAG_Data_Products@cms.hhs.gov)"
#> 
#> $description
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                               by_geography_and_service 
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                               "The Medicare Inpatient Hospitals by Geography and Service dataset provides information on hospital discharges for Original Medicare Part A beneficiaries by IPPS hospitals. This dataset contains information on the number of discharges, payments, and submitted charges organized by geography and Medicare Severity Diagnosis Related Group (DRG)." 
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                            by_provider 
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                "The Medicare Inpatient Hospitals by Provider dataset provides information on inpatient discharges for Original Medicare Part A beneficiaries by IPPS hospitals. It includes information on the use, payment, and hospital charges for more than 3,000 U.S. hospitals that received IPPS payments. The data are organized by hospital." 
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                by_provider_and_service 
#> "The Medicare Inpatient Hospitals by Provider and Service dataset provides information on inpatient discharges for Original Medicare Part A beneficiaries by IPPS hospitals. It includes information on the use, payment, and hospital charges for more than 3,000 U.S. hospitals that received IPPS payments. The data are organized by hospital and Medicare Severity Diagnosis Related Group (DRG).\n\n \n\nHospitals determine what they will charge for items and services provided to patients, and these charges are the amount the hospital bills for an item or service. The Total Payment Amount includes the DRG amount, claim per diem amount, beneficiary primary payer claim payment amount, beneficiary Part A (Hospital Insurance) coinsurance amount, beneficiary deductible amount, beneficiary blood deductible amount and diagnosis related group outlier amount." 
#> 
#> $dictionary
#>                                                                                 by_geography_and_service 
#> "https://data.cms.gov/resources/medicare-inpatient-hospitals-by-geography-and-service-data-dictionary-0" 
#>                                                                                              by_provider 
#>                "https://data.cms.gov/resources/medicare-inpatient-hospitals-by-provider-data-dictionary" 
#>                                                                                  by_provider_and_service 
#>  "https://data.cms.gov/resources/medicare-inpatient-hospitals-by-provider-and-service-data-dictionary-0" 
#> 
#> $site
#>                                                                                                                      by_geography_and_service 
#> "https://data.cms.gov/provider-summary-by-type-of-service/medicare-inpatient-hospitals/medicare-inpatient-hospitals-by-geography-and-service" 
#>                                                                                                                                   by_provider 
#>              "https://data.cms.gov/provider-summary-by-type-of-service/medicare-inpatient-hospitals/medicare-inpatient-hospitals-by-provider" 
#>                                                                                                                       by_provider_and_service 
#>  "https://data.cms.gov/provider-summary-by-type-of-service/medicare-inpatient-hospitals/medicare-inpatient-hospitals-by-provider-and-service" 
#> 
#> $by_geography_and_service
#> # A tibble: 10 × 7
#>     year modified   temporal              identifier download filetype resources
#>    <int> <date>     <chr>                 <chr>      <chr>    <chr>    <chr>    
#>  1  2022 2024-06-04 2022-01-01 • 2022-12… https://d… https:/… text/csv https://…
#>  2  2021 2023-05-10 2021-01-01 • 2021-12… https://d… https:/… text/csv https://…
#>  3  2020 2023-05-10 2020-01-01 • 2020-12… https://d… https:/… text/csv https://…
#>  4  2019 2023-05-10 2019-01-01 • 2019-12… https://d… https:/… text/csv https://…
#>  5  2018 2023-05-10 2018-01-01 • 2018-12… https://d… https:/… text/csv https://…
#>  6  2017 2023-05-10 2017-01-01 • 2017-12… https://d… https:/… text/csv https://…
#>  7  2016 2023-05-10 2016-01-01 • 2016-12… https://d… https:/… text/csv https://…
#>  8  2015 2023-05-10 2015-01-01 • 2015-12… https://d… https:/… text/csv https://…
#>  9  2014 2023-05-10 2014-01-01 • 2014-12… https://d… https:/… text/csv https://…
#> 10  2013 2023-05-10 2013-01-01 • 2013-12… https://d… NA       NA       https://…
#> 
#> $by_provider
#> # A tibble: 10 × 7
#>     year modified   temporal              identifier download filetype resources
#>    <int> <date>     <chr>                 <chr>      <chr>    <chr>    <chr>    
#>  1  2022 2024-06-04 2022-01-01 • 2022-12… https://d… https:/… text/csv https://…
#>  2  2021 2024-06-04 2021-01-01 • 2021-12… https://d… https:/… text/csv https://…
#>  3  2020 2024-06-04 2020-01-01 • 2020-12… https://d… https:/… text/csv https://…
#>  4  2019 2024-06-04 2019-01-01 • 2019-12… https://d… https:/… text/csv https://…
#>  5  2018 2024-06-04 2018-01-01 • 2018-12… https://d… https:/… text/csv https://…
#>  6  2017 2024-06-04 2017-01-01 • 2017-12… https://d… https:/… text/csv https://…
#>  7  2016 2024-06-04 2016-01-01 • 2016-12… https://d… https:/… text/csv https://…
#>  8  2015 2024-06-04 2015-01-01 • 2015-12… https://d… https:/… text/csv https://…
#>  9  2014 2024-06-04 2014-01-01 • 2014-12… https://d… https:/… text/csv https://…
#> 10  2013 2024-06-04 2013-01-01 • 2013-12… https://d… NA       NA       https://…
#> 
#> $by_provider_and_service
#> # A tibble: 10 × 7
#>     year modified   temporal              identifier download filetype resources
#>    <int> <date>     <chr>                 <chr>      <chr>    <chr>    <chr>    
#>  1  2022 2024-06-04 2022-01-01 • 2022-12… https://d… https:/… text/csv https://…
#>  2  2021 2023-09-05 2021-01-01 • 2021-12… https://d… https:/… text/csv https://…
#>  3  2020 2023-05-10 2020-01-01 • 2020-12… https://d… https:/… text/csv https://…
#>  4  2019 2023-05-10 2019-01-01 • 2019-12… https://d… https:/… text/csv https://…
#>  5  2018 2023-05-10 2018-01-01 • 2018-12… https://d… https:/… text/csv https://…
#>  6  2017 2023-05-10 2017-01-01 • 2017-12… https://d… https:/… text/csv https://…
#>  7  2016 2023-05-10 2016-01-01 • 2016-12… https://d… https:/… text/csv https://…
#>  8  2015 2023-05-10 2015-01-01 • 2015-12… https://d… https:/… text/csv https://…
#>  9  2014 2023-05-10 2014-01-01 • 2014-12… https://d… https:/… text/csv https://…
#> 10  2013 2023-05-10 2013-01-01 • 2013-12… https://d… NA       NA       https://…
#>