Medicaid API Endpoint Classes

Medicaid API Endpoint Classes

Arguments

alias

<chr> endpoint alias

Value

An S7 <caid_endpoint>, <caid_group>, or <caid_temporal> object

Examples

caid_endpoint("MLR_summary")
#> <caid_endpoint>
#>  @ identifier: chr "https://data.medicaid.gov/api/1/datastore/query/743f9f04-4473-41e2-9da2-9a89db65ee55/0"
#>  @ metadata  :List of 5
#>  .. $ title      : chr "MLR Summary Reports"
#>  .. $ description: chr "The MLR Summary Reports provide plan-specific MLRs for Medicaid managed care programs. The MLR summary report d"| __truncated__
#>  .. $ modified   : Date[1:1], format: "2024-11-04"
#>  .. $ dictionary : chr NA
#>  .. $ download   : chr "https://download.medicaid.gov/data/mlr-public-use-file-07302024final.csv"
#>  @ dimensions: <class_dimensions>
#>  .. @ limit : int 8000
#>  .. @ rows  : int 1719
#>  .. @ pages : int 1
#>  .. @ fields:List of 14
#>  .. .. $ state                                            : chr "state"
#>  .. .. $ program_name                                     : chr "program_name"
#>  .. .. $ program_type                                     : chr "program_type"
#>  .. .. $ eligibility_group                                : chr "eligibility_group"
#>  .. .. $ if_other_describe_eligibility_group              : chr "if_other_describe_eligibility_group"
#>  .. .. $ mco_pihp_or_pahp_name                            : chr "mco_pihp_or_pahp_name"
#>  .. .. $ mlr_reporting_period_start_date                  : chr "mlr_reporting_period_start_date"
#>  .. .. $ mlr_reporting_period_end_date                    : chr "mlr_reporting_period_end_date"
#>  .. .. $ report_year                                      : chr "report_year"
#>  .. .. $ mlr_numerator                                    : chr "mlr_numerator"
#>  .. .. $ mlr_denominator                                  : chr "mlr_denominator"
#>  .. .. $ statereported_member_months                      : chr "statereported_member_months"
#>  .. .. $ adjusted_mlr                                     : chr "adjusted_mlr"
#>  .. .. $ remittance_dollar_amount_for_mlr_reporting_period: chr "remittance_dollar_amount_for_mlr_reporting_period"
caid_group("demographics")
#> <caid_group>
#>  @ group  : chr "Medicaid and CHIP Enrollee Demographics"
#>  @ members:List of 8
#>  .. $ DEMO_well      : <caid_endpoint>
#>  ..  ..@ identifier: chr "https://data.medicaid.gov/api/1/datastore/query/289182e5-44c8-4fcc-96c4-4af69be9863f/0"
#>  ..  ..@ metadata  :List of 5
#>  .. .. .. $ title      : chr "Medicaid and CHIP enrollees who received a well-child visit"
#>  .. .. .. $ description: chr "This data set includes annual counts and percentages of Medicaid and Childrens Health Insurance Program (CHIP) "| __truncated__
#>  .. .. .. $ modified   : Date[1:1], format: "2025-01-17"
#>  .. .. .. $ dictionary : chr NA
#>  .. .. .. $ download   : chr "https://download.medicaid.gov/data/well-child-visits-2020-2022-01162025.csv"
#>  ..  ..@ dimensions: <class_dimensions>
#>  .. .. .. @ limit : int 8000
#>  .. .. .. @ rows  : int 63
#>  .. .. .. @ pages : int 1
#>  .. .. .. @ fields:List of 9
#>  .. .. .. .. $ year                          : chr "year"
#>  .. .. .. .. $ geography                     : chr "geography"
#>  .. .. .. .. $ subpopulation_topic           : chr "subpopulation_topic"
#>  .. .. .. .. $ subpopulation                 : chr "subpopulation"
#>  .. .. .. .. $ category                      : chr "category"
#>  .. .. .. .. $ count_of_enrollees            : chr "count_of_enrollees"
#>  .. .. .. .. $ denominator_count_of_enrollees: chr "denominator_count_of_enrollees"
#>  .. .. .. .. $ percentage_of_enrollees       : chr "percentage_of_enrollees"
#>  .. .. .. .. $ data_version                  : chr "data_version"
#>  .. $ DEMO_sud       : <caid_endpoint>
#>  ..  ..@ identifier: chr "https://data.medicaid.gov/api/1/datastore/query/8062e2f4-4c0a-41c9-8217-979468a80986/0"
#>  ..  ..@ metadata  :List of 5
#>  .. .. .. $ title      : chr "Medicaid and CHIP enrollees who received mental health or SUD services"
#>  .. .. .. $ description: chr "This data set includes annual counts and percentages of Medicaid and Childrens Health Insurance Program (CHIP) "| __truncated__
#>  .. .. .. $ modified   : Date[1:1], format: "2025-01-17"
#>  .. .. .. $ dictionary : chr NA
#>  .. .. .. $ download   : chr "https://download.medicaid.gov/data/mh-and-sud-services-2020-2022-01162025.csv"
#>  ..  ..@ dimensions: <class_dimensions>
#>  .. .. .. @ limit : int 8000
#>  .. .. .. @ rows  : int 216
#>  .. .. .. @ pages : int 1
#>  .. .. .. @ fields:List of 9
#>  .. .. .. .. $ year                          : chr "year"
#>  .. .. .. .. $ geography                     : chr "geography"
#>  .. .. .. .. $ subpopulation_topic           : chr "subpopulation_topic"
#>  .. .. .. .. $ subpopulation                 : chr "subpopulation"
#>  .. .. .. .. $ category                      : chr "category"
#>  .. .. .. .. $ count_of_enrollees            : chr "count_of_enrollees"
#>  .. .. .. .. $ denominator_count_of_enrollees: chr "denominator_count_of_enrollees"
#>  .. .. .. .. $ percentage_of_enrollees       : chr "percentage_of_enrollees"
#>  .. .. .. .. $ data_version                  : chr "data_version"
#>  .. $ DEMO_disability: <caid_endpoint>
#>  ..  ..@ identifier: chr "https://data.medicaid.gov/api/1/datastore/query/b12e390f-1721-47c0-bb38-b0b1805e0446/0"
#>  ..  ..@ metadata  :List of 5
#>  .. .. .. $ title      : chr "Medicaid enrollees who qualify for benefits based on disability"
#>  .. .. .. $ description: chr "This data set includes annual counts and percentages of Medicaid enrollees who are eligible for benefits based "| __truncated__
#>  .. .. .. $ modified   : Date[1:1], format: "2025-01-17"
#>  .. .. .. $ dictionary : chr NA
#>  .. .. .. $ download   : chr "https://download.medicaid.gov/data/disability-2020-2022-01162025.csv"
#>  ..  ..@ dimensions: <class_dimensions>
#>  .. .. .. @ limit : int 8000
#>  .. .. .. @ rows  : int 114
#>  .. .. .. @ pages : int 1
#>  .. .. .. @ fields:List of 9
#>  .. .. .. .. $ year                          : chr "year"
#>  .. .. .. .. $ geography                     : chr "geography"
#>  .. .. .. .. $ subpopulation_topic           : chr "subpopulation_topic"
#>  .. .. .. .. $ subpopulation                 : chr "subpopulation"
#>  .. .. .. .. $ category                      : chr "category"
#>  .. .. .. .. $ count_of_enrollees            : chr "count_of_enrollees"
#>  .. .. .. .. $ denominator_count_of_enrollees: chr "denominator_count_of_enrollees"
#>  .. .. .. .. $ percentage_of_enrollees       : chr "percentage_of_enrollees"
#>  .. .. .. .. $ data_version                  : chr "data_version"
#>  .. $ DEMO_premature : <caid_endpoint>
#>  ..  ..@ identifier: chr "https://data.medicaid.gov/api/1/datastore/query/ee3b9534-0d19-4c1b-bf74-43f898d5de7c/0"
#>  ..  ..@ metadata  :List of 5
#>  .. .. .. $ title      : chr "Prematurity and severe maternal morbidity among Medicaid- and CHIP-covered live births"
#>  .. .. .. $ description: chr "This data set includes annual counts and rates of Medicaid- and Childrens Health Insurance Program (CHIP)-cover"| __truncated__
#>  .. .. .. $ modified   : Date[1:1], format: "2025-01-17"
#>  .. .. .. $ dictionary : chr NA
#>  .. .. .. $ download   : chr "https://download.medicaid.gov/data/preterm-birth-and-smm-2020-2022-01172025.csv"
#>  ..  ..@ dimensions: <class_dimensions>
#>  .. .. .. @ limit : int 8000
#>  .. .. .. @ rows  : int 435
#>  .. .. .. @ pages : int 1
#>  .. .. .. @ fields:List of 9
#>  .. .. .. .. $ year                           : chr "year"
#>  .. .. .. .. $ geography                      : chr "geography"
#>  .. .. .. .. $ subpopulation_topic            : chr "subpopulation_topic"
#>  .. .. .. .. $ subpopulation                  : chr "subpopulation"
#>  .. .. .. .. $ category                       : chr "category"
#>  .. .. .. .. $ count_of_deliveries            : chr "count_of_deliveries"
#>  .. .. .. .. $ denominator_count_of_deliveries: chr "denominator_count_of_deliveries"
#>  .. .. .. .. $ rate_of_deliveries             : chr "rate_of_deliveries"
#>  .. .. .. .. $ data_version                   : chr "data_version"
#>  .. $ DEMO_language  : <caid_endpoint>
#>  ..  ..@ identifier: chr "https://data.medicaid.gov/api/1/datastore/query/3df86994-c0e4-4106-a404-e95e87c824e6/0"
#>  ..  ..@ metadata  :List of 5
#>  .. .. .. $ title      : chr "Primary language spoken by the Medicaid and CHIP population"
#>  .. .. .. $ description: chr "This data set includes annual counts and percentages of Medicaid and Childrens Health Insurance Program (CHIP) "| __truncated__
#>  .. .. .. $ modified   : Date[1:1], format: "2025-01-18"
#>  .. .. .. $ dictionary : chr NA
#>  .. .. .. $ download   : chr "https://download.medicaid.gov/data/primary-language-2020-2022-01162025.csv"
#>  ..  ..@ dimensions: <class_dimensions>
#>  .. .. .. @ limit : int 8000
#>  .. .. .. @ rows  : int 675
#>  .. .. .. @ pages : int 1
#>  .. .. .. @ fields:List of 9
#>  .. .. .. .. $ year                          : chr "year"
#>  .. .. .. .. $ geography                     : chr "geography"
#>  .. .. .. .. $ subpopulation_topic           : chr "subpopulation_topic"
#>  .. .. .. .. $ subpopulation                 : chr "subpopulation"
#>  .. .. .. .. $ category                      : chr "category"
#>  .. .. .. .. $ count_of_enrollees            : chr "count_of_enrollees"
#>  .. .. .. .. $ denominator_count_of_enrollees: chr "denominator_count_of_enrollees"
#>  .. .. .. .. $ percentage_of_enrollees       : chr "percentage_of_enrollees"
#>  .. .. .. .. $ data_version                  : chr "data_version"
#>  .. $ DEMO_race      : <caid_endpoint>
#>  ..  ..@ identifier: chr "https://data.medicaid.gov/api/1/datastore/query/b9e18d80-c765-4fb9-95e7-81951e01a6a3/0"
#>  ..  ..@ metadata  :List of 5
#>  .. .. .. $ title      : chr "Race and ethnicity of the national Medicaid and CHIP population"
#>  .. .. .. $ description: chr "This data set includes annual counts and percentages of Medicaid and Childrens Health Insurance Program (CHIP) "| __truncated__
#>  .. .. .. $ modified   : Date[1:1], format: "2025-01-16"
#>  .. .. .. $ dictionary : chr NA
#>  .. .. .. $ download   : chr "https://download.medicaid.gov/data/race-and-ethnicity-2020-2022-01162025.csv"
#>  ..  ..@ dimensions: <class_dimensions>
#>  .. .. .. @ limit : int 8000
#>  .. .. .. @ rows  : int 216
#>  .. .. .. @ pages : int 1
#>  .. .. .. @ fields:List of 9
#>  .. .. .. .. $ year                          : chr "year"
#>  .. .. .. .. $ geography                     : chr "geography"
#>  .. .. .. .. $ subpopulation_topic           : chr "subpopulation_topic"
#>  .. .. .. .. $ subpopulation                 : chr "subpopulation"
#>  .. .. .. .. $ category                      : chr "category"
#>  .. .. .. .. $ count_of_enrollees            : chr "count_of_enrollees"
#>  .. .. .. .. $ denominator_count_of_enrollees: chr "denominator_count_of_enrollees"
#>  .. .. .. .. $ percentage_of_enrollees       : chr "percentage_of_enrollees"
#>  .. .. .. .. $ data_version                  : chr "data_version"
#>  .. $ DEMO_rural     : <caid_endpoint>
#>  ..  ..@ identifier: chr "https://data.medicaid.gov/api/1/datastore/query/73fdbf7c-569e-427b-9bff-eeed34a51dcb/0"
#>  ..  ..@ metadata  :List of 5
#>  .. .. .. $ title      : chr "Rural Medicaid and CHIP enrollees"
#>  .. .. .. $ description: chr "This data set includes annual counts and percentages of Medicaid and Childrens Health Insurance Program (CHIP) "| __truncated__
#>  .. .. .. $ modified   : Date[1:1], format: "2025-01-17"
#>  .. .. .. $ dictionary : chr NA
#>  .. .. .. $ download   : chr "https://download.medicaid.gov/data/rural-residence-2020-2022-01162025.csv"
#>  ..  ..@ dimensions: <class_dimensions>
#>  .. .. .. @ limit : int 8000
#>  .. .. .. @ rows  : int 603
#>  .. .. .. @ pages : int 1
#>  .. .. .. @ fields:List of 9
#>  .. .. .. .. $ year                          : chr "year"
#>  .. .. .. .. $ geography                     : chr "geography"
#>  .. .. .. .. $ subpopulation_topic           : chr "subpopulation_topic"
#>  .. .. .. .. $ subpopulation                 : chr "subpopulation"
#>  .. .. .. .. $ category                      : chr "category"
#>  .. .. .. .. $ count_of_enrollees            : chr "count_of_enrollees"
#>  .. .. .. .. $ denominator_count_of_enrollees: chr "denominator_count_of_enrollees"
#>  .. .. .. .. $ percentage_of_enrollees       : chr "percentage_of_enrollees"
#>  .. .. .. .. $ data_version                  : chr "data_version"
#>  .. $ DEMO_waive     : <caid_endpoint>
#>  ..  ..@ identifier: chr "https://data.medicaid.gov/api/1/datastore/query/4d4eaf55-33d3-4468-80b4-63553f4530ae/0"
#>  ..  ..@ metadata  :List of 5
#>  .. .. .. $ title      : chr "Section 1915(c) waiver program participants"
#>  .. .. .. $ description: chr "This data set includes annual counts and percentages of Medicaid and Childrens Health Insurance Program (CHIP) "| __truncated__
#>  .. .. .. $ modified   : Date[1:1], format: "2025-01-17"
#>  .. .. .. $ dictionary : chr NA
#>  .. .. .. $ download   : chr "https://download.medicaid.gov/data/1915c-waiver-2020-2022-01162025.csv"
#>  ..  ..@ dimensions: <class_dimensions>
#>  .. .. .. @ limit : int 8000
#>  .. .. .. @ rows  : int 450
#>  .. .. .. @ pages : int 1
#>  .. .. .. @ fields:List of 9
#>  .. .. .. .. $ year                          : chr "year"
#>  .. .. .. .. $ geography                     : chr "geography"
#>  .. .. .. .. $ subpopulation_topic           : chr "subpopulation_topic"
#>  .. .. .. .. $ subpopulation                 : chr "subpopulation"
#>  .. .. .. .. $ category                      : chr "category"
#>  .. .. .. .. $ count_of_enrollees            : chr "count_of_enrollees"
#>  .. .. .. .. $ denominator_count_of_enrollees: chr "denominator_count_of_enrollees"
#>  .. .. .. .. $ percentage_of_enrollees       : chr "percentage_of_enrollees"
#>  .. .. .. .. $ data_version                  : chr "data_version"
caid_temporal("MCP")
#> <caid_temporal>
#>  @ metadata  :List of 3
#>  .. $ title      : chr "Managed Care Programs by State"
#>  .. $ description: chr "Provides program names, program features, population enrolled, benefits covered, quality assurance and improvem"| __truncated__
#>  .. $ modified   : NULL
#>  @ dimensions: <class_dimensions>
#>  .. @ limit : int 8000
#>  .. @ rows  : int 179
#>  .. @ pages : int 1
#>  .. @ fields:List of 60
#>  .. .. $ features                                                        : chr "features"
#>  .. .. $ program_type                                                    : chr "program_type"
#>  .. .. $ statewide_or_regionspecific                                     : chr "statewide_or_regionspecific"
#>  .. .. $ federal_operating_authority                                     : chr "federal_operating_authority"
#>  .. .. $ program_start_date                                              : chr "program_start_date"
#>  .. .. $ waiver_expiration_date_if_applicable                            : chr "waiver_expiration_date_if_applicable"
#>  .. .. $ if_the_program_ended_in_2021_indicate_the_end_date              : chr "if_the_program_ended_in_2021_indicate_the_end_date"
#>  .. .. $ populations_enrolled_lowincome_adults_not_covered_under_aca_0778: chr "populations_enrolled_lowincome_adults_not_covered_under_aca_0778"
#>  .. .. $ populations_enrolled_lowincome_adults_covered_under_aca_sec_03d9: chr "populations_enrolled_lowincome_adults_covered_under_aca_sec_03d9"
#>  .. .. $ populations_enrolled_aged_blind_or_disabled_children_or_adults  : chr "populations_enrolled_aged_blind_or_disabled_children_or_adults"
#>  .. .. $ populations_enrolled_nondisabled_children_excludes_children_1dda: chr "populations_enrolled_nondisabled_children_excludes_children_1dda"
#>  .. .. $ populations_enrolled_individuals_receiving_limited_benefits_facc: chr "populations_enrolled_individuals_receiving_limited_benefits_facc"
#>  .. .. $ populations_enrolled_full_duals                                 : chr "populations_enrolled_full_duals"
#>  .. .. $ populations_enrolled_children_with_special_health_care_needs    : chr "populations_enrolled_children_with_special_health_care_needs"
#>  .. .. $ populations_enrolled_native_americanalaskan_natives             : chr "populations_enrolled_native_americanalaskan_natives"
#>  .. .. $ populations_enrolled_foster_care_and_adoption_assistance_ch_b424: chr "populations_enrolled_foster_care_and_adoption_assistance_ch_b424"
#>  .. .. $ populations_enrolled_enrollment_choice_period                   : chr "populations_enrolled_enrollment_choice_period"
#>  .. .. $ populations_enrolled_enrollment_broker_name_if_applicable       : chr "populations_enrolled_enrollment_broker_name_if_applicable"
#>  .. .. $ populations_enrolled_notes_on_enrollment_choice_period          : chr "populations_enrolled_notes_on_enrollment_choice_period"
#>  .. .. $ benefits_covered_inpatient_hospital_physical_health             : chr "benefits_covered_inpatient_hospital_physical_health"
#>  .. .. $ benefits_covered_inpatient_hospital_behavioral_health_mh_an_4041: chr "benefits_covered_inpatient_hospital_behavioral_health_mh_an_4041"
#>  .. .. $ benefits_covered_outpatient_hospital_physical_health            : chr "benefits_covered_outpatient_hospital_physical_health"
#>  .. .. $ benefits_covered_outpatient_hospital_behavioral_health_mh_a_ceee: chr "benefits_covered_outpatient_hospital_behavioral_health_mh_a_ceee"
#>  .. .. $ benefits_covered_partial_hospitalization                        : chr "benefits_covered_partial_hospitalization"
#>  .. .. $ benefits_covered_physician                                      : chr "benefits_covered_physician"
#>  .. .. $ benefits_covered_nurse_practitioner                             : chr "benefits_covered_nurse_practitioner"
#>  .. .. $ benefits_covered_rural_health_clinics_and_fqhcs                 : chr "benefits_covered_rural_health_clinics_and_fqhcs"
#>  .. .. $ benefits_covered_clinic_services                                : chr "benefits_covered_clinic_services"
#>  .. .. $ benefits_covered_lab_and_xray                                   : chr "benefits_covered_lab_and_xray"
#>  .. .. $ benefits_covered_prescription_drugs                             : chr "benefits_covered_prescription_drugs"
#>  .. .. $ benefits_covered_prosthetic_devices                             : chr "benefits_covered_prosthetic_devices"
#>  .. .. $ benefits_covered_epsdt                                          : chr "benefits_covered_epsdt"
#>  .. .. $ benefits_covered_case_management                                : chr "benefits_covered_case_management"
#>  .. .. $ benefits_covered_ssa_section_1945authorized_health_home         : chr "benefits_covered_ssa_section_1945authorized_health_home"
#>  .. .. $ benefits_covered_home_health_services_services_in_home          : chr "benefits_covered_home_health_services_services_in_home"
#>  .. .. $ benefits_covered_family_planning                                : chr "benefits_covered_family_planning"
#>  .. .. $ benefits_covered_dental_services_medicalsurgical                : chr "benefits_covered_dental_services_medicalsurgical"
#>  .. .. $ benefits_covered_dental_preventative_or_corrective              : chr "benefits_covered_dental_preventative_or_corrective"
#>  .. .. $ benefits_covered_personal_care_state_plan_option                : chr "benefits_covered_personal_care_state_plan_option"
#>  .. .. $ benefits_covered_hcbs_waiver_services                           : chr "benefits_covered_hcbs_waiver_services"
#>  .. .. $ benefits_covered_private_duty_nursing                           : chr "benefits_covered_private_duty_nursing"
#>  .. .. $ benefits_covered_icfidd                                         : chr "benefits_covered_icfidd"
#>  .. .. $ benefits_covered_nursing_facility_services                      : chr "benefits_covered_nursing_facility_services"
#>  .. .. $ benefits_covered_hospice_care                                   : chr "benefits_covered_hospice_care"
#>  .. .. $ benefits_covered_nonemergency_medical_transportation            : chr "benefits_covered_nonemergency_medical_transportation"
#>  .. .. $ benefits_covered_institution_for_mental_disease_inpatient_t_448a: chr "benefits_covered_institution_for_mental_disease_inpatient_t_448a"
#>  .. .. $ benefits_covered_other_eg_nurse_midwife_services_freestandi_d504: chr "benefits_covered_other_eg_nurse_midwife_services_freestandi_d504"
#>  .. .. $ quality_assurance_and_improvement_hedis_data_required           : chr "quality_assurance_and_improvement_hedis_data_required"
#>  .. .. $ quality_assurance_and_improvement_cahps_data_required           : chr "quality_assurance_and_improvement_cahps_data_required"
#>  .. .. $ quality_assurance_and_improvement_accreditation_required        : chr "quality_assurance_and_improvement_accreditation_required"
#>  .. .. $ quality_assurance_and_improvement_accrediting_organization      : chr "quality_assurance_and_improvement_accrediting_organization"
#>  .. .. $ quality_assurance_and_improvement_eqro_contractor_name_if_a_f2dc: chr "quality_assurance_and_improvement_eqro_contractor_name_if_a_f2dc"
#>  .. .. $ performance_incentives_payment_bonusesdifferentials_to_rewa_eece: chr "performance_incentives_payment_bonusesdifferentials_to_rewa_eece"
#>  .. .. $ performance_incentives_preferential_autoenrollment_to_rewar_ab96: chr "performance_incentives_preferential_autoenrollment_to_rewar_ab96"
#>  .. .. $ performance_incentives_public_reports_comparing_plan_perfor_42d0: chr "performance_incentives_public_reports_comparing_plan_perfor_42d0"
#>  .. .. $ performance_incentives_withholds_tied_to_performance_metrics    : chr "performance_incentives_withholds_tied_to_performance_metrics"
#>  .. .. $ performance_incentives_mcosphps_required_or_encouraged_to_p_bf59: chr "performance_incentives_mcosphps_required_or_encouraged_to_p_bf59"
#>  .. .. $ participating_plans_plans_in_program                            : chr "participating_plans_plans_in_program"
#>  .. .. $ notes_program_notes                                             : chr "notes_program_notes"
#>  .. .. $ state                                                           : chr "state"
#>  @ endpoints : tibble [6 × 5] (S3: tbl_df/tbl/data.frame)
#>  $ year      : int [1:6] 2022 2021 2020 2019 2018 2017
#>  $ identifier: chr [1:6] "https://data.medicaid.gov/api/1/datastore/query/ac74549c-a621-4f6b-a42d-1a05bc0eb5a3/0" "https://data.medicaid.gov/api/1/datastore/query/7459d190-e592-42e0-9b66-6f4e7e05eb9b/0" "https://data.medicaid.gov/api/1/datastore/query/8483cef4-972a-4c39-8290-c9d19bead88d/0" "https://data.medicaid.gov/api/1/datastore/query/72135da5-f8b4-475b-b602-91621f4205d9/0" ...
#>  $ modified  : Date[1:6], format: "2024-10-16" "2023-07-21" ...
#>  $ download  : chr [1:6] "https://data.medicaid.gov/sites/default/files/uploaded_resources/2022-Managed-care-prgrm-features-by-state.csv" "https://data.medicaid.gov/sites/default/files/uploaded_resources/2021-Table9_07212023_1.csv" "https://data.medicaid.gov/sites/default/files/uploaded_resources/2020-Table9_06132022.csv" "https://data.medicaid.gov/sites/default/files/uploaded_resources/2019_Table9_10_0.csv" ...
#>  $ dictionary: chr [1:6] NA NA NA NA ...