Medicare API Endpoint Classes

Medicare API Endpoint Classes

Usage

care_endpoint(alias)

care_temporal(alias)

pro_endpoint(alias)

Arguments

alias

<chr> endpoint alias

Value

An S7 <care_endpoint>, or <care_temporal> object

Examples

care_endpoint("care_dialysis")
#> <care_endpoint>
#>  @ identifier: chr "https://data.cms.gov/data-api/v1/dataset/f8610e87-ba25-43a3-a49e-927dbc8701ae/data-viewer"
#>  @ metadata  :List of 10
#>  .. $ title      : chr "Medicare Dialysis Facilities"
#>  .. $ description: chr "The Medicare Dialysis Facilities data provides information on clinical and patient measures for Medicare-certif"| __truncated__
#>  .. $ modified   : Date[1:1], format: "2024-02-29"
#>  .. $ temporal   : chr "2009-01-01 • 2024-12-31"
#>  .. $ periodicity: chr "Annually [R/P1Y]"
#>  .. $ download   : chr "https://data.cms.gov/sites/default/files/2024-02/b2c2658d-e19d-48db-8b3b-9d343a4d395f/dfr_facility_socrata_fy2024.csv"
#>  .. $ resources  : chr "https://data.cms.gov/data-api/v1/dataset-resources/f8610e87-ba25-43a3-a49e-927dbc8701ae"
#>  .. $ dictionary : chr "https://data.cms.gov/sites/default/files/2024-02/FY2024%20Data%20Dictionary.xlsx"
#>  .. $ site       : chr "https://data.cms.gov/quality-of-care/medicare-dialysis-facilities"
#>  .. $ references : chr "https://data.cms.gov/sites/default/files/2024-02/FY2024_DFR_Guide%20and%20Methodology.pdf"
#>  @ dimensions: <class_dimensions>
#>  .. @ limit : int 5000
#>  .. @ rows  : int 11679243
#>  .. @ pages : int 2336
#>  .. @ fields:List of 14
#>  .. .. $ state         : chr(0) 
#>  .. .. $ CCN           : chr(0) 
#>  .. .. $ Provider_Name : chr(0) 
#>  .. .. $ city          : chr(0) 
#>  .. .. $ Ownership_Type: chr(0) 
#>  .. .. $ ESRD_Network  : chr(0) 
#>  .. .. $ NPI           : chr(0) 
#>  .. .. $ Chain         : chr(0) 
#>  .. .. $ Modality      : chr(0) 
#>  .. .. $ Alternate_CCNs: chr(0) 
#>  .. .. $ Measure       : chr(0) 
#>  .. .. $ Measure_Score : chr(0) 
#>  .. .. $ year          : chr(0) 
#>  .. .. $ Measure_ID    : chr(0) 
care_temporal("quality_payment")
#> <care_temporal>
#>  @ metadata  :List of 5
#>  .. $ title      : chr "Quality Payment Program Experience"
#>  .. $ description: chr "The Quality Payment Program (QPP) Experience dataset provides participation and performance information in the "| __truncated__
#>  .. $ periodicity: chr "Annually [R/P1Y]"
#>  .. $ dictionary : chr "https://data.cms.gov/sites/default/files/2024-04/2022%20QPP%20Experience%20Report%20Public%20Use%20Files%20Data"| __truncated__
#>  .. $ site       : chr "https://data.cms.gov/quality-of-care/quality-payment-program-experience"
#>  @ dimensions: <class_dimensions>
#>  .. @ limit : int 5000
#>  .. @ rows  : int 624200
#>  .. @ pages : int 125
#>  .. @ fields:List of 165
#>  .. .. $ provider key                                     : chr(0) 
#>  .. .. $ practice state or us territory                   : chr(0) 
#>  .. .. $ practice size                                    : chr(0) 
#>  .. .. $ clinician type                                   : chr(0) 
#>  .. .. $ clinician specialty                              : chr(0) 
#>  .. .. $ years in medicare                                : chr(0) 
#>  .. .. $ npi                                              : chr(0) 
#>  .. .. $ non-reporting                                    : chr(0) 
#>  .. .. $ participation option                             : chr(0) 
#>  .. .. $ medicare patients                                : chr(0) 
#>  .. .. $ allowed charges                                  : chr(0) 
#>  .. .. $ services                                         : chr(0) 
#>  .. .. $ opted into mips                                  : chr(0) 
#>  .. .. $ small practice status                            : chr(0) 
#>  .. .. $ rural status                                     : chr(0) 
#>  .. .. $ health professional shortage area status         : chr(0) 
#>  .. .. $ ambulatory surgical center-based status          : chr(0) 
#>  .. .. $ hospital-based status                            : chr(0) 
#>  .. .. $ non-patient facing status                        : chr(0) 
#>  .. .. $ facility-based status                            : chr(0) 
#>  .. .. $ dual eligibility ratio                           : chr(0) 
#>  .. .. $ safety-net status                                : chr(0) 
#>  .. .. $ extreme uncontrollable circumstance (euc)        : chr(0) 
#>  .. .. $ final score                                      : chr(0) 
#>  .. .. $ payment adjustment percentage                    : chr(0) 
#>  .. .. $ complex patient bonus                            : chr(0) 
#>  .. .. $ quality reweighting (euc)                        : chr(0) 
#>  .. .. $ quality category score                           : chr(0) 
#>  .. .. $ quality improvement score                        : chr(0) 
#>  .. .. $ small practice bonus                             : chr(0) 
#>  .. .. $ quality measure id 1                             : chr(0) 
#>  .. .. $ quality measure collection type 1                : chr(0) 
#>  .. .. $ quality measure score 1                          : chr(0) 
#>  .. .. $ quality measure id 2                             : chr(0) 
#>  .. .. $ quality measure collection type 2                : chr(0) 
#>  .. .. $ quality measure score 2                          : chr(0) 
#>  .. .. $ quality measure id 3                             : chr(0) 
#>  .. .. $ quality measure collection type 3                : chr(0) 
#>  .. .. $ quality measure score 3                          : chr(0) 
#>  .. .. $ quality measure id 4                             : chr(0) 
#>  .. .. $ quality measure collection type 4                : chr(0) 
#>  .. .. $ quality measure score 4                          : chr(0) 
#>  .. .. $ quality measure id 5                             : chr(0) 
#>  .. .. $ quality measure collection type 5                : chr(0) 
#>  .. .. $ quality measure score 5                          : chr(0) 
#>  .. .. $ quality measure id 6                             : chr(0) 
#>  .. .. $ quality measure collection type 6                : chr(0) 
#>  .. .. $ quality measure score 6                          : chr(0) 
#>  .. .. $ quality measure id 7                             : chr(0) 
#>  .. .. $ quality measure collection type 7                : chr(0) 
#>  .. .. $ quality measure score 7                          : chr(0) 
#>  .. .. $ quality measure id 8                             : chr(0) 
#>  .. .. $ quality measure collection type 8                : chr(0) 
#>  .. .. $ quality measure score 8                          : chr(0) 
#>  .. .. $ quality measure id 9                             : chr(0) 
#>  .. .. $ quality measure collection type 9                : chr(0) 
#>  .. .. $ quality measure score 9                          : chr(0) 
#>  .. .. $ quality measure id 10                            : chr(0) 
#>  .. .. $ quality measure collection type 10               : chr(0) 
#>  .. .. $ quality measure score 10                         : chr(0) 
#>  .. .. $ quality measure id 11                            : chr(0) 
#>  .. .. $ quality measure collection type 11               : chr(0) 
#>  .. .. $ quality measure score 11                         : chr(0) 
#>  .. .. $ quality measure id 12                            : chr(0) 
#>  .. .. $ quality measure collection type 12               : chr(0) 
#>  .. .. $ quality measure score 12                         : chr(0) 
#>  .. .. $ promoting interoperability (pi) category score   : chr(0) 
#>  .. .. $ pi reweighting (euc)                             : chr(0) 
#>  .. .. $ pi reweighting (hardship exception)              : chr(0) 
#>  .. .. $ pi reweighting (special status or clinician type): chr(0) 
#>  .. .. $ cehrt id                                         : chr(0) 
#>  .. .. $ pi measure id 1                                  : chr(0) 
#>  .. .. $ pi measure type 1                                : chr(0) 
#>  .. .. $ pi measure score 1                               : chr(0) 
#>  .. .. $ pi measure id 2                                  : chr(0) 
#>  .. .. $ pi measure type 2                                : chr(0) 
#>  .. .. $ pi measure score 2                               : chr(0) 
#>  .. .. $ pi measure id 3                                  : chr(0) 
#>  .. .. $ pi measure type 3                                : chr(0) 
#>  .. .. $ pi measure score 3                               : chr(0) 
#>  .. .. $ pi measure id 4                                  : chr(0) 
#>  .. .. $ pi measure type 4                                : chr(0) 
#>  .. .. $ pi measure score 4                               : chr(0) 
#>  .. .. $ pi measure id 5                                  : chr(0) 
#>  .. .. $ pi measure type 5                                : chr(0) 
#>  .. .. $ pi measure score 5                               : chr(0) 
#>  .. .. $ pi measure id 6                                  : chr(0) 
#>  .. .. $ pi measure type 6                                : chr(0) 
#>  .. .. $ pi measure score 6                               : chr(0) 
#>  .. .. $ pi measure id 7                                  : chr(0) 
#>  .. .. $ pi measure type 7                                : chr(0) 
#>  .. .. $ pi measure score 7                               : chr(0) 
#>  .. .. $ pi measure id 8                                  : chr(0) 
#>  .. .. $ pi measure type 8                                : chr(0) 
#>  .. .. $ pi measure score 8                               : chr(0) 
#>  .. .. $ pi measure id 9                                  : chr(0) 
#>  .. .. $ pi measure type 9                                : chr(0) 
#>  .. .. $ pi measure score 9                               : chr(0) 
#>  .. .. $ pi measure id 10                                 : chr(0) 
#>  .. ..  [list output truncated]
#>  @ endpoints : tibble [6 × 6] (S3: tbl_df/tbl/data.frame)
#>  $ year      : int [1:6] 2022 2021 2020 2019 2018 2017
#>  $ modified  : Date[1:6], format: "2024-05-08" "2023-06-12" ...
#>  $ temporal  : chr [1:6] "2022-01-01 • 2022-12-31" "2021-01-01 • 2021-12-31" "2020-01-01 • 2020-12-31" "2019-01-01 • 2019-12-31" ...
#>  $ identifier: chr [1:6] "https://data.cms.gov/data-api/v1/dataset/b3438273-b4a6-44ca-8fb2-9e6026b74642/data" "https://data.cms.gov/data-api/v1/dataset/8c2ab3e4-597e-43c9-9d94-7223c66d7f25/data" "https://data.cms.gov/data-api/v1/dataset/48a484e6-ff02-4284-8666-6045b3cac17f/data" "https://data.cms.gov/data-api/v1/dataset/5776e77d-a844-49fd-b697-a2b01bdacdc7/data" ...
#>  $ download  : chr [1:6] "https://data.cms.gov/sites/default/files/2023-05/0bac88bc-a4c8-4dff-bdbf-660d24e6a689/2021_puf_QPPAR-18260.csv" "https://data.cms.gov/sites/default/files/2023-04/fb3758dd-6ace-4ddb-863b-1b895cef4d7e/2020_puf_qppar_QPPAR-17660.csv" "https://data.cms.gov/sites/default/files/2023-04/f9ef9f64-91d8-49a2-a1be-5a48bbdcc1e9/2019_puf_qppar_QPPAR-17661.csv" "https://data.cms.gov/sites/default/files/2023-04/d2a19b56-7a5f-49c7-b78d-4d0dd49b9391/2018_puf_qppar-QPPAR-17662.csv" ...
#>  $ resources : chr [1:6] "https://data.cms.gov/data-api/v1/dataset-resources/b3438273-b4a6-44ca-8fb2-9e6026b74642" "https://data.cms.gov/data-api/v1/dataset-resources/8c2ab3e4-597e-43c9-9d94-7223c66d7f25" "https://data.cms.gov/data-api/v1/dataset-resources/48a484e6-ff02-4284-8666-6045b3cac17f" "https://data.cms.gov/data-api/v1/dataset-resources/5776e77d-a844-49fd-b697-a2b01bdacdc7" ...
pro_endpoint("asc_facility")
#> <pro_endpoint>
#>  @ identifier: chr "https://data.cms.gov/provider-data/api/1/datastore/query/4jcv-atw7/0"
#>  @ metadata  :List of 9
#>  .. $ title      : chr "Ambulatory Surgical Center Quality Measures - Facility"
#>  .. $ description: chr "A list of ambulatory surgical centers participating in the Ambulatory Surgical Center Quality Reporting (ASCQR)"| __truncated__
#>  .. $ modified   : Date[1:1], format: "2025-04-08"
#>  .. $ group      : chr "Hospitals"
#>  .. $ issued     : Date[1:1], format: "2024-10-10"
#>  .. $ released   : Date[1:1], format: "2025-04-30"
#>  .. $ download   : chr "https://data.cms.gov/provider-data/sites/default/files/resources/dd03994fc93e296bb0297f1cd43cc987_1744668308/ASC_Facility.csv"
#>  .. $ dictionary : chr "https://data.cms.gov/provider-data/dataset/4jcv-atw7#data-dictionary"
#>  .. $ site       : chr "https://data.cms.gov/provider-data/dataset/4jcv-atw7"
#>  @ dimensions: <class_dimensions>
#>  .. @ limit : int 2000
#>  .. @ rows  : int 11288
#>  .. @ pages : int 6
#>  .. @ fields:List of 50
#>  .. .. $ facility_name             : chr(0) 
#>  .. .. $ facility_id               : chr(0) 
#>  .. .. $ npi                       : chr(0) 
#>  .. .. $ citytown                  : chr(0) 
#>  .. .. $ state                     : chr(0) 
#>  .. .. $ zip_code                  : chr(0) 
#>  .. .. $ year                      : chr(0) 
#>  .. .. $ asc1_rate                 : chr(0) 
#>  .. .. $ asc1_footnote             : chr(0) 
#>  .. .. $ asc2_rate                 : chr(0) 
#>  .. .. $ asc2_footnote             : chr(0) 
#>  .. .. $ asc3_rate                 : chr(0) 
#>  .. .. $ asc3_footnote             : chr(0) 
#>  .. .. $ asc4_rate                 : chr(0) 
#>  .. .. $ asc4_footnote             : chr(0) 
#>  .. .. $ asc9_rate                 : chr(0) 
#>  .. .. $ asc9_footnote             : chr(0) 
#>  .. .. $ asc11_rate                : chr(0) 
#>  .. .. $ asc11_footnote            : chr(0) 
#>  .. .. $ asc12_total_cases         : chr(0) 
#>  .. .. $ asc12_performance_category: chr(0) 
#>  .. .. $ asc12_rshv_rate           : chr(0) 
#>  .. .. $ asc12_interval_lower_limit: chr(0) 
#>  .. .. $ asc12_interval_upper_limit: chr(0) 
#>  .. .. $ asc12_footnote            : chr(0) 
#>  .. .. $ asc13_rate                : chr(0) 
#>  .. .. $ asc13_footnote            : chr(0) 
#>  .. .. $ asc14_rate                : chr(0) 
#>  .. .. $ asc14_footnote            : chr(0) 
#>  .. .. $ asc17_total_cases         : chr(0) 
#>  .. .. $ asc17_performance_category: chr(0) 
#>  .. .. $ asc17_rshv_rate           : chr(0) 
#>  .. .. $ asc17_interval_lower_limit: chr(0) 
#>  .. .. $ asc17_interval_upper_limit: chr(0) 
#>  .. .. $ asc17_footnote            : chr(0) 
#>  .. .. $ asc18_total_cases         : chr(0) 
#>  .. .. $ asc18_performance_category: chr(0) 
#>  .. .. $ asc18_rshv_rate           : chr(0) 
#>  .. .. $ asc18_interval_lower_limit: chr(0) 
#>  .. .. $ asc18_interval_upper_limit: chr(0) 
#>  .. .. $ asc18_footnote            : chr(0) 
#>  .. .. $ asc19_total_cases         : chr(0) 
#>  .. .. $ asc19_performance_category: chr(0) 
#>  .. .. $ asc19_rshv_rate           : chr(0) 
#>  .. .. $ asc19_interval_lower_limit: chr(0) 
#>  .. .. $ asc19_interval_upper_limit: chr(0) 
#>  .. .. $ asc19_footnote            : chr(0) 
#>  .. .. $ asc20_sample              : chr(0) 
#>  .. .. $ asc20_rate                : chr(0) 
#>  .. .. $ asc20_footnote            : chr(0)