Create Collection of Endpoints

Create Collection of Endpoints

Usage

new_collection(alias, call = caller_env())

Arguments

alias

<chr> Alias representing the CMS data endpoint.

call

<env> Environment to use for error reporting.

Value

S7 class_collection object.

Examples

new_collection("caid_demographics")
#> <class_collection>
#>  @ group  : chr "Medicaid and CHIP Enrollee Demographics"
#>  @ members:List of 8
#>  .. $ demo_wellvisit  : <class_endpoint>
#>  ..  ..@ catalog   : <class_clog>
#>  .. .. .. @ id: chr "caid"
#>  ..  ..@ metadata  :List of 5
#>  .. .. .. $ clog       : chr "caid"
#>  .. .. .. $ api        : chr "end"
#>  .. .. .. $ title      : chr "Medicaid and CHIP enrollees who received a well-child visit"
#>  .. .. .. $ description: chr "This data set includes annual counts and percentages of Medicaid and Childrens Health Insurance Program (CHIP) "| __truncated__
#>  .. .. .. $ modified   : Date[1:1], format: "2025-01-17"
#>  ..  ..@ dimensions: <class_dimensions>
#>  .. .. .. @ limit : int 8000
#>  .. .. .. @ rows  : int 63
#>  .. .. .. @ pages : int 1
#>  .. .. .. @ fields:List of 9
#>  .. .. .. .. $ year                          : chr(0) 
#>  .. .. .. .. $ geography                     : chr(0) 
#>  .. .. .. .. $ subpopulation_topic           : chr(0) 
#>  .. .. .. .. $ subpopulation                 : chr(0) 
#>  .. .. .. .. $ category                      : chr(0) 
#>  .. .. .. .. $ count_of_enrollees            : chr(0) 
#>  .. .. .. .. $ denominator_count_of_enrollees: chr(0) 
#>  .. .. .. .. $ percentage_of_enrollees       : chr(0) 
#>  .. .. .. .. $ data_version                  : chr(0) 
#>  ..  ..@ identifier: chr "https://data.medicaid.gov/api/1/datastore/query/289182e5-44c8-4fcc-96c4-4af69be9863f/0"
#>  .. $ demo_mental     : <class_endpoint>
#>  ..  ..@ catalog   : <class_clog>
#>  .. .. .. @ id: chr "caid"
#>  ..  ..@ metadata  :List of 5
#>  .. .. .. $ clog       : chr "caid"
#>  .. .. .. $ api        : chr "end"
#>  .. .. .. $ title      : chr "Medicaid and CHIP enrollees who received mental health or SUD services"
#>  .. .. .. $ description: chr "This data set includes annual counts and percentages of Medicaid and Childrens Health Insurance Program (CHIP) "| __truncated__
#>  .. .. .. $ modified   : Date[1:1], format: "2025-01-17"
#>  ..  ..@ dimensions: <class_dimensions>
#>  .. .. .. @ limit : int 8000
#>  .. .. .. @ rows  : int 216
#>  .. .. .. @ pages : int 1
#>  .. .. .. @ fields:List of 9
#>  .. .. .. .. $ year                          : chr(0) 
#>  .. .. .. .. $ geography                     : chr(0) 
#>  .. .. .. .. $ subpopulation_topic           : chr(0) 
#>  .. .. .. .. $ subpopulation                 : chr(0) 
#>  .. .. .. .. $ category                      : chr(0) 
#>  .. .. .. .. $ count_of_enrollees            : chr(0) 
#>  .. .. .. .. $ denominator_count_of_enrollees: chr(0) 
#>  .. .. .. .. $ percentage_of_enrollees       : chr(0) 
#>  .. .. .. .. $ data_version                  : chr(0) 
#>  ..  ..@ identifier: chr "https://data.medicaid.gov/api/1/datastore/query/8062e2f4-4c0a-41c9-8217-979468a80986/0"
#>  .. $ demo_disability : <class_endpoint>
#>  ..  ..@ catalog   : <class_clog>
#>  .. .. .. @ id: chr "caid"
#>  ..  ..@ metadata  :List of 5
#>  .. .. .. $ clog       : chr "caid"
#>  .. .. .. $ api        : chr "end"
#>  .. .. .. $ title      : chr "Medicaid enrollees who qualify for benefits based on disability"
#>  .. .. .. $ description: chr "This data set includes annual counts and percentages of Medicaid enrollees who are eligible for benefits based "| __truncated__
#>  .. .. .. $ modified   : Date[1:1], format: "2025-01-17"
#>  ..  ..@ dimensions: <class_dimensions>
#>  .. .. .. @ limit : int 8000
#>  .. .. .. @ rows  : int 114
#>  .. .. .. @ pages : int 1
#>  .. .. .. @ fields:List of 9
#>  .. .. .. .. $ year                          : chr(0) 
#>  .. .. .. .. $ geography                     : chr(0) 
#>  .. .. .. .. $ subpopulation_topic           : chr(0) 
#>  .. .. .. .. $ subpopulation                 : chr(0) 
#>  .. .. .. .. $ category                      : chr(0) 
#>  .. .. .. .. $ count_of_enrollees            : chr(0) 
#>  .. .. .. .. $ denominator_count_of_enrollees: chr(0) 
#>  .. .. .. .. $ percentage_of_enrollees       : chr(0) 
#>  .. .. .. .. $ data_version                  : chr(0) 
#>  ..  ..@ identifier: chr "https://data.medicaid.gov/api/1/datastore/query/b12e390f-1721-47c0-bb38-b0b1805e0446/0"
#>  .. $ demo_prematurity: <class_endpoint>
#>  ..  ..@ catalog   : <class_clog>
#>  .. .. .. @ id: chr "caid"
#>  ..  ..@ metadata  :List of 5
#>  .. .. .. $ clog       : chr "caid"
#>  .. .. .. $ api        : chr "end"
#>  .. .. .. $ title      : chr "Prematurity and severe maternal morbidity among Medicaid- and CHIP-covered live births"
#>  .. .. .. $ description: chr "This data set includes annual counts and rates of Medicaid- and Childrens Health Insurance Program (CHIP)-cover"| __truncated__
#>  .. .. .. $ modified   : Date[1:1], format: "2025-01-17"
#>  ..  ..@ dimensions: <class_dimensions>
#>  .. .. .. @ limit : int 8000
#>  .. .. .. @ rows  : int 435
#>  .. .. .. @ pages : int 1
#>  .. .. .. @ fields:List of 9
#>  .. .. .. .. $ year                           : chr(0) 
#>  .. .. .. .. $ geography                      : chr(0) 
#>  .. .. .. .. $ subpopulation_topic            : chr(0) 
#>  .. .. .. .. $ subpopulation                  : chr(0) 
#>  .. .. .. .. $ category                       : chr(0) 
#>  .. .. .. .. $ count_of_deliveries            : chr(0) 
#>  .. .. .. .. $ denominator_count_of_deliveries: chr(0) 
#>  .. .. .. .. $ rate_of_deliveries             : chr(0) 
#>  .. .. .. .. $ data_version                   : chr(0) 
#>  ..  ..@ identifier: chr "https://data.medicaid.gov/api/1/datastore/query/ee3b9534-0d19-4c1b-bf74-43f898d5de7c/0"
#>  .. $ demo_language   : <class_endpoint>
#>  ..  ..@ catalog   : <class_clog>
#>  .. .. .. @ id: chr "caid"
#>  ..  ..@ metadata  :List of 5
#>  .. .. .. $ clog       : chr "caid"
#>  .. .. .. $ api        : chr "end"
#>  .. .. .. $ title      : chr "Primary language spoken by the Medicaid and CHIP population"
#>  .. .. .. $ description: chr "This data set includes annual counts and percentages of Medicaid and Childrens Health Insurance Program (CHIP) "| __truncated__
#>  .. .. .. $ modified   : Date[1:1], format: "2025-01-18"
#>  ..  ..@ dimensions: <class_dimensions>
#>  .. .. .. @ limit : int 8000
#>  .. .. .. @ rows  : int 675
#>  .. .. .. @ pages : int 1
#>  .. .. .. @ fields:List of 9
#>  .. .. .. .. $ year                          : chr(0) 
#>  .. .. .. .. $ geography                     : chr(0) 
#>  .. .. .. .. $ subpopulation_topic           : chr(0) 
#>  .. .. .. .. $ subpopulation                 : chr(0) 
#>  .. .. .. .. $ category                      : chr(0) 
#>  .. .. .. .. $ count_of_enrollees            : chr(0) 
#>  .. .. .. .. $ denominator_count_of_enrollees: chr(0) 
#>  .. .. .. .. $ percentage_of_enrollees       : chr(0) 
#>  .. .. .. .. $ data_version                  : chr(0) 
#>  ..  ..@ identifier: chr "https://data.medicaid.gov/api/1/datastore/query/3df86994-c0e4-4106-a404-e95e87c824e6/0"
#>  .. $ demo_ethnicity  : <class_endpoint>
#>  ..  ..@ catalog   : <class_clog>
#>  .. .. .. @ id: chr "caid"
#>  ..  ..@ metadata  :List of 5
#>  .. .. .. $ clog       : chr "caid"
#>  .. .. .. $ api        : chr "end"
#>  .. .. .. $ title      : chr "Race and ethnicity of the national Medicaid and CHIP population"
#>  .. .. .. $ description: chr "This data set includes annual counts and percentages of Medicaid and Childrens Health Insurance Program (CHIP) "| __truncated__
#>  .. .. .. $ modified   : Date[1:1], format: "2025-01-16"
#>  ..  ..@ dimensions: <class_dimensions>
#>  .. .. .. @ limit : int 8000
#>  .. .. .. @ rows  : int 216
#>  .. .. .. @ pages : int 1
#>  .. .. .. @ fields:List of 9
#>  .. .. .. .. $ year                          : chr(0) 
#>  .. .. .. .. $ geography                     : chr(0) 
#>  .. .. .. .. $ subpopulation_topic           : chr(0) 
#>  .. .. .. .. $ subpopulation                 : chr(0) 
#>  .. .. .. .. $ category                      : chr(0) 
#>  .. .. .. .. $ count_of_enrollees            : chr(0) 
#>  .. .. .. .. $ denominator_count_of_enrollees: chr(0) 
#>  .. .. .. .. $ percentage_of_enrollees       : chr(0) 
#>  .. .. .. .. $ data_version                  : chr(0) 
#>  ..  ..@ identifier: chr "https://data.medicaid.gov/api/1/datastore/query/b9e18d80-c765-4fb9-95e7-81951e01a6a3/0"
#>  .. $ demo_rural      : <class_endpoint>
#>  ..  ..@ catalog   : <class_clog>
#>  .. .. .. @ id: chr "caid"
#>  ..  ..@ metadata  :List of 5
#>  .. .. .. $ clog       : chr "caid"
#>  .. .. .. $ api        : chr "end"
#>  .. .. .. $ title      : chr "Rural Medicaid and CHIP enrollees"
#>  .. .. .. $ description: chr "This data set includes annual counts and percentages of Medicaid and Childrens Health Insurance Program (CHIP) "| __truncated__
#>  .. .. .. $ modified   : Date[1:1], format: "2025-01-17"
#>  ..  ..@ dimensions: <class_dimensions>
#>  .. .. .. @ limit : int 8000
#>  .. .. .. @ rows  : int 603
#>  .. .. .. @ pages : int 1
#>  .. .. .. @ fields:List of 9
#>  .. .. .. .. $ year                          : chr(0) 
#>  .. .. .. .. $ geography                     : chr(0) 
#>  .. .. .. .. $ subpopulation_topic           : chr(0) 
#>  .. .. .. .. $ subpopulation                 : chr(0) 
#>  .. .. .. .. $ category                      : chr(0) 
#>  .. .. .. .. $ count_of_enrollees            : chr(0) 
#>  .. .. .. .. $ denominator_count_of_enrollees: chr(0) 
#>  .. .. .. .. $ percentage_of_enrollees       : chr(0) 
#>  .. .. .. .. $ data_version                  : chr(0) 
#>  ..  ..@ identifier: chr "https://data.medicaid.gov/api/1/datastore/query/73fdbf7c-569e-427b-9bff-eeed34a51dcb/0"
#>  .. $ demo_waiver     : <class_endpoint>
#>  ..  ..@ catalog   : <class_clog>
#>  .. .. .. @ id: chr "caid"
#>  ..  ..@ metadata  :List of 5
#>  .. .. .. $ clog       : chr "caid"
#>  .. .. .. $ api        : chr "end"
#>  .. .. .. $ title      : chr "Section 1915(c) waiver program participants"
#>  .. .. .. $ description: chr "This data set includes annual counts and percentages of Medicaid and Childrens Health Insurance Program (CHIP) "| __truncated__
#>  .. .. .. $ modified   : Date[1:1], format: "2025-01-17"
#>  ..  ..@ dimensions: <class_dimensions>
#>  .. .. .. @ limit : int 8000
#>  .. .. .. @ rows  : int 450
#>  .. .. .. @ pages : int 1
#>  .. .. .. @ fields:List of 9
#>  .. .. .. .. $ year                          : chr(0) 
#>  .. .. .. .. $ geography                     : chr(0) 
#>  .. .. .. .. $ subpopulation_topic           : chr(0) 
#>  .. .. .. .. $ subpopulation                 : chr(0) 
#>  .. .. .. .. $ category                      : chr(0) 
#>  .. .. .. .. $ count_of_enrollees            : chr(0) 
#>  .. .. .. .. $ denominator_count_of_enrollees: chr(0) 
#>  .. .. .. .. $ percentage_of_enrollees       : chr(0) 
#>  .. .. .. .. $ data_version                  : chr(0) 
#>  ..  ..@ identifier: chr "https://data.medicaid.gov/api/1/datastore/query/4d4eaf55-33d3-4468-80b4-63553f4530ae/0"
new_collection("caid_unwind")
#> <class_collection>
#>  @ group  : chr "Medicaid Unwinding Report"
#>  @ members:List of 4
#>  .. $ unwind_marketplace: <class_endpoint>
#>  ..  ..@ catalog   : <class_clog>
#>  .. .. .. @ id: chr "caid"
#>  ..  ..@ metadata  :List of 6
#>  .. .. .. $ clog       : chr "caid"
#>  .. .. .. $ api        : chr "end"
#>  .. .. .. $ title      : chr "HealthCare.gov Marketplace Medicaid Unwinding Report"
#>  .. .. .. $ description: chr "Metrics from individual Marketplaces during the current reporting period. The report includes data for the stat"| __truncated__
#>  .. .. .. $ modified   : Date[1:1], format: "2024-08-30"
#>  .. .. .. $ periodicity: chr "Decennially [R/P10Y]"
#>  ..  ..@ dimensions: <class_dimensions>
#>  .. .. .. @ limit : int 8000
#>  .. .. .. @ rows  : int 204
#>  .. .. .. @ pages : int 1
#>  .. .. .. @ fields:List of 30
#>  .. .. .. .. $ state                           : chr(0) 
#>  .. .. .. .. $ metric                          : chr(0) 
#>  .. .. .. .. $ april_2023_net_count            : chr(0) 
#>  .. .. .. .. $ april_2023_net_percentage       : chr(0) 
#>  .. .. .. .. $ may_2023_net_count              : chr(0) 
#>  .. .. .. .. $ may_2023_net_percentage         : chr(0) 
#>  .. .. .. .. $ june_2023_net_count             : chr(0) 
#>  .. .. .. .. $ june_2023_net_percentage        : chr(0) 
#>  .. .. .. .. $ july_2023_net_count             : chr(0) 
#>  .. .. .. .. $ july_2023_net_percentage        : chr(0) 
#>  .. .. .. .. $ august_2023_net_count           : chr(0) 
#>  .. .. .. .. $ august_2023_net_percentage      : chr(0) 
#>  .. .. .. .. $ september_2023_net_count        : chr(0) 
#>  .. .. .. .. $ september_2023_net_percentage   : chr(0) 
#>  .. .. .. .. $ october_2023_net_count          : chr(0) 
#>  .. .. .. .. $ october_2023_net_percentage     : chr(0) 
#>  .. .. .. .. $ november_2023_net_count         : chr(0) 
#>  .. .. .. .. $ november_2023_net_percentage    : chr(0) 
#>  .. .. .. .. $ december_2023_net_count         : chr(0) 
#>  .. .. .. .. $ december_2023_net_percentage    : chr(0) 
#>  .. .. .. .. $ january_2024_net_count          : chr(0) 
#>  .. .. .. .. $ january_2024_net_percentage     : chr(0) 
#>  .. .. .. .. $ february_2024_net_count         : chr(0) 
#>  .. .. .. .. $ february_2024_net_percentage    : chr(0) 
#>  .. .. .. .. $ march_2024_net_count            : chr(0) 
#>  .. .. .. .. $ march_2024_net_percentage       : chr(0) 
#>  .. .. .. .. $ april_2024_net_count            : chr(0) 
#>  .. .. .. .. $ april_2024_net_percentage       : chr(0) 
#>  .. .. .. .. $ april_2024_cumulative_count     : chr(0) 
#>  .. .. .. .. $ april_cumulative_2024_percentage: chr(0) 
#>  ..  ..@ identifier: chr "https://data.medicaid.gov/api/1/datastore/query/9a83ba5e-05f5-47f5-82de-f3a59233a912/0"
#>  .. $ unwind_transition : <class_endpoint>
#>  ..  ..@ catalog   : <class_clog>
#>  .. .. .. @ id: chr "caid"
#>  ..  ..@ metadata  :List of 6
#>  .. .. .. $ clog       : chr "caid"
#>  .. .. .. $ api        : chr "end"
#>  .. .. .. $ title      : chr "HealthCare.gov Transitions Marketplace Medicaid Unwinding Report"
#>  .. .. .. $ description: chr "Metrics from individual Marketplaces during the current reporting period. The report includes data for the stat"| __truncated__
#>  .. .. .. $ modified   : Date[1:1], format: "2024-12-27"
#>  .. .. .. $ periodicity: chr "Decennially [R/P10Y]"
#>  ..  ..@ dimensions: <class_dimensions>
#>  .. .. .. @ limit : int 8000
#>  .. .. .. @ rows  : int 59527
#>  .. .. .. @ pages : int 8
#>  .. .. .. @ fields:List of 9
#>  .. .. .. .. $ state                                                           : chr(0) 
#>  .. .. .. .. $ data_record_note                                                : chr(0) 
#>  .. .. .. .. $ metric                                                          : chr(0) 
#>  .. .. .. .. $ time_period                                                     : chr(0) 
#>  .. .. .. .. $ release_through                                                 : chr(0) 
#>  .. .. .. .. $ count_of_individuals_whose_medicaid_or_chip_coverage_was_te_dd11: chr(0) 
#>  .. .. .. .. $ percentage_of_individuals_whose_medicaid_or_chip_coverage_w_da6e: chr(0) 
#>  .. .. .. .. $ cumulative_count_of_individuals_whose_medicaid_or_chip_cove_28d9: chr(0) 
#>  .. .. .. .. $ cumulative_percentage_of_individuals_whose_medicaid_or_chip_6ac2: chr(0) 
#>  ..  ..@ identifier: chr "https://data.medicaid.gov/api/1/datastore/query/5636a78c-fe18-4229-aee1-e40fa910a8a0/0"
#>  .. $ unwind_historic   : <class_endpoint>
#>  ..  ..@ catalog   : <class_clog>
#>  .. .. .. @ id: chr "caid"
#>  ..  ..@ metadata  :List of 6
#>  .. .. .. $ clog       : chr "caid"
#>  .. .. .. $ api        : chr "end"
#>  .. .. .. $ title      : chr "Separate CHIP Enrollment by Month and State Historic CAA/Unwinding Period"
#>  .. .. .. $ description: chr "<p>This historic dataset with total enrollment in separate CHIP programs by month and state was created to fulf"| __truncated__
#>  .. .. .. $ modified   : Date[1:1], format: "2024-12-27"
#>  .. .. .. $ periodicity: chr "Monthly [R/P1M]"
#>  ..  ..@ dimensions: <class_dimensions>
#>  .. .. .. @ limit : int 8000
#>  .. .. .. @ rows  : int 780
#>  .. .. .. @ pages : int 1
#>  .. .. .. @ fields:List of 5
#>  .. .. .. .. $ state           : chr(0) 
#>  .. .. .. .. $ reporting_period: chr(0) 
#>  .. .. .. .. $ release_date    : chr(0) 
#>  .. .. .. .. $ value           : chr(0) 
#>  .. .. .. .. $ data_notes      : chr(0) 
#>  ..  ..@ identifier: chr "https://data.medicaid.gov/api/1/datastore/query/d30cfc7c-4b32-4df1-b2bf-e0a850befd77/0"
#>  .. $ unwind_sbm        : <class_endpoint>
#>  ..  ..@ catalog   : <class_clog>
#>  .. .. .. @ id: chr "caid"
#>  ..  ..@ metadata  :List of 6
#>  .. .. .. $ clog       : chr "caid"
#>  .. .. .. $ api        : chr "end"
#>  .. .. .. $ title      : chr "State-based Marketplace (SBM) Medicaid Unwinding Report"
#>  .. .. .. $ description: chr "Metrics from individual Marketplaces during the current reporting period. The report includes data for the stat"| __truncated__
#>  .. .. .. $ modified   : Date[1:1], format: "2024-12-27"
#>  .. .. .. $ periodicity: chr "Decennially [R/P10Y]"
#>  ..  ..@ dimensions: <class_dimensions>
#>  .. .. .. @ limit : int 8000
#>  .. .. .. @ rows  : int 128
#>  .. .. .. @ pages : int 1
#>  .. .. .. @ fields:List of 59
#>  .. .. .. .. $ state                                      : chr(0) 
#>  .. .. .. .. $ metric                                     : chr(0) 
#>  .. .. .. .. $ april_2023_count                           : chr(0) 
#>  .. .. .. .. $ april_2023_percentage_applications         : chr(0) 
#>  .. .. .. .. $ april_2023_percentage_account_transfers    : chr(0) 
#>  .. .. .. .. $ may_2023_count                             : chr(0) 
#>  .. .. .. .. $ may_2023_percentage_applications           : chr(0) 
#>  .. .. .. .. $ may_2023_percentage_account_transfers      : chr(0) 
#>  .. .. .. .. $ june_2023_count                            : chr(0) 
#>  .. .. .. .. $ june_2023_percentage_applications          : chr(0) 
#>  .. .. .. .. $ june_2023_percentage_account_transfers     : chr(0) 
#>  .. .. .. .. $ july_2023_count                            : chr(0) 
#>  .. .. .. .. $ july_2023_percentage_applications          : chr(0) 
#>  .. .. .. .. $ july_2023_percentage_account_transfers     : chr(0) 
#>  .. .. .. .. $ august_2023_count                          : chr(0) 
#>  .. .. .. .. $ august_2023_percentage_applications        : chr(0) 
#>  .. .. .. .. $ august_2023_percentage_account_transfers   : chr(0) 
#>  .. .. .. .. $ september_2023_count                       : chr(0) 
#>  .. .. .. .. $ september_2023_percentage_applications     : chr(0) 
#>  .. .. .. .. $ september_2023_percentage_account_transfers: chr(0) 
#>  .. .. .. .. $ october_2023_count                         : chr(0) 
#>  .. .. .. .. $ october_2023_percentage_applications       : chr(0) 
#>  .. .. .. .. $ october_2023_percentage_account_transfers  : chr(0) 
#>  .. .. .. .. $ november_2023_count                        : chr(0) 
#>  .. .. .. .. $ november_2023_percentage_applications      : chr(0) 
#>  .. .. .. .. $ november_2023_percentage_account_transfers : chr(0) 
#>  .. .. .. .. $ december_2023_count                        : chr(0) 
#>  .. .. .. .. $ december_2023_percentage_applications      : chr(0) 
#>  .. .. .. .. $ december_2023_percentage_account_transfers : chr(0) 
#>  .. .. .. .. $ january_2024_count                         : chr(0) 
#>  .. .. .. .. $ january_2024_percentage_applications       : chr(0) 
#>  .. .. .. .. $ january_2024_percentage_account_transfers  : chr(0) 
#>  .. .. .. .. $ february_2024_count                        : chr(0) 
#>  .. .. .. .. $ february_2024_percentage_applications      : chr(0) 
#>  .. .. .. .. $ february_2024_percentage_account_transfers : chr(0) 
#>  .. .. .. .. $ march_2024_count                           : chr(0) 
#>  .. .. .. .. $ march_2024_percentage_applications         : chr(0) 
#>  .. .. .. .. $ march_2024_percentage_account_transfers    : chr(0) 
#>  .. .. .. .. $ april_2024_count                           : chr(0) 
#>  .. .. .. .. $ april_2024_percentage_applications         : chr(0) 
#>  .. .. .. .. $ april_2024_percentage_account_transfers    : chr(0) 
#>  .. .. .. .. $ may_2024_count                             : chr(0) 
#>  .. .. .. .. $ may_2024_percentage_applications           : chr(0) 
#>  .. .. .. .. $ may_2024_percentage_account_transfers      : chr(0) 
#>  .. .. .. .. $ june_2024_count                            : chr(0) 
#>  .. .. .. .. $ june_2024_percentage_applications          : chr(0) 
#>  .. .. .. .. $ june_2024_percentage_account_transfers     : chr(0) 
#>  .. .. .. .. $ july_2024_count                            : chr(0) 
#>  .. .. .. .. $ july_2024_percentage_applications          : chr(0) 
#>  .. .. .. .. $ july_2024_percentage_account_transfers     : chr(0) 
#>  .. .. .. .. $ august_2024_count                          : chr(0) 
#>  .. .. .. .. $ august_2024_percentage_applications        : chr(0) 
#>  .. .. .. .. $ august_2024_percentage_account_transfers   : chr(0) 
#>  .. .. .. .. $ september_2024_count                       : chr(0) 
#>  .. .. .. .. $ september_2024_percentage_applications     : chr(0) 
#>  .. .. .. .. $ september_2024_percentage_account_transfers: chr(0) 
#>  .. .. .. .. $ total_count                                : chr(0) 
#>  .. .. .. .. $ total_percentage_applications              : chr(0) 
#>  .. .. .. .. $ total_percentage_account_transfers         : chr(0) 
#>  ..  ..@ identifier: chr "https://data.medicaid.gov/api/1/datastore/query/5670e72c-e44e-4282-ab67-4ebebaba3cbd/0"
new_collection("caid_services") |> query_nresults()
#> $service_acute
#> [1] 25440
#> 
#> $service_behavior
#> [1] 31800
#> 
#> $service_perinatal
#> [1] 12720
#> 
#> $service_screening
#> [1] 3180
#> 
#> $service_contraceptive
#> [1] 6360
#> 
#> $service_dental
#> [1] 3180
#> 
#> $service_pregnancy
#> [1] 6360
#> 
#> $service_telehealth
#> [1] 12720
#> 
#> $service_vaccination
#> [1] 47700
#> 
#> $service_bloodlead
#> [1] 3180
#> 
#> $service_respiratory
#> [1] 28620
#>