Create Groups of Endpoints

Create Groups of Endpoints

Usage

new_group(
  member_names,
  group_name = NULL,
  quick = FALSE,
  offset = 0,
  limit = 10
)

Arguments

member_names

<chr> Alias representing the CMS data endpoint.

group_name

<chr> Name of the group.

quick

<lgl> call quick_() method on group if TRUE.

offset

<int> Offset for pagination.

limit

<int> Limit for pagination.

Value

S7 class_group object.

Examples

new_group(c("local_help", "managed_care_share"))
#> <class_group>
#>  @ group  : chr "local_help | managed_care_share"
#>  @ members:List of 2
#>  .. $ local_help        : <class_endpoint>
#>  ..  ..@ identifier: chr "https://data.healthcare.gov/api/1/datastore/query/3ddf85bc-f71b-4417-b271-410cbf9e0905/0"
#>  ..  ..@ metadata  :List of 6
#>  .. .. .. $ api        : chr "HealthcareGov"
#>  .. .. .. $ title      : chr "Find Local Help"
#>  .. .. .. $ description: chr "CSV output of the Assister and Agent/Broker listing dataset used within the Find Local Help application [1]. Th"| __truncated__
#>  .. .. .. $ modified   : Date[1:1], format: "2025-05-24"
#>  .. .. .. $ periodicity: chr "Daily [R/P1D]"
#>  .. .. .. $ download   : chr "https://data.healthcare.gov/datafile/localhelp/prod/helpers.csv"
#>  ..  ..@ dimensions: <class_dimensions>
#>  .. .. .. @ limit : int 500
#>  .. .. .. @ rows  : int 82707
#>  .. .. .. @ pages : int 166
#>  .. .. .. @ fields:List of 23
#>  .. .. .. .. $ record_id                                   : chr(0) 
#>  .. .. .. .. $ data_set_name                               : chr(0) 
#>  .. .. .. .. $ organization_name                           : chr(0) 
#>  .. .. .. .. $ main_phone_number                           : chr(0) 
#>  .. .. .. .. $ street_address                              : chr(0) 
#>  .. .. .. .. $ city                                        : chr(0) 
#>  .. .. .. .. $ state_or_territory                          : chr(0) 
#>  .. .. .. .. $ zip_code                                    : chr(0) 
#>  .. .. .. .. $ non_english_assistance_offered              : chr(0) 
#>  .. .. .. .. $ organization_acronym                        : chr(0) 
#>  .. .. .. .. $ office_hours                                : chr(0) 
#>  .. .. .. .. $ toll_free_phone                             : chr(0) 
#>  .. .. .. .. $ tty_phone                                   : chr(0) 
#>  .. .. .. .. $ language_line_phone_for_non_english_speakers: chr(0) 
#>  .. .. .. .. $ location_name                               : chr(0) 
#>  .. .. .. .. $ website_url                                 : chr(0) 
#>  .. .. .. .. $ additional_information                      : chr(0) 
#>  .. .. .. .. $ accessibility_service_offered               : chr(0) 
#>  .. .. .. .. $ geo_location                                : chr(0) 
#>  .. .. .. .. $ email_address                               : chr(0) 
#>  .. .. .. .. $ display                                     : chr(0) 
#>  .. .. .. .. $ display_address                             : chr(0) 
#>  .. .. .. .. $ address_comment                             : chr(0) 
#>  .. $ managed_care_share: <class_endpoint>
#>  ..  ..@ identifier: chr "https://data.medicaid.gov/api/1/datastore/query/79692ea5-21e1-56bf-8149-97d437120c4b/0"
#>  ..  ..@ metadata  :List of 6
#>  .. .. .. $ api        : chr "Medicaid"
#>  .. .. .. $ title      : chr "Share of Medicaid Enrollees in Managed Care"
#>  .. .. .. $ description: chr "The Share of Medicaid Enrollees in any Managed Care and in Comprehensive Managed CaAre profiles state-level enr"| __truncated__
#>  .. .. .. $ modified   : Date[1:1], format: "2024-10-16"
#>  .. .. .. $ periodicity: chr "Decennially [R/P10Y]"
#>  .. .. .. $ download   : chr "https://data.medicaid.gov/sites/default/files/uploaded_resources/share-of-medicaid-enrollees-in-managed-care.2022.csv"
#>  ..  ..@ dimensions: <class_dimensions>
#>  .. .. .. @ limit : int 8000
#>  .. .. .. @ rows  : int 399
#>  .. .. .. @ pages : int 1
#>  .. .. .. @ fields:List of 8
#>  .. .. .. .. $ state                                          : chr(0) 
#>  .. .. .. .. $ notes                                          : chr(0) 
#>  .. .. .. .. $ total_medicaid_enrollees                       : chr(0) 
#>  .. .. .. .. $ individuals_enrolled_any                       : chr(0) 
#>  .. .. .. .. $ percent_of_all_medicaid_enrollees_any          : chr(0) 
#>  .. .. .. .. $ individuals_enrolled_comprehensive             : chr(0) 
#>  .. .. .. .. $ percent_of_all_medicaid_enrollees_comprehensive: chr(0) 
#>  .. .. .. .. $ year                                           : chr(0) 
new_group(c("local_help", "managed_care_share"), quick = TRUE)
#> $local_help
#> # A tibble: 10 × 23
#>    record_id data_set_name    organization_name main_phone_number street_address
#>    <chr>     <chr>            <chr>             <chr>             <chr>         
#>  1 1428646   Agent/Broker (A… NA                7085291060        NA            
#>  2 1424965   Agent/Broker (A… NA                7866704302        NA            
#>  3 21643     Certified Appli… Central Texas Co… 5129789015        1210 W. Brake…
#>  4 1362234   Agent/Broker (A… NA                8643990510        28 Plainfield…
#>  5 1307876   Agent/Broker (A… NA                6822945885        971 W 80 TH P…
#>  6 28359     Certified Appli… Community Health… 8019552388        980 Medical D…
#>  7 1450279   Agent/Broker (A… NA                3162086089        1605 E 2nd St…
#>  8 1525446   Agent/Broker (A… NA                4078790006        2796 Club Cor…
#>  9 68200     Certified Appli… JeffCare - a pro… 5048385257        3616 S. I-10 …
#> 10 1373616   Agent/Broker (A… NA                5122988536        NA            
#> # ℹ 18 more variables: city <chr>, state_or_territory <chr>, zip_code <chr>,
#> #   non_english_assistance_offered <chr>, organization_acronym <chr>,
#> #   office_hours <chr>, toll_free_phone <chr>, tty_phone <chr>,
#> #   language_line_phone_for_non_english_speakers <chr>, location_name <chr>,
#> #   website_url <chr>, additional_information <chr>,
#> #   accessibility_service_offered <chr>, geo_location <chr>,
#> #   email_address <chr>, display <chr>, display_address <chr>, …
#> 
#> $managed_care_share
#> # A tibble: 10 × 8
#>    state                notes      total_medicaid_enrol…¹ individuals_enrolled…²
#>    <chr>                <chr>      <chr>                  <chr>                 
#>  1 TOTALS               NA         80,263,839             65,034,032            
#>  2 Arkansas5            Arkansas … 1,026,621              512,041               
#>  3 District of Columbia NA         251,791                180,942               
#>  4 Wyoming              NA         64,442                 391                   
#>  5 Pennsylvania         NA         2,753,618              2,521,421             
#>  6 Massachusetts        NA         1,889,306              1,268,120             
#>  7 North Dakota         NA         93,422                 48,621                
#>  8 Michigan             NA         4,448,582              4,370,138             
#>  9 Virginia             NA         1,111,999              761,019               
#> 10 Hawaii               NA         358,302                354,289               
#> # ℹ abbreviated names: ¹​total_medicaid_enrollees, ²​individuals_enrolled_any
#> # ℹ 4 more variables: percent_of_all_medicaid_enrollees_any <chr>,
#> #   individuals_enrolled_comprehensive <chr>,
#> #   percent_of_all_medicaid_enrollees_comprehensive <chr>, year <chr>
#>