Search the NLM NPI Registry

Search the NLM NPI Registry

Search the NPPES NPI Registry

Usage

npi_nlm(terms, npi = NULL)

npi_nppes(
  npi = NULL,
  entity = NULL,
  first = NULL,
  last = NULL,
  organization = NULL,
  name_type = NULL,
  taxonomy_desc = NULL,
  city = NULL,
  state = NULL,
  zip = NULL,
  country = NULL
)

Source

API Documentation: Individuals

API Documentation: Organizations

API Documentation

Arguments

terms

<chr> Search terms, separated by spaces

npi

<chr> Unique 10-digit National Provider Identifier number issued by CMS to US healthcare providers through NPPES.

entity

<chr> Entity type; one of either I for Individual (NPI-1) or O for Organizational (NPI-2)

first, last

<chr> Individual provider's name

organization

<chr> Organizational provider's name

name_type

<chr> Type of individual the first and last name parameters refer to; one of either AO for Authorized Officials or Provider for Individual Providers.

taxonomy_desc

<chr> Provider's taxonomy description, e.g. Pharmacist, Pediatrics

city

<chr> City name. For military addresses, search for either APO or FPO.

state

<chr> 2-character state abbreviation. If it is the only input, one other parameter besides entype and country is required.

zip

<chr> WC 5- to 9-digit zip code, without a hyphen.

country

<chr> 2-character country abbreviation. Can be the only input, as long as it is not US.

Value

<tibble> of search results

<tibble> of search results

Examples

if (FALSE) { # interactive()
npi_nlm("john bethesda")
npi_nlm("Wiregrass Georgia")
npi_nlm("Dentist Valdosta")
npi_nlm("Valdosta")
npi_nlm("Atlanta")
}
if (FALSE) { # interactive()
npi_nppes(npi = npi_ex$k[1:2]) |> str()
npi_nppes(npi = npi_ex$k[1]) |> str()
}