Open Payments Endpoint Group

Open Payments Endpoint Group

Usage

openGroup(alias)

Arguments

alias

<chr> title alias

Value

An S7 <openGroup> object.

Examples

openGroup("profile")
#> <openGroup>
#>  @ group  : chr "Open Payments Profiles"
#>  @ members:List of 6
#>  .. $ PROF_covered    : <openMain>
#>  ..  ..@ title     : chr "Covered Recipient Profile Supplement"
#>  ..  ..@ metadata  : <openMeta>
#>  .. .. .. @ description: chr "The Covered Recipient Profile Supplement file contains information about physicians and non-physician practitio"| __truncated__
#>  .. .. .. @ modified   : Date[1:1], format: "2025-01-30"
#>  .. .. .. @ download   : chr "https://download.cms.gov/openpayments/PHPRFL_P01302025_01212025/OP_CVRD_RCPNT_PRFL_SPLMTL_P01302025_01212025.csv"
#>  ..  ..@ dimensions: <openDim>
#>  .. .. .. @ rows  : int 1530412
#>  .. .. .. @ pages : int 3061
#>  .. .. .. @ fields: chr [1:32] "covered_recipient_profile_type" "covered_recipient_profile_id" "associated_covered_recipient_profile_id_1" "associated_covered_recipient_profile_id_2" ...
#>  ..  ..@ identifier: chr "https://openpaymentsdata.cms.gov/api/1/datastore/query/23160558-6742-54ff-8b9f-cac7d514ff4e/0"
#>  .. $ PROF_physician  : <openMain>
#>  ..  ..@ title     : chr "Physician (Distinct) Profile Information"
#>  ..  ..@ metadata  : <openMeta>
#>  .. .. .. @ description: chr "This dataset is pre-filtered based on the most frequent searches of Open Payments data."
#>  .. .. .. @ modified   : Date[1:1], format: "2025-01-30"
#>  .. .. .. @ download   : chr "https://download.cms.gov/openpayments/SMRY_RPTS_P01302025_01212025/PBLCTN_PHYSN_NON_PHYSN_PRCTNR_PRFL_SRCH_P013"| __truncated__
#>  ..  ..@ dimensions: <openDim>
#>  .. .. .. @ rows  : int 1528054
#>  .. .. .. @ pages : int 3057
#>  .. .. .. @ fields: chr [1:33] "covered_recipient_profile_type" "covered_recipient_profile_id" "covered_recipient_npi" "covered_recipient_profile_first_name" ...
#>  ..  ..@ identifier: chr "https://openpaymentsdata.cms.gov/api/1/datastore/query/6ed6ae76-2999-49da-b0b2-d7df150ac754/0"
#>  .. $ PROF_information: <openMain>
#>  ..  ..@ title     : chr "Profile Information"
#>  ..  ..@ metadata  : <openMeta>
#>  .. .. .. @ description: chr "This dataset provides profile information to display on payment search results."
#>  .. .. .. @ modified   : Date[1:1], format: "2025-01-30"
#>  .. .. .. @ download   : chr "https://download.cms.gov/openpayments/SMRY_RPTS_P01302025_01212025/all-entities-profile-search.csv"
#>  ..  ..@ dimensions: <openDim>
#>  .. .. .. @ rows  : int 1534695
#>  .. .. .. @ pages : int 3070
#>  .. .. .. @ fields: chr [1:20] "entity_id" "entity_provider_type" "entity_npi" "entity_alternate_id" ...
#>  ..  ..@ identifier: chr "https://openpaymentsdata.cms.gov/api/1/datastore/query/f2ed17cc-f045-4e25-9b28-75c72f0c7efe/0"
#>  .. $ PROF_mapping    : <openMain>
#>  ..  ..@ title     : chr "Provider Profile ID Mapping Table"
#>  ..  ..@ metadata  : <openMeta>
#>  .. .. .. @ description: chr "Maps secondary/duplicate provider profile IDs to the primary ID in the Open Payments system. For multiple reaso"| __truncated__
#>  .. .. .. @ modified   : Date[1:1], format: "2025-01-30"
#>  .. .. .. @ download   : chr "https://download.cms.gov/openpayments/SMRY_RPTS_P01302025_01212025/PBLCTN_PRVDR_PRFL_MAPPING_P01302025_01212025.csv"
#>  ..  ..@ dimensions: <openDim>
#>  .. .. .. @ rows  : int 2358
#>  .. .. .. @ pages : int 5
#>  .. .. .. @ fields: chr [1:2] "primary_provider_profile_id" "secondary_provider_profile_id"
#>  ..  ..@ identifier: chr "https://openpaymentsdata.cms.gov/api/1/datastore/query/d19e5be2-5b8c-562c-af22-c243503bcfac/0"
#>  .. $ PROF_entity     : <openMain>
#>  ..  ..@ title     : chr "Reporting Entity Profile Information"
#>  ..  ..@ metadata  : <openMeta>
#>  .. .. .. @ description: chr "This dataset is pre-filtered based on the most frequent searches of Open Payments data."
#>  .. .. .. @ modified   : Date[1:1], format: "2025-01-30"
#>  .. .. .. @ download   : chr "https://download.cms.gov/openpayments/SMRY_RPTS_P01302025_01212025/PBLCTN_RPTG_ORG_PRFL_SRCH_P01302025_01212025.csv"
#>  ..  ..@ dimensions: <openDim>
#>  .. .. .. @ rows  : int 2843
#>  .. .. .. @ pages : int 6
#>  .. .. .. @ fields: chr [1:9] "amgpo_making_payment_id" "amgpo_making_payment_name" "amgpo_making_payment_state" "amgpo_making_payment_country" ...
#>  ..  ..@ identifier: chr "https://openpaymentsdata.cms.gov/api/1/datastore/query/1cf0c6c0-c377-466e-b78f-037b442559f8/0"
#>  .. $ PROF_teaching   : <openMain>
#>  ..  ..@ title     : chr "Teaching Hospital Profile Information"
#>  ..  ..@ metadata  : <openMeta>
#>  .. .. .. @ description: chr "This dataset is pre-filtered based on the most frequent searches of Open Payments data."
#>  .. .. .. @ modified   : Date[1:1], format: "2025-01-30"
#>  .. .. .. @ download   : chr "https://download.cms.gov/openpayments/SMRY_RPTS_P01302025_01212025/PBLCTN_TH_PRFL_SRCH_P01302025_01212025.csv"
#>  ..  ..@ dimensions: <openDim>
#>  .. .. .. @ rows  : int 1440
#>  .. .. .. @ pages : int 3
#>  .. .. .. @ fields: chr [1:12] "teaching_hospital_ccn" "teaching_hospital_name" "teaching_hospital_address1" "teaching_hospital_address2" ...
#>  ..  ..@ identifier: chr "https://openpaymentsdata.cms.gov/api/1/datastore/query/8219fcb3-fb3d-4b51-9f65-be8a048c0608/0"