Open Payments Endpoint

Open Payments Endpoint

Usage

openMain(alias)

Arguments

alias

<chr> endpoint alias

Value

An S7 <openMain> object.

Examples

openMain("profile_covered")
#> ── Covered Recipient Profile Supplement ──
#> 
#>   The Covered Recipient Profile Supplement
#>   file contains information about physicians
#>   and non-physician practitioners who have
#>   been indicated as recipients of payments,
#>   other transfers of value, or ownership and
#>   investment interest in payment records, as
#>   well as physicians and non-physician
#>   practitioners who have been identified as
#>   principal investigators associated with
#>   research payment records published by Open
#>   Payments. This file contains only those
#>   physicians that have at least one
#>   published payment record in this cycle of
#>   the publication as of January 30, 2025.
#>   The criteria used by the Centers for
#>   Medicare & Medicaid Services (CMS) to
#>   determine which payment records are
#>   eligible for publication is available in
#>   the Open Payments Methodology and Data
#>   Dictionary Document. This document can be
#>   found on the Resources page of the Open
#>   Payments website
#>   (https://www.cms.gov/OpenPayments/Resources).
#>   The Methodology and Data Dictionary
#>   Document also includes information on the
#>   data collection and reporting methodology,
#>   data fields included in the files, and any
#>   notes or special considerations that users
#>   should be aware of.
#> 
#> • Rows 1,530,412
#> • Pages 3,061
#> • Fields 32
#> 
#> • Modified ❯ 2025-01-30
openMain("profile_physician")
#> ── Physician (Distinct) Profile Information ──
#> 
#>   This dataset is pre-filtered based on the
#>   most frequent searches of Open Payments
#>   data.
#> 
#> • Rows 1,528,054
#> • Pages 3,057
#> • Fields 33
#> 
#> • Modified ❯ 2025-01-30
openMain("profile_information")
#> ── Profile Information ──
#> 
#>   This dataset provides profile information
#>   to display on payment search results.
#> 
#> • Rows 1,534,695
#> • Pages 3,070
#> • Fields 20
#> 
#> • Modified ❯ 2025-01-30
openMain("profile_mapping")
#> ── Provider Profile ID Mapping Table ──
#> 
#>   Maps secondary/duplicate provider profile
#>   IDs to the primary ID in the Open Payments
#>   system. For multiple reasons, a small
#>   number of providers have more than one ID
#>   in the payments records.
#> 
#> • Rows 2,358
#> • Pages 5
#> • Fields 2
#> 
#> • Modified ❯ 2025-01-30
openMain("profile_entity")
#> ── Reporting Entity Profile Information ──
#> 
#>   This dataset is pre-filtered based on the
#>   most frequent searches of Open Payments
#>   data.
#> 
#> • Rows 2,843
#> • Pages 6
#> • Fields 9
#> 
#> • Modified ❯ 2025-01-30
openMain("profile_teaching")
#> ── Teaching Hospital Profile Information ──
#> 
#>   This dataset is pre-filtered based on the
#>   most frequent searches of Open Payments
#>   data.
#> 
#> • Rows 1,440
#> • Pages 3
#> • Fields 12
#> 
#> • Modified ❯ 2025-01-30
openMain("state_total")
#> ── State Level Payment Total and Averages for all Years ──
#> 
#>   This dataset is pre-filtered based on the
#>   most frequent searches of Open Payments
#>   data.
#> 
#> • Rows 2,765
#> • Pages 6
#> • Fields 28
#> 
#> • Modified ❯ 2025-01-30
openMain("state_group")
#> ── State Payment Totals and Averages Grouped by Nature of Payment for all Years ──
#> 
#>   Groups payment data by state, recipient
#>   type, and nature of payment.
#> 
#> • Rows 16,758
#> • Pages 34
#> • Fields 28
#> 
#> • Modified ❯ 2025-01-30
openMain("national_group")
#> ── National Level Payment Total and Averages by Provider Specialty for all Years ──
#> 
#>   This dataset is pre-filtered based on the
#>   most frequent searches of Open Payments
#>   data.
#> 
#> • Rows 4,188
#> • Pages 9
#> • Fields 21
#> 
#> • Modified ❯ 2025-01-30
openMain("national_total")
#> ── National Level Payment Total and Averages for all Years ──
#> 
#>   This dataset is pre-filtered based on the
#>   most frequent searches of Open Payments
#>   data.
#> 
#> • Rows 24
#> • Pages 1
#> • Fields 18
#> 
#> • Modified ❯ 2025-01-30