Load Open Payments API Dataset

Load Open Payments API Dataset

Usage

open_Dataset(period = NULL, group = NULL, dataset = NULL)

Arguments

period

<chr> dataset year

group

<chr> dataset group keyword

dataset

<chr> dataset title

Value

<Dataset> object

Examples

open_Dataset(period = "All", group = "Summary", dataset = "Covered Recipient Profile Supplement")
#> $year
#> [1] "All"
#> 
#> $theme
#> [1] "Summary"
#> 
#> $title
#> [1] "Covered Recipient Profile Supplement"
#> 
#> $description
#> [1] "The Covered Recipient Profile Supplement file contains information about physicians and non-physician practitioners who have been indicated as recipients of payments, other transfers of value, or ownership and investment interest in payment records, as well as physicians and non-physician practitioners who have been identified as principal investigators associated with research payment records published by Open Payments. This file contains only those physicians that have at least one published payment record in this cycle of the publication as of January 30, 2025. The criteria used by the Centers for Medicare & Medicaid Services (CMS) to determine which payment records are eligible for publication is available in the Open Payments Methodology and Data Dictionary Document. This document can be found on the Resources page of the Open Payments website (https://www.cms.gov/OpenPayments/Resources). The Methodology and Data Dictionary Document also includes information on the data collection and reporting methodology, data fields included in the files, and any notes or special considerations that users should be aware of."
#> 
#> $modified
#> [1] "2025-01-30"
#> 
#> $temporal
#> [1] "2017-01-01/2023-12-31"
#> 
#> $identifier
#> [1] "https://openpaymentsdata.cms.gov/api/1/datastore/query/23160558-6742-54ff-8b9f-cac7d514ff4e/0"
#> 
#> $downloadURL
#> [1] "https://download.cms.gov/openpayments/PHPRFL_P01302025_01212025/OP_CVRD_RCPNT_PRFL_SPLMTL_P01302025_01212025.csv"
#> 
#> $describedBy
#> [1] "https://openpaymentsdata.cms.gov/api/1/metastore/schemas/data-dictionary/items/80796264-d7e9-57cd-8133-2a2a14b1647b"
#>