Load Open Payments Grouped Summaries

Load Open Payments Grouped Summaries

Usage

open_grouped_summary()

Value

<list> of Open Payments API grouped information

Examples

open_grouped_summary()
#> $modified
#> [1] "2025-01-30"
#> 
#> $summary
#> # A tibble: 12 × 3
#>    title                                                  identifier downloadurl
#>    <chr>                                                  <chr>      <chr>      
#>  1 National Level Payment Total and Averages by Provider… 04c3355f-… https://do…
#>  2 National Level Payment Total and Averages for all Yea… aaaddc12-… https://do…
#>  3 Payments Grouped by Covered Recipient and Nature of P… 8d4d1fa0-… https://do…
#>  4 Payments Grouped by Covered Recipient and Reporting E… e49defac-… https://do…
#>  5 Payments Grouped by Physician (Distinct) for all Years 895260f5-… https://do…
#>  6 Payments Grouped by Reporting Entities                 d5762bed-… https://do…
#>  7 Payments Grouped by Reporting Entities and Nature of … 61efcbef-… https://do…
#>  8 Payments Grouped by Reporting Entities, Covered Recip… ea5eea39-… https://do…
#>  9 Payments Grouped by Teaching Hospital                  217ad1a7-… https://do…
#> 10 State Level Payment Total and Averages for all Years   02157882-… https://do…
#> 11 State Payment Totals and Averages Grouped by Nature o… 7c5144d9-… https://do…
#> 12 Summary Dashboard - all Years                          e0d225fc-… https://do…
#>