Open Payments API Endpoint Classes

Open Payments API Endpoint Classes

Usage

open_endpoint(alias)

open_temporal(alias)

Arguments

alias

<chr> endpoint alias

Value

An S7 <open_endpoint> or <open_temporal> object

Examples

open_endpoint("profile_covered")
#> <open_endpoint>
#>  @ identifier: chr "https://openpaymentsdata.cms.gov/api/1/datastore/query/23160558-6742-54ff-8b9f-cac7d514ff4e/0"
#>  @ metadata  :List of 4
#>  .. $ title      : chr "Covered Recipient Profile Supplement"
#>  .. $ description: chr "The Covered Recipient Profile Supplement file contains information about physicians and non-physician practitio"| __truncated__
#>  .. $ modified   : Date[1:1], format: "2025-01-30"
#>  .. $ download   : chr "https://download.cms.gov/openpayments/PHPRFL_P01302025_01212025/OP_CVRD_RCPNT_PRFL_SPLMTL_P01302025_01212025.csv"
#>  @ dimensions: <class_dimensions>
#>  .. @ limit : int 500
#>  .. @ rows  : int 1530412
#>  .. @ pages : int 3061
#>  .. @ fields:List of 32
#>  .. .. $ covered_recipient_profile_type                 : chr(0) 
#>  .. .. $ covered_recipient_profile_id                   : chr(0) 
#>  .. .. $ associated_covered_recipient_profile_id_1      : chr(0) 
#>  .. .. $ associated_covered_recipient_profile_id_2      : chr(0) 
#>  .. .. $ covered_recipient_npi                          : chr(0) 
#>  .. .. $ covered_recipient_profile_first_name           : chr(0) 
#>  .. .. $ covered_recipient_profile_middle_name          : chr(0) 
#>  .. .. $ covered_recipient_profile_last_name            : chr(0) 
#>  .. .. $ covered_recipient_profile_suffix               : chr(0) 
#>  .. .. $ covered_recipient_profile_alternate_first_name : chr(0) 
#>  .. .. $ covered_recipient_profile_alternate_middle_name: chr(0) 
#>  .. .. $ covered_recipient_profile_alternate_last_name  : chr(0) 
#>  .. .. $ covered_recipient_profile_alternate_suffix     : chr(0) 
#>  .. .. $ covered_recipient_profile_address_line_1       : chr(0) 
#>  .. .. $ covered_recipient_profile_address_line_2       : chr(0) 
#>  .. .. $ covered_recipient_profile_city                 : chr(0) 
#>  .. .. $ covered_recipient_profile_state                : chr(0) 
#>  .. .. $ covered_recipient_profile_zipcode              : chr(0) 
#>  .. .. $ covered_recipient_profile_country_name         : chr(0) 
#>  .. .. $ covered_recipient_profile_province_name        : chr(0) 
#>  .. .. $ covered_recipient_profile_primary_specialty    : chr(0) 
#>  .. .. $ covered_recipient_profile_ops_taxonomy_1       : chr(0) 
#>  .. .. $ covered_recipient_profile_ops_taxonomy_2       : chr(0) 
#>  .. .. $ covered_recipient_profile_ops_taxonomy_3       : chr(0) 
#>  .. .. $ covered_recipient_profile_ops_taxonomy_4       : chr(0) 
#>  .. .. $ covered_recipient_profile_ops_taxonomy_5       : chr(0) 
#>  .. .. $ covered_recipient_profile_ops_taxonomy_6       : chr(0) 
#>  .. .. $ covered_recipient_profile_license_state_code_1 : chr(0) 
#>  .. .. $ covered_recipient_profile_license_state_code_2 : chr(0) 
#>  .. .. $ covered_recipient_profile_license_state_code_3 : chr(0) 
#>  .. .. $ covered_recipient_profile_license_state_code_4 : chr(0) 
#>  .. .. $ covered_recipient_profile_license_state_code_5 : chr(0) 
open_temporal("grouped_state_nature")
#> <open_temporal>
#>  @ metadata  :List of 3
#>  .. $ title      : chr "State Payment Totals Grouped by Nature of Payment for all Years"
#>  .. $ description: chr "Groups payment data by state, recipient type, and nature of payment."
#>  .. $ modified   : Date[1:1], format: "2025-01-30"
#>  @ dimensions: <class_dimensions>
#>  .. @ limit : int 500
#>  .. @ rows  : int 16758
#>  .. @ pages : int 34
#>  .. @ fields:List of 28
#>  .. .. $ country_code                                          : chr(0) 
#>  .. .. $ country_name                                          : chr(0) 
#>  .. .. $ state_code                                            : chr(0) 
#>  .. .. $ state_name                                            : chr(0) 
#>  .. .. $ nature_of_payment                                     : chr(0) 
#>  .. .. $ recipient_type                                        : chr(0) 
#>  .. .. $ program_year                                          : chr(0) 
#>  .. .. $ total_number_of_physicians                            : chr(0) 
#>  .. .. $ total_number_of_non_physician_practitioners           : chr(0) 
#>  .. .. $ total_number_of_teaching_hospitals                    : chr(0) 
#>  .. .. $ total_payment_amount_physician                        : chr(0) 
#>  .. .. $ total_payment_amount_non_physician_practitioner       : chr(0) 
#>  .. .. $ total_payment_amount_teaching_hospital                : chr(0) 
#>  .. .. $ mean_total_payment_amount_physician                   : chr(0) 
#>  .. .. $ mean_total_payment_amount_non_physician_practitioner  : chr(0) 
#>  .. .. $ mean_total_payment_amount_teaching_hospital           : chr(0) 
#>  .. .. $ median_total_payment_amount_physician                 : chr(0) 
#>  .. .. $ median_total_payment_amount_non_physician_practitioner: chr(0) 
#>  .. .. $ median_total_payment_amount_teaching_hospital         : chr(0) 
#>  .. .. $ total_payment_count_physician                         : chr(0) 
#>  .. .. $ total_payment_count_non_physician_practitioner        : chr(0) 
#>  .. .. $ total_payment_count_teaching_hospital                 : chr(0) 
#>  .. .. $ mean_total_payment_count_physician                    : chr(0) 
#>  .. .. $ mean_total_payment_count_non_physician_practitioner   : chr(0) 
#>  .. .. $ mean_total_payment_count_teaching_hospital            : chr(0) 
#>  .. .. $ median_total_payment_count_physician                  : chr(0) 
#>  .. .. $ median_total_payment_count_non_physician_practitioner : chr(0) 
#>  .. .. $ median_total_payment_count_teaching_hospital          : chr(0) 
#>  @ endpoints : tibble [7 × 4] (S3: tbl_df/tbl/data.frame)
#>  $ year      : int [1:7] 2023 2022 2021 2020 2019 2018 2017
#>  $ identifier: chr [1:7] "https://openpaymentsdata.cms.gov/api/1/datastore/query/f83aed3a-7504-4fca-aabb-ac0338c79688/0" "https://openpaymentsdata.cms.gov/api/1/datastore/query/c64c40af-23a5-4419-8b64-692684483664/0" "https://openpaymentsdata.cms.gov/api/1/datastore/query/7fad5b56-2cfe-52ab-ac4d-3200780287d4/0" "https://openpaymentsdata.cms.gov/api/1/datastore/query/7f3055b6-73d8-5aff-be15-55873552e2cd/0" ...
#>  $ contact   : chr [1:7] "Open Payments (mailto:openpayments@cms.hhs.gov)" "Open Payments (mailto:openpayments@cms.hhs.gov)" "Open Payments (mailto:openpayments@cms.hhs.gov)" "Open Payments (mailto:openpayments@cms.hhs.gov)" ...
#>  $ download  : chr [1:7] "https://download.cms.gov/openpayments/SMRY_RPTS_P01302025_01212025/PBLCTN_NTR_OF_PYMT_BY_STATE_SMRY_PGYR2023_P0"| __truncated__ "https://download.cms.gov/openpayments/SMRY_RPTS_P01302025_01212025/PBLCTN_NTR_OF_PYMT_BY_STATE_SMRY_PGYR2022_P0"| __truncated__ "https://download.cms.gov/openpayments/SMRY_RPTS_P01302025_01212025/PBLCTN_NTR_OF_PYMT_BY_STATE_SMRY_PGYR2021_P0"| __truncated__ "https://download.cms.gov/openpayments/SMRY_RPTS_P01302025_01212025/PBLCTN_NTR_OF_PYMT_BY_STATE_SMRY_PGYR2020_P0"| __truncated__ ...