Provider Endpoint

Provider Endpoint

Provider Endpoint Group

Usage

proMain(alias)

proGroup(alias)

Arguments

alias

<chr> title alias

Value

An S7 <proMain> object.

An S7 <proGroup> object.

Examples

proMain("affiliations")
#> <proMain>
#>  @ title     : chr "Facility Affiliation Data"
#>  @ metadata  : <proMeta>
#>  .. @ description: chr "This is the facility affiliations data publicly reported in the Provider Data Catalog."
#>  .. @ issued     : Date[1:1], format: "2023-08-17"
#>  .. @ modified   : Date[1:1], format: "2025-04-04"
#>  .. @ released   : Date[1:1], format: "2025-04-17"
#>  .. @ site       : chr "https://data.cms.gov/provider-data/dataset/27ea-46a8"
#>  .. @ dictionary : chr "https://data.cms.gov/provider-data/dataset/27ea-46a8#data-dictionary"
#>  .. @ download   : chr "https://data.cms.gov/provider-data/sites/default/files/resources/b7c4080ae144663e43353a9c35cd3f53_1743782712/Fa"| __truncated__
#>  @ dimensions: <proDim>
#>  .. @ rows  : int 1577099
#>  .. @ pages : int 789
#>  .. @ fields: chr [1:9] "npi" "ind_pac_id" "provider_last_name" ...
#>  @ identifier: chr "https://data.cms.gov/provider-data/api/1/datastore/query/27ea-46a8/0"
proMain("clinicians")
#> <proMain>
#>  @ title     : chr "National Downloadable File"
#>  @ metadata  : <proMeta>
#>  .. @ description: chr "The Doctors and Clinicians national downloadable file is organized such that each line is unique at the clinici"| __truncated__
#>  .. @ issued     : Date[1:1], format: "2023-08-17"
#>  .. @ modified   : Date[1:1], format: "2025-04-04"
#>  .. @ released   : Date[1:1], format: "2025-04-17"
#>  .. @ site       : chr "https://data.cms.gov/provider-data/dataset/mj5m-pzi6"
#>  .. @ dictionary : chr "https://data.cms.gov/provider-data/dataset/mj5m-pzi6#data-dictionary"
#>  .. @ download   : chr "https://data.cms.gov/provider-data/sites/default/files/resources/52c3f098d7e56028a298fd297cb0b38d_1743782711/DA"| __truncated__
#>  @ dimensions: <proDim>
#>  .. @ rows  : int 2642756
#>  .. @ pages : int 1322
#>  .. @ fields: chr [1:31] "npi" "ind_pac_id" "ind_enrl_id" "provider_last_name" ...
#>  @ identifier: chr "https://data.cms.gov/provider-data/api/1/datastore/query/mj5m-pzi6/0"
proMain("utilization")
#> <proMain>
#>  @ title     : chr "Utilization Data"
#>  @ metadata  : <proMeta>
#>  .. @ description: chr "The Doctors and Clinicians utilization data file reports volume information for procedures of interest on clini"| __truncated__
#>  .. @ issued     : Date[1:1], format: "2024-01-04"
#>  .. @ modified   : Date[1:1], format: "2025-03-24"
#>  .. @ released   : Date[1:1], format: "2025-04-03"
#>  .. @ site       : chr "https://data.cms.gov/provider-data/dataset/n0yb-util"
#>  .. @ dictionary : chr "https://data.cms.gov/provider-data/dataset/n0yb-util#data-dictionary"
#>  .. @ download   : chr "https://data.cms.gov/provider-data/sites/default/files/resources/72eedb23fcbe1ef3666f89db01b85e1d_1742861112/Utilization.csv"
#>  @ dimensions: <proDim>
#>  .. @ rows  : int 301894
#>  .. @ pages : int 151
#>  .. @ fields: chr [1:10] "npi" "ind_pac_id" "provider_last_name" ...
#>  @ identifier: chr "https://data.cms.gov/provider-data/api/1/datastore/query/n0yb-util/0"
proGroup("PDC")
#> <proGroup>
#>  @ group  : chr "Provider Data Catalog"
#>  @ members:List of 3
#>  .. $ affiliations: <proMain>
#>  ..  ..@ title     : chr "Facility Affiliation Data"
#>  ..  ..@ metadata  : <proMeta>
#>  .. .. .. @ description: chr "This is the facility affiliations data publicly reported in the Provider Data Catalog."
#>  .. .. .. @ issued     : Date[1:1], format: "2023-08-17"
#>  .. .. .. @ modified   : Date[1:1], format: "2025-04-04"
#>  .. .. .. @ released   : Date[1:1], format: "2025-04-17"
#>  .. .. .. @ site       : chr "https://data.cms.gov/provider-data/dataset/27ea-46a8"
#>  .. .. .. @ dictionary : chr "https://data.cms.gov/provider-data/dataset/27ea-46a8#data-dictionary"
#>  .. .. .. @ download   : chr "https://data.cms.gov/provider-data/sites/default/files/resources/b7c4080ae144663e43353a9c35cd3f53_1743782712/Fa"| __truncated__
#>  ..  ..@ dimensions: <proDim>
#>  .. .. .. @ rows  : int 1577099
#>  .. .. .. @ pages : int 789
#>  .. .. .. @ fields: chr [1:9] "npi" "ind_pac_id" "provider_last_name" "provider_first_name" ...
#>  ..  ..@ identifier: chr "https://data.cms.gov/provider-data/api/1/datastore/query/27ea-46a8/0"
#>  .. $ clinicians  : <proMain>
#>  ..  ..@ title     : chr "National Downloadable File"
#>  ..  ..@ metadata  : <proMeta>
#>  .. .. .. @ description: chr "The Doctors and Clinicians national downloadable file is organized such that each line is unique at the clinici"| __truncated__
#>  .. .. .. @ issued     : Date[1:1], format: "2023-08-17"
#>  .. .. .. @ modified   : Date[1:1], format: "2025-04-04"
#>  .. .. .. @ released   : Date[1:1], format: "2025-04-17"
#>  .. .. .. @ site       : chr "https://data.cms.gov/provider-data/dataset/mj5m-pzi6"
#>  .. .. .. @ dictionary : chr "https://data.cms.gov/provider-data/dataset/mj5m-pzi6#data-dictionary"
#>  .. .. .. @ download   : chr "https://data.cms.gov/provider-data/sites/default/files/resources/52c3f098d7e56028a298fd297cb0b38d_1743782711/DA"| __truncated__
#>  ..  ..@ dimensions: <proDim>
#>  .. .. .. @ rows  : int 2642756
#>  .. .. .. @ pages : int 1322
#>  .. .. .. @ fields: chr [1:31] "npi" "ind_pac_id" "ind_enrl_id" "provider_last_name" ...
#>  ..  ..@ identifier: chr "https://data.cms.gov/provider-data/api/1/datastore/query/mj5m-pzi6/0"
#>  .. $ utilization : <proMain>
#>  ..  ..@ title     : chr "Utilization Data"
#>  ..  ..@ metadata  : <proMeta>
#>  .. .. .. @ description: chr "The Doctors and Clinicians utilization data file reports volume information for procedures of interest on clini"| __truncated__
#>  .. .. .. @ issued     : Date[1:1], format: "2024-01-04"
#>  .. .. .. @ modified   : Date[1:1], format: "2025-03-24"
#>  .. .. .. @ released   : Date[1:1], format: "2025-04-03"
#>  .. .. .. @ site       : chr "https://data.cms.gov/provider-data/dataset/n0yb-util"
#>  .. .. .. @ dictionary : chr "https://data.cms.gov/provider-data/dataset/n0yb-util#data-dictionary"
#>  .. .. .. @ download   : chr "https://data.cms.gov/provider-data/sites/default/files/resources/72eedb23fcbe1ef3666f89db01b85e1d_1742861112/Utilization.csv"
#>  ..  ..@ dimensions: <proDim>
#>  .. .. .. @ rows  : int 301894
#>  .. .. .. @ pages : int 151
#>  .. .. .. @ fields: chr [1:10] "npi" "ind_pac_id" "provider_last_name" "provider_first_name" ...
#>  ..  ..@ identifier: chr "https://data.cms.gov/provider-data/api/1/datastore/query/n0yb-util/0"
proGroup("MIPS")
#> <proGroup>
#>  @ group  : chr "PY 2022 MIPS Public Reporting"
#>  @ members:List of 5
#>  .. $ MIPS_performance       : <proMain>
#>  ..  ..@ title     : chr "PY 2022 Clinician Public Reporting: Overall MIPS Performance"
#>  ..  ..@ metadata  : <proMeta>
#>  .. .. .. @ description: chr "This file contains Merit-Based Incentive Payment System (MIPS) Final Scores and performance category scores for"| __truncated__
#>  .. .. .. @ issued     : Date[1:1], format: "2023-08-17"
#>  .. .. .. @ modified   : Date[1:1], format: "2024-07-25"
#>  .. .. .. @ released   : Date[1:1], format: "2024-08-01"
#>  .. .. .. @ site       : chr "https://data.cms.gov/provider-data/dataset/a174-a962"
#>  .. .. .. @ dictionary : chr "https://data.cms.gov/provider-data/dataset/a174-a962#data-dictionary"
#>  .. .. .. @ download   : chr "https://data.cms.gov/provider-data/sites/default/files/resources/6b9e57db797c95853b034b329b1212b2_1721952317/ec_score_file.csv"
#>  ..  ..@ dimensions: <proDim>
#>  .. .. .. @ rows  : int 624209
#>  .. .. .. @ pages : int 313
#>  .. .. .. @ fields: chr [1:13] "npi" "org_pac_id" "provider_last_name" "provider_first_name" ...
#>  ..  ..@ identifier: chr "https://data.cms.gov/provider-data/api/1/datastore/query/a174-a962/0"
#>  .. $ MIPS_patient           : <proMain>
#>  ..  ..@ title     : chr "PY 2022 Group Public Reporting: Patient Experience"
#>  ..  ..@ metadata  : <proMeta>
#>  .. .. .. @ description: chr "This file contains information for the Consumer Assessment of Healthcare Providers and Systems (CAHPS) for MIPS"| __truncated__
#>  .. .. .. @ issued     : Date[1:1], format: "2023-08-17"
#>  .. .. .. @ modified   : Date[1:1], format: "2024-05-16"
#>  .. .. .. @ released   : Date[1:1], format: "2024-05-30"
#>  .. .. .. @ site       : chr "https://data.cms.gov/provider-data/dataset/8c70-d353"
#>  .. .. .. @ dictionary : chr "https://data.cms.gov/provider-data/dataset/8c70-d353#data-dictionary"
#>  .. .. .. @ download   : chr "https://data.cms.gov/provider-data/sites/default/files/resources/6083bb351293e72eda42881d6a93d803_1715976323/gr"| __truncated__
#>  ..  ..@ dimensions: <proDim>
#>  .. .. .. @ rows  : int 432
#>  .. .. .. @ pages : int 1
#>  .. .. .. @ fields: chr [1:7] "facility_name" "org_pac_id" "measure_cd" "measure_title" ...
#>  ..  ..@ identifier: chr "https://data.cms.gov/provider-data/api/1/datastore/query/8c70-d353/0"
#>  .. $ MIPS_measures_clinician: <proMain>
#>  ..  ..@ title     : chr "PY 2022 Clinician Public Reporting: MIPS Measures and Attestations"
#>  ..  ..@ metadata  : <proMeta>
#>  .. .. .. @ description: chr "This file contains performance information for Merit-Based Incentive Payment System (MIPS) Quality, Promoting I"| __truncated__
#>  .. .. .. @ issued     : Date[1:1], format: "2020-11-30"
#>  .. .. .. @ modified   : Date[1:1], format: "2024-05-16"
#>  .. .. .. @ released   : Date[1:1], format: "2024-05-30"
#>  .. .. .. @ site       : chr "https://data.cms.gov/provider-data/dataset/7d6a-e7a6"
#>  .. .. .. @ dictionary : chr "https://data.cms.gov/provider-data/dataset/7d6a-e7a6#data-dictionary"
#>  .. .. .. @ download   : chr "https://data.cms.gov/provider-data/sites/default/files/resources/b9c4d902c52ae898d3f422fffbdc59fe_1715976323/ec"| __truncated__
#>  ..  ..@ dimensions: <proDim>
#>  .. .. .. @ rows  : int 523282
#>  .. .. .. @ pages : int 262
#>  .. .. .. @ fields: chr [1:18] "npi" "ind_pac_id" "provider_last_name" "provider_first_name" ...
#>  ..  ..@ identifier: chr "https://data.cms.gov/provider-data/api/1/datastore/query/7d6a-e7a6/0"
#>  .. $ MIPS_measures_group    : <proMain>
#>  ..  ..@ title     : chr "PY 2022 Group Public Reporting: MIPS Measures and Attestations"
#>  ..  ..@ metadata  : <proMeta>
#>  .. .. .. @ description: chr "This file contains performance information for Merit-Based Incentive Payment System (MIPS) Quality, Promoting I"| __truncated__
#>  .. .. .. @ issued     : Date[1:1], format: "2023-08-17"
#>  .. .. .. @ modified   : Date[1:1], format: "2024-05-16"
#>  .. .. .. @ released   : Date[1:1], format: "2024-05-30"
#>  .. .. .. @ site       : chr "https://data.cms.gov/provider-data/dataset/0ba7-2cb0"
#>  .. .. .. @ dictionary : chr "https://data.cms.gov/provider-data/dataset/0ba7-2cb0#data-dictionary"
#>  .. .. .. @ download   : chr "https://data.cms.gov/provider-data/sites/default/files/resources/410d73868de5a4ea585a61ca1365a256_1715976322/gr"| __truncated__
#>  ..  ..@ dimensions: <proDim>
#>  .. .. .. @ rows  : int 193488
#>  .. .. .. @ pages : int 97
#>  .. .. .. @ fields: chr [1:16] "facility_name" "org_pac_id" "aco_id_1" "aco_nm_1" ...
#>  ..  ..@ identifier: chr "https://data.cms.gov/provider-data/api/1/datastore/query/0ba7-2cb0/0"
#>  .. $ MIPS_measures_virtual  : <proMain>
#>  ..  ..@ title     : chr "PY 2022 Virtual Group Public Reporting: MIPS Measures and Attestations"
#>  ..  ..@ metadata  : <proMeta>
#>  .. .. .. @ description: chr "This file contains performance information for Merit-Based Incentive Payment System (MIPS) Quality, Promoting I"| __truncated__
#>  .. .. .. @ issued     : Date[1:1], format: "2021-06-04"
#>  .. .. .. @ modified   : Date[1:1], format: "2024-05-16"
#>  .. .. .. @ released   : Date[1:1], format: "2024-05-30"
#>  .. .. .. @ site       : chr "https://data.cms.gov/provider-data/dataset/46dc-a66c"
#>  .. .. .. @ dictionary : chr "https://data.cms.gov/provider-data/dataset/46dc-a66c#data-dictionary"
#>  .. .. .. @ download   : chr "https://data.cms.gov/provider-data/sites/default/files/resources/635d18fe1428158bf5b033cd27da6cc3_1715976333/vg"| __truncated__
#>  ..  ..@ dimensions: <proDim>
#>  .. .. .. @ rows  : int 57
#>  .. .. .. @ pages : int 1
#>  .. .. .. @ fields: chr [1:8] "virtual_group_id" "measure_cd" "measure_title" "invs_msr" ...
#>  ..  ..@ identifier: chr "https://data.cms.gov/provider-data/api/1/datastore/query/46dc-a66c/0"