Provider Endpoint

Provider Endpoint

Usage

proMain(alias)

Arguments

alias

<chr> endpoint alias

Value

An S7 <proMain> object.

Examples

proMain("affiliations") |> str()
#> <proMain>
#>  @ title      : chr "Facility Affiliation Data"
#>  @ description: chr "This is the facility affiliations data publicly reported in the Provider Data Catalog."
#>  @ modified   : Date[1:1], format: "2025-03-24"
#>  @ rows       : int 1566867
#>  @ pages      : int 784
#>  @ fields     : chr [1:9] "npi" "ind_pac_id" "provider_last_name" ...
#>  @ download   : chr "https://data.cms.gov/provider-data/sites/default/files/resources/b7c4080ae144663e43353a9c35cd3f53_1742861113/Fa"| __truncated__
#>  @ issued     : Date[1:1], format: "2023-08-17"
#>  @ released   : Date[1:1], format: "2025-04-03"
#>  @ uuid       : chr "27ea-46a8"
#>  @ site       : chr "https://data.cms.gov/provider-data/dataset/27ea-46a8"
#>  @ identifier : chr "https://data.cms.gov/provider-data/api/1/datastore/query/27ea-46a8/0"
proMain("clinicians") |> str()
#> <proMain>
#>  @ title      : chr "National Downloadable File"
#>  @ description: chr "The Doctors and Clinicians national downloadable file is organized such that each line is unique at the clinici"| __truncated__
#>  @ modified   : Date[1:1], format: "2025-03-24"
#>  @ rows       : int 2631270
#>  @ pages      : int 1316
#>  @ fields     : chr [1:31] "npi" "ind_pac_id" "ind_enrl_id" "provider_last_name" ...
#>  @ download   : chr "https://data.cms.gov/provider-data/sites/default/files/resources/52c3f098d7e56028a298fd297cb0b38d_1742861111/DA"| __truncated__
#>  @ issued     : Date[1:1], format: "2023-08-17"
#>  @ released   : Date[1:1], format: "2025-04-03"
#>  @ uuid       : chr "mj5m-pzi6"
#>  @ site       : chr "https://data.cms.gov/provider-data/dataset/mj5m-pzi6"
#>  @ identifier : chr "https://data.cms.gov/provider-data/api/1/datastore/query/mj5m-pzi6/0"
proMain("utilization") |> str()
#> <proMain>
#>  @ title      : chr "Utilization Data"
#>  @ description: chr "The Doctors and Clinicians utilization data file reports volume information for procedures of interest on clini"| __truncated__
#>  @ modified   : Date[1:1], format: "2025-03-24"
#>  @ rows       : int 301894
#>  @ pages      : int 151
#>  @ fields     : chr [1:10] "npi" "ind_pac_id" "provider_last_name" ...
#>  @ download   : chr "https://data.cms.gov/provider-data/sites/default/files/resources/72eedb23fcbe1ef3666f89db01b85e1d_1742861112/Utilization.csv"
#>  @ issued     : Date[1:1], format: "2024-01-04"
#>  @ released   : Date[1:1], format: "2025-04-03"
#>  @ uuid       : chr "n0yb-util"
#>  @ site       : chr "https://data.cms.gov/provider-data/dataset/n0yb-util"
#>  @ identifier : chr "https://data.cms.gov/provider-data/api/1/datastore/query/n0yb-util/0"