Provider API Endpoints

Facility affiliations data publicly reported in the Provider Data Catalog.

The Doctors and Clinicians national downloadable file is organized such that each line is unique at the clinician/enrollment record/group/address level. Clinicians with multiple Medicare enrollment records and/or single enrollments linking to multiple practice locations are listed on multiple lines.

The Doctors and Clinicians utilization data file reports volume information for procedures of interest on clinician profile pages and in the provider data catalog (PDC) to inform patients and caregivers about clinicians' experience.

Usage

prov_endpoint(alias, call = caller_env())

prov_group(alias, call = caller_env(), ...)

affiliations(
  npi = NULL,
  pac = NULL,
  last_name = NULL,
  first_name = NULL,
  middle_name = NULL,
  suffix = NULL,
  facility_type = NULL,
  ccn_facility = NULL,
  ccn_primary = NULL
)

clinicians(
  npi = NULL,
  pac = NULL,
  enid = NULL,
  last_name = NULL,
  first_name = NULL,
  middle_name = NULL,
  suffix = NULL,
  gender = NULL,
  credential = NULL,
  med_school = NULL,
  grad_year = NULL,
  spec_prim = NULL,
  spec_sec = NULL,
  telehealth = NULL,
  pac_org = NULL,
  org_n_members = NULL,
  facility_name = NULL,
  add1 = NULL,
  add2 = NULL,
  city = NULL,
  state = NULL,
  zip = NULL,
  phone = NULL,
  ind_assign = NULL,
  grp_assign = NULL,
  add_id = NULL
)

utilization(
  npi = NULL,
  pac = NULL,
  last_name = NULL,
  first_name = NULL,
  middle_name = NULL,
  suffix = NULL,
  procedure = NULL,
  count = NULL,
  percentile = NULL
)

Arguments

alias

<chr> endpoint or group alias

call

<env> environment to use for error reporting

...

Additional arguments passed to the group constructor

npi

<chr> 10-digit Individual NPI

pac

<chr> 10-digit PECOS Associate Control (PAC) ID

first_name, middle_name, last_name, suffix

<chr> Individual provider's name

facility_type

<chr> Type of facility, one of the following:

"Hospital"

"Long-term care hospital"

"Nursing home"

"Inpatient rehabilitation facility"

"Home health agency"

"Skilled nursing facility"

"Hospice"

"Dialysis facility"

ccn_facility

<chr> 6-digit CMS Certification Number (CCN) of a facility or unit within hospital where an individual provider provides service.

ccn_primary

<int> 6-digit CMS Certification Number (CCN) of a sub-unit's primary hospital, should the provider provide services in said unit.

enid

<chr> 15-digit Medicare Enrollment ID

gender

Individual provider's gender

credential

Individual provider's credentials

med_school

Medical school attended

grad_year

Year of graduation from medical school

spec_prim

Primary specialty

spec_sec

Secondary specialty

telehealth

Telehealth services

pac_org

<chr> 10-digit Organization PECOS Associate Control (PAC) ID

org_n_members

Number of members in the organization

facility_name

Facility name

add1

Address line 1

add2

Address line 2

city

City

state

State

zip

Zip code

phone

Phone number

ind_assign

Individual assignment

grp_assign

Group assignment

add_id

Address ID

procedure

<chr> Procedure category

count

<int> Number of procedures performed

percentile

<int> Percentile of procedures performed

Value

An S7 <class_endpoint> or <class_group> object

<tibble> of search results

<tibble> of search results

<tibble> of search results

Periodicity

Monthly (R/P1M)

Monthly (R/P1M)

Examples

if (FALSE) { # rlang::is_interactive()
prov_endpoint("asc_facility")
prov_group("pro_dialysis")
}
if (FALSE) { # interactive()
affiliations(facility_type = "Home health agency")
affiliations(last_name = "CURRY")
affiliations(ccn_primary = "670055")
affiliations(ccn_facility = "370781")
affiliations(ccn_facility = "331302")
affiliations(ccn_facility = "33Z302")
affiliations(npi = "1043245657")
}
if (FALSE) { # interactive()
clinicians(last_name = "CURRY")
clinicians(npi = "1043245657")
clinicians(state = "GA")
}
if (FALSE) { # interactive()
utilization(last_name = "CURRY")
utilization(npi = "1043245657")
utilization(npi = "1003001785")
}