Public Provider Enrollment Experiment

Public Provider Enrollment Experiment

Usage

provider_enrollment()

Value

<tibble> of search results

Examples

provider_enrollment()
#> $title
#> [1] "Medicare Fee-For-Service  Public Provider Enrollment"
#> 
#> $modified
#> [1] "2024-10-16"
#> 
#> $temporal
#> [1] "2023-01-01"
#> 
#> $accrualPeriodicity
#> [1] "[R/P3M] Quarterly"
#> 
#> $identifier
#> [1] "https://data.cms.gov/data-api/v1/dataset/2457ea29-fc82-48b0-86ec-3b0755de7515/data-viewer"
#> 
#> $describedBy
#> [1] "https://data.cms.gov/resources/medicare-fee-for-service-public-provider-enrollment-data-dictionary"
#> 
#> $description
#> [1] "The Medicare Fee-For-Service Public Provider Enrollment dataset includes information on providers who are actively approved to bill Medicare or have completed the 855O at the time the data was pulled from the Provider Enrollment, Chain, and Ownership System (PECOS). The release of this provider enrollment data is not related to other provider information releases such as Physician Compare or Data Transparency.\n\n \n\nNote: This full dataset contains more records than most spreadsheet programs can handle, which will result in an incomplete load of data. Use of a database or statistical software is required."
#> 
#> $landingPage
#> [1] "https://data.cms.gov/provider-characteristics/medicare-provider-supplier-enrollment/medicare-fee-for-service-public-provider-enrollment"
#> 
#> $title
#> [1] "Medicare Fee-For-Service Public Provider Enrollment : 2024-09-01"
#> [2] "Medicare Fee-For-Service Public Provider Enrollment : 2024-04-01"
#> [3] "Medicare Fee-For-Service Public Provider Enrollment : 2024-01-01"
#> [4] "Medicare Fee-For-Service Public Provider Enrollment : 2023-12-31"
#> [5] "Medicare Fee-For-Service Public Provider Enrollment : 2023-09-01"
#> [6] "Medicare Fee-For-Service Public Provider Enrollment : 2023-06-01"
#> [7] "Medicare Fee-For-Service Public Provider Enrollment : 2023-03-31"
#> 
#> $modified
#> [1] "2024-10-16" "2024-07-15" "2024-04-15" "2024-01-16" "2023-10-16"
#> [6] "2023-09-22" "2023-04-19"
#> 
#> $temporal
#> [1] NA NA NA NA NA NA NA
#> 
#> $accessURL
#> [1] "https://data.cms.gov/data-api/v1/dataset/57312fe7-f2cc-4cda-b4a0-78d9edb42f67/data"
#> [2] "https://data.cms.gov/data-api/v1/dataset/8d9854d3-6b9f-49df-b315-5e4e50df7aa7/data"
#> [3] "https://data.cms.gov/data-api/v1/dataset/a5cac43c-010d-4190-9395-d8841b1c5cc6/data"
#> [4] "https://data.cms.gov/data-api/v1/dataset/3d7326b6-f5cc-49ba-967b-20c484af40ce/data"
#> [5] "https://data.cms.gov/data-api/v1/dataset/cf2b32cc-c0e8-4d3c-bfdf-7752be357f6b/data"
#> [6] "https://data.cms.gov/data-api/v1/dataset/20fd773a-7f15-48ed-8164-f543706269a4/data"
#> [7] "https://data.cms.gov/data-api/v1/dataset/f783ac85-fd89-4db9-a7e9-288cda155258/data"
#> 
#> $resourcesAPI
#> [1] "https://data.cms.gov/data-api/v1/dataset-resources/57312fe7-f2cc-4cda-b4a0-78d9edb42f67"
#> [2] "https://data.cms.gov/data-api/v1/dataset-resources/8d9854d3-6b9f-49df-b315-5e4e50df7aa7"
#> [3] "https://data.cms.gov/data-api/v1/dataset-resources/a5cac43c-010d-4190-9395-d8841b1c5cc6"
#> [4] "https://data.cms.gov/data-api/v1/dataset-resources/3d7326b6-f5cc-49ba-967b-20c484af40ce"
#> [5] "https://data.cms.gov/data-api/v1/dataset-resources/cf2b32cc-c0e8-4d3c-bfdf-7752be357f6b"
#> [6] "https://data.cms.gov/data-api/v1/dataset-resources/20fd773a-7f15-48ed-8164-f543706269a4"
#> [7] "https://data.cms.gov/data-api/v1/dataset-resources/f783ac85-fd89-4db9-a7e9-288cda155258"
#> 
#> $downloadURL
#> [1] "https://data.cms.gov/sites/default/files/2024-10/8e1bbada-e9a4-4a4c-9017-539c5cb1fff8/PPEF_Enrollment_Extract_2024.10.07.csv"
#> [2] "https://data.cms.gov/sites/default/files/2024-07/0c28f2f8-4153-421c-a9b7-72a5be89d9e6/PPEF_Enrollment_Extract_2024.07.01.csv"
#> [3] "https://data.cms.gov/sites/default/files/2024-04/07dedcbc-ff73-4097-a83d-7be34d526aa9/PPEF_Enrollment_Extract_2024.04.01.csv"
#> [4] "https://data.cms.gov/sites/default/files/2024-01/22961024-cac8-43fc-86d1-163e19dca5cf/PPEF_Enrollment_Extract_2024.01.05.csv"
#> [5] "https://data.cms.gov/sites/default/files/2023-10/d2c13c22-d68b-4d46-8436-0e2a6abac4c9/PPEF_Enrollment_Extract_2023.10.02.csv"
#> [6] "https://data.cms.gov/sites/default/files/2023-07/dc5e6313-b126-4f77-9846-2a63d32787d7/PPEF_Enrollment_Extract_2023.07.06.csv"
#> [7] "https://data.cms.gov/sites/default/files/2023-04/0239bace-c79e-461c-9fed-04e165280f25/PPEF_Enrollment_Extract_2023.03.31.csv"
#>