Load Public API Dataset

Load Public API Dataset

Usage

public_Dataset(dataset, fname = TRUE)

Arguments

dataset

<chr> dataset title

fname

<lgl> Is dataset a function name?; default is TRUE

Value

<Dataset> object

Examples

public_Dataset("enrollees")
#> 
#> ── Dataset Medicare Fee-For-Service Public Provider Enrollment 
#> → Rows: 2,945,846 | Fields: 11 | Resources: 7 files
#> 
#> • Periodicity: Quarterly [R/P3M]
#> • Last Modified: 2025-01-15
#> • Time Period: 2023-01-01 - 2024-12-31
#> • Theme: Medicare
#> • Keywords: Medicare, Original Medicare, Provider Enrollment
#> 
#> The Medicare Fee-For-Service Public Provider Enrollment dataset includes
#> information on providers who are actively approved to bill Medicare or have
#> completed the 855O at the time the data was pulled from the Provider
#> Enrollment, Chain, and Ownership System (PECOS). The release of this provider
#> enrollment data is not related to other provider information releases such as
#> Physician Compare or Data ...[truncated]
#> 
#> ℹ Data Dictionary | Landing Page | References

public_Dataset("hospitals")
#> 
#> ── Dataset Hospital Enrollments 
#> → Rows: 9,272 | Fields: 39 | Resources: 5 files
#> 
#> • Periodicity: Monthly [R/P1M]
#> • Last Modified: 2025-01-15
#> • Time Period: 2022-11-01 - 2025-01-31
#> • Theme: Medicare
#> • Keywords: Medicare, Original Medicare, Medicare Advantage, Hospitals &
#>   Facilities, Provider Enrollment
#> 
#> The Hospital Enrollments dataset provides enrollment information of all
#> Hospitals currently enrolled in Medicare. This data includes information on the
#> Hospital's sub-group type, legal business name, doing business as name,
#> organization type and address. 
#> 
#> ℹ Data Dictionary | Landing Page | References

public_Dataset("reassignments")
#> 
#> ── Dataset Revalidation Reassignment List 
#> → Rows: 3,691,875 | Fields: 16 | Resources: 3 files
#> 
#> • Periodicity: Monthly [R/P1M]
#> • Last Modified: 2025-01-02
#> • Time Period: 2022-02-01 - 2025-01-31
#> • Theme: Medicare
#> • Keywords: States & Territories, ZIP Code, Medicare, Original Medicare,
#>   Medicare Advantage, Provider Enrollment
#> 
#> The Revalidation Reassignment List dataset provides information on
#> reassignments of providers who are due for revalidation. Note: This full
#> dataset contains more records than most spreadsheet programs can handle, which
#> will result in an incomplete load of data. Use of a database or statistical
#> software is required.
#> 
#> ℹ Data Dictionary | Landing Page | References

public_Dataset("opt_out")
#> 
#> ── Dataset Opt Out Affidavits 
#> → Rows: 48,702 | Fields: 13 | Resources: 3 files
#> 
#> • Periodicity: Monthly [R/P1M]
#> • Last Modified: 2025-01-22
#> • Time Period: 2023-04-01 - 2024-12-31
#> • Theme: Medicare
#> • Keywords: Medicare, Original Medicare, Medicare Advantage, Provider
#>   Enrollment
#> 
#> ATTENTION USERSSome Providers' Opt-Out Status may end early due to COVID 19
#> waivers. Please contact your respective MAC for further information. For more
#> information on the opt-out process, see Manage Your Enrollment or view the FAQ
#> section below. The Opt Out Affidavits dataset provides information on providers
#> who have decided not to participate in Medicare. It contains provider's NPI,
#> specialty,...[truncated]
#> 
#> ℹ Data Dictionary | Landing Page | References

public_Dataset("laboratories")
#> 
#> ── Dataset Provider of Services File - Clinical Laboratories 
#> → Rows: 653,223 | Fields: 103 | Resources: 3 files
#> 
#> • Periodicity: Quarterly [R/P3M]
#> • Last Modified: 2025-01-16
#> • Time Period: 2011-10-01 - 2024-12-31
#> • Theme: Medicare
#> • Keywords: Medicare, Original Medicare, Medicare Advantage, Hospitals &
#>   Facilities, Outpatient Facilities, Clinics, Health Care Use & Payments
#> 
#> The Provider of Services (POS) Clinical Laboratories (CLIA) data provides
#> information on CLIA demographics and types of testing services the facility
#> provides. In this file you will find provider number, name, address and
#> characteristics of the participating institution providers.
#> 
#> ℹ Data Dictionary | Landing Page | References