Load Public API Distribution

Load Public API Distribution

Usage

public_Distribution(dataset, fname = TRUE)

Arguments

dataset

<chr> dataset title

fname

<lgl> Is dataset a function name?; default is TRUE

Value

<Distribution> object

Examples

public_Distribution("utilization_provider")
#> <provider::Distribution>
#>  @ title             : chr "Medicare Physician & Other Practitioners - by Provider and Service"
#>  @ type              : chr "dcat:Distribution"
#>  @ accessLevel       : chr "public"
#>  @ accrualPeriodicity: chr "Annually [R/P1Y]"
#>  @ bureauCode        : chr "009:38"
#>  @ contactPoint      : <provider::ContactPoint>
#>  .. @ type    : chr "vcard:Contact"
#>  .. @ fn      : chr "Medicare Provider Data - OEDA"
#>  .. @ hasEmail: chr "mailto:IPAG_Data_Products@cms.hhs.gov"
#>  @ describedBy       : chr "https://data.cms.gov/resources/medicare-physician-other-practitioners-by-provider-and-service-data-dictionary"
#>  @ description       : chr "The Medicare Physician & Other Practitioners by Provider and Service dataset provides information on use, payme"| __truncated__
#>  @ keyword           : chr "Medicare, Original Medicare, Physicians & Practitioners, Health Care Use & Payments, Health Equity"
#>  @ landingPage       : chr "https://data.cms.gov/provider-summary-by-type-of-service/medicare-physician-other-practitioners/medicare-physic"| __truncated__
#>  @ programCode       : chr "009:000"
#>  @ publisher         : <provider::Publisher>
#>  .. @ type: chr "org:Organization"
#>  .. @ name: chr "Centers for Medicare & Medicaid Services (CMS)"
#>  @ references        : chr "https://data.cms.gov/resources/medicare-physician-other-practitioners-methodology"
#>  @ distributions     : tibble [20 × 6] (S3: tbl_df/tbl/data.frame)
#>  $ title       : chr [1:20] "Medicare Physician & Other Practitioners - by Provider and Service : 2022-12-01" "Medicare Physician & Other Practitioners - by Provider and Service : 2021-12-31" "Medicare Physician & Other Practitioners - by Provider and Service : 2020-12-31" "Medicare Physician & Other Practitioners - by Provider and Service : 2019-12-31" ...
#>  $ modified    : Date[1:20], format: "2024-06-04" "2023-05-10" ...
#>  $ temporal    : chr [1:20] "2022-01-01/2022-12-31" "2021-01-01/2021-12-31" "2020-01-01/2020-12-31" "2019-01-01/2019-12-31" ...
#>  $ accessURL   : chr [1:20] "https://data.cms.gov/data-api/v1/dataset/4f307be4-6868-4a9e-ae92-acf3fd4b5543/data" "https://data.cms.gov/data-api/v1/dataset/5c67d835-3862-4f63-897d-85d3eac82d5b/data" "https://data.cms.gov/data-api/v1/dataset/862ed658-1f38-4b2f-b02b-0b359e12c78a/data" "https://data.cms.gov/data-api/v1/dataset/5fccd951-9538-48a7-9075-6f02b9867868/data" ...
#>  $ resourcesAPI: chr [1:20] "https://data.cms.gov/data-api/v1/dataset-resources/4f307be4-6868-4a9e-ae92-acf3fd4b5543" "https://data.cms.gov/data-api/v1/dataset-resources/5c67d835-3862-4f63-897d-85d3eac82d5b" "https://data.cms.gov/data-api/v1/dataset-resources/862ed658-1f38-4b2f-b02b-0b359e12c78a" "https://data.cms.gov/data-api/v1/dataset-resources/5fccd951-9538-48a7-9075-6f02b9867868" ...
#>  $ downloadURL : chr [1:20] "https://data.cms.gov/sites/default/files/2024-05/1570d9f0-59ef-416f-bb37-e78a7afe6f88/MUP_PHY_R24_P05_V10_D22_Prov_Svc.csv" "https://data.cms.gov/sites/default/files/2023-05/914a4463-7af3-423f-83a7-b343794e20ee/MUP_PHY_R23_P05_V10_D21_Prov_Svc.csv" "https://data.cms.gov/sites/default/files/2022-11/2260d1cc-9635-4c36-b890-61e672c3fea1/MUP_PHY_R22_P05_V10_D20_Prov_Svc.csv" "https://data.cms.gov/sites/default/files/2021-08/MUP_PHY_R21_P04_V10_D19_Prov_Svc.csv" ...

public_Distribution("quality_payment")
#> <provider::Distribution>
#>  @ title             : chr "Quality Payment Program Experience"
#>  @ type              : chr "dcat:Distribution"
#>  @ accessLevel       : chr "public"
#>  @ accrualPeriodicity: chr "Annually [R/P1Y]"
#>  @ bureauCode        : chr "009:38"
#>  @ contactPoint      : <provider::ContactPoint>
#>  .. @ type    : chr "vcard:Contact"
#>  .. @ fn      : chr "Quality Payment Program - CCSQ"
#>  .. @ hasEmail: chr "mailto:qpp@cms.hhs.gov"
#>  @ describedBy       : chr "https://data.cms.gov/sites/default/files/2024-04/2022%20QPP%20Experience%20Report%20Public%20Use%20Files%20Data"| __truncated__
#>  @ description       : chr "The Quality Payment Program (QPP) Experience dataset provides participation and performance information in the "| __truncated__
#>  @ keyword           : chr "Medicare, Original Medicare, Health Care Use & Payments"
#>  @ landingPage       : chr "https://data.cms.gov/quality-of-care/quality-payment-program-experience"
#>  @ programCode       : chr "009:000"
#>  @ publisher         : <provider::Publisher>
#>  .. @ type: chr "org:Organization"
#>  .. @ name: chr "Centers for Medicare & Medicaid Services (CMS)"
#>  @ references        : chr "https://data.cms.gov/sites/default/files/2024-04/2022%20QPP%20Experience%20Report%20Public%20Use%20Files%20Methodology.pdf"
#>  @ distributions     : tibble [6 × 6] (S3: tbl_df/tbl/data.frame)
#>  $ title       : chr [1:6] "Quality Payment Program Experience : 2022-01-01" "Quality Payment Program Experience : 2021-05-30" "Quality Payment Program Experience : 2020-12-31" "Quality Payment Program Experience : 2019-12-31" ...
#>  $ modified    : Date[1:6], format: "2024-05-08" "2023-06-12" ...
#>  $ temporal    : chr [1:6] "2022-01-01/2022-12-31" "2021-01-01/2021-12-31" "2020-01-01/2020-12-31" "2019-01-01/2019-12-31" ...
#>  $ accessURL   : chr [1:6] "https://data.cms.gov/data-api/v1/dataset/b3438273-b4a6-44ca-8fb2-9e6026b74642/data" "https://data.cms.gov/data-api/v1/dataset/8c2ab3e4-597e-43c9-9d94-7223c66d7f25/data" "https://data.cms.gov/data-api/v1/dataset/48a484e6-ff02-4284-8666-6045b3cac17f/data" "https://data.cms.gov/data-api/v1/dataset/5776e77d-a844-49fd-b697-a2b01bdacdc7/data" ...
#>  $ resourcesAPI: chr [1:6] "https://data.cms.gov/data-api/v1/dataset-resources/b3438273-b4a6-44ca-8fb2-9e6026b74642" "https://data.cms.gov/data-api/v1/dataset-resources/8c2ab3e4-597e-43c9-9d94-7223c66d7f25" "https://data.cms.gov/data-api/v1/dataset-resources/48a484e6-ff02-4284-8666-6045b3cac17f" "https://data.cms.gov/data-api/v1/dataset-resources/5776e77d-a844-49fd-b697-a2b01bdacdc7" ...
#>  $ downloadURL : chr [1:6] "https://data.cms.gov/sites/default/files/2024-04/a963b1e9-d6d2-47a5-a385-80e37884aca2/2022_puf.csv" "https://data.cms.gov/sites/default/files/2023-05/0bac88bc-a4c8-4dff-bdbf-660d24e6a689/2021_puf_QPPAR-18260.csv" "https://data.cms.gov/sites/default/files/2023-04/fb3758dd-6ace-4ddb-863b-1b895cef4d7e/2020_puf_qppar_QPPAR-17660.csv" "https://data.cms.gov/sites/default/files/2023-04/f9ef9f64-91d8-49a2-a1be-5a48bbdcc1e9/2019_puf_qppar_QPPAR-17661.csv" ...