Return the number of results for a given query

Return the number of results for a given query

Usage

query_nresults(obj, query = NULL, ...)

Arguments

obj

An object of class <class_endpoint> or <class_temporal>.

query

A <chr> string representing the query to be passed to the API. If NULL, the default query for the endpoint is used.

...

Additional arguments passed to the query.

Value

An <int> representing the number of results found.

Examples

query_nresults(
   endpoint("care_dial_end"),
   query(state = any_of(c("CA", "GA", "NY", "TX"))))
#> Medicare Dialysis Facilities
#> 3.5M

query_nresults(
   endpoint("quality_payment"),
   query(year = 2020:2023,
   `practice state or us territory` = any_of(c("CA", "GA", "NY", "TX"))))
#> Quality Payment Program Experience
#> $counts
#> # A tibble: 4 × 5
#>    year  found  total  prop url                                                 
#>   <int>  <int>  <int> <dbl> <chr>                                               
#> 1  2023 146120 524998 0.278 https://data.cms.gov/data-api/v1/dataset/896eab17-3…
#> 2  2022 152680 624200 0.245 https://data.cms.gov/data-api/v1/dataset/b3438273-b…
#> 3  2021 179288 698730 0.257 https://data.cms.gov/data-api/v1/dataset/8c2ab3e4-5…
#> 4  2020 241326 921517 0.262 https://data.cms.gov/data-api/v1/dataset/48a484e6-f…
#>