Quickly access CMS data

quick() is a convenience function to quickly access various CMS data endpoints. It provides a simple way to retrieve data from different CMS programs and categories.

Usage

quick(x, offset = 0L, limit = 1L, call = caller_env())

Arguments

x

A string representing the CMS data endpoint or category.

offset

An integer representing the offset for pagination. Default is 0.

limit

An integer representing the maximum number of records to retrieve. Default is 10,000.

call

The calling environment. Default is the current environment.

Value

A data frame or object containing the requested CMS data.

Examples

quick("pro_dialysis") |> str()
#> List of 3
#>  $ dialysis_facility: tibble [1 × 142] (S3: tbl_df/tbl/data.frame)
#>   ..$ cms_certification_number_ccn                                    : chr "012306"
#>   ..$ network                                                         : chr "8"
#>   ..$ facility_name                                                   : chr "CHILDRENS HOSPITAL DIALYSIS"
#>   ..$ five_star_date                                                  : chr "01Jan2020-31Dec2023"
#>   ..$ five_star                                                       : chr NA
#>   ..$ five_star_data_availability_code                                : chr "260"
#>   ..$ address_line_1                                                  : chr "1600 7TH AVENUE SOUTH"
#>   ..$ address_line_2                                                  : chr NA
#>   ..$ citytown                                                        : chr "BIRMINGHAM"
#>   ..$ state                                                           : chr "AL"
#>   ..$ zip_code                                                        : chr "35233"
#>   ..$ countyparish                                                    : chr "Jefferson"
#>   ..$ telephone_number                                                : chr "(205) 638-9275"
#>   ..$ profit_or_nonprofit                                             : chr "Non-profit"
#>   ..$ chain_owned                                                     : chr "No"
#>   ..$ chain_organization                                              : chr "Independent"
#>   ..$ late_shift                                                      : chr "No"
#>   ..$ of_dialysis_stations                                            : chr "7"
#>   ..$ offers_incenter_hemodialysis                                    : chr "Yes"
#>   ..$ offers_peritoneal_dialysis                                      : chr "Yes"
#>   ..$ offers_home_hemodialysis_training                               : chr "Yes"
#>   ..$ certification_date                                              : chr "17NOV1982"
#>   ..$ claims_date                                                     : chr "01JUL2023-30JUN2024"
#>   ..$ eqrs_date                                                       : chr "01JUL2023-30JUN2024"
#>   ..$ smr_date                                                        : chr "01Jan2020-31Dec2023"
#>   ..$ patient_survival_category_text                                  : chr "Not Available"
#>   ..$ patient_survival_data_availability_code                         : chr "199"
#>   ..$ number_of_patients_included_in_survival_summary                 : chr "64"
#>   ..$ mortality_rate_facility                                         : chr NA
#>   ..$ mortality_rate_upper_confidence_limit_975                       : chr NA
#>   ..$ mortality_rate_lower_confidence_limit_25                        : chr NA
#>   ..$ shr_date                                                        : chr "01Jan2023-31Dec2023"
#>   ..$ patient_hospitalization_category_text                           : chr "Worse than Expected"
#>   ..$ patient_hospitalization_data_availability_code                  : chr "001"
#>   ..$ number_of_patients_included_in_hospitalization_summary          : chr "18"
#>   ..$ hospitalization_rate_facility                                   : chr "255.5"
#>   ..$ hospitalization_rate_upper_confidence_limit_975                 : chr "391.9"
#>   ..$ hospitalization_rate_lower_confidence_limit_25                  : chr "176.7"
#>   ..$ srr_date                                                        : chr "01Jan2023-31Dec2023"
#>   ..$ patient_hospital_readmission_category                           : chr "As Expected"
#>   ..$ patient_hospital_readmission_data_availability_code             : chr "001"
#>   ..$ number_of_hospitalizations_included_in_hospital_readmission_fc2b: chr "63"
#>   ..$ readmission_rate_facility                                       : chr "37.4"
#>   ..$ readmission_rate_upper_confidence_limit_975                     : chr "53.5"
#>   ..$ readmission_rate_lower_confidence_limit_25                      : chr "23.9"
#>   ..$ strr_date                                                       : chr "01Jan2023-31Dec2023"
#>   ..$ patient_transfusion_category_text                               : chr "Not Available"
#>   ..$ patient_transfusion_data_availability_code                      : chr "201"
#>   ..$ number_of_patients_included_in_the_transfusion_summary          : chr "0"
#>   ..$ transfusion_rate_facility                                       : chr NA
#>   ..$ transfusion_rate_upper_confidence_limit_975                     : chr NA
#>   ..$ transfusion_rate_lower_confidence_limit_25                      : chr NA
#>   ..$ swr_date                                                        : chr "01Jan2020-31Dec2022"
#>   ..$ swr_category_text                                               : chr "As Expected"
#>   ..$ patient_transplant_waitlist_data_availability_code              : chr "001"
#>   ..$ number_of_patients_in_this_facility_for_swr                     : chr "50"
#>   ..$ standardized_first_kidney_transplant_waitlist_ratio             : chr "1.32"
#>   ..$ 95_ci_upper_limit_for_swr                                       : chr "2.09"
#>   ..$ 95_ci_lower_limit_for_swr                                       : chr "0.78"
#>   ..$ pppw_category_text                                              : chr "As Expected"
#>   ..$ patient_prevalent_transplant_waitlist_data_availability_code    : chr "001"
#>   ..$ number_of_patients_for_pppw                                     : chr "44"
#>   ..$ percentage_of_prevalent_patients_waitlisted                     : chr "20.6"
#>   ..$ 95_ci_upper_limit_for_pppw                                      : chr "42.1"
#>   ..$ 95_ci_lower_limit_for_pppw                                      : chr "8.5"
#>   ..$ sedr_date                                                       : chr "01Jan2023-31Dec2023"
#>   ..$ sedr_category_text                                              : chr "Not Available"
#>   ..$ emergency_department_encounter_data_availability_code           : chr "201"
#>   ..$ number_of_patients_included_in_sedr_summary                     : chr "0"
#>   ..$ standardized_ed_visits_ratio_facility                           : chr NA
#>   ..$ sedr_upper_confidence_limit_975                                 : chr NA
#>   ..$ sedr_lower_confidence_limit_25                                  : chr NA
#>   ..$ ed30_date                                                       : chr "01Jan2022-31Dec2023"
#>   ..$ ed30_category_text                                              : chr "Not Available"
#>   ..$ emergency_department_encounter_ratio_occurring_within_30_da_f8e3: chr "199"
#>   ..$ number_of_hospitalizations_included_in_ed30_summary             : chr "0"
#>   ..$ standardized_ed_visits_within_30_days_of_hospital_discharge_6307: chr NA
#>   ..$ ed30_upper_confidence_limit_975                                 : chr NA
#>   ..$ ed30_lower_confidence_limit_25                                  : chr NA
#>   ..$ years_modality_switch_based_upon                                : chr "01Jan2020-31Dec2022"
#>   ..$ smosr_classification_category_facility                          : chr "Not Available"
#>   ..$ modality_switch_data_availability_code                          : chr "201"
#>   ..$ smosr_number_of_eligible_patients_facility                      : chr "0"
#>   ..$ smosr_standardized_modality_switch_ratio_facility               : chr NA
#>   ..$ smosr_upper_confidence_limit_facility                           : chr NA
#>   ..$ smosr_lower_confidence_limit_facility                           : chr NA
#>   ..$ sir_date                                                        : chr "01Jan2023-31Dec2023"
#>   ..$ patient_infection_category_text                                 : chr "As Expected"
#>   ..$ patient_infection_data_availability_code                        : chr "001"
#>   ..$ standard_infection_ratio                                        : chr "0.86"
#>   ..$ sir_upper_confidence_limit_975                                  : chr "2.33"
#>   ..$ sir_lower_confidence_limit_25                                   : chr "0.22"
#>   ..$ fistula_category_text                                           : chr "Not Available"
#>   ..$ fistula_data_availability_code                                  : chr "199"
#>   ..$ number_of_patients_included_in_fistula_summary                  : chr "1"
#>   ..$ fistula_rate_facility                                           : chr NA
#>   ..$ fistula_rate_upper_confidence_limit_975                         : chr NA
#>   ..$ fistula_rate_lower_confidence_limit_25                          : chr NA
#>   ..$ hcp_vaccination_data_collection_dates                           : chr "01APR2024-30JUN2024"
#>   .. [list output truncated]
#>  $ dialysis_national: tibble [1 × 59] (S3: tbl_df/tbl/data.frame)
#>   ..$ country                                                         : chr "NATION"
#>   ..$ transfusion_rate_us                                             : chr "32.8"
#>   ..$ transfusions__better_than_expected_us                           : chr "4"
#>   ..$ transfusions__as_expected_us                                    : chr "4668"
#>   ..$ transfusions__worse_than_expected_us                            : chr "341"
#>   ..$ percentage_of_adult_hd_patients_with_ktv12                      : chr "97"
#>   ..$ percentage_of_adult_pd_patients_with_ktv17                      : chr "92"
#>   ..$ percentage_of_pediatric_hd_patients_with_ktv12                  : chr "86"
#>   ..$ percentage_of_pediatric_pd_patients_with_ktv18                  : chr "77"
#>   ..$ percentage_of_pediatric_hd_patients_with_npcr                   : chr "88"
#>   ..$ fistula_rate_us                                                 : chr "58.1"
#>   ..$ fistula_rate__better_than_expected_us                           : chr "278"
#>   ..$ fistula_rate__as_expected_us                                    : chr "6512"
#>   ..$ fistula_rate__worse_than_expected_us                            : chr "257"
#>   ..$ percentage_of_adult_patients_with_long_term_catheter_in_use     : chr "18"
#>   ..$ percentage_of_adult_patients_with_hypercalcemia_serum_calci_044d: chr "2"
#>   ..$ percentage_of_adult_patients_with_serum_phosphorus_less_tha_c222: chr "7"
#>   ..$ percentage_of_adult_patients_with_serum_phosphorus_between__85e8: chr "23"
#>   ..$ percentage_of_adult_patients_with_serum_phosphorus_between__fad7: chr "28"
#>   ..$ percentage_of_adult_patients_with_serum_phosphorus_between__ff32: chr "25"
#>   ..$ percentage_of_adult_patients_with_serum_phosphorus_greater__d8e3: chr "17"
#>   ..$ hospitalization_rate_us                                         : chr "144.7"
#>   ..$ hospitalizations__better_than_expected_us                       : chr "97"
#>   ..$ hospitalizations__as_expected_us                                : chr "6825"
#>   ..$ hospitalizations__worse_than_expected_us                        : chr "340"
#>   ..$ readmission_rate_us                                             : chr "26.5"
#>   ..$ hospital_readmission__better_than_expected_us                   : chr "132"
#>   ..$ hospital_readmission__as_expected_us                            : chr "6709"
#>   ..$ hospital__readmission__worse_than_expected_us                   : chr "237"
#>   ..$ mortality_rate_us                                               : chr "22.5"
#>   ..$ survival__better_than_expected_us                               : chr "144"
#>   ..$ survival__as_expected_us                                        : chr "6867"
#>   ..$ survival__worse_than_expected_us                                : chr "174"
#>   ..$ standardized_first_kidney_transplant_waitlist_ratio_us          : chr "1"
#>   ..$ incident_patients_transplant_waitlisting__better_than_expec_a24d: chr "277"
#>   ..$ incident_patients_transplant_waitlisting__as_expected_us        : chr "3660"
#>   ..$ incident_patients_transplant_waitlisting__worse_than_expected_us: chr "196"
#>   ..$ percentage_of_prevalent_patients_waitlisted_us                  : chr "16"
#>   ..$ prevalent_patients_transplant_waitlisting_better_than_expec_c0cd: chr "433"
#>   ..$ prevalent_patients_transplant_waitlisting__as_expected_us       : chr "6730"
#>   ..$ prevalent_patients_transplant_waitlisting__worse_than_expec_e6e9: chr "159"
#>   ..$ healthcare_worker_covid19_vaccination_adherence_percentage_us   : chr "3"
#>   ..$ percentage_of_patients_with_hgb10_gdl_us                        : chr "21"
#>   ..$ percentage_of_patients_with_hgb12_gdl_us                        : chr "1"
#>   ..$ infection__better_than_expected_us                              : chr "3319"
#>   ..$ infection__as_expected_us                                       : chr "3106"
#>   ..$ infection__worse_than_expected_us                               : chr "39"
#>   ..$ standardized_ed_visits_ratio_us                                 : chr "0.99"
#>   ..$ sedr__better_than_expected_us                                   : chr "31"
#>   ..$ sedr__as_expected_us                                            : chr "6472"
#>   ..$ sedr__worse_than_expected_us                                    : chr "418"
#>   ..$ standardized_ed_visits_within_30_days_of_hospital_discharge_d0c6: chr "1.04"
#>   ..$ ed30__better_than_expected_us                                   : chr "138"
#>   ..$ ed30__as_expected_us                                            : chr "6588"
#>   ..$ ed30__worse_than_expected_us                                    : chr "203"
#>   ..$ standardized_modality_switch_ratio                              : chr "1"
#>   ..$ n_for_smsr_category_1_smsrbetter_than_expected_us               : chr "585"
#>   ..$ n_for_smsr_category_2_smsras_expected_us                        : chr "5518"
#>   ..$ n_for_smsr_category_3_smsrworse_than_expected_us                : chr "441"
#>  $ dialysis_state   : tibble [1 × 49] (S3: tbl_df/tbl/data.frame)
#>   ..$ state                                                           : chr "AK"
#>   ..$ survival_better_than_expected_state                             : chr "0"
#>   ..$ survival_as_expected_state                                      : chr "8"
#>   ..$ survival_worse_than_expected_state                              : chr "0"
#>   ..$ hospitalizations_better_than_expected_state                     : chr "0"
#>   ..$ hospitalizations_as_expected_state                              : chr "8"
#>   ..$ hospitalizations_worse_than_expected_state                      : chr "0"
#>   ..$ hospital_readmission__better_than_expected_state                : chr "0"
#>   ..$ hospital_readmission__as_expected_state                         : chr "7"
#>   ..$ hospital__readmission__worse_than_expected_state                : chr "0"
#>   ..$ transfusions_better_than_expected_state                         : chr "0"
#>   ..$ transfusions_as_expected_state                                  : chr "8"
#>   ..$ transfusions_worse_than_expected_state                          : chr "0"
#>   ..$ infection_better_than_expected_state                            : chr "0"
#>   ..$ infection_as_expected_state                                     : chr "8"
#>   ..$ infection_worse_than_expected_state                             : chr "0"
#>   ..$ percentage_of_adult_hd_patients_with_ktv12                      : chr "98"
#>   ..$ percentage_of_adult_pd_patients_with_ktv17                      : chr "95"
#>   ..$ percentage_of_pediatric_hd_patients_with_ktv12                  : chr NA
#>   ..$ percentage_of_pediatric_pd_patients_with_ktv18                  : chr NA
#>   ..$ percentage_of_pediatric_hd_patients_with_npcr_in_use            : chr NA
#>   ..$ fistula_rate__better_than_expected_state                        : chr "2"
#>   ..$ fistula_rate__as_expected_state                                 : chr "6"
#>   ..$ fistula_rate__worse_than_expected_state                         : chr "0"
#>   ..$ percentage_of_adult_patients_with_long_term_catheter_in_use     : chr "13"
#>   ..$ percentage_of_adult_patients_with_hypercalcemia_serum_calci_044d: chr "0"
#>   ..$ percentage_of_adult_patients_with_serum_phosphorus_less_tha_c222: chr "9"
#>   ..$ percentage_of_adult_patients_with_serum_phosphorus_between__85e8: chr "25"
#>   ..$ percentage_of_adult_patients_with_serum_phosphorus_between__fad7: chr "28"
#>   ..$ percentage_of_adult_patients_with_serum_phosphorus_between__ff32: chr "22"
#>   ..$ percentage_of_adult_patients_with_serum_phosphorus_greater__d8e3: chr "17"
#>   ..$ incident_patients_transplant_waitlisting_better_than_expect_49e4: chr "0"
#>   ..$ incident_patients_transplant_waitlisting__as_expected_state     : chr "5"
#>   ..$ incident_patients_transplant_waitlisting__worse_than_expect_9b13: chr "0"
#>   ..$ prevalent_patients_transplant_waitlisting_better_than_expec_c046: chr "1"
#>   ..$ prevalent_patients_transplant_waitlisting__as_expected_state    : chr "6"
#>   ..$ prevalent_patients_transplant_waitlisting__worse_than_expec_eeb0: chr "1"
#>   ..$ healthcare_worker_covid19_vaccination_adherence_percentage_state: chr "11"
#>   ..$ percentage_of_patients_with_hgb10_gdl_state                     : chr "16"
#>   ..$ percentage_of_patients_with_hgb12_gdl_state                     : chr "0"
#>   ..$ sedr__better_than_expected_state                                : chr "0"
#>   ..$ sedr__as_expected_state                                         : chr "4"
#>   ..$ sedr__worse_than_expected_state                                 : chr "4"
#>   ..$ ed30__better_than_expected_state                                : chr "0"
#>   ..$ ed30__as_expected_state                                         : chr "7"
#>   ..$ ed30__worse_than_expected_state                                 : chr "1"
#>   ..$ n_for_smsr_category_1_smsrbetter_than_expected_state            : chr "2"
#>   ..$ n_for_smsr_category_2_smsras_expected_state                     : chr "6"
#>   ..$ n_for_smsr_category_3_smsrworse_than_expected_state             : chr "0"