Standardise a query object to match the fields of an endpoint

Standardise a query object to match the fields of an endpoint

Usage

standardise(obj, qry)

Arguments

obj

A <class_care/caid/prov/open/hgov> object.

qry

A <class_query> object.

Value

A named list of query parameters that match the fields of the endpoint.

Examples

qry <- new_query(
   first_name = starts_with("Andr"),
   last_name  = contains("J"),
   state_cd   = any_of(c("CA", "GA", "NY")),
   city       = equals(c("Atlanta", "Los Angeles"), negate = TRUE),
   npi        = npi_ex$k,
   covered_recipient_npi = npi_ex$k,
   ccn        = "01256",
   rate       = between(0.45, 0.67),
   year       = 2014:2025)

standardise(endpoint("enroll_prov"), qry)
#> $params
#> $params$NPI
#>  [1] 1225701881 1174270805 1235702796 1962116806 1013647569 1306500665
#>  [7] 1982296737 1083295638 1841967825 1891390084 1275117269 1992338701
#> [13] 1891355863 1548743511 1023473279 1861857013 1689182859 1982059275
#> 
#> $params$STATE_CD
#> <any_of>
#>  @ operator: chr "IN"
#>  @ value   : chr [1:3] "CA" "GA" "NY"
#> 
#> $params$FIRST_NAME
#> <starts_with>
#>  @ operator: chr "STARTS_WITH"
#>  @ value   : chr "Andr"
#> 
#> $params$LAST_NAME
#> <contains>
#>  @ operator: chr "CONTAINS"
#>  @ value   : chr "J"
#> 
#> 
#> $query
#> $query$NPI
#>  [1] "filter[1][condition][path]=NPI"            
#>  [2] "filter[1][condition][operator]=="          
#>  [3] "filter[1][condition][value][1]=1225701881" 
#>  [4] "filter[1][condition][value][2]=1174270805" 
#>  [5] "filter[1][condition][value][3]=1235702796" 
#>  [6] "filter[1][condition][value][4]=1962116806" 
#>  [7] "filter[1][condition][value][5]=1013647569" 
#>  [8] "filter[1][condition][value][6]=1306500665" 
#>  [9] "filter[1][condition][value][7]=1982296737" 
#> [10] "filter[1][condition][value][8]=1083295638" 
#> [11] "filter[1][condition][value][9]=1841967825" 
#> [12] "filter[1][condition][value][10]=1891390084"
#> [13] "filter[1][condition][value][11]=1275117269"
#> [14] "filter[1][condition][value][12]=1992338701"
#> [15] "filter[1][condition][value][13]=1891355863"
#> [16] "filter[1][condition][value][14]=1548743511"
#> [17] "filter[1][condition][value][15]=1023473279"
#> [18] "filter[1][condition][value][16]=1861857013"
#> [19] "filter[1][condition][value][17]=1689182859"
#> [20] "filter[1][condition][value][18]=1982059275"
#> 
#> $query$STATE_CD
#> [1] "filter[2][condition][path]=STATE_CD" "filter[2][condition][operator]=IN"  
#> [3] "filter[2][condition][value][1]=CA"   "filter[2][condition][value][2]=GA"  
#> [5] "filter[2][condition][value][3]=NY"  
#> 
#> $query$FIRST_NAME
#> [1] "filter[3][condition][path]=FIRST_NAME"     
#> [2] "filter[3][condition][operator]=STARTS_WITH"
#> [3] "filter[3][condition][value]=Andr"          
#> 
#> $query$LAST_NAME
#> [1] "filter[4][condition][path]=LAST_NAME"   
#> [2] "filter[4][condition][operator]=CONTAINS"
#> [3] "filter[4][condition][value]=J"          
#> 
#> 
#> $identifier
#> [1] "https://data.cms.gov/data-api/v1/dataset/2457ea29-fc82-48b0-86ec-3b0755de7515/data-viewer?offset=0&size=1"
#> 

standardise(endpoint("dial_facility"), qry)
#> $params
#> $params$CCN
#> [1] "01256"
#> 
#> $params$city
#> <equals>
#>  @ operator: chr "<>"
#>  @ value   : chr [1:2] "Atlanta" "Los Angeles"
#> 
#> $params$NPI
#>  [1] 1225701881 1174270805 1235702796 1962116806 1013647569 1306500665
#>  [7] 1982296737 1083295638 1841967825 1891390084 1275117269 1992338701
#> [13] 1891355863 1548743511 1023473279 1861857013 1689182859 1982059275
#> 
#> $params$year
#>  [1] 2014 2015 2016 2017 2018 2019 2020 2021 2022 2023 2024 2025
#> 
#> 
#> $query
#> $query$CCN
#> [1] "filter[1][condition][path]=CCN"    "filter[1][condition][operator]==" 
#> [3] "filter[1][condition][value]=01256"
#> 
#> $query$city
#> [1] "filter[2][condition][path]=city"           
#> [2] "filter[2][condition][operator]=<>"         
#> [3] "filter[2][condition][value][1]=Atlanta"    
#> [4] "filter[2][condition][value][2]=Los Angeles"
#> 
#> $query$NPI
#>  [1] "filter[3][condition][path]=NPI"            
#>  [2] "filter[3][condition][operator]=="          
#>  [3] "filter[3][condition][value][1]=1225701881" 
#>  [4] "filter[3][condition][value][2]=1174270805" 
#>  [5] "filter[3][condition][value][3]=1235702796" 
#>  [6] "filter[3][condition][value][4]=1962116806" 
#>  [7] "filter[3][condition][value][5]=1013647569" 
#>  [8] "filter[3][condition][value][6]=1306500665" 
#>  [9] "filter[3][condition][value][7]=1982296737" 
#> [10] "filter[3][condition][value][8]=1083295638" 
#> [11] "filter[3][condition][value][9]=1841967825" 
#> [12] "filter[3][condition][value][10]=1891390084"
#> [13] "filter[3][condition][value][11]=1275117269"
#> [14] "filter[3][condition][value][12]=1992338701"
#> [15] "filter[3][condition][value][13]=1891355863"
#> [16] "filter[3][condition][value][14]=1548743511"
#> [17] "filter[3][condition][value][15]=1023473279"
#> [18] "filter[3][condition][value][16]=1861857013"
#> [19] "filter[3][condition][value][17]=1689182859"
#> [20] "filter[3][condition][value][18]=1982059275"
#> 
#> $query$year
#>  [1] "filter[4][condition][path]=year"     
#>  [2] "filter[4][condition][operator]=="    
#>  [3] "filter[4][condition][value][1]=2014" 
#>  [4] "filter[4][condition][value][2]=2015" 
#>  [5] "filter[4][condition][value][3]=2016" 
#>  [6] "filter[4][condition][value][4]=2017" 
#>  [7] "filter[4][condition][value][5]=2018" 
#>  [8] "filter[4][condition][value][6]=2019" 
#>  [9] "filter[4][condition][value][7]=2020" 
#> [10] "filter[4][condition][value][8]=2021" 
#> [11] "filter[4][condition][value][9]=2022" 
#> [12] "filter[4][condition][value][10]=2023"
#> [13] "filter[4][condition][value][11]=2024"
#> [14] "filter[4][condition][value][12]=2025"
#> 
#> 
#> $identifier
#>                                                                                                 2025 
#> "https://data.cms.gov/data-api/v1/dataset/7d4e7bd1-f08a-4fca-8085-6af24026a19a/data?offset=0&size=1" 
#>                                                                                                 2024 
#> "https://data.cms.gov/data-api/v1/dataset/62d125ff-c959-4c5c-8a46-8cacba50bfd4/data?offset=0&size=1" 
#>                                                                                                 2023 
#> "https://data.cms.gov/data-api/v1/dataset/200e7e7e-a5a6-4450-ba18-ab53d237cc45/data?offset=0&size=1" 
#>                                                                                                 2022 
#> "https://data.cms.gov/data-api/v1/dataset/ebc64ad1-164c-455a-80ca-cb0753e44a46/data?offset=0&size=1" 
#>                                                                                                 2021 
#> "https://data.cms.gov/data-api/v1/dataset/e66eaf03-398f-4a8c-9d1b-6aaf25fb6425/data?offset=0&size=1" 
#>                                                                                                 2020 
#> "https://data.cms.gov/data-api/v1/dataset/44aedde0-c535-497d-ab27-b415ec12ff5a/data?offset=0&size=1" 
#> 

standardise(endpoint("quality_payment"), qry)
#> $params
#> $params$npi
#>  [1] 1225701881 1174270805 1235702796 1962116806 1013647569 1306500665
#>  [7] 1982296737 1083295638 1841967825 1891390084 1275117269 1992338701
#> [13] 1891355863 1548743511 1023473279 1861857013 1689182859 1982059275
#> 
#> 
#> $query
#> $query$npi
#>  [1] "filter[1][condition][path]=npi"            
#>  [2] "filter[1][condition][operator]=="          
#>  [3] "filter[1][condition][value][1]=1225701881" 
#>  [4] "filter[1][condition][value][2]=1174270805" 
#>  [5] "filter[1][condition][value][3]=1235702796" 
#>  [6] "filter[1][condition][value][4]=1962116806" 
#>  [7] "filter[1][condition][value][5]=1013647569" 
#>  [8] "filter[1][condition][value][6]=1306500665" 
#>  [9] "filter[1][condition][value][7]=1982296737" 
#> [10] "filter[1][condition][value][8]=1083295638" 
#> [11] "filter[1][condition][value][9]=1841967825" 
#> [12] "filter[1][condition][value][10]=1891390084"
#> [13] "filter[1][condition][value][11]=1275117269"
#> [14] "filter[1][condition][value][12]=1992338701"
#> [15] "filter[1][condition][value][13]=1891355863"
#> [16] "filter[1][condition][value][14]=1548743511"
#> [17] "filter[1][condition][value][15]=1023473279"
#> [18] "filter[1][condition][value][16]=1861857013"
#> [19] "filter[1][condition][value][17]=1689182859"
#> [20] "filter[1][condition][value][18]=1982059275"
#> 
#> 
#> $identifier
#>                                                                                                 2023 
#> "https://data.cms.gov/data-api/v1/dataset/3545dc8e-7037-41ee-8a1d-9f3198fe4927/data?offset=0&size=1" 
#>                                                                                                 2022 
#> "https://data.cms.gov/data-api/v1/dataset/b3438273-b4a6-44ca-8fb2-9e6026b74642/data?offset=0&size=1" 
#>                                                                                                 2021 
#> "https://data.cms.gov/data-api/v1/dataset/8c2ab3e4-597e-43c9-9d94-7223c66d7f25/data?offset=0&size=1" 
#>                                                                                                 2020 
#> "https://data.cms.gov/data-api/v1/dataset/48a484e6-ff02-4284-8666-6045b3cac17f/data?offset=0&size=1" 
#>                                                                                                 2019 
#> "https://data.cms.gov/data-api/v1/dataset/5776e77d-a844-49fd-b697-a2b01bdacdc7/data?offset=0&size=1" 
#>                                                                                                 2018 
#> "https://data.cms.gov/data-api/v1/dataset/4cae2e39-693d-4699-ac23-2611b29c5da7/data?offset=0&size=1" 
#>                                                                                                 2017 
#> "https://data.cms.gov/data-api/v1/dataset/d87865f9-70b6-41fb-be3a-db735c4a05d2/data?offset=0&size=1" 
#> 

standardise(endpoint("pdc_clinician"), qry)
#> $params
#> $params$npi
#>  [1] 1225701881 1174270805 1235702796 1962116806 1013647569 1306500665
#>  [7] 1982296737 1083295638 1841967825 1891390084 1275117269 1992338701
#> [13] 1891355863 1548743511 1023473279 1861857013 1689182859 1982059275
#> 
#> 
#> $query
#> $query$npi
#>  [1] "conditions[0][property]=npi"         "conditions[0][operator]=="          
#>  [3] "conditions[0][value][1]=1225701881"  "conditions[0][value][2]=1174270805" 
#>  [5] "conditions[0][value][3]=1235702796"  "conditions[0][value][4]=1962116806" 
#>  [7] "conditions[0][value][5]=1013647569"  "conditions[0][value][6]=1306500665" 
#>  [9] "conditions[0][value][7]=1982296737"  "conditions[0][value][8]=1083295638" 
#> [11] "conditions[0][value][9]=1841967825"  "conditions[0][value][10]=1891390084"
#> [13] "conditions[0][value][11]=1275117269" "conditions[0][value][12]=1992338701"
#> [15] "conditions[0][value][13]=1891355863" "conditions[0][value][14]=1548743511"
#> [17] "conditions[0][value][15]=1023473279" "conditions[0][value][16]=1861857013"
#> [19] "conditions[0][value][17]=1689182859" "conditions[0][value][18]=1982059275"
#> 
#> 
#> $identifier
#> [1] "https://data.cms.gov/provider-data/api/1/datastore/query/mj5m-pzi6/0?count=true&results=true&offset=0&limit=1"
#> 

standardise(endpoint("grp_cover_nature"), qry)
#> $params
#> $params$covered_recipient_npi
#>  [1] 1225701881 1174270805 1235702796 1962116806 1013647569 1306500665
#>  [7] 1982296737 1083295638 1841967825 1891390084 1275117269 1992338701
#> [13] 1891355863 1548743511 1023473279 1861857013 1689182859 1982059275
#> 
#> 
#> $query
#> $query$covered_recipient_npi
#>  [1] "conditions[0][property]=covered_recipient_npi"
#>  [2] "conditions[0][operator]=="                    
#>  [3] "conditions[0][value][1]=1225701881"           
#>  [4] "conditions[0][value][2]=1174270805"           
#>  [5] "conditions[0][value][3]=1235702796"           
#>  [6] "conditions[0][value][4]=1962116806"           
#>  [7] "conditions[0][value][5]=1013647569"           
#>  [8] "conditions[0][value][6]=1306500665"           
#>  [9] "conditions[0][value][7]=1982296737"           
#> [10] "conditions[0][value][8]=1083295638"           
#> [11] "conditions[0][value][9]=1841967825"           
#> [12] "conditions[0][value][10]=1891390084"          
#> [13] "conditions[0][value][11]=1275117269"          
#> [14] "conditions[0][value][12]=1992338701"          
#> [15] "conditions[0][value][13]=1891355863"          
#> [16] "conditions[0][value][14]=1548743511"          
#> [17] "conditions[0][value][15]=1023473279"          
#> [18] "conditions[0][value][16]=1861857013"          
#> [19] "conditions[0][value][17]=1689182859"          
#> [20] "conditions[0][value][18]=1982059275"          
#> 
#> 
#> $identifier
#>                                                                                                                                     2024 
#> "https://openpaymentsdata.cms.gov/api/1/datastore/query/88822473-9093-46fc-ad38-81330eb8de4b/0?count=true&results=true&offset=0&limit=1" 
#>                                                                                                                                     2023 
#> "https://openpaymentsdata.cms.gov/api/1/datastore/query/4017acfa-0eeb-4b8a-b68c-3b41e779c4a2/0?count=true&results=true&offset=0&limit=1" 
#>                                                                                                                                     2022 
#> "https://openpaymentsdata.cms.gov/api/1/datastore/query/686b765e-f4b7-4d35-a050-fb5af926dbbe/0?count=true&results=true&offset=0&limit=1" 
#>                                                                                                                                     2021 
#> "https://openpaymentsdata.cms.gov/api/1/datastore/query/d35243be-55cb-5087-931e-a294e0e0917f/0?count=true&results=true&offset=0&limit=1" 
#>                                                                                                                                     2020 
#> "https://openpaymentsdata.cms.gov/api/1/datastore/query/1fd2be47-cb72-435a-b2df-142f38e99a05/0?count=true&results=true&offset=0&limit=1" 
#>                                                                                                                                     2019 
#> "https://openpaymentsdata.cms.gov/api/1/datastore/query/0ff0961f-8295-477c-8044-85b9a0753893/0?count=true&results=true&offset=0&limit=1" 
#>                                                                                                                                     2018 
#> "https://openpaymentsdata.cms.gov/api/1/datastore/query/07c1e377-eb5b-4830-ad9b-61f04258036e/0?count=true&results=true&offset=0&limit=1" 
#>