Standardise a query object to match the fields of an endpoint

Standardise a query object to match the fields of an endpoint

Usage

standardise_query(obj, qry)

Arguments

obj

A <class_care/caid/prov/open/hgov> object.

qry

A <class_query> object.

Value

A named list of query parameters that match the fields of the endpoint.

Examples

qry <- new_query(
   first_name = starts_with("Andr"),
   last_name  = contains("J"),
   state_cd   = any_of(c("CA", "GA", "NY")),
   city       = equals(c("Atlanta", "Los Angeles"), negate = TRUE),
   npi        = npi_ex$k,
   ccn        = "01256",
   rate       = between(0.45, 0.67),
   year       = 2014:2025)

obj <- endpoint("enroll_prov")
standardise_query(obj, qry)
#> $NPI
#>  [1] 1225701881 1174270805 1235702796 1962116806 1013647569 1306500665
#>  [7] 1982296737 1083295638 1841967825 1891390084 1275117269 1992338701
#> [13] 1891355863 1548743511 1023473279 1861857013 1689182859 1982059275
#> 
#> $STATE_CD
#> <any_of>
#>  @ operator: chr "IN"
#>  @ value   : chr [1:3] "CA" "GA" "NY"
#> 
#> $FIRST_NAME
#> <starts_with>
#>  @ operator: chr "STARTS_WITH"
#>  @ value   : chr "Andr"
#> 
#> $LAST_NAME
#> <contains>
#>  @ operator: chr "CONTAINS"
#>  @ value   : chr "J"
#> 

obj <- endpoint("dial_facility")
standardise_query(obj, qry)
#> $CCN
#> [1] "01256"
#> 
#> $city
#> <equals>
#>  @ operator: chr "<>"
#>  @ value   : chr [1:2] "Atlanta" "Los Angeles"
#> 
#> $NPI
#>  [1] 1225701881 1174270805 1235702796 1962116806 1013647569 1306500665
#>  [7] 1982296737 1083295638 1841967825 1891390084 1275117269 1992338701
#> [13] 1891355863 1548743511 1023473279 1861857013 1689182859 1982059275
#> 
#> $year
#>  [1] 2014 2015 2016 2017 2018 2019 2020 2021 2022 2023 2024 2025
#> 

obj <- endpoint("quality_payment")
standardise_query(obj, qry)
#> $npi
#>  [1] 1225701881 1174270805 1235702796 1962116806 1013647569 1306500665
#>  [7] 1982296737 1083295638 1841967825 1891390084 1275117269 1992338701
#> [13] 1891355863 1548743511 1023473279 1861857013 1689182859 1982059275
#>