PDC Utilization

The Doctors and Clinicians utilization data file reports volume information for procedures of interest on clinician profile pages and in the provider data catalog (PDC) to inform patients and caregivers about clinicians' experience.

Usage

utilization(
  npi = NULL,
  pac = NULL,
  last_name = NULL,
  first_name = NULL,
  middle_name = NULL,
  suffix = NULL,
  procedure = NULL,
  count = NULL,
  percentile = NULL
)

Arguments

npi

<chr> 10-digit Individual NPI

pac

<chr> 10-digit PECOS Associate Control (PAC) ID

first_name, middle_name, last_name, suffix

<chr> Individual provider's name

procedure

<chr> Procedure category

count

<int> Number of procedures performed

percentile

<int> Percentile of procedures performed

Examples

utilization()
#> # A tibble: 2,000 × 10
#>    npi     pac   last_name first_name middle_name suff  procedure_category count
#>  * <chr>   <chr> <chr>     <chr>      <chr>       <chr> <chr>              <chr>
#>  1 100300… 9931… KHALIL    RASHID     NA          NA    Spinal fusion      1-10 
#>  2 100300… 0446… ROTHCHILD KEVIN      B           NA    Hernia repair - g… 1-10 
#>  3 100300… 0446… ROTHCHILD KEVIN      B           NA    Hernia repair (mi… 1-10 
#>  4 100300… 4082… KIM       DAE        NA          NA    Prostate resection 21   
#>  5 100300… 8527… BENHARASH PEYMAN     NA          NA    Coronary artery b… 17   
#>  6 100300… 2860… WARREN    KASEY      W           NA    Cataract surgery   1-10 
#>  7 100300… 5799… JONES     JACLYN     C           NA    Lower limb (leg) … 17   
#>  8 100300… 5799… JONES     JACLYN     C           NA    Upper limb (arm) … 27   
#>  9 100300… 6002… HAMMAN    DANIEL     RICHARD     NA    Knee replacement   1-10 
#> 10 100300… 6002… HAMMAN    DANIEL     RICHARD     NA    Lower limb (leg) … 1-10 
#> # ℹ 1,990 more rows
#> # ℹ 2 more variables: percentile <chr>, profile <chr>