Provider Facility Affiliations

Facility affiliations data publicly reported in the Provider Data Catalog.

Usage

affiliations(
  npi = NULL,
  pac = NULL,
  last_name = NULL,
  first_name = NULL,
  middle_name = NULL,
  suffix = NULL,
  facility_type = NULL,
  ccn_facility = NULL,
  ccn_primary = NULL
)

Arguments

npi

<chr> 10-digit Individual NPI

pac

<chr> 10-digit PECOS Associate Control (PAC) ID

first_name, middle_name, last_name, suffix

<chr> Individual provider's name

facility_type

<chr> Type of facility, one of the following:

"Hospital"

"Long-term care hospital"

"Nursing home"

"Inpatient rehabilitation facility"

"Home health agency"

"Skilled nursing facility"

"Hospice"

"Dialysis facility"

ccn_facility

<chr> 6-digit CMS Certification Number (CCN) of a facility or unit within hospital where an individual provider provides service.

ccn_primary

<int> 6-digit CMS Certification Number (CCN) of a sub-unit's primary hospital, should the provider provide services in said unit.

Value

<tibble> of search results

Periodicity

Monthly (R/P1M)

Examples

affiliations(facility_type = "Home health agency")
#> # A tibble: 2,000 × 9
#>    npi   pac   last_name first_name middle_name suff  facility_type ccn_facility
#>  * <chr> <chr> <chr>     <chr>      <chr>       <chr> <chr>         <chr>       
#>  1 1003… 8527… BENHARASH PEYMAN     NA          NA    Home health … 057761      
#>  2 1003… 8527… BENHARASH PEYMAN     NA          NA    Home health … 058097      
#>  3 1003… 1153… LOHANO    JAIVANTI   NA          NA    Home health … 157152      
#>  4 1003… 6305… SANTIAGO  CYNTHIA    NA          NA    Home health … 337147      
#>  5 1003… 0547… MILLER    CAROLYN    K           NA    Home health … 397006      
#>  6 1003… 8325… SABODASH  VALERIY    NA          NA    Home health … 109119      
#>  7 1003… 8325… SABODASH  VALERIY    NA          NA    Home health … 109793      
#>  8 1003… 8325… SABODASH  VALERIY    NA          NA    Home health … 687176      
#>  9 1003… 8325… SABODASH  VALERIY    NA          NA    Home health … 687173      
#> 10 1003… 8820… DUNN      GRACE      EIZABETH    NA    Home health … 107486      
#> # ℹ 1,990 more rows
#> # ℹ 1 more variable: ccn_primary <chr>
affiliations(last_name = "CURRY")
#> # A tibble: 252 × 9
#>    npi   pac   last_name first_name middle_name suff  facility_type ccn_facility
#>  * <chr> <chr> <chr>     <chr>      <chr>       <chr> <chr>         <chr>       
#>  1 1033… 4082… CURRY     ROBIN      G           NA    Hospital      180088      
#>  2 1033… 4082… CURRY     ROBIN      G           NA    Hospital      180012      
#>  3 1033… 4082… CURRY     ROBIN      G           NA    Hospital      180130      
#>  4 1033… 4082… CURRY     ROBIN      G           NA    Hospital      180138      
#>  5 1033… 8628… CURRY     SONDRA     NA          NA    Hospital      010033      
#>  6 1043… 5890… CURRY     JOSEPH     M           NA    Home health … 317073      
#>  7 1043… 5890… CURRY     JOSEPH     M           NA    Hospital      390174      
#>  8 1043… 5890… CURRY     JOSEPH     M           NA    Hospital      310086      
#>  9 1043… 5890… CURRY     JOSEPH     M           NA    Hospital      390115      
#> 10 1043… 5395… CURRY     SHAWN      J           NA    Home health … 157217      
#> # ℹ 242 more rows
#> # ℹ 1 more variable: ccn_primary <chr>
affiliations(ccn_primary = "670055")
#> # A tibble: 6 × 9
#>   npi    pac   last_name first_name middle_name suff  facility_type ccn_facility
#> * <chr>  <chr> <chr>     <chr>      <chr>       <chr> <chr>         <chr>       
#> 1 10831… 3072… ASAMOA    JOSHUA     J           NA    Inpatient re… 67T055      
#> 2 15585… 8921… ALFONSO   JOHN       D           NA    Inpatient re… 67T055      
#> 3 17101… 9739… AMARANENI VIDYA      NA          NA    Inpatient re… 67T055      
#> 4 17200… 7012… JANES     WILLIAM    WARREN      NA    Inpatient re… 67T055      
#> 5 17405… 0345… WEIKLE    GEOFF      R           NA    Inpatient re… 67T055      
#> 6 19122… 5092… MANSOOR   SAAD       MD          NA    Inpatient re… 67T055      
#> # ℹ 1 more variable: ccn_primary <chr>
affiliations(ccn_facility = "370781")
#> # A tibble: 18 × 9
#>    npi   pac   last_name first_name middle_name suff  facility_type ccn_facility
#>  * <chr> <chr> <chr>     <chr>      <chr>       <chr> <chr>         <chr>       
#>  1 1003… 1254… LEE       NICOLE     NA          NA    Hospital      370781      
#>  2 1043… 8123… BARLOW    JEFFERY    D           NA    Hospital      370781      
#>  3 1053… 7618… SIZELOVE  AARON      S           NA    Hospital      370781      
#>  4 1184… 2567… GLENN     JAMES      A           NA    Hospital      370781      
#>  5 1225… 2466… HUENERGA… BRENDA     K           NA    Hospital      370781      
#>  6 1245… 4082… MADRID    STEVE      M           NA    Hospital      370781      
#>  7 1245… 8426… BEDNARCZ… STEVEN     NA          NA    Hospital      370781      
#>  8 1316… 3072… FAYED     ISLAM      NA          NA    Hospital      370781      
#>  9 1316… 2769… BEAM      CHESTER    WRAY        NA    Hospital      370781      
#> 10 1427… 7517… MCGARRY   THOMAS     NA          NA    Hospital      370781      
#> 11 1487… 8426… RUSSENBE… NICHOLAS   NA          NA    Hospital      370781      
#> 12 1568… 3870… SCHMIDT   DWAYNE     A           NA    Hospital      370781      
#> 13 1629… 6103… MORRIS    JANA       NIKOLE      NA    Hospital      370781      
#> 14 1710… 2163… COSTELLO  RICHARD    F           JR.   Hospital      370781      
#> 15 1740… 4880… SCRUGGS   DEANNA     LYNN        NA    Hospital      370781      
#> 16 1891… 4981… STUART    DANNA      K           NA    Hospital      370781      
#> 17 1922… 9032… ARMOR     JESS       F           NA    Hospital      370781      
#> 18 1952… 1951… MORGAN    JARED      KEITH       NA    Hospital      370781      
#> # ℹ 1 more variable: ccn_primary <chr>
affiliations(ccn_facility = "331302")
#> # A tibble: 180 × 9
#>    npi   pac   last_name first_name middle_name suff  facility_type ccn_facility
#>  * <chr> <chr> <chr>     <chr>      <chr>       <chr> <chr>         <chr>       
#>  1 1003… 6204… CARTER-K… STACI      L           NA    Hospital      331302      
#>  2 1003… 1759… GREENE    LAURA      A           NA    Hospital      331302      
#>  3 1013… 3375… TRIPLETT  EMILY      NA          NA    Hospital      331302      
#>  4 1023… 6901… WILHELM   LINDSEY    B           NA    Hospital      331302      
#>  5 1043… 7214… FIORINI … VANESSA    NA          NA    Hospital      331302      
#>  6 1073… 0547… MAGGY     BROOKE     A           NA    Hospital      331302      
#>  7 1083… 0042… KRULEWITZ NEIL       A           NA    Hospital      331302      
#>  8 1083… 0840… COBB      CARL       W           NA    Hospital      331302      
#>  9 1083… 2860… ELLIOTT   STEVEN     J           NA    Hospital      331302      
#> 10 1093… 5597… MAHONEY   ANDREW     C           NA    Hospital      331302      
#> # ℹ 170 more rows
#> # ℹ 1 more variable: ccn_primary <chr>
affiliations(ccn_facility = "33Z302")
#> # A tibble: 3 × 9
#>   npi    pac   last_name first_name middle_name suff  facility_type ccn_facility
#> * <chr>  <chr> <chr>     <chr>      <chr>       <chr> <chr>         <chr>       
#> 1 13969… 8921… HALLORAN  MARY       K           NA    Nursing home  33Z302      
#> 2 15381… 0547… CHON      IL         JUN         NA    Nursing home  33Z302      
#> 3 18018… 3577… DEMURO    ROB        L           NA    Nursing home  33Z302      
#> # ℹ 1 more variable: ccn_primary <chr>
affiliations(npi = "1043245657")
#> # A tibble: 2 × 9
#>   npi    pac   last_name first_name middle_name suff  facility_type ccn_facility
#> * <chr>  <chr> <chr>     <chr>      <chr>       <chr> <chr>         <chr>       
#> 1 10432… 7810… FUNG      MARK       K           NA    Hospital      470003      
#> 2 10432… 7810… FUNG      MARK       K           NA    Hospital      330250      
#> # ℹ 1 more variable: ccn_primary <chr>