Build a Query for an Endpoint

Build a Query for an Endpoint

Usage

build(obj, qry)

Arguments

obj

An <endpoint>, <collection> or <group> object.

qry

A <query> object.

Value

A list of query parameters matched to an endpoint's fields.

Examples

qry <- new_query(
   first_name            = starts_with("Andr"),
   middle_name           = ends_with("e"),
   last_name             = contains("J"),
   state_cd              = any_of(c("CA", "GA", "NY")),
   state                 = none_of(c("GA", "NY")),
   city                  = all_but(c("Atlanta", "Los Angeles")),
   npi                   = npi_ex$k,
   covered_recipient_npi = npi_ex$k[1:5],
   ccn                   = "105101256",
   rate                  = between(0.45, 0.67),
   year                  = 2021:2025)

build(endpoint("enroll_prov"), qry)
#> $params
#> $params$NPI
#> [1] "filter[1][condition][path]=NPI&filter[1][condition][operator]==&filter[1][condition][value][1]=1225701881&filter[1][condition][value][2]=1174270805&filter[1][condition][value][3]=1235702796&filter[1][condition][value][4]=1962116806&filter[1][condition][value][5]=1013647569&filter[1][condition][value][6]=1306500665&filter[1][condition][value][7]=1982296737&filter[1][condition][value][8]=1083295638&filter[1][condition][value][9]=1841967825&filter[1][condition][value][10]=1891390084&filter[1][condition][value][11]=1275117269&filter[1][condition][value][12]=1992338701&filter[1][condition][value][13]=1891355863&filter[1][condition][value][14]=1548743511&filter[1][condition][value][15]=1023473279&filter[1][condition][value][16]=1861857013&filter[1][condition][value][17]=1689182859&filter[1][condition][value][18]=1982059275"
#> 
#> $params$STATE_CD
#> [1] "filter[2][condition][path]=STATE_CD&filter[2][condition][operator]=IN&filter[2][condition][value][1]=CA&filter[2][condition][value][2]=GA&filter[2][condition][value][3]=NY"
#> 
#> $params$FIRST_NAME
#> [1] "filter[3][condition][path]=FIRST_NAME&filter[3][condition][operator]=STARTS_WITH&filter[3][condition][value]=Andr"
#> 
#> $params$LAST_NAME
#> [1] "filter[4][condition][path]=LAST_NAME&filter[4][condition][operator]=CONTAINS&filter[4][condition][value]=J"
#> 
#> 
#> $identifier
#> [1] "https://data.cms.gov/data-api/v1/dataset/2457ea29-fc82-48b0-86ec-3b0755de7515/data-viewer?offset=0&size=1"
#> 

build(endpoint("dial_facility"), qry)
#> $params
#> $params$state
#> [1] "filter[1][condition][path]=state&filter[1][condition][operator]=NOT+IN&filter[1][condition][value][1]=GA&filter[1][condition][value][2]=NY"
#> 
#> $params$CCN
#> [1] "filter[2][condition][path]=CCN&filter[2][condition][operator]==&filter[2][condition][value]=105101256"
#> 
#> $params$city
#> [1] "filter[3][condition][path]=city&filter[3][condition][operator]=<>&filter[3][condition][value][1]=Atlanta&filter[3][condition][value][2]=Los Angeles"
#> 
#> $params$NPI
#> [1] "filter[4][condition][path]=NPI&filter[4][condition][operator]==&filter[4][condition][value][1]=1225701881&filter[4][condition][value][2]=1174270805&filter[4][condition][value][3]=1235702796&filter[4][condition][value][4]=1962116806&filter[4][condition][value][5]=1013647569&filter[4][condition][value][6]=1306500665&filter[4][condition][value][7]=1982296737&filter[4][condition][value][8]=1083295638&filter[4][condition][value][9]=1841967825&filter[4][condition][value][10]=1891390084&filter[4][condition][value][11]=1275117269&filter[4][condition][value][12]=1992338701&filter[4][condition][value][13]=1891355863&filter[4][condition][value][14]=1548743511&filter[4][condition][value][15]=1023473279&filter[4][condition][value][16]=1861857013&filter[4][condition][value][17]=1689182859&filter[4][condition][value][18]=1982059275"
#> 
#> 
#> $identifier
#>                                                                                                 2025 
#> "https://data.cms.gov/data-api/v1/dataset/7d4e7bd1-f08a-4fca-8085-6af24026a19a/data?offset=0&size=1" 
#>                                                                                                 2024 
#> "https://data.cms.gov/data-api/v1/dataset/62d125ff-c959-4c5c-8a46-8cacba50bfd4/data?offset=0&size=1" 
#>                                                                                                 2023 
#> "https://data.cms.gov/data-api/v1/dataset/200e7e7e-a5a6-4450-ba18-ab53d237cc45/data?offset=0&size=1" 
#>                                                                                                 2022 
#> "https://data.cms.gov/data-api/v1/dataset/ebc64ad1-164c-455a-80ca-cb0753e44a46/data?offset=0&size=1" 
#>                                                                                                 2021 
#> "https://data.cms.gov/data-api/v1/dataset/e66eaf03-398f-4a8c-9d1b-6aaf25fb6425/data?offset=0&size=1" 
#> 

build(endpoint("quality_payment"), qry)
#> $params
#> $params$npi
#> [1] "filter[1][condition][path]=npi&filter[1][condition][operator]==&filter[1][condition][value][1]=1225701881&filter[1][condition][value][2]=1174270805&filter[1][condition][value][3]=1235702796&filter[1][condition][value][4]=1962116806&filter[1][condition][value][5]=1013647569&filter[1][condition][value][6]=1306500665&filter[1][condition][value][7]=1982296737&filter[1][condition][value][8]=1083295638&filter[1][condition][value][9]=1841967825&filter[1][condition][value][10]=1891390084&filter[1][condition][value][11]=1275117269&filter[1][condition][value][12]=1992338701&filter[1][condition][value][13]=1891355863&filter[1][condition][value][14]=1548743511&filter[1][condition][value][15]=1023473279&filter[1][condition][value][16]=1861857013&filter[1][condition][value][17]=1689182859&filter[1][condition][value][18]=1982059275"
#> 
#> 
#> $identifier
#>                                                                                                 2023 
#> "https://data.cms.gov/data-api/v1/dataset/3545dc8e-7037-41ee-8a1d-9f3198fe4927/data?offset=0&size=1" 
#>                                                                                                 2022 
#> "https://data.cms.gov/data-api/v1/dataset/b3438273-b4a6-44ca-8fb2-9e6026b74642/data?offset=0&size=1" 
#>                                                                                                 2021 
#> "https://data.cms.gov/data-api/v1/dataset/8c2ab3e4-597e-43c9-9d94-7223c66d7f25/data?offset=0&size=1" 
#> 

build(endpoint("pdc_clinician"), qry)
#> $params
#> $params$npi
#> [1] "conditions[0][property]=npi&conditions[0][operator]==&conditions[0][value][1]=1225701881&conditions[0][value][2]=1174270805&conditions[0][value][3]=1235702796&conditions[0][value][4]=1962116806&conditions[0][value][5]=1013647569&conditions[0][value][6]=1306500665&conditions[0][value][7]=1982296737&conditions[0][value][8]=1083295638&conditions[0][value][9]=1841967825&conditions[0][value][10]=1891390084&conditions[0][value][11]=1275117269&conditions[0][value][12]=1992338701&conditions[0][value][13]=1891355863&conditions[0][value][14]=1548743511&conditions[0][value][15]=1023473279&conditions[0][value][16]=1861857013&conditions[0][value][17]=1689182859&conditions[0][value][18]=1982059275"
#> 
#> $params$state
#> [1] "conditions[1][property]=state&conditions[1][operator]=not+in&conditions[1][value][1]=GA&conditions[1][value][2]=NY"
#> 
#> 
#> $identifier
#> [1] "https://data.cms.gov/provider-data/api/1/datastore/query/mj5m-pzi6/0?count=true&results=true&offset=0&limit=1"
#> 

build(endpoint("grp_cover_nature"), qry)
#> $params
#> $params$covered_recipient_npi
#> [1] "conditions[0][property]=covered_recipient_npi&conditions[0][operator]==&conditions[0][value][1]=1225701881&conditions[0][value][2]=1174270805&conditions[0][value][3]=1235702796&conditions[0][value][4]=1962116806&conditions[0][value][5]=1013647569"
#> 
#> 
#> $identifier
#>                                                                                                                                     2024 
#> "https://openpaymentsdata.cms.gov/api/1/datastore/query/88822473-9093-46fc-ad38-81330eb8de4b/0?count=true&results=true&offset=0&limit=1" 
#>                                                                                                                                     2023 
#> "https://openpaymentsdata.cms.gov/api/1/datastore/query/4017acfa-0eeb-4b8a-b68c-3b41e779c4a2/0?count=true&results=true&offset=0&limit=1" 
#>                                                                                                                                     2022 
#> "https://openpaymentsdata.cms.gov/api/1/datastore/query/686b765e-f4b7-4d35-a050-fb5af926dbbe/0?count=true&results=true&offset=0&limit=1" 
#>                                                                                                                                     2021 
#> "https://openpaymentsdata.cms.gov/api/1/datastore/query/d35243be-55cb-5087-931e-a294e0e0917f/0?count=true&results=true&offset=0&limit=1" 
#> 

build(collection("hha"), qry)
#> ! No fields matched in Home Health Agency All Owners.
#> $hha_owner
#> $hha_owner$params
#> NULL
#> 
#> $hha_owner$identifier
#> [1] "https://data.cms.gov/data-api/v1/dataset/fc009b2d-7846-44b1-b4a1-692f0c143879/data-viewer?offset=0&size=1"
#> 
#> 
#> $hha_enroll
#> $hha_enroll$params
#> $hha_enroll$params$NPI
#> [1] "filter[1][condition][path]=NPI&filter[1][condition][operator]==&filter[1][condition][value][1]=1225701881&filter[1][condition][value][2]=1174270805&filter[1][condition][value][3]=1235702796&filter[1][condition][value][4]=1962116806&filter[1][condition][value][5]=1013647569&filter[1][condition][value][6]=1306500665&filter[1][condition][value][7]=1982296737&filter[1][condition][value][8]=1083295638&filter[1][condition][value][9]=1841967825&filter[1][condition][value][10]=1891390084&filter[1][condition][value][11]=1275117269&filter[1][condition][value][12]=1992338701&filter[1][condition][value][13]=1891355863&filter[1][condition][value][14]=1548743511&filter[1][condition][value][15]=1023473279&filter[1][condition][value][16]=1861857013&filter[1][condition][value][17]=1689182859&filter[1][condition][value][18]=1982059275"
#> 
#> $hha_enroll$params$CCN
#> [1] "filter[2][condition][path]=CCN&filter[2][condition][operator]==&filter[2][condition][value]=105101256"
#> 
#> $hha_enroll$params$CITY
#> [1] "filter[3][condition][path]=CITY&filter[3][condition][operator]=<>&filter[3][condition][value][1]=Atlanta&filter[3][condition][value][2]=Los Angeles"
#> 
#> $hha_enroll$params$STATE
#> [1] "filter[4][condition][path]=STATE&filter[4][condition][operator]=NOT+IN&filter[4][condition][value][1]=GA&filter[4][condition][value][2]=NY"
#> 
#> 
#> $hha_enroll$identifier
#> [1] "https://data.cms.gov/data-api/v1/dataset/15f64ab4-3172-4a27-b589-ebd67a6d28aa/data-viewer?offset=0&size=1"
#> 
#> 
#> $hha_cost
#> $hha_cost$params
#> $hha_cost$params$City
#> [1] "filter[1][condition][path]=City&filter[1][condition][operator]=<>&filter[1][condition][value][1]=Atlanta&filter[1][condition][value][2]=Los Angeles"
#> 
#> 
#> $hha_cost$identifier
#>                                                                                                 2022 
#> "https://data.cms.gov/data-api/v1/dataset/1adf48d9-9ff2-4641-ac51-d827aabf791c/data?offset=0&size=1" 
#>                                                                                                 2021 
#> "https://data.cms.gov/data-api/v1/dataset/a02a324e-a8ad-4933-8bcd-e6a10ab28c8e/data?offset=0&size=1" 
#> 
#> 

build(collection("hhc"), qry)
#> ! No fields matched in Home Health Care - Measure Date Range.
#> ! No fields matched in Home Health Care - National Data.
#> $hhc_range
#> $hhc_range$params
#> NULL
#> 
#> $hhc_range$identifier
#> [1] "https://data.cms.gov/provider-data/api/1/datastore/query/c886-nwpj/0?count=true&results=true&offset=0&limit=1"
#> 
#> 
#> $hhc_nation
#> $hhc_nation$params
#> NULL
#> 
#> $hhc_nation$identifier
#> [1] "https://data.cms.gov/provider-data/api/1/datastore/query/97z8-de96/0?count=true&results=true&offset=0&limit=1"
#> 
#> 
#> $hhc_state
#> $hhc_state$params
#> $hhc_state$params$state
#> [1] "conditions[0][property]=state&conditions[0][operator]=not+in&conditions[0][value][1]=GA&conditions[0][value][2]=NY"
#> 
#> 
#> $hhc_state$identifier
#> [1] "https://data.cms.gov/provider-data/api/1/datastore/query/tee5-ixt5/0?count=true&results=true&offset=0&limit=1"
#> 
#> 
#> $hhc_zip
#> $hhc_zip$params
#> $hhc_zip$params$state
#> [1] "conditions[0][property]=state&conditions[0][operator]=not+in&conditions[0][value][1]=GA&conditions[0][value][2]=NY"
#> 
#> 
#> $hhc_zip$identifier
#> [1] "https://data.cms.gov/provider-data/api/1/datastore/query/m5eg-upu5/0?count=true&results=true&offset=0&limit=1"
#> 
#> 
#> $hhc_agency
#> $hhc_agency$params
#> $hhc_agency$params$state
#> [1] "conditions[0][property]=state&conditions[0][operator]=not+in&conditions[0][value][1]=GA&conditions[0][value][2]=NY"
#> 
#> 
#> $hhc_agency$identifier
#> [1] "https://data.cms.gov/provider-data/api/1/datastore/query/6jpm-sxkc/0?count=true&results=true&offset=0&limit=1"
#> 
#>