Medicare Endpoint

Medicare Endpoint

Usage

careMain(alias)

Arguments

alias

<chr> endpoint alias

Value

An S7 <careMain> object.

Examples

careMain("contact")
#> <careMain>
#>  @ title     : chr "Public Reporting of Missing Digital Contact Information"
#>  @ metadata  : <careMeta>
#>  .. @ description: chr "In the May 2020 CMS Interoperability and Patient Access final rule, CMS finalized the policy to publicly report"| __truncated__
#>  .. @ temporal   : chr "2023-01-01 • 2024-09-30"
#>  .. @ periodicity: chr "Quarterly [R/P3M]"
#>  .. @ modified   : Date[1:1], format: "2024-10-25"
#>  .. @ dictionary : chr "https://data.cms.gov/resources/public-reporting-of-missing-digital-contact-information-data-dictionary"
#>  .. @ site       : chr "https://data.cms.gov/provider-compliance/public-reporting-of-missing-digital-contact-information"
#>  .. @ references : chr "https://data.cms.gov/sites/default/files/2021-12/8eb2b4bf-6e5f-4e05-bcdb-39c07ad8f77a/Missing_Digital_Contact_I"| __truncated__
#>  .. @ download   : chr "https://data.cms.gov/sites/default/files/2024-10/fb284594-a4ff-442f-8b66-f6838c4355ce/Providers%20without%20End"| __truncated__
#>  @ dimensions: <careDim>
#>  .. @ rows  : int 3505696
#>  .. @ pages : int 702
#>  .. @ fields: chr [1:2] "NPI" "Provider Name"
#>  @ identifier: chr "https://data.cms.gov/data-api/v1/dataset/63a83bb1-4c02-43b3-8ef4-e3d3c6cf62fa/data-viewer"
#>  @ resources : chr "https://data.cms.gov/data-api/v1/dataset-resources/63a83bb1-4c02-43b3-8ef4-e3d3c6cf62fa"
careMain("crosswalk")
#> <careMain>
#>  @ title     : chr "Medicare Provider and Supplier Taxonomy Crosswalk"
#>  @ metadata  : <careMeta>
#>  .. @ description: chr "The Medicare Provider and Supplier Taxonomy Crosswalk dataset lists the providers and suppliers eligible to enr"| __truncated__
#>  .. @ temporal   : chr "2021-07-01 • 2025-06-30"
#>  .. @ periodicity: chr "Biannually [R/P6M]"
#>  .. @ modified   : Date[1:1], format: "2025-04-22"
#>  .. @ dictionary : chr "https://data.cms.gov/resources/medicare-provider-and-supplier-taxonomy-crosswalk-data-dictionary"
#>  .. @ site       : chr "https://data.cms.gov/provider-characteristics/medicare-provider-supplier-enrollment/medicare-provider-and-suppl"| __truncated__
#>  .. @ references : chr "https://data.cms.gov/resources/medicare-provider-and-supplier-taxonomy-crosswalk-methodology"
#>  .. @ download   : chr "https://data.cms.gov/sites/default/files/2025-04/d7c24ae8-908c-4174-8470-c31c1c55e2a8/Medicare_Provider_and_Sup"| __truncated__
#>  @ dimensions: <careDim>
#>  .. @ rows  : int 561
#>  .. @ pages : int 1
#>  .. @ fields: chr [1:4] "MEDICARE SPECIALTY CODE" ...
#>  @ identifier: chr "https://data.cms.gov/data-api/v1/dataset/113eb0bc-0c9a-4d91-9f93-3f6b28c0bf6b/data-viewer"
#>  @ resources : chr "https://data.cms.gov/data-api/v1/dataset-resources/113eb0bc-0c9a-4d91-9f93-3f6b28c0bf6b"
careMain("dialysis")
#> <careMain>
#>  @ title     : chr "Medicare Dialysis Facilities"
#>  @ metadata  : <careMeta>
#>  .. @ description: chr "The Medicare Dialysis Facilities data provides information on clinical and patient measures for Medicare-certif"| __truncated__
#>  .. @ temporal   : chr "2009-01-01 • 2024-12-31"
#>  .. @ periodicity: chr "Annually [R/P1Y]"
#>  .. @ modified   : Date[1:1], format: "2024-02-29"
#>  .. @ dictionary : chr "https://data.cms.gov/sites/default/files/2024-02/FY2024%20Data%20Dictionary.xlsx"
#>  .. @ site       : chr "https://data.cms.gov/quality-of-care/medicare-dialysis-facilities"
#>  .. @ references : chr "https://data.cms.gov/sites/default/files/2024-02/FY2024_DFR_Guide%20and%20Methodology.pdf"
#>  .. @ download   : chr "https://data.cms.gov/sites/default/files/2024-02/b2c2658d-e19d-48db-8b3b-9d343a4d395f/dfr_facility_socrata_fy2024.csv"
#>  @ dimensions: <careDim>
#>  .. @ rows  : int 11679243
#>  .. @ pages : int 2336
#>  .. @ fields: chr [1:14] "state" "CCN" "Provider_Name" "city" "Ownership_Type" ...
#>  @ identifier: chr "https://data.cms.gov/data-api/v1/dataset/f8610e87-ba25-43a3-a49e-927dbc8701ae/data-viewer"
#>  @ resources : chr "https://data.cms.gov/data-api/v1/dataset-resources/f8610e87-ba25-43a3-a49e-927dbc8701ae"
careMain("enrollees")
#> <careMain>
#>  @ title     : chr "Medicare Fee-For-Service Public Provider Enrollment"
#>  @ metadata  : <careMeta>
#>  .. @ description: chr "The Medicare Fee-For-Service Public Provider Enrollment dataset includes information on providers who are activ"| __truncated__
#>  .. @ temporal   : chr "2023-01-01 • 2025-03-31"
#>  .. @ periodicity: chr "Quarterly [R/P3M]"
#>  .. @ modified   : Date[1:1], format: "2025-04-15"
#>  .. @ dictionary : chr "https://data.cms.gov/resources/medicare-fee-for-service-public-provider-enrollment-data-dictionary"
#>  .. @ site       : chr "https://data.cms.gov/provider-characteristics/medicare-provider-supplier-enrollment/medicare-fee-for-service-pu"| __truncated__
#>  .. @ references : chr "https://data.cms.gov/resources/fee-for-service-public-provider-enrollment-methodology"
#>  .. @ download   : chr "https://data.cms.gov/sites/default/files/2025-04/5d10bc83-29bd-498e-a1ff-3e6e2065a7e3/PPEF_Enrollment_Extract_2025.04.01.csv"
#>  @ dimensions: <careDim>
#>  .. @ rows  : int 2922752
#>  .. @ pages : int 585
#>  .. @ fields: chr [1:10] "NPI" "PECOS_ASCT_CNTL_ID" "ENRLMT_ID" ...
#>  @ identifier: chr "https://data.cms.gov/data-api/v1/dataset/2457ea29-fc82-48b0-86ec-3b0755de7515/data-viewer"
#>  @ resources : chr "https://data.cms.gov/data-api/v1/dataset-resources/2457ea29-fc82-48b0-86ec-3b0755de7515"
careMain("facilities")
#> <careMain>
#>  @ title     : chr "Provider of Services File - Hospital & Non-Hospital Facilities"
#>  @ metadata  : <careMeta>
#>  .. @ description: chr "Please be advised that as of Q4 2023 there is a new Provider of Service file (POS) that contains the provider a"| __truncated__
#>  .. @ temporal   : chr "2011-10-01 • 2025-03-31"
#>  .. @ periodicity: chr "Quarterly [R/P3M]"
#>  .. @ modified   : Date[1:1], format: "2025-04-25"
#>  .. @ dictionary : chr "https://data.cms.gov/sites/default/files/2023-07/0ca58d5d-7914-4532-b22d-41741d3e6151/P.QWB.POSQ.OTHER.LAYOUT.MAR23.pdf"
#>  .. @ site       : chr "https://data.cms.gov/provider-characteristics/hospitals-and-other-facilities/provider-of-services-file-hospital"| __truncated__
#>  .. @ references : chr "https://data.cms.gov/resources/provider-of-services-file-methodology"
#>  .. @ download   : chr "https://data.cms.gov/sites/default/files/2025-04/52ac594d-c42d-4377-9c03-309d6a783dc4/PQWB.POSQ.CLIA.DATA.MAR25.csv"
#>  @ dimensions: <careDim>
#>  .. @ rows  : int 116047
#>  .. @ pages : int 24
#>  .. @ fields: chr [1:473] "PRVDR_CTGRY_SBTYP_CD" "PRVDR_CTGRY_CD" "CHOW_CNT" ...
#>  @ identifier: chr "https://data.cms.gov/data-api/v1/dataset/8ba0f9b4-9493-4aa0-9f82-44ea9468d1b5/data-viewer"
#>  @ resources : chr "https://data.cms.gov/data-api/v1/dataset-resources/8ba0f9b4-9493-4aa0-9f82-44ea9468d1b5"
careMain("IQIES")
#> <careMain>
#>  @ title     : chr "Provider of Services File - Internet Quality Improvement and Evaluation System - Home Health Agency, Ambulatory"| __truncated__
#>  @ metadata  : <careMeta>
#>  .. @ description: chr "The Provider of Services File (POS) - Internet Quality Improvement and Evaluation System (iQIES) - Home Health "| __truncated__
#>  .. @ temporal   : chr "2023-10-01 • 2025-03-31"
#>  .. @ periodicity: chr "Quarterly [R/P3M]"
#>  .. @ modified   : Date[1:1], format: "2025-04-15"
#>  .. @ dictionary : chr "https://data.cms.gov/sites/default/files/2025-01/47cff7a0-5575-4a62-be85-ae72dcedf677/iQIES%20POS%20Data%20Dictionary.zip"
#>  .. @ site       : chr "https://data.cms.gov/provider-characteristics/hospitals-and-other-facilities/provider-of-services-file-internet"| __truncated__
#>  .. @ references : chr "https://data.cms.gov/resources/pos-file-iqies-for-hha-asc-and-hospice-providers-methodology"
#>  .. @ download   : chr "https://data.cms.gov/sites/default/files/2025-04/6aeb9473-03a7-4db9-9b71-057d60cbcada/iQIES%20HHA_HOSPICE_ASC_DATA.MAR25.csv"
#>  @ dimensions: <careDim>
#>  .. @ rows  : int 39912
#>  .. @ pages : int 8
#>  .. @ fields: chr [1:103] "prvdr_num" "fac_name" "prvdr_sbtyp_id" ...
#>  @ identifier: chr "https://data.cms.gov/data-api/v1/dataset/086e48c4-87a6-4be1-8823-29e8da8f225b/data-viewer"
#>  @ resources : chr "https://data.cms.gov/data-api/v1/dataset-resources/086e48c4-87a6-4be1-8823-29e8da8f225b"
careMain("laboratories")
#> <careMain>
#>  @ title     : chr "Provider of Services File - Clinical Laboratories"
#>  @ metadata  : <careMeta>
#>  .. @ description: chr "The Provider of Services (POS) Clinical Laboratories (CLIA) data provides information on CLIA demographics and "| __truncated__
#>  .. @ temporal   : chr "2011-10-01 • 2025-03-31"
#>  .. @ periodicity: chr "Quarterly [R/P3M]"
#>  .. @ modified   : Date[1:1], format: "2025-04-15"
#>  .. @ dictionary : chr "https://data.cms.gov/sites/default/files/2023-07/50eacc07-7930-4e62-97e3-e04a2dc1e2a3/P.QWB.POSQ.CLIA.LAYOUT.MAR23.pdf"
#>  .. @ site       : chr "https://data.cms.gov/provider-characteristics/hospitals-and-other-facilities/provider-of-services-file-clinical-laboratories"
#>  .. @ references : chr "https://data.cms.gov/resources/provider-of-services-file-methodology"
#>  .. @ download   : chr "https://data.cms.gov/sites/default/files/2025-04/d0fc9f7a-b8d2-4988-a6e9-eb4cff2d45c3/PQWB.POSQ.CLIA.DATA.MAR25.csv"
#>  @ dimensions: <careDim>
#>  .. @ rows  : int 657403
#>  .. @ pages : int 132
#>  .. @ fields: chr [1:103] "PRVDR_CTGRY_SBTYP_CD" "PRVDR_CTGRY_CD" "CHOW_CNT" ...
#>  @ identifier: chr "https://data.cms.gov/data-api/v1/dataset/d3eb38ac-d8e9-40d3-b7b7-6205d3d1dc16/data-viewer"
#>  @ resources : chr "https://data.cms.gov/data-api/v1/dataset-resources/d3eb38ac-d8e9-40d3-b7b7-6205d3d1dc16"
careMain("long_term")
#> <careMain>
#>  @ title     : chr "Long-Term Care Facility Characteristics"
#>  @ metadata  : <careMeta>
#>  .. @ description: chr "The Long-Term Care Facility Characteristics, CMS Form 671, dataset provides information submitted by nursing ho"| __truncated__
#>  .. @ temporal   : chr "2023-10-01 • 2024-12-31"
#>  .. @ periodicity: chr "Quarterly [R/P3M]"
#>  .. @ modified   : Date[1:1], format: "2025-04-02"
#>  .. @ dictionary : chr "https://data.cms.gov/resources/long-term-care-facility-characteristics-data-dictionary"
#>  .. @ site       : chr "https://data.cms.gov/quality-of-care/long-term-care-facility-characteristics"
#>  .. @ references : chr "https://data.cms.gov/resources/long-term-care-facility-characteristics-methodology"
#>  .. @ download   : chr "https://data.cms.gov/sites/default/files/2025-01/d51e2acf-28d6-49cf-9be0-a306c1b17837/Form_671_Q4_2024.csv"
#>  @ dimensions: <careDim>
#>  .. @ rows  : int 14799
#>  .. @ pages : int 3
#>  .. @ fields: chr [1:33] "Provider Number" "Facility Name" "City" "State" ...
#>  @ identifier: chr "https://data.cms.gov/data-api/v1/dataset/129a6503-c0f1-4132-b186-4c0232c2d894/data-viewer"
#>  @ resources : chr "https://data.cms.gov/data-api/v1/dataset-resources/129a6503-c0f1-4132-b186-4c0232c2d894"
careMain("opt_out")
#> <careMain>
#>  @ title     : chr "Opt Out Affidavits"
#>  @ metadata  : <careMeta>
#>  .. @ description: chr "The Opt Out Affidavits dataset provides information on providers who have decided not to participate in Medicar"| __truncated__
#>  .. @ temporal   : chr "2023-04-01 • 2025-03-31"
#>  .. @ periodicity: chr "Monthly [R/P1M]"
#>  .. @ modified   : Date[1:1], format: "2025-04-24"
#>  .. @ dictionary : chr "https://data.cms.gov/resources/opt-out-affidavits-data-dictionary"
#>  .. @ site       : chr "https://data.cms.gov/provider-characteristics/medicare-provider-supplier-enrollment/opt-out-affidavits"
#>  .. @ references : chr "https://data.cms.gov/resources/opt-out-affidavits-methodology"
#>  .. @ download   : chr "https://data.cms.gov/sites/default/files/2025-04/fcfeb052-e800-4934-844d-0bd2e0b48937/Opt_Out_List_March.csv"
#>  @ dimensions: <careDim>
#>  .. @ rows  : int 50473
#>  .. @ pages : int 11
#>  .. @ fields: chr [1:13] "First Name" "Last Name" "NPI" "Specialty" ...
#>  @ identifier: chr "https://data.cms.gov/data-api/v1/dataset/9887a515-7552-4693-bf58-735c77af46d7/data-viewer"
#>  @ resources : chr "https://data.cms.gov/data-api/v1/dataset-resources/9887a515-7552-4693-bf58-735c77af46d7"
careMain("order_refer")
#> <careMain>
#>  @ title     : chr "Order and Referring"
#>  @ metadata  : <careMeta>
#>  .. @ description: chr "The Order and Referring dataset provides information on all physicians and non-physician practitioners, by thei"| __truncated__
#>  .. @ temporal   : chr "2023-05-21 • 2025-04-19"
#>  .. @ periodicity: chr "Twice a Week [R/P3.5D]"
#>  .. @ modified   : Date[1:1], format: "2025-04-25"
#>  .. @ dictionary : chr "https://data.cms.gov/resources/order-and-referring-data-dictionary"
#>  .. @ site       : chr "https://data.cms.gov/provider-characteristics/medicare-provider-supplier-enrollment/order-and-referring"
#>  .. @ references : chr "https://data.cms.gov/resources/order-and-referring-methodology"
#>  .. @ download   : chr "https://data.cms.gov/sites/default/files/2025-04/d3b6f85a-5985-468b-bf04-4222abcb504f/OrderReferring_20250424.csv"
#>  @ dimensions: <careDim>
#>  .. @ rows  : int 1944844
#>  .. @ pages : int 389
#>  .. @ fields: chr [1:8] "NPI" "LAST_NAME" "FIRST_NAME" "PARTB" "DME" "HHA" ...
#>  @ identifier: chr "https://data.cms.gov/data-api/v1/dataset/c99b5865-1119-4436-bb80-c5af2773ea1f/data-viewer"
#>  @ resources : chr "https://data.cms.gov/data-api/v1/dataset-resources/c99b5865-1119-4436-bb80-c5af2773ea1f"
careMain("RBCS")
#> <careMain>
#>  @ title     : chr "Restructured BETOS Classification System"
#>  @ metadata  : <careMeta>
#>  .. @ description: chr "The Restructured BETOS Classification System (RBCS) dataset is a taxonomy that allows researchers to group heal"| __truncated__
#>  .. @ temporal   : chr "2024-01-01 • 2024-12-31"
#>  .. @ periodicity: chr "Annually [R/P1Y]"
#>  .. @ modified   : Date[1:1], format: "2024-10-22"
#>  .. @ dictionary : chr "https://data.cms.gov/resources/rbcs-data-dictionary-2024"
#>  .. @ site       : chr "https://data.cms.gov/provider-summary-by-type-of-service/provider-service-classifications/restructured-betos-cl"| __truncated__
#>  .. @ references : chr "https://data.cms.gov/sites/default/files/2024-10/2024%20RBCS%20Final%20Report.pdf"
#>  .. @ download   : chr "https://data.cms.gov/sites/default/files/2024-10/30b74146-c5ce-47e0-8ae2-ad1204c81906/2024%20RBCS%20Taxonomy_CSV.csv"
#>  @ dimensions: <careDim>
#>  .. @ rows  : int 18077
#>  .. @ pages : int 4
#>  .. @ fields: chr [1:14] "HCPCS_Cd" "RBCS_ID" "RBCS_Cat" "RBCS_Cat_Desc" ...
#>  @ identifier: chr "https://data.cms.gov/data-api/v1/dataset/e3db6e56-149f-49ce-b374-40aecda2357b/data-viewer"
#>  @ resources : chr "https://data.cms.gov/data-api/v1/dataset-resources/e3db6e56-149f-49ce-b374-40aecda2357b"
careMain("transparency")
#> <careMain>
#>  @ title     : chr "Hospital Price Transparency Enforcement Activities and Outcomes"
#>  @ metadata  : <careMeta>
#>  .. @ description: chr "The Hospital Price Transparency Enforcement Activities and Outcomes dataset contains information related to enf"| __truncated__
#>  .. @ temporal   : chr "2023-10-01 • 2024-12-31"
#>  .. @ periodicity: chr "Quarterly [R/P3M]"
#>  .. @ modified   : Date[1:1], format: "2025-02-06"
#>  .. @ dictionary : chr "https://data.cms.gov/resources/hospital-price-transparency-enforcement-activities-and-outcomes-data-dictionary"
#>  .. @ site       : chr "https://data.cms.gov/provider-characteristics/hospitals-and-other-facilities/hospital-price-transparency-enforc"| __truncated__
#>  .. @ references : chr "https://data.cms.gov/resources/hospital-price-transparency-enforcement-activities-and-outcomes-methodology"
#>  .. @ download   : chr "https://data.cms.gov/sites/default/files/2025-01/3febdccd-aae4-4d5b-be37-7fdcb10b67ea/Hospital_Price_Transparen"| __truncated__
#>  @ dimensions: <careDim>
#>  .. @ rows  : int 5290
#>  .. @ pages : int 2
#>  .. @ fields: chr [1:7] "Case_ID" "Hosp_Name" "Hosp_Address" "City" "State" ...
#>  @ identifier: chr "https://data.cms.gov/data-api/v1/dataset/6a3aa708-3c9d-411a-a1a4-e046d3ade7ef/data-viewer"
#>  @ resources : chr "https://data.cms.gov/data-api/v1/dataset-resources/6a3aa708-3c9d-411a-a1a4-e046d3ade7ef"