Medicare Endpoint

Medicare Endpoint

Usage

careMain(alias)

Arguments

alias

<chr> endpoint alias

Value

An S7 <careMain> object.

Examples

careMain("contact")
#> ── Public Reporting of Missing Digital Contact Information ──
#> 
#>   In the May 2020 CMS Interoperability and
#>   Patient Access final rule, CMS finalized
#>   the policy to publicly report the names
#>   and NPIs of those providers who do not
#>   have digital contact information included
#>   in the NPPES system (85 FR 25584). This
#>   data includes the NPI and provider name of
#>   providers and clinicians without digital
#>   contact information in NPPES.
#> 
#> • Rows 3,505,696
#> • Pages 702
#> • Fields 2
#> 
#> • Modified ❯ 2024-10-25
#> • Timespan ❯ 2023-01-01 • 2024-09-30
#> • Periodicity ❯ Quarterly [R/P3M]
careMain("crosswalk")
#> ── Medicare Provider and Supplier Taxonomy Crosswalk ──
#> 
#>   The Medicare Provider and Supplier
#>   Taxonomy Crosswalk dataset lists the
#>   providers and suppliers eligible to enroll
#>   in Medicare programs with the proper
#>   healthcare provider taxonomy code. This
#>   data includes the Medicare speciality
#>   codes, if available, provider/supplier
#>   type description, taxonomy code, and the
#>   taxonomy description.
#> 
#> • Rows 561
#> • Pages 1
#> • Fields 4
#> 
#> • Modified ❯ 2025-01-17
#> • Timespan ❯ 2021-07-01 • 2025-06-30
#> • Periodicity ❯ Biannually [R/P6M]
careMain("dialysis")
#> ── Medicare Dialysis Facilities ──
#> 
#>   The Medicare Dialysis Facilities data
#>   provides information on clinical and
#>   patient measures for
#>   Medicare-certified ESRD facilities, also
#>   known as dialysis facilities. It contains
#>   data on patient characteristics, treatment
#>   patterns, hospitalization, mortality, and
#>   transplantation patterns in
#>   Medicare-certified dialysis facilities.
#> 
#> • Rows 11,679,243
#> • Pages 2,336
#> • Fields 14
#> 
#> • Modified ❯ 2024-02-29
#> • Timespan ❯ 2009-01-01 • 2024-12-31
#> • Periodicity ❯ Annually [R/P1Y]
careMain("enrollees")
#> ── Medicare Fee-For-Service Public Provider Enrollment ──
#> 
#>   The Medicare Fee-For-Service Public
#>   Provider Enrollment dataset includes
#>   information on providers who are actively
#>   approved to bill Medicare or have
#>   completed the 855O at the time the data
#>   was pulled from the Provider Enrollment,
#>   Chain, and Ownership System (PECOS). The
#>   release of this provider enrollment data
#>   is not related to other provider
#>   information releases such as Physician
#>   Compare or Data Transparency.
#> 
#> • Rows 2,945,846
#> • Pages 590
#> • Fields 10
#> 
#> • Modified ❯ 2025-03-18
#> • Timespan ❯ 2023-01-01 • 2024-12-31
#> • Periodicity ❯ Quarterly [R/P3M]
careMain("facilities")
#> ── Provider of Services File - Hospital & Non-Hospital Facilities ──
#> 
#>   Please be advised that as of Q4 2023 there
#>   is a new Provider of Service file (POS)
#>   that contains the provider and
#>   certification details for Home Health
#>   Agencies (HHAs), Hospices, and Ambulatory
#>   Surgical Centers (ASCs). Data contained in
#>   this file are extracted from the Internet
#>   Quality Improvement and Evaluation System
#>   (iQIES) environment and will be updated
#>   quarterly along with the other two POS
#>   files. The Provider of Services File -
#>   Hospital & Non-Hospital Facilities
#>   data provide critical resources for other
#>   federal regulator requirements as well as
#>   supports the ongoing quality & research
#>   efforts sponsored by CMS. In this file you
#>   will find provider certification,
#>   termination, accreditation, ownership,
#>   name, location and other characteristics
#>   organized by CMS Certification Number.
#> 
#> • Rows 115,647
#> • Pages 24
#> • Fields 473
#> 
#> • Modified ❯ 2025-01-31
#> • Timespan ❯ 2011-10-01 • 2024-12-31
#> • Periodicity ❯ Quarterly [R/P3M]
careMain("hospice_acute")
#> ── Medicare Post-Acute Care and Hospice - by Geography & Provider ──
#> 
#>   The Medicare Post-Acute Care and Hospice
#>   Provider Utilization and Payment Public
#>   Use Files (PAC PUF) contains information
#>   on demographic and clinical
#>   characteristics of beneficiaries served,
#>   professional and paraprofessional service
#>   utilization, and payment information at
#>   the provider, state, and national levels
#>   for each PAC setting (i.e. HHA, hospices,
#>   SNF, IRF, and LTCH). There are additional
#>   datasets which can be found as a
#>   downloadable report under ‘Resources,
#>   which include information specific to the
#>   unique variables (e.g., case-mix groups)
#>   for HHAs, SNFs and IRFs. More information
#>   can be found below in the resources
#>   section.
#> 
#> • Rows 89,392
#> • Pages 18
#> • Fields 88
#> 
#> • Modified ❯ 2024-06-14
#> • Timespan ❯ 2022-01-01 • 2022-12-31
#> • Periodicity ❯ Annually [R/P1Y]
careMain("IQIES")
#> ── Provider of Services File - Internet Quality Improvement and Evaluation System - Home Health Agency, Ambulatory Surgical Center, and Hospice Providers ──
#> 
#>   The Provider of Services File (POS) -
#>   Internet Quality Improvement and
#>   Evaluation System (iQIES) - Home Health
#>   Agency (HHA), Ambulatory Surgical Center
#>   (ASC), and Hospice Providers data provides
#>   information on provider demographic and
#>   associated certification information. In
#>   this file you will find provider number
#>   (CMS Certification Number), name, address,
#>   and other characteristics of the
#>   participating institution providers.
#> 
#> • Rows 39,612
#> • Pages 8
#> • Fields 103
#> 
#> • Modified ❯ 2025-01-16
#> • Timespan ❯ 2023-10-01 • 2024-12-31
#> • Periodicity ❯ Quarterly [R/P3M]
careMain("laboratories")
#> ── Provider of Services File - Clinical Laboratories ──
#> 
#>   The Provider of Services (POS) Clinical
#>   Laboratories (CLIA) data provides
#>   information on CLIA demographics and types
#>   of testing services the facility provides.
#>   In this file you will find provider
#>   number, name, address and characteristics
#>   of the participating institution
#>   providers.
#> 
#> • Rows 653,223
#> • Pages 131
#> • Fields 103
#> 
#> • Modified ❯ 2025-01-16
#> • Timespan ❯ 2011-10-01 • 2024-12-31
#> • Periodicity ❯ Quarterly [R/P3M]
careMain("long_term")
#> ── Long-Term Care Facility Characteristics ──
#> 
#>   The Long-Term Care Facility
#>   Characteristics, CMS Form 671, dataset
#>   provides information submitted by nursing
#>   homes on the CMS Form 671 collected during
#>   annual surveys. The data include
#>   information about resident census,
#>   ownership, dedicated special care units,
#>   facility characteristics, and staffing.
#>   Note: Annual surveys are conducted every 9
#>   to 15 months. Additionally, some states
#>   are experiencing delays in conducting
#>   annual surveys, resulting in longer
#>   periods of time since the last annual
#>   survey. As such, some of the data included
#>   in these files may not be up to date.
#> 
#> • Rows 14,799
#> • Pages 3
#> • Fields 33
#> 
#> • Modified ❯ 2025-04-02
#> • Timespan ❯ 2023-10-01 • 2024-12-31
#> • Periodicity ❯ Quarterly [R/P3M]
careMain("opt_out")
#> ── Opt Out Affidavits ──
#> 
#>   The Opt Out Affidavits dataset provides
#>   information on providers who have decided
#>   not to participate in Medicare. It
#>   contains providers NPI, specialty,
#>   address, and effective dates.
#> 
#> • Rows 49,923
#> • Pages 10
#> • Fields 13
#> 
#> • Modified ❯ 2025-03-24
#> • Timespan ❯ 2023-04-01 • 2025-02-28
#> • Periodicity ❯ Monthly [R/P1M]
careMain("order_refer")
#> ── Order and Referring ──
#> 
#>   The Order and Referring dataset provides
#>   information on all physicians and
#>   non-physician practitioners, by their
#>   National Provider Identifier (NPI), who
#>   are of a type/specialty that is legally
#>   eligible to order and refer in the
#>   Medicare program and who have current
#>   enrollment records in Medicare.
#> 
#> • Rows 1,942,928
#> • Pages 389
#> • Fields 8
#> 
#> • Modified ❯ 2025-04-11
#> • Timespan ❯ 2023-05-21 • 2025-04-05
#> • Periodicity ❯ Twice a Week [R/P3.5D]
careMain("rbcs")
#> ── Restructured BETOS Classification System ──
#> 
#>   The Restructured BETOS Classification
#>   System (RBCS) dataset is a taxonomy that
#>   allows researchers to group healthcare
#>   service codes for Medicare Part B services
#>   (i.e., HCPCS codes) into clinically
#>   meaningful categories and subcategories.
#>   It is based on the original
#>   Berenson-Eggers Type of Service (BETOS)
#>   classification created in the 1980s, and
#>   includes notable updates such as Part B
#>   non-physician services. The RBCS will
#>   undergo annual updates by a technical
#>   expert panel of researchers and
#>   clinicians.
#> 
#> • Rows 18,077
#> • Pages 4
#> • Fields 14
#> 
#> • Modified ❯ 2024-10-22
#> • Timespan ❯ 2024-01-01 • 2024-12-31
#> • Periodicity ❯ Annually [R/P1Y]
careMain("transparency")
#> ── Hospital Price Transparency Enforcement Activities and Outcomes ──
#> 
#>   The Hospital Price Transparency
#>   Enforcement Activities and Outcomes
#>   dataset contains information related to
#>   enforcement actions taken by CMS following
#>   a compliance review of a hospitals
#>   obligation to establish, update and make
#>   public a list of the hospital’s standard
#>   charges for items and services provided by
#>   the hospital, in accordance with
#>   regulation (45 CFR 180). This data set
#>   includes the name of each hospital or
#>   hospital location, the hospital or
#>   hospital location address, the outcome or
#>   action following a CMS compliance review
#>   and the date of the outcome or action
#>   taken.
#> 
#> • Rows 5,290
#> • Pages 2
#> • Fields 7
#> 
#> • Modified ❯ 2025-02-06
#> • Timespan ❯ 2023-10-01 • 2024-12-31
#> • Periodicity ❯ Quarterly [R/P3M]