Group of Medicare Temporal Endpoints

Group of Medicare Temporal Endpoints

Usage

careTempGroup(alias)

Arguments

alias

<chr> title alias

Value

An S7 <careTempGroup> object.

Examples

careTempGroup("inpatient")
#> <careTempGroup>
#>  @ title      : chr "Medicare Inpatient Hospitals"
#>  @ periodicity: chr "Annually [R/P1Y]"
#>  @ years      : int [1:10] 2022 2021 2020 2019 2018 2017 2016 2015 2014 2013
#>  @ groups     :List of 3
#>  .. $ by_geography_and_service:List of 7
#>  ..  ..$ description: chr "The Medicare Inpatient Hospitals by Geography and Service dataset provides information on hospital discharges f"| __truncated__
#>  ..  ..$ dictionary : chr "https://data.cms.gov/resources/medicare-inpatient-hospitals-by-geography-and-service-data-dictionary-0"
#>  ..  ..$ site       : chr "https://data.cms.gov/provider-summary-by-type-of-service/medicare-inpatient-hospitals/medicare-inpatient-hospit"| __truncated__
#>  ..  ..$ rows       : int 26260
#>  ..  ..$ pages      : int 6
#>  ..  ..$ fields     : chr [1:9] "Rndrng_Prvdr_Geo_Lvl" "Rndrng_Prvdr_Geo_Cd" "Rndrng_Prvdr_Geo_Desc" "DRG_Cd" ...
#>  ..  ..$ endpoints  : tibble [10 × 3] (S3: tbl_df/tbl/data.frame)
#>  ..  .. ..$ year      : int [1:10] 2022 2021 2020 2019 2018 2017 2016 2015 2014 2013
#>  ..  .. ..$ identifier: chr [1:10] "https://data.cms.gov/data-api/v1/dataset/5c8b315f-a2a3-4b2f-bba6-7be19f8d7edc/data" "https://data.cms.gov/data-api/v1/dataset/68739613-1a4f-40ae-9664-49da07abd81f/data" "https://data.cms.gov/data-api/v1/dataset/e583de38-33d7-4895-a115-0d8af74f0795/data" "https://data.cms.gov/data-api/v1/dataset/337617e5-c18b-4cb9-8152-34851747584b/data" ...
#>  ..  .. ..$ resources : chr [1:10] "https://data.cms.gov/data-api/v1/dataset-resources/5c8b315f-a2a3-4b2f-bba6-7be19f8d7edc" "https://data.cms.gov/data-api/v1/dataset-resources/68739613-1a4f-40ae-9664-49da07abd81f" "https://data.cms.gov/data-api/v1/dataset-resources/e583de38-33d7-4895-a115-0d8af74f0795" "https://data.cms.gov/data-api/v1/dataset-resources/337617e5-c18b-4cb9-8152-34851747584b" ...
#>  .. $ by_provider             :List of 7
#>  ..  ..$ description: chr "The Medicare Inpatient Hospitals by Provider dataset provides information on inpatient discharges for Original "| __truncated__
#>  ..  ..$ dictionary : chr "https://data.cms.gov/resources/medicare-inpatient-hospitals-by-provider-data-dictionary"
#>  ..  ..$ site       : chr "https://data.cms.gov/provider-summary-by-type-of-service/medicare-inpatient-hospitals/medicare-inpatient-hospitals-by-provider"
#>  ..  ..$ rows       : int 3120
#>  ..  ..$ pages      : int 1
#>  ..  ..$ fields     : chr [1:57] "Rndrng_Prvdr_CCN" "Rndrng_Prvdr_Org_Name" "Rndrng_Prvdr_St" "Rndrng_Prvdr_City" ...
#>  ..  ..$ endpoints  : tibble [10 × 3] (S3: tbl_df/tbl/data.frame)
#>  ..  .. ..$ year      : int [1:10] 2022 2021 2020 2019 2018 2017 2016 2015 2014 2013
#>  ..  .. ..$ identifier: chr [1:10] "https://data.cms.gov/data-api/v1/dataset/cc73e77b-7367-48c9-b766-fafefe3f8ffa/data" "https://data.cms.gov/data-api/v1/dataset/c7e1501c-a770-4e2a-beb6-32b3525491a7/data" "https://data.cms.gov/data-api/v1/dataset/d7b2810d-ac10-498f-8794-1c3d9f72ee96/data" "https://data.cms.gov/data-api/v1/dataset/c349419e-3c6a-419e-8c74-0641b64cca12/data" ...
#>  ..  .. ..$ resources : chr [1:10] "https://data.cms.gov/data-api/v1/dataset-resources/cc73e77b-7367-48c9-b766-fafefe3f8ffa" "https://data.cms.gov/data-api/v1/dataset-resources/c7e1501c-a770-4e2a-beb6-32b3525491a7" "https://data.cms.gov/data-api/v1/dataset-resources/d7b2810d-ac10-498f-8794-1c3d9f72ee96" "https://data.cms.gov/data-api/v1/dataset-resources/c349419e-3c6a-419e-8c74-0641b64cca12" ...
#>  .. $ by_provider_and_service :List of 7
#>  ..  ..$ description: chr "The Medicare Inpatient Hospitals by Provider and Service dataset provides information on inpatient discharges f"| __truncated__
#>  ..  ..$ dictionary : chr "https://data.cms.gov/resources/medicare-inpatient-hospitals-by-provider-and-service-data-dictionary-0"
#>  ..  ..$ site       : chr "https://data.cms.gov/provider-summary-by-type-of-service/medicare-inpatient-hospitals/medicare-inpatient-hospit"| __truncated__
#>  ..  ..$ rows       : int 145742
#>  ..  ..$ pages      : int 30
#>  ..  ..$ fields     : chr [1:15] "Rndrng_Prvdr_CCN" "Rndrng_Prvdr_Org_Name" "Rndrng_Prvdr_City" "Rndrng_Prvdr_St" ...
#>  ..  ..$ endpoints  : tibble [10 × 3] (S3: tbl_df/tbl/data.frame)
#>  ..  .. ..$ year      : int [1:10] 2022 2021 2020 2019 2018 2017 2016 2015 2014 2013
#>  ..  .. ..$ identifier: chr [1:10] "https://data.cms.gov/data-api/v1/dataset/46bf50f8-0983-4ca2-b8d5-f2afbbf2e589/data" "https://data.cms.gov/data-api/v1/dataset/117d93f2-ce81-40fe-a4d4-8c03203b95e1/data" "https://data.cms.gov/data-api/v1/dataset/ef5bdbe1-27b4-4296-b320-52bd5d2183d7/data" "https://data.cms.gov/data-api/v1/dataset/6f6d93e1-ecf8-4b93-9845-091faf20f274/data" ...
#>  ..  .. ..$ resources : chr [1:10] "https://data.cms.gov/data-api/v1/dataset-resources/46bf50f8-0983-4ca2-b8d5-f2afbbf2e589" "https://data.cms.gov/data-api/v1/dataset-resources/117d93f2-ce81-40fe-a4d4-8c03203b95e1" "https://data.cms.gov/data-api/v1/dataset-resources/ef5bdbe1-27b4-4296-b320-52bd5d2183d7" "https://data.cms.gov/data-api/v1/dataset-resources/6f6d93e1-ecf8-4b93-9845-091faf20f274" ...
careTempGroup("outpatient")
#> <careTempGroup>
#>  @ title      : chr "Medicare Outpatient Hospitals"
#>  @ periodicity: chr "Annually [R/P1Y]"
#>  @ years      : int [1:8] 2022 2021 2020 2019 2018 2017 2016 2015
#>  @ groups     :List of 2
#>  .. $ by_geography_and_service:List of 7
#>  ..  ..$ description: chr "The Medicare Outpatient Hospitals by Geography and Service dataset provides information on services for Origina"| __truncated__
#>  ..  ..$ dictionary : chr "https://data.cms.gov/resources/medicare-outpatient-hospitals-by-geography-and-service-data-dictionary"
#>  ..  ..$ site       : chr "https://data.cms.gov/provider-summary-by-type-of-service/medicare-outpatient-hospitals/medicare-outpatient-hosp"| __truncated__
#>  ..  ..$ rows       : int 90591
#>  ..  ..$ pages      : int 19
#>  ..  ..$ fields     : chr [1:15] "Rndrng_Prvdr_Geo_Lvl" "Rndrng_Prvdr_Geo_Cd" "Rndrng_Prvdr_Geo_Desc" "Srvc_Lvl" ...
#>  ..  ..$ endpoints  : tibble [8 × 3] (S3: tbl_df/tbl/data.frame)
#>  ..  .. ..$ year      : int [1:8] 2022 2021 2020 2019 2018 2017 2016 2015
#>  ..  .. ..$ identifier: chr [1:8] "https://data.cms.gov/data-api/v1/dataset/6af002ba-b934-4491-96dd-d041989b843c/data" "https://data.cms.gov/data-api/v1/dataset/b3537e90-c3c5-4895-914d-71f54123a0a4/data" "https://data.cms.gov/data-api/v1/dataset/31db4293-f75c-4940-bfae-32d554cd164f/data" "https://data.cms.gov/data-api/v1/dataset/46ae348d-8ba4-43df-868e-74ff87eada39/data" ...
#>  ..  .. ..$ resources : chr [1:8] "https://data.cms.gov/data-api/v1/dataset-resources/6af002ba-b934-4491-96dd-d041989b843c" "https://data.cms.gov/data-api/v1/dataset-resources/b3537e90-c3c5-4895-914d-71f54123a0a4" "https://data.cms.gov/data-api/v1/dataset-resources/31db4293-f75c-4940-bfae-32d554cd164f" "https://data.cms.gov/data-api/v1/dataset-resources/46ae348d-8ba4-43df-868e-74ff87eada39" ...
#>  .. $ by_provider_and_service :List of 7
#>  ..  ..$ description: chr "The Medicare Outpatient Hospitals by Provider and Service dataset provides information on services for Original"| __truncated__
#>  ..  ..$ dictionary : chr "https://data.cms.gov/resources/medicare-outpatient-hospitals-by-provider-and-service-data-dictionary"
#>  ..  ..$ site       : chr "https://data.cms.gov/provider-summary-by-type-of-service/medicare-outpatient-hospitals/medicare-outpatient-hosp"| __truncated__
#>  ..  ..$ rows       : int 116459
#>  ..  ..$ pages      : int 24
#>  ..  ..$ fields     : chr [1:18] "Rndrng_Prvdr_CCN" "Rndrng_Prvdr_Org_Name" "Rndrng_Prvdr_St" "Rndrng_Prvdr_City" ...
#>  ..  ..$ endpoints  : tibble [8 × 3] (S3: tbl_df/tbl/data.frame)
#>  ..  .. ..$ year      : int [1:8] 2022 2021 2020 2019 2018 2017 2016 2015
#>  ..  .. ..$ identifier: chr [1:8] "https://data.cms.gov/data-api/v1/dataset/456dbe68-8878-49dd-889b-bb8bf432fcbe/data" "https://data.cms.gov/data-api/v1/dataset/a7617aef-046f-4557-8a2a-54f9c9a88b57/data" "https://data.cms.gov/data-api/v1/dataset/8281e9f7-1874-4bd9-a183-a74bec2debd3/data" "https://data.cms.gov/data-api/v1/dataset/f3fc4327-f66e-49e1-ba06-f2607a164406/data" ...
#>  ..  .. ..$ resources : chr [1:8] "https://data.cms.gov/data-api/v1/dataset-resources/456dbe68-8878-49dd-889b-bb8bf432fcbe" "https://data.cms.gov/data-api/v1/dataset-resources/a7617aef-046f-4557-8a2a-54f9c9a88b57" "https://data.cms.gov/data-api/v1/dataset-resources/8281e9f7-1874-4bd9-a183-a74bec2debd3" "https://data.cms.gov/data-api/v1/dataset-resources/f3fc4327-f66e-49e1-ba06-f2607a164406" ...
careTempGroup("utilization")
#> <careTempGroup>
#>  @ title      : chr "Medicare Physician & Other Practitioners"
#>  @ periodicity: chr "Annually [R/P1Y]"
#>  @ years      : int [1:10] 2022 2021 2020 2019 2018 2017 2016 2015 2014 2013
#>  @ groups     :List of 3
#>  .. $ by_geography_and_service:List of 7
#>  ..  ..$ description: chr "The Medicare Physician & Other Practitioners by Geography and Service dataset contains information on use, paym"| __truncated__
#>  ..  ..$ dictionary : chr "https://data.cms.gov/resources/medicare-physician-other-practitioners-by-geography-and-service-data-dictionary"
#>  ..  ..$ site       : chr "https://data.cms.gov/provider-summary-by-type-of-service/medicare-physician-other-practitioners/medicare-physic"| __truncated__
#>  ..  ..$ rows       : int 270673
#>  ..  ..$ pages      : int 55
#>  ..  ..$ fields     : chr [1:15] "Rndrng_Prvdr_Geo_Lvl" "Rndrng_Prvdr_Geo_Cd" "Rndrng_Prvdr_Geo_Desc" "HCPCS_Cd" ...
#>  ..  ..$ endpoints  : tibble [10 × 3] (S3: tbl_df/tbl/data.frame)
#>  ..  .. ..$ year      : int [1:10] 2022 2021 2020 2019 2018 2017 2016 2015 2014 2013
#>  ..  .. ..$ identifier: chr [1:10] "https://data.cms.gov/data-api/v1/dataset/87304f15-9ed0-41dc-a141-6141a0327453/data" "https://data.cms.gov/data-api/v1/dataset/f00f462f-bb61-48b1-a321-1c50d78ce175/data" "https://data.cms.gov/data-api/v1/dataset/31eb3018-43e0-4259-a0d8-e7a1112ffd08/data" "https://data.cms.gov/data-api/v1/dataset/673030ae-ceed-4561-8fca-b1275395a86a/data" ...
#>  ..  .. ..$ resources : chr [1:10] "https://data.cms.gov/data-api/v1/dataset-resources/87304f15-9ed0-41dc-a141-6141a0327453" "https://data.cms.gov/data-api/v1/dataset-resources/f00f462f-bb61-48b1-a321-1c50d78ce175" "https://data.cms.gov/data-api/v1/dataset-resources/31eb3018-43e0-4259-a0d8-e7a1112ffd08" "https://data.cms.gov/data-api/v1/dataset-resources/673030ae-ceed-4561-8fca-b1275395a86a" ...
#>  .. $ by_provider             :List of 7
#>  ..  ..$ description: chr "The Medicare Physician & Other Practitioners by Provider dataset provides information on use, payments, submitt"| __truncated__
#>  ..  ..$ dictionary : chr "https://data.cms.gov/resources/medicare-physician-other-practitioners-by-provider-data-dictionary"
#>  ..  ..$ site       : chr "https://data.cms.gov/provider-summary-by-type-of-service/medicare-physician-other-practitioners/medicare-physic"| __truncated__
#>  ..  ..$ rows       : int 1230293
#>  ..  ..$ pages      : int 247
#>  ..  ..$ fields     : chr [1:81] "Rndrng_NPI" "Rndrng_Prvdr_Last_Org_Name" "Rndrng_Prvdr_First_Name" "Rndrng_Prvdr_MI" ...
#>  ..  ..$ endpoints  : tibble [10 × 3] (S3: tbl_df/tbl/data.frame)
#>  ..  .. ..$ year      : int [1:10] 2022 2021 2020 2019 2018 2017 2016 2015 2014 2013
#>  ..  .. ..$ identifier: chr [1:10] "https://data.cms.gov/data-api/v1/dataset/21555c17-ec1b-4e74-b2c6-925c6cbb3147/data" "https://data.cms.gov/data-api/v1/dataset/44e0a489-666c-4ea4-a1a2-360b6cdc19db/data" "https://data.cms.gov/data-api/v1/dataset/29d799aa-c660-44fe-a51a-72c4b3e661ac/data" "https://data.cms.gov/data-api/v1/dataset/6a53afe5-1cbc-4b33-9dc8-926ee532dc66/data" ...
#>  ..  .. ..$ resources : chr [1:10] "https://data.cms.gov/data-api/v1/dataset-resources/21555c17-ec1b-4e74-b2c6-925c6cbb3147" "https://data.cms.gov/data-api/v1/dataset-resources/44e0a489-666c-4ea4-a1a2-360b6cdc19db" "https://data.cms.gov/data-api/v1/dataset-resources/29d799aa-c660-44fe-a51a-72c4b3e661ac" "https://data.cms.gov/data-api/v1/dataset-resources/6a53afe5-1cbc-4b33-9dc8-926ee532dc66" ...
#>  .. $ by_provider_and_service :List of 7
#>  ..  ..$ description: chr "The Medicare Physician & Other Practitioners by Provider and Service dataset provides information on use, payme"| __truncated__
#>  ..  ..$ dictionary : chr "https://data.cms.gov/resources/medicare-physician-other-practitioners-by-provider-and-service-data-dictionary"
#>  ..  ..$ site       : chr "https://data.cms.gov/provider-summary-by-type-of-service/medicare-physician-other-practitioners/medicare-physic"| __truncated__
#>  ..  ..$ rows       : int 9755427
#>  ..  ..$ pages      : int 1952
#>  ..  ..$ fields     : chr [1:28] "Rndrng_NPI" "Rndrng_Prvdr_Last_Org_Name" "Rndrng_Prvdr_First_Name" "Rndrng_Prvdr_MI" ...
#>  ..  ..$ endpoints  : tibble [10 × 3] (S3: tbl_df/tbl/data.frame)
#>  ..  .. ..$ year      : int [1:10] 2022 2021 2020 2019 2018 2017 2016 2015 2014 2013
#>  ..  .. ..$ identifier: chr [1:10] "https://data.cms.gov/data-api/v1/dataset/e650987d-01b7-4f09-b75e-b0b075afbf98/data" "https://data.cms.gov/data-api/v1/dataset/31dc2c47-f297-4948-bfb4-075e1bec3a02/data" "https://data.cms.gov/data-api/v1/dataset/c957b49e-1323-49e7-8678-c09da387551d/data" "https://data.cms.gov/data-api/v1/dataset/867b8ac7-ccb7-4cc9-873d-b24340d89e32/data" ...
#>  ..  .. ..$ resources : chr [1:10] "https://data.cms.gov/data-api/v1/dataset-resources/e650987d-01b7-4f09-b75e-b0b075afbf98" "https://data.cms.gov/data-api/v1/dataset-resources/31dc2c47-f297-4948-bfb4-075e1bec3a02" "https://data.cms.gov/data-api/v1/dataset-resources/c957b49e-1323-49e7-8678-c09da387551d" "https://data.cms.gov/data-api/v1/dataset-resources/867b8ac7-ccb7-4cc9-873d-b24340d89e32" ...
careTempGroup("suppliers")
#> <careTempGroup>
#>  @ title      : chr "Medicare Durable Medical Equipment, Devices & Supplies"
#>  @ periodicity: chr "Annually [R/P1Y]"
#>  @ years      : int [1:9] 2022 2021 2020 2019 2018 2017 2016 2015 2014
#>  @ groups     :List of 5
#>  .. $ by_geography_and_service         :List of 7
#>  ..  ..$ description: chr "The Medicare Durable Medical Equipment, Devices & Supplies by Geography and Service dataset contains informatio"| __truncated__
#>  ..  ..$ dictionary : chr "https://data.cms.gov/resources/medicare-durable-medical-equipment-devices-supplies-by-geography-and-service-dat"| __truncated__
#>  ..  ..$ site       : chr "https://data.cms.gov/provider-summary-by-type-of-service/medicare-durable-medical-equipment-devices-supplies/me"| __truncated__
#>  ..  ..$ rows       : int 39129
#>  ..  ..$ pages      : int 8
#>  ..  ..$ fields     : chr [1:18] "Rfrg_Prvdr_Geo_Lvl" "Rfrg_Prvdr_Geo_Cd" "Rfrg_Prvdr_Geo_Desc" "RBCS_Lvl" ...
#>  ..  ..$ endpoints  : tibble [9 × 3] (S3: tbl_df/tbl/data.frame)
#>  ..  .. ..$ year      : int [1:9] 2022 2021 2020 2019 2018 2017 2016 2015 2014
#>  ..  .. ..$ identifier: chr [1:9] "https://data.cms.gov/data-api/v1/dataset/a9d1461a-9c63-49e1-8d91-696082e20e47/data" "https://data.cms.gov/data-api/v1/dataset/daedccd6-5881-4997-93cb-ad6f3d1dd2fd/data" "https://data.cms.gov/data-api/v1/dataset/c5da241c-71d2-4b04-9f44-b324ff4b4c01/data" "https://data.cms.gov/data-api/v1/dataset/3931609c-3f1d-4ab4-95f9-0e31d43613f0/data" ...
#>  ..  .. ..$ resources : chr [1:9] "https://data.cms.gov/data-api/v1/dataset-resources/a9d1461a-9c63-49e1-8d91-696082e20e47" "https://data.cms.gov/data-api/v1/dataset-resources/daedccd6-5881-4997-93cb-ad6f3d1dd2fd" "https://data.cms.gov/data-api/v1/dataset-resources/c5da241c-71d2-4b04-9f44-b324ff4b4c01" "https://data.cms.gov/data-api/v1/dataset-resources/3931609c-3f1d-4ab4-95f9-0e31d43613f0" ...
#>  .. $ by_referring_provider            :List of 7
#>  ..  ..$ description: chr "The Medicare Durable Medical Equipment, Devices & Supplies by Referring Provider dataset contains information o"| __truncated__
#>  ..  ..$ dictionary : chr "https://data.cms.gov/resources/medicare-durable-medical-equipment-devices-supplies-by-referring-provider-data-dictionary"
#>  ..  ..$ site       : chr "https://data.cms.gov/provider-summary-by-type-of-service/medicare-durable-medical-equipment-devices-supplies/me"| __truncated__
#>  ..  ..$ rows       : int 385637
#>  ..  ..$ pages      : int 78
#>  ..  ..$ fields     : chr [1:97] "Rfrg_NPI" "Rfrg_Prvdr_Last_Name_Org" "Rfrg_Prvdr_First_Name" "Rfrg_Prvdr_MI" ...
#>  ..  ..$ endpoints  : tibble [9 × 3] (S3: tbl_df/tbl/data.frame)
#>  ..  .. ..$ year      : int [1:9] 2022 2021 2020 2019 2018 2017 2016 2015 2014
#>  ..  .. ..$ identifier: chr [1:9] "https://data.cms.gov/data-api/v1/dataset/3dccdda1-c3fb-4ac5-908f-18b387495cb3/data" "https://data.cms.gov/data-api/v1/dataset/ac919fb0-dc48-4290-8abf-a4edb8e09139/data" "https://data.cms.gov/data-api/v1/dataset/d3c866db-3caf-4f3c-acc6-086a1770fb6b/data" "https://data.cms.gov/data-api/v1/dataset/d3e0a332-082d-4479-bb3e-f2911eb86497/data" ...
#>  ..  .. ..$ resources : chr [1:9] "https://data.cms.gov/data-api/v1/dataset-resources/3dccdda1-c3fb-4ac5-908f-18b387495cb3" "https://data.cms.gov/data-api/v1/dataset-resources/ac919fb0-dc48-4290-8abf-a4edb8e09139" "https://data.cms.gov/data-api/v1/dataset-resources/d3c866db-3caf-4f3c-acc6-086a1770fb6b" "https://data.cms.gov/data-api/v1/dataset-resources/d3e0a332-082d-4479-bb3e-f2911eb86497" ...
#>  .. $ by_referring_provider_and_service:List of 7
#>  ..  ..$ description: chr "The Medicare Durable Medical Equipment, Devices & Supplies by Referring Provider and Service dataset contains i"| __truncated__
#>  ..  ..$ dictionary : chr "https://data.cms.gov/resources/medicare-durable-medical-equipment-devices-supplies-by-referring-provider-and-se"| __truncated__
#>  ..  ..$ site       : chr "https://data.cms.gov/provider-summary-by-type-of-service/medicare-durable-medical-equipment-devices-supplies/me"| __truncated__
#>  ..  ..$ rows       : int 1457378
#>  ..  ..$ pages      : int 292
#>  ..  ..$ fields     : chr [1:33] "Rfrg_NPI" "Rfrg_Prvdr_Last_Name_Org" "Rfrg_Prvdr_First_Name" "Rfrg_Prvdr_MI" ...
#>  ..  ..$ endpoints  : tibble [9 × 3] (S3: tbl_df/tbl/data.frame)
#>  ..  .. ..$ year      : int [1:9] 2022 2021 2020 2019 2018 2017 2016 2015 2014
#>  ..  .. ..$ identifier: chr [1:9] "https://data.cms.gov/data-api/v1/dataset/0dd53b4b-67ba-48c7-b8fa-fecbdfc83b70/data" "https://data.cms.gov/data-api/v1/dataset/46ae675c-bc81-40ca-aa79-64da1c1ec9d9/data" "https://data.cms.gov/data-api/v1/dataset/323df359-ceac-4525-a350-e2cd9eb128fe/data" "https://data.cms.gov/data-api/v1/dataset/eb0019f6-791d-4065-ae4e-4761d2f6c9f2/data" ...
#>  ..  .. ..$ resources : chr [1:9] "https://data.cms.gov/data-api/v1/dataset-resources/0dd53b4b-67ba-48c7-b8fa-fecbdfc83b70" "https://data.cms.gov/data-api/v1/dataset-resources/46ae675c-bc81-40ca-aa79-64da1c1ec9d9" "https://data.cms.gov/data-api/v1/dataset-resources/323df359-ceac-4525-a350-e2cd9eb128fe" "https://data.cms.gov/data-api/v1/dataset-resources/eb0019f6-791d-4065-ae4e-4761d2f6c9f2" ...
#>  .. $ by_supplier                      :List of 7
#>  ..  ..$ description: chr "The Medicare Durable Medical Equipment, Devices & Supplies by Supplier dataset contains information on usage, p"| __truncated__
#>  ..  ..$ dictionary : chr "https://data.cms.gov/resources/medicare-durable-medical-equipment-devices-supplies-by-supplier-data-dictionary"
#>  ..  ..$ site       : chr "https://data.cms.gov/provider-summary-by-type-of-service/medicare-durable-medical-equipment-devices-supplies/me"| __truncated__
#>  ..  ..$ rows       : int 66406
#>  ..  ..$ pages      : int 14
#>  ..  ..$ fields     : chr [1:93] "Suplr_NPI" "Suplr_Prvdr_Last_Name_Org" "Suplr_Prvdr_First_Name" "Suplr_Prvdr_MI" ...
#>  ..  ..$ endpoints  : tibble [9 × 3] (S3: tbl_df/tbl/data.frame)
#>  ..  .. ..$ year      : int [1:9] 2022 2021 2020 2019 2018 2017 2016 2015 2014
#>  ..  .. ..$ identifier: chr [1:9] "https://data.cms.gov/data-api/v1/dataset/471f9c0d-12c4-4601-a547-559dcfa2d0e2/data" "https://data.cms.gov/data-api/v1/dataset/887561f5-1676-4c93-84b6-d2e50eb609d3/data" "https://data.cms.gov/data-api/v1/dataset/04f11092-de71-4a40-93b1-da2b41e4e938/data" "https://data.cms.gov/data-api/v1/dataset/e8c8eb83-a1d8-49d3-ac99-dda79b147021/data" ...
#>  ..  .. ..$ resources : chr [1:9] "https://data.cms.gov/data-api/v1/dataset-resources/471f9c0d-12c4-4601-a547-559dcfa2d0e2" "https://data.cms.gov/data-api/v1/dataset-resources/887561f5-1676-4c93-84b6-d2e50eb609d3" "https://data.cms.gov/data-api/v1/dataset-resources/04f11092-de71-4a40-93b1-da2b41e4e938" "https://data.cms.gov/data-api/v1/dataset-resources/e8c8eb83-a1d8-49d3-ac99-dda79b147021" ...
#>  .. $ by_supplier_and_service          :List of 7
#>  ..  ..$ description: chr "The Medicare Durable Medical Equipment, Devices & Supplies by Supplier and Service dataset contains information"| __truncated__
#>  ..  ..$ dictionary : chr "https://data.cms.gov/resources/medicare-durable-medical-equipment-devices-supplies-by-supplier-and-service-data-dictionary"
#>  ..  ..$ site       : chr "https://data.cms.gov/provider-summary-by-type-of-service/medicare-durable-medical-equipment-devices-supplies/me"| __truncated__
#>  ..  ..$ rows       : int 482638
#>  ..  ..$ pages      : int 97
#>  ..  ..$ fields     : chr [1:32] "Suplr_NPI" "Suplr_Prvdr_Last_Name_Org" "Suplr_Prvdr_First_Name" "Suplr_Prvdr_MI" ...
#>  ..  ..$ endpoints  : tibble [9 × 3] (S3: tbl_df/tbl/data.frame)
#>  ..  .. ..$ year      : int [1:9] 2022 2021 2020 2019 2018 2017 2016 2015 2014
#>  ..  .. ..$ identifier: chr [1:9] "https://data.cms.gov/data-api/v1/dataset/dad703df-2e52-4ac6-a26f-51ecbf463d77/data" "https://data.cms.gov/data-api/v1/dataset/e531ffc9-57ad-48db-8d88-f686edb8b8e1/data" "https://data.cms.gov/data-api/v1/dataset/1fe3a8c3-ecfd-44b9-9418-fda1843f72a5/data" "https://data.cms.gov/data-api/v1/dataset/5165dbdd-eae3-4b94-82cd-da635e00e0fa/data" ...
#>  ..  .. ..$ resources : chr [1:9] "https://data.cms.gov/data-api/v1/dataset-resources/dad703df-2e52-4ac6-a26f-51ecbf463d77" "https://data.cms.gov/data-api/v1/dataset-resources/e531ffc9-57ad-48db-8d88-f686edb8b8e1" "https://data.cms.gov/data-api/v1/dataset-resources/1fe3a8c3-ecfd-44b9-9418-fda1843f72a5" "https://data.cms.gov/data-api/v1/dataset-resources/5165dbdd-eae3-4b94-82cd-da635e00e0fa" ...
careTempGroup("prescribers")
#> <careTempGroup>
#>  @ title      : chr "Medicare Part D Prescribers"
#>  @ periodicity: chr "Annually [R/P1Y]"
#>  @ years      : int [1:10] 2022 2021 2020 2019 2018 2017 2016 2015 2014 2013
#>  @ groups     :List of 3
#>  .. $ by_geography_and_drug:List of 7
#>  ..  ..$ description: chr "The Medicare Part D Prescribers by Geography and Drug dataset contains information on prescription drugs prescr"| __truncated__
#>  ..  ..$ dictionary : chr "https://data.cms.gov/resources/medicare-part-d-prescribers-by-geography-and-drug-data-dictionary"
#>  ..  ..$ site       : chr "https://data.cms.gov/provider-summary-by-type-of-service/medicare-part-d-prescribers/medicare-part-d-prescriber"| __truncated__
#>  ..  ..$ rows       : int 115396
#>  ..  ..$ pages      : int 24
#>  ..  ..$ fields     : chr [1:22] "Prscrbr_Geo_Lvl" "Prscrbr_Geo_Cd" "Prscrbr_Geo_Desc" "Brnd_Name" ...
#>  ..  ..$ endpoints  : tibble [10 × 3] (S3: tbl_df/tbl/data.frame)
#>  ..  .. ..$ year      : int [1:10] 2022 2021 2020 2019 2018 2017 2016 2015 2014 2013
#>  ..  .. ..$ identifier: chr [1:10] "https://data.cms.gov/data-api/v1/dataset/1fc57194-a51d-4864-aee6-de0889488151/data" "https://data.cms.gov/data-api/v1/dataset/7dda2a9d-034a-446a-b4b3-e1254e0127b2/data" "https://data.cms.gov/data-api/v1/dataset/83891e77-99cf-4865-b60a-97703b916e09/data" "https://data.cms.gov/data-api/v1/dataset/73a6335e-f16f-4c81-a84b-6b5a986e2bf8/data" ...
#>  ..  .. ..$ resources : chr [1:10] "https://data.cms.gov/data-api/v1/dataset-resources/1fc57194-a51d-4864-aee6-de0889488151" "https://data.cms.gov/data-api/v1/dataset-resources/7dda2a9d-034a-446a-b4b3-e1254e0127b2" "https://data.cms.gov/data-api/v1/dataset-resources/83891e77-99cf-4865-b60a-97703b916e09" "https://data.cms.gov/data-api/v1/dataset-resources/73a6335e-f16f-4c81-a84b-6b5a986e2bf8" ...
#>  .. $ by_provider          :List of 7
#>  ..  ..$ description: chr "The Medicare Part D Prescribers by Provider dataset contains information on prescription drugs prescribed by in"| __truncated__
#>  ..  ..$ dictionary : chr "https://data.cms.gov/resources/medicare-part-d-prescribers-by-provider-data-dictionary"
#>  ..  ..$ site       : chr "https://data.cms.gov/provider-summary-by-type-of-service/medicare-part-d-prescribers/medicare-part-d-prescribers-by-provider"
#>  ..  ..$ rows       : int 1332309
#>  ..  ..$ pages      : int 267
#>  ..  ..$ fields     : chr [1:84] "Prscrbr_NPI" "Prscrbr_Last_Org_Name" "Prscrbr_First_Name" "Prscrbr_MI" ...
#>  ..  ..$ endpoints  : tibble [10 × 3] (S3: tbl_df/tbl/data.frame)
#>  ..  .. ..$ year      : int [1:10] 2022 2021 2020 2019 2018 2017 2016 2015 2014 2013
#>  ..  .. ..$ identifier: chr [1:10] "https://data.cms.gov/data-api/v1/dataset/bed99012-c527-4d9d-92ea-67ec2510abea/data" "https://data.cms.gov/data-api/v1/dataset/3f7ab9ce-6fb6-4e6b-9af3-b681f2d3a95e/data" "https://data.cms.gov/data-api/v1/dataset/ed6548c1-c905-4fdf-84cc-1c897ca6210d/data" "https://data.cms.gov/data-api/v1/dataset/007f61da-2c20-4c80-90a5-88be67c8d022/data" ...
#>  ..  .. ..$ resources : chr [1:10] "https://data.cms.gov/data-api/v1/dataset-resources/bed99012-c527-4d9d-92ea-67ec2510abea" "https://data.cms.gov/data-api/v1/dataset-resources/3f7ab9ce-6fb6-4e6b-9af3-b681f2d3a95e" "https://data.cms.gov/data-api/v1/dataset-resources/ed6548c1-c905-4fdf-84cc-1c897ca6210d" "https://data.cms.gov/data-api/v1/dataset-resources/007f61da-2c20-4c80-90a5-88be67c8d022" ...
#>  .. $ by_provider_and_drug :List of 7
#>  ..  ..$ description: chr "The Medicare Part D Prescribers by Provider and Drug dataset provides information on prescription drugs prescri"| __truncated__
#>  ..  ..$ dictionary : chr "https://data.cms.gov/resources/medicare-part-d-prescribers-by-provider-and-drug-data-dictionary"
#>  ..  ..$ site       : chr "https://data.cms.gov/provider-summary-by-type-of-service/medicare-part-d-prescribers/medicare-part-d-prescriber"| __truncated__
#>  ..  ..$ rows       : int 25869521
#>  ..  ..$ pages      : int 5174
#>  ..  ..$ fields     : chr [1:22] "Prscrbr_NPI" "Prscrbr_Last_Org_Name" "Prscrbr_First_Name" "Prscrbr_City" ...
#>  ..  ..$ endpoints  : tibble [10 × 3] (S3: tbl_df/tbl/data.frame)
#>  ..  .. ..$ year      : int [1:10] 2022 2021 2020 2019 2018 2017 2016 2015 2014 2013
#>  ..  .. ..$ identifier: chr [1:10] "https://data.cms.gov/data-api/v1/dataset/b101b457-ffa4-49bb-8fd9-27c1266086e2/data" "https://data.cms.gov/data-api/v1/dataset/f68114ed-f854-4ffc-9c6e-ed78b5e2f8d0/data" "https://data.cms.gov/data-api/v1/dataset/7795fe20-e80e-435a-a9ed-d2d65e05feeb/data" "https://data.cms.gov/data-api/v1/dataset/2a6705e6-7a1e-460c-ba22-35249a531918/data" ...
#>  ..  .. ..$ resources : chr [1:10] "https://data.cms.gov/data-api/v1/dataset-resources/b101b457-ffa4-49bb-8fd9-27c1266086e2" "https://data.cms.gov/data-api/v1/dataset-resources/f68114ed-f854-4ffc-9c6e-ed78b5e2f8d0" "https://data.cms.gov/data-api/v1/dataset-resources/7795fe20-e80e-435a-a9ed-d2d65e05feeb" "https://data.cms.gov/data-api/v1/dataset-resources/2a6705e6-7a1e-460c-ba22-35249a531918" ...
careTempGroup("staffing")
#> <careTempGroup>
#>  @ title      : chr "Payroll Based Journal"
#>  @ periodicity: chr "Quarterly [R/P3M]"
#>  @ years      : int [1:8] 2024 2023 2022 2021 2020 2019 2018 2017
#>  @ groups     :List of 3
#>  .. $ daily_non_nurse_staffing             :List of 7
#>  ..  ..$ description: chr "The Payroll Based Journal (PBJ) Nurse Staffing and Non-Nurse Staffing datasets provide information submitted by"| __truncated__
#>  ..  ..$ dictionary : chr "https://data.cms.gov/resources/payroll-based-journal-daily-non-nurse-staffing-data-dictionary"
#>  ..  ..$ site       : chr "https://data.cms.gov/quality-of-care/payroll-based-journal-daily-non-nurse-staffing"
#>  ..  ..$ rows       : int 1338416
#>  ..  ..$ pages      : int 268
#>  ..  ..$ fields     : chr [1:82] "PROVNUM" "PROVNAME" "CITY" "STATE" ...
#>  ..  ..$ endpoints  : tibble [8 × 3] (S3: tbl_df/tbl/data.frame)
#>  ..  .. ..$ year      : int [1:8] 2024 2023 2022 2021 2020 2019 2018 2017
#>  ..  .. ..$ identifier: chr [1:8] "https://data.cms.gov/data-api/v1/dataset/293788f3-20a9-4ebd-ab4a-9989325ce3e7/data" "https://data.cms.gov/data-api/v1/dataset/80e14392-7640-4fef-b226-309929273962/data" "https://data.cms.gov/data-api/v1/dataset/161d21ed-2f94-46da-aeb3-6bdd58d57264/data" "https://data.cms.gov/data-api/v1/dataset/d805bf46-faa2-48d8-89af-1f40def16c36/data" ...
#>  ..  .. ..$ resources : chr [1:8] "https://data.cms.gov/data-api/v1/dataset-resources/293788f3-20a9-4ebd-ab4a-9989325ce3e7" "https://data.cms.gov/data-api/v1/dataset-resources/80e14392-7640-4fef-b226-309929273962" "https://data.cms.gov/data-api/v1/dataset-resources/161d21ed-2f94-46da-aeb3-6bdd58d57264" "https://data.cms.gov/data-api/v1/dataset-resources/d805bf46-faa2-48d8-89af-1f40def16c36" ...
#>  .. $ daily_nurse_staffing                 :List of 7
#>  ..  ..$ description: chr "The Payroll Based Journal (PBJ) Nurse Staffing and Non-Nurse Staffing datasets provide information submitted by"| __truncated__
#>  ..  ..$ dictionary : chr "https://data.cms.gov/resources/payroll-based-journal-daily-nurse-staffing-data-dictionary"
#>  ..  ..$ site       : chr "https://data.cms.gov/quality-of-care/payroll-based-journal-daily-nurse-staffing"
#>  ..  ..$ rows       : int 1338416
#>  ..  ..$ pages      : int 268
#>  ..  ..$ fields     : chr [1:33] "PROVNUM" "PROVNAME" "CITY" "STATE" ...
#>  ..  ..$ endpoints  : tibble [8 × 3] (S3: tbl_df/tbl/data.frame)
#>  ..  .. ..$ year      : int [1:8] 2024 2023 2022 2021 2020 2019 2018 2017
#>  ..  .. ..$ identifier: chr [1:8] "https://data.cms.gov/data-api/v1/dataset/989fbc78-1655-487d-9f24-d68e9a0ab3af/data" "https://data.cms.gov/data-api/v1/dataset/ef20ab44-833c-4611-9600-2862e460072c/data" "https://data.cms.gov/data-api/v1/dataset/7a678e92-0418-44e0-9d72-043fd23f3959/data" "https://data.cms.gov/data-api/v1/dataset/93881636-0cde-4118-a110-ba2fd8cc890a/data" ...
#>  ..  .. ..$ resources : chr [1:8] "https://data.cms.gov/data-api/v1/dataset-resources/989fbc78-1655-487d-9f24-d68e9a0ab3af" "https://data.cms.gov/data-api/v1/dataset-resources/ef20ab44-833c-4611-9600-2862e460072c" "https://data.cms.gov/data-api/v1/dataset-resources/7a678e92-0418-44e0-9d72-043fd23f3959" "https://data.cms.gov/data-api/v1/dataset-resources/93881636-0cde-4118-a110-ba2fd8cc890a" ...
#>  .. $ employee_detail_nursing_home_staffing:List of 7
#>  ..  ..$ description: chr "The Payroll Based Journal (PBJ) Employee Detail Nursing Home Staffing dataset provides information submitted by"| __truncated__
#>  ..  ..$ dictionary : chr "https://data.cms.gov/resources/payroll-based-journal-employee-detail-nursing-home-staffing-data-dictionary"
#>  ..  ..$ site       : chr "https://data.cms.gov/quality-of-care/payroll-based-journal-employee-detail-nursing-home-staffing"
#>  ..  ..$ rows       : int 63260729
#>  ..  ..$ pages      : int 12653
#>  ..  ..$ fields     : chr [1:9] "PROVNUM" "STATE" "CY_Qtr" "WorkDate" ...
#>  ..  ..$ endpoints  : tibble [5 × 3] (S3: tbl_df/tbl/data.frame)
#>  ..  .. ..$ year      : int [1:5] 2024 2023 2022 2021 2020
#>  ..  .. ..$ identifier: chr [1:5] "https://data.cms.gov/data-api/v1/dataset/f513e55c-236c-437c-9a02-b53cc4acad72/data" "https://data.cms.gov/data-api/v1/dataset/86f6f9ad-2cb2-447c-b460-cf4f819fea7c/data" "https://data.cms.gov/data-api/v1/dataset/91c55823-f84a-4b68-b770-9f594f9e776f/data" "https://data.cms.gov/data-api/v1/dataset/24e7953e-ed2a-4931-a272-a4c0995c4865/data" ...
#>  ..  .. ..$ resources : chr [1:5] "https://data.cms.gov/data-api/v1/dataset-resources/f513e55c-236c-437c-9a02-b53cc4acad72" "https://data.cms.gov/data-api/v1/dataset-resources/86f6f9ad-2cb2-447c-b460-cf4f819fea7c" "https://data.cms.gov/data-api/v1/dataset-resources/91c55823-f84a-4b68-b770-9f594f9e776f" "https://data.cms.gov/data-api/v1/dataset-resources/24e7953e-ed2a-4931-a272-a4c0995c4865" ...