API Catalogs

List of API catalogs:

care: CMS Medicare API

prov: CMS Provider API

open: CMS Open Payments API

caid: CMS Medicaid API

hgov: CMS HealthCare.gov API

Usage

catalogs()

Value

<list> of catalogs

Examples

catalogs()
#> ! The following functions will be optimised package-wide:
#> `sum`
#> `prod`
#> `mean`
#> `median`
#> `min`
#> `max`
#> `sd`
#> `var`
#> `dplyr::n`
#> `dplyr::first`
#> `dplyr::last`
#> `dplyr::n_distinct`
#> `dplyr::row_number`
#> `dplyr::lag`
#> `dplyr::lead`
#> `dplyr::cur_group`
#> `dplyr::cur_group_id`
#> `dplyr::cur_group_rows`
#> 
#> Optimised expressions are independent from each other and typical data-masking
#> rules may not apply
#> 
#> Run `options(fastplyr.optimise = FALSE)` to disable optimisations globally
#> 
#> Run `options(fastplyr.inform = FALSE)` to disable this message
#> This message is displayed once per session.
#> left join: x[id] 318/318 (100%) <1.01:1st> w[id] 315/315 (100%)
#> $care
#> $care$end
#> # A tibble: 139 × 14
#>    clog  api   title         description modified   periodicity temporal contact
#>  * <chr> <chr> <chr>         <chr>       <date>     <chr>       <chr>    <chr>  
#>  1 care  end   ACO REACH Al… The Accoun… 2025-05-21 Annually [… 2021-01… ACO RE…
#>  2 care  end   ACO REACH El… Accountabl… 2025-05-21 Annually [… 2021-01… ACO RE…
#>  3 care  end   ACO REACH Fi… The Accoun… 2025-05-21 Annually [… 2021-01… ACO RE…
#>  4 care  end   ACO REACH Pr… The Accoun… 2025-05-21 Annually [… 2021-01… ACO RE…
#>  5 care  end   Accountable … The Accoun… 2025-05-16 Annually [… 2014-01… Shared…
#>  6 care  end   Accountable … The Accoun… 2025-05-16 Annually [… 2017-01… Shared…
#>  7 care  end   Accountable … TheAccount… 2025-05-16 Annually [… 2014-01… Shared…
#>  8 care  end   Advance Inve… The Advanc… 2025-03-18 Annually [… 2024-01… Shared…
#>  9 care  end   Agency for H… The Agency… 2020-12-08 Annually [… 2015-01… Defici…
#> 10 care  end   CMS Program … The CMS Pr… 2025-05-14 Annually [… 2013-01… CMS Pr…
#> # ℹ 129 more rows
#> # ℹ 6 more variables: identifier <chr>, dictionary <chr>, site <chr>,
#> #   references <chr>, download <chr>, resources <chr>
#> 
#> $care$tmp
#> # A tibble: 79 × 10
#>    clog  api   title description periodicity contact dictionary site  references
#>  * <chr> <chr> <chr> <chr>       <chr>       <chr>   <chr>      <chr> <chr>     
#>  1 care  tmp   Acco… The Accoun… Annually [… Shared… https://d… http… https://d…
#>  2 care  tmp   Acco… The Accoun… Annually [… Shared… https://d… http… https://d…
#>  3 care  tmp   Acco… TheAccount… Annually [… Shared… https://d… http… https://d…
#>  4 care  tmp   Adva… The Advanc… Annually [… Shared… https://d… http… https://d…
#>  5 care  tmp   Agen… The Agency… Annually [… Defici… https://d… http… https://d…
#>  6 care  tmp   Coun… The Shared… Annually [… Shared… https://d… http… https://d…
#>  7 care  tmp   Defi… The Public… Annually [… Defici… https://d… http… https://d…
#>  8 care  tmp   End-… The End-St… Monthly [R… ETC Mo… https://d… http… https://d…
#>  9 care  tmp   Faci… The Facili… Quarterly … NH Fac… https://d… http… https://d…
#> 10 care  tmp   Fede… TheFederal… Quarterly … Provid… https://d… http… https://d…
#> # ℹ 69 more rows
#> # ℹ 1 more variable: endpoints <list<tibble[,6]>>
#> 
#> 
#> $prov
#> $prov$end
#> # A tibble: 152 × 13
#>    clog  api   title          group description issued     modified   released  
#>    <chr> <chr> <chr>          <chr> <chr>       <date>     <date>     <date>    
#>  1 prov  end   Dialysis Faci… Dial… "A list of… 2020-03-14 2025-03-24 2025-04-23
#>  2 prov  end   Dialysis Faci… Dial… " National… 2020-03-14 2025-03-24 2025-04-23
#>  3 prov  end   Dialysis Faci… Dial… "State ave… 2020-03-14 2025-03-24 2025-04-23
#>  4 prov  end   ESRD QIP - Cl… Dial… "This data… 2021-01-20 2024-12-03 2025-01-15
#>  5 prov  end   ESRD QIP - Co… Dial… "This data… 2021-01-20 2024-12-03 2025-01-15
#>  6 prov  end   ESRD QIP - Co… Dial… "This data… 2025-01-15 2024-12-03 2025-01-15
#>  7 prov  end   ESRD QIP - Di… Dial… "This data… 2021-01-20 2024-12-03 2025-01-15
#>  8 prov  end   ESRD QIP - Fo… Dial… "This tabl… 2021-01-20 2024-12-03 2025-01-15
#>  9 prov  end   ESRD QIP - Hy… Dial… "This data… 2021-01-20 2024-12-03 2025-01-15
#> 10 prov  end   ESRD QIP - In… Dial… "This data… 2021-01-20 2024-12-03 2025-01-15
#> # ℹ 142 more rows
#> # ℹ 5 more variables: identifier <chr>, contact <chr>, download <chr>,
#> #   site <chr>, dictionary <chr>
#> 
#> 
#> $open
#> $open$end
#> # A tibble: 18 × 8
#>    clog  api   title          description modified   identifier contact download
#>    <chr> <chr> <chr>          <chr>       <date>     <chr>      <chr>   <chr>   
#>  1 open  end   Covered Recip… The Covere… 2025-01-30 https://o… Open P… https:/…
#>  2 open  end   National Leve… This datas… 2025-01-30 https://o… Open P… https:/…
#>  3 open  end   National Leve… This datas… 2025-01-30 https://o… Open P… https:/…
#>  4 open  end   Payments Grou… This datas… 2025-01-30 https://o… Open P… https:/…
#>  5 open  end   Payments Grou… This datas… 2025-01-30 https://o… Open P… https:/…
#>  6 open  end   Payments Grou… This datas… 2025-01-30 https://o… Open P… https:/…
#>  7 open  end   Payments Grou… This datas… 2025-01-30 https://o… Open P… https:/…
#>  8 open  end   Payments Grou… This datas… 2025-01-30 https://o… Open P… https:/…
#>  9 open  end   Payments Grou… This datas… 2025-01-30 https://o… Open P… https:/…
#> 10 open  end   Payments Grou… This datas… 2025-01-30 https://o… Open P… https:/…
#> 11 open  end   Physician (Di… This datas… 2025-01-30 https://o… Open P… https:/…
#> 12 open  end   Profile Infor… This datas… 2025-01-30 https://o… Open P… https:/…
#> 13 open  end   Provider Prof… Maps secon… 2025-01-30 https://o… Open P… https:/…
#> 14 open  end   Reporting Ent… This datas… 2025-01-30 https://o… Open P… https:/…
#> 15 open  end   State Level P… This datas… 2025-01-30 https://o… Open P… https:/…
#> 16 open  end   State Payment… Groups pay… 2025-01-30 https://o… Open P… https:/…
#> 17 open  end   Summary Dashb… Summary da… 2025-01-30 https://o… Open P… https:/…
#> 18 open  end   Teaching Hosp… This datas… 2025-01-30 https://o… Open P… https:/…
#> 
#> $open$tmp
#> # A tibble: 8 × 6
#>   clog  api   title                             description modified   endpoints
#>   <chr> <chr> <chr>                             <chr>       <date>     <list<ti>
#> 1 open  tmp   General Payment Data              All genera… 2025-01-30   [7 × 4]
#> 2 open  tmp   Ownership Payment Data            All owners… 2025-01-30   [7 × 4]
#> 3 open  tmp   Payments Grouped by Covered Reci… This datas… 2025-01-30   [7 × 4]
#> 4 open  tmp   Payments Grouped by Covered Reci… This datas… 2025-01-30   [7 × 4]
#> 5 open  tmp   Payments Grouped by Reporting En… This datas… 2025-01-30   [7 × 4]
#> 6 open  tmp   Payments Grouped by Reporting En… This datas… 2025-01-30   [7 × 4]
#> 7 open  tmp   Research Payment Data             All resear… 2025-01-30   [7 × 4]
#> 8 open  tmp   State Payment Totals Grouped by … Groups pay… 2025-01-30   [7 × 4]
#> 
#> 
#> $caid
#> $caid$end
#> # A tibble: 94 × 8
#>    clog  api   title       identifier description periodicity modified   contact
#>  * <chr> <chr> <chr>       <chr>      <chr>       <chr>       <date>     <chr>  
#>  1 caid  end   Drug Produ… https://d… "Active dr… Decenniall… 2025-05-07 Medica…
#>  2 caid  end   State Medi… https://d… "All state… Monthly [R… 2025-05-30 Medica…
#>  3 caid  end   Drug Manuf… https://d… "Drug Manu… Decenniall… 2025-05-07 Medica…
#>  4 caid  end   Medicaid E… https://d… "Total Med… Decenniall… 2024-12-11 Medica…
#>  5 caid  end   Managed Ca… https://d… "The Medic… Decenniall… 2024-10-16 Jeff C…
#>  6 caid  end   Managed Ca… https://d… "The Medic… Decenniall… 2024-10-16 Medica…
#>  7 caid  end   Managed Lo… https://d… "1. Enroll… Decenniall… 2024-10-16 Jeff C…
#>  8 caid  end   Managed Ca… https://d… "The Medic… Decenniall… 2024-10-16 Medica…
#>  9 caid  end   Managed Ca… https://d… "Number of… Decenniall… 2024-10-16 Medica…
#> 10 caid  end   Medicaid C… https://d… "This data… Decenniall… 2024-10-28 Medica…
#> # ℹ 84 more rows
#> 
#> $caid$tmp
#> # A tibble: 6 × 6
#>   clog  api   title                            description periodicity endpoints
#>   <chr> <chr> <chr>                            <chr>       <chr>       <list<ti>
#> 1 caid  end   Child and Adult Health Care Qua… Performanc… Annually […  [10 × 3]
#> 2 caid  end   Managed Care Programs by State   Provides p… Decenniall…   [1 × 3]
#> 3 caid  end   NADAC                            National A… Decenniall…  [13 × 3]
#> 4 caid  end   Pricing Comparison for Blood Di… The Refere… Decenniall…   [7 × 3]
#> 5 caid  end   Product Data for Newly Reported… The data b… Decenniall…  [99 × 3]
#> 6 caid  end   State Drug Utilization Data      Drug utili… Decenniall…  [34 × 3]
#> 
#> $caid$download
#> # A tibble: 423 × 4
#>    title                                                    download ext       N
#>    <chr>                                                    <chr>    <chr> <int>
#>  1 Acute Care Services Provided to the Medicaid and CHIP P… https:/… csv       1
#>  2 April 2023 Updated Medicaid and CHIP Renewal Outcomes    https:/… xlsx      1
#>  3 April 2024 Medicaid and CHIP CAA Reporting Metrics       https:/… xlsx      1
#>  4 August 2023 Medicaid and CHIP CAA Reporting Metrics      https:/… xlsx      1
#>  5 August 2023 Updated Medicaid and CHIP Renewal Outcomes   https:/… xlsx      1
#>  6 Behavioral Health Services Provided to the Medicaid and… https:/… csv       1
#>  7 Beneficiaries receiving a behavioral health service by … https:/… csv       1
#>  8 Beneficiaries receiving a physical health service among… https:/… csv       1
#>  9 Beneficiaries receiving a physical hlth serv among bene… https:/… csv       1
#> 10 Beneficiaries who could benefit from integrated care, 2… https:/… csv       1
#> # ℹ 413 more rows
#> 
#> 
#> $hgov
#> $hgov$end
#> # A tibble: 28 × 11
#>    clog  api   title    identifier description periodicity issued     modified  
#>  * <chr> <chr> <chr>    <chr>      <chr>       <chr>       <date>     <date>    
#>  1 hgov  end   Agent B… https://d… NA          Decenniall… 2018-07-31 2025-06-13
#>  2 hgov  end   AB Regi… https://d… Access Dat… Decenniall… 2016-03-30 2025-06-13
#>  3 hgov  end   AB Susp… https://d… Access Dat… Decenniall… 2016-05-03 2025-06-13
#>  4 hgov  end   Issuer_… https://d… The datase… Decenniall… 2020-06-05 2025-01-13
#>  5 hgov  end   Rolling… https://d… NA          Decenniall… 2016-08-22 2024-08-09
#>  6 hgov  end   Auto-po… https://d… ECP Autofi… Decenniall… 2016-08-22 2024-08-09
#>  7 hgov  end   Issuer … https://d… NA          NA          2019-10-09 2020-09-29
#>  8 hgov  end   NIPR Va… https://d… This datas… Decenniall… 2016-11-17 2020-08-10
#>  9 hgov  end   Direct … https://d… NA          Decenniall… 2018-09-27 2024-08-07
#> 10 hgov  end   Issuer … https://d… NA          NA          2019-10-09 2019-11-22
#> # ℹ 18 more rows
#> # ℹ 3 more variables: contact <chr>, download <chr>, resources <chr>
#> 
#> $hgov$tmp
#> # A tibble: 15 × 4
#>    clog  api   title                                            endpoints
#>    <chr> <chr> <chr>                                   <list<tibble[,9]>>
#>  1 hgov  tmp   Benefits and Cost Sharing PUF                     [12 × 9]
#>  2 hgov  tmp   Business Rules PUF                                [12 × 9]
#>  3 hgov  tmp   MLR Dataset                                        [9 × 9]
#>  4 hgov  tmp   Machine Readable PUF                               [4 × 9]
#>  5 hgov  tmp   Network PUF                                       [12 × 9]
#>  6 hgov  tmp   Plan Attributes PUF                               [12 × 9]
#>  7 hgov  tmp   Plan ID Crosswalk PUF                             [11 × 9]
#>  8 hgov  tmp   QHP Landscape Individual Market Dental             [9 × 9]
#>  9 hgov  tmp   QHP Landscape Individual Market Medical            [9 × 9]
#> 10 hgov  tmp   QHP Landscape SHOP Market Dental                   [9 × 9]
#> 11 hgov  tmp   QHP Landscape SHOP Market Medical                 [10 × 9]
#> 12 hgov  tmp   Quality PUF                                        [8 × 9]
#> 13 hgov  tmp   Rate PUF                                          [12 × 9]
#> 14 hgov  tmp   Service Area PUF                                  [12 × 9]
#> 15 hgov  tmp   Transparency in Coverage PUF                       [5 × 9]
#> 
#>