Endpoint Collection Load a collection of endpoints by alias.

Endpoint Collection Load a collection of endpoints by alias.

Usage

collection(alias)

Arguments

alias

<chr> collection alias

Value

An S7 <class_collection> object.

Examples

collection("caid_unwind")
#> <class_collection>
#>  @ name   : chr "Medicaid Unwinding Report"
#>  @ members:List of 4
#>  .. $ unwind_market    : <class_caid>
#>  ..  ..@ metadata  :List of 4
#>  .. .. .. $ title      : chr "HealthCare.gov Marketplace Medicaid Unwinding Report"
#>  .. .. .. $ description: chr "Metrics from individual Marketplaces during the current reporting period. The report includes data for the stat"| __truncated__
#>  .. .. .. $ modified   : Date[1:1], format: "2024-08-30"
#>  .. .. .. $ periodicity: chr "Decennially [R/P10Y]"
#>  ..  ..@ dimensions: <class_dimensions>
#>  .. .. .. @ limit : int 8000
#>  .. .. .. @ rows  : int 204
#>  .. .. .. @ pages : int 1
#>  .. .. .. @ fields:List of 30
#>  .. .. .. .. $ state                           : chr(0) 
#>  .. .. .. .. $ metric                          : chr(0) 
#>  .. .. .. .. $ april_2023_net_count            : chr(0) 
#>  .. .. .. .. $ april_2023_net_percentage       : chr(0) 
#>  .. .. .. .. $ may_2023_net_count              : chr(0) 
#>  .. .. .. .. $ may_2023_net_percentage         : chr(0) 
#>  .. .. .. .. $ june_2023_net_count             : chr(0) 
#>  .. .. .. .. $ june_2023_net_percentage        : chr(0) 
#>  .. .. .. .. $ july_2023_net_count             : chr(0) 
#>  .. .. .. .. $ july_2023_net_percentage        : chr(0) 
#>  .. .. .. .. $ august_2023_net_count           : chr(0) 
#>  .. .. .. .. $ august_2023_net_percentage      : chr(0) 
#>  .. .. .. .. $ september_2023_net_count        : chr(0) 
#>  .. .. .. .. $ september_2023_net_percentage   : chr(0) 
#>  .. .. .. .. $ october_2023_net_count          : chr(0) 
#>  .. .. .. .. $ october_2023_net_percentage     : chr(0) 
#>  .. .. .. .. $ november_2023_net_count         : chr(0) 
#>  .. .. .. .. $ november_2023_net_percentage    : chr(0) 
#>  .. .. .. .. $ december_2023_net_count         : chr(0) 
#>  .. .. .. .. $ december_2023_net_percentage    : chr(0) 
#>  .. .. .. .. $ january_2024_net_count          : chr(0) 
#>  .. .. .. .. $ january_2024_net_percentage     : chr(0) 
#>  .. .. .. .. $ february_2024_net_count         : chr(0) 
#>  .. .. .. .. $ february_2024_net_percentage    : chr(0) 
#>  .. .. .. .. $ march_2024_net_count            : chr(0) 
#>  .. .. .. .. $ march_2024_net_percentage       : chr(0) 
#>  .. .. .. .. $ april_2024_net_count            : chr(0) 
#>  .. .. .. .. $ april_2024_net_percentage       : chr(0) 
#>  .. .. .. .. $ april_2024_cumulative_count     : chr(0) 
#>  .. .. .. .. $ april_cumulative_2024_percentage: chr(0) 
#>  ..  ..@ identifier: <class_current> chr "https://data.medicaid.gov/api/1/datastore/query/9a83ba5e-05f5-47f5-82de-f3a59233a912/0?count=true&results=true&"| __truncated__
#>  .. $ unwind_transition: <class_caid>
#>  ..  ..@ metadata  :List of 4
#>  .. .. .. $ title      : chr "HealthCare.gov Transitions Marketplace Medicaid Unwinding Report"
#>  .. .. .. $ description: chr "Metrics from individual Marketplaces during the current reporting period. The report includes data for the stat"| __truncated__
#>  .. .. .. $ modified   : Date[1:1], format: "2024-12-27"
#>  .. .. .. $ periodicity: chr "Decennially [R/P10Y]"
#>  ..  ..@ dimensions: <class_dimensions>
#>  .. .. .. @ limit : int 8000
#>  .. .. .. @ rows  : int 59527
#>  .. .. .. @ pages : int 8
#>  .. .. .. @ fields:List of 9
#>  .. .. .. .. $ state                                                           : chr(0) 
#>  .. .. .. .. $ data_record_note                                                : chr(0) 
#>  .. .. .. .. $ metric                                                          : chr(0) 
#>  .. .. .. .. $ time_period                                                     : chr(0) 
#>  .. .. .. .. $ release_through                                                 : chr(0) 
#>  .. .. .. .. $ count_of_individuals_whose_medicaid_or_chip_coverage_was_te_dd11: chr(0) 
#>  .. .. .. .. $ percentage_of_individuals_whose_medicaid_or_chip_coverage_w_da6e: chr(0) 
#>  .. .. .. .. $ cumulative_count_of_individuals_whose_medicaid_or_chip_cove_28d9: chr(0) 
#>  .. .. .. .. $ cumulative_percentage_of_individuals_whose_medicaid_or_chip_6ac2: chr(0) 
#>  ..  ..@ identifier: <class_current> chr "https://data.medicaid.gov/api/1/datastore/query/5636a78c-fe18-4229-aee1-e40fa910a8a0/0?count=true&results=true&"| __truncated__
#>  .. $ unwind_historic  : <class_caid>
#>  ..  ..@ metadata  :List of 4
#>  .. .. .. $ title      : chr "Separate CHIP Enrollment by Month and State Historic CAA/Unwinding Period"
#>  .. .. .. $ description: chr "<p>This historic dataset with total enrollment in separate CHIP programs by month and state was created to fulf"| __truncated__
#>  .. .. .. $ modified   : Date[1:1], format: "2024-12-27"
#>  .. .. .. $ periodicity: chr "Monthly [R/P1M]"
#>  ..  ..@ dimensions: <class_dimensions>
#>  .. .. .. @ limit : int 8000
#>  .. .. .. @ rows  : int 780
#>  .. .. .. @ pages : int 1
#>  .. .. .. @ fields:List of 5
#>  .. .. .. .. $ state           : chr(0) 
#>  .. .. .. .. $ reporting_period: chr(0) 
#>  .. .. .. .. $ release_date    : chr(0) 
#>  .. .. .. .. $ value           : chr(0) 
#>  .. .. .. .. $ data_notes      : chr(0) 
#>  ..  ..@ identifier: <class_current> chr "https://data.medicaid.gov/api/1/datastore/query/d30cfc7c-4b32-4df1-b2bf-e0a850befd77/0?count=true&results=true&"| __truncated__
#>  .. $ unwind_sbm       : <class_caid>
#>  ..  ..@ metadata  :List of 4
#>  .. .. .. $ title      : chr "State-based Marketplace (SBM) Medicaid Unwinding Report"
#>  .. .. .. $ description: chr "Metrics from individual Marketplaces during the current reporting period. The report includes data for the stat"| __truncated__
#>  .. .. .. $ modified   : Date[1:1], format: "2024-12-27"
#>  .. .. .. $ periodicity: chr "Decennially [R/P10Y]"
#>  ..  ..@ dimensions: <class_dimensions>
#>  .. .. .. @ limit : int 8000
#>  .. .. .. @ rows  : int 128
#>  .. .. .. @ pages : int 1
#>  .. .. .. @ fields:List of 59
#>  .. .. .. .. $ state                                      : chr(0) 
#>  .. .. .. .. $ metric                                     : chr(0) 
#>  .. .. .. .. $ april_2023_count                           : chr(0) 
#>  .. .. .. .. $ april_2023_percentage_applications         : chr(0) 
#>  .. .. .. .. $ april_2023_percentage_account_transfers    : chr(0) 
#>  .. .. .. .. $ may_2023_count                             : chr(0) 
#>  .. .. .. .. $ may_2023_percentage_applications           : chr(0) 
#>  .. .. .. .. $ may_2023_percentage_account_transfers      : chr(0) 
#>  .. .. .. .. $ june_2023_count                            : chr(0) 
#>  .. .. .. .. $ june_2023_percentage_applications          : chr(0) 
#>  .. .. .. .. $ june_2023_percentage_account_transfers     : chr(0) 
#>  .. .. .. .. $ july_2023_count                            : chr(0) 
#>  .. .. .. .. $ july_2023_percentage_applications          : chr(0) 
#>  .. .. .. .. $ july_2023_percentage_account_transfers     : chr(0) 
#>  .. .. .. .. $ august_2023_count                          : chr(0) 
#>  .. .. .. .. $ august_2023_percentage_applications        : chr(0) 
#>  .. .. .. .. $ august_2023_percentage_account_transfers   : chr(0) 
#>  .. .. .. .. $ september_2023_count                       : chr(0) 
#>  .. .. .. .. $ september_2023_percentage_applications     : chr(0) 
#>  .. .. .. .. $ september_2023_percentage_account_transfers: chr(0) 
#>  .. .. .. .. $ october_2023_count                         : chr(0) 
#>  .. .. .. .. $ october_2023_percentage_applications       : chr(0) 
#>  .. .. .. .. $ october_2023_percentage_account_transfers  : chr(0) 
#>  .. .. .. .. $ november_2023_count                        : chr(0) 
#>  .. .. .. .. $ november_2023_percentage_applications      : chr(0) 
#>  .. .. .. .. $ november_2023_percentage_account_transfers : chr(0) 
#>  .. .. .. .. $ december_2023_count                        : chr(0) 
#>  .. .. .. .. $ december_2023_percentage_applications      : chr(0) 
#>  .. .. .. .. $ december_2023_percentage_account_transfers : chr(0) 
#>  .. .. .. .. $ january_2024_count                         : chr(0) 
#>  .. .. .. .. $ january_2024_percentage_applications       : chr(0) 
#>  .. .. .. .. $ january_2024_percentage_account_transfers  : chr(0) 
#>  .. .. .. .. $ february_2024_count                        : chr(0) 
#>  .. .. .. .. $ february_2024_percentage_applications      : chr(0) 
#>  .. .. .. .. $ february_2024_percentage_account_transfers : chr(0) 
#>  .. .. .. .. $ march_2024_count                           : chr(0) 
#>  .. .. .. .. $ march_2024_percentage_applications         : chr(0) 
#>  .. .. .. .. $ march_2024_percentage_account_transfers    : chr(0) 
#>  .. .. .. .. $ april_2024_count                           : chr(0) 
#>  .. .. .. .. $ april_2024_percentage_applications         : chr(0) 
#>  .. .. .. .. $ april_2024_percentage_account_transfers    : chr(0) 
#>  .. .. .. .. $ may_2024_count                             : chr(0) 
#>  .. .. .. .. $ may_2024_percentage_applications           : chr(0) 
#>  .. .. .. .. $ may_2024_percentage_account_transfers      : chr(0) 
#>  .. .. .. .. $ june_2024_count                            : chr(0) 
#>  .. .. .. .. $ june_2024_percentage_applications          : chr(0) 
#>  .. .. .. .. $ june_2024_percentage_account_transfers     : chr(0) 
#>  .. .. .. .. $ july_2024_count                            : chr(0) 
#>  .. .. .. .. $ july_2024_percentage_applications          : chr(0) 
#>  .. .. .. .. $ july_2024_percentage_account_transfers     : chr(0) 
#>  .. .. .. .. $ august_2024_count                          : chr(0) 
#>  .. .. .. .. $ august_2024_percentage_applications        : chr(0) 
#>  .. .. .. .. $ august_2024_percentage_account_transfers   : chr(0) 
#>  .. .. .. .. $ september_2024_count                       : chr(0) 
#>  .. .. .. .. $ september_2024_percentage_applications     : chr(0) 
#>  .. .. .. .. $ september_2024_percentage_account_transfers: chr(0) 
#>  .. .. .. .. $ total_count                                : chr(0) 
#>  .. .. .. .. $ total_percentage_applications              : chr(0) 
#>  .. .. .. .. $ total_percentage_account_transfers         : chr(0) 
#>  ..  ..@ identifier: <class_current> chr "https://data.medicaid.gov/api/1/datastore/query/5670e72c-e44e-4282-ab67-4ebebaba3cbd/0?count=true&results=true&"| __truncated__
collection("caid_services")
#> <class_collection>
#>  @ name   : chr "Services Provided to the Medicaid and CHIP Population"
#>  @ members:List of 11
#>  .. $ service_acute        : <class_caid>
#>  ..  ..@ metadata  :List of 4
#>  .. .. .. $ title      : chr "Acute Care Services Provided to the Medicaid and CHIP Population"
#>  .. .. .. $ description: chr "This data set includes monthly counts and rates (per 1,000 beneficiaries) of acute care services, including eme"| __truncated__
#>  .. .. .. $ modified   : Date[1:1], format: "2024-01-05"
#>  .. .. .. $ periodicity: chr "Decennially [R/P10Y]"
#>  ..  ..@ dimensions: <class_dimensions>
#>  .. .. .. @ limit : int 8000
#>  .. .. .. @ rows  : int 25440
#>  .. .. .. @ pages : int 4
#>  .. .. .. @ fields:List of 8
#>  .. .. .. .. $ state                   : chr(0) 
#>  .. .. .. .. $ year                    : chr(0) 
#>  .. .. .. .. $ month                   : chr(0) 
#>  .. .. .. .. $ _condition              : chr(0) 
#>  .. .. .. .. $ acutecareservice        : chr(0) 
#>  .. .. .. .. $ servicecount            : chr(0) 
#>  .. .. .. .. $ rateper1000beneficiaries: chr(0) 
#>  .. .. .. .. $ dataquality             : chr(0) 
#>  ..  ..@ identifier: <class_current> chr "https://data.medicaid.gov/api/1/datastore/query/d55e5ef5-9a2e-4d6c-884e-65a40a70756b/0?count=true&results=true&"| __truncated__
#>  .. $ service_behavior     : <class_caid>
#>  ..  ..@ metadata  :List of 4
#>  .. .. .. $ title      : chr "Behavioral Health ServicesProvided to the Medicaid and CHIP Population"
#>  .. .. .. $ description: chr "This data set includes monthly counts and rates (per 1,000 beneficiaries) of behavioral health services, includ"| __truncated__
#>  .. .. .. $ modified   : Date[1:1], format: "2024-01-05"
#>  .. .. .. $ periodicity: chr "Decennially [R/P10Y]"
#>  ..  ..@ dimensions: <class_dimensions>
#>  .. .. .. @ limit : int 8000
#>  .. .. .. @ rows  : int 31800
#>  .. .. .. @ pages : int 4
#>  .. .. .. @ fields:List of 8
#>  .. .. .. .. $ state                   : chr(0) 
#>  .. .. .. .. $ year                    : chr(0) 
#>  .. .. .. .. $ month                   : chr(0) 
#>  .. .. .. .. $ _condition              : chr(0) 
#>  .. .. .. .. $ behavioralhealthservice : chr(0) 
#>  .. .. .. .. $ servicecount            : chr(0) 
#>  .. .. .. .. $ rateper1000beneficiaries: chr(0) 
#>  .. .. .. .. $ dataquality             : chr(0) 
#>  ..  ..@ identifier: <class_current> chr "https://data.medicaid.gov/api/1/datastore/query/f403f019-48f5-48f0-b0ce-c051cfe04a5e/0?count=true&results=true&"| __truncated__
#>  .. $ service_perinatal    : <class_caid>
#>  ..  ..@ metadata  :List of 4
#>  .. .. .. $ title      : chr "Perinatal Care Services Provided to Medicaid and CHIP Beneficiaries ages 15 to 44"
#>  .. .. .. $ description: chr "This data set includes monthly counts and rates (per 1,000 beneficiaries) of perinatal care, including prenatal"| __truncated__
#>  .. .. .. $ modified   : Date[1:1], format: "2024-01-05"
#>  .. .. .. $ periodicity: chr "Decennially [R/P10Y]"
#>  ..  ..@ dimensions: <class_dimensions>
#>  .. .. .. @ limit : int 8000
#>  .. .. .. @ rows  : int 12720
#>  .. .. .. @ pages : int 2
#>  .. .. .. @ fields:List of 7
#>  .. .. .. .. $ state                   : chr(0) 
#>  .. .. .. .. $ year                    : chr(0) 
#>  .. .. .. .. $ month                   : chr(0) 
#>  .. .. .. .. $ perinatalcaretype       : chr(0) 
#>  .. .. .. .. $ servicecount            : chr(0) 
#>  .. .. .. .. $ rateper1000beneficiaries: chr(0) 
#>  .. .. .. .. $ dataquality             : chr(0) 
#>  ..  ..@ identifier: <class_current> chr "https://data.medicaid.gov/api/1/datastore/query/ed67e610-aed3-4bed-842f-e6044511dd64/0?count=true&results=true&"| __truncated__
#>  .. $ service_screening    : <class_caid>
#>  ..  ..@ metadata  :List of 4
#>  .. .. .. $ title      : chr "Health Screenings Provided to Medicaid and CHIP Beneficiaries Under Age 19"
#>  .. .. .. $ description: chr "This data set includes monthly counts and rates (per 1,000 beneficiaries) ofhealth screenings provided to Medic"| __truncated__
#>  .. .. .. $ modified   : Date[1:1], format: "2024-01-05"
#>  .. .. .. $ periodicity: chr "Decennially [R/P10Y]"
#>  ..  ..@ dimensions: <class_dimensions>
#>  .. .. .. @ limit : int 8000
#>  .. .. .. @ rows  : int 3180
#>  .. .. .. @ pages : int 1
#>  .. .. .. @ fields:List of 7
#>  .. .. .. .. $ state                   : chr(0) 
#>  .. .. .. .. $ year                    : chr(0) 
#>  .. .. .. .. $ month                   : chr(0) 
#>  .. .. .. .. $ screeningservice        : chr(0) 
#>  .. .. .. .. $ servicecount            : chr(0) 
#>  .. .. .. .. $ rateper1000beneficiaries: chr(0) 
#>  .. .. .. .. $ dataquality             : chr(0) 
#>  ..  ..@ identifier: <class_current> chr "https://data.medicaid.gov/api/1/datastore/query/37b55242-7cd4-4818-b783-348adf39e50c/0?count=true&results=true&"| __truncated__
#>  .. $ service_contraceptive: <class_caid>
#>  ..  ..@ metadata  :List of 4
#>  .. .. .. $ title      : chr "Contraceptive Care Services Provided to Medicaid and CHIP Beneficiaries ages 15 to 44"
#>  .. .. .. $ description: chr "This data set includes monthly counts and rates (per 1,000 beneficiaries) of contraceptives, including any cont"| __truncated__
#>  .. .. .. $ modified   : Date[1:1], format: "2024-01-05"
#>  .. .. .. $ periodicity: chr "Decennially [R/P10Y]"
#>  ..  ..@ dimensions: <class_dimensions>
#>  .. .. .. @ limit : int 8000
#>  .. .. .. @ rows  : int 6360
#>  .. .. .. @ pages : int 1
#>  .. .. .. @ fields:List of 7
#>  .. .. .. .. $ state                   : chr(0) 
#>  .. .. .. .. $ year                    : chr(0) 
#>  .. .. .. .. $ month                   : chr(0) 
#>  .. .. .. .. $ contraceptivetype       : chr(0) 
#>  .. .. .. .. $ servicecount            : chr(0) 
#>  .. .. .. .. $ rateper1000beneficiaries: chr(0) 
#>  .. .. .. .. $ dataquality             : chr(0) 
#>  ..  ..@ identifier: <class_current> chr "https://data.medicaid.gov/api/1/datastore/query/eb88943c-309d-440f-a2d5-ef93e066d83c/0?count=true&results=true&"| __truncated__
#>  .. $ service_dental       : <class_caid>
#>  ..  ..@ metadata  :List of 4
#>  .. .. .. $ title      : chr "Dental Services Provided to Medicaid and CHIP Beneficiaries Under Age 19"
#>  .. .. .. $ description: chr "This data set includes monthly counts and rates (per 1,000 beneficiaries) of dental services provided to Medica"| __truncated__
#>  .. .. .. $ modified   : Date[1:1], format: "2024-01-05"
#>  .. .. .. $ periodicity: chr "Decennially [R/P10Y]"
#>  ..  ..@ dimensions: <class_dimensions>
#>  .. .. .. @ limit : int 8000
#>  .. .. .. @ rows  : int 3180
#>  .. .. .. @ pages : int 1
#>  .. .. .. @ fields:List of 7
#>  .. .. .. .. $ state                   : chr(0) 
#>  .. .. .. .. $ year                    : chr(0) 
#>  .. .. .. .. $ month                   : chr(0) 
#>  .. .. .. .. $ dentalservice           : chr(0) 
#>  .. .. .. .. $ servicecount            : chr(0) 
#>  .. .. .. .. $ rateper1000beneficiaries: chr(0) 
#>  .. .. .. .. $ dataquality             : chr(0) 
#>  ..  ..@ identifier: <class_current> chr "https://data.medicaid.gov/api/1/datastore/query/2aa0ed3d-6ef8-47a8-90b0-06b6d2fa3953/0?count=true&results=true&"| __truncated__
#>  .. $ service_pregnancy    : <class_caid>
#>  ..  ..@ metadata  :List of 4
#>  .. .. .. $ title      : chr "Pregnancy Outcomes for Medicaid and CHIP Beneficiaries ages 15 to 44"
#>  .. .. .. $ description: chr "This data set includes monthly counts and rates (per 1,000 beneficiaries) of pregnancy outcomes, including (1) "| __truncated__
#>  .. .. .. $ modified   : Date[1:1], format: "2024-01-05"
#>  .. .. .. $ periodicity: chr "Decennially [R/P10Y]"
#>  ..  ..@ dimensions: <class_dimensions>
#>  .. .. .. @ limit : int 8000
#>  .. .. .. @ rows  : int 6360
#>  .. .. .. @ pages : int 1
#>  .. .. .. @ fields:List of 7
#>  .. .. .. .. $ state                   : chr(0) 
#>  .. .. .. .. $ year                    : chr(0) 
#>  .. .. .. .. $ month                   : chr(0) 
#>  .. .. .. .. $ pregnancyoutcome        : chr(0) 
#>  .. .. .. .. $ servicecount            : chr(0) 
#>  .. .. .. .. $ rateper1000beneficiaries: chr(0) 
#>  .. .. .. .. $ dataquality             : chr(0) 
#>  ..  ..@ identifier: <class_current> chr "https://data.medicaid.gov/api/1/datastore/query/8dd46b48-f004-4245-a51d-a8d864ea586f/0?count=true&results=true&"| __truncated__
#>  .. $ service_telehealth   : <class_caid>
#>  ..  ..@ metadata  :List of 4
#>  .. .. .. $ title      : chr "Telehealth Services Provided to the Medicaid and CHIP Population"
#>  .. .. .. $ description: chr "This data set includes monthly counts and rates (per 1,000 beneficiaries) ofservices provided via telehealth, i"| __truncated__
#>  .. .. .. $ modified   : Date[1:1], format: "2024-01-05"
#>  .. .. .. $ periodicity: chr "Decennially [R/P10Y]"
#>  ..  ..@ dimensions: <class_dimensions>
#>  .. .. .. @ limit : int 8000
#>  .. .. .. @ rows  : int 12720
#>  .. .. .. @ pages : int 2
#>  .. .. .. @ fields:List of 8
#>  .. .. .. .. $ state                   : chr(0) 
#>  .. .. .. .. $ year                    : chr(0) 
#>  .. .. .. .. $ month                   : chr(0) 
#>  .. .. .. .. $ telehealthtype          : chr(0) 
#>  .. .. .. .. $ servicetype             : chr(0) 
#>  .. .. .. .. $ servicecount            : chr(0) 
#>  .. .. .. .. $ rateper1000beneficiaries: chr(0) 
#>  .. .. .. .. $ dataquality             : chr(0) 
#>  ..  ..@ identifier: <class_current> chr "https://data.medicaid.gov/api/1/datastore/query/651fa253-4dd4-4867-8725-2b5ae1dd5ce9/0?count=true&results=true&"| __truncated__
#>  .. $ service_vaccination  : <class_caid>
#>  ..  ..@ metadata  :List of 4
#>  .. .. .. $ title      : chr "Vaccinations Provided to the Medicaid and CHIP Population under age 19"
#>  .. .. .. $ description: chr "This data set includes monthly counts and rates (per 1,000 beneficiaries) of vaccinations provided to Medicaid "| __truncated__
#>  .. .. .. $ modified   : Date[1:1], format: "2024-01-05"
#>  .. .. .. $ periodicity: chr "Decennially [R/P10Y]"
#>  ..  ..@ dimensions: <class_dimensions>
#>  .. .. .. @ limit : int 8000
#>  .. .. .. @ rows  : int 47700
#>  .. .. .. @ pages : int 6
#>  .. .. .. @ fields:List of 7
#>  .. .. .. .. $ state                   : chr(0) 
#>  .. .. .. .. $ year                    : chr(0) 
#>  .. .. .. .. $ month                   : chr(0) 
#>  .. .. .. .. $ vaccinetype             : chr(0) 
#>  .. .. .. .. $ servicecount            : chr(0) 
#>  .. .. .. .. $ rateper1000beneficiaries: chr(0) 
#>  .. .. .. .. $ dataquality             : chr(0) 
#>  ..  ..@ identifier: <class_current> chr "https://data.medicaid.gov/api/1/datastore/query/43d419f9-f863-4f68-b622-311a8800100d/0?count=true&results=true&"| __truncated__
#>  .. $ service_bloodlead    : <class_caid>
#>  ..  ..@ metadata  :List of 4
#>  .. .. .. $ title      : chr "Blood Lead Screening Services Provided to Medicaid and CHIP Beneficiaries Ages 1-2"
#>  .. .. .. $ description: chr "This data set includes monthly counts and rates (per 1,000 beneficiaries) of behavioral health services, includ"| __truncated__
#>  .. .. .. $ modified   : Date[1:1], format: "2024-01-05"
#>  .. .. .. $ periodicity: chr "Decennially [R/P10Y]"
#>  ..  ..@ dimensions: <class_dimensions>
#>  .. .. .. @ limit : int 8000
#>  .. .. .. @ rows  : int 3180
#>  .. .. .. @ pages : int 1
#>  .. .. .. @ fields:List of 7
#>  .. .. .. .. $ state                    : chr(0) 
#>  .. .. .. .. $ year                     : chr(0) 
#>  .. .. .. .. $ month                    : chr(0) 
#>  .. .. .. .. $ bloodleadscreeningservice: chr(0) 
#>  .. .. .. .. $ servicecount             : chr(0) 
#>  .. .. .. .. $ rateper1000beneficiaries : chr(0) 
#>  .. .. .. .. $ dataquality              : chr(0) 
#>  ..  ..@ identifier: <class_current> chr "https://data.medicaid.gov/api/1/datastore/query/34fc0456-e28c-4762-91e1-49cd88462e7a/0?count=true&results=true&"| __truncated__
#>  .. $ service_respiratory  : <class_caid>
#>  ..  ..@ metadata  :List of 4
#>  .. .. .. $ title      : chr "Respiratory Conditions in the Medicaid and CHIP Population"
#>  .. .. .. $ description: chr "This data set includes monthly counts and percentages of Medicaid and CHIP beneficiaries, by state, who receive"| __truncated__
#>  .. .. .. $ modified   : Date[1:1], format: "2024-01-05"
#>  .. .. .. $ periodicity: chr "Decennially [R/P10Y]"
#>  ..  ..@ dimensions: <class_dimensions>
#>  .. .. .. @ limit : int 8000
#>  .. .. .. @ rows  : int 28620
#>  .. .. .. @ pages : int 4
#>  .. .. .. @ fields:List of 7
#>  .. .. .. .. $ state                    : chr(0) 
#>  .. .. .. .. $ year                     : chr(0) 
#>  .. .. .. .. $ month                    : chr(0) 
#>  .. .. .. .. $ _condition               : chr(0) 
#>  .. .. .. .. $ beneficiarycount         : chr(0) 
#>  .. .. .. .. $ percentageofbeneficiaries: chr(0) 
#>  .. .. .. .. $ dataquality              : chr(0) 
#>  ..  ..@ identifier: <class_current> chr "https://data.medicaid.gov/api/1/datastore/query/ed35fa7f-279b-4d68-b624-b6ff32e956df/0?count=true&results=true&"| __truncated__