Load a collection of endpoints by alias

Load a collection of endpoints by alias

Usage

collection(alias)

Arguments

alias

<chr> collection alias

Value

An S7 <class_collection> object.

Examples

collection("unwind")
#> <class_collection>
#>  @ name   : chr "Medicaid Unwinding Report"
#>  @ members:List of 4
#>  .. $ wind_mark    : <class_caid>
#>  ..  ..@ access: <class_current>
#>  .. .. .. @ identifier: chr "https://data.medicaid.gov/api/1/datastore/query/9a83ba5e-05f5-47f5-82de-f3a59233a912/0?count=true&results=true&"| __truncated__
#>  .. .. .. @ metadata  :List of 4
#>  .. .. .. .. $ title      : chr "HealthCare.gov Marketplace Medicaid Unwinding Report"
#>  .. .. .. .. $ description: chr "Metrics from individual Marketplaces during the current reporting period. The report includes data for the stat"| __truncated__
#>  .. .. .. .. $ modified   : Date[1:1], format: "2024-08-30"
#>  .. .. .. .. $ periodicity: chr "Decennially [R/P10Y]"
#>  .. .. .. @ dimensions: <class_dimensions>
#>  .. .. .. .. @ limit : int 8000
#>  .. .. .. .. @ rows  : int 204
#>  .. .. .. .. @ pages : int 1
#>  .. .. .. .. @ fields: chr [1:30] "state" "metric" "april_2023_net_count" "april_2023_net_percentage" ...
#>  .. $ wind_trans   : <class_caid>
#>  ..  ..@ access: <class_current>
#>  .. .. .. @ identifier: chr "https://data.medicaid.gov/api/1/datastore/query/5636a78c-fe18-4229-aee1-e40fa910a8a0/0?count=true&results=true&"| __truncated__
#>  .. .. .. @ metadata  :List of 4
#>  .. .. .. .. $ title      : chr "HealthCare.gov Transitions Marketplace Medicaid Unwinding Report"
#>  .. .. .. .. $ description: chr "Metrics from individual Marketplaces during the current reporting period. The report includes data for the stat"| __truncated__
#>  .. .. .. .. $ modified   : Date[1:1], format: "2024-12-27"
#>  .. .. .. .. $ periodicity: chr "Decennially [R/P10Y]"
#>  .. .. .. @ dimensions: <class_dimensions>
#>  .. .. .. .. @ limit : int 8000
#>  .. .. .. .. @ rows  : int 59527
#>  .. .. .. .. @ pages : int 8
#>  .. .. .. .. @ fields: chr [1:9] "state" "data_record_note" "metric" "time_period" ...
#>  .. $ wind_historic: <class_caid>
#>  ..  ..@ access: <class_current>
#>  .. .. .. @ identifier: chr "https://data.medicaid.gov/api/1/datastore/query/d30cfc7c-4b32-4df1-b2bf-e0a850befd77/0?count=true&results=true&"| __truncated__
#>  .. .. .. @ metadata  :List of 4
#>  .. .. .. .. $ title      : chr "Separate CHIP Enrollment by Month and State Historic CAA/Unwinding Period"
#>  .. .. .. .. $ description: chr "<p>This historic dataset with total enrollment in separate CHIP programs by month and state was created to fulf"| __truncated__
#>  .. .. .. .. $ modified   : Date[1:1], format: "2024-12-27"
#>  .. .. .. .. $ periodicity: chr "Monthly [R/P1M]"
#>  .. .. .. @ dimensions: <class_dimensions>
#>  .. .. .. .. @ limit : int 8000
#>  .. .. .. .. @ rows  : int 780
#>  .. .. .. .. @ pages : int 1
#>  .. .. .. .. @ fields: chr [1:5] "state" "reporting_period" "release_date" "value" ...
#>  .. $ wind_sbm     : <class_caid>
#>  ..  ..@ access: <class_current>
#>  .. .. .. @ identifier: chr "https://data.medicaid.gov/api/1/datastore/query/5670e72c-e44e-4282-ab67-4ebebaba3cbd/0?count=true&results=true&"| __truncated__
#>  .. .. .. @ metadata  :List of 4
#>  .. .. .. .. $ title      : chr "State-based Marketplace (SBM) Medicaid Unwinding Report"
#>  .. .. .. .. $ description: chr "Metrics from individual Marketplaces during the current reporting period. The report includes data for the stat"| __truncated__
#>  .. .. .. .. $ modified   : Date[1:1], format: "2024-12-27"
#>  .. .. .. .. $ periodicity: chr "Decennially [R/P10Y]"
#>  .. .. .. @ dimensions: <class_dimensions>
#>  .. .. .. .. @ limit : int 8000
#>  .. .. .. .. @ rows  : int 128
#>  .. .. .. .. @ pages : int 1
#>  .. .. .. .. @ fields: chr [1:59] "state" "metric" "april_2023_count" "april_2023_percentage_applications" ...
collection("managed")
#> <class_collection>
#>  @ name   : chr "Managed Care Enrollment"
#>  @ members:List of 12
#>  .. $ man_mltss     : <class_caid>
#>  ..  ..@ access: <class_current>
#>  .. .. .. @ identifier: chr "https://data.medicaid.gov/api/1/datastore/query/5394bcab-c748-5e4b-af07-b5bf77ed3aa3/0?count=true&results=true&"| __truncated__
#>  .. .. .. @ metadata  :List of 4
#>  .. .. .. .. $ title      : chr "Managed Long Term Services and Supports (MLTSS) Enrollees"
#>  .. .. .. .. $ description: chr "1. Enrollment includes both Medicaid-only and Medicare-Medicaid (dual) enrollees. For both types of enrollees, "| __truncated__
#>  .. .. .. .. $ modified   : Date[1:1], format: "2024-10-16"
#>  .. .. .. .. $ periodicity: chr "Decennially [R/P10Y]"
#>  .. .. .. @ dimensions: <class_dimensions>
#>  .. .. .. .. @ limit : int 8000
#>  .. .. .. .. @ rows  : int 399
#>  .. .. .. .. @ pages : int 1
#>  .. .. .. .. @ fields: chr [1:8] "state" "notes" "total_any_managed_care_enrollees" "comprehensive_managed_care_ltss_enrollees" ...
#>  .. $ man_summary   : <class_caid>
#>  ..  ..@ access: <class_current>
#>  .. .. .. @ identifier: chr "https://data.medicaid.gov/api/1/datastore/query/52ed908b-0cb8-5dd2-846d-99d4af12b369/0?count=true&results=true&"| __truncated__
#>  .. .. .. @ metadata  :List of 4
#>  .. .. .. .. $ title      : chr "Managed Care Enrollment Summary"
#>  .. .. .. .. $ description: chr "The Medicaid Managed Care Enrollment Report profiles enrollment statistics on Medicaid managed care programs on"| __truncated__
#>  .. .. .. .. $ modified   : Date[1:1], format: "2024-10-16"
#>  .. .. .. .. $ periodicity: chr "Decennially [R/P10Y]"
#>  .. .. .. @ dimensions: <class_dimensions>
#>  .. .. .. .. @ limit : int 8000
#>  .. .. .. .. @ rows  : int 399
#>  .. .. .. .. @ pages : int 1
#>  .. .. .. .. @ fields: chr [1:7] "state" "notes" "total_medicaid_enrollees" "total_medicaid_enrollment_in_any_type_of_managed_care" ...
#>  .. $ man_state_end : <class_caid>
#>  ..  ..@ access: <class_current>
#>  .. .. .. @ identifier: chr "https://data.medicaid.gov/api/1/datastore/query/0e112ea8-8e8e-5dee-a7e2-7ed551c3baa4/0?count=true&results=true&"| __truncated__
#>  .. .. .. @ metadata  :List of 4
#>  .. .. .. .. $ title      : chr "Managed Care Programs by State"
#>  .. .. .. .. $ description: chr "Provides program names, program features, population enrolled, benefits covered, quality assurance and improvem"| __truncated__
#>  .. .. .. .. $ modified   : Date[1:1], format: "2018-08-22"
#>  .. .. .. .. $ periodicity: chr "Decennially [R/P10Y]"
#>  .. .. .. @ dimensions: <class_dimensions>
#>  .. .. .. .. @ limit : int 8000
#>  .. .. .. .. @ rows  : int 195
#>  .. .. .. .. @ pages : int 1
#>  .. .. .. .. @ fields: chr [1:69] "state" "program_name" "program_type" "statewide_or_regionspecific" ...
#>  .. $ man_program   : <class_caid>
#>  ..  ..@ access: <class_current>
#>  .. .. .. @ identifier: chr "https://data.medicaid.gov/api/1/datastore/query/0bef7b8a-c663-5b14-9a46-0b5c2b86b0fe/0?count=true&results=true&"| __truncated__
#>  .. .. .. @ metadata  :List of 4
#>  .. .. .. .. $ title      : chr "Managed Care Enrollment by Program and Plan"
#>  .. .. .. .. $ description: chr "The Medicaid Managed Care Enrollment Report profiles enrollment statistics on Medicaid managed care programs on"| __truncated__
#>  .. .. .. .. $ modified   : Date[1:1], format: "2024-10-16"
#>  .. .. .. .. $ periodicity: chr "Decennially [R/P10Y]"
#>  .. .. .. @ dimensions: <class_dimensions>
#>  .. .. .. .. @ limit : int 8000
#>  .. .. .. .. @ rows  : int 5974
#>  .. .. .. .. @ pages : int 1
#>  .. .. .. .. @ fields: chr [1:10] "state" "notes" "program_name" "plan_name" ...
#>  .. $ man_pop       : <class_caid>
#>  ..  ..@ access: <class_current>
#>  .. .. .. @ identifier: chr "https://data.medicaid.gov/api/1/datastore/query/e2ce0d2f-07c5-5213-947a-31e19bc649f6/0?count=true&results=true&"| __truncated__
#>  .. .. .. @ metadata  :List of 4
#>  .. .. .. .. $ title      : chr "Managed Care Enrollment by Program and Population (All)"
#>  .. .. .. .. $ description: chr "The Medicaid Managed Care Enrollment Report profiles enrollment statistics on Medicaid managed care programs on"| __truncated__
#>  .. .. .. .. $ modified   : Date[1:1], format: "2024-10-16"
#>  .. .. .. .. $ periodicity: chr "Decennially [R/P10Y]"
#>  .. .. .. @ dimensions: <class_dimensions>
#>  .. .. .. .. @ limit : int 8000
#>  .. .. .. .. @ rows  : int 401
#>  .. .. .. .. @ pages : int 1
#>  .. .. .. .. @ fields: chr [1:13] "state" "notes" "total_medicaid_enrollees" "comprehensive_mco_with_or_without_mltss" ...
#>  .. $ man_dual      : <class_caid>
#>  ..  ..@ access: <class_current>
#>  .. .. .. @ identifier: chr "https://data.medicaid.gov/api/1/datastore/query/8e7be65b-97ba-5ecf-8394-ca8e6f63685a/0?count=true&results=true&"| __truncated__
#>  .. .. .. @ metadata  :List of 4
#>  .. .. .. .. $ title      : chr "Managed Care Enrollment by Program and Population (Duals)"
#>  .. .. .. .. $ description: chr "The Medicaid Managed Care Enrollment Report profiles enrollment statistics on Medicaid managed care programs on"| __truncated__
#>  .. .. .. .. $ modified   : Date[1:1], format: "2024-10-16"
#>  .. .. .. .. $ periodicity: chr "Decennially [R/P10Y]"
#>  .. .. .. @ dimensions: <class_dimensions>
#>  .. .. .. .. @ limit : int 8000
#>  .. .. .. .. @ rows  : int 399
#>  .. .. .. .. @ pages : int 1
#>  .. .. .. .. @ fields: chr [1:13] "state" "notes" "total_beneficiaries" "comprehensive_mco_with_or_without_mltss" ...
#>  .. $ man_feat_pop  : <class_caid>
#>  ..  ..@ access: <class_current>
#>  .. .. .. @ identifier: chr "https://data.medicaid.gov/api/1/datastore/query/927f4847-2c0a-50c1-8f50-9103de7d048b/0?count=true&results=true&"| __truncated__
#>  .. .. .. @ metadata  :List of 4
#>  .. .. .. .. $ title      : chr "Managed Care Features By Enrollment Population"
#>  .. .. .. .. $ description: chr "Number of Managed Care Programs Enrolling Certain Populations on a Mandatory or Voluntary Basis at any point in 2022"
#>  .. .. .. .. $ modified   : Date[1:1], format: "2024-10-16"
#>  .. .. .. .. $ periodicity: chr "Decennially [R/P10Y]"
#>  .. .. .. @ dimensions: <class_dimensions>
#>  .. .. .. .. @ limit : int 8000
#>  .. .. .. .. @ rows  : int 74
#>  .. .. .. .. @ pages : int 1
#>  .. .. .. .. @ fields: chr [1:21] "features" "notes" "comprehensive_mco_with_or_without_mltss" "comprehensive_mco_with_or_without_mltss_2" ...
#>  .. $ man_feat_qa   : <class_caid>
#>  ..  ..@ access: <class_current>
#>  .. .. .. @ identifier: chr "https://data.medicaid.gov/api/1/datastore/query/a6d2c261-210d-587c-8fa4-892444b7de42/0?count=true&results=true&"| __truncated__
#>  .. .. .. @ metadata  :List of 4
#>  .. .. .. .. $ title      : chr "Managed Care Features by QA and Performance Incentive"
#>  .. .. .. .. $ description: chr "Number of Managed Care Program Types, by Quality Assurance Requirements, Performance Incentives, and Provider V"| __truncated__
#>  .. .. .. .. $ modified   : Date[1:1], format: "2024-10-16"
#>  .. .. .. .. $ periodicity: chr "Decennially [R/P10Y]"
#>  .. .. .. @ dimensions: <class_dimensions>
#>  .. .. .. .. @ limit : int 8000
#>  .. .. .. .. @ rows  : int 63
#>  .. .. .. .. @ pages : int 1
#>  .. .. .. .. @ fields: chr [1:11] "features" "comprehensive_mco_with_or_without_mltss" "pccm" "pccm_entity" ...
#>  .. $ man_bene_month: <class_caid>
#>  ..  ..@ access: <class_current>
#>  .. .. .. @ identifier: chr "https://data.medicaid.gov/api/1/datastore/query/89baf100-259b-4763-b9e2-337972f988c4/0?count=true&results=true&"| __truncated__
#>  .. .. .. @ metadata  :List of 5
#>  .. .. .. .. $ title      : chr "Managed Care Information for Medicaid and CHIP Beneficiaries by Month"
#>  .. .. .. .. $ description: chr "This data set includes monthly enrollment counts of Medicaid and CHIP beneficiaries by managed care participati"| __truncated__
#>  .. .. .. .. $ modified   : Date[1:1], format: "2024-01-05"
#>  .. .. .. .. $ periodicity: chr "Decennially [R/P10Y]"
#>  .. .. .. .. $ download   : chr "https://download.medicaid.gov/data/ManagedCare-montly.csv"
#>  .. .. .. @ dimensions: <class_dimensions>
#>  .. .. .. .. @ limit : int 8000
#>  .. .. .. .. @ rows  : int 31164
#>  .. .. .. .. @ pages : int 4
#>  .. .. .. .. @ fields: chr [1:5] "state" "month" "managedcareparticipation" "countenrolled" ...
#>  .. $ man_bene_year : <class_caid>
#>  ..  ..@ access: <class_current>
#>  .. .. .. @ identifier: chr "https://data.medicaid.gov/api/1/datastore/query/c0bef49e-46ea-4b9d-9153-76b4078d58b7/0?count=true&results=true&"| __truncated__
#>  .. .. .. @ metadata  :List of 5
#>  .. .. .. .. $ title      : chr "Managed Care Information for Medicaid and CHIP Beneficiaries by Year"
#>  .. .. .. .. $ description: chr "This data set presents annual enrollment counts of Medicaid and CHIP beneficiaries by managed care participatio"| __truncated__
#>  .. .. .. .. $ modified   : Date[1:1], format: "2023-03-20"
#>  .. .. .. .. $ periodicity: chr "Decennially [R/P10Y]"
#>  .. .. .. .. $ download   : chr "https://download.medicaid.gov/data/ManagedCare-anul.csv"
#>  .. .. .. @ dimensions: <class_dimensions>
#>  .. .. .. .. @ limit : int 8000
#>  .. .. .. .. @ rows  : int 2597
#>  .. .. .. .. @ pages : int 1
#>  .. .. .. .. @ fields: chr [1:7] "state" "year" "managedcareparticipation" "counteverenrolled" ...
#>  .. $ man_share     : <class_caid>
#>  ..  ..@ access: <class_current>
#>  .. .. .. @ identifier: chr "https://data.medicaid.gov/api/1/datastore/query/79692ea5-21e1-56bf-8149-97d437120c4b/0?count=true&results=true&"| __truncated__
#>  .. .. .. @ metadata  :List of 4
#>  .. .. .. .. $ title      : chr "Share of Medicaid Enrollees in Managed Care"
#>  .. .. .. .. $ description: chr "The Share of Medicaid Enrollees in any Managed Care and in Comprehensive Managed CaAre profiles state-level enr"| __truncated__
#>  .. .. .. .. $ modified   : Date[1:1], format: "2024-10-16"
#>  .. .. .. .. $ periodicity: chr "Decennially [R/P10Y]"
#>  .. .. .. @ dimensions: <class_dimensions>
#>  .. .. .. .. @ limit : int 8000
#>  .. .. .. .. @ rows  : int 399
#>  .. .. .. .. @ pages : int 1
#>  .. .. .. .. @ fields: chr [1:8] "state" "notes" "total_medicaid_enrollees" "individuals_enrolled_any" ...
#>  .. $ man_tmp       : <class_caid>
#>  ..  ..@ access: <class_temporal>
#>  .. .. .. @ identifier: tibble [1 × 4] (S3: tbl_df/tbl/data.frame)
#>  $ year      : int 2017
#>  $ modified  : Date[1:1], format: "2019-05-03"
#>  $ identifier: chr "https://data.medicaid.gov/api/1/datastore/query/8f39b637-9bb1-5894-9062-2c4f2ad70fba/0?count=true&results=true&"| __truncated__
#>  $ download  : chr NA
#>  .. .. .. @ metadata  :List of 3
#>  .. .. .. .. $ title      : chr "Managed Care Programs by State"
#>  .. .. .. .. $ description: chr "Provides program names, program features, population enrolled, benefits covered, quality assurance and improvem"| __truncated__
#>  .. .. .. .. $ periodicity: chr "Decennially [R/P10Y]"
#>  .. .. .. @ dimensions: <class_dimensions>
#>  .. .. .. .. @ limit : int 8000
#>  .. .. .. .. @ rows  : int 175
#>  .. .. .. .. @ pages : int 1
#>  .. .. .. .. @ fields: chr [1:61] "state" "program_name" "program_type" "statewide_or_regionspecific" ...