Load a collection of endpoints by alias

Load a collection of endpoints by alias

Usage

collection(alias)

Arguments

alias

<chr> collection alias

Value

An S7 <class_collection> object.

Examples

collection("caid_unwind")
#> <class_collection>
#>  @ name   : chr "Medicaid Unwinding Report"
#>  @ members:List of 4
#>  .. $ unwind_market    : <class_current>
#>  ..  ..@ access    : <class_caid>
#>  .. .. .. @ identifier: chr "https://data.medicaid.gov/api/1/datastore/query/9a83ba5e-05f5-47f5-82de-f3a59233a912/0?count=true&results=true&"| __truncated__
#>  ..  ..@ metadata  :List of 5
#>  .. .. .. $ title      : chr "HealthCare.gov Marketplace Medicaid Unwinding Report"
#>  .. .. .. $ description: chr "Metrics from individual Marketplaces during the current reporting period. The report includes data for the stat"| __truncated__
#>  .. .. .. $ modified   : Date[1:1], format: "2024-08-30"
#>  .. .. .. $ periodicity: chr "Decennially [R/P10Y]"
#>  .. .. .. $ download   : chr NA
#>  ..  ..@ dimensions: <class_dimensions>
#>  .. .. .. @ limit : int 8000
#>  .. .. .. @ rows  : int 204
#>  .. .. .. @ pages : int 1
#>  .. .. .. @ fields:List of 30
#>  .. .. .. .. $ state                           : chr NA
#>  .. .. .. .. $ metric                          : chr NA
#>  .. .. .. .. $ april_2023_net_count            : chr NA
#>  .. .. .. .. $ april_2023_net_percentage       : chr NA
#>  .. .. .. .. $ may_2023_net_count              : chr NA
#>  .. .. .. .. $ may_2023_net_percentage         : chr NA
#>  .. .. .. .. $ june_2023_net_count             : chr NA
#>  .. .. .. .. $ june_2023_net_percentage        : chr NA
#>  .. .. .. .. $ july_2023_net_count             : chr NA
#>  .. .. .. .. $ july_2023_net_percentage        : chr NA
#>  .. .. .. .. $ august_2023_net_count           : chr NA
#>  .. .. .. .. $ august_2023_net_percentage      : chr NA
#>  .. .. .. .. $ september_2023_net_count        : chr NA
#>  .. .. .. .. $ september_2023_net_percentage   : chr NA
#>  .. .. .. .. $ october_2023_net_count          : chr NA
#>  .. .. .. .. $ october_2023_net_percentage     : chr NA
#>  .. .. .. .. $ november_2023_net_count         : chr NA
#>  .. .. .. .. $ november_2023_net_percentage    : chr NA
#>  .. .. .. .. $ december_2023_net_count         : chr NA
#>  .. .. .. .. $ december_2023_net_percentage    : chr NA
#>  .. .. .. .. $ january_2024_net_count          : chr NA
#>  .. .. .. .. $ january_2024_net_percentage     : chr NA
#>  .. .. .. .. $ february_2024_net_count         : chr NA
#>  .. .. .. .. $ february_2024_net_percentage    : chr NA
#>  .. .. .. .. $ march_2024_net_count            : chr NA
#>  .. .. .. .. $ march_2024_net_percentage       : chr NA
#>  .. .. .. .. $ april_2024_net_count            : chr NA
#>  .. .. .. .. $ april_2024_net_percentage       : chr NA
#>  .. .. .. .. $ april_2024_cumulative_count     : chr NA
#>  .. .. .. .. $ april_cumulative_2024_percentage: chr NA
#>  .. $ unwind_transition: <class_current>
#>  ..  ..@ access    : <class_caid>
#>  .. .. .. @ identifier: chr "https://data.medicaid.gov/api/1/datastore/query/5636a78c-fe18-4229-aee1-e40fa910a8a0/0?count=true&results=true&"| __truncated__
#>  ..  ..@ metadata  :List of 5
#>  .. .. .. $ title      : chr "HealthCare.gov Transitions Marketplace Medicaid Unwinding Report"
#>  .. .. .. $ description: chr "Metrics from individual Marketplaces during the current reporting period. The report includes data for the stat"| __truncated__
#>  .. .. .. $ modified   : Date[1:1], format: "2024-12-27"
#>  .. .. .. $ periodicity: chr "Decennially [R/P10Y]"
#>  .. .. .. $ download   : chr NA
#>  ..  ..@ dimensions: <class_dimensions>
#>  .. .. .. @ limit : int 8000
#>  .. .. .. @ rows  : int 59527
#>  .. .. .. @ pages : int 8
#>  .. .. .. @ fields:List of 9
#>  .. .. .. .. $ state                                                           : chr NA
#>  .. .. .. .. $ data_record_note                                                : chr NA
#>  .. .. .. .. $ metric                                                          : chr NA
#>  .. .. .. .. $ time_period                                                     : chr NA
#>  .. .. .. .. $ release_through                                                 : chr NA
#>  .. .. .. .. $ count_of_individuals_whose_medicaid_or_chip_coverage_was_te_dd11: chr NA
#>  .. .. .. .. $ percentage_of_individuals_whose_medicaid_or_chip_coverage_w_da6e: chr NA
#>  .. .. .. .. $ cumulative_count_of_individuals_whose_medicaid_or_chip_cove_28d9: chr NA
#>  .. .. .. .. $ cumulative_percentage_of_individuals_whose_medicaid_or_chip_6ac2: chr NA
#>  .. $ unwind_historic  : <class_current>
#>  ..  ..@ access    : <class_caid>
#>  .. .. .. @ identifier: chr "https://data.medicaid.gov/api/1/datastore/query/d30cfc7c-4b32-4df1-b2bf-e0a850befd77/0?count=true&results=true&"| __truncated__
#>  ..  ..@ metadata  :List of 5
#>  .. .. .. $ title      : chr "Separate CHIP Enrollment by Month and State Historic CAA/Unwinding Period"
#>  .. .. .. $ description: chr "<p>This historic dataset with total enrollment in separate CHIP programs by month and state was created to fulf"| __truncated__
#>  .. .. .. $ modified   : Date[1:1], format: "2024-12-27"
#>  .. .. .. $ periodicity: chr "Monthly [R/P1M]"
#>  .. .. .. $ download   : chr NA
#>  ..  ..@ dimensions: <class_dimensions>
#>  .. .. .. @ limit : int 8000
#>  .. .. .. @ rows  : int 780
#>  .. .. .. @ pages : int 1
#>  .. .. .. @ fields:List of 5
#>  .. .. .. .. $ state           : chr NA
#>  .. .. .. .. $ reporting_period: chr NA
#>  .. .. .. .. $ release_date    : chr NA
#>  .. .. .. .. $ value           : chr NA
#>  .. .. .. .. $ data_notes      : chr NA
#>  .. $ unwind_sbm       : <class_current>
#>  ..  ..@ access    : <class_caid>
#>  .. .. .. @ identifier: chr "https://data.medicaid.gov/api/1/datastore/query/5670e72c-e44e-4282-ab67-4ebebaba3cbd/0?count=true&results=true&"| __truncated__
#>  ..  ..@ metadata  :List of 5
#>  .. .. .. $ title      : chr "State-based Marketplace (SBM) Medicaid Unwinding Report"
#>  .. .. .. $ description: chr "Metrics from individual Marketplaces during the current reporting period. The report includes data for the stat"| __truncated__
#>  .. .. .. $ modified   : Date[1:1], format: "2024-12-27"
#>  .. .. .. $ periodicity: chr "Decennially [R/P10Y]"
#>  .. .. .. $ download   : chr NA
#>  ..  ..@ dimensions: <class_dimensions>
#>  .. .. .. @ limit : int 8000
#>  .. .. .. @ rows  : int 128
#>  .. .. .. @ pages : int 1
#>  .. .. .. @ fields:List of 59
#>  .. .. .. .. $ state                                      : chr NA
#>  .. .. .. .. $ metric                                     : chr NA
#>  .. .. .. .. $ april_2023_count                           : chr NA
#>  .. .. .. .. $ april_2023_percentage_applications         : chr NA
#>  .. .. .. .. $ april_2023_percentage_account_transfers    : chr NA
#>  .. .. .. .. $ may_2023_count                             : chr NA
#>  .. .. .. .. $ may_2023_percentage_applications           : chr NA
#>  .. .. .. .. $ may_2023_percentage_account_transfers      : chr NA
#>  .. .. .. .. $ june_2023_count                            : chr NA
#>  .. .. .. .. $ june_2023_percentage_applications          : chr NA
#>  .. .. .. .. $ june_2023_percentage_account_transfers     : chr NA
#>  .. .. .. .. $ july_2023_count                            : chr NA
#>  .. .. .. .. $ july_2023_percentage_applications          : chr NA
#>  .. .. .. .. $ july_2023_percentage_account_transfers     : chr NA
#>  .. .. .. .. $ august_2023_count                          : chr NA
#>  .. .. .. .. $ august_2023_percentage_applications        : chr NA
#>  .. .. .. .. $ august_2023_percentage_account_transfers   : chr NA
#>  .. .. .. .. $ september_2023_count                       : chr NA
#>  .. .. .. .. $ september_2023_percentage_applications     : chr NA
#>  .. .. .. .. $ september_2023_percentage_account_transfers: chr NA
#>  .. .. .. .. $ october_2023_count                         : chr NA
#>  .. .. .. .. $ october_2023_percentage_applications       : chr NA
#>  .. .. .. .. $ october_2023_percentage_account_transfers  : chr NA
#>  .. .. .. .. $ november_2023_count                        : chr NA
#>  .. .. .. .. $ november_2023_percentage_applications      : chr NA
#>  .. .. .. .. $ november_2023_percentage_account_transfers : chr NA
#>  .. .. .. .. $ december_2023_count                        : chr NA
#>  .. .. .. .. $ december_2023_percentage_applications      : chr NA
#>  .. .. .. .. $ december_2023_percentage_account_transfers : chr NA
#>  .. .. .. .. $ january_2024_count                         : chr NA
#>  .. .. .. .. $ january_2024_percentage_applications       : chr NA
#>  .. .. .. .. $ january_2024_percentage_account_transfers  : chr NA
#>  .. .. .. .. $ february_2024_count                        : chr NA
#>  .. .. .. .. $ february_2024_percentage_applications      : chr NA
#>  .. .. .. .. $ february_2024_percentage_account_transfers : chr NA
#>  .. .. .. .. $ march_2024_count                           : chr NA
#>  .. .. .. .. $ march_2024_percentage_applications         : chr NA
#>  .. .. .. .. $ march_2024_percentage_account_transfers    : chr NA
#>  .. .. .. .. $ april_2024_count                           : chr NA
#>  .. .. .. .. $ april_2024_percentage_applications         : chr NA
#>  .. .. .. .. $ april_2024_percentage_account_transfers    : chr NA
#>  .. .. .. .. $ may_2024_count                             : chr NA
#>  .. .. .. .. $ may_2024_percentage_applications           : chr NA
#>  .. .. .. .. $ may_2024_percentage_account_transfers      : chr NA
#>  .. .. .. .. $ june_2024_count                            : chr NA
#>  .. .. .. .. $ june_2024_percentage_applications          : chr NA
#>  .. .. .. .. $ june_2024_percentage_account_transfers     : chr NA
#>  .. .. .. .. $ july_2024_count                            : chr NA
#>  .. .. .. .. $ july_2024_percentage_applications          : chr NA
#>  .. .. .. .. $ july_2024_percentage_account_transfers     : chr NA
#>  .. .. .. .. $ august_2024_count                          : chr NA
#>  .. .. .. .. $ august_2024_percentage_applications        : chr NA
#>  .. .. .. .. $ august_2024_percentage_account_transfers   : chr NA
#>  .. .. .. .. $ september_2024_count                       : chr NA
#>  .. .. .. .. $ september_2024_percentage_applications     : chr NA
#>  .. .. .. .. $ september_2024_percentage_account_transfers: chr NA
#>  .. .. .. .. $ total_count                                : chr NA
#>  .. .. .. .. $ total_percentage_applications              : chr NA
#>  .. .. .. .. $ total_percentage_account_transfers         : chr NA