Load an endpoint by alias

Load an endpoint by alias

Usage

endpoint(alias)

Arguments

alias

<chr> endpoint alias

Value

An S7 <class_care/caid/prov/open/hgov> object.

Examples

endpoint("dial_facility")
#> <class_care>
#>  @ access: <care_temporal>
#>  .. @ identifier: tibble [6 × 6] (S3: tbl_df/tbl/data.frame)
#>  $ year      : int [1:6] 2025 2024 2023 2022 2021 2020
#>  $ modified  : Date[1:6], format: "2025-06-24" "2024-02-29" ...
#>  $ temporal  : chr [1:6] "2025-01-01 • 2025-12-31" "2024-01-01 • 2024-12-31" "2023-01-01 • 2023-12-31" "2022-01-01 • 2022-12-31" ...
#>  $ identifier: chr [1:6] "https://data.cms.gov/data-api/v1/dataset/7d4e7bd1-f08a-4fca-8085-6af24026a19a/data?offset=0&size=1" "https://data.cms.gov/data-api/v1/dataset/62d125ff-c959-4c5c-8a46-8cacba50bfd4/data?offset=0&size=1" "https://data.cms.gov/data-api/v1/dataset/200e7e7e-a5a6-4450-ba18-ab53d237cc45/data?offset=0&size=1" "https://data.cms.gov/data-api/v1/dataset/ebc64ad1-164c-455a-80ca-cb0753e44a46/data?offset=0&size=1" ...
#>  $ download  : chr [1:6] "https://data.cms.gov/sites/default/files/2024-02/b2c2658d-e19d-48db-8b3b-9d343a4d395f/dfr_facility_socrata_fy2024.csv" "https://data.cms.gov/sites/default/files/2023-02/2eeb2aa7-05cd-4b7a-a4b2-62e82da03b96/dfr_facility_socrata_fy2023.csv" "https://data.cms.gov/sites/default/files/2022-02/48802694-7450-47c3-a61f-dad7d392037c/dfr_facility_socrata_fy2022.csv" "https://data.cms.gov/sites/default/files/2021-01/FY_2021_Facility_Level_Dialysis_Facility_Reports.csv" ...
#>  $ resources : chr [1:6] "https://data.cms.gov/data-api/v1/dataset-resources/7d4e7bd1-f08a-4fca-8085-6af24026a19a" "https://data.cms.gov/data-api/v1/dataset-resources/62d125ff-c959-4c5c-8a46-8cacba50bfd4" "https://data.cms.gov/data-api/v1/dataset-resources/200e7e7e-a5a6-4450-ba18-ab53d237cc45" "https://data.cms.gov/data-api/v1/dataset-resources/ebc64ad1-164c-455a-80ca-cb0753e44a46" ...
#>  .. @ metadata  :List of 6
#>  .. .. $ title      : chr "Medicare Dialysis Facilities"
#>  .. .. $ description: chr "The Medicare Dialysis Facilities data provides information on clinical and patient measures for Medicare-certif"| __truncated__
#>  .. .. $ periodicity: chr "Annually [R/P1Y]"
#>  .. .. $ dictionary : chr "https://data.cms.gov/sites/default/files/2025-03/FY2025%20PUF%20Data%20Dictionary.xlsx"
#>  .. .. $ site       : chr "https://data.cms.gov/quality-of-care/medicare-dialysis-facilities"
#>  .. .. $ references : chr "https://data.cms.gov/sites/default/files/2025-03/FY2025_DFR_Guide.pdf"
#>  .. @ dimensions: <class_dimensions>
#>  .. .. @ limit : int 5000
#>  .. .. @ rows  : int 11985275
#>  .. .. @ pages : int 2398
#>  .. .. @ fields: chr [1:14] "state" "CCN" "Provider_Name" "city" ...
endpoint("man_mltss")
#> <class_caid>
#>  @ access: <class_current>
#>  .. @ identifier: chr "https://data.medicaid.gov/api/1/datastore/query/5394bcab-c748-5e4b-af07-b5bf77ed3aa3/0?count=true&results=true&"| __truncated__
#>  .. @ metadata  :List of 4
#>  .. .. $ title      : chr "Managed Long Term Services and Supports (MLTSS) Enrollees"
#>  .. .. $ description: chr "1. Enrollment includes both Medicaid-only and Medicare-Medicaid (dual) enrollees. For both types of enrollees, "| __truncated__
#>  .. .. $ modified   : Date[1:1], format: "2024-10-16"
#>  .. .. $ periodicity: chr "Decennially [R/P10Y]"
#>  .. @ dimensions: <class_dimensions>
#>  .. .. @ limit : int 8000
#>  .. .. @ rows  : int 399
#>  .. .. @ pages : int 1
#>  .. .. @ fields: chr [1:8] "state" "notes" ...
endpoint("ab_reg_comp")
#> <class_hgov>
#>  @ access: <class_current>
#>  .. @ identifier: chr "https://data.healthcare.gov/api/1/datastore/query/wb6u-x2ny/0?count=true&results=true&offset=0&limit=1"
#>  .. @ metadata  :List of 6
#>  .. .. $ title      : chr "AB Registration Completion List"
#>  .. .. $ description: chr "Access Data Dictionary Here - https://www.cms.gov/CCIIO/Programs-and-Initiatives/Health-Insurance-Marketplaces/"| __truncated__
#>  .. .. $ modified   : Date[1:1], format: "2025-07-25"
#>  .. .. $ issued     : Date[1:1], format: "2016-03-30"
#>  .. .. $ periodicity: chr "Decennially [R/P10Y]"
#>  .. .. $ download   : chr "https://data.healthcare.gov/sites/default/files/uploaded_resources/TBL_RCL_179.csv"
#>  .. @ dimensions: <class_dimensions>
#>  .. .. @ limit : int 500
#>  .. .. @ rows  : int 1025734
#>  .. .. @ pages : int 2052
#>  .. .. @ fields: chr [1:7] "npn" "applicable_plan_year" ...
endpoint("asc_facility")
#> <class_prov>
#>  @ access: <class_current>
#>  .. @ identifier: chr "https://data.cms.gov/provider-data/api/1/datastore/query/4jcv-atw7/0?count=true&results=true&offset=0&limit=1"
#>  .. @ metadata  :List of 9
#>  .. .. $ title      : chr "Ambulatory Surgical Center Quality Measures - Facility"
#>  .. .. $ description: chr "A list of ambulatory surgical centers participating in the Ambulatory Surgical Center Quality Reporting (ASCQR)"| __truncated__
#>  .. .. $ modified   : Date[1:1], format: "2025-04-08"
#>  .. .. $ group      : chr "Hospitals"
#>  .. .. $ issued     : Date[1:1], format: "2024-10-10"
#>  .. .. $ released   : Date[1:1], format: "2025-04-30"
#>  .. .. $ download   : chr "https://data.cms.gov/provider-data/sites/default/files/resources/dd03994fc93e296bb0297f1cd43cc987_1744668308/ASC_Facility.csv"
#>  .. .. $ dictionary : chr "https://data.cms.gov/provider-data/dataset/4jcv-atw7#data-dictionary"
#>  .. .. $ site       : chr "https://data.cms.gov/provider-data/dataset/4jcv-atw7"
#>  .. @ dimensions: <class_dimensions>
#>  .. .. @ limit : int 1500
#>  .. .. @ rows  : int 11288
#>  .. .. @ pages : int 8
#>  .. .. @ fields: chr [1:50] "facility_name" "facility_id" "npi" "citytown" ...
endpoint("dial_listing")
#> <class_prov>
#>  @ access: <class_current>
#>  .. @ identifier: chr "https://data.cms.gov/provider-data/api/1/datastore/query/23ew-n7w9/0?count=true&results=true&offset=0&limit=1"
#>  .. @ metadata  :List of 9
#>  .. .. $ title      : chr "Dialysis Facility - Listing by Facility"
#>  .. .. $ description: chr "A list of all dialysis facilities registered with Medicare that includes addresses and phone numbers, as well a"| __truncated__
#>  .. .. $ modified   : Date[1:1], format: "2025-06-17"
#>  .. .. $ group      : chr "Dialysis facilities"
#>  .. .. $ issued     : Date[1:1], format: "2020-03-14"
#>  .. .. $ released   : Date[1:1], format: "2025-07-16"
#>  .. .. $ download   : chr "https://data.cms.gov/provider-data/sites/default/files/resources/c04d84bc5c641284494bee4f20f17f9c_1750176303/DFC_FACILITY.csv"
#>  .. .. $ dictionary : chr "https://data.cms.gov/provider-data/dataset/23ew-n7w9#data-dictionary"
#>  .. .. $ site       : chr "https://data.cms.gov/provider-data/dataset/23ew-n7w9"
#>  .. @ dimensions: <class_dimensions>
#>  .. .. @ limit : int 1500
#>  .. .. @ rows  : int 7561
#>  .. .. @ pages : int 6
#>  .. .. @ fields: chr [1:142] "cms_certification_number_ccn" "network" ...