Load an endpoint by alias

Load an endpoint by alias

Usage

endpoint(alias)

Arguments

alias

<chr> endpoint alias

Value

An S7 <class_care/caid/prov/open/hgov> object.

Examples

endpoint("care_dial_end")
#> <class_care>
#>  @ metadata  :List of 10
#>  .. $ title      : chr "Medicare Dialysis Facilities"
#>  .. $ description: chr "The Medicare Dialysis Facilities data provides information on clinical and patient measures for Medicare-certif"| __truncated__
#>  .. $ modified   : Date[1:1], format: "2025-06-24"
#>  .. $ temporal   : chr "2009-01-01 • 2025-12-31"
#>  .. $ periodicity: chr "Annually [R/P1Y]"
#>  .. $ download   : chr "https://data.cms.gov/sites/default/files/2025-04/94d61c1e-b20f-47a9-b236-521016b54b0f/dfr_facility_socrata_fy2025_resubmit.csv"
#>  .. $ resources  : chr "https://data.cms.gov/data-api/v1/dataset-resources/f8610e87-ba25-43a3-a49e-927dbc8701ae"
#>  .. $ dictionary : chr "https://data.cms.gov/sites/default/files/2025-03/FY2025%20PUF%20Data%20Dictionary.xlsx"
#>  .. $ site       : chr "https://data.cms.gov/quality-of-care/medicare-dialysis-facilities"
#>  .. $ references : chr "https://data.cms.gov/sites/default/files/2025-03/FY2025_DFR_Guide.pdf"
#>  @ dimensions: <class_dimensions>
#>  .. @ limit : int 5000
#>  .. @ rows  : int 11985275
#>  .. @ pages : int 2398
#>  .. @ fields:List of 14
#>  .. .. $ state         : chr(0) 
#>  .. .. $ CCN           : chr(0) 
#>  .. .. $ Provider_Name : chr(0) 
#>  .. .. $ city          : chr(0) 
#>  .. .. $ Ownership_Type: chr(0) 
#>  .. .. $ ESRD_Network  : chr(0) 
#>  .. .. $ NPI           : chr(0) 
#>  .. .. $ Chain         : chr(0) 
#>  .. .. $ Modality      : chr(0) 
#>  .. .. $ Alternate_CCNs: chr(0) 
#>  .. .. $ Measure       : chr(0) 
#>  .. .. $ Measure_Score : chr(0) 
#>  .. .. $ year          : chr(0) 
#>  .. .. $ Measure_ID    : chr(0) 
#>  @ identifier: <class_current> chr "https://data.cms.gov/data-api/v1/dataset/f8610e87-ba25-43a3-a49e-927dbc8701ae/data-viewer?offset=0&size=1"
#>  @ resources : <class_current> chr "https://data.cms.gov/data-api/v1/dataset-resources/f8610e87-ba25-43a3-a49e-927dbc8701ae"
endpoint("care_dial_tmp")
#> <class_care>
#>  @ metadata  :List of 6
#>  .. $ title      : chr "Medicare Dialysis Facilities"
#>  .. $ description: chr "The Medicare Dialysis Facilities data provides information on clinical and patient measures for Medicare-certif"| __truncated__
#>  .. $ periodicity: chr "Annually [R/P1Y]"
#>  .. $ dictionary : chr "https://data.cms.gov/sites/default/files/2025-03/FY2025%20PUF%20Data%20Dictionary.xlsx"
#>  .. $ site       : chr "https://data.cms.gov/quality-of-care/medicare-dialysis-facilities"
#>  .. $ references : chr "https://data.cms.gov/sites/default/files/2025-03/FY2025_DFR_Guide.pdf"
#>  @ dimensions: <class_dimensions>
#>  .. @ limit : int 5000
#>  .. @ rows  : int 11985275
#>  .. @ pages : int 2398
#>  .. @ fields:List of 14
#>  .. .. $ state         : chr(0) 
#>  .. .. $ CCN           : chr(0) 
#>  .. .. $ Provider_Name : chr(0) 
#>  .. .. $ city          : chr(0) 
#>  .. .. $ Ownership_Type: chr(0) 
#>  .. .. $ ESRD_Network  : chr(0) 
#>  .. .. $ NPI           : chr(0) 
#>  .. .. $ Chain         : chr(0) 
#>  .. .. $ Modality      : chr(0) 
#>  .. .. $ Alternate_CCNs: chr(0) 
#>  .. .. $ Measure       : chr(0) 
#>  .. .. $ Measure_Score : chr(0) 
#>  .. .. $ year          : chr(0) 
#>  .. .. $ Measure_ID    : chr(0) 
#>  @ identifier: <class_temporal> List of 5
#>  .. $ year      : int [1:6] 2025 2024 2023 2022 2021 2020
#>  .. $ modified  : Date[1:6], format: "2025-06-24" "2024-02-29" ...
#>  .. $ temporal  : chr [1:6] "2025-01-01 • 2025-12-31" "2024-01-01 • 2024-12-31" "2023-01-01 • 2023-12-31" "2022-01-01 • 2022-12-31" ...
#>  .. $ identifier: chr [1:6] "https://data.cms.gov/data-api/v1/dataset/7d4e7bd1-f08a-4fca-8085-6af24026a19a/data?offset=0&size=1" "https://data.cms.gov/data-api/v1/dataset/62d125ff-c959-4c5c-8a46-8cacba50bfd4/data?offset=0&size=1" "https://data.cms.gov/data-api/v1/dataset/200e7e7e-a5a6-4450-ba18-ab53d237cc45/data?offset=0&size=1" "https://data.cms.gov/data-api/v1/dataset/ebc64ad1-164c-455a-80ca-cb0753e44a46/data?offset=0&size=1" ...
#>  .. $ download  : chr [1:6] "https://data.cms.gov/sites/default/files/2024-02/b2c2658d-e19d-48db-8b3b-9d343a4d395f/dfr_facility_socrata_fy2024.csv" "https://data.cms.gov/sites/default/files/2023-02/2eeb2aa7-05cd-4b7a-a4b2-62e82da03b96/dfr_facility_socrata_fy2023.csv" "https://data.cms.gov/sites/default/files/2022-02/48802694-7450-47c3-a61f-dad7d392037c/dfr_facility_socrata_fy2022.csv" "https://data.cms.gov/sites/default/files/2021-01/FY_2021_Facility_Level_Dialysis_Facility_Reports.csv" ...
#>  @ resources : <class_temporal> List of 2
#>  .. $ year     : int [1:6] 2025 2024 2023 2022 2021 2020
#>  .. $ resources: chr [1:6] "https://data.cms.gov/data-api/v1/dataset-resources/7d4e7bd1-f08a-4fca-8085-6af24026a19a" "https://data.cms.gov/data-api/v1/dataset-resources/62d125ff-c959-4c5c-8a46-8cacba50bfd4" "https://data.cms.gov/data-api/v1/dataset-resources/200e7e7e-a5a6-4450-ba18-ab53d237cc45" "https://data.cms.gov/data-api/v1/dataset-resources/ebc64ad1-164c-455a-80ca-cb0753e44a46" ...
endpoint("managed_mltss")
#> <class_caid>
#>  @ metadata  :List of 5
#>  .. $ title      : chr "Managed Long Term Services and Supports (MLTSS) Enrollees"
#>  .. $ description: chr "1. Enrollment includes both Medicaid-only and Medicare-Medicaid (dual) enrollees. For both types of enrollees, "| __truncated__
#>  .. $ modified   : Date[1:1], format: "2024-10-16"
#>  .. $ periodicity: chr "Decennially [R/P10Y]"
#>  .. $ download   : chr "https://data.medicaid.gov/sites/default/files/uploaded_resources/managed-long-term-services-and-supports-mltss-"| __truncated__
#>  @ dimensions: <class_dimensions>
#>  .. @ limit : int 8000
#>  .. @ rows  : int 399
#>  .. @ pages : int 1
#>  .. @ fields:List of 8
#>  .. .. $ state                                    : chr(0) 
#>  .. .. $ notes                                    : chr(0) 
#>  .. .. $ total_any_managed_care_enrollees         : chr(0) 
#>  .. .. $ comprehensive_managed_care_ltss_enrollees: chr(0) 
#>  .. .. $ comprehensive_managed_care_ltss_percent  : chr(0) 
#>  .. .. $ managed_ltss_only_enrollees              : chr(0) 
#>  .. .. $ managed_ltss_only_percent                : chr(0) 
#>  .. .. $ year                                     : chr(0) 
#>  @ identifier: <class_current> chr "https://data.medicaid.gov/api/1/datastore/query/5394bcab-c748-5e4b-af07-b5bf77ed3aa3/0?count=true&results=true&"| __truncated__
endpoint("hgov_ab_reg_comp")
#> <class_hgov>
#>  @ metadata  :List of 7
#>  .. $ title      : chr "AB Registration Completion List"
#>  .. $ description: chr "Access Data Dictionary Here - https://www.cms.gov/CCIIO/Programs-and-Initiatives/Health-Insurance-Marketplaces/"| __truncated__
#>  .. $ modified   : Date[1:1], format: "2025-07-08"
#>  .. $ issued     : Date[1:1], format: "2016-03-30"
#>  .. $ periodicity: chr "Decennially [R/P10Y]"
#>  .. $ download   : chr "https://data.healthcare.gov/sites/default/files/uploaded_resources/TBL_RCL_173.csv"
#>  .. $ resources  : chr NA
#>  @ dimensions: <class_dimensions>
#>  .. @ limit : int 500
#>  .. @ rows  : int 1024945
#>  .. @ pages : int 2050
#>  .. @ fields:List of 7
#>  .. .. $ npn                                    : chr(0) 
#>  .. .. $ applicable_plan_year                   : chr(0) 
#>  .. .. $ individual_registration_completion_date: chr(0) 
#>  .. .. $ individual_marketplace_end_date        : chr(0) 
#>  .. .. $ shop_registration_completion_date      : chr(0) 
#>  .. .. $ shop_end_date                          : chr(0) 
#>  .. .. $ npn_valid_current_plan_year_only       : chr(0) 
#>  @ identifier: <class_current> chr "https://data.healthcare.gov/api/1/datastore/query/wb6u-x2ny/0?count=true&results=true&offset=0&limit=1&"
endpoint("asc_facility")
#> <class_prov>
#>  @ metadata  :List of 9
#>  .. $ title      : chr "Ambulatory Surgical Center Quality Measures - Facility"
#>  .. $ description: chr "A list of ambulatory surgical centers participating in the Ambulatory Surgical Center Quality Reporting (ASCQR)"| __truncated__
#>  .. $ modified   : Date[1:1], format: "2025-04-08"
#>  .. $ group      : chr "Hospitals"
#>  .. $ issued     : Date[1:1], format: "2024-10-10"
#>  .. $ released   : Date[1:1], format: "2025-04-30"
#>  .. $ download   : chr "https://data.cms.gov/provider-data/sites/default/files/resources/dd03994fc93e296bb0297f1cd43cc987_1744668308/ASC_Facility.csv"
#>  .. $ dictionary : chr "https://data.cms.gov/provider-data/dataset/4jcv-atw7#data-dictionary"
#>  .. $ site       : chr "https://data.cms.gov/provider-data/dataset/4jcv-atw7"
#>  @ dimensions: <class_dimensions>
#>  .. @ limit : int 1500
#>  .. @ rows  : int 11288
#>  .. @ pages : int 8
#>  .. @ fields:List of 50
#>  .. .. $ facility_name             : chr(0) 
#>  .. .. $ facility_id               : chr(0) 
#>  .. .. $ npi                       : chr(0) 
#>  .. .. $ citytown                  : chr(0) 
#>  .. .. $ state                     : chr(0) 
#>  .. .. $ zip_code                  : chr(0) 
#>  .. .. $ year                      : chr(0) 
#>  .. .. $ asc1_rate                 : chr(0) 
#>  .. .. $ asc1_footnote             : chr(0) 
#>  .. .. $ asc2_rate                 : chr(0) 
#>  .. .. $ asc2_footnote             : chr(0) 
#>  .. .. $ asc3_rate                 : chr(0) 
#>  .. .. $ asc3_footnote             : chr(0) 
#>  .. .. $ asc4_rate                 : chr(0) 
#>  .. .. $ asc4_footnote             : chr(0) 
#>  .. .. $ asc9_rate                 : chr(0) 
#>  .. .. $ asc9_footnote             : chr(0) 
#>  .. .. $ asc11_rate                : chr(0) 
#>  .. .. $ asc11_footnote            : chr(0) 
#>  .. .. $ asc12_total_cases         : chr(0) 
#>  .. .. $ asc12_performance_category: chr(0) 
#>  .. .. $ asc12_rshv_rate           : chr(0) 
#>  .. .. $ asc12_interval_lower_limit: chr(0) 
#>  .. .. $ asc12_interval_upper_limit: chr(0) 
#>  .. .. $ asc12_footnote            : chr(0) 
#>  .. .. $ asc13_rate                : chr(0) 
#>  .. .. $ asc13_footnote            : chr(0) 
#>  .. .. $ asc14_rate                : chr(0) 
#>  .. .. $ asc14_footnote            : chr(0) 
#>  .. .. $ asc17_total_cases         : chr(0) 
#>  .. .. $ asc17_performance_category: chr(0) 
#>  .. .. $ asc17_rshv_rate           : chr(0) 
#>  .. .. $ asc17_interval_lower_limit: chr(0) 
#>  .. .. $ asc17_interval_upper_limit: chr(0) 
#>  .. .. $ asc17_footnote            : chr(0) 
#>  .. .. $ asc18_total_cases         : chr(0) 
#>  .. .. $ asc18_performance_category: chr(0) 
#>  .. .. $ asc18_rshv_rate           : chr(0) 
#>  .. .. $ asc18_interval_lower_limit: chr(0) 
#>  .. .. $ asc18_interval_upper_limit: chr(0) 
#>  .. .. $ asc18_footnote            : chr(0) 
#>  .. .. $ asc19_total_cases         : chr(0) 
#>  .. .. $ asc19_performance_category: chr(0) 
#>  .. .. $ asc19_rshv_rate           : chr(0) 
#>  .. .. $ asc19_interval_lower_limit: chr(0) 
#>  .. .. $ asc19_interval_upper_limit: chr(0) 
#>  .. .. $ asc19_footnote            : chr(0) 
#>  .. .. $ asc20_sample              : chr(0) 
#>  .. .. $ asc20_rate                : chr(0) 
#>  .. .. $ asc20_footnote            : chr(0) 
#>  @ identifier: <class_current> chr "https://data.cms.gov/provider-data/api/1/datastore/query/4jcv-atw7/0?count=true&results=true&offset=0&limit=1&"
endpoint("dialysis_by_facility")
#> <class_prov>
#>  @ metadata  :List of 9
#>  .. $ title      : chr "Dialysis Facility - Listing by Facility"
#>  .. $ description: chr "A list of all dialysis facilities registered with Medicare that includes addresses and phone numbers, as well a"| __truncated__
#>  .. $ modified   : Date[1:1], format: "2025-03-24"
#>  .. $ group      : chr "Dialysis facilities"
#>  .. $ issued     : Date[1:1], format: "2020-03-14"
#>  .. $ released   : Date[1:1], format: "2025-04-23"
#>  .. $ download   : chr "https://data.cms.gov/provider-data/sites/default/files/resources/c04d84bc5c641284494bee4f20f17f9c_1742846704/DFC_FACILITY.csv"
#>  .. $ dictionary : chr "https://data.cms.gov/provider-data/dataset/23ew-n7w9#data-dictionary"
#>  .. $ site       : chr "https://data.cms.gov/provider-data/dataset/23ew-n7w9"
#>  @ dimensions: <class_dimensions>
#>  .. @ limit : int 1500
#>  .. @ rows  : int 7556
#>  .. @ pages : int 6
#>  .. @ fields:List of 142
#>  .. .. $ cms_certification_number_ccn                                    : chr(0) 
#>  .. .. $ network                                                         : chr(0) 
#>  .. .. $ facility_name                                                   : chr(0) 
#>  .. .. $ five_star_date                                                  : chr(0) 
#>  .. .. $ five_star                                                       : chr(0) 
#>  .. .. $ five_star_data_availability_code                                : chr(0) 
#>  .. .. $ address_line_1                                                  : chr(0) 
#>  .. .. $ address_line_2                                                  : chr(0) 
#>  .. .. $ citytown                                                        : chr(0) 
#>  .. .. $ state                                                           : chr(0) 
#>  .. .. $ zip_code                                                        : chr(0) 
#>  .. .. $ countyparish                                                    : chr(0) 
#>  .. .. $ telephone_number                                                : chr(0) 
#>  .. .. $ profit_or_nonprofit                                             : chr(0) 
#>  .. .. $ chain_owned                                                     : chr(0) 
#>  .. .. $ chain_organization                                              : chr(0) 
#>  .. .. $ late_shift                                                      : chr(0) 
#>  .. .. $ of_dialysis_stations                                            : chr(0) 
#>  .. .. $ offers_incenter_hemodialysis                                    : chr(0) 
#>  .. .. $ offers_peritoneal_dialysis                                      : chr(0) 
#>  .. .. $ offers_home_hemodialysis_training                               : chr(0) 
#>  .. .. $ certification_date                                              : chr(0) 
#>  .. .. $ claims_date                                                     : chr(0) 
#>  .. .. $ eqrs_date                                                       : chr(0) 
#>  .. .. $ smr_date                                                        : chr(0) 
#>  .. .. $ patient_survival_category_text                                  : chr(0) 
#>  .. .. $ patient_survival_data_availability_code                         : chr(0) 
#>  .. .. $ number_of_patients_included_in_survival_summary                 : chr(0) 
#>  .. .. $ mortality_rate_facility                                         : chr(0) 
#>  .. .. $ mortality_rate_upper_confidence_limit_975                       : chr(0) 
#>  .. .. $ mortality_rate_lower_confidence_limit_25                        : chr(0) 
#>  .. .. $ shr_date                                                        : chr(0) 
#>  .. .. $ patient_hospitalization_category_text                           : chr(0) 
#>  .. .. $ patient_hospitalization_data_availability_code                  : chr(0) 
#>  .. .. $ number_of_patients_included_in_hospitalization_summary          : chr(0) 
#>  .. .. $ hospitalization_rate_facility                                   : chr(0) 
#>  .. .. $ hospitalization_rate_upper_confidence_limit_975                 : chr(0) 
#>  .. .. $ hospitalization_rate_lower_confidence_limit_25                  : chr(0) 
#>  .. .. $ srr_date                                                        : chr(0) 
#>  .. .. $ patient_hospital_readmission_category                           : chr(0) 
#>  .. .. $ patient_hospital_readmission_data_availability_code             : chr(0) 
#>  .. .. $ number_of_hospitalizations_included_in_hospital_readmission_fc2b: chr(0) 
#>  .. .. $ readmission_rate_facility                                       : chr(0) 
#>  .. .. $ readmission_rate_upper_confidence_limit_975                     : chr(0) 
#>  .. .. $ readmission_rate_lower_confidence_limit_25                      : chr(0) 
#>  .. .. $ strr_date                                                       : chr(0) 
#>  .. .. $ patient_transfusion_category_text                               : chr(0) 
#>  .. .. $ patient_transfusion_data_availability_code                      : chr(0) 
#>  .. .. $ number_of_patients_included_in_the_transfusion_summary          : chr(0) 
#>  .. .. $ transfusion_rate_facility                                       : chr(0) 
#>  .. .. $ transfusion_rate_upper_confidence_limit_975                     : chr(0) 
#>  .. .. $ transfusion_rate_lower_confidence_limit_25                      : chr(0) 
#>  .. .. $ swr_date                                                        : chr(0) 
#>  .. .. $ swr_category_text                                               : chr(0) 
#>  .. .. $ patient_transplant_waitlist_data_availability_code              : chr(0) 
#>  .. .. $ number_of_patients_in_this_facility_for_swr                     : chr(0) 
#>  .. .. $ standardized_first_kidney_transplant_waitlist_ratio             : chr(0) 
#>  .. .. $ 95_ci_upper_limit_for_swr                                       : chr(0) 
#>  .. .. $ 95_ci_lower_limit_for_swr                                       : chr(0) 
#>  .. .. $ pppw_category_text                                              : chr(0) 
#>  .. .. $ patient_prevalent_transplant_waitlist_data_availability_code    : chr(0) 
#>  .. .. $ number_of_patients_for_pppw                                     : chr(0) 
#>  .. .. $ percentage_of_prevalent_patients_waitlisted                     : chr(0) 
#>  .. .. $ 95_ci_upper_limit_for_pppw                                      : chr(0) 
#>  .. .. $ 95_ci_lower_limit_for_pppw                                      : chr(0) 
#>  .. .. $ sedr_date                                                       : chr(0) 
#>  .. .. $ sedr_category_text                                              : chr(0) 
#>  .. .. $ emergency_department_encounter_data_availability_code           : chr(0) 
#>  .. .. $ number_of_patients_included_in_sedr_summary                     : chr(0) 
#>  .. .. $ standardized_ed_visits_ratio_facility                           : chr(0) 
#>  .. .. $ sedr_upper_confidence_limit_975                                 : chr(0) 
#>  .. .. $ sedr_lower_confidence_limit_25                                  : chr(0) 
#>  .. .. $ ed30_date                                                       : chr(0) 
#>  .. .. $ ed30_category_text                                              : chr(0) 
#>  .. .. $ emergency_department_encounter_ratio_occurring_within_30_da_f8e3: chr(0) 
#>  .. .. $ number_of_hospitalizations_included_in_ed30_summary             : chr(0) 
#>  .. .. $ standardized_ed_visits_within_30_days_of_hospital_discharge_6307: chr(0) 
#>  .. .. $ ed30_upper_confidence_limit_975                                 : chr(0) 
#>  .. .. $ ed30_lower_confidence_limit_25                                  : chr(0) 
#>  .. .. $ years_modality_switch_based_upon                                : chr(0) 
#>  .. .. $ smosr_classification_category_facility                          : chr(0) 
#>  .. .. $ modality_switch_data_availability_code                          : chr(0) 
#>  .. .. $ smosr_number_of_eligible_patients_facility                      : chr(0) 
#>  .. .. $ smosr_standardized_modality_switch_ratio_facility               : chr(0) 
#>  .. .. $ smosr_upper_confidence_limit_facility                           : chr(0) 
#>  .. .. $ smosr_lower_confidence_limit_facility                           : chr(0) 
#>  .. .. $ sir_date                                                        : chr(0) 
#>  .. .. $ patient_infection_category_text                                 : chr(0) 
#>  .. .. $ patient_infection_data_availability_code                        : chr(0) 
#>  .. .. $ standard_infection_ratio                                        : chr(0) 
#>  .. .. $ sir_upper_confidence_limit_975                                  : chr(0) 
#>  .. .. $ sir_lower_confidence_limit_25                                   : chr(0) 
#>  .. .. $ fistula_category_text                                           : chr(0) 
#>  .. .. $ fistula_data_availability_code                                  : chr(0) 
#>  .. .. $ number_of_patients_included_in_fistula_summary                  : chr(0) 
#>  .. .. $ fistula_rate_facility                                           : chr(0) 
#>  .. .. $ fistula_rate_upper_confidence_limit_975                         : chr(0) 
#>  .. .. $ fistula_rate_lower_confidence_limit_25                          : chr(0) 
#>  .. .. $ hcp_vaccination_data_collection_dates                           : chr(0) 
#>  .. ..  [list output truncated]
#>  @ identifier: <class_current> chr "https://data.cms.gov/provider-data/api/1/datastore/query/23ew-n7w9/0?count=true&results=true&offset=0&limit=1&"