Medicare Provider Enrollees

Information on a point in time snapshot of enrollment level data for providers actively enrolled in Medicare.

Usage

enrollees(
  npi = NULL,
  pac = NULL,
  enid = NULL,
  specialty_code = NULL,
  specialty_desc = NULL,
  state = NULL,
  first_name = NULL,
  middle_name = NULL,
  last_name = NULL,
  org_name = NULL
)

Arguments

npi

<chr> 10-digit Individual NPI

pac

<chr> 10-digit PECOS Associate Control (PAC) ID

enid

<chr> 15-digit Medicare Enrollment ID

specialty_code

<chr> Enrollment specialty code

specialty_desc

<chr> Enrollment specialty description

state

<chr> Enrollment state abbreviation

first_name, middle_name, last_name

<chr> Individual provider's name

org_name

<chr> Organizational provider's name

Value

<tibble> of search results

Details

Individual & organizational enrollment-level data on providers actively approved to bill Medicare.

Periodicity

Quarterly (R/P3M)

Links

Enrollment API

Enrollment Data Dictionary

Examples

enrollees(enid = "I20040309000221")
#> # A tibble: 5,000 × 10
#>    NPI        PECOS_ASCT_CNTL_ID ENRLMT_ID   PROVIDER_TYPE_CD PROVIDER_TYPE_DESC
#>    <chr>      <chr>              <chr>       <chr>            <chr>             
#>  1 1003879883 8022920719         I200311030… 14-16            PRACTITIONER - OB…
#>  2 1003976986 7113839812         I200311030… 14-68            PRACTITIONER - CL…
#>  3 1407802119 8022920727         I200311030… 14-93            PRACTITIONER - EM…
#>  4 1831165075 5193637890         I200311030… 14-16            PRACTITIONER - OB…
#>  5 1851357214 2466364161         I200311030… 14-30            PRACTITIONER - DI…
#>  6 1083766935 5092627794         I200311030… 14-35            PRACTITIONER - CH…
#>  7 1083835177 5991617698         I200311030… 14-41            PRACTITIONER - OP…
#>  8 1720297963 7618889312         I200311030… 14-26            PRACTITIONER - PS…
#>  9 1497881189 1254243959         I200311030… 14-01            PRACTITIONER - GE…
#> 10 1437155520 9234041930         I200311030… 14-68            PRACTITIONER - CL…
#> # ℹ 4,990 more rows
#> # ℹ 5 more variables: STATE_CD <chr>, FIRST_NAME <chr>, MDL_NAME <chr>,
#> #   LAST_NAME <chr>, ORG_NAME <chr>
# enrollees(npi = "1417918293", specialty_code = "14-41")
# enrollees(pac = "2860305554")
# enrollees(state = "GA")