Load a group of endpoints by alias

Load a group of endpoints by alias

Usage

group(alias, description = NULL, call = caller_env())

Arguments

alias

<chr> Alias representing the CMS data endpoint.

description

<chr> Group description. Defaults to NULL, which will use the aliases as the description.

call

<env> Environment from which to call the function.

Value

S7 <class_group> object.

Examples

group(c("asc_facility", "care_dial_end"))
#> <class_group>
#>  @ name   : chr "asc_facility | care_dial_end"
#>  @ members:List of 2
#>  .. $ asc_facility : <class_prov>
#>  ..  ..@ metadata  :List of 9
#>  .. .. .. $ title      : chr "Ambulatory Surgical Center Quality Measures - Facility"
#>  .. .. .. $ description: chr "A list of ambulatory surgical centers participating in the Ambulatory Surgical Center Quality Reporting (ASCQR)"| __truncated__
#>  .. .. .. $ modified   : Date[1:1], format: "2025-04-08"
#>  .. .. .. $ group      : chr "Hospitals"
#>  .. .. .. $ issued     : Date[1:1], format: "2024-10-10"
#>  .. .. .. $ released   : Date[1:1], format: "2025-04-30"
#>  .. .. .. $ download   : chr "https://data.cms.gov/provider-data/sites/default/files/resources/dd03994fc93e296bb0297f1cd43cc987_1744668308/ASC_Facility.csv"
#>  .. .. .. $ dictionary : chr "https://data.cms.gov/provider-data/dataset/4jcv-atw7#data-dictionary"
#>  .. .. .. $ site       : chr "https://data.cms.gov/provider-data/dataset/4jcv-atw7"
#>  ..  ..@ dimensions: <class_dimensions>
#>  .. .. .. @ limit : int 1500
#>  .. .. .. @ rows  : int 11288
#>  .. .. .. @ pages : int 8
#>  .. .. .. @ fields:List of 50
#>  .. .. .. .. $ facility_name             : chr(0) 
#>  .. .. .. .. $ facility_id               : chr(0) 
#>  .. .. .. .. $ npi                       : chr(0) 
#>  .. .. .. .. $ citytown                  : chr(0) 
#>  .. .. .. .. $ state                     : chr(0) 
#>  .. .. .. .. $ zip_code                  : chr(0) 
#>  .. .. .. .. $ year                      : chr(0) 
#>  .. .. .. .. $ asc1_rate                 : chr(0) 
#>  .. .. .. .. $ asc1_footnote             : chr(0) 
#>  .. .. .. .. $ asc2_rate                 : chr(0) 
#>  .. .. .. .. $ asc2_footnote             : chr(0) 
#>  .. .. .. .. $ asc3_rate                 : chr(0) 
#>  .. .. .. .. $ asc3_footnote             : chr(0) 
#>  .. .. .. .. $ asc4_rate                 : chr(0) 
#>  .. .. .. .. $ asc4_footnote             : chr(0) 
#>  .. .. .. .. $ asc9_rate                 : chr(0) 
#>  .. .. .. .. $ asc9_footnote             : chr(0) 
#>  .. .. .. .. $ asc11_rate                : chr(0) 
#>  .. .. .. .. $ asc11_footnote            : chr(0) 
#>  .. .. .. .. $ asc12_total_cases         : chr(0) 
#>  .. .. .. .. $ asc12_performance_category: chr(0) 
#>  .. .. .. .. $ asc12_rshv_rate           : chr(0) 
#>  .. .. .. .. $ asc12_interval_lower_limit: chr(0) 
#>  .. .. .. .. $ asc12_interval_upper_limit: chr(0) 
#>  .. .. .. .. $ asc12_footnote            : chr(0) 
#>  .. .. .. .. $ asc13_rate                : chr(0) 
#>  .. .. .. .. $ asc13_footnote            : chr(0) 
#>  .. .. .. .. $ asc14_rate                : chr(0) 
#>  .. .. .. .. $ asc14_footnote            : chr(0) 
#>  .. .. .. .. $ asc17_total_cases         : chr(0) 
#>  .. .. .. .. $ asc17_performance_category: chr(0) 
#>  .. .. .. .. $ asc17_rshv_rate           : chr(0) 
#>  .. .. .. .. $ asc17_interval_lower_limit: chr(0) 
#>  .. .. .. .. $ asc17_interval_upper_limit: chr(0) 
#>  .. .. .. .. $ asc17_footnote            : chr(0) 
#>  .. .. .. .. $ asc18_total_cases         : chr(0) 
#>  .. .. .. .. $ asc18_performance_category: chr(0) 
#>  .. .. .. .. $ asc18_rshv_rate           : chr(0) 
#>  .. .. .. .. $ asc18_interval_lower_limit: chr(0) 
#>  .. .. .. .. $ asc18_interval_upper_limit: chr(0) 
#>  .. .. .. .. $ asc18_footnote            : chr(0) 
#>  .. .. .. .. $ asc19_total_cases         : chr(0) 
#>  .. .. .. .. $ asc19_performance_category: chr(0) 
#>  .. .. .. .. $ asc19_rshv_rate           : chr(0) 
#>  .. .. .. .. $ asc19_interval_lower_limit: chr(0) 
#>  .. .. .. .. $ asc19_interval_upper_limit: chr(0) 
#>  .. .. .. .. $ asc19_footnote            : chr(0) 
#>  .. .. .. .. $ asc20_sample              : chr(0) 
#>  .. .. .. .. $ asc20_rate                : chr(0) 
#>  .. .. .. .. $ asc20_footnote            : chr(0) 
#>  ..  ..@ identifier: <class_current> chr "https://data.cms.gov/provider-data/api/1/datastore/query/4jcv-atw7/0?count=true&results=true&offset=0&limit=1&"
#>  .. $ care_dial_end: <class_care>
#>  ..  ..@ metadata  :List of 10
#>  .. .. .. $ title      : chr "Medicare Dialysis Facilities"
#>  .. .. .. $ description: chr "The Medicare Dialysis Facilities data provides information on clinical and patient measures for Medicare-certif"| __truncated__
#>  .. .. .. $ modified   : Date[1:1], format: "2025-06-24"
#>  .. .. .. $ temporal   : chr "2009-01-01 • 2025-12-31"
#>  .. .. .. $ periodicity: chr "Annually [R/P1Y]"
#>  .. .. .. $ download   : chr "https://data.cms.gov/sites/default/files/2025-04/94d61c1e-b20f-47a9-b236-521016b54b0f/dfr_facility_socrata_fy2025_resubmit.csv"
#>  .. .. .. $ resources  : chr "https://data.cms.gov/data-api/v1/dataset-resources/f8610e87-ba25-43a3-a49e-927dbc8701ae"
#>  .. .. .. $ dictionary : chr "https://data.cms.gov/sites/default/files/2025-03/FY2025%20PUF%20Data%20Dictionary.xlsx"
#>  .. .. .. $ site       : chr "https://data.cms.gov/quality-of-care/medicare-dialysis-facilities"
#>  .. .. .. $ references : chr "https://data.cms.gov/sites/default/files/2025-03/FY2025_DFR_Guide.pdf"
#>  ..  ..@ dimensions: <class_dimensions>
#>  .. .. .. @ limit : int 5000
#>  .. .. .. @ rows  : int 11985275
#>  .. .. .. @ pages : int 2398
#>  .. .. .. @ fields:List of 14
#>  .. .. .. .. $ state         : chr(0) 
#>  .. .. .. .. $ CCN           : chr(0) 
#>  .. .. .. .. $ Provider_Name : chr(0) 
#>  .. .. .. .. $ city          : chr(0) 
#>  .. .. .. .. $ Ownership_Type: chr(0) 
#>  .. .. .. .. $ ESRD_Network  : chr(0) 
#>  .. .. .. .. $ NPI           : chr(0) 
#>  .. .. .. .. $ Chain         : chr(0) 
#>  .. .. .. .. $ Modality      : chr(0) 
#>  .. .. .. .. $ Alternate_CCNs: chr(0) 
#>  .. .. .. .. $ Measure       : chr(0) 
#>  .. .. .. .. $ Measure_Score : chr(0) 
#>  .. .. .. .. $ year          : chr(0) 
#>  .. .. .. .. $ Measure_ID    : chr(0) 
#>  ..  ..@ identifier: <class_current> chr "https://data.cms.gov/data-api/v1/dataset/f8610e87-ba25-43a3-a49e-927dbc8701ae/data-viewer?offset=0&size=1"
#>  ..  ..@ resources : <class_current> chr "https://data.cms.gov/data-api/v1/dataset-resources/f8610e87-ba25-43a3-a49e-927dbc8701ae"
try(group("asc_facility"))
#> Error in eval(expr, envir) : 
#>   A <class_group> must have more than one `alias`.
#> ℹ Run `endpoint("asc_facility")` to load this endpoint.