Load a group of endpoints by alias

Load a group of endpoints by alias

Usage

group(alias, description = NULL, call = caller_env())

Arguments

alias

<chr> Alias representing the CMS data endpoint.

description

<chr> Group description. Defaults to NULL, which will use the aliases as the description.

call

<env> Environment from which to call the function.

Value

S7 <class_group> object.

Examples

group(c("asc_facility", "enterprise"))
#> <class_group>
#>  @ name   : chr "[asc_facility, enterprise]"
#>  @ members:List of 2
#>  .. $ asc_facility: <class_prov>
#>  ..  ..@ access: <class_current>
#>  .. .. .. @ identifier: chr "https://data.cms.gov/provider-data/api/1/datastore/query/4jcv-atw7/0?count=true&results=true&offset=0&limit=1"
#>  .. .. .. @ metadata  :List of 9
#>  .. .. .. .. $ title      : chr "Ambulatory Surgical Center Quality Measures - Facility"
#>  .. .. .. .. $ description: chr "A list of ambulatory surgical centers participating in the Ambulatory Surgical Center Quality Reporting (ASCQR)"| __truncated__
#>  .. .. .. .. $ modified   : Date[1:1], format: "2025-04-08"
#>  .. .. .. .. $ group      : chr "Hospitals"
#>  .. .. .. .. $ issued     : Date[1:1], format: "2024-10-10"
#>  .. .. .. .. $ released   : Date[1:1], format: "2025-04-30"
#>  .. .. .. .. $ download   : chr "https://data.cms.gov/provider-data/sites/default/files/resources/dd03994fc93e296bb0297f1cd43cc987_1744668308/ASC_Facility.csv"
#>  .. .. .. .. $ dictionary : chr "https://data.cms.gov/provider-data/dataset/4jcv-atw7#data-dictionary"
#>  .. .. .. .. $ site       : chr "https://data.cms.gov/provider-data/dataset/4jcv-atw7"
#>  .. .. .. @ dimensions: <class_dimensions>
#>  .. .. .. .. @ limit : int 1500
#>  .. .. .. .. @ rows  : int 11288
#>  .. .. .. .. @ pages : int 8
#>  .. .. .. .. @ fields:List of 50
#>  .. .. .. .. .. $ facility_name             : chr NA
#>  .. .. .. .. .. $ facility_id               : chr NA
#>  .. .. .. .. .. $ npi                       : chr NA
#>  .. .. .. .. .. $ citytown                  : chr NA
#>  .. .. .. .. .. $ state                     : chr NA
#>  .. .. .. .. .. $ zip_code                  : chr NA
#>  .. .. .. .. .. $ year                      : chr NA
#>  .. .. .. .. .. $ asc1_rate                 : chr NA
#>  .. .. .. .. .. $ asc1_footnote             : chr NA
#>  .. .. .. .. .. $ asc2_rate                 : chr NA
#>  .. .. .. .. .. $ asc2_footnote             : chr NA
#>  .. .. .. .. .. $ asc3_rate                 : chr NA
#>  .. .. .. .. .. $ asc3_footnote             : chr NA
#>  .. .. .. .. .. $ asc4_rate                 : chr NA
#>  .. .. .. .. .. $ asc4_footnote             : chr NA
#>  .. .. .. .. .. $ asc9_rate                 : chr NA
#>  .. .. .. .. .. $ asc9_footnote             : chr NA
#>  .. .. .. .. .. $ asc11_rate                : chr NA
#>  .. .. .. .. .. $ asc11_footnote            : chr NA
#>  .. .. .. .. .. $ asc12_total_cases         : chr NA
#>  .. .. .. .. .. $ asc12_performance_category: chr NA
#>  .. .. .. .. .. $ asc12_rshv_rate           : chr NA
#>  .. .. .. .. .. $ asc12_interval_lower_limit: chr NA
#>  .. .. .. .. .. $ asc12_interval_upper_limit: chr NA
#>  .. .. .. .. .. $ asc12_footnote            : chr NA
#>  .. .. .. .. .. $ asc13_rate                : chr NA
#>  .. .. .. .. .. $ asc13_footnote            : chr NA
#>  .. .. .. .. .. $ asc14_rate                : chr NA
#>  .. .. .. .. .. $ asc14_footnote            : chr NA
#>  .. .. .. .. .. $ asc17_total_cases         : chr NA
#>  .. .. .. .. .. $ asc17_performance_category: chr NA
#>  .. .. .. .. .. $ asc17_rshv_rate           : chr NA
#>  .. .. .. .. .. $ asc17_interval_lower_limit: chr NA
#>  .. .. .. .. .. $ asc17_interval_upper_limit: chr NA
#>  .. .. .. .. .. $ asc17_footnote            : chr NA
#>  .. .. .. .. .. $ asc18_total_cases         : chr NA
#>  .. .. .. .. .. $ asc18_performance_category: chr NA
#>  .. .. .. .. .. $ asc18_rshv_rate           : chr NA
#>  .. .. .. .. .. $ asc18_interval_lower_limit: chr NA
#>  .. .. .. .. .. $ asc18_interval_upper_limit: chr NA
#>  .. .. .. .. .. $ asc18_footnote            : chr NA
#>  .. .. .. .. .. $ asc19_total_cases         : chr NA
#>  .. .. .. .. .. $ asc19_performance_category: chr NA
#>  .. .. .. .. .. $ asc19_rshv_rate           : chr NA
#>  .. .. .. .. .. $ asc19_interval_lower_limit: chr NA
#>  .. .. .. .. .. $ asc19_interval_upper_limit: chr NA
#>  .. .. .. .. .. $ asc19_footnote            : chr NA
#>  .. .. .. .. .. $ asc20_sample              : chr NA
#>  .. .. .. .. .. $ asc20_rate                : chr NA
#>  .. .. .. .. .. $ asc20_footnote            : chr NA
#>  .. $ enterprise  : <class_caid>
#>  ..  ..@ access: <class_current>
#>  .. .. .. @ identifier: chr "https://data.medicaid.gov/api/1/datastore/query/13a06cdb-6bbb-4f86-bba7-9d6f3db41090/0?count=true&results=true&"| __truncated__
#>  .. .. .. @ metadata  :List of 5
#>  .. .. .. .. $ title      : chr "Medicaid Enterprise System Datatable"
#>  .. .. .. .. $ description: chr "Medicaid Enterprise System Datatable"
#>  .. .. .. .. $ modified   : Date[1:1], format: "2025-05-21"
#>  .. .. .. .. $ periodicity: chr NA
#>  .. .. .. .. $ download   : chr NA
#>  .. .. .. @ dimensions: <class_dimensions>
#>  .. .. .. .. @ limit : int 8000
#>  .. .. .. .. @ rows  : int 68
#>  .. .. .. .. @ pages : int 1
#>  .. .. .. .. @ fields:List of 13
#>  .. .. .. .. .. $ implementation_year              : chr NA
#>  .. .. .. .. .. $ policyupdate_type                : chr NA
#>  .. .. .. .. .. $ policyupdate_title               : chr NA
#>  .. .. .. .. .. $ potential_mes_impact             : chr NA
#>  .. .. .. .. .. $ statute_or_regulation_citation   : chr NA
#>  .. .. .. .. .. $ compliance_month                 : chr NA
#>  .. .. .. .. .. $ compliance_day                   : chr NA
#>  .. .. .. .. .. $ type_of_potential_system_impact  : chr NA
#>  .. .. .. .. .. $ category_of_potential_impact     : chr NA
#>  .. .. .. .. .. $ link_to__source_document         : chr NA
#>  .. .. .. .. .. $ impact_to_medicaid_chip_andor_bhp: chr NA
#>  .. .. .. .. .. $ tshirt_size                      : chr NA
#>  .. .. .. .. .. $ certification_activities_required: chr NA
try(group("asc_facility"))
#> Error in eval(expr, envir) : 
#>   A <class_group> must have more than one `alias`.
#> ℹ Run `endpoint("asc_facility")` to load this endpoint.