Create Groups of Endpoints

Convenience function to quickly access various CMS data endpoints. Mostly for debugging purposes.

Usage

new_group(
  member_names,
  group_name = NULL,
  quick_call = FALSE,
  offset = 0,
  limit = 10
)

quick(x, offset = 0, limit = 10000, call = caller_env())

Arguments

member_names

<chr> Alias representing the CMS data endpoint.

group_name

<chr> Name of the group.

quick_call

<lgl> call quick_() method on group if TRUE.

offset

<int> The offset for pagination. Default is 0.

limit

<int> The maximum number of records to retrieve. Default is 10000.

x

<chr> Alias representing the CMS data endpoint or category.

call

Call environment. Default is the current environment.

Value

S7 class_group object.

A data frame or list containing the requested CMS data.

Examples

new_group(c("qhp_bus_rule_variables", "managed_care_share"))
#> <class_group>
#>  @ group  : chr "ALIAS( qhp_bus_rule_variables ) | ALIAS( managed_care_share )"
#>  @ members:List of 2
#>  .. $ qhp_bus_rule_variables: <class_endpoint>
#>  ..  ..@ identifier: chr "https://data.healthcare.gov/api/1/datastore/query/3yfc-yz8a/0"
#>  ..  ..@ metadata  :List of 5
#>  .. .. .. $ title      : chr "QHP Landscape Health Plan Business Rule Variables"
#>  .. .. .. $ description: chr "QHP Landscape Health Plan Business Rule Variables"
#>  .. .. .. $ modified   : Date[1:1], format: "2014-04-16"
#>  .. .. .. $ issued     : Date[1:1], format: "2013-11-26"
#>  .. .. .. $ download   : chr "https://data.healthcare.gov/datafile/qhp-landscape-health-plan-business-rule-variables.3yfc-yz8a.csv"
#>  ..  ..@ dimensions: <class_dimensions>
#>  .. .. .. @ limit : int 500
#>  .. .. .. @ rows  : int 5891
#>  .. .. .. @ pages : int 12
#>  .. .. .. @ fields:List of 19
#>  .. .. .. .. $ standard_component_id                                           : chr(0) 
#>  .. .. .. .. $ issuer_id                                                       : chr(0) 
#>  .. .. .. .. $ issuer_legal_name                                               : chr(0) 
#>  .. .. .. .. $ state                                                           : chr(0) 
#>  .. .. .. .. $ metal_level                                                     : chr(0) 
#>  .. .. .. .. $ market_coverage                                                 : chr(0) 
#>  .. .. .. .. $ on_or_off_exchange                                              : chr(0) 
#>  .. .. .. .. $ sadp_or_qhp                                                     : chr(0) 
#>  .. .. .. .. $ child_only_offering                                             : chr(0) 
#>  .. .. .. .. $ guaranteed_or_estimated_rate                                    : chr(0) 
#>  .. .. .. .. $ how_are_rates_for_contracts_covering_two_or_more_enrollees__e264: chr(0) 
#>  .. .. .. .. $ what_are_the_maximum_number_of_under_age_under_21_dependent_364a: chr(0) 
#>  .. .. .. .. $ what_are_the_maximum_number_of_under_age_under_21_dependent_e6fb: chr(0) 
#>  .. .. .. .. $ is_there_a_maximum_age_for_a_dependent                          : chr(0) 
#>  .. .. .. .. $ what_are_the_maximum_number_of_children_used_to_quote_a_chi_733a: chr(0) 
#>  .. .. .. .. $ are_domestic_partners_treated_the_same_as_secondary_subscribers : chr(0) 
#>  .. .. .. .. $ are_samesex_partners_treated_the_same_as_secondary_subscribers  : chr(0) 
#>  .. .. .. .. $ how_is_age_determined_for_rating_and_eligibility_purposes       : chr(0) 
#>  .. .. .. .. $ what_relationships_between_primary_and_dependent_are_allowe_e280: chr(0) 
#>  .. $ managed_care_share    : <class_endpoint>
#>  ..  ..@ identifier: chr "https://data.medicaid.gov/api/1/datastore/query/79692ea5-21e1-56bf-8149-97d437120c4b/0"
#>  ..  ..@ metadata  :List of 5
#>  .. .. .. $ title      : chr "Share of Medicaid Enrollees in Managed Care"
#>  .. .. .. $ description: chr "The Share of Medicaid Enrollees in any Managed Care and in Comprehensive Managed CaAre profiles state-level enr"| __truncated__
#>  .. .. .. $ modified   : Date[1:1], format: "2024-10-16"
#>  .. .. .. $ periodicity: chr "Decennially [R/P10Y]"
#>  .. .. .. $ download   : chr "https://data.medicaid.gov/sites/default/files/uploaded_resources/share-of-medicaid-enrollees-in-managed-care.2022.csv"
#>  ..  ..@ dimensions: <class_dimensions>
#>  .. .. .. @ limit : int 8000
#>  .. .. .. @ rows  : int 399
#>  .. .. .. @ pages : int 1
#>  .. .. .. @ fields:List of 8
#>  .. .. .. .. $ state                                          : chr(0) 
#>  .. .. .. .. $ notes                                          : chr(0) 
#>  .. .. .. .. $ total_medicaid_enrollees                       : chr(0) 
#>  .. .. .. .. $ individuals_enrolled_any                       : chr(0) 
#>  .. .. .. .. $ percent_of_all_medicaid_enrollees_any          : chr(0) 
#>  .. .. .. .. $ individuals_enrolled_comprehensive             : chr(0) 
#>  .. .. .. .. $ percent_of_all_medicaid_enrollees_comprehensive: chr(0) 
#>  .. .. .. .. $ year                                           : chr(0) 
new_group(c("qhp_bus_rule_variables", "managed_care_share"), quick_call = TRUE)
#> $qhp_bus_rule_variables
#> # A tibble: 10 × 19
#>    standard_component_id issuer_id issuer_legal_name           state metal_level
#>    <chr>                 <chr>     <chr>                       <chr> <chr>      
#>  1 10191NJ0050001        10191     Freelancers CO-OP of New J… NJ    Silver     
#>  2 10191NJ0040002        10191     Freelancers CO-OP of New J… NJ    Silver     
#>  3 10191NJ0070004        10191     Freelancers CO-OP of New J… NJ    Silver     
#>  4 10191NJ0050003        10191     Freelancers CO-OP of New J… NJ    Platinum   
#>  5 10191NJ0120001        10191     Freelancers CO-OP of New J… NJ    Silver     
#>  6 10191NJ0140002        10191     Freelancers CO-OP of New J… NJ    Silver     
#>  7 10191NJ0050002        10191     Freelancers CO-OP of New J… NJ    Gold       
#>  8 10191NJ0070003        10191     Freelancers CO-OP of New J… NJ    Gold       
#>  9 10191NJ0070001        10191     Freelancers CO-OP of New J… NJ    Bronze     
#> 10 10191NJ0100002        10191     Freelancers CO-OP of New J… NJ    Silver     
#> # ℹ 14 more variables: market_coverage <chr>, on_or_off_exchange <chr>,
#> #   sadp_or_qhp <chr>, child_only_offering <chr>,
#> #   guaranteed_or_estimated_rate <chr>,
#> #   how_are_rates_for_contracts_covering_two_or_more_enrollees__e264 <chr>,
#> #   what_are_the_maximum_number_of_under_age_under_21_dependent_364a <chr>,
#> #   what_are_the_maximum_number_of_under_age_under_21_dependent_e6fb <chr>,
#> #   is_there_a_maximum_age_for_a_dependent <chr>, …
#> 
#> $managed_care_share
#> # A tibble: 10 × 8
#>    state                notes      total_medicaid_enrol…¹ individuals_enrolled…²
#>    <chr>                <chr>      <chr>                  <chr>                 
#>  1 TOTALS               NA         80,263,839             65,034,032            
#>  2 Arkansas5            Arkansas … 1,026,621              512,041               
#>  3 District of Columbia NA         251,791                180,942               
#>  4 Wyoming              NA         64,442                 391                   
#>  5 Pennsylvania         NA         2,753,618              2,521,421             
#>  6 Massachusetts        NA         1,889,306              1,268,120             
#>  7 North Dakota         NA         93,422                 48,621                
#>  8 Michigan             NA         4,448,582              4,370,138             
#>  9 Virginia             NA         1,111,999              761,019               
#> 10 Hawaii               NA         358,302                354,289               
#> # ℹ abbreviated names: ¹​total_medicaid_enrollees, ²​individuals_enrolled_any
#> # ℹ 4 more variables: percent_of_all_medicaid_enrollees_any <chr>,
#> #   individuals_enrolled_comprehensive <chr>,
#> #   percent_of_all_medicaid_enrollees_comprehensive <chr>, year <chr>
#> 
quick("revalid_group")
#> # A tibble: 5,000 × 15
#>    `Group PAC ID` `Group Enrollment ID` `Group Legal Business Name`            
#>    <chr>          <chr>                 <chr>                                  
#>  1 6305933532     O20071103000162       'a' Street Clinic Of Chiropractic, Pllc
#>  2 1850813254     O20250314001303       1 Care Partners Llc                    
#>  3 3274795547     O20120502000499       1 Community Wellness & Health Inc      
#>  4 1658693460     O20141205001656       1 Cp Place Pllc                        
#>  5 1658693460     O20141205001656       1 Cp Place Pllc                        
#>  6 5890010177     O20150209000022       1 Hour Eyedocs, Llc                    
#>  7 6204378193     O20240606002709       1 Identity Counseling, Llc             
#>  8 4789081530     O20210922000917       1 Love Wellness Society Llc            
#>  9 4789081530     O20210922000917       1 Love Wellness Society Llc            
#> 10 4789081530     O20210922000917       1 Love Wellness Society Llc            
#> # ℹ 4,990 more rows
#> # ℹ 12 more variables: `Group State Code` <chr>, `Group Due Date` <chr>,
#> #   `Group Reassignments and Physician Assistants` <chr>, `Record Type` <chr>,
#> #   `Individual Enrollment ID` <chr>, `Individual NPI` <chr>,
#> #   `Individual First Name` <chr>, `Individual Last Name` <chr>,
#> #   `Individual State Code` <chr>, `Individual Specialty Description` <chr>,
#> #   `Individual Due Date` <chr>, …
quick("out_img_national")
#> # A tibble: 4 × 6
#>   measure_id measure_name                     score footnote start_date end_date
#>   <chr>      <chr>                            <chr> <chr>    <chr>      <chr>   
#> 1 OP_10      Abdomen CT Use of Contrast Mate… 5.8   NA       07/01/2022 06/30/2…
#> 2 OP_13      Outpatients who got cardiac ima… 3.7   NA       07/01/2022 06/30/2…
#> 3 OP_39      Breast Cancer Screening Recall … 9     NA       07/01/2022 06/30/2…
#> 4 OP_8       MRI Lumbar Spine for Low Back P… 36.2  NA       07/01/2022 06/30/2…