Load Main API Dataset

Load Main API Dataset

Usage

main_current(alias)

Arguments

alias

<chr> endpoint alias

Value

<Dataset> object

Examples

main_current("enrollees")
#> $title
#> [1] "Medicare Fee-For-Service Public Provider Enrollment"
#> 
#> $description
#> [1] "The Medicare Fee-For-Service Public Provider Enrollment dataset includes information on providers who are actively approved to bill Medicare or have completed the 855O at the time the data was pulled from the Provider Enrollment, Chain, and Ownership System (PECOS). The release of this provider enrollment data is not related to other provider information releases such as Physician Compare or Data Transparency."
#> 
#> $modified
#> [1] "2025-01-15"
#> 
#> $periodicity
#> [1] "Quarterly [R/P3M]"
#> 
#> $temporal
#> [1] "2023-01-01/2024-12-31"
#> 
#> $contact
#> [1] "Provider Enrollment Data Requests - CPI (mailto:ProviderEnrollmentDataRequests@cms.hhs.gov)"
#> 
#> $identifier
#> [1] "https://data.cms.gov/data-api/v1/dataset/2457ea29-fc82-48b0-86ec-3b0755de7515/data-viewer"
#> 
#> $dictionary
#> [1] "https://data.cms.gov/resources/medicare-fee-for-service-public-provider-enrollment-data-dictionary"
#> 
#> $site
#> [1] "https://data.cms.gov/provider-characteristics/medicare-provider-supplier-enrollment/medicare-fee-for-service-public-provider-enrollment"
#> 
#> $references
#> [1] "https://data.cms.gov/resources/fee-for-service-public-provider-enrollment-methodology"
#> 
#> $download
#> [1] "https://data.cms.gov/sites/default/files/2025-01/8ced283e-4017-4b90-b511-142aabb8531b/PPEF_Enrollment_Extract_2025.01.02.csv"
#> 
#> $filetype
#> [1] "text/csv"
#> 
#> $resources
#> [1] "https://data.cms.gov/data-api/v1/dataset-resources/2457ea29-fc82-48b0-86ec-3b0755de7515"
#> 
main_current("opt_out")
#> $title
#> [1] "Opt Out Affidavits"
#> 
#> $description
#> [1] "ATTENTION USERS\n\n\nSome Providers Opt-Out Status may end early due to COVID 19 waivers. Please contact your respective MAC for further information.\n\n \n\nFor more information on the opt-out process, see Manage Your Enrollment or view the FAQ section below.\n\n \n\nThe Opt Out Affidavits dataset provides information on providers who have decided not to participate in Medicare. It contains providers NPI, specialty, address, and effective dates."
#> 
#> $modified
#> [1] "2025-02-27"
#> 
#> $periodicity
#> [1] "Monthly [R/P1M]"
#> 
#> $temporal
#> [1] "2023-04-01/2025-01-31"
#> 
#> $contact
#> [1] "Provider Enrollment Data Requests - CPI (mailto:ProviderEnrollmentDataRequests@cms.hhs.gov)"
#> 
#> $identifier
#> [1] "https://data.cms.gov/data-api/v1/dataset/9887a515-7552-4693-bf58-735c77af46d7/data-viewer"
#> 
#> $dictionary
#> [1] "https://data.cms.gov/resources/opt-out-affidavits-data-dictionary"
#> 
#> $site
#> [1] "https://data.cms.gov/provider-characteristics/medicare-provider-supplier-enrollment/opt-out-affidavits"
#> 
#> $references
#> [1] "https://data.cms.gov/resources/opt-out-affidavits-methodology"
#> 
#> $download
#> [1] "https://data.cms.gov/sites/default/files/2025-02/101001b7-73df-419e-b101-602baa5015c5/Opt_Out_List_January.csv"
#> 
#> $filetype
#> [1] "text/csv"
#> 
#> $resources
#> [1] "https://data.cms.gov/data-api/v1/dataset-resources/9887a515-7552-4693-bf58-735c77af46d7"
#> 
main_current("order_refer")
#> $title
#> [1] "Order and Referring"
#> 
#> $description
#> [1] "The Order and Referring dataset provides information on all physicians and non-physician practitioners, by their National Provider Identifier (NPI), who are of a type/specialty that is legally eligible to order and refer in the Medicare program and who have current enrollment records in Medicare."
#> 
#> $modified
#> [1] "2025-03-07"
#> 
#> $periodicity
#> [1] "Twice a Week [R/P3.5D]"
#> 
#> $temporal
#> [1] "2023-05-21/2025-03-01"
#> 
#> $contact
#> [1] "Provider Enrollment Data Requests - CPI (mailto:ProviderEnrollmentDataRequests@cms.hhs.gov)"
#> 
#> $identifier
#> [1] "https://data.cms.gov/data-api/v1/dataset/c99b5865-1119-4436-bb80-c5af2773ea1f/data-viewer"
#> 
#> $dictionary
#> [1] "https://data.cms.gov/resources/order-and-referring-data-dictionary"
#> 
#> $site
#> [1] "https://data.cms.gov/provider-characteristics/medicare-provider-supplier-enrollment/order-and-referring"
#> 
#> $references
#> [1] "https://data.cms.gov/resources/order-and-referring-methodology"
#> 
#> $download
#> [1] "https://data.cms.gov/sites/default/files/2025-03/90cbe4f6-0c7a-4a1f-94cc-7302836f3167/OrderReferring_20250306.csv"
#> 
#> $filetype
#> [1] "text/csv"
#> 
#> $resources
#> [1] "https://data.cms.gov/data-api/v1/dataset-resources/c99b5865-1119-4436-bb80-c5af2773ea1f"
#> 
main_current("reassignments")
#> $title
#> [1] "Revalidation Reassignment List"
#> 
#> $description
#> [1] "The Revalidation Reassignment List dataset provides information on reassignments of providers who are due for revalidation."
#> 
#> $modified
#> [1] "2025-01-02"
#> 
#> $periodicity
#> [1] "Monthly [R/P1M]"
#> 
#> $temporal
#> [1] "2022-02-01/2025-01-31"
#> 
#> $contact
#> [1] "Provider Enrollment Data Requests - CPI (mailto:ProviderEnrollmentDataRequests@cms.hhs.gov)"
#> 
#> $identifier
#> [1] "https://data.cms.gov/data-api/v1/dataset/20f51cff-4137-4f3a-b6b7-bfc9ad57983b/data-viewer"
#> 
#> $dictionary
#> [1] "https://data.cms.gov/resources/revalidation-reassignments-list-data-dictionary"
#> 
#> $site
#> [1] "https://data.cms.gov/provider-characteristics/medicare-provider-supplier-enrollment/revalidation-reassignment-list"
#> 
#> $references
#> [1] "https://data.cms.gov/resources/additional-information-on-revalidation"
#> 
#> $download
#> [1] "https://data.cms.gov/sites/default/files/2024-12/20b356f5-23c7-4f6e-8d71-55aa623f19fc/revalidation_group_reassignment_pa_xref.csv"
#> 
#> $filetype
#> [1] "text/csv"
#> 
#> $resources
#> [1] "https://data.cms.gov/data-api/v1/dataset-resources/20f51cff-4137-4f3a-b6b7-bfc9ad57983b"
#> 
main_current("hospitals")
#> $title
#> [1] "Hospital Enrollments"
#> 
#> $description
#> [1] "The Hospital Enrollments dataset provides enrollment information of all Hospitals currently enrolled in Medicare. This data includes information on the Hospitals sub-group type, legal business name, doing business as name, organization type and address."
#> 
#> $modified
#> [1] "2025-01-15"
#> 
#> $periodicity
#> [1] "Monthly [R/P1M]"
#> 
#> $temporal
#> [1] "2022-11-01/2025-01-31"
#> 
#> $contact
#> [1] "Provider Enrollment Data Requests - CPI (mailto:ProviderEnrollmentDataRequests@cms.hhs.gov)"
#> 
#> $identifier
#> [1] "https://data.cms.gov/data-api/v1/dataset/f6f6505c-e8b0-4d57-b258-e2b94133aaf2/data-viewer"
#> 
#> $dictionary
#> [1] "https://data.cms.gov/sites/default/files/2023-07/Hospital_Enrollments_Data_Dictionary.pdf"
#> 
#> $site
#> [1] "https://data.cms.gov/provider-characteristics/hospitals-and-other-facilities/hospital-enrollments"
#> 
#> $references
#> [1] "https://data.cms.gov/sites/default/files/2024-10/Hospital_Data_Guidance.pdf"
#> 
#> $download
#> [1] "https://data.cms.gov/sites/default/files/2025-01/0ef8ef0d-bf93-4b00-a542-a4e2853947cc/Hospital_Enrollments_2025.01.02.csv"
#> 
#> $filetype
#> [1] "text/csv"
#> 
#> $resources
#> [1] "https://data.cms.gov/data-api/v1/dataset-resources/f6f6505c-e8b0-4d57-b258-e2b94133aaf2"
#> 
main_current("laboratories")
#> $title
#> [1] "Provider of Services File - Clinical Laboratories"
#> 
#> $description
#> [1] "The Provider of Services (POS) Clinical Laboratories (CLIA) data provides information on CLIA demographics and types of testing services the facility provides. In this file you will find provider number, name, address and characteristics of the participating institution providers."
#> 
#> $modified
#> [1] "2025-01-16"
#> 
#> $periodicity
#> [1] "Quarterly [R/P3M]"
#> 
#> $temporal
#> [1] "2011-10-01/2024-12-31"
#> 
#> $contact
#> [1] "QIESData - CCSQ (mailto:QIESData@cms.hhs.gov)"
#> 
#> $identifier
#> [1] "https://data.cms.gov/data-api/v1/dataset/d3eb38ac-d8e9-40d3-b7b7-6205d3d1dc16/data-viewer"
#> 
#> $dictionary
#> [1] "https://data.cms.gov/sites/default/files/2023-07/50eacc07-7930-4e62-97e3-e04a2dc1e2a3/P.QWB.POSQ.CLIA.LAYOUT.MAR23.pdf"
#> 
#> $site
#> [1] "https://data.cms.gov/provider-characteristics/hospitals-and-other-facilities/provider-of-services-file-clinical-laboratories"
#> 
#> $references
#> [1] "https://data.cms.gov/resources/provider-of-services-file-methodology"
#> 
#> $download
#> [1] "https://data.cms.gov/sites/default/files/2025-01/f16b0884-32a4-427f-b679-b48d3ef3146f/clia_data_Q4%202024.csv"
#> 
#> $filetype
#> [1] "text/csv"
#> 
#> $resources
#> [1] "https://data.cms.gov/data-api/v1/dataset-resources/d3eb38ac-d8e9-40d3-b7b7-6205d3d1dc16"
#> 
main_current("crosswalk")
#> $title
#> [1] "Medicare Provider and Supplier Taxonomy Crosswalk"
#> 
#> $description
#> [1] "The Medicare Provider and Supplier Taxonomy Crosswalk dataset lists the providers and suppliers eligible to enroll in Medicare programs with the proper healthcare provider taxonomy code. This data includes the Medicare speciality codes, if available, provider/supplier type description, taxonomy code, and the taxonomy description."
#> 
#> $modified
#> [1] "2025-01-17"
#> 
#> $periodicity
#> [1] "Biannually [R/P6M]"
#> 
#> $temporal
#> [1] "2021-07-01/2025-06-30"
#> 
#> $contact
#> [1] "Provider Enrollment Data Requests - CPI (mailto:ProviderEnrollmentDataRequests@cms.hhs.gov)"
#> 
#> $identifier
#> [1] "https://data.cms.gov/data-api/v1/dataset/113eb0bc-0c9a-4d91-9f93-3f6b28c0bf6b/data-viewer"
#> 
#> $dictionary
#> [1] "https://data.cms.gov/resources/medicare-provider-and-supplier-taxonomy-crosswalk-data-dictionary"
#> 
#> $site
#> [1] "https://data.cms.gov/provider-characteristics/medicare-provider-supplier-enrollment/medicare-provider-and-supplier-taxonomy-crosswalk"
#> 
#> $references
#> [1] "https://data.cms.gov/resources/medicare-provider-and-supplier-taxonomy-crosswalk-methodology"
#> 
#> $download
#> [1] "https://data.cms.gov/sites/default/files/2025-01/7b0ea92d-2f70-424e-bec4-3fd35dc9a597/Medicare_Provider_and_Supplier_Taxonomy_Crosswalk_JAN__2025.csv"
#> 
#> $filetype
#> [1] "text/csv"
#> 
#> $resources
#> [1] "https://data.cms.gov/data-api/v1/dataset-resources/113eb0bc-0c9a-4d91-9f93-3f6b28c0bf6b"
#> 
main_current("rbcs")
#> $title
#> [1] "Restructured BETOS Classification System"
#> 
#> $description
#> [1] "The Restructured BETOS Classification System (RBCS) dataset is a taxonomy that allows researchers to group healthcare service codes for Medicare Part B services (i.e., HCPCS codes) into clinically meaningful categories and subcategories. It is based on the original Berenson-Eggers Type of Service (BETOS) classification created in the 1980s, and includes notable updates such as Part B non-physician services. The RBCS will undergo annual updates by a technical expert panel of researchers and clinicians."
#> 
#> $modified
#> [1] "2024-10-22"
#> 
#> $periodicity
#> [1] "Annually [R/P1Y]"
#> 
#> $temporal
#> [1] "2024-01-01/2024-12-31"
#> 
#> $contact
#> [1] "PDAG Data Products - OEDA (mailto:PDAG_Data_Products@cms.hhs.gov)"
#> 
#> $identifier
#> [1] "https://data.cms.gov/data-api/v1/dataset/e3db6e56-149f-49ce-b374-40aecda2357b/data-viewer"
#> 
#> $dictionary
#> [1] "https://data.cms.gov/resources/rbcs-data-dictionary-2024"
#> 
#> $site
#> [1] "https://data.cms.gov/provider-summary-by-type-of-service/provider-service-classifications/restructured-betos-classification-system"
#> 
#> $references
#> [1] "https://data.cms.gov/sites/default/files/2024-10/2024%20RBCS%20Final%20Report.pdf"
#> 
#> $download
#> [1] "https://data.cms.gov/sites/default/files/2024-10/30b74146-c5ce-47e0-8ae2-ad1204c81906/2024%20RBCS%20Taxonomy_CSV.csv"
#> 
#> $filetype
#> [1] "text/csv"
#> 
#> $resources
#> [1] "https://data.cms.gov/data-api/v1/dataset-resources/e3db6e56-149f-49ce-b374-40aecda2357b"
#>