Load Main API Dataset

Load Main API Dataset

Usage

main_current(alias)

Arguments

alias

<chr> endpoint alias

Value

<Dataset> object

Examples

main_current("enrollees")
#> $title
#> [1] "Medicare Fee-For-Service  Public Provider Enrollment"
#> 
#> $description
#> [1] "The Medicare Fee-For-Service Public Provider Enrollment dataset includes information on providers who are actively approved to bill Medicare or have completed the 855O at the time the data was pulled from the Provider Enrollment, Chain, and Ownership System (PECOS). The release of this provider enrollment data is not related to other provider information releases such as Physician Compare or Data Transparency.. .  . . Note: This full dataset contains more records than most spreadsheet programs can handle, which will result in an incomplete load of data. Use of a database or statistical software is required."
#> 
#> $periodicity
#> [1] "Quarterly [R/P3M]"
#> 
#> $contact
#> [1] "Provider Enrollment Data Requests - CPI (mailto:ProviderEnrollmentDataRequests@cms.hhs.gov)"
#> 
#> $dictionary
#> [1] "https://data.cms.gov/resources/medicare-fee-for-service-public-provider-enrollment-data-dictionary"
#> 
#> $identifier
#> [1] "https://data.cms.gov/data-api/v1/dataset/2457ea29-fc82-48b0-86ec-3b0755de7515/data-viewer"
#> 
#> $modified
#> [1] "2025-01-15"
#> 
#> $site
#> [1] "https://data.cms.gov/provider-characteristics/medicare-provider-supplier-enrollment/medicare-fee-for-service-public-provider-enrollment"
#> 
#> $temporal
#> [1] "2023-01-01/2024-12-31"
#> 
#> $references
#> [1] "https://data.cms.gov/resources/fee-for-service-public-provider-enrollment-methodology"
#> 
main_current("hospitals")
#> $title
#> [1] "Hospital Enrollments"
#> 
#> $description
#> [1] "The Hospital Enrollments dataset provides enrollment information of all Hospitals currently enrolled in Medicare. This data includes information on the Hospital's sub-group type, legal business name, doing business as name, organization type and address. "
#> 
#> $periodicity
#> [1] "Monthly [R/P1M]"
#> 
#> $contact
#> [1] "Provider Enrollment Data Requests - CPI (mailto:ProviderEnrollmentDataRequests@cms.hhs.gov)"
#> 
#> $dictionary
#> [1] "https://data.cms.gov/sites/default/files/2023-07/Hospital_Enrollments_Data_Dictionary.pdf"
#> 
#> $identifier
#> [1] "https://data.cms.gov/data-api/v1/dataset/f6f6505c-e8b0-4d57-b258-e2b94133aaf2/data-viewer"
#> 
#> $modified
#> [1] "2025-01-15"
#> 
#> $site
#> [1] "https://data.cms.gov/provider-characteristics/hospitals-and-other-facilities/hospital-enrollments"
#> 
#> $temporal
#> [1] "2022-11-01/2025-01-31"
#> 
#> $references
#> [1] "https://data.cms.gov/sites/default/files/2024-10/Hospital_Data_Guidance.pdf"
#> 
main_current("reassignments")
#> $title
#> [1] "Revalidation Reassignment List"
#> 
#> $description
#> [1] "The Revalidation Reassignment List dataset provides information on reassignments of providers who are due for revalidation.. .  . . Note: This full dataset contains more records than most spreadsheet programs can handle, which will result in an incomplete load of data. Use of a database or statistical software is required."
#> 
#> $periodicity
#> [1] "Monthly [R/P1M]"
#> 
#> $contact
#> [1] "Provider Enrollment Data Requests - CPI (mailto:ProviderEnrollmentDataRequests@cms.hhs.gov)"
#> 
#> $dictionary
#> [1] "https://data.cms.gov/resources/revalidation-reassignments-list-data-dictionary"
#> 
#> $identifier
#> [1] "https://data.cms.gov/data-api/v1/dataset/20f51cff-4137-4f3a-b6b7-bfc9ad57983b/data-viewer"
#> 
#> $modified
#> [1] "2025-01-02"
#> 
#> $site
#> [1] "https://data.cms.gov/provider-characteristics/medicare-provider-supplier-enrollment/revalidation-reassignment-list"
#> 
#> $temporal
#> [1] "2022-02-01/2025-01-31"
#> 
#> $references
#> [1] "https://data.cms.gov/resources/additional-information-on-revalidation"
#> 
main_current("opt_out")
#> $title
#> [1] "Opt Out Affidavits"
#> 
#> $description
#> [1] "ATTENTION USERS. . . Some Providers' Opt-Out Status may end early due to COVID 19 waivers. Please contact your respective MAC for further information.. .  . . For more information on the opt-out process, see Manage Your Enrollment or view the FAQ section below.. .  . . The Opt Out Affidavits dataset provides information on providers who have decided not to participate in Medicare. It contains provider's NPI, specialty, address, and effective dates."
#> 
#> $periodicity
#> [1] "Monthly [R/P1M]"
#> 
#> $contact
#> [1] "Provider Enrollment Data Requests - CPI (mailto:ProviderEnrollmentDataRequests@cms.hhs.gov)"
#> 
#> $dictionary
#> [1] "https://data.cms.gov/resources/opt-out-affidavits-data-dictionary"
#> 
#> $identifier
#> [1] "https://data.cms.gov/data-api/v1/dataset/9887a515-7552-4693-bf58-735c77af46d7/data-viewer"
#> 
#> $modified
#> [1] "2025-02-27"
#> 
#> $site
#> [1] "https://data.cms.gov/provider-characteristics/medicare-provider-supplier-enrollment/opt-out-affidavits"
#> 
#> $temporal
#> [1] "2023-04-01/2025-01-31"
#> 
#> $references
#> [1] "https://data.cms.gov/resources/opt-out-affidavits-methodology"
#> 
main_current("laboratories")
#> $title
#> [1] "Provider of Services File - Clinical Laboratories"
#> 
#> $description
#> [1] "The Provider of Services (POS) Clinical Laboratories (CLIA) data provides information on CLIA demographics and types of testing services the facility provides. In this file you will find provider number, name, address and characteristics of the participating institution providers."
#> 
#> $periodicity
#> [1] "Quarterly [R/P3M]"
#> 
#> $contact
#> [1] "QIESData - CCSQ (mailto:QIESData@cms.hhs.gov)"
#> 
#> $dictionary
#> [1] "https://data.cms.gov/sites/default/files/2023-07/50eacc07-7930-4e62-97e3-e04a2dc1e2a3/P.QWB.POSQ.CLIA.LAYOUT.MAR23.pdf"
#> 
#> $identifier
#> [1] "https://data.cms.gov/data-api/v1/dataset/d3eb38ac-d8e9-40d3-b7b7-6205d3d1dc16/data-viewer"
#> 
#> $modified
#> [1] "2025-01-16"
#> 
#> $site
#> [1] "https://data.cms.gov/provider-characteristics/hospitals-and-other-facilities/provider-of-services-file-clinical-laboratories"
#> 
#> $temporal
#> [1] "2011-10-01/2024-12-31"
#> 
#> $references
#> [1] "https://data.cms.gov/resources/provider-of-services-file-methodology"
#>