Load <TemporalMainGroup> API Endpoints

Load <TemporalMainGroup> API Endpoints

Usage

main_temporal_group(alias)

Arguments

alias

<chr> dataset title

Value

<TemporalMainGroup> object

Examples

main_temporal_group("utilization")
#> $title
#> [1] "Medicare Physician & Other Practitioners"
#> 
#> $periodicity
#> [1] "Annually [R/P1Y]"
#> 
#> $contact
#> [1] "Medicare Provider Data - OEDA (mailto:IPAG_Data_Products@cms.hhs.gov)"
#> 
#> $by_geography_and_service
#> $by_geography_and_service$description
#> [1] "The Medicare Physician & Other Practitioners by Geography and Service dataset contains information on use, payments, submitted charges, and beneficiary demographic and health characteristics organized by geography, Healthcare Common Procedure Coding System (HCPCS) code, and place of service."
#> 
#> $by_geography_and_service$dictionary
#> [1] "https://data.cms.gov/resources/medicare-physician-other-practitioners-by-geography-and-service-data-dictionary"
#> 
#> $by_geography_and_service$site
#> [1] "https://data.cms.gov/provider-summary-by-type-of-service/medicare-physician-other-practitioners/medicare-physician-other-practitioners-by-geography-and-service"
#> 
#> $by_geography_and_service$rows
#> [1] 270673
#> 
#> $by_geography_and_service$pages
#> [1] 55
#> 
#> $by_geography_and_service$fields
#>  [1] "Rndrng_Prvdr_Geo_Lvl"  "Rndrng_Prvdr_Geo_Cd"   "Rndrng_Prvdr_Geo_Desc"
#>  [4] "HCPCS_Cd"              "HCPCS_Desc"            "HCPCS_Drug_Ind"       
#>  [7] "Place_Of_Srvc"         "Tot_Rndrng_Prvdrs"     "Tot_Benes"            
#> [10] "Tot_Srvcs"             "Tot_Bene_Day_Srvcs"    "Avg_Sbmtd_Chrg"       
#> [13] "Avg_Mdcr_Alowd_Amt"    "Avg_Mdcr_Pymt_Amt"     "Avg_Mdcr_Stdzd_Amt"   
#> 
#> $by_geography_and_service$endpoints
#> # A tibble: 10 × 3
#>     year identifier                                                    resources
#>    <int> <chr>                                                         <chr>    
#>  1  2022 https://data.cms.gov/data-api/v1/dataset/87304f15-9ed0-41dc-… https://…
#>  2  2021 https://data.cms.gov/data-api/v1/dataset/f00f462f-bb61-48b1-… https://…
#>  3  2020 https://data.cms.gov/data-api/v1/dataset/31eb3018-43e0-4259-… https://…
#>  4  2019 https://data.cms.gov/data-api/v1/dataset/673030ae-ceed-4561-… https://…
#>  5  2018 https://data.cms.gov/data-api/v1/dataset/05a85700-052f-4509-… https://…
#>  6  2017 https://data.cms.gov/data-api/v1/dataset/8e96a9f2-ce6e-46fd-… https://…
#>  7  2016 https://data.cms.gov/data-api/v1/dataset/c7d3f18c-2f00-4553-… https://…
#>  8  2015 https://data.cms.gov/data-api/v1/dataset/dbee9609-2c90-43ca-… https://…
#>  9  2014 https://data.cms.gov/data-api/v1/dataset/28181bd2-b377-4003-… https://…
#> 10  2013 https://data.cms.gov/data-api/v1/dataset/3c2a4756-0a8c-4e4d-… https://…
#> 
#> 
#> $by_provider
#> $by_provider$description
#> [1] "The Medicare Physician & Other Practitioners by Provider dataset provides information on use, payments, submitted charges and beneficiary demographic and health characteristics organized by National Provider Identifier (NPI)."
#> 
#> $by_provider$dictionary
#> [1] "https://data.cms.gov/resources/medicare-physician-other-practitioners-by-provider-data-dictionary"
#> 
#> $by_provider$site
#> [1] "https://data.cms.gov/provider-summary-by-type-of-service/medicare-physician-other-practitioners/medicare-physician-other-practitioners-by-provider"
#> 
#> $by_provider$rows
#> [1] 1230293
#> 
#> $by_provider$pages
#> [1] 247
#> 
#> $by_provider$fields
#>  [1] "Rndrng_NPI"                       "Rndrng_Prvdr_Last_Org_Name"      
#>  [3] "Rndrng_Prvdr_First_Name"          "Rndrng_Prvdr_MI"                 
#>  [5] "Rndrng_Prvdr_Crdntls"             "Rndrng_Prvdr_Gndr"               
#>  [7] "Rndrng_Prvdr_Ent_Cd"              "Rndrng_Prvdr_St1"                
#>  [9] "Rndrng_Prvdr_St2"                 "Rndrng_Prvdr_City"               
#> [11] "Rndrng_Prvdr_State_Abrvtn"        "Rndrng_Prvdr_State_FIPS"         
#> [13] "Rndrng_Prvdr_Zip5"                "Rndrng_Prvdr_RUCA"               
#> [15] "Rndrng_Prvdr_RUCA_Desc"           "Rndrng_Prvdr_Cntry"              
#> [17] "Rndrng_Prvdr_Type"                "Rndrng_Prvdr_Mdcr_Prtcptg_Ind"   
#> [19] "Tot_HCPCS_Cds"                    "Tot_Benes"                       
#> [21] "Tot_Srvcs"                        "Tot_Sbmtd_Chrg"                  
#> [23] "Tot_Mdcr_Alowd_Amt"               "Tot_Mdcr_Pymt_Amt"               
#> [25] "Tot_Mdcr_Stdzd_Amt"               "Drug_Sprsn_Ind"                  
#> [27] "Drug_Tot_HCPCS_Cds"               "Drug_Tot_Benes"                  
#> [29] "Drug_Tot_Srvcs"                   "Drug_Sbmtd_Chrg"                 
#> [31] "Drug_Mdcr_Alowd_Amt"              "Drug_Mdcr_Pymt_Amt"              
#> [33] "Drug_Mdcr_Stdzd_Amt"              "Med_Sprsn_Ind"                   
#> [35] "Med_Tot_HCPCS_Cds"                "Med_Tot_Benes"                   
#> [37] "Med_Tot_Srvcs"                    "Med_Sbmtd_Chrg"                  
#> [39] "Med_Mdcr_Alowd_Amt"               "Med_Mdcr_Pymt_Amt"               
#> [41] "Med_Mdcr_Stdzd_Amt"               "Bene_Avg_Age"                    
#> [43] "Bene_Age_LT_65_Cnt"               "Bene_Age_65_74_Cnt"              
#> [45] "Bene_Age_75_84_Cnt"               "Bene_Age_GT_84_Cnt"              
#> [47] "Bene_Feml_Cnt"                    "Bene_Male_Cnt"                   
#> [49] "Bene_Race_Wht_Cnt"                "Bene_Race_Black_Cnt"             
#> [51] "Bene_Race_API_Cnt"                "Bene_Race_Hspnc_Cnt"             
#> [53] "Bene_Race_NatInd_Cnt"             "Bene_Race_Othr_Cnt"              
#> [55] "Bene_Dual_Cnt"                    "Bene_Ndual_Cnt"                  
#> [57] "Bene_CC_BH_ADHD_OthCD_V1_Pct"     "Bene_CC_BH_Alcohol_Drug_V1_Pct"  
#> [59] "Bene_CC_BH_Tobacco_V1_Pct"        "Bene_CC_BH_Alz_NonAlzdem_V2_Pct" 
#> [61] "Bene_CC_BH_Anxiety_V1_Pct"        "Bene_CC_BH_Bipolar_V1_Pct"       
#> [63] "Bene_CC_BH_Mood_V2_Pct"           "Bene_CC_BH_Depress_V1_Pct"       
#> [65] "Bene_CC_BH_PD_V1_Pct"             "Bene_CC_BH_PTSD_V1_Pct"          
#> [67] "Bene_CC_BH_Schizo_OthPsy_V1_Pct"  "Bene_CC_PH_Asthma_V2_Pct"        
#> [69] "Bene_CC_PH_Afib_V2_Pct"           "Bene_CC_PH_Cancer6_V2_Pct"       
#> [71] "Bene_CC_PH_CKD_V2_Pct"            "Bene_CC_PH_COPD_V2_Pct"          
#> [73] "Bene_CC_PH_Diabetes_V2_Pct"       "Bene_CC_PH_HF_NonIHD_V2_Pct"     
#> [75] "Bene_CC_PH_Hyperlipidemia_V2_Pct" "Bene_CC_PH_Hypertension_V2_Pct"  
#> [77] "Bene_CC_PH_IschemicHeart_V2_Pct"  "Bene_CC_PH_Osteoporosis_V2_Pct"  
#> [79] "Bene_CC_PH_Parkinson_V2_Pct"      "Bene_CC_PH_Arthritis_V2_Pct"     
#> [81] "Bene_CC_PH_Stroke_TIA_V2_Pct"     "Bene_Avg_Risk_Scre"              
#> 
#> $by_provider$endpoints
#> # A tibble: 10 × 3
#>     year identifier                                                    resources
#>    <int> <chr>                                                         <chr>    
#>  1  2022 https://data.cms.gov/data-api/v1/dataset/8568be52-dfc9-4fe0-… https://…
#>  2  2021 https://data.cms.gov/data-api/v1/dataset/1a3a5f39-09ac-48e2-… https://…
#>  3  2020 https://data.cms.gov/data-api/v1/dataset/6c0feedd-12a5-401c-… https://…
#>  4  2019 https://data.cms.gov/data-api/v1/dataset/6859b742-ef67-4930-… https://…
#>  5  2018 https://data.cms.gov/data-api/v1/dataset/1d2877f2-2d19-454b-… https://…
#>  6  2017 https://data.cms.gov/data-api/v1/dataset/6a4c7840-eb12-400d-… https://…
#>  7  2016 https://data.cms.gov/data-api/v1/dataset/373bec8b-d766-4201-… https://…
#>  8  2015 https://data.cms.gov/data-api/v1/dataset/686ce936-9324-4241-… https://…
#>  9  2014 https://data.cms.gov/data-api/v1/dataset/6a5a5529-9b34-48b7-… https://…
#> 10  2013 https://data.cms.gov/data-api/v1/dataset/20f32dd0-a755-43cd-… https://…
#> 
#> 
#> $by_provider_and_service
#> $by_provider_and_service$description
#> [1] "The Medicare Physician & Other Practitioners by Provider and Service dataset provides information on use, payments, and submitted charges organized by National Provider Identifier (NPI), Healthcare Common Procedure Coding System (HCPCS) code, and place of service."
#> 
#> $by_provider_and_service$dictionary
#> [1] "https://data.cms.gov/resources/medicare-physician-other-practitioners-by-provider-and-service-data-dictionary"
#> 
#> $by_provider_and_service$site
#> [1] "https://data.cms.gov/provider-summary-by-type-of-service/medicare-physician-other-practitioners/medicare-physician-other-practitioners-by-provider-and-service"
#> 
#> $by_provider_and_service$rows
#> [1] 9755427
#> 
#> $by_provider_and_service$pages
#> [1] 1952
#> 
#> $by_provider_and_service$fields
#>  [1] "Rndrng_NPI"                    "Rndrng_Prvdr_Last_Org_Name"   
#>  [3] "Rndrng_Prvdr_First_Name"       "Rndrng_Prvdr_MI"              
#>  [5] "Rndrng_Prvdr_Crdntls"          "Rndrng_Prvdr_Gndr"            
#>  [7] "Rndrng_Prvdr_Ent_Cd"           "Rndrng_Prvdr_St1"             
#>  [9] "Rndrng_Prvdr_St2"              "Rndrng_Prvdr_City"            
#> [11] "Rndrng_Prvdr_State_Abrvtn"     "Rndrng_Prvdr_State_FIPS"      
#> [13] "Rndrng_Prvdr_Zip5"             "Rndrng_Prvdr_RUCA"            
#> [15] "Rndrng_Prvdr_RUCA_Desc"        "Rndrng_Prvdr_Cntry"           
#> [17] "Rndrng_Prvdr_Type"             "Rndrng_Prvdr_Mdcr_Prtcptg_Ind"
#> [19] "HCPCS_Cd"                      "HCPCS_Desc"                   
#> [21] "HCPCS_Drug_Ind"                "Place_Of_Srvc"                
#> [23] "Tot_Benes"                     "Tot_Srvcs"                    
#> [25] "Tot_Bene_Day_Srvcs"            "Avg_Sbmtd_Chrg"               
#> [27] "Avg_Mdcr_Alowd_Amt"            "Avg_Mdcr_Pymt_Amt"            
#> [29] "Avg_Mdcr_Stdzd_Amt"           
#> 
#> $by_provider_and_service$endpoints
#> # A tibble: 10 × 3
#>     year identifier                                                    resources
#>    <int> <chr>                                                         <chr>    
#>  1  2022 https://data.cms.gov/data-api/v1/dataset/4f307be4-6868-4a9e-… https://…
#>  2  2021 https://data.cms.gov/data-api/v1/dataset/5c67d835-3862-4f63-… https://…
#>  3  2020 https://data.cms.gov/data-api/v1/dataset/862ed658-1f38-4b2f-… https://…
#>  4  2019 https://data.cms.gov/data-api/v1/dataset/5fccd951-9538-48a7-… https://…
#>  5  2018 https://data.cms.gov/data-api/v1/dataset/02c0692d-e2d9-4714-… https://…
#>  6  2017 https://data.cms.gov/data-api/v1/dataset/7ebc578d-c2c7-46fd-… https://…
#>  7  2016 https://data.cms.gov/data-api/v1/dataset/5055d307-4fb3-4474-… https://…
#>  8  2015 https://data.cms.gov/data-api/v1/dataset/0ccba18d-b821-47c6-… https://…
#>  9  2014 https://data.cms.gov/data-api/v1/dataset/e6aacd22-1b89-4914-… https://…
#> 10  2013 https://data.cms.gov/data-api/v1/dataset/ebaf67d7-1572-4419-… https://…
#> 
#> 
main_temporal_group("prescribers")
#> $title
#> [1] "Medicare Part D Prescribers"
#> 
#> $periodicity
#> [1] "Annually [R/P1Y]"
#> 
#> $contact
#> [1] "Medicare Provider Data - OEDA (mailto:IPAG_Data_Products@cms.hhs.gov)"
#> 
#> $by_geography_and_drug
#> $by_geography_and_drug$description
#> [1] "The Medicare Part D Prescribers by Geography and Drug dataset contains information on prescription drugs prescribed by individual physicians and other health care providers and paid for under the Medicare Part D Prescription Drug Program. For each drug, the dataset includes the total number of prescriptions that were dispensed, which include original prescriptions and any refills, and the total drug cost. The total drug cost includes the ingredient cost of the medication, dispensing fees, sales tax, and any applicable administration fees and is based on the amount paid by the Part D plan, Medicare beneficiary, government subsidies, and any other third-party payers."
#> 
#> $by_geography_and_drug$dictionary
#> [1] "https://data.cms.gov/resources/medicare-part-d-prescribers-by-geography-and-drug-data-dictionary"
#> 
#> $by_geography_and_drug$site
#> [1] "https://data.cms.gov/provider-summary-by-type-of-service/medicare-part-d-prescribers/medicare-part-d-prescribers-by-geography-and-drug"
#> 
#> $by_geography_and_drug$rows
#> [1] 115396
#> 
#> $by_geography_and_drug$pages
#> [1] 24
#> 
#> $by_geography_and_drug$fields
#>  [1] "Prscrbr_Geo_Lvl"      "Prscrbr_Geo_Cd"       "Prscrbr_Geo_Desc"    
#>  [4] "Brnd_Name"            "Gnrc_Name"            "Tot_Prscrbrs"        
#>  [7] "Tot_Clms"             "Tot_30day_Fills"      "Tot_Drug_Cst"        
#> [10] "Tot_Benes"            "GE65_Sprsn_Flag"      "GE65_Tot_Clms"       
#> [13] "GE65_Tot_30day_Fills" "GE65_Tot_Drug_Cst"    "GE65_Bene_Sprsn_Flag"
#> [16] "GE65_Tot_Benes"       "LIS_Bene_Cst_Shr"     "NonLIS_Bene_Cst_Shr" 
#> [19] "Opioid_Drug_Flag"     "Opioid_LA_Drug_Flag"  "Antbtc_Drug_Flag"    
#> [22] "Antpsyct_Drug_Flag"  
#> 
#> $by_geography_and_drug$endpoints
#> # A tibble: 10 × 3
#>     year identifier                                                    resources
#>    <int> <chr>                                                         <chr>    
#>  1  2022 https://data.cms.gov/data-api/v1/dataset/1fc57194-a51d-4864-… https://…
#>  2  2021 https://data.cms.gov/data-api/v1/dataset/7dda2a9d-034a-446a-… https://…
#>  3  2020 https://data.cms.gov/data-api/v1/dataset/83891e77-99cf-4865-… https://…
#>  4  2019 https://data.cms.gov/data-api/v1/dataset/73a6335e-f16f-4c81-… https://…
#>  5  2018 https://data.cms.gov/data-api/v1/dataset/b083c9b6-b841-4676-… https://…
#>  6  2017 https://data.cms.gov/data-api/v1/dataset/7863e41c-3f6f-4889-… https://…
#>  7  2016 https://data.cms.gov/data-api/v1/dataset/c5b3c840-5eb2-4e4c-… https://…
#>  8  2015 https://data.cms.gov/data-api/v1/dataset/73748d0b-2acb-420a-… https://…
#>  9  2014 https://data.cms.gov/data-api/v1/dataset/63099eb4-6c9e-4fec-… https://…
#> 10  2013 https://data.cms.gov/data-api/v1/dataset/9aec2a70-4ef2-438a-… https://…
#> 
#> 
#> $by_provider
#> $by_provider$description
#> [1] "The Medicare Part D Prescribers by Provider dataset contains information on prescription drugs prescribed by individual physicians and other health care providers and paid for under the Medicare Part D Prescription Drug Program. The dataset identifies providers by their National Provider Identifier (NPI) and summarizes for each prescriber the total number of prescriptions that were dispensed, which include original prescriptions and any refills, and the total drug cost."
#> 
#> $by_provider$dictionary
#> [1] "https://data.cms.gov/resources/medicare-part-d-prescribers-by-provider-data-dictionary"
#> 
#> $by_provider$site
#> [1] "https://data.cms.gov/provider-summary-by-type-of-service/medicare-part-d-prescribers/medicare-part-d-prescribers-by-provider"
#> 
#> $by_provider$rows
#> [1] 1332309
#> 
#> $by_provider$pages
#> [1] 267
#> 
#> $by_provider$fields
#>  [1] "Prscrbr_NPI"                    "Prscrbr_Last_Org_Name"         
#>  [3] "Prscrbr_First_Name"             "Prscrbr_MI"                    
#>  [5] "Prscrbr_Crdntls"                "Prscrbr_Gndr"                  
#>  [7] "Prscrbr_Ent_Cd"                 "Prscrbr_St1"                   
#>  [9] "Prscrbr_St2"                    "Prscrbr_City"                  
#> [11] "Prscrbr_State_Abrvtn"           "Prscrbr_State_FIPS"            
#> [13] "Prscrbr_zip5"                   "Prscrbr_RUCA"                  
#> [15] "Prscrbr_RUCA_Desc"              "Prscrbr_Cntry"                 
#> [17] "Prscrbr_Type"                   "Prscrbr_Type_Src"              
#> [19] "Tot_Clms"                       "Tot_30day_Fills"               
#> [21] "Tot_Drug_Cst"                   "Tot_Day_Suply"                 
#> [23] "Tot_Benes"                      "GE65_Sprsn_Flag"               
#> [25] "GE65_Tot_Clms"                  "GE65_Tot_30day_Fills"          
#> [27] "GE65_Tot_Drug_Cst"              "GE65_Tot_Day_Suply"            
#> [29] "GE65_Bene_Sprsn_Flag"           "GE65_Tot_Benes"                
#> [31] "Brnd_Sprsn_Flag"                "Brnd_Tot_Clms"                 
#> [33] "Brnd_Tot_Drug_Cst"              "Gnrc_Sprsn_Flag"               
#> [35] "Gnrc_Tot_Clms"                  "Gnrc_Tot_Drug_Cst"             
#> [37] "Othr_Sprsn_Flag"                "Othr_Tot_Clms"                 
#> [39] "Othr_Tot_Drug_Cst"              "MAPD_Sprsn_Flag"               
#> [41] "MAPD_Tot_Clms"                  "MAPD_Tot_Drug_Cst"             
#> [43] "PDP_Sprsn_Flag"                 "PDP_Tot_Clms"                  
#> [45] "PDP_Tot_Drug_Cst"               "LIS_Sprsn_Flag"                
#> [47] "LIS_Tot_Clms"                   "LIS_Drug_Cst"                  
#> [49] "NonLIS_Sprsn_Flag"              "NonLIS_Tot_Clms"               
#> [51] "NonLIS_Drug_Cst"                "Opioid_Tot_Clms"               
#> [53] "Opioid_Tot_Drug_Cst"            "Opioid_Tot_Suply"              
#> [55] "Opioid_Tot_Benes"               "Opioid_Prscrbr_Rate"           
#> [57] "Opioid_LA_Tot_Clms"             "Opioid_LA_Tot_Drug_Cst"        
#> [59] "Opioid_LA_Tot_Suply"            "Opioid_LA_Tot_Benes"           
#> [61] "Opioid_LA_Prscrbr_Rate"         "Antbtc_Tot_Clms"               
#> [63] "Antbtc_Tot_Drug_Cst"            "Antbtc_Tot_Benes"              
#> [65] "Antpsyct_GE65_Sprsn_Flag"       "Antpsyct_GE65_Tot_Clms"        
#> [67] "Antpsyct_GE65_Tot_Drug_Cst"     "Antpsyct_GE65_Bene_Suprsn_Flag"
#> [69] "Antpsyct_GE65_Tot_Benes"        "Bene_Avg_Age"                  
#> [71] "Bene_Age_LT_65_Cnt"             "Bene_Age_65_74_Cnt"            
#> [73] "Bene_Age_75_84_Cnt"             "Bene_Age_GT_84_Cnt"            
#> [75] "Bene_Feml_Cnt"                  "Bene_Male_Cnt"                 
#> [77] "Bene_Race_Wht_Cnt"              "Bene_Race_Black_Cnt"           
#> [79] "Bene_Race_Api_Cnt"              "Bene_Race_Hspnc_Cnt"           
#> [81] "Bene_Race_Natind_Cnt"           "Bene_Race_Othr_Cnt"            
#> [83] "Bene_Dual_Cnt"                  "Bene_Ndual_Cnt"                
#> [85] "Bene_Avg_Risk_Scre"            
#> 
#> $by_provider$endpoints
#> # A tibble: 10 × 3
#>     year identifier                                                    resources
#>    <int> <chr>                                                         <chr>    
#>  1  2022 https://data.cms.gov/data-api/v1/dataset/5b7298b5-4a3b-4517-… https://…
#>  2  2021 https://data.cms.gov/data-api/v1/dataset/c55024e4-4b4f-4106-… https://…
#>  3  2020 https://data.cms.gov/data-api/v1/dataset/eca92c9a-cc1b-4d6e-… https://…
#>  4  2019 https://data.cms.gov/data-api/v1/dataset/b6c4cbfd-0f23-4a38-… https://…
#>  5  2018 https://data.cms.gov/data-api/v1/dataset/d9bc3449-89ff-4d8f-… https://…
#>  6  2017 https://data.cms.gov/data-api/v1/dataset/a850b96d-3a91-4833-… https://…
#>  7  2016 https://data.cms.gov/data-api/v1/dataset/19f9331a-01d9-411b-… https://…
#>  8  2015 https://data.cms.gov/data-api/v1/dataset/a993311d-d020-47e8-… https://…
#>  9  2014 https://data.cms.gov/data-api/v1/dataset/759dba06-95bf-46d6-… https://…
#> 10  2013 https://data.cms.gov/data-api/v1/dataset/6e8cdefe-5e6d-4b4b-… https://…
#> 
#> 
#> $by_provider_and_drug
#> $by_provider_and_drug$description
#> [1] "The Medicare Part D Prescribers by Provider and Drug dataset provides information on prescription drugs prescribed to Medicare beneficiaries enrolled in Part D by physicians and other health care providers. This dataset contains the total number of prescription fills that were dispensed and the total drug cost paid organized by prescribing National Provider Identifier (NPI), drug brand name (if applicable) and drug generic name."
#> 
#> $by_provider_and_drug$dictionary
#> [1] "https://data.cms.gov/resources/medicare-part-d-prescribers-by-provider-and-drug-data-dictionary"
#> 
#> $by_provider_and_drug$site
#> [1] "https://data.cms.gov/provider-summary-by-type-of-service/medicare-part-d-prescribers/medicare-part-d-prescribers-by-provider-and-drug"
#> 
#> $by_provider_and_drug$rows
#> [1] 25869521
#> 
#> $by_provider_and_drug$pages
#> [1] 5174
#> 
#> $by_provider_and_drug$fields
#>  [1] "Prscrbr_NPI"           "Prscrbr_Last_Org_Name" "Prscrbr_First_Name"   
#>  [4] "Prscrbr_City"          "Prscrbr_State_Abrvtn"  "Prscrbr_State_FIPS"   
#>  [7] "Prscrbr_Type"          "Prscrbr_Type_Src"      "Brnd_Name"            
#> [10] "Gnrc_Name"             "Tot_Clms"              "Tot_30day_Fills"      
#> [13] "Tot_Day_Suply"         "Tot_Drug_Cst"          "Tot_Benes"            
#> [16] "GE65_Sprsn_Flag"       "GE65_Tot_Clms"         "GE65_Tot_30day_Fills" 
#> [19] "GE65_Tot_Drug_Cst"     "GE65_Tot_Day_Suply"    "GE65_Bene_Sprsn_Flag" 
#> [22] "GE65_Tot_Benes"       
#> 
#> $by_provider_and_drug$endpoints
#> # A tibble: 10 × 3
#>     year identifier                                                    resources
#>    <int> <chr>                                                         <chr>    
#>  1  2022 https://data.cms.gov/data-api/v1/dataset/b101b457-ffa4-49bb-… https://…
#>  2  2021 https://data.cms.gov/data-api/v1/dataset/f68114ed-f854-4ffc-… https://…
#>  3  2020 https://data.cms.gov/data-api/v1/dataset/7795fe20-e80e-435a-… https://…
#>  4  2019 https://data.cms.gov/data-api/v1/dataset/2a6705e6-7a1e-460c-… https://…
#>  5  2018 https://data.cms.gov/data-api/v1/dataset/802fe556-311f-4962-… https://…
#>  6  2017 https://data.cms.gov/data-api/v1/dataset/05f108dd-76c4-49f4-… https://…
#>  7  2016 https://data.cms.gov/data-api/v1/dataset/25106f9d-0eb8-4ba7-… https://…
#>  8  2015 https://data.cms.gov/data-api/v1/dataset/1d650894-8afe-4056-… https://…
#>  9  2014 https://data.cms.gov/data-api/v1/dataset/0779bc8d-18dd-40b8-… https://…
#> 10  2013 https://data.cms.gov/data-api/v1/dataset/c6905d43-45de-470d-… https://…
#> 
#> 
main_temporal_group("suppliers")
#> $title
#> [1] "Medicare Durable Medical Equipment, Devices & Supplies"
#> 
#> $periodicity
#> [1] "Annually [R/P1Y]"
#> 
#> $contact
#> [1] "Medicare Provider Data - OEDA (mailto:IPAG_Data_Products@cms.hhs.gov)"
#> 
#> $by_geography_and_service
#> $by_geography_and_service$description
#> [1] "The Medicare Durable Medical Equipment, Devices & Supplies by Geography and Service dataset contains information on usage, payments, and submitted charges organized by geography, Healthcare Common Procedure Coding System (HCPCS) code, and supplier rental indicator."
#> 
#> $by_geography_and_service$dictionary
#> [1] "https://data.cms.gov/resources/medicare-durable-medical-equipment-devices-supplies-by-geography-and-service-data-dictionary-2022"
#> 
#> $by_geography_and_service$site
#> [1] "https://data.cms.gov/provider-summary-by-type-of-service/medicare-durable-medical-equipment-devices-supplies/medicare-durable-medical-equipment-devices-supplies-by-geography-and-service"
#> 
#> $by_geography_and_service$rows
#> [1] 39129
#> 
#> $by_geography_and_service$pages
#> [1] 8
#> 
#> $by_geography_and_service$fields
#>  [1] "Rfrg_Prvdr_Geo_Lvl"       "Rfrg_Prvdr_Geo_Cd"       
#>  [3] "Rfrg_Prvdr_Geo_Desc"      "RBCS_Lvl"                
#>  [5] "RBCS_Id"                  "RBCS_Desc"               
#>  [7] "HCPCS_Cd"                 "HCPCS_Desc"              
#>  [9] "Suplr_Rentl_Ind"          "Tot_Rfrg_Prvdrs"         
#> [11] "Tot_Suplrs"               "Tot_Suplr_Benes"         
#> [13] "Tot_Suplr_Clms"           "Tot_Suplr_Srvcs"         
#> [15] "Avg_Suplr_Sbmtd_Chrg"     "Avg_Suplr_Mdcr_Alowd_Amt"
#> [17] "Avg_Suplr_Mdcr_Pymt_Amt"  "Avg_Suplr_Mdcr_Stdzd_Amt"
#> 
#> $by_geography_and_service$endpoints
#> # A tibble: 9 × 3
#>    year identifier                                                     resources
#>   <int> <chr>                                                          <chr>    
#> 1  2022 https://data.cms.gov/data-api/v1/dataset/a9d1461a-9c63-49e1-8… https://…
#> 2  2021 https://data.cms.gov/data-api/v1/dataset/daedccd6-5881-4997-9… https://…
#> 3  2020 https://data.cms.gov/data-api/v1/dataset/c5da241c-71d2-4b04-9… https://…
#> 4  2019 https://data.cms.gov/data-api/v1/dataset/3931609c-3f1d-4ab4-9… https://…
#> 5  2018 https://data.cms.gov/data-api/v1/dataset/bc473128-1fae-409f-8… https://…
#> 6  2017 https://data.cms.gov/data-api/v1/dataset/55a97c7d-082f-4605-a… https://…
#> 7  2016 https://data.cms.gov/data-api/v1/dataset/cb61a768-3ded-47f6-b… https://…
#> 8  2015 https://data.cms.gov/data-api/v1/dataset/e679ad08-6e14-4857-9… https://…
#> 9  2014 https://data.cms.gov/data-api/v1/dataset/52e4df6d-ea54-4fe3-b… https://…
#> 
#> 
#> $by_referring_provider
#> $by_referring_provider$description
#> [1] "The Medicare Durable Medical Equipment, Devices & Supplies by Referring Provider dataset contains information on usage, payments, submitted charges and beneficiary demographic and health characteristics organized by National Provider Identifier (NPI)."
#> 
#> $by_referring_provider$dictionary
#> [1] "https://data.cms.gov/resources/medicare-durable-medical-equipment-devices-supplies-by-referring-provider-data-dictionary"
#> 
#> $by_referring_provider$site
#> [1] "https://data.cms.gov/provider-summary-by-type-of-service/medicare-durable-medical-equipment-devices-supplies/medicare-durable-medical-equipment-devices-supplies-by-referring-provider"
#> 
#> $by_referring_provider$rows
#> [1] 385637
#> 
#> $by_referring_provider$pages
#> [1] 78
#> 
#> $by_referring_provider$fields
#>  [1] "Rfrg_NPI"                         "Rfrg_Prvdr_Last_Name_Org"        
#>  [3] "Rfrg_Prvdr_First_Name"            "Rfrg_Prvdr_MI"                   
#>  [5] "Rfrg_Prvdr_Crdntls"               "Rfrg_Prvdr_Gndr"                 
#>  [7] "Rfrg_Prvdr_Ent_Cd"                "Rfrg_Prvdr_St1"                  
#>  [9] "Rfrg_Prvdr_St2"                   "Rfrg_Prvdr_City"                 
#> [11] "Rfrg_Prvdr_State_Abrvtn"          "Rfrg_Prvdr_State_FIPS"           
#> [13] "Rfrg_Prvdr_Zip5"                  "Rfrg_Prvdr_RUCA"                 
#> [15] "Rfrg_Prvdr_RUCA_Desc"             "Rfrg_Prvdr_Cntry"                
#> [17] "Rfrg_Prvdr_Spclty_Desc"           "Rfrg_Prvdr_Spclty_Srce"          
#> [19] "Tot_Suplrs"                       "Tot_Suplr_HCPCS_Cds"             
#> [21] "Tot_Suplr_Benes"                  "Tot_Suplr_Clms"                  
#> [23] "Tot_Suplr_Srvcs"                  "Suplr_Sbmtd_Chrgs"               
#> [25] "Suplr_Mdcr_Alowd_Amt"             "Suplr_Mdcr_Pymt_Amt"             
#> [27] "Suplr_Mdcr_Stdzd_Pymt_Amt"        "DME_Sprsn_Ind"                   
#> [29] "DME_Tot_Suplrs"                   "DME_Tot_Suplr_HCPCS_Cds"         
#> [31] "DME_Tot_Suplr_Benes"              "DME_Tot_Suplr_Clms"              
#> [33] "DME_Tot_Suplr_Srvcs"              "DME_Suplr_Sbmtd_Chrgs"           
#> [35] "DME_Suplr_Mdcr_Alowd_Amt"         "DME_Suplr_Mdcr_Pymt_Amt"         
#> [37] "DME_Suplr_Mdcr_Stdzd_Pymt_Amt"    "POS_Sprsn_Ind"                   
#> [39] "POS_Tot_Suplrs"                   "POS_Tot_Suplr_HCPCS_Cds"         
#> [41] "POS_Tot_Suplr_Benes"              "POS_Tot_Suplr_Clms"              
#> [43] "POS_Tot_Suplr_Srvcs"              "POS_Suplr_Sbmtd_Chrgs"           
#> [45] "POS_Suplr_Mdcr_Alowd_Amt"         "POS_Suplr_Mdcr_Pymt_Amt"         
#> [47] "POS_Suplr_Mdcr_Stdzd_Pymt_Amt"    "Drug_Sprsn_Ind"                  
#> [49] "Drug_Tot_Suplrs"                  "Drug_Tot_Suplr_HCPCS_Cds"        
#> [51] "Drug_Tot_Suplr_Benes"             "Drug_Tot_Suplr_Clms"             
#> [53] "Drug_Tot_Suplr_Srvcs"             "Drug_Suplr_Sbmtd_Chrgs"          
#> [55] "Drug_Suplr_Mdcr_Alowd_Amt"        "Drug_Suplr_Mdcr_Pymt_Amt"        
#> [57] "Drug_Suplr_Mdcr_Stdzd_Pymt_Amt"   "Bene_Avg_Age"                    
#> [59] "Bene_Age_LT_65_Cnt"               "Bene_Age_65_74_Cnt"              
#> [61] "Bene_Age_75_84_Cnt"               "Bene_Age_GT_84_Cnt"              
#> [63] "Bene_Feml_Cnt"                    "Bene_Male_Cnt"                   
#> [65] "Bene_Race_Wht_Cnt"                "Bene_Race_Black_Cnt"             
#> [67] "Bene_Race_Api_Cnt"                "Bene_Race_Hspnc_Cnt"             
#> [69] "Bene_Race_Natind_Cnt"             "Bene_Race_Othr_Cnt"              
#> [71] "Bene_Ndual_Cnt"                   "Bene_Dual_Cnt"                   
#> [73] "Bene_CC_BH_ADHD_OthCD_V1_Pct"     "Bene_CC_BH_Alcohol_Drug_V1_Pct"  
#> [75] "Bene_CC_BH_Tobacco_V1_Pct"        "Bene_CC_BH_Alz_NonAlzdem_V2_Pct" 
#> [77] "Bene_CC_BH_Anxiety_V1_Pct"        "Bene_CC_BH_Bipolar_V1_Pct"       
#> [79] "Bene_CC_BH_Mood_V2_Pct"           "Bene_CC_BH_Depress_V1_Pct"       
#> [81] "Bene_CC_BH_PD_V1_Pct"             "Bene_CC_BH_PTSD_V1_Pct"          
#> [83] "Bene_CC_BH_Schizo_OthPsy_V1_Pct"  "Bene_CC_PH_Asthma_V2_Pct"        
#> [85] "Bene_CC_PH_Afib_V2_Pct"           "Bene_CC_PH_Cancer6_V2_Pct"       
#> [87] "Bene_CC_PH_CKD_V2_Pct"            "Bene_CC_PH_COPD_V2_Pct"          
#> [89] "Bene_CC_PH_Diabetes_V2_Pct"       "Bene_CC_PH_HF_NonIHD_V2_Pct"     
#> [91] "Bene_CC_PH_Hyperlipidemia_V2_Pct" "Bene_CC_PH_Hypertension_V2_Pct"  
#> [93] "Bene_CC_PH_IschemicHeart_V2_Pct"  "Bene_CC_PH_Osteoporosis_V2_Pct"  
#> [95] "Bene_CC_PH_Parkinson_V2_Pct"      "Bene_CC_PH_Arthritis_V2_Pct"     
#> [97] "Bene_CC_PH_Stroke_TIA_V2_Pct"     "Bene_Avg_Risk_Scre"              
#> 
#> $by_referring_provider$endpoints
#> # A tibble: 9 × 3
#>    year identifier                                                     resources
#>   <int> <chr>                                                          <chr>    
#> 1  2022 https://data.cms.gov/data-api/v1/dataset/e0eb5710-5287-431b-9… https://…
#> 2  2021 https://data.cms.gov/data-api/v1/dataset/527aab51-3316-4cb1-8… https://…
#> 3  2020 https://data.cms.gov/data-api/v1/dataset/5fb6377b-41b9-4002-8… https://…
#> 4  2019 https://data.cms.gov/data-api/v1/dataset/c8cd9ce3-c005-4823-a… https://…
#> 5  2018 https://data.cms.gov/data-api/v1/dataset/78202b62-ef13-43e3-8… https://…
#> 6  2017 https://data.cms.gov/data-api/v1/dataset/28980e93-6484-4c3b-b… https://…
#> 7  2016 https://data.cms.gov/data-api/v1/dataset/4f478343-3637-420d-b… https://…
#> 8  2015 https://data.cms.gov/data-api/v1/dataset/b6469102-0edd-4ea3-a… https://…
#> 9  2014 https://data.cms.gov/data-api/v1/dataset/6215b30b-4ee3-46a4-a… https://…
#> 
#> 
#> $by_referring_provider_and_service
#> $by_referring_provider_and_service$description
#> [1] "The Medicare Durable Medical Equipment, Devices & Supplies by Referring Provider and Service dataset contains information on usage, payments and submitted charges organized by National Provider Identifier (NPI), Healthcare Common Procedure Coding System (HCPCS) code, and supplier rental indicator."
#> 
#> $by_referring_provider_and_service$dictionary
#> [1] "https://data.cms.gov/resources/medicare-durable-medical-equipment-devices-supplies-by-referring-provider-and-service-data-dictionary"
#> 
#> $by_referring_provider_and_service$site
#> [1] "https://data.cms.gov/provider-summary-by-type-of-service/medicare-durable-medical-equipment-devices-supplies/medicare-durable-medical-equipment-devices-supplies-by-referring-provider-and-service"
#> 
#> $by_referring_provider_and_service$rows
#> [1] 1457378
#> 
#> $by_referring_provider_and_service$pages
#> [1] 292
#> 
#> $by_referring_provider_and_service$fields
#>  [1] "Rfrg_NPI"                 "Rfrg_Prvdr_Last_Name_Org"
#>  [3] "Rfrg_Prvdr_First_Name"    "Rfrg_Prvdr_MI"           
#>  [5] "Rfrg_Prvdr_Crdntls"       "Rfrg_Prvdr_Gndr"         
#>  [7] "Rfrg_Prvdr_Ent_Cd"        "Rfrg_Prvdr_St1"          
#>  [9] "Rfrg_Prvdr_St2"           "Rfrg_Prvdr_City"         
#> [11] "Rfrg_Prvdr_State_Abrvtn"  "Rfrg_Prvdr_State_FIPS"   
#> [13] "Rfrg_Prvdr_Zip5"          "Rfrg_Prvdr_RUCA_Cat"     
#> [15] "Rfrg_Prvdr_RUCA"          "Rfrg_Prvdr_RUCA_Desc"    
#> [17] "Rfrg_Prvdr_Cntry"         "Rfrg_Prvdr_Spclty_Cd"    
#> [19] "Rfrg_Prvdr_Spclty_Desc"   "Rfrg_Prvdr_Spclty_Srce"  
#> [21] "RBCS_Lvl"                 "RBCS_Id"                 
#> [23] "RBCS_Desc"                "HCPCS_CD"                
#> [25] "HCPCS_Desc"               "Suplr_Rentl_Ind"         
#> [27] "Tot_Suplrs"               "Tot_Suplr_Benes"         
#> [29] "Tot_Suplr_Clms"           "Tot_Suplr_Srvcs"         
#> [31] "Avg_Suplr_Sbmtd_Chrg"     "Avg_Suplr_Mdcr_Alowd_Amt"
#> [33] "Avg_Suplr_Mdcr_Pymt_Amt"  "Avg_Suplr_Mdcr_Stdzd_Amt"
#> 
#> $by_referring_provider_and_service$endpoints
#> # A tibble: 9 × 3
#>    year identifier                                                     resources
#>   <int> <chr>                                                          <chr>    
#> 1  2022 https://data.cms.gov/data-api/v1/dataset/7f98d91f-f2a6-44cd-a… https://…
#> 2  2021 https://data.cms.gov/data-api/v1/dataset/442050e6-3d96-47b9-b… https://…
#> 3  2020 https://data.cms.gov/data-api/v1/dataset/72ce1cd5-efa9-484e-b… https://…
#> 4  2019 https://data.cms.gov/data-api/v1/dataset/5d2c51d1-50cb-4b10-8… https://…
#> 5  2018 https://data.cms.gov/data-api/v1/dataset/f848042c-b0b8-4e7d-b… https://…
#> 6  2017 https://data.cms.gov/data-api/v1/dataset/49a59bf3-78a7-4346-a… https://…
#> 7  2016 https://data.cms.gov/data-api/v1/dataset/3749c9ac-a576-4c7e-b… https://…
#> 8  2015 https://data.cms.gov/data-api/v1/dataset/da444e4c-a713-4739-b… https://…
#> 9  2014 https://data.cms.gov/data-api/v1/dataset/04695917-1285-4063-9… https://…
#> 
#> 
#> $by_supplier
#> $by_supplier$description
#> [1] "The Medicare Durable Medical Equipment, Devices & Supplies by Supplier dataset contains information on usage, payments, submitted charges and beneficiary demographic and health characteristics organized by supplier National Provider Identifier (NPI)."
#> 
#> $by_supplier$dictionary
#> [1] "https://data.cms.gov/resources/medicare-durable-medical-equipment-devices-supplies-by-supplier-data-dictionary"
#> 
#> $by_supplier$site
#> [1] "https://data.cms.gov/provider-summary-by-type-of-service/medicare-durable-medical-equipment-devices-supplies/medicare-durable-medical-equipment-devices-supplies-by-supplier"
#> 
#> $by_supplier$rows
#> [1] 66406
#> 
#> $by_supplier$pages
#> [1] 14
#> 
#> $by_supplier$fields
#>  [1] "Suplr_NPI"                        "Suplr_Prvdr_Last_Name_Org"       
#>  [3] "Suplr_Prvdr_First_Name"           "Suplr_Prvdr_MI"                  
#>  [5] "Suplr_Prvdr_Crdntls"              "Suplr_Prvdr_Gndr"                
#>  [7] "Suplr_Prvdr_Ent_Cd"               "Suplr_Prvdr_St1"                 
#>  [9] "Suplr_Prvdr_St2"                  "Suplr_Prvdr_City"                
#> [11] "Suplr_Prvdr_State_Abrvtn"         "Suplr_Prvdr_State_FIPS"          
#> [13] "Suplr_Prvdr_Zip5"                 "Suplr_Prvdr_RUCA"                
#> [15] "Suplr_Prvdr_RUCA_Desc"            "Suplr_Prvdr_Cntry"               
#> [17] "Suplr_Prvdr_Spclty_Desc"          "Suplr_Prvdr_Spclty_Srce"         
#> [19] "Tot_Suplr_HCPCS_Cds"              "Tot_Suplr_Benes"                 
#> [21] "Tot_Suplr_Clms"                   "Tot_Suplr_Srvcs"                 
#> [23] "Suplr_Sbmtd_Chrgs"                "Suplr_Mdcr_Alowd_Amt"            
#> [25] "Suplr_Mdcr_Pymt_Amt"              "Suplr_Mdcr_Stdzd_Pymt_Amt"       
#> [27] "DME_Sprsn_Ind"                    "DME_Tot_Suplr_HCPCS_Cds"         
#> [29] "DME_Tot_Suplr_Benes"              "DME_Tot_Suplr_Clms"              
#> [31] "DME_Tot_Suplr_Srvcs"              "DME_Suplr_Sbmtd_Chrgs"           
#> [33] "DME_Suplr_Mdcr_Alowd_Amt"         "DME_Suplr_Mdcr_Pymt_Amt"         
#> [35] "DME_Suplr_Mdcr_Stdzd_Pymt_Amt"    "POS_Sprsn_Ind"                   
#> [37] "POS_Tot_Suplr_HCPCS_Cds"          "POS_Tot_Suplr_Benes"             
#> [39] "POS_Tot_Suplr_Clms"               "POS_Tot_Suplr_Srvcs"             
#> [41] "POS_Suplr_Sbmtd_Chrgs"            "POS_Suplr_Mdcr_Alowd_Amt"        
#> [43] "POS_Suplr_Mdcr_Pymt_Amt"          "POS_Suplr_Mdcr_Stdzd_Pymt_Amt"   
#> [45] "Drug_Sprsn_Ind"                   "Drug_Tot_Suplr_HCPCS_Cds"        
#> [47] "Drug_Tot_Suplr_Benes"             "Drug_Tot_Suplr_Clms"             
#> [49] "Drug_Tot_Suplr_Srvcs"             "Drug_Suplr_Sbmtd_Chrgs"          
#> [51] "Drug_Suplr_Mdcr_Alowd_Amt"        "Drug_Suplr_Mdcr_Pymt_Amt"        
#> [53] "Drug_Suplr_Mdcr_Stdzd_Pymt_Amt"   "Bene_Avg_Age"                    
#> [55] "Bene_Age_LT_65_Cnt"               "Bene_Age_65_74_Cnt"              
#> [57] "Bene_Age_75_84_Cnt"               "Bene_Age_GT_84_Cnt"              
#> [59] "Bene_Feml_Cnt"                    "Bene_Male_Cnt"                   
#> [61] "Bene_Race_Wht_Cnt"                "Bene_Race_Black_Cnt"             
#> [63] "Bene_Race_Api_Cnt"                "Bene_Race_Hspnc_Cnt"             
#> [65] "Bene_Race_Natind_Cnt"             "Bene_Race_Othr_Cnt"              
#> [67] "Bene_Ndual_Cnt"                   "Bene_Dual_Cnt"                   
#> [69] "Bene_CC_BH_ADHD_OthCD_V1_Pct"     "Bene_CC_BH_Alcohol_Drug_V1_Pct"  
#> [71] "Bene_CC_BH_Tobacco_V1_Pct"        "Bene_CC_BH_Alz_NonAlzdem_V2_Pct" 
#> [73] "Bene_CC_BH_Anxiety_V1_Pct"        "Bene_CC_BH_Bipolar_V1_Pct"       
#> [75] "Bene_CC_BH_Mood_V2_Pct"           "Bene_CC_BH_Depress_V1_Pct"       
#> [77] "Bene_CC_BH_PD_V1_Pct"             "Bene_CC_BH_PTSD_V1_Pct"          
#> [79] "Bene_CC_BH_Schizo_OthPsy_V1_Pct"  "Bene_CC_PH_Asthma_V2_Pct"        
#> [81] "Bene_CC_PH_Afib_V2_Pct"           "Bene_CC_PH_Cancer6_V2_Pct"       
#> [83] "Bene_CC_PH_CKD_V2_Pct"            "Bene_CC_PH_COPD_V2_Pct"          
#> [85] "Bene_CC_PH_Diabetes_V2_Pct"       "Bene_CC_PH_HF_NonIHD_V2_Pct"     
#> [87] "Bene_CC_PH_Hyperlipidemia_V2_Pct" "Bene_CC_PH_Hypertension_V2_Pct"  
#> [89] "Bene_CC_PH_IschemicHeart_V2_Pct"  "Bene_CC_PH_Osteoporosis_V2_Pct"  
#> [91] "Bene_CC_PH_Parkinson_V2_Pct"      "Bene_CC_PH_Arthritis_V2_Pct"     
#> [93] "Bene_CC_PH_Stroke_TIA_V2_Pct"     "Bene_Avg_Risk_Scre"              
#> 
#> $by_supplier$endpoints
#> # A tibble: 9 × 3
#>    year identifier                                                     resources
#>   <int> <chr>                                                          <chr>    
#> 1  2022 https://data.cms.gov/data-api/v1/dataset/70b8ae50-1919-4468-9… https://…
#> 2  2021 https://data.cms.gov/data-api/v1/dataset/107436f3-cc94-4469-8… https://…
#> 3  2020 https://data.cms.gov/data-api/v1/dataset/40c9de9d-18a3-4476-b… https://…
#> 4  2019 https://data.cms.gov/data-api/v1/dataset/ce5ee6de-c08b-4cf4-9… https://…
#> 5  2018 https://data.cms.gov/data-api/v1/dataset/f210bbbc-2cd3-437a-a… https://…
#> 6  2017 https://data.cms.gov/data-api/v1/dataset/e46dfb9a-e8ad-4701-a… https://…
#> 7  2016 https://data.cms.gov/data-api/v1/dataset/4f332b42-067c-46f9-b… https://…
#> 8  2015 https://data.cms.gov/data-api/v1/dataset/7b7cc52c-8cdc-4c5a-9… https://…
#> 9  2014 https://data.cms.gov/data-api/v1/dataset/07c7442a-cdc0-4b79-b… https://…
#> 
#> 
#> $by_supplier_and_service
#> $by_supplier_and_service$description
#> [1] "The Medicare Durable Medical Equipment, Devices & Supplies by Supplier and Service dataset contains information on usage, payments and submitted charges organized by supplier National Provider Identifier (NPI), Healthcare Common Procedure Coding System (HCPCS) code, and supplier rental indicator."
#> 
#> $by_supplier_and_service$dictionary
#> [1] "https://data.cms.gov/resources/medicare-durable-medical-equipment-devices-supplies-by-supplier-and-service-data-dictionary"
#> 
#> $by_supplier_and_service$site
#> [1] "https://data.cms.gov/provider-summary-by-type-of-service/medicare-durable-medical-equipment-devices-supplies/medicare-durable-medical-equipment-devices-supplies-by-supplier-and-service"
#> 
#> $by_supplier_and_service$rows
#> [1] 482638
#> 
#> $by_supplier_and_service$pages
#> [1] 97
#> 
#> $by_supplier_and_service$fields
#>  [1] "Suplr_NPI"                 "Suplr_Prvdr_Last_Name_Org"
#>  [3] "Suplr_Prvdr_First_Name"    "Suplr_Prvdr_MI"           
#>  [5] "Suplr_Prvdr_Crdntls"       "Suplr_Prvdr_Gndr"         
#>  [7] "Suplr_Prvdr_Ent_Cd"        "Suplr_Prvdr_St1"          
#>  [9] "Suplr_Prvdr_St2"           "Suplr_Prvdr_City"         
#> [11] "Suplr_Prvdr_State_Abrvtn"  "Suplr_Prvdr_State_FIPS"   
#> [13] "Suplr_Prvdr_Zip5"          "Suplr_Prvdr_RUCA_Cat"     
#> [15] "Suplr_Prvdr_RUCA"          "Suplr_Prvdr_RUCA_Desc"    
#> [17] "Suplr_Prvdr_Cntry"         "Suplr_Prvdr_Spclty_Cd"    
#> [19] "Suplr_Prvdr_Spclty_Desc"   "Suplr_Prvdr_Spclty_Srce"  
#> [21] "RBCS_Lvl"                  "RBCS_Id"                  
#> [23] "RBCS_Desc"                 "HCPCS_Cd"                 
#> [25] "HCPCS_Desc"                "Suplr_Rentl_Ind"          
#> [27] "Tot_Suplr_Benes"           "Tot_Suplr_Clms"           
#> [29] "Tot_Suplr_Srvcs"           "Avg_Suplr_Sbmtd_Chrg"     
#> [31] "Avg_Suplr_Mdcr_Alowd_Amt"  "Avg_Suplr_Mdcr_Pymt_Amt"  
#> [33] "Avg_Suplr_Mdcr_Stdzd_Amt" 
#> 
#> $by_supplier_and_service$endpoints
#> # A tibble: 9 × 3
#>    year identifier                                                     resources
#>   <int> <chr>                                                          <chr>    
#> 1  2022 https://data.cms.gov/data-api/v1/dataset/ce9ec4d2-d404-4463-9… https://…
#> 2  2021 https://data.cms.gov/data-api/v1/dataset/1ca065e5-adfa-4a43-9… https://…
#> 3  2020 https://data.cms.gov/data-api/v1/dataset/6160f96b-710a-4380-a… https://…
#> 4  2019 https://data.cms.gov/data-api/v1/dataset/ea4a5834-aeaa-4176-8… https://…
#> 5  2018 https://data.cms.gov/data-api/v1/dataset/7cd2e7bb-3843-433a-a… https://…
#> 6  2017 https://data.cms.gov/data-api/v1/dataset/bb6672c2-c1db-4b51-9… https://…
#> 7  2016 https://data.cms.gov/data-api/v1/dataset/4bf75e4a-a46c-4171-8… https://…
#> 8  2015 https://data.cms.gov/data-api/v1/dataset/4e0ad5ae-31b0-433b-a… https://…
#> 9  2014 https://data.cms.gov/data-api/v1/dataset/90aefd0d-c6ab-4306-a… https://…
#> 
#> 
main_temporal_group("outpatient")
#> $title
#> [1] "Medicare Outpatient Hospitals"
#> 
#> $periodicity
#> [1] "Annually [R/P1Y]"
#> 
#> $contact
#> [1] "Medicare Provider Data - OEDA (mailto:IPAG_Data_Products@cms.hhs.gov)"
#> 
#> $by_geography_and_service
#> $by_geography_and_service$description
#> [1] "The Medicare Outpatient Hospitals by Geography and Service dataset provides information on services for Original Medicare Part B beneficiaries by OPPS hospitals. These datasets contain information on the number of services, payments, and submitted charges organized by geography and comprehensive Ambulatory Payment Classification (APC)."
#> 
#> $by_geography_and_service$dictionary
#> [1] "https://data.cms.gov/resources/medicare-outpatient-hospitals-by-geography-and-service-data-dictionary"
#> 
#> $by_geography_and_service$site
#> [1] "https://data.cms.gov/provider-summary-by-type-of-service/medicare-outpatient-hospitals/medicare-outpatient-hospitals-by-geography-and-service"
#> 
#> $by_geography_and_service$rows
#> [1] 90591
#> 
#> $by_geography_and_service$pages
#> [1] 19
#> 
#> $by_geography_and_service$fields
#>  [1] "Rndrng_Prvdr_Geo_Lvl"  "Rndrng_Prvdr_Geo_Cd"   "Rndrng_Prvdr_Geo_Desc"
#>  [4] "Srvc_Lvl"              "APC_Cd"                "APC_Desc"             
#>  [7] "HCPCS_Cd"              "HCPCS_Desc"            "Bene_Cnt"             
#> [10] "CAPC_Srvcs"            "Avg_Tot_Sbmtd_Chrgs"   "Avg_Mdcr_Alowd_Amt"   
#> [13] "Avg_Mdcr_Pymt_Amt"     "Outlier_Srvcs"         "Avg_Mdcr_Outlier_Amt" 
#> 
#> $by_geography_and_service$endpoints
#> # A tibble: 8 × 3
#>    year identifier                                                     resources
#>   <int> <chr>                                                          <chr>    
#> 1  2022 https://data.cms.gov/data-api/v1/dataset/6af002ba-b934-4491-9… https://…
#> 2  2021 https://data.cms.gov/data-api/v1/dataset/b3537e90-c3c5-4895-9… https://…
#> 3  2020 https://data.cms.gov/data-api/v1/dataset/31db4293-f75c-4940-b… https://…
#> 4  2019 https://data.cms.gov/data-api/v1/dataset/46ae348d-8ba4-43df-8… https://…
#> 5  2018 https://data.cms.gov/data-api/v1/dataset/bb7e2342-b0cd-4c79-b… https://…
#> 6  2017 https://data.cms.gov/data-api/v1/dataset/23f747b4-381d-4887-b… https://…
#> 7  2016 https://data.cms.gov/data-api/v1/dataset/301f5c5e-4ec3-4d59-8… https://…
#> 8  2015 https://data.cms.gov/data-api/v1/dataset/4409fc98-6ae1-4e1a-8… https://…
#> 
#> 
#> $by_provider_and_service
#> $by_provider_and_service$description
#> [1] "The Medicare Outpatient Hospitals by Provider and Service dataset provides information on services for Original Medicare Part B beneficiaries by OPPS hospitals. These datasets contain information on the number of services, payments, and submitted charges organized by provider CMS Certified Number (CCN) and comprehensive Ambulatory Payment Classification (APC)."
#> 
#> $by_provider_and_service$dictionary
#> [1] "https://data.cms.gov/resources/medicare-outpatient-hospitals-by-provider-and-service-data-dictionary"
#> 
#> $by_provider_and_service$site
#> [1] "https://data.cms.gov/provider-summary-by-type-of-service/medicare-outpatient-hospitals/medicare-outpatient-hospitals-by-provider-and-service"
#> 
#> $by_provider_and_service$rows
#> [1] 116459
#> 
#> $by_provider_and_service$pages
#> [1] 24
#> 
#> $by_provider_and_service$fields
#>  [1] "Rndrng_Prvdr_CCN"          "Rndrng_Prvdr_Org_Name"    
#>  [3] "Rndrng_Prvdr_St"           "Rndrng_Prvdr_City"        
#>  [5] "Rndrng_Prvdr_State_Abrvtn" "Rndrng_Prvdr_State_FIPS"  
#>  [7] "Rndrng_Prvdr_Zip5"         "Rndrng_Prvdr_RUCA"        
#>  [9] "Rndrng_Prvdr_RUCA_Desc"    "APC_Cd"                   
#> [11] "APC_Desc"                  "Bene_Cnt"                 
#> [13] "CAPC_Srvcs"                "Avg_Tot_Sbmtd_Chrgs"      
#> [15] "Avg_Mdcr_Alowd_Amt"        "Avg_Mdcr_Pymt_Amt"        
#> [17] "Outlier_Srvcs"             "Avg_Mdcr_Outlier_Amt"     
#> 
#> $by_provider_and_service$endpoints
#> # A tibble: 8 × 3
#>    year identifier                                                     resources
#>   <int> <chr>                                                          <chr>    
#> 1  2022 https://data.cms.gov/data-api/v1/dataset/456dbe68-8878-49dd-8… https://…
#> 2  2021 https://data.cms.gov/data-api/v1/dataset/a7617aef-046f-4557-8… https://…
#> 3  2020 https://data.cms.gov/data-api/v1/dataset/8281e9f7-1874-4bd9-a… https://…
#> 4  2019 https://data.cms.gov/data-api/v1/dataset/f3fc4327-f66e-49e1-b… https://…
#> 5  2018 https://data.cms.gov/data-api/v1/dataset/c52aa19e-38d1-40a3-b… https://…
#> 6  2017 https://data.cms.gov/data-api/v1/dataset/e5b339e2-0caa-4a01-b… https://…
#> 7  2016 https://data.cms.gov/data-api/v1/dataset/d7112a9d-aeb6-4242-a… https://…
#> 8  2015 https://data.cms.gov/data-api/v1/dataset/42381a9e-676d-4392-b… https://…
#> 
#> 
main_temporal_group("inpatient")
#> $title
#> [1] "Medicare Inpatient Hospitals"
#> 
#> $periodicity
#> [1] "Annually [R/P1Y]"
#> 
#> $contact
#> [1] "Medicare Provider Data - OEDA (mailto:IPAG_Data_Products@cms.hhs.gov)"
#> 
#> $by_geography_and_service
#> $by_geography_and_service$description
#> [1] "The Medicare Inpatient Hospitals by Geography and Service dataset provides information on hospital discharges for Original Medicare Part A beneficiaries by IPPS hospitals. This dataset contains information on the number of discharges, payments, and submitted charges organized by geography and Medicare Severity Diagnosis Related Group (DRG)."
#> 
#> $by_geography_and_service$dictionary
#> [1] "https://data.cms.gov/resources/medicare-inpatient-hospitals-by-geography-and-service-data-dictionary-0"
#> 
#> $by_geography_and_service$site
#> [1] "https://data.cms.gov/provider-summary-by-type-of-service/medicare-inpatient-hospitals/medicare-inpatient-hospitals-by-geography-and-service"
#> 
#> $by_geography_and_service$rows
#> [1] 26260
#> 
#> $by_geography_and_service$pages
#> [1] 6
#> 
#> $by_geography_and_service$fields
#> [1] "Rndrng_Prvdr_Geo_Lvl"  "Rndrng_Prvdr_Geo_Cd"   "Rndrng_Prvdr_Geo_Desc"
#> [4] "DRG_Cd"                "DRG_Desc"              "Tot_Dschrgs"          
#> [7] "Avg_Submtd_Cvrd_Chrg"  "Avg_Tot_Pymt_Amt"      "Avg_Mdcr_Pymt_Amt"    
#> 
#> $by_geography_and_service$endpoints
#> # A tibble: 10 × 3
#>     year identifier                                                    resources
#>    <int> <chr>                                                         <chr>    
#>  1  2022 https://data.cms.gov/data-api/v1/dataset/5c8b315f-a2a3-4b2f-… https://…
#>  2  2021 https://data.cms.gov/data-api/v1/dataset/68739613-1a4f-40ae-… https://…
#>  3  2020 https://data.cms.gov/data-api/v1/dataset/e583de38-33d7-4895-… https://…
#>  4  2019 https://data.cms.gov/data-api/v1/dataset/337617e5-c18b-4cb9-… https://…
#>  5  2018 https://data.cms.gov/data-api/v1/dataset/25b5b6c9-977d-4f7a-… https://…
#>  6  2017 https://data.cms.gov/data-api/v1/dataset/7c638c5f-7e8c-4346-… https://…
#>  7  2016 https://data.cms.gov/data-api/v1/dataset/81dca731-b739-4a0a-… https://…
#>  8  2015 https://data.cms.gov/data-api/v1/dataset/98069693-d269-46a1-… https://…
#>  9  2014 https://data.cms.gov/data-api/v1/dataset/fe4fa35f-d27b-4bed-… https://…
#> 10  2013 https://data.cms.gov/data-api/v1/dataset/c7019ecb-8f6f-4548-… https://…
#> 
#> 
#> $by_provider
#> $by_provider$description
#> [1] "The Medicare Inpatient Hospitals by Provider dataset provides information on inpatient discharges for Original Medicare Part A beneficiaries by IPPS hospitals. It includes information on the use, payment, and hospital charges for more than 3,000 U.S. hospitals that received IPPS payments. The data are organized by hospital."
#> 
#> $by_provider$dictionary
#> [1] "https://data.cms.gov/resources/medicare-inpatient-hospitals-by-provider-data-dictionary"
#> 
#> $by_provider$site
#> [1] "https://data.cms.gov/provider-summary-by-type-of-service/medicare-inpatient-hospitals/medicare-inpatient-hospitals-by-provider"
#> 
#> $by_provider$rows
#> [1] 3120
#> 
#> $by_provider$pages
#> [1] 1
#> 
#> $by_provider$fields
#>  [1] "Rndrng_Prvdr_CCN"                 "Rndrng_Prvdr_Org_Name"           
#>  [3] "Rndrng_Prvdr_St"                  "Rndrng_Prvdr_City"               
#>  [5] "Rndrng_Prvdr_Zip5"                "Rndrng_Prvdr_State_Abrvtn"       
#>  [7] "Rndrng_Prvdr_State_FIPS"          "Rndrng_Prvdr_RUCA"               
#>  [9] "Rndrng_Prvdr_RUCA_Desc"           "Tot_Benes"                       
#> [11] "Tot_Submtd_Cvrd_Chrg"             "Tot_Pymt_Amt"                    
#> [13] "Tot_Mdcr_Pymt_Amt"                "Tot_Dschrgs"                     
#> [15] "Tot_Cvrd_Days"                    "Tot_Days"                        
#> [17] "Bene_Avg_Age"                     "Bene_Age_LT_65_Cnt"              
#> [19] "Bene_Age_65_74_Cnt"               "Bene_Age_75_84_Cnt"              
#> [21] "Bene_Age_GT_84_Cnt"               "Bene_Feml_Cnt"                   
#> [23] "Bene_Male_Cnt"                    "Bene_Race_Wht_Cnt"               
#> [25] "Bene_Race_Black_Cnt"              "Bene_Race_API_Cnt"               
#> [27] "Bene_Race_Hspnc_Cnt"              "Bene_Race_NatInd_Cnt"            
#> [29] "Bene_Race_Othr_Cnt"               "Bene_Dual_Cnt"                   
#> [31] "Bene_Ndual_Cnt"                   "Bene_CC_BH_ADHD_OthCD_V1_Pct"    
#> [33] "Bene_CC_BH_Alcohol_Drug_V1_Pct"   "Bene_CC_BH_Tobacco_V1_Pct"       
#> [35] "Bene_CC_BH_Alz_NonAlzdem_V2_Pct"  "Bene_CC_BH_Anxiety_V1_Pct"       
#> [37] "Bene_CC_BH_Bipolar_V1_Pct"        "Bene_CC_BH_Mood_V2_Pct"          
#> [39] "Bene_CC_BH_Depress_V1_Pct"        "Bene_CC_BH_PD_V1_Pct"            
#> [41] "Bene_CC_BH_PTSD_V1_Pct"           "Bene_CC_BH_Schizo_OthPsy_V1_Pct" 
#> [43] "Bene_CC_PH_Asthma_V2_Pct"         "Bene_CC_PH_Afib_V2_Pct"          
#> [45] "Bene_CC_PH_Cancer6_V2_Pct"        "Bene_CC_PH_CKD_V2_Pct"           
#> [47] "Bene_CC_PH_COPD_V2_Pct"           "Bene_CC_PH_Diabetes_V2_Pct"      
#> [49] "Bene_CC_PH_HF_NonIHD_V2_Pct"      "Bene_CC_PH_Hyperlipidemia_V2_Pct"
#> [51] "Bene_CC_PH_Hypertension_V2_Pct"   "Bene_CC_PH_IschemicHeart_V2_Pct" 
#> [53] "Bene_CC_PH_Osteoporosis_V2_Pct"   "Bene_CC_PH_Parkinson_V2_Pct"     
#> [55] "Bene_CC_PH_Arthritis_V2_Pct"      "Bene_CC_PH_Stroke_TIA_V2_Pct"    
#> [57] "Bene_Avg_Risk_Scre"              
#> 
#> $by_provider$endpoints
#> # A tibble: 10 × 3
#>     year identifier                                                    resources
#>    <int> <chr>                                                         <chr>    
#>  1  2022 https://data.cms.gov/data-api/v1/dataset/cc73e77b-7367-48c9-… https://…
#>  2  2021 https://data.cms.gov/data-api/v1/dataset/c7e1501c-a770-4e2a-… https://…
#>  3  2020 https://data.cms.gov/data-api/v1/dataset/d7b2810d-ac10-498f-… https://…
#>  4  2019 https://data.cms.gov/data-api/v1/dataset/c349419e-3c6a-419e-… https://…
#>  5  2018 https://data.cms.gov/data-api/v1/dataset/f8f65555-1540-44ef-… https://…
#>  6  2017 https://data.cms.gov/data-api/v1/dataset/be4e1a28-8afe-40c3-… https://…
#>  7  2016 https://data.cms.gov/data-api/v1/dataset/f0fb93a4-3475-42de-… https://…
#>  8  2015 https://data.cms.gov/data-api/v1/dataset/5cae9291-4be8-4e3c-… https://…
#>  9  2014 https://data.cms.gov/data-api/v1/dataset/b685142b-cc56-4ff1-… https://…
#> 10  2013 https://data.cms.gov/data-api/v1/dataset/c5a33eaf-a076-4513-… https://…
#> 
#> 
#> $by_provider_and_service
#> $by_provider_and_service$description
#> [1] "The Medicare Inpatient Hospitals by Provider and Service dataset provides information on inpatient discharges for Original Medicare Part A beneficiaries by IPPS hospitals. It includes information on the use, payment, and hospital charges for more than 3,000 U.S. hospitals that received IPPS payments. The data are organized by hospital and Medicare Severity Diagnosis Related Group (DRG).\n\n \n\nHospitals determine what they will charge for items and services provided to patients, and these charges are the amount the hospital bills for an item or service. The Total Payment Amount includes the DRG amount, claim per diem amount, beneficiary primary payer claim payment amount, beneficiary Part A (Hospital Insurance) coinsurance amount, beneficiary deductible amount, beneficiary blood deductible amount and diagnosis related group outlier amount."
#> 
#> $by_provider_and_service$dictionary
#> [1] "https://data.cms.gov/resources/medicare-inpatient-hospitals-by-provider-and-service-data-dictionary-0"
#> 
#> $by_provider_and_service$site
#> [1] "https://data.cms.gov/provider-summary-by-type-of-service/medicare-inpatient-hospitals/medicare-inpatient-hospitals-by-provider-and-service"
#> 
#> $by_provider_and_service$rows
#> [1] 145742
#> 
#> $by_provider_and_service$pages
#> [1] 30
#> 
#> $by_provider_and_service$fields
#>  [1] "Rndrng_Prvdr_CCN"          "Rndrng_Prvdr_Org_Name"    
#>  [3] "Rndrng_Prvdr_City"         "Rndrng_Prvdr_St"          
#>  [5] "Rndrng_Prvdr_State_FIPS"   "Rndrng_Prvdr_Zip5"        
#>  [7] "Rndrng_Prvdr_State_Abrvtn" "Rndrng_Prvdr_RUCA"        
#>  [9] "Rndrng_Prvdr_RUCA_Desc"    "DRG_Cd"                   
#> [11] "DRG_Desc"                  "Tot_Dschrgs"              
#> [13] "Avg_Submtd_Cvrd_Chrg"      "Avg_Tot_Pymt_Amt"         
#> [15] "Avg_Mdcr_Pymt_Amt"        
#> 
#> $by_provider_and_service$endpoints
#> # A tibble: 10 × 3
#>     year identifier                                                    resources
#>    <int> <chr>                                                         <chr>    
#>  1  2022 https://data.cms.gov/data-api/v1/dataset/46bf50f8-0983-4ca2-… https://…
#>  2  2021 https://data.cms.gov/data-api/v1/dataset/117d93f2-ce81-40fe-… https://…
#>  3  2020 https://data.cms.gov/data-api/v1/dataset/ef5bdbe1-27b4-4296-… https://…
#>  4  2019 https://data.cms.gov/data-api/v1/dataset/6f6d93e1-ecf8-4b93-… https://…
#>  5  2018 https://data.cms.gov/data-api/v1/dataset/09c12f06-e3fe-4cb0-… https://…
#>  6  2017 https://data.cms.gov/data-api/v1/dataset/b61ba5eb-021b-4510-… https://…
#>  7  2016 https://data.cms.gov/data-api/v1/dataset/ca9e33a4-e46c-4de9-… https://…
#>  8  2015 https://data.cms.gov/data-api/v1/dataset/e70315f5-4b02-46a8-… https://…
#>  9  2014 https://data.cms.gov/data-api/v1/dataset/635d7ccd-3dd7-4f1d-… https://…
#> 10  2013 https://data.cms.gov/data-api/v1/dataset/cf60c282-a006-444c-… https://…
#> 
#>