Medicaid API Endpoint Classes

Medicaid API Endpoint Classes

Usage

caid_endpoint(alias)

caid_group(alias)

caid_temporal(alias)

Arguments

alias

<chr> endpoint alias

Value

An S7 <caid_endpoint>, <caid_group>, or <caid_temporal> object

Examples

caid_endpoint("managed_longterm")
#> <caid_endpoint>
#>  @ identifier: chr "https://data.medicaid.gov/api/1/datastore/query/5394bcab-c748-5e4b-af07-b5bf77ed3aa3/0"
#>  @ metadata  :List of 6
#>  .. $ title      : chr "Managed Long Term Services and Supports (MLTSS) Enrollees"
#>  .. $ description: chr "1. Enrollment includes both Medicaid-only and Medicare-Medicaid (dual) enrollees. For both types of enrollees, "| __truncated__
#>  .. $ modified   : Date[1:1], format: "2024-10-16"
#>  .. $ periodicity: chr "Decennially [R/P10Y]"
#>  .. $ download   : chr "https://data.medicaid.gov/sites/default/files/uploaded_resources/managed-long-term-services-and-supports-mltss-"| __truncated__
#>  .. $ dictionary : chr "https://data.medicaid.gov/api/1/metastore/schemas/data-dictionary/items/7c6b1168-3353-4027-8b9c-03d622d7e999"
#>  @ dimensions: <class_dimensions>
#>  .. @ limit : int 8000
#>  .. @ rows  : int 399
#>  .. @ pages : int 1
#>  .. @ fields:List of 8
#>  .. .. $ state                                    : chr(0) 
#>  .. .. $ notes                                    : chr(0) 
#>  .. .. $ total_any_managed_care_enrollees         : chr(0) 
#>  .. .. $ comprehensive_managed_care_ltss_enrollees: chr(0) 
#>  .. .. $ comprehensive_managed_care_ltss_percent  : chr(0) 
#>  .. .. $ managed_ltss_only_enrollees              : chr(0) 
#>  .. .. $ managed_ltss_only_percent                : chr(0) 
#>  .. .. $ year                                     : chr(0) 
caid_group("caid_dual_status")
#> <caid_group>
#>  @ group  : chr "Dual Status Information for Medicaid and CHIP Beneficiaries"
#>  @ members:List of 2
#>  .. $ dual_status_month: <caid_endpoint>
#>  ..  ..@ identifier: chr "https://data.medicaid.gov/api/1/datastore/query/36ed6909-ab49-4ca5-a38e-6790138cd613/0"
#>  ..  ..@ metadata  :List of 6
#>  .. .. .. $ title      : chr "Dual Status Information for Medicaid and CHIP Beneficiaries by Month"
#>  .. .. .. $ description: chr "This data set includes monthly enrollment counts of Medicaid and CHIP beneficiaries by dual eligibility status "| __truncated__
#>  .. .. .. $ modified   : Date[1:1], format: "2024-01-05"
#>  .. .. .. $ periodicity: chr "Decennially [R/P10Y]"
#>  .. .. .. $ download   : chr "https://download.medicaid.gov/data/DualStatus-montly.csv"
#>  .. .. .. $ dictionary : chr "https://data.medicaid.gov/api/1/metastore/schemas/data-dictionary/items/de58d5ec-4515-494e-ae3f-620200f72989"
#>  ..  ..@ dimensions: <class_dimensions>
#>  .. .. .. @ limit : int 8000
#>  .. .. .. @ rows  : int 13356
#>  .. .. .. @ pages : int 2
#>  .. .. .. @ fields:List of 5
#>  .. .. .. .. $ state        : chr(0) 
#>  .. .. .. .. $ month        : chr(0) 
#>  .. .. .. .. $ dualstatus   : chr(0) 
#>  .. .. .. .. $ countenrolled: chr(0) 
#>  .. .. .. .. $ dqunusable   : chr(0) 
#>  .. $ dual_status_year : <caid_endpoint>
#>  ..  ..@ identifier: chr "https://data.medicaid.gov/api/1/datastore/query/93b36a8e-4dd5-4ff4-9a8b-8c6537684705/0"
#>  ..  ..@ metadata  :List of 6
#>  .. .. .. $ title      : chr "Dual Status Information for Medicaid and CHIP Beneficiaries by Year"
#>  .. .. .. $ description: chr "This data set presents annual enrollment counts of Medicaid and CHIP beneficiaries by dual eligibility status f"| __truncated__
#>  .. .. .. $ modified   : Date[1:1], format: "2024-01-05"
#>  .. .. .. $ periodicity: chr "Decennially [R/P10Y]"
#>  .. .. .. $ download   : chr "https://download.medicaid.gov/data/DualStatus-anul.csv"
#>  .. .. .. $ dictionary : chr "https://data.medicaid.gov/api/1/metastore/schemas/data-dictionary/items/de58d5ec-4515-494e-ae3f-620200f72989"
#>  ..  ..@ dimensions: <class_dimensions>
#>  .. .. .. @ limit : int 8000
#>  .. .. .. @ rows  : int 1113
#>  .. .. .. @ pages : int 1
#>  .. .. .. @ fields:List of 7
#>  .. .. .. .. $ state                    : chr(0) 
#>  .. .. .. .. $ year                     : chr(0) 
#>  .. .. .. .. $ dualstatus               : chr(0) 
#>  .. .. .. .. $ counteverenrolled        : chr(0) 
#>  .. .. .. .. $ countlastmonthenrollment : chr(0) 
#>  .. .. .. .. $ averageenrollmentpermonth: chr(0) 
#>  .. .. .. .. $ dqunusable               : chr(0) 
caid_temporal("state_drug_util")
#> <caid_temporal>
#>  @ metadata  :List of 3
#>  .. $ title      : chr "State Drug Utilization Data"
#>  .. $ description: chr "Drug utilization data are reported by states for covered outpatient drugs that are paid for by state Medicaid a"| __truncated__
#>  .. $ periodicity: chr "Decennially [R/P10Y]"
#>  @ dimensions: <class_dimensions>
#>  .. @ limit : int 8000
#>  .. @ rows  : int 5205065
#>  .. @ pages : int 651
#>  .. @ fields:List of 15
#>  .. .. $ utilization_type              : chr(0) 
#>  .. .. $ state                         : chr(0) 
#>  .. .. $ ndc                           : chr(0) 
#>  .. .. $ labeler_code                  : chr(0) 
#>  .. .. $ product_code                  : chr(0) 
#>  .. .. $ package_size                  : chr(0) 
#>  .. .. $ year                          : chr(0) 
#>  .. .. $ quarter                       : chr(0) 
#>  .. .. $ suppression_used              : chr(0) 
#>  .. .. $ product_name                  : chr(0) 
#>  .. .. $ units_reimbursed              : chr(0) 
#>  .. .. $ number_of_prescriptions       : chr(0) 
#>  .. .. $ total_amount_reimbursed       : chr(0) 
#>  .. .. $ medicaid_amount_reimbursed    : chr(0) 
#>  .. .. $ non_medicaid_amount_reimbursed: chr(0) 
#>  @ endpoints : tibble [34 × 5] (S3: tbl_df/tbl/data.frame)
#>  $ year      : int [1:34] 2024 2023 2022 2021 2020 2019 2018 2017 2016 2015 ...
#>  $ modified  : Date[1:34], format: "2025-05-14" "2025-05-14" ...
#>  $ identifier: chr [1:34] "https://data.medicaid.gov/api/1/datastore/query/61729e5a-7aa8-448c-8903-ba3e0cd0ea3c/0" "https://data.medicaid.gov/api/1/datastore/query/d890d3a9-6b00-43fd-8b31-fcba4c8e2909/0" "https://data.medicaid.gov/api/1/datastore/query/200c2cba-e58d-4a95-aa60-14b99736808d/0" "https://data.medicaid.gov/api/1/datastore/query/eec7fbe6-c4c4-5915-b3d0-be5828ef4e9d/0" ...
#>  $ download  : chr [1:34] "https://download.medicaid.gov/data/SDUD-2024.csv" "https://download.medicaid.gov/data/SDUD-2023.csv" "https://download.medicaid.gov/data/SDUD-2022.csv" "https://download.medicaid.gov/data/SDUD-2021.csv" ...
#>  $ dictionary: chr [1:34] "https://data.medicaid.gov/api/1/metastore/schemas/data-dictionary/items/66f843c8-1180-4244-8adc-554d02584fa4" "https://data.medicaid.gov/api/1/metastore/schemas/data-dictionary/items/de58d5ec-4515-494e-ae3f-620200f72989" "https://data.medicaid.gov/api/1/metastore/schemas/data-dictionary/items/de58d5ec-4515-494e-ae3f-620200f72989" "https://data.medicaid.gov/api/1/metastore/schemas/data-dictionary/items/75ffd9ed-1e27-44c0-ad49-ea39648dff02" ...