Medicare API Endpoint Classes

Medicare API Endpoint Classes

Usage

care_endpoint(alias)

care_temporal(alias)

care_group(alias, call = caller_env(), ...)

Arguments

alias

<chr> endpoint alias

call

<env> environment to use for error reporting

...

Additional arguments passed to the group constructor

Value

An S7 <care_endpoint>, or <care_temporal> object

Examples

care_endpoint("care_dialysis")
#> <care_endpoint>
#>  @ identifier: chr "https://data.cms.gov/data-api/v1/dataset/f8610e87-ba25-43a3-a49e-927dbc8701ae/data-viewer"
#>  @ metadata  :List of 11
#>  .. $ api        : chr "Medicare"
#>  .. $ title      : chr "Medicare Dialysis Facilities"
#>  .. $ description: chr "The Medicare Dialysis Facilities data provides information on clinical and patient measures for Medicare-certif"| __truncated__
#>  .. $ modified   : Date[1:1], format: "2024-02-29"
#>  .. $ temporal   : chr "2009-01-01 • 2024-12-31"
#>  .. $ periodicity: chr "Annually [R/P1Y]"
#>  .. $ download   : chr "https://data.cms.gov/sites/default/files/2024-02/b2c2658d-e19d-48db-8b3b-9d343a4d395f/dfr_facility_socrata_fy2024.csv"
#>  .. $ resources  : chr "https://data.cms.gov/data-api/v1/dataset-resources/f8610e87-ba25-43a3-a49e-927dbc8701ae"
#>  .. $ dictionary : chr "https://data.cms.gov/sites/default/files/2024-02/FY2024%20Data%20Dictionary.xlsx"
#>  .. $ site       : chr "https://data.cms.gov/quality-of-care/medicare-dialysis-facilities"
#>  .. $ references : chr "https://data.cms.gov/sites/default/files/2024-02/FY2024_DFR_Guide%20and%20Methodology.pdf"
#>  @ dimensions: <class_dimensions>
#>  .. @ limit : int 5000
#>  .. @ rows  : int 11679243
#>  .. @ pages : int 2336
#>  .. @ fields:List of 14
#>  .. .. $ state         : chr(0) 
#>  .. .. $ CCN           : chr(0) 
#>  .. .. $ Provider_Name : chr(0) 
#>  .. .. $ city          : chr(0) 
#>  .. .. $ Ownership_Type: chr(0) 
#>  .. .. $ ESRD_Network  : chr(0) 
#>  .. .. $ NPI           : chr(0) 
#>  .. .. $ Chain         : chr(0) 
#>  .. .. $ Modality      : chr(0) 
#>  .. .. $ Alternate_CCNs: chr(0) 
#>  .. .. $ Measure       : chr(0) 
#>  .. .. $ Measure_Score : chr(0) 
#>  .. .. $ year          : chr(0) 
#>  .. .. $ Measure_ID    : chr(0) 
care_temporal("quality_payment")
#> <care_temporal>
#>  @ metadata  :List of 7
#>  .. $ api        : chr "Medicare [Temporal]"
#>  .. $ title      : chr "Quality Payment Program Experience"
#>  .. $ description: chr "The Quality Payment Program (QPP) Experience dataset provides participation and performance information in the "| __truncated__
#>  .. $ periodicity: chr "Annually [R/P1Y]"
#>  .. $ dictionary : chr "https://data.cms.gov/sites/default/files/2024-04/2022%20QPP%20Experience%20Report%20Public%20Use%20Files%20Data"| __truncated__
#>  .. $ site       : chr "https://data.cms.gov/quality-of-care/quality-payment-program-experience"
#>  .. $ references : chr "https://data.cms.gov/sites/default/files/2024-04/2022%20QPP%20Experience%20Report%20Public%20Use%20Files%20Methodology.pdf"
#>  @ dimensions: <class_dimensions>
#>  .. @ limit : int 5000
#>  .. @ rows  : int 624200
#>  .. @ pages : int 125
#>  .. @ fields:List of 165
#>  .. .. $ provider key                                     : chr(0) 
#>  .. .. $ practice state or us territory                   : chr(0) 
#>  .. .. $ practice size                                    : chr(0) 
#>  .. .. $ clinician type                                   : chr(0) 
#>  .. .. $ clinician specialty                              : chr(0) 
#>  .. .. $ years in medicare                                : chr(0) 
#>  .. .. $ npi                                              : chr(0) 
#>  .. .. $ non-reporting                                    : chr(0) 
#>  .. .. $ participation option                             : chr(0) 
#>  .. .. $ medicare patients                                : chr(0) 
#>  .. .. $ allowed charges                                  : chr(0) 
#>  .. .. $ services                                         : chr(0) 
#>  .. .. $ opted into mips                                  : chr(0) 
#>  .. .. $ small practice status                            : chr(0) 
#>  .. .. $ rural status                                     : chr(0) 
#>  .. .. $ health professional shortage area status         : chr(0) 
#>  .. .. $ ambulatory surgical center-based status          : chr(0) 
#>  .. .. $ hospital-based status                            : chr(0) 
#>  .. .. $ non-patient facing status                        : chr(0) 
#>  .. .. $ facility-based status                            : chr(0) 
#>  .. .. $ dual eligibility ratio                           : chr(0) 
#>  .. .. $ safety-net status                                : chr(0) 
#>  .. .. $ extreme uncontrollable circumstance (euc)        : chr(0) 
#>  .. .. $ final score                                      : chr(0) 
#>  .. .. $ payment adjustment percentage                    : chr(0) 
#>  .. .. $ complex patient bonus                            : chr(0) 
#>  .. .. $ quality reweighting (euc)                        : chr(0) 
#>  .. .. $ quality category score                           : chr(0) 
#>  .. .. $ quality improvement score                        : chr(0) 
#>  .. .. $ small practice bonus                             : chr(0) 
#>  .. .. $ quality measure id 1                             : chr(0) 
#>  .. .. $ quality measure collection type 1                : chr(0) 
#>  .. .. $ quality measure score 1                          : chr(0) 
#>  .. .. $ quality measure id 2                             : chr(0) 
#>  .. .. $ quality measure collection type 2                : chr(0) 
#>  .. .. $ quality measure score 2                          : chr(0) 
#>  .. .. $ quality measure id 3                             : chr(0) 
#>  .. .. $ quality measure collection type 3                : chr(0) 
#>  .. .. $ quality measure score 3                          : chr(0) 
#>  .. .. $ quality measure id 4                             : chr(0) 
#>  .. .. $ quality measure collection type 4                : chr(0) 
#>  .. .. $ quality measure score 4                          : chr(0) 
#>  .. .. $ quality measure id 5                             : chr(0) 
#>  .. .. $ quality measure collection type 5                : chr(0) 
#>  .. .. $ quality measure score 5                          : chr(0) 
#>  .. .. $ quality measure id 6                             : chr(0) 
#>  .. .. $ quality measure collection type 6                : chr(0) 
#>  .. .. $ quality measure score 6                          : chr(0) 
#>  .. .. $ quality measure id 7                             : chr(0) 
#>  .. .. $ quality measure collection type 7                : chr(0) 
#>  .. .. $ quality measure score 7                          : chr(0) 
#>  .. .. $ quality measure id 8                             : chr(0) 
#>  .. .. $ quality measure collection type 8                : chr(0) 
#>  .. .. $ quality measure score 8                          : chr(0) 
#>  .. .. $ quality measure id 9                             : chr(0) 
#>  .. .. $ quality measure collection type 9                : chr(0) 
#>  .. .. $ quality measure score 9                          : chr(0) 
#>  .. .. $ quality measure id 10                            : chr(0) 
#>  .. .. $ quality measure collection type 10               : chr(0) 
#>  .. .. $ quality measure score 10                         : chr(0) 
#>  .. .. $ quality measure id 11                            : chr(0) 
#>  .. .. $ quality measure collection type 11               : chr(0) 
#>  .. .. $ quality measure score 11                         : chr(0) 
#>  .. .. $ quality measure id 12                            : chr(0) 
#>  .. .. $ quality measure collection type 12               : chr(0) 
#>  .. .. $ quality measure score 12                         : chr(0) 
#>  .. .. $ promoting interoperability (pi) category score   : chr(0) 
#>  .. .. $ pi reweighting (euc)                             : chr(0) 
#>  .. .. $ pi reweighting (hardship exception)              : chr(0) 
#>  .. .. $ pi reweighting (special status or clinician type): chr(0) 
#>  .. .. $ cehrt id                                         : chr(0) 
#>  .. .. $ pi measure id 1                                  : chr(0) 
#>  .. .. $ pi measure type 1                                : chr(0) 
#>  .. .. $ pi measure score 1                               : chr(0) 
#>  .. .. $ pi measure id 2                                  : chr(0) 
#>  .. .. $ pi measure type 2                                : chr(0) 
#>  .. .. $ pi measure score 2                               : chr(0) 
#>  .. .. $ pi measure id 3                                  : chr(0) 
#>  .. .. $ pi measure type 3                                : chr(0) 
#>  .. .. $ pi measure score 3                               : chr(0) 
#>  .. .. $ pi measure id 4                                  : chr(0) 
#>  .. .. $ pi measure type 4                                : chr(0) 
#>  .. .. $ pi measure score 4                               : chr(0) 
#>  .. .. $ pi measure id 5                                  : chr(0) 
#>  .. .. $ pi measure type 5                                : chr(0) 
#>  .. .. $ pi measure score 5                               : chr(0) 
#>  .. .. $ pi measure id 6                                  : chr(0) 
#>  .. .. $ pi measure type 6                                : chr(0) 
#>  .. .. $ pi measure score 6                               : chr(0) 
#>  .. .. $ pi measure id 7                                  : chr(0) 
#>  .. .. $ pi measure type 7                                : chr(0) 
#>  .. .. $ pi measure score 7                               : chr(0) 
#>  .. .. $ pi measure id 8                                  : chr(0) 
#>  .. .. $ pi measure type 8                                : chr(0) 
#>  .. .. $ pi measure score 8                               : chr(0) 
#>  .. .. $ pi measure id 9                                  : chr(0) 
#>  .. .. $ pi measure type 9                                : chr(0) 
#>  .. .. $ pi measure score 9                               : chr(0) 
#>  .. .. $ pi measure id 10                                 : chr(0) 
#>  .. ..  [list output truncated]
#>  @ endpoints : tibble [6 × 6] (S3: tbl_df/tbl/data.frame)
#>  $ year      : int [1:6] 2022 2021 2020 2019 2018 2017
#>  $ modified  : Date[1:6], format: "2024-05-08" "2023-06-12" ...
#>  $ temporal  : chr [1:6] "2022-01-01 • 2022-12-31" "2021-01-01 • 2021-12-31" "2020-01-01 • 2020-12-31" "2019-01-01 • 2019-12-31" ...
#>  $ identifier: chr [1:6] "https://data.cms.gov/data-api/v1/dataset/b3438273-b4a6-44ca-8fb2-9e6026b74642/data" "https://data.cms.gov/data-api/v1/dataset/8c2ab3e4-597e-43c9-9d94-7223c66d7f25/data" "https://data.cms.gov/data-api/v1/dataset/48a484e6-ff02-4284-8666-6045b3cac17f/data" "https://data.cms.gov/data-api/v1/dataset/5776e77d-a844-49fd-b697-a2b01bdacdc7/data" ...
#>  $ download  : chr [1:6] "https://data.cms.gov/sites/default/files/2023-05/0bac88bc-a4c8-4dff-bdbf-660d24e6a689/2021_puf_QPPAR-18260.csv" "https://data.cms.gov/sites/default/files/2023-04/fb3758dd-6ace-4ddb-863b-1b895cef4d7e/2020_puf_qppar_QPPAR-17660.csv" "https://data.cms.gov/sites/default/files/2023-04/f9ef9f64-91d8-49a2-a1be-5a48bbdcc1e9/2019_puf_qppar_QPPAR-17661.csv" "https://data.cms.gov/sites/default/files/2023-04/d2a19b56-7a5f-49c7-b78d-4d0dd49b9391/2018_puf_qppar-QPPAR-17662.csv" ...
#>  $ resources : chr [1:6] "https://data.cms.gov/data-api/v1/dataset-resources/b3438273-b4a6-44ca-8fb2-9e6026b74642" "https://data.cms.gov/data-api/v1/dataset-resources/8c2ab3e4-597e-43c9-9d94-7223c66d7f25" "https://data.cms.gov/data-api/v1/dataset-resources/48a484e6-ff02-4284-8666-6045b3cac17f" "https://data.cms.gov/data-api/v1/dataset-resources/5776e77d-a844-49fd-b697-a2b01bdacdc7" ...
care_group("care_hospital")
#> <class_group>
#>  @ group  : chr "Hospitals"
#>  @ members:List of 6
#>  .. $ hospital_owners      : <care_endpoint>
#>  ..  ..@ identifier: chr "https://data.cms.gov/data-api/v1/dataset/029c119f-f79c-49be-9100-344d31d10344/data-viewer"
#>  ..  ..@ metadata  :List of 11
#>  .. .. .. $ api        : chr "Medicare"
#>  .. .. .. $ title      : chr "Hospital All Owners"
#>  .. .. .. $ description: chr "The Hospital All Owners Information dataset provides information on all owners of the hospitals. This data incl"| __truncated__
#>  .. .. .. $ modified   : Date[1:1], format: "2025-05-15"
#>  .. .. .. $ temporal   : chr "2022-11-01 • 2025-05-31"
#>  .. .. .. $ periodicity: chr "Monthly [R/P1M]"
#>  .. .. .. $ download   : chr "https://data.cms.gov/sites/default/files/2025-05/f65adce5-fa22-4c0d-b148-dec15500590f/Hospital_All_Owners_2025.05.01.csv"
#>  .. .. .. $ resources  : chr "https://data.cms.gov/data-api/v1/dataset-resources/029c119f-f79c-49be-9100-344d31d10344"
#>  .. .. .. $ dictionary : chr "https://data.cms.gov/sites/default/files/2025-05/90b341cf-89fd-41c6-bf8a-a74e0fc4d1a7/Hospital_All_Owners_Data_Dictionary.pdf"
#>  .. .. .. $ site       : chr "https://data.cms.gov/provider-characteristics/hospitals-and-other-facilities/hospital-all-owners"
#>  .. .. .. $ references : chr "https://data.cms.gov/sites/default/files/2024-10/Hospital_Data_Guidance.pdf"
#>  ..  ..@ dimensions: <class_dimensions>
#>  .. .. .. @ limit : int 5000
#>  .. .. .. @ rows  : int 148341
#>  .. .. .. @ pages : int 30
#>  .. .. .. @ fields:List of 38
#>  .. .. .. .. $ ENROLLMENT ID                      : chr(0) 
#>  .. .. .. .. $ ASSOCIATE ID                       : chr(0) 
#>  .. .. .. .. $ ORGANIZATION NAME                  : chr(0) 
#>  .. .. .. .. $ ASSOCIATE ID - OWNER               : chr(0) 
#>  .. .. .. .. $ TYPE - OWNER                       : chr(0) 
#>  .. .. .. .. $ ROLE CODE - OWNER                  : chr(0) 
#>  .. .. .. .. $ ROLE TEXT - OWNER                  : chr(0) 
#>  .. .. .. .. $ ASSOCIATION DATE - OWNER           : chr(0) 
#>  .. .. .. .. $ FIRST NAME - OWNER                 : chr(0) 
#>  .. .. .. .. $ MIDDLE NAME - OWNER                : chr(0) 
#>  .. .. .. .. $ LAST NAME - OWNER                  : chr(0) 
#>  .. .. .. .. $ TITLE - OWNER                      : chr(0) 
#>  .. .. .. .. $ ORGANIZATION NAME - OWNER          : chr(0) 
#>  .. .. .. .. $ DOING BUSINESS AS NAME - OWNER     : chr(0) 
#>  .. .. .. .. $ ADDRESS LINE 1 - OWNER             : chr(0) 
#>  .. .. .. .. $ ADDRESS LINE 2 - OWNER             : chr(0) 
#>  .. .. .. .. $ CITY - OWNER                       : chr(0) 
#>  .. .. .. .. $ STATE - OWNER                      : chr(0) 
#>  .. .. .. .. $ ZIP CODE - OWNER                   : chr(0) 
#>  .. .. .. .. $ PERCENTAGE OWNERSHIP               : chr(0) 
#>  .. .. .. .. $ CREATED FOR ACQUISITION - OWNER    : chr(0) 
#>  .. .. .. .. $ CORPORATION - OWNER                : chr(0) 
#>  .. .. .. .. $ LLC - OWNER                        : chr(0) 
#>  .. .. .. .. $ MEDICAL PROVIDER SUPPLIER - OWNER  : chr(0) 
#>  .. .. .. .. $ MANAGEMENT SERVICES COMPANY - OWNER: chr(0) 
#>  .. .. .. .. $ MEDICAL STAFFING COMPANY - OWNER   : chr(0) 
#>  .. .. .. .. $ HOLDING COMPANY - OWNER            : chr(0) 
#>  .. .. .. .. $ INVESTMENT FIRM - OWNER            : chr(0) 
#>  .. .. .. .. $ FINANCIAL INSTITUTION - OWNER      : chr(0) 
#>  .. .. .. .. $ CONSULTING FIRM - OWNER            : chr(0) 
#>  .. .. .. .. $ FOR PROFIT - OWNER                 : chr(0) 
#>  .. .. .. .. $ NON PROFIT - OWNER                 : chr(0) 
#>  .. .. .. .. $ PRIVATE EQUITY COMPANY - OWNER     : chr(0) 
#>  .. .. .. .. $ REIT - OWNER                       : chr(0) 
#>  .. .. .. .. $ CHAIN HOME OFFICE - OWNER          : chr(0) 
#>  .. .. .. .. $ OTHER TYPE - OWNER                 : chr(0) 
#>  .. .. .. .. $ OTHER TYPE TEXT - OWNER            : chr(0) 
#>  .. .. .. .. $ OWNED BY ANOTHER ORG OR IND - OWNER: chr(0) 
#>  .. $ hospital_chow        : <care_endpoint>
#>  ..  ..@ identifier: chr "https://data.cms.gov/data-api/v1/dataset/c04031db-54ce-461c-85d1-d2613d71f167/data-viewer"
#>  ..  ..@ metadata  :List of 11
#>  .. .. .. $ api        : chr "Medicare"
#>  .. .. .. $ title      : chr "Hospital Change of Ownership"
#>  .. .. .. $ description: chr "The Hospital Change of Ownership (CHOW) dataset provides information on the hospital ownership changes that occ"| __truncated__
#>  .. .. .. $ modified   : Date[1:1], format: "2025-05-15"
#>  .. .. .. $ temporal   : chr "2022-01-01 • 2025-03-31"
#>  .. .. .. $ periodicity: chr "Quarterly [R/P3M]"
#>  .. .. .. $ download   : chr "https://data.cms.gov/sites/default/files/2025-05/211e6b65-6a6d-4e08-8ef2-1735ab22ffd6/Hospital_CHOW_2025.04.01.csv"
#>  .. .. .. $ resources  : chr "https://data.cms.gov/data-api/v1/dataset-resources/c04031db-54ce-461c-85d1-d2613d71f167"
#>  .. .. .. $ dictionary : chr "https://data.cms.gov/sites/default/files/2023-01/Hospital_CHOW_Data_Dictionary_2022.12.30.pdf"
#>  .. .. .. $ site       : chr "https://data.cms.gov/provider-characteristics/hospitals-and-other-facilities/hospital-change-of-ownership"
#>  .. .. .. $ references : chr "https://data.cms.gov/sites/default/files/2024-10/Hospital_CHOW_Data_Guidance.pdf"
#>  ..  ..@ dimensions: <class_dimensions>
#>  .. .. .. @ limit : int 5000
#>  .. .. .. @ rows  : int 715
#>  .. .. .. @ pages : int 1
#>  .. .. .. @ fields:List of 23
#>  .. .. .. .. $ ENROLLMENT ID - BUYER          : chr(0) 
#>  .. .. .. .. $ ENROLLMENT STATE - BUYER       : chr(0) 
#>  .. .. .. .. $ PROVIDER TYPE CODE - BUYER     : chr(0) 
#>  .. .. .. .. $ PROVIDER TYPE TEXT - BUYER     : chr(0) 
#>  .. .. .. .. $ NPI - BUYER                    : chr(0) 
#>  .. .. .. .. $ MULTIPLE NPI FLAG - BUYER      : chr(0) 
#>  .. .. .. .. $ CCN - BUYER                    : chr(0) 
#>  .. .. .. .. $ ASSOCIATE ID - BUYER           : chr(0) 
#>  .. .. .. .. $ ORGANIZATION NAME - BUYER      : chr(0) 
#>  .. .. .. .. $ DOING BUSINESS AS NAME - BUYER : chr(0) 
#>  .. .. .. .. $ CHOW TYPE CODE                 : chr(0) 
#>  .. .. .. .. $ CHOW TYPE TEXT                 : chr(0) 
#>  .. .. .. .. $ EFFECTIVE DATE                 : chr(0) 
#>  .. .. .. .. $ ENROLLMENT ID - SELLER         : chr(0) 
#>  .. .. .. .. $ ENROLLMENT STATE - SELLER      : chr(0) 
#>  .. .. .. .. $ PROVIDER TYPE CODE - SELLER    : chr(0) 
#>  .. .. .. .. $ PROVIDER TYPE TEXT - SELLER    : chr(0) 
#>  .. .. .. .. $ NPI - SELLER                   : chr(0) 
#>  .. .. .. .. $ MULTIPLE NPI FLAG - SELLER     : chr(0) 
#>  .. .. .. .. $ CCN - SELLER                   : chr(0) 
#>  .. .. .. .. $ ASSOCIATE ID - SELLER          : chr(0) 
#>  .. .. .. .. $ ORGANIZATION NAME - SELLER     : chr(0) 
#>  .. .. .. .. $ DOING BUSINESS AS NAME - SELLER: chr(0) 
#>  .. $ hospital_chow_owner  : <care_endpoint>
#>  ..  ..@ identifier: chr "https://data.cms.gov/data-api/v1/dataset/60625dc8-b621-45f0-9423-077fd133b13e/data-viewer"
#>  ..  ..@ metadata  :List of 11
#>  .. .. .. $ api        : chr "Medicare"
#>  .. .. .. $ title      : chr "Hospital Change of Ownership - Owner Information"
#>  .. .. .. $ description: chr "The Hospital Change of Ownership (CHOW) - Owner Information dataset provides information on individual and orga"| __truncated__
#>  .. .. .. $ modified   : Date[1:1], format: "2025-05-15"
#>  .. .. .. $ temporal   : chr "2022-01-01 • 2025-03-31"
#>  .. .. .. $ periodicity: chr "Quarterly [R/P3M]"
#>  .. .. .. $ download   : chr "https://data.cms.gov/sites/default/files/2025-05/5af7a510-ed7c-48ab-a61e-7bcd0c88550e/Hospital_CHOW_Owners_2025.04.01.csv"
#>  .. .. .. $ resources  : chr "https://data.cms.gov/data-api/v1/dataset-resources/60625dc8-b621-45f0-9423-077fd133b13e"
#>  .. .. .. $ dictionary : chr "https://data.cms.gov/sites/default/files/2024-10/Hospital_CHOW_Owners_Data_Dictionary.pdf"
#>  .. .. .. $ site       : chr "https://data.cms.gov/provider-characteristics/hospitals-and-other-facilities/hospital-change-of-ownership-owner-information"
#>  .. .. .. $ references : chr "https://data.cms.gov/sites/default/files/2024-10/Hospital_CHOW_Data_Guidance.pdf"
#>  ..  ..@ dimensions: <class_dimensions>
#>  .. .. .. @ limit : int 5000
#>  .. .. .. @ rows  : int 21289
#>  .. .. .. @ pages : int 5
#>  .. .. .. @ fields:List of 34
#>  .. .. .. .. $ ENROLLMENT ID                      : chr(0) 
#>  .. .. .. .. $ ASSOCIATE ID                       : chr(0) 
#>  .. .. .. .. $ ORGANIZATION NAME                  : chr(0) 
#>  .. .. .. .. $ ASSOCIATE ID - OWNER               : chr(0) 
#>  .. .. .. .. $ TYPE - OWNER                       : chr(0) 
#>  .. .. .. .. $ ROLE CODE - OWNER                  : chr(0) 
#>  .. .. .. .. $ ROLE TEXT - OWNER                  : chr(0) 
#>  .. .. .. .. $ ASSOCIATION DATE - OWNER           : chr(0) 
#>  .. .. .. .. $ FIRST NAME - OWNER                 : chr(0) 
#>  .. .. .. .. $ MIDDLE NAME - OWNER                : chr(0) 
#>  .. .. .. .. $ LAST NAME - OWNER                  : chr(0) 
#>  .. .. .. .. $ TITLE - OWNER                      : chr(0) 
#>  .. .. .. .. $ ORGANIZATION NAME - OWNER          : chr(0) 
#>  .. .. .. .. $ DOING BUSINESS AS NAME - OWNER     : chr(0) 
#>  .. .. .. .. $ ADDRESS LINE 1 - OWNER             : chr(0) 
#>  .. .. .. .. $ ADDRESS LINE 2 - OWNER             : chr(0) 
#>  .. .. .. .. $ CITY - OWNER                       : chr(0) 
#>  .. .. .. .. $ STATE - OWNER                      : chr(0) 
#>  .. .. .. .. $ ZIP CODE - OWNER                   : chr(0) 
#>  .. .. .. .. $ PERCENTAGE OWNERSHIP               : chr(0) 
#>  .. .. .. .. $ CREATED FOR ACQUISITION - OWNER    : chr(0) 
#>  .. .. .. .. $ CORPORATION - OWNER                : chr(0) 
#>  .. .. .. .. $ LLC - OWNER                        : chr(0) 
#>  .. .. .. .. $ MEDICAL PROVIDER SUPPLIER - OWNER  : chr(0) 
#>  .. .. .. .. $ MANAGEMENT SERVICES COMPANY - OWNER: chr(0) 
#>  .. .. .. .. $ MEDICAL STAFFING COMPANY - OWNER   : chr(0) 
#>  .. .. .. .. $ HOLDING COMPANY - OWNER            : chr(0) 
#>  .. .. .. .. $ INVESTMENT FIRM - OWNER            : chr(0) 
#>  .. .. .. .. $ FINANCIAL INSTITUTION - OWNER      : chr(0) 
#>  .. .. .. .. $ CONSULTING FIRM - OWNER            : chr(0) 
#>  .. .. .. .. $ FOR PROFIT - OWNER                 : chr(0) 
#>  .. .. .. .. $ NON PROFIT - OWNER                 : chr(0) 
#>  .. .. .. .. $ OTHER TYPE - OWNER                 : chr(0) 
#>  .. .. .. .. $ OTHER TYPE TEXT - OWNER            : chr(0) 
#>  .. $ hospital_enrollments : <care_endpoint>
#>  ..  ..@ identifier: chr "https://data.cms.gov/data-api/v1/dataset/f6f6505c-e8b0-4d57-b258-e2b94133aaf2/data-viewer"
#>  ..  ..@ metadata  :List of 11
#>  .. .. .. $ api        : chr "Medicare"
#>  .. .. .. $ title      : chr "Hospital Enrollments"
#>  .. .. .. $ description: chr "The Hospital Enrollments dataset provides enrollment information of all Hospitals currently enrolled in Medicar"| __truncated__
#>  .. .. .. $ modified   : Date[1:1], format: "2025-05-15"
#>  .. .. .. $ temporal   : chr "2022-11-01 • 2025-05-31"
#>  .. .. .. $ periodicity: chr "Monthly [R/P1M]"
#>  .. .. .. $ download   : chr "https://data.cms.gov/sites/default/files/2025-05/37dd6e7b-69e4-46d8-b0b3-31e789b2a94f/Hospital_Enrollments_2025.05.01.csv"
#>  .. .. .. $ resources  : chr "https://data.cms.gov/data-api/v1/dataset-resources/f6f6505c-e8b0-4d57-b258-e2b94133aaf2"
#>  .. .. .. $ dictionary : chr "https://data.cms.gov/sites/default/files/2023-07/Hospital_Enrollments_Data_Dictionary.pdf"
#>  .. .. .. $ site       : chr "https://data.cms.gov/provider-characteristics/hospitals-and-other-facilities/hospital-enrollments"
#>  .. .. .. $ references : chr "https://data.cms.gov/sites/default/files/2024-10/Hospital_Data_Guidance.pdf"
#>  ..  ..@ dimensions: <class_dimensions>
#>  .. .. .. @ limit : int 5000
#>  .. .. .. @ rows  : int 9246
#>  .. .. .. @ pages : int 2
#>  .. .. .. @ fields:List of 39
#>  .. .. .. .. $ ENROLLMENT ID                 : chr(0) 
#>  .. .. .. .. $ ENROLLMENT STATE              : chr(0) 
#>  .. .. .. .. $ PROVIDER TYPE CODE            : chr(0) 
#>  .. .. .. .. $ PROVIDER TYPE TEXT            : chr(0) 
#>  .. .. .. .. $ NPI                           : chr(0) 
#>  .. .. .. .. $ MULTIPLE NPI FLAG             : chr(0) 
#>  .. .. .. .. $ CCN                           : chr(0) 
#>  .. .. .. .. $ ASSOCIATE ID                  : chr(0) 
#>  .. .. .. .. $ ORGANIZATION NAME             : chr(0) 
#>  .. .. .. .. $ DOING BUSINESS AS NAME        : chr(0) 
#>  .. .. .. .. $ INCORPORATION DATE            : chr(0) 
#>  .. .. .. .. $ INCORPORATION STATE           : chr(0) 
#>  .. .. .. .. $ ORGANIZATION TYPE STRUCTURE   : chr(0) 
#>  .. .. .. .. $ ORGANIZATION OTHER TYPE TEXT  : chr(0) 
#>  .. .. .. .. $ PROPRIETARY NONPROFIT         : chr(0) 
#>  .. .. .. .. $ ADDRESS LINE 1                : chr(0) 
#>  .. .. .. .. $ ADDRESS LINE 2                : chr(0) 
#>  .. .. .. .. $ CITY                          : chr(0) 
#>  .. .. .. .. $ STATE                         : chr(0) 
#>  .. .. .. .. $ ZIP CODE                      : chr(0) 
#>  .. .. .. .. $ PRACTICE LOCATION TYPE        : chr(0) 
#>  .. .. .. .. $ LOCATION OTHER TYPE TEXT      : chr(0) 
#>  .. .. .. .. $ SUBGROUP - GENERAL            : chr(0) 
#>  .. .. .. .. $ SUBGROUP - ACUTE CARE         : chr(0) 
#>  .. .. .. .. $ SUBGROUP - ALCOHOL DRUG       : chr(0) 
#>  .. .. .. .. $ SUBGROUP - CHILDRENS          : chr(0) 
#>  .. .. .. .. $ SUBGROUP - LONG-TERM          : chr(0) 
#>  .. .. .. .. $ SUBGROUP - PSYCHIATRIC        : chr(0) 
#>  .. .. .. .. $ SUBGROUP - REHABILITATION     : chr(0) 
#>  .. .. .. .. $ SUBGROUP - SHORT-TERM         : chr(0) 
#>  .. .. .. .. $ SUBGROUP - SWING-BED APPROVED : chr(0) 
#>  .. .. .. .. $ SUBGROUP - PSYCHIATRIC UNIT   : chr(0) 
#>  .. .. .. .. $ SUBGROUP - REHABILITATION UNIT: chr(0) 
#>  .. .. .. .. $ SUBGROUP - SPECIALTY HOSPITAL : chr(0) 
#>  .. .. .. .. $ SUBGROUP - OTHER              : chr(0) 
#>  .. .. .. .. $ SUBGROUP - OTHER TEXT         : chr(0) 
#>  .. .. .. .. $ REH CONVERSION FLAG           : chr(0) 
#>  .. .. .. .. $ REH CONVERSION DATE           : chr(0) 
#>  .. .. .. .. $ CAH OR HOSPITAL CCN           : chr(0) 
#>  .. $ hospital_costreport  : <care_endpoint>
#>  ..  ..@ identifier: chr "https://data.cms.gov/data-api/v1/dataset/44060663-47d8-4ced-a115-b53b4c270acb/data-viewer"
#>  ..  ..@ metadata  :List of 11
#>  .. .. .. $ api        : chr "Medicare"
#>  .. .. .. $ title      : chr "Hospital Provider Cost Report"
#>  .. .. .. $ description: chr "The Hospital Provider Cost Report dataset provides select measures from the hospital annual cost report. This d"| __truncated__
#>  .. .. .. $ modified   : Date[1:1], format: "2024-10-23"
#>  .. .. .. $ temporal   : chr "2011-01-01 • 2022-12-31"
#>  .. .. .. $ periodicity: chr "Annually [R/P1Y]"
#>  .. .. .. $ download   : chr "https://data.cms.gov/sites/default/files/2024-10/b220a101-219f-47d5-acfe-1685596bc727/CostReport_2022_Final.csv"
#>  .. .. .. $ resources  : chr "https://data.cms.gov/data-api/v1/dataset-resources/44060663-47d8-4ced-a115-b53b4c270acb"
#>  .. .. .. $ dictionary : chr "https://data.cms.gov/sites/default/files/2024-03/9756088d-5280-4090-80b9-449d31ef25a3/Cost%20Report%20Data%20Di"| __truncated__
#>  .. .. .. $ site       : chr "https://data.cms.gov/provider-compliance/cost-report/hospital-provider-cost-report"
#>  .. .. .. $ references : chr "https://data.cms.gov/resources/hospital-provider-cost-report-methodology"
#>  ..  ..@ dimensions: <class_dimensions>
#>  .. .. .. @ limit : int 5000
#>  .. .. .. @ rows  : int 6064
#>  .. .. .. @ pages : int 2
#>  .. .. .. @ fields:List of 117
#>  .. .. .. .. $ rpt_rec_num                                                            : chr(0) 
#>  .. .. .. .. $ Provider CCN                                                           : chr(0) 
#>  .. .. .. .. $ Hospital Name                                                          : chr(0) 
#>  .. .. .. .. $ Street Address                                                         : chr(0) 
#>  .. .. .. .. $ City                                                                   : chr(0) 
#>  .. .. .. .. $ State Code                                                             : chr(0) 
#>  .. .. .. .. $ Zip Code                                                               : chr(0) 
#>  .. .. .. .. $ County                                                                 : chr(0) 
#>  .. .. .. .. $ Medicare CBSA Number                                                   : chr(0) 
#>  .. .. .. .. $ Rural Versus Urban                                                     : chr(0) 
#>  .. .. .. .. $ CCN Facility Type                                                      : chr(0) 
#>  .. .. .. .. $ Provider Type                                                          : chr(0) 
#>  .. .. .. .. $ Type of Control                                                        : chr(0) 
#>  .. .. .. .. $ Fiscal Year Begin Date                                                 : chr(0) 
#>  .. .. .. .. $ Fiscal Year End Date                                                   : chr(0) 
#>  .. .. .. .. $ FTE - Employees on Payroll                                             : chr(0) 
#>  .. .. .. .. $ Number of Interns and Residents (FTE)                                  : chr(0) 
#>  .. .. .. .. $ Total Days Title V                                                     : chr(0) 
#>  .. .. .. .. $ Total Days Title XVIII                                                 : chr(0) 
#>  .. .. .. .. $ Total Days Title XIX                                                   : chr(0) 
#>  .. .. .. .. $ Total Days (V + XVIII + XIX + Unknown)                                 : chr(0) 
#>  .. .. .. .. $ Number of Beds                                                         : chr(0) 
#>  .. .. .. .. $ Total Bed Days Available                                               : chr(0) 
#>  .. .. .. .. $ Total Discharges Title V                                               : chr(0) 
#>  .. .. .. .. $ Total Discharges Title XVIII                                           : chr(0) 
#>  .. .. .. .. $ Total Discharges Title XIX                                             : chr(0) 
#>  .. .. .. .. $ Total Discharges (V + XVIII + XIX + Unknown)                           : chr(0) 
#>  .. .. .. .. $ Number of Beds + Total for all Subproviders                            : chr(0) 
#>  .. .. .. .. $ Hospital Total Days Title V For Adults & Peds                          : chr(0) 
#>  .. .. .. .. $ Hospital Total Days Title XVIII For Adults & Peds                      : chr(0) 
#>  .. .. .. .. $ Hospital Total Days Title XIX For Adults & Peds                        : chr(0) 
#>  .. .. .. .. $ Hospital Total Days (V + XVIII + XIX + Unknown) For Adults & Peds      : chr(0) 
#>  .. .. .. .. $ Hospital Number of Beds For Adults & Peds                              : chr(0) 
#>  .. .. .. .. $ Hospital Total Bed Days Available For Adults & Peds                    : chr(0) 
#>  .. .. .. .. $ Hospital Total Discharges Title V For Adults & Peds                    : chr(0) 
#>  .. .. .. .. $ Hospital Total Discharges Title XVIII For Adults & Peds                : chr(0) 
#>  .. .. .. .. $ Hospital Total Discharges Title XIX For Adults & Peds                  : chr(0) 
#>  .. .. .. .. $ Hospital Total Discharges (V + XVIII + XIX + Unknown) For Adults & Peds: chr(0) 
#>  .. .. .. .. $ Cost of Charity Care                                                   : chr(0) 
#>  .. .. .. .. $ Total Bad Debt Expense                                                 : chr(0) 
#>  .. .. .. .. $ Cost of Uncompensated Care                                             : chr(0) 
#>  .. .. .. .. $ Total Unreimbursed and Uncompensated Care                              : chr(0) 
#>  .. .. .. .. $ Total Salaries From Worksheet A                                        : chr(0) 
#>  .. .. .. .. $ Overhead Non-Salary Costs                                              : chr(0) 
#>  .. .. .. .. $ Depreciation Cost                                                      : chr(0) 
#>  .. .. .. .. $ Total Costs                                                            : chr(0) 
#>  .. .. .. .. $ Inpatient Total Charges                                                : chr(0) 
#>  .. .. .. .. $ Outpatient Total Charges                                               : chr(0) 
#>  .. .. .. .. $ Combined Outpatient + Inpatient Total Charges                          : chr(0) 
#>  .. .. .. .. $ Wage-Related Costs (Core)                                              : chr(0) 
#>  .. .. .. .. $ Wage-Related Costs (RHC/FQHC)                                          : chr(0) 
#>  .. .. .. .. $ Total Salaries (adjusted)                                              : chr(0) 
#>  .. .. .. .. $ Contract Labor: Direct Patient Care                                    : chr(0) 
#>  .. .. .. .. $ Wage Related Costs for Part - A Teaching Physicians                    : chr(0) 
#>  .. .. .. .. $ Wage Related Costs for Interns and Residents                           : chr(0) 
#>  .. .. .. .. $ Cash on Hand and in Banks                                              : chr(0) 
#>  .. .. .. .. $ Temporary Investments                                                  : chr(0) 
#>  .. .. .. .. $ Notes Receivable                                                       : chr(0) 
#>  .. .. .. .. $ Accounts Receivable                                                    : chr(0) 
#>  .. .. .. .. $ Less: Allowances for Uncollectible Notes and Accounts Receivable       : chr(0) 
#>  .. .. .. .. $ Inventory                                                              : chr(0) 
#>  .. .. .. .. $ Prepaid Expenses                                                       : chr(0) 
#>  .. .. .. .. $ Other Current Assets                                                   : chr(0) 
#>  .. .. .. .. $ Total Current Assets                                                   : chr(0) 
#>  .. .. .. .. $ Land                                                                   : chr(0) 
#>  .. .. .. .. $ Land Improvements                                                      : chr(0) 
#>  .. .. .. .. $ Buildings                                                              : chr(0) 
#>  .. .. .. .. $ Leasehold Improvements                                                 : chr(0) 
#>  .. .. .. .. $ Fixed Equipment                                                        : chr(0) 
#>  .. .. .. .. $ Major Movable Equipment                                                : chr(0) 
#>  .. .. .. .. $ Minor Equipment Depreciable                                            : chr(0) 
#>  .. .. .. .. $ Health Information Technology Designated Assets                        : chr(0) 
#>  .. .. .. .. $ Total Fixed Assets                                                     : chr(0) 
#>  .. .. .. .. $ Investments                                                            : chr(0) 
#>  .. .. .. .. $ Other Assets                                                           : chr(0) 
#>  .. .. .. .. $ Total Other Assets                                                     : chr(0) 
#>  .. .. .. .. $ Total Assets                                                           : chr(0) 
#>  .. .. .. .. $ Accounts Payable                                                       : chr(0) 
#>  .. .. .. .. $ Salaries, Wages, and Fees Payable                                      : chr(0) 
#>  .. .. .. .. $ Payroll Taxes Payable                                                  : chr(0) 
#>  .. .. .. .. $ Notes and Loans Payable (Short Term)                                   : chr(0) 
#>  .. .. .. .. $ Deferred Income                                                        : chr(0) 
#>  .. .. .. .. $ Other Current Liabilities                                              : chr(0) 
#>  .. .. .. .. $ Total Current Liabilities                                              : chr(0) 
#>  .. .. .. .. $ Mortgage Payable                                                       : chr(0) 
#>  .. .. .. .. $ Notes Payable                                                          : chr(0) 
#>  .. .. .. .. $ Unsecured Loans                                                        : chr(0) 
#>  .. .. .. .. $ Other Long Term Liabilities                                            : chr(0) 
#>  .. .. .. .. $ Total Long Term Liabilities                                            : chr(0) 
#>  .. .. .. .. $ Total Liabilities                                                      : chr(0) 
#>  .. .. .. .. $ General Fund Balance                                                   : chr(0) 
#>  .. .. .. .. $ Total Fund Balances                                                    : chr(0) 
#>  .. .. .. .. $ Total Liabilities and Fund Balances                                    : chr(0) 
#>  .. .. .. .. $ DRG Amounts Other Than Outlier Payments                                : chr(0) 
#>  .. .. .. .. $ DRG Amounts Before October 1                                           : chr(0) 
#>  .. .. .. .. $ DRG Amounts After October 1                                            : chr(0) 
#>  .. .. .. .. $ Outlier Payments For Discharges                                        : chr(0) 
#>  .. .. .. .. $ Disproportionate Share Adjustment                                      : chr(0) 
#>  .. .. .. .. $ Allowable DSH Percentage                                               : chr(0) 
#>  .. .. .. ..  [list output truncated]
#>  .. $ hospital_service_area: <care_endpoint>
#>  ..  ..@ identifier: chr "https://data.cms.gov/data-api/v1/dataset/8708ca8b-8636-44ed-8303-724cbfaf78ad/data-viewer"
#>  ..  ..@ metadata  :List of 11
#>  .. .. .. $ api        : chr "Medicare"
#>  .. .. .. $ title      : chr "Hospital Service Area"
#>  .. .. .. $ description: chr "The Hospital Service Area data is a summary of calendar year Medicare inpatient hospital fee-for-service and Me"| __truncated__
#>  .. .. .. $ modified   : Date[1:1], format: "2024-07-26"
#>  .. .. .. $ temporal   : chr "2015-01-01 • 2023-12-31"
#>  .. .. .. $ periodicity: chr "Annually [R/P1Y]"
#>  .. .. .. $ download   : chr "https://data.cms.gov/sites/default/files/2024-07/88791b91-c9e4-448f-b961-60313b43bc40/Hospital_Service_Area_2023.csv"
#>  .. .. .. $ resources  : chr "https://data.cms.gov/data-api/v1/dataset-resources/8708ca8b-8636-44ed-8303-724cbfaf78ad"
#>  .. .. .. $ dictionary : chr "https://data.cms.gov/resources/hospital-service-area-data-dictionary"
#>  .. .. .. $ site       : chr "https://data.cms.gov/provider-summary-by-type-of-service/medicare-inpatient-hospitals/hospital-service-area"
#>  .. .. .. $ references : chr "https://data.cms.gov/resources/hospital-service-area-methodology"
#>  ..  ..@ dimensions: <class_dimensions>
#>  .. .. .. @ limit : int 5000
#>  .. .. .. @ rows  : int 1139666
#>  .. .. .. @ pages : int 228
#>  .. .. .. @ fields:List of 5
#>  .. .. .. .. $ MEDICARE_PROV_NUM  : chr(0) 
#>  .. .. .. .. $ ZIP_CD_OF_RESIDENCE: chr(0) 
#>  .. .. .. .. $ TOTAL_DAYS_OF_CARE : chr(0) 
#>  .. .. .. .. $ TOTAL_CHARGES      : chr(0) 
#>  .. .. .. .. $ TOTAL_CASES        : chr(0)