Search the NLM NPI Registry

Search the NLM NPI Registry

Search the NPPES NPI Registry

Usage

npi_nlm(terms, npi = NULL)

npi_nppes(
  npi = NULL,
  entity = NULL,
  first = NULL,
  last = NULL,
  organization = NULL,
  name_type = NULL,
  taxonomy_desc = NULL,
  city = NULL,
  state = NULL,
  zip = NULL,
  country = NULL
)

Source

API Documentation: Individuals

API Documentation: Organizations

API Documentation

Arguments

terms

<chr> Search terms, separated by spaces

npi

<chr> Unique 10-digit National Provider Identifier number issued by CMS to US healthcare providers through NPPES.

entity

<chr> Entity type; one of either I for Individual (NPI-1) or O for Organizational (NPI-2)

first, last

<chr> Individual provider's name

organization

<chr> Organizational provider's name

name_type

<chr> Type of individual the first and last name parameters refer to; one of either AO for Authorized Officials or Provider for Individual Providers.

taxonomy_desc

<chr> Provider's taxonomy description, e.g. Pharmacist, Pediatrics

city

<chr> City name. For military addresses, search for either APO or FPO.

state

<chr> 2-character state abbreviation. If it is the only input, one other parameter besides entype and country is required.

zip

<chr> WC 5- to 9-digit zip code, without a hyphen.

country

<chr> 2-character country abbreviation. Can be the only input, as long as it is not US.

Value

<tibble> of search results

<tibble> of search results

Examples

npi_nlm("john bethesda")
#> ◼ NPPES [NLM]
#> ◼ 138 RWS [1 PG]
#> ◼ 138 RWS [1 PG]
#> # A tibble: 138 × 4
#>    full_name      npi        specialty                              full_address
#>    <chr>          <chr>      <chr>                                  <chr>       
#>  1 KELLY, JOHN    1760880173 Dentist                                4833 BETHES…
#>  2 VAVREK, JOHN   1801864624 Military Health Care Provider          NNMC BETHES…
#>  3 SCHACHT, JOHN  1760801047 Medical Genetics and Genomics          WR, 8901 RO…
#>  4 MAYE, JOHN     1225008147 Certified Registered Nurse Anesthetis… 8901 WISCON…
#>  5 WEEKS, JOHN    1750077236 Military Health Care Provider          8901 WISCON…
#>  6 CARAWAY, JOHN  1639923378 Student in an Organized Health Care E… 8901 ROCKVI…
#>  7 ATWOOD, JOHN   1023117090 Physician/Cardiovascular Disease (Car… 8901 ROCKVI…
#>  8 CHOI, JOHN     1205862869 Physician/Neurology                    8901 ROCKVI…
#>  9 DANG, JOHN     1497245765 Physician/Internal Medicine            8901 WISCON…
#> 10 LINABURY, JOHN 1942564885 Physician/Dermatology                  8901 WISCON…
#> # ℹ 128 more rows
npi_nlm("Wiregrass Georgia")
#> ◼ NPPES [NLM]
#> ◼ 1 RW [1 PG]
#> ◼ 1 RW [1 PG]
#> # A tibble: 1 × 4
#>   full_name        npi        specialty full_address                            
#>   <chr>            <chr>      <chr>     <chr>                                   
#> 1 STEWART, CHARLES 1881965564 Dentist   4089 VAL TECH RD WIREGRASS GEORGIA TECH…
npi_nlm("Dentist Valdosta")
#> ◼ NPPES [NLM]
#> ◼ 56 RWS [1 PG]
#> ◼ 56 RWS [1 PG]
#> # A tibble: 56 × 4
#>    full_name           npi        specialty full_address                        
#>    <chr>               <chr>      <chr>     <chr>                               
#>  1 DEPTA, BRENT        1649370339 Dentist   2711 WINDEMERE DR, VALDOSTA, GA 316…
#>  2 GRAY, HARDY         1497104871 Dentist   3362 GREYSTONE WAY, VALDOSTA, GA 31…
#>  3 MAXWELL, DOUGLAS    1205999133 Dentist   2415 ANDOVER DR, VALDOSTA, GA 31602 
#>  4 MARABLE, STANLEY    1295883957 Dentist   308 JANET ST, VALDOSTA, GA 31602    
#>  5 ROWELL, EUGENE, III 1790347185 Dentist   2711 WINDEMERE DR, VALDOSTA, GA 316…
#>  6 RASMUSSEN, DUHWAN   1366117186 Dentist   282 NORMAN DR, VALDOSTA, GA 31601   
#>  7 BROADFOOT, JAMES    1174704837 Dentist   414 PENDLETON PL, VALDOSTA, GA 31602
#>  8 GRANTHAM, TIMOTHY   1063580181 Dentist   103 W CRANFORD AVE, VALDOSTA, GA 31…
#>  9 LAIRSEY, JASON      1528128683 Dentist   103 W CRANFORD AVE, VALDOSTA, GA 31…
#> 10 ADAMS, BRADLEY      1922026111 Dentist   816 NORTHWOOD PARK DRIVE, VALDOSTA,…
#> # ℹ 46 more rows
npi_nlm("Valdosta")
#> ◼ NPPES [NLM]
#> ◼ 1711 RWS [4 PGS]
#> ◼ 1711 RWS [4 PGS]
#> # A tibble: 1,711 × 4
#>    full_name           npi        specialty                         full_address
#>    <chr>               <chr>      <chr>                             <chr>       
#>  1 CRISCUOLA, ROBERT   1487632246 Physician/Pediatric Medicine      3331 N VALD…
#>  2 ALVARADO, ALAN      1972581767 Physician/Pediatric Medicine      3440 N VALD…
#>  3 STIEFEL, ERIC       1023219672 Physician/General Surgery         3527 N VALD…
#>  4 SPALDING, JOSHUA    1134828551 Physical Therapist in Private Pr… 3527 N VALD…
#>  5 PEACOCK, KIMBERLY   1942347919 Certified Registered Nurse Anest… 4280 N VALD…
#>  6 KENDRICK, JOHN      1164426136 Physician/General Surgery         3527 N VALD…
#>  7 AGUERO, OSCAR, JR.  1962406918 Physician/General Surgery         3527 N VALD…
#>  8 HASKINS, ERIN       1568769479 Speech Language Pathologist       3256 N VALD…
#>  9 MCNEAL, SHELBY      1265020358 Nurse Practitioner                4274 N VALD…
#> 10 CHAVARRIA, KAYLEIGH 1205695301 Speech Language Pathologist       3256 N VALD…
#> # ℹ 1,701 more rows
npi_nlm("Atlanta")
#> Warning: ! 10000 Results Found
#> ✔ Returning API limit of [7500].
#> ◼ NPPES [NLM]
#> ◼ 7500 RWS [16 PGS]
#> ◼ 7500 RWS [16 PGS]
#> # A tibble: 7,500 × 4
#>    full_name          npi        specialty                          full_address
#>    <chr>              <chr>      <chr>                              <chr>       
#>  1 ADKINS, ATLANTA    1356023022 Technician                         224 MILL ST…
#>  2 HUFF, ATLANTA      1851157887 Counselor                          446 MORGAN …
#>  3 MORELIK, ATLANTA   1407621188 Case Manager/Care Coordinator      60 W SUNBRI…
#>  4 CRANIAS, ATLANTA   1720656531 Physician Assistant                2824 ENTERP…
#>  5 MCCRAVEY, ATLANTA  1538777081 Nurse Practitioner                 2166 E BROA…
#>  6 WILLIAMS, ATLANTA  1023670940 Registered Nurse                   240 BUCKTAI…
#>  7 MACKLIN, ATLANTA   1083207948 Registered Dietitian or Nutrition… 2115 STEPHE…
#>  8 ST.MARTIN, ATLANTA 1114544004 Counselor                          20331 FLANA…
#>  9 BLACK, ATLANTA     1992516892 Specialist/Technologist            5716 S ARRO…
#> 10 ROLOFF, ATLANTA    1073242244 Dentist                            13000 TECHN…
#> # ℹ 7,490 more rows
npi_nppes(npi = npi_ex$k[1:2]) |> str()
#> tibble [2 × 9] (S3: tbl_df/tbl/data.frame)
#>  $ enumeration_type : chr [1:2] "NPI-2" "NPI-2"
#>  $ number           : chr [1:2] "1225701881" "1174270805"
#>  $ addresses        :List of 2
#>   ..$ :'data.frame':	2 obs. of  9 variables:
#>   .. ..$ country_code    : chr [1:2] "US" "US"
#>   .. ..$ country_name    : chr [1:2] "United States" "United States"
#>   .. ..$ address_purpose : chr [1:2] "LOCATION" "MAILING"
#>   .. ..$ address_type    : chr [1:2] "DOM" "DOM"
#>   .. ..$ address_1       : chr [1:2] "26400 W 12 MILE RD STE 140" "26400 W 12 MILE RD STE 140"
#>   .. ..$ city            : chr [1:2] "SOUTHFIELD" "SOUTHFIELD"
#>   .. ..$ state           : chr [1:2] "MI" "MI"
#>   .. ..$ postal_code     : chr [1:2] "480341753" "480341753"
#>   .. ..$ telephone_number: chr [1:2] "866-658-8214" "866-658-8214"
#>   ..$ :'data.frame':	2 obs. of  9 variables:
#>   .. ..$ country_code    : chr [1:2] "US" "US"
#>   .. ..$ country_name    : chr [1:2] "United States" "United States"
#>   .. ..$ address_purpose : chr [1:2] "LOCATION" "MAILING"
#>   .. ..$ address_type    : chr [1:2] "DOM" "DOM"
#>   .. ..$ address_1       : chr [1:2] "1414 WALNUT ST" "333 S DESPLAINES ST STE 201"
#>   .. ..$ city            : chr [1:2] "PHILADELPHIA" "CHICAGO"
#>   .. ..$ state           : chr [1:2] "PA" "IL"
#>   .. ..$ postal_code     : chr [1:2] "191023824" "606615514"
#>   .. ..$ telephone_number: chr [1:2] "855-563-2639" "855-563-2639"
#>  $ practiceLocations:List of 2
#>   ..$ : chr NA
#>   ..$ : chr NA
#>  $ basic            :List of 2
#>   ..$ :List of 10
#>   .. ..$ organization_name                    : chr "VIOS FERTILITY INSTITUTE MICHIGAN PLLC"
#>   .. ..$ organizational_subpart               : chr "NO"
#>   .. ..$ enumeration_date                     : chr "2021-07-29"
#>   .. ..$ last_updated                         : chr "2024-02-12"
#>   .. ..$ certification_date                   : chr "2024-02-12"
#>   .. ..$ status                               : chr "A"
#>   .. ..$ authorized_official_first_name       : chr "KATRINA"
#>   .. ..$ authorized_official_last_name        : chr "MARSHALL"
#>   .. ..$ authorized_official_telephone_number : chr "7734359036"
#>   .. ..$ authorized_official_title_or_position: chr "Dir Ins Cred & Payor Contracts"
#>   ..$ :List of 12
#>   .. ..$ organization_name                    : chr "PEACH STATE MEDICAL PRACTICE, PC"
#>   .. ..$ organizational_subpart               : chr "NO"
#>   .. ..$ enumeration_date                     : chr "2022-03-02"
#>   .. ..$ last_updated                         : chr "2024-05-20"
#>   .. ..$ certification_date                   : chr "2024-05-20"
#>   .. ..$ status                               : chr "A"
#>   .. ..$ authorized_official_first_name       : chr "FAHIMEH"
#>   .. ..$ authorized_official_last_name        : chr "SASAN"
#>   .. ..$ authorized_official_telephone_number : chr "8555632639"
#>   .. ..$ authorized_official_title_or_position: chr "Founding OBGYN"
#>   .. ..$ authorized_official_name_prefix      : chr "Dr."
#>   .. ..$ authorized_official_credential       : chr "DO"
#>  $ taxonomies       :List of 2
#>   ..$ :'data.frame':	1 obs. of  6 variables:
#>   .. ..$ code          : chr "207VE0102X"
#>   .. ..$ taxonomy_group: chr "193400000X - Multiple Single Specialty Group"
#>   .. ..$ desc          : chr "Obstetrics & Gynecology, Reproductive Endocrinology"
#>   .. ..$ state         : logi NA
#>   .. ..$ license       : logi NA
#>   .. ..$ primary       : logi TRUE
#>   ..$ :'data.frame':	2 obs. of  6 variables:
#>   .. ..$ code          : chr [1:2] "207VE0102X" "207VG0400X"
#>   .. ..$ taxonomy_group: chr [1:2] "193200000X - Multi-Specialty Group" "193200000X - Multi-Specialty Group"
#>   .. ..$ desc          : chr [1:2] "Obstetrics & Gynecology, Reproductive Endocrinology" "Obstetrics & Gynecology, Gynecology"
#>   .. ..$ state         : logi [1:2] NA NA
#>   .. ..$ license       : logi [1:2] NA NA
#>   .. ..$ primary       : logi [1:2] TRUE FALSE
#>  $ identifiers      :List of 2
#>   ..$ : chr NA
#>   ..$ : chr NA
#>  $ endpoints        :List of 2
#>   ..$ : chr NA
#>   ..$ : chr NA
#>  $ other_names      :List of 2
#>   ..$ : chr NA
#>   ..$ :'data.frame':	1 obs. of  3 variables:
#>   .. ..$ organization_name: chr "KINDBODY OF PA MEDICAL PRACTICE, PC"
#>   .. ..$ type             : chr "Doing Business As"
#>   .. ..$ code             : chr "3"
npi_nppes(npi = npi_ex$k[1]) |> str()
#> tibble [1 × 9] (S3: tbl_df/tbl/data.frame)
#>  $ enumeration_type : chr "NPI-2"
#>  $ number           : chr "1225701881"
#>  $ addresses        :List of 1
#>   ..$ :'data.frame':	2 obs. of  9 variables:
#>   .. ..$ country_code    : chr [1:2] "US" "US"
#>   .. ..$ country_name    : chr [1:2] "United States" "United States"
#>   .. ..$ address_purpose : chr [1:2] "LOCATION" "MAILING"
#>   .. ..$ address_type    : chr [1:2] "DOM" "DOM"
#>   .. ..$ address_1       : chr [1:2] "26400 W 12 MILE RD STE 140" "26400 W 12 MILE RD STE 140"
#>   .. ..$ city            : chr [1:2] "SOUTHFIELD" "SOUTHFIELD"
#>   .. ..$ state           : chr [1:2] "MI" "MI"
#>   .. ..$ postal_code     : chr [1:2] "480341753" "480341753"
#>   .. ..$ telephone_number: chr [1:2] "866-658-8214" "866-658-8214"
#>  $ practiceLocations:List of 1
#>   ..$ : list()
#>  $ basic            :'data.frame':	1 obs. of  10 variables:
#>   ..$ organization_name                    : chr "VIOS FERTILITY INSTITUTE MICHIGAN PLLC"
#>   ..$ organizational_subpart               : chr "NO"
#>   ..$ enumeration_date                     : chr "2021-07-29"
#>   ..$ last_updated                         : chr "2024-02-12"
#>   ..$ certification_date                   : chr "2024-02-12"
#>   ..$ status                               : chr "A"
#>   ..$ authorized_official_first_name       : chr "KATRINA"
#>   ..$ authorized_official_last_name        : chr "MARSHALL"
#>   ..$ authorized_official_telephone_number : chr "7734359036"
#>   ..$ authorized_official_title_or_position: chr "Dir Ins Cred & Payor Contracts"
#>  $ taxonomies       :List of 1
#>   ..$ :'data.frame':	1 obs. of  6 variables:
#>   .. ..$ code          : chr "207VE0102X"
#>   .. ..$ taxonomy_group: chr "193400000X - Multiple Single Specialty Group"
#>   .. ..$ desc          : chr "Obstetrics & Gynecology, Reproductive Endocrinology"
#>   .. ..$ state         : logi NA
#>   .. ..$ license       : logi NA
#>   .. ..$ primary       : logi TRUE
#>  $ identifiers      :List of 1
#>   ..$ : list()
#>  $ endpoints        :List of 1
#>   ..$ : list()
#>  $ other_names      :List of 1
#>   ..$ : list()