Open Payments API Endpoint Classes

Open Payments API Endpoint Classes

Usage

open_endpoint(alias)

open_group(alias)

open_temporal(alias)

open_troup(alias)

Arguments

alias

<chr> endpoint alias

Value

An S7 <open_endpoint>, <open_group>, <open_temporal>, or <open_troup> object

Examples

open_endpoint("PROF_covered")
#> <open_endpoint>
#>  @ identifier: chr "https://openpaymentsdata.cms.gov/api/1/datastore/query/23160558-6742-54ff-8b9f-cac7d514ff4e/0"
#>  @ metadata  :List of 4
#>  .. $ title      : chr "Covered Recipient Profile Supplement"
#>  .. $ description: chr "The Covered Recipient Profile Supplement file contains information about physicians and non-physician practitio"| __truncated__
#>  .. $ modified   : Date[1:1], format: "2025-01-30"
#>  .. $ download   : chr "https://download.cms.gov/openpayments/PHPRFL_P01302025_01212025/OP_CVRD_RCPNT_PRFL_SPLMTL_P01302025_01212025.csv"
#>  @ dimensions: <class_dimensions>
#>  .. @ limit : int 500
#>  .. @ rows  : int 1530412
#>  .. @ pages : int 3061
#>  .. @ fields:List of 32
#>  .. .. $ covered_recipient_profile_type                 : chr "covered_recipient_profile_type"
#>  .. .. $ covered_recipient_profile_id                   : chr "covered_recipient_profile_id"
#>  .. .. $ associated_covered_recipient_profile_id_1      : chr "associated_covered_recipient_profile_id_1"
#>  .. .. $ associated_covered_recipient_profile_id_2      : chr "associated_covered_recipient_profile_id_2"
#>  .. .. $ covered_recipient_npi                          : chr "covered_recipient_npi"
#>  .. .. $ covered_recipient_profile_first_name           : chr "covered_recipient_profile_first_name"
#>  .. .. $ covered_recipient_profile_middle_name          : chr "covered_recipient_profile_middle_name"
#>  .. .. $ covered_recipient_profile_last_name            : chr "covered_recipient_profile_last_name"
#>  .. .. $ covered_recipient_profile_suffix               : chr "covered_recipient_profile_suffix"
#>  .. .. $ covered_recipient_profile_alternate_first_name : chr "covered_recipient_profile_alternate_first_name"
#>  .. .. $ covered_recipient_profile_alternate_middle_name: chr "covered_recipient_profile_alternate_middle_name"
#>  .. .. $ covered_recipient_profile_alternate_last_name  : chr "covered_recipient_profile_alternate_last_name"
#>  .. .. $ covered_recipient_profile_alternate_suffix     : chr "covered_recipient_profile_alternate_suffix"
#>  .. .. $ covered_recipient_profile_address_line_1       : chr "covered_recipient_profile_address_line_1"
#>  .. .. $ covered_recipient_profile_address_line_2       : chr "covered_recipient_profile_address_line_2"
#>  .. .. $ covered_recipient_profile_city                 : chr "covered_recipient_profile_city"
#>  .. .. $ covered_recipient_profile_state                : chr "covered_recipient_profile_state"
#>  .. .. $ covered_recipient_profile_zipcode              : chr "covered_recipient_profile_zipcode"
#>  .. .. $ covered_recipient_profile_country_name         : chr "covered_recipient_profile_country_name"
#>  .. .. $ covered_recipient_profile_province_name        : chr "covered_recipient_profile_province_name"
#>  .. .. $ covered_recipient_profile_primary_specialty    : chr "covered_recipient_profile_primary_specialty"
#>  .. .. $ covered_recipient_profile_ops_taxonomy_1       : chr "covered_recipient_profile_ops_taxonomy_1"
#>  .. .. $ covered_recipient_profile_ops_taxonomy_2       : chr "covered_recipient_profile_ops_taxonomy_2"
#>  .. .. $ covered_recipient_profile_ops_taxonomy_3       : chr "covered_recipient_profile_ops_taxonomy_3"
#>  .. .. $ covered_recipient_profile_ops_taxonomy_4       : chr "covered_recipient_profile_ops_taxonomy_4"
#>  .. .. $ covered_recipient_profile_ops_taxonomy_5       : chr "covered_recipient_profile_ops_taxonomy_5"
#>  .. .. $ covered_recipient_profile_ops_taxonomy_6       : chr "covered_recipient_profile_ops_taxonomy_6"
#>  .. .. $ covered_recipient_profile_license_state_code_1 : chr "covered_recipient_profile_license_state_code_1"
#>  .. .. $ covered_recipient_profile_license_state_code_2 : chr "covered_recipient_profile_license_state_code_2"
#>  .. .. $ covered_recipient_profile_license_state_code_3 : chr "covered_recipient_profile_license_state_code_3"
#>  .. .. $ covered_recipient_profile_license_state_code_4 : chr "covered_recipient_profile_license_state_code_4"
#>  .. .. $ covered_recipient_profile_license_state_code_5 : chr "covered_recipient_profile_license_state_code_5"
open_group("profile")
#> <open_group>
#>  @ group  : chr "Open Payments Profiles"
#>  @ members:List of 6
#>  .. $ PROF_covered    : <open_endpoint>
#>  ..  ..@ identifier: chr "https://openpaymentsdata.cms.gov/api/1/datastore/query/23160558-6742-54ff-8b9f-cac7d514ff4e/0"
#>  ..  ..@ metadata  :List of 4
#>  .. .. .. $ title      : chr "Covered Recipient Profile Supplement"
#>  .. .. .. $ description: chr "The Covered Recipient Profile Supplement file contains information about physicians and non-physician practitio"| __truncated__
#>  .. .. .. $ modified   : Date[1:1], format: "2025-01-30"
#>  .. .. .. $ download   : chr "https://download.cms.gov/openpayments/PHPRFL_P01302025_01212025/OP_CVRD_RCPNT_PRFL_SPLMTL_P01302025_01212025.csv"
#>  ..  ..@ dimensions: <class_dimensions>
#>  .. .. .. @ limit : int 500
#>  .. .. .. @ rows  : int 1530412
#>  .. .. .. @ pages : int 3061
#>  .. .. .. @ fields:List of 32
#>  .. .. .. .. $ covered_recipient_profile_type                 : chr "covered_recipient_profile_type"
#>  .. .. .. .. $ covered_recipient_profile_id                   : chr "covered_recipient_profile_id"
#>  .. .. .. .. $ associated_covered_recipient_profile_id_1      : chr "associated_covered_recipient_profile_id_1"
#>  .. .. .. .. $ associated_covered_recipient_profile_id_2      : chr "associated_covered_recipient_profile_id_2"
#>  .. .. .. .. $ covered_recipient_npi                          : chr "covered_recipient_npi"
#>  .. .. .. .. $ covered_recipient_profile_first_name           : chr "covered_recipient_profile_first_name"
#>  .. .. .. .. $ covered_recipient_profile_middle_name          : chr "covered_recipient_profile_middle_name"
#>  .. .. .. .. $ covered_recipient_profile_last_name            : chr "covered_recipient_profile_last_name"
#>  .. .. .. .. $ covered_recipient_profile_suffix               : chr "covered_recipient_profile_suffix"
#>  .. .. .. .. $ covered_recipient_profile_alternate_first_name : chr "covered_recipient_profile_alternate_first_name"
#>  .. .. .. .. $ covered_recipient_profile_alternate_middle_name: chr "covered_recipient_profile_alternate_middle_name"
#>  .. .. .. .. $ covered_recipient_profile_alternate_last_name  : chr "covered_recipient_profile_alternate_last_name"
#>  .. .. .. .. $ covered_recipient_profile_alternate_suffix     : chr "covered_recipient_profile_alternate_suffix"
#>  .. .. .. .. $ covered_recipient_profile_address_line_1       : chr "covered_recipient_profile_address_line_1"
#>  .. .. .. .. $ covered_recipient_profile_address_line_2       : chr "covered_recipient_profile_address_line_2"
#>  .. .. .. .. $ covered_recipient_profile_city                 : chr "covered_recipient_profile_city"
#>  .. .. .. .. $ covered_recipient_profile_state                : chr "covered_recipient_profile_state"
#>  .. .. .. .. $ covered_recipient_profile_zipcode              : chr "covered_recipient_profile_zipcode"
#>  .. .. .. .. $ covered_recipient_profile_country_name         : chr "covered_recipient_profile_country_name"
#>  .. .. .. .. $ covered_recipient_profile_province_name        : chr "covered_recipient_profile_province_name"
#>  .. .. .. .. $ covered_recipient_profile_primary_specialty    : chr "covered_recipient_profile_primary_specialty"
#>  .. .. .. .. $ covered_recipient_profile_ops_taxonomy_1       : chr "covered_recipient_profile_ops_taxonomy_1"
#>  .. .. .. .. $ covered_recipient_profile_ops_taxonomy_2       : chr "covered_recipient_profile_ops_taxonomy_2"
#>  .. .. .. .. $ covered_recipient_profile_ops_taxonomy_3       : chr "covered_recipient_profile_ops_taxonomy_3"
#>  .. .. .. .. $ covered_recipient_profile_ops_taxonomy_4       : chr "covered_recipient_profile_ops_taxonomy_4"
#>  .. .. .. .. $ covered_recipient_profile_ops_taxonomy_5       : chr "covered_recipient_profile_ops_taxonomy_5"
#>  .. .. .. .. $ covered_recipient_profile_ops_taxonomy_6       : chr "covered_recipient_profile_ops_taxonomy_6"
#>  .. .. .. .. $ covered_recipient_profile_license_state_code_1 : chr "covered_recipient_profile_license_state_code_1"
#>  .. .. .. .. $ covered_recipient_profile_license_state_code_2 : chr "covered_recipient_profile_license_state_code_2"
#>  .. .. .. .. $ covered_recipient_profile_license_state_code_3 : chr "covered_recipient_profile_license_state_code_3"
#>  .. .. .. .. $ covered_recipient_profile_license_state_code_4 : chr "covered_recipient_profile_license_state_code_4"
#>  .. .. .. .. $ covered_recipient_profile_license_state_code_5 : chr "covered_recipient_profile_license_state_code_5"
#>  .. $ PROF_physician  : <open_endpoint>
#>  ..  ..@ identifier: chr "https://openpaymentsdata.cms.gov/api/1/datastore/query/6ed6ae76-2999-49da-b0b2-d7df150ac754/0"
#>  ..  ..@ metadata  :List of 4
#>  .. .. .. $ title      : chr "Physician (Distinct) Profile Information"
#>  .. .. .. $ description: chr "This dataset is pre-filtered based on the most frequent searches of Open Payments data."
#>  .. .. .. $ modified   : Date[1:1], format: "2025-01-30"
#>  .. .. .. $ download   : chr "https://download.cms.gov/openpayments/SMRY_RPTS_P01302025_01212025/PBLCTN_PHYSN_NON_PHYSN_PRCTNR_PRFL_SRCH_P013"| __truncated__
#>  ..  ..@ dimensions: <class_dimensions>
#>  .. .. .. @ limit : int 500
#>  .. .. .. @ rows  : int 1528054
#>  .. .. .. @ pages : int 3057
#>  .. .. .. @ fields:List of 33
#>  .. .. .. .. $ covered_recipient_profile_type                  : chr "covered_recipient_profile_type"
#>  .. .. .. .. $ covered_recipient_profile_id                    : chr "covered_recipient_profile_id"
#>  .. .. .. .. $ covered_recipient_npi                           : chr "covered_recipient_npi"
#>  .. .. .. .. $ covered_recipient_profile_first_name            : chr "covered_recipient_profile_first_name"
#>  .. .. .. .. $ covered_recipient_profile_middle_name           : chr "covered_recipient_profile_middle_name"
#>  .. .. .. .. $ covered_recipient_profile_last_name             : chr "covered_recipient_profile_last_name"
#>  .. .. .. .. $ covered_recipient_profile_suffix                : chr "covered_recipient_profile_suffix"
#>  .. .. .. .. $ covered_recipient_profile_alternate_first_name1 : chr "covered_recipient_profile_alternate_first_name1"
#>  .. .. .. .. $ covered_recipient_profile_alternate_middle_name1: chr "covered_recipient_profile_alternate_middle_name1"
#>  .. .. .. .. $ covered_recipient_profile_alternate_last_name1  : chr "covered_recipient_profile_alternate_last_name1"
#>  .. .. .. .. $ covered_recipient_profile_alternate_suffix1     : chr "covered_recipient_profile_alternate_suffix1"
#>  .. .. .. .. $ covered_recipient_profile_address_line_1        : chr "covered_recipient_profile_address_line_1"
#>  .. .. .. .. $ covered_recipient_profile_address_line_2        : chr "covered_recipient_profile_address_line_2"
#>  .. .. .. .. $ covered_recipient_profile_city                  : chr "covered_recipient_profile_city"
#>  .. .. .. .. $ covered_recipient_profile_state                 : chr "covered_recipient_profile_state"
#>  .. .. .. .. $ covered_recipient_profile_zipcode               : chr "covered_recipient_profile_zipcode"
#>  .. .. .. .. $ covered_recipient_profile_country_name          : chr "covered_recipient_profile_country_name"
#>  .. .. .. .. $ covered_recipient_profile_province_name         : chr "covered_recipient_profile_province_name"
#>  .. .. .. .. $ covered_recipient_profile_primary_specialty     : chr "covered_recipient_profile_primary_specialty"
#>  .. .. .. .. $ covered_recipient_profile_license_state_code_1  : chr "covered_recipient_profile_license_state_code_1"
#>  .. .. .. .. $ covered_recipient_profile_license_state_code_2  : chr "covered_recipient_profile_license_state_code_2"
#>  .. .. .. .. $ covered_recipient_profile_license_state_code_3  : chr "covered_recipient_profile_license_state_code_3"
#>  .. .. .. .. $ covered_recipient_profile_license_state_code_4  : chr "covered_recipient_profile_license_state_code_4"
#>  .. .. .. .. $ covered_recipient_profile_license_state_code_5  : chr "covered_recipient_profile_license_state_code_5"
#>  .. .. .. .. $ covered_recipient_profile_alternate_first_name2 : chr "covered_recipient_profile_alternate_first_name2"
#>  .. .. .. .. $ covered_recipient_profile_alternate_last_name2  : chr "covered_recipient_profile_alternate_last_name2"
#>  .. .. .. .. $ covered_recipient_profile_alternate_first_name3 : chr "covered_recipient_profile_alternate_first_name3"
#>  .. .. .. .. $ covered_recipient_profile_alternate_last_name3  : chr "covered_recipient_profile_alternate_last_name3"
#>  .. .. .. .. $ covered_recipient_profile_alternate_first_name4 : chr "covered_recipient_profile_alternate_first_name4"
#>  .. .. .. .. $ covered_recipient_profile_alternate_last_name4  : chr "covered_recipient_profile_alternate_last_name4"
#>  .. .. .. .. $ covered_recipient_profile_alternate_first_name5 : chr "covered_recipient_profile_alternate_first_name5"
#>  .. .. .. .. $ covered_recipient_profile_alternate_last_name5  : chr "covered_recipient_profile_alternate_last_name5"
#>  .. .. .. .. $ has_multiple_ids                                : chr "has_multiple_ids"
#>  .. $ PROF_information: <open_endpoint>
#>  ..  ..@ identifier: chr "https://openpaymentsdata.cms.gov/api/1/datastore/query/f2ed17cc-f045-4e25-9b28-75c72f0c7efe/0"
#>  ..  ..@ metadata  :List of 4
#>  .. .. .. $ title      : chr "Profile Information"
#>  .. .. .. $ description: chr "This dataset provides profile information to display on payment search results."
#>  .. .. .. $ modified   : Date[1:1], format: "2025-01-30"
#>  .. .. .. $ download   : chr "https://download.cms.gov/openpayments/SMRY_RPTS_P01302025_01212025/all-entities-profile-search.csv"
#>  ..  ..@ dimensions: <class_dimensions>
#>  .. .. .. @ limit : int 500
#>  .. .. .. @ rows  : int 1534695
#>  .. .. .. @ pages : int 3070
#>  .. .. .. @ fields:List of 20
#>  .. .. .. .. $ entity_id             : chr "entity_id"
#>  .. .. .. .. $ entity_provider_type  : chr "entity_provider_type"
#>  .. .. .. .. $ entity_npi            : chr "entity_npi"
#>  .. .. .. .. $ entity_alternate_id   : chr "entity_alternate_id"
#>  .. .. .. .. $ entity_type           : chr "entity_type"
#>  .. .. .. .. $ entity_name           : chr "entity_name"
#>  .. .. .. .. $ entity_alternate_name1: chr "entity_alternate_name1"
#>  .. .. .. .. $ entity_alternate_name2: chr "entity_alternate_name2"
#>  .. .. .. .. $ entity_alternate_name3: chr "entity_alternate_name3"
#>  .. .. .. .. $ entity_alternate_name4: chr "entity_alternate_name4"
#>  .. .. .. .. $ entity_alternate_name5: chr "entity_alternate_name5"
#>  .. .. .. .. $ entity_address_line_1 : chr "entity_address_line_1"
#>  .. .. .. .. $ entity_address_line_2 : chr "entity_address_line_2"
#>  .. .. .. .. $ entity_city           : chr "entity_city"
#>  .. .. .. .. $ entity_state          : chr "entity_state"
#>  .. .. .. .. $ entity_zipcode        : chr "entity_zipcode"
#>  .. .. .. .. $ entity_country_name   : chr "entity_country_name"
#>  .. .. .. .. $ entity_province_name  : chr "entity_province_name"
#>  .. .. .. .. $ entity_specialty      : chr "entity_specialty"
#>  .. .. .. .. $ location              : chr "location"
#>  .. $ PROF_mapping    : <open_endpoint>
#>  ..  ..@ identifier: chr "https://openpaymentsdata.cms.gov/api/1/datastore/query/d19e5be2-5b8c-562c-af22-c243503bcfac/0"
#>  ..  ..@ metadata  :List of 4
#>  .. .. .. $ title      : chr "Provider Profile ID Mapping Table"
#>  .. .. .. $ description: chr "Maps secondary/duplicate provider profile IDs to the primary ID in the Open Payments system. For multiple reaso"| __truncated__
#>  .. .. .. $ modified   : Date[1:1], format: "2025-01-30"
#>  .. .. .. $ download   : chr "https://download.cms.gov/openpayments/SMRY_RPTS_P01302025_01212025/PBLCTN_PRVDR_PRFL_MAPPING_P01302025_01212025.csv"
#>  ..  ..@ dimensions: <class_dimensions>
#>  .. .. .. @ limit : int 500
#>  .. .. .. @ rows  : int 2358
#>  .. .. .. @ pages : int 5
#>  .. .. .. @ fields:List of 2
#>  .. .. .. .. $ primary_provider_profile_id  : chr "primary_provider_profile_id"
#>  .. .. .. .. $ secondary_provider_profile_id: chr "secondary_provider_profile_id"
#>  .. $ PROF_entity     : <open_endpoint>
#>  ..  ..@ identifier: chr "https://openpaymentsdata.cms.gov/api/1/datastore/query/1cf0c6c0-c377-466e-b78f-037b442559f8/0"
#>  ..  ..@ metadata  :List of 4
#>  .. .. .. $ title      : chr "Reporting Entity Profile Information"
#>  .. .. .. $ description: chr "This dataset is pre-filtered based on the most frequent searches of Open Payments data."
#>  .. .. .. $ modified   : Date[1:1], format: "2025-01-30"
#>  .. .. .. $ download   : chr "https://download.cms.gov/openpayments/SMRY_RPTS_P01302025_01212025/PBLCTN_RPTG_ORG_PRFL_SRCH_P01302025_01212025.csv"
#>  ..  ..@ dimensions: <class_dimensions>
#>  .. .. .. @ limit : int 500
#>  .. .. .. @ rows  : int 2843
#>  .. .. .. @ pages : int 6
#>  .. .. .. @ fields:List of 9
#>  .. .. .. .. $ amgpo_making_payment_id             : chr "amgpo_making_payment_id"
#>  .. .. .. .. $ amgpo_making_payment_name           : chr "amgpo_making_payment_name"
#>  .. .. .. .. $ amgpo_making_payment_state          : chr "amgpo_making_payment_state"
#>  .. .. .. .. $ amgpo_making_payment_country        : chr "amgpo_making_payment_country"
#>  .. .. .. .. $ amgpo_making_payment_alternate_name1: chr "amgpo_making_payment_alternate_name1"
#>  .. .. .. .. $ amgpo_making_payment_alternate_name2: chr "amgpo_making_payment_alternate_name2"
#>  .. .. .. .. $ amgpo_making_payment_alternate_name3: chr "amgpo_making_payment_alternate_name3"
#>  .. .. .. .. $ amgpo_making_payment_alternate_name4: chr "amgpo_making_payment_alternate_name4"
#>  .. .. .. .. $ amgpo_making_payment_alternate_name5: chr "amgpo_making_payment_alternate_name5"
#>  .. $ PROF_teaching   : <open_endpoint>
#>  ..  ..@ identifier: chr "https://openpaymentsdata.cms.gov/api/1/datastore/query/8219fcb3-fb3d-4b51-9f65-be8a048c0608/0"
#>  ..  ..@ metadata  :List of 4
#>  .. .. .. $ title      : chr "Teaching Hospital Profile Information"
#>  .. .. .. $ description: chr "This dataset is pre-filtered based on the most frequent searches of Open Payments data."
#>  .. .. .. $ modified   : Date[1:1], format: "2025-01-30"
#>  .. .. .. $ download   : chr "https://download.cms.gov/openpayments/SMRY_RPTS_P01302025_01212025/PBLCTN_TH_PRFL_SRCH_P01302025_01212025.csv"
#>  ..  ..@ dimensions: <class_dimensions>
#>  .. .. .. @ limit : int 500
#>  .. .. .. @ rows  : int 1440
#>  .. .. .. @ pages : int 3
#>  .. .. .. @ fields:List of 12
#>  .. .. .. .. $ teaching_hospital_ccn            : chr "teaching_hospital_ccn"
#>  .. .. .. .. $ teaching_hospital_name           : chr "teaching_hospital_name"
#>  .. .. .. .. $ teaching_hospital_address1       : chr "teaching_hospital_address1"
#>  .. .. .. .. $ teaching_hospital_address2       : chr "teaching_hospital_address2"
#>  .. .. .. .. $ teaching_hospital_city           : chr "teaching_hospital_city"
#>  .. .. .. .. $ teaching_hospital_state          : chr "teaching_hospital_state"
#>  .. .. .. .. $ teaching_hospital_zip_code       : chr "teaching_hospital_zip_code"
#>  .. .. .. .. $ teaching_hospital_alternate_name1: chr "teaching_hospital_alternate_name1"
#>  .. .. .. .. $ teaching_hospital_alternate_name2: chr "teaching_hospital_alternate_name2"
#>  .. .. .. .. $ teaching_hospital_alternate_name3: chr "teaching_hospital_alternate_name3"
#>  .. .. .. .. $ teaching_hospital_alternate_name4: chr "teaching_hospital_alternate_name4"
#>  .. .. .. .. $ teaching_hospital_alternate_name5: chr "teaching_hospital_alternate_name5"
open_temporal("GROUP_all")
#> <open_temporal>
#>  @ metadata  :List of 3
#>  .. $ title      : chr "Payments Grouped by Reporting Entities, Covered Recipient, and Nature of Payments"
#>  .. $ description: chr "This dataset is pre-filtered based on the most frequent searches of Open Payments data."
#>  .. $ modified   : Date[1:1], format: "2025-01-30"
#>  @ dimensions: <class_dimensions>
#>  .. @ limit : int 500
#>  .. @ rows  : int 4588015
#>  .. @ pages : int 9177
#>  .. @ fields:List of 8
#>  .. .. $ amgpo_id                   : chr "amgpo_id"
#>  .. .. $ recipient_id               : chr "recipient_id"
#>  .. .. $ recipient_type             : chr "recipient_type"
#>  .. .. $ nature_of_payment_type_code: chr "nature_of_payment_type_code"
#>  .. .. $ number_of_transaction      : chr "number_of_transaction"
#>  .. .. $ total_amount               : chr "total_amount"
#>  .. .. $ recipient_name             : chr "recipient_name"
#>  .. .. $ amgpo_name                 : chr "amgpo_name"
#>  @ endpoints : tibble [7 × 3] (S3: tbl_df/tbl/data.frame)
#>  $ year      : int [1:7] 2023 2022 2021 2020 2019 2018 2017
#>  $ identifier: chr [1:7] "https://openpaymentsdata.cms.gov/api/1/datastore/query/29c23b6a-dd1e-4d82-a606-8d792c7f229e/0" "https://openpaymentsdata.cms.gov/api/1/datastore/query/171e369a-6f25-49eb-aa36-37f0a497e5ca/0" "https://openpaymentsdata.cms.gov/api/1/datastore/query/f240c437-9d8a-5be7-95a5-8ce8e0d970bf/0" "https://openpaymentsdata.cms.gov/api/1/datastore/query/532df191-473f-4bcc-9056-a6260b9437b5/0" ...
#>  $ download  : chr [1:7] "https://download.cms.gov/openpayments/SMRY_RPTS_P01302025_01212025/PBLCTN_SMRY_BY_AMGPO_BY_CR_BY_NTR_OF_PYMT_PG"| __truncated__ "https://download.cms.gov/openpayments/SMRY_RPTS_P01302025_01212025/PBLCTN_SMRY_BY_AMGPO_BY_CR_BY_NTR_OF_PYMT_PG"| __truncated__ "https://download.cms.gov/openpayments/SMRY_RPTS_P01302025_01212025/PBLCTN_SMRY_BY_AMGPO_BY_CR_BY_NTR_OF_PYMT_PG"| __truncated__ "https://download.cms.gov/openpayments/SMRY_RPTS_P01302025_01212025/PBLCTN_SMRY_BY_AMGPO_BY_CR_BY_NTR_OF_PYMT_PG"| __truncated__ ...
open_troup("grouped_payment")
#> <open_troup>
#>  @ group  : chr "Payments Grouped by Year"
#>  @ members:List of 4
#>  .. $ GROUP_recip_nature : <open_temporal>
#>  ..  ..@ metadata  :List of 3
#>  .. .. .. $ title      : chr "Payments Grouped by Covered Recipient and Nature of Payments"
#>  .. .. .. $ description: chr "This dataset is pre-filtered based on the most frequent searches of Open Payments data."
#>  .. .. .. $ modified   : Date[1:1], format: "2025-01-30"
#>  ..  ..@ dimensions: <class_dimensions>
#>  .. .. .. @ limit : int 500
#>  .. .. .. @ rows  : int 1184051
#>  .. .. .. @ pages : int 2369
#>  .. .. .. @ fields:List of 10
#>  .. .. .. .. $ recipient_id                         : chr "recipient_id"
#>  .. .. .. .. $ recipient_type                       : chr "recipient_type"
#>  .. .. .. .. $ covered_recipient_npi                : chr "covered_recipient_npi"
#>  .. .. .. .. $ covered_recipient_profile_first_name : chr "covered_recipient_profile_first_name"
#>  .. .. .. .. $ covered_recipient_profile_middle_name: chr "covered_recipient_profile_middle_name"
#>  .. .. .. .. $ covered_recipient_profile_last_name  : chr "covered_recipient_profile_last_name"
#>  .. .. .. .. $ teaching_hospital_name               : chr "teaching_hospital_name"
#>  .. .. .. .. $ nature_of_payment_type_code          : chr "nature_of_payment_type_code"
#>  .. .. .. .. $ number_of_transaction                : chr "number_of_transaction"
#>  .. .. .. .. $ total_amount                         : chr "total_amount"
#>  ..  ..@ endpoints : tibble [7 × 3] (S3: tbl_df/tbl/data.frame)
#>  $ year      : int [1:7] 2023 2022 2021 2020 2019 2018 2017
#>  $ identifier: chr [1:7] "https://openpaymentsdata.cms.gov/api/1/datastore/query/4017acfa-0eeb-4b8a-b68c-3b41e779c4a2/0" "https://openpaymentsdata.cms.gov/api/1/datastore/query/686b765e-f4b7-4d35-a050-fb5af926dbbe/0" "https://openpaymentsdata.cms.gov/api/1/datastore/query/d35243be-55cb-5087-931e-a294e0e0917f/0" "https://openpaymentsdata.cms.gov/api/1/datastore/query/1fd2be47-cb72-435a-b2df-142f38e99a05/0" ...
#>  $ download  : chr [1:7] "https://download.cms.gov/openpayments/SMRY_RPTS_P01302025_01212025/PBLCTN_SMRY_BY_CR_BY_NTR_OF_PYMT_PGYR2023_P0"| __truncated__ "https://download.cms.gov/openpayments/SMRY_RPTS_P01302025_01212025/PBLCTN_SMRY_BY_CR_BY_NTR_OF_PYMT_PGYR2022_P0"| __truncated__ "https://download.cms.gov/openpayments/SMRY_RPTS_P01302025_01212025/PBLCTN_SMRY_BY_CR_BY_NTR_OF_PYMT_PGYR2021_P0"| __truncated__ "https://download.cms.gov/openpayments/SMRY_RPTS_P01302025_01212025/PBLCTN_SMRY_BY_CR_BY_NTR_OF_PYMT_PGYR2020_P0"| __truncated__ ...
#>  .. $ GROUP_recip_entity : <open_temporal>
#>  ..  ..@ metadata  :List of 3
#>  .. .. .. $ title      : chr "Payments Grouped by Covered Recipient and Reporting Entities"
#>  .. .. .. $ description: chr "This dataset is pre-filtered based on the most frequent searches of Open Payments data."
#>  .. .. .. $ modified   : Date[1:1], format: "2025-01-30"
#>  ..  ..@ dimensions: <class_dimensions>
#>  .. .. .. @ limit : int 500
#>  .. .. .. @ rows  : int 4416875
#>  .. .. .. @ pages : int 8834
#>  .. .. .. @ fields:List of 12
#>  .. .. .. .. $ recipient_id                         : chr "recipient_id"
#>  .. .. .. .. $ recipient_type                       : chr "recipient_type"
#>  .. .. .. .. $ covered_recipient_npi                : chr "covered_recipient_npi"
#>  .. .. .. .. $ covered_recipient_profile_first_name : chr "covered_recipient_profile_first_name"
#>  .. .. .. .. $ covered_recipient_profile_middle_name: chr "covered_recipient_profile_middle_name"
#>  .. .. .. .. $ covered_recipient_profile_last_name  : chr "covered_recipient_profile_last_name"
#>  .. .. .. .. $ teaching_hospital_name               : chr "teaching_hospital_name"
#>  .. .. .. .. $ payment_type                         : chr "payment_type"
#>  .. .. .. .. $ amgpo_name                           : chr "amgpo_name"
#>  .. .. .. .. $ amgpo_id                             : chr "amgpo_id"
#>  .. .. .. .. $ number_of_transaction                : chr "number_of_transaction"
#>  .. .. .. .. $ total_amount                         : chr "total_amount"
#>  ..  ..@ endpoints : tibble [7 × 3] (S3: tbl_df/tbl/data.frame)
#>  $ year      : int [1:7] 2023 2022 2021 2020 2019 2018 2017
#>  $ identifier: chr [1:7] "https://openpaymentsdata.cms.gov/api/1/datastore/query/4fbf0094-3c7b-425a-aabe-5778052559ad/0" "https://openpaymentsdata.cms.gov/api/1/datastore/query/d2d87f39-b117-4ac8-9f36-a088fb22a25b/0" "https://openpaymentsdata.cms.gov/api/1/datastore/query/7c9ebb83-8f88-5462-ab7d-1925b0113cfc/0" "https://openpaymentsdata.cms.gov/api/1/datastore/query/78ad9c8d-d7ae-4066-bf2d-a66e3c9d6e3b/0" ...
#>  $ download  : chr [1:7] "https://download.cms.gov/openpayments/SMRY_RPTS_P01302025_01212025/PBLCTN_SMRY_BY_CR_BY_AMGPO_PGYR2023_P01302025_01212025.csv" "https://download.cms.gov/openpayments/SMRY_RPTS_P01302025_01212025/PBLCTN_SMRY_BY_CR_BY_AMGPO_PGYR2022_P01302025_01212025.csv" "https://download.cms.gov/openpayments/SMRY_RPTS_P01302025_01212025/PBLCTN_SMRY_BY_CR_BY_AMGPO_PGYR2021_P01302025_01212025.csv" "https://download.cms.gov/openpayments/SMRY_RPTS_P01302025_01212025/PBLCTN_SMRY_BY_CR_BY_AMGPO_PGYR2020_P01302025_01212025.csv" ...
#>  .. $ GROUP_entity_nature: <open_temporal>
#>  ..  ..@ metadata  :List of 3
#>  .. .. .. $ title      : chr "Payments Grouped by Reporting Entities and Nature of Payments"
#>  .. .. .. $ description: chr "This dataset is pre-filtered based on the most frequent searches of Open Payments data."
#>  .. .. .. $ modified   : Date[1:1], format: "2025-01-30"
#>  ..  ..@ dimensions: <class_dimensions>
#>  .. .. .. @ limit : int 500
#>  .. .. .. @ rows  : int 6192
#>  .. .. .. @ pages : int 13
#>  .. .. .. @ fields:List of 5
#>  .. .. .. .. $ amgpo_id                   : chr "amgpo_id"
#>  .. .. .. .. $ nature_of_payment_type_code: chr "nature_of_payment_type_code"
#>  .. .. .. .. $ number_of_transaction      : chr "number_of_transaction"
#>  .. .. .. .. $ total_amount               : chr "total_amount"
#>  .. .. .. .. $ amgpo_name                 : chr "amgpo_name"
#>  ..  ..@ endpoints : tibble [7 × 3] (S3: tbl_df/tbl/data.frame)
#>  $ year      : int [1:7] 2023 2022 2021 2020 2019 2018 2017
#>  $ identifier: chr [1:7] "https://openpaymentsdata.cms.gov/api/1/datastore/query/72008dab-0953-4226-a4cd-9f1872e8170d/0" "https://openpaymentsdata.cms.gov/api/1/datastore/query/cedcd327-4e5d-43f9-8eb1-c11850fa7c66/0" "https://openpaymentsdata.cms.gov/api/1/datastore/query/a2714020-f599-52b0-b12b-8f5b510e547a/0" "https://openpaymentsdata.cms.gov/api/1/datastore/query/7dfe6254-2aee-454f-9ea5-df611a9a7b55/0" ...
#>  $ download  : chr [1:7] "https://download.cms.gov/openpayments/SMRY_RPTS_P01302025_01212025/PBLCTN_SMRY_BY_AMGPO_BY_NTR_OF_PYMT_PGYR2023"| __truncated__ "https://download.cms.gov/openpayments/SMRY_RPTS_P01302025_01212025/PBLCTN_SMRY_BY_AMGPO_BY_NTR_OF_PYMT_PGYR2022"| __truncated__ "https://download.cms.gov/openpayments/SMRY_RPTS_P01302025_01212025/PBLCTN_SMRY_BY_AMGPO_BY_NTR_OF_PYMT_PGYR2021"| __truncated__ "https://download.cms.gov/openpayments/SMRY_RPTS_P01302025_01212025/PBLCTN_SMRY_BY_AMGPO_BY_NTR_OF_PYMT_PGYR2020"| __truncated__ ...
#>  .. $ GROUP_all          : <open_temporal>
#>  ..  ..@ metadata  :List of 3
#>  .. .. .. $ title      : chr "Payments Grouped by Reporting Entities, Covered Recipient, and Nature of Payments"
#>  .. .. .. $ description: chr "This dataset is pre-filtered based on the most frequent searches of Open Payments data."
#>  .. .. .. $ modified   : Date[1:1], format: "2025-01-30"
#>  ..  ..@ dimensions: <class_dimensions>
#>  .. .. .. @ limit : int 500
#>  .. .. .. @ rows  : int 4588015
#>  .. .. .. @ pages : int 9177
#>  .. .. .. @ fields:List of 8
#>  .. .. .. .. $ amgpo_id                   : chr "amgpo_id"
#>  .. .. .. .. $ recipient_id               : chr "recipient_id"
#>  .. .. .. .. $ recipient_type             : chr "recipient_type"
#>  .. .. .. .. $ nature_of_payment_type_code: chr "nature_of_payment_type_code"
#>  .. .. .. .. $ number_of_transaction      : chr "number_of_transaction"
#>  .. .. .. .. $ total_amount               : chr "total_amount"
#>  .. .. .. .. $ recipient_name             : chr "recipient_name"
#>  .. .. .. .. $ amgpo_name                 : chr "amgpo_name"
#>  ..  ..@ endpoints : tibble [7 × 3] (S3: tbl_df/tbl/data.frame)
#>  $ year      : int [1:7] 2023 2022 2021 2020 2019 2018 2017
#>  $ identifier: chr [1:7] "https://openpaymentsdata.cms.gov/api/1/datastore/query/29c23b6a-dd1e-4d82-a606-8d792c7f229e/0" "https://openpaymentsdata.cms.gov/api/1/datastore/query/171e369a-6f25-49eb-aa36-37f0a497e5ca/0" "https://openpaymentsdata.cms.gov/api/1/datastore/query/f240c437-9d8a-5be7-95a5-8ce8e0d970bf/0" "https://openpaymentsdata.cms.gov/api/1/datastore/query/532df191-473f-4bcc-9056-a6260b9437b5/0" ...
#>  $ download  : chr [1:7] "https://download.cms.gov/openpayments/SMRY_RPTS_P01302025_01212025/PBLCTN_SMRY_BY_AMGPO_BY_CR_BY_NTR_OF_PYMT_PG"| __truncated__ "https://download.cms.gov/openpayments/SMRY_RPTS_P01302025_01212025/PBLCTN_SMRY_BY_AMGPO_BY_CR_BY_NTR_OF_PYMT_PG"| __truncated__ "https://download.cms.gov/openpayments/SMRY_RPTS_P01302025_01212025/PBLCTN_SMRY_BY_AMGPO_BY_CR_BY_NTR_OF_PYMT_PG"| __truncated__ "https://download.cms.gov/openpayments/SMRY_RPTS_P01302025_01212025/PBLCTN_SMRY_BY_AMGPO_BY_CR_BY_NTR_OF_PYMT_PG"| __truncated__ ...