Open Payments API Endpoints

Open Payments API Endpoints

Usage

open_endpoint(alias, call = caller_env())

open_temporal(alias, call = caller_env())

open_group(alias, call = caller_env(), ...)

Arguments

alias

<chr> endpoint or group alias

call

<env> environment to use for error reporting

...

Additional arguments passed to the group constructor

Value

An S7 <class_endpoint>, <class_temporal> or <class_group> object

Examples

open_endpoint("profile_covered")
#> <class_endpoint>
#>  @ identifier: chr "https://openpaymentsdata.cms.gov/api/1/datastore/query/23160558-6742-54ff-8b9f-cac7d514ff4e/0"
#>  @ metadata  :List of 5
#>  .. $ api        : chr "Open Payments"
#>  .. $ title      : chr "Covered Recipient Profile Supplement"
#>  .. $ description: chr "The Covered Recipient Profile Supplement file contains information about physicians and non-physician practitio"| __truncated__
#>  .. $ modified   : Date[1:1], format: "2025-01-30"
#>  .. $ download   : chr "https://download.cms.gov/openpayments/PHPRFL_P01302025_01212025/OP_CVRD_RCPNT_PRFL_SPLMTL_P01302025_01212025.csv"
#>  @ dimensions: <class_dimensions>
#>  .. @ limit : int 500
#>  .. @ rows  : int 1530412
#>  .. @ pages : int 3061
#>  .. @ fields:List of 32
#>  .. .. $ covered_recipient_profile_type                 : chr(0) 
#>  .. .. $ covered_recipient_profile_id                   : chr(0) 
#>  .. .. $ associated_covered_recipient_profile_id_1      : chr(0) 
#>  .. .. $ associated_covered_recipient_profile_id_2      : chr(0) 
#>  .. .. $ covered_recipient_npi                          : chr(0) 
#>  .. .. $ covered_recipient_profile_first_name           : chr(0) 
#>  .. .. $ covered_recipient_profile_middle_name          : chr(0) 
#>  .. .. $ covered_recipient_profile_last_name            : chr(0) 
#>  .. .. $ covered_recipient_profile_suffix               : chr(0) 
#>  .. .. $ covered_recipient_profile_alternate_first_name : chr(0) 
#>  .. .. $ covered_recipient_profile_alternate_middle_name: chr(0) 
#>  .. .. $ covered_recipient_profile_alternate_last_name  : chr(0) 
#>  .. .. $ covered_recipient_profile_alternate_suffix     : chr(0) 
#>  .. .. $ covered_recipient_profile_address_line_1       : chr(0) 
#>  .. .. $ covered_recipient_profile_address_line_2       : chr(0) 
#>  .. .. $ covered_recipient_profile_city                 : chr(0) 
#>  .. .. $ covered_recipient_profile_state                : chr(0) 
#>  .. .. $ covered_recipient_profile_zipcode              : chr(0) 
#>  .. .. $ covered_recipient_profile_country_name         : chr(0) 
#>  .. .. $ covered_recipient_profile_province_name        : chr(0) 
#>  .. .. $ covered_recipient_profile_primary_specialty    : chr(0) 
#>  .. .. $ covered_recipient_profile_ops_taxonomy_1       : chr(0) 
#>  .. .. $ covered_recipient_profile_ops_taxonomy_2       : chr(0) 
#>  .. .. $ covered_recipient_profile_ops_taxonomy_3       : chr(0) 
#>  .. .. $ covered_recipient_profile_ops_taxonomy_4       : chr(0) 
#>  .. .. $ covered_recipient_profile_ops_taxonomy_5       : chr(0) 
#>  .. .. $ covered_recipient_profile_ops_taxonomy_6       : chr(0) 
#>  .. .. $ covered_recipient_profile_license_state_code_1 : chr(0) 
#>  .. .. $ covered_recipient_profile_license_state_code_2 : chr(0) 
#>  .. .. $ covered_recipient_profile_license_state_code_3 : chr(0) 
#>  .. .. $ covered_recipient_profile_license_state_code_4 : chr(0) 
#>  .. .. $ covered_recipient_profile_license_state_code_5 : chr(0) 
open_temporal("grouped_state_nature")
#> <class_temporal>
#>  @ metadata  :List of 4
#>  .. $ api        : chr "Open Payments [Temporal]"
#>  .. $ title      : chr "State Payment Totals Grouped by Nature of Payment for all Years"
#>  .. $ description: chr "Groups payment data by state, recipient type, and nature of payment."
#>  .. $ modified   : Date[1:1], format: "2025-01-30"
#>  @ dimensions: <class_dimensions>
#>  .. @ limit : int 500
#>  .. @ rows  : int 16758
#>  .. @ pages : int 34
#>  .. @ fields:List of 28
#>  .. .. $ country_code                                          : chr(0) 
#>  .. .. $ country_name                                          : chr(0) 
#>  .. .. $ state_code                                            : chr(0) 
#>  .. .. $ state_name                                            : chr(0) 
#>  .. .. $ nature_of_payment                                     : chr(0) 
#>  .. .. $ recipient_type                                        : chr(0) 
#>  .. .. $ program_year                                          : chr(0) 
#>  .. .. $ total_number_of_physicians                            : chr(0) 
#>  .. .. $ total_number_of_non_physician_practitioners           : chr(0) 
#>  .. .. $ total_number_of_teaching_hospitals                    : chr(0) 
#>  .. .. $ total_payment_amount_physician                        : chr(0) 
#>  .. .. $ total_payment_amount_non_physician_practitioner       : chr(0) 
#>  .. .. $ total_payment_amount_teaching_hospital                : chr(0) 
#>  .. .. $ mean_total_payment_amount_physician                   : chr(0) 
#>  .. .. $ mean_total_payment_amount_non_physician_practitioner  : chr(0) 
#>  .. .. $ mean_total_payment_amount_teaching_hospital           : chr(0) 
#>  .. .. $ median_total_payment_amount_physician                 : chr(0) 
#>  .. .. $ median_total_payment_amount_non_physician_practitioner: chr(0) 
#>  .. .. $ median_total_payment_amount_teaching_hospital         : chr(0) 
#>  .. .. $ total_payment_count_physician                         : chr(0) 
#>  .. .. $ total_payment_count_non_physician_practitioner        : chr(0) 
#>  .. .. $ total_payment_count_teaching_hospital                 : chr(0) 
#>  .. .. $ mean_total_payment_count_physician                    : chr(0) 
#>  .. .. $ mean_total_payment_count_non_physician_practitioner   : chr(0) 
#>  .. .. $ mean_total_payment_count_teaching_hospital            : chr(0) 
#>  .. .. $ median_total_payment_count_physician                  : chr(0) 
#>  .. .. $ median_total_payment_count_non_physician_practitioner : chr(0) 
#>  .. .. $ median_total_payment_count_teaching_hospital          : chr(0) 
#>  @ endpoints : tibble [7 × 3] (S3: tbl_df/tbl/data.frame)
#>  $ year      : int [1:7] 2023 2022 2021 2020 2019 2018 2017
#>  $ identifier: chr [1:7] "https://openpaymentsdata.cms.gov/api/1/datastore/query/f83aed3a-7504-4fca-aabb-ac0338c79688/0" "https://openpaymentsdata.cms.gov/api/1/datastore/query/c64c40af-23a5-4419-8b64-692684483664/0" "https://openpaymentsdata.cms.gov/api/1/datastore/query/7fad5b56-2cfe-52ab-ac4d-3200780287d4/0" "https://openpaymentsdata.cms.gov/api/1/datastore/query/7f3055b6-73d8-5aff-be15-55873552e2cd/0" ...
#>  $ download  : chr [1:7] "https://download.cms.gov/openpayments/SMRY_RPTS_P01302025_01212025/PBLCTN_NTR_OF_PYMT_BY_STATE_SMRY_PGYR2023_P0"| __truncated__ "https://download.cms.gov/openpayments/SMRY_RPTS_P01302025_01212025/PBLCTN_NTR_OF_PYMT_BY_STATE_SMRY_PGYR2022_P0"| __truncated__ "https://download.cms.gov/openpayments/SMRY_RPTS_P01302025_01212025/PBLCTN_NTR_OF_PYMT_BY_STATE_SMRY_PGYR2021_P0"| __truncated__ "https://download.cms.gov/openpayments/SMRY_RPTS_P01302025_01212025/PBLCTN_NTR_OF_PYMT_BY_STATE_SMRY_PGYR2020_P0"| __truncated__ ...
open_group("payment_detailed")
#> <class_group>
#>  @ group  : chr "Open Payments by Year (Detailed)"
#>  @ members:List of 3
#>  .. $ payment_general  : <class_temporal>
#>  ..  ..@ metadata  :List of 4
#>  .. .. .. $ api        : chr "Open Payments [Temporal]"
#>  .. .. .. $ title      : chr "General Payment Data"
#>  .. .. .. $ description: chr "All general (non-research, non-ownership related) payments from the program year"
#>  .. .. .. $ modified   : Date[1:1], format: "2025-01-30"
#>  ..  ..@ dimensions: <class_dimensions>
#>  .. .. .. @ limit : int 500
#>  .. .. .. @ rows  : int 14607336
#>  .. .. .. @ pages : int 29215
#>  .. .. .. @ fields:List of 91
#>  .. .. .. .. $ change_type                                                     : chr(0) 
#>  .. .. .. .. $ covered_recipient_type                                          : chr(0) 
#>  .. .. .. .. $ teaching_hospital_ccn                                           : chr(0) 
#>  .. .. .. .. $ teaching_hospital_id                                            : chr(0) 
#>  .. .. .. .. $ teaching_hospital_name                                          : chr(0) 
#>  .. .. .. .. $ covered_recipient_profile_id                                    : chr(0) 
#>  .. .. .. .. $ covered_recipient_npi                                           : chr(0) 
#>  .. .. .. .. $ covered_recipient_first_name                                    : chr(0) 
#>  .. .. .. .. $ covered_recipient_middle_name                                   : chr(0) 
#>  .. .. .. .. $ covered_recipient_last_name                                     : chr(0) 
#>  .. .. .. .. $ covered_recipient_name_suffix                                   : chr(0) 
#>  .. .. .. .. $ recipient_primary_business_street_address_line1                 : chr(0) 
#>  .. .. .. .. $ recipient_primary_business_street_address_line2                 : chr(0) 
#>  .. .. .. .. $ recipient_city                                                  : chr(0) 
#>  .. .. .. .. $ recipient_state                                                 : chr(0) 
#>  .. .. .. .. $ recipient_zip_code                                              : chr(0) 
#>  .. .. .. .. $ recipient_country                                               : chr(0) 
#>  .. .. .. .. $ recipient_province                                              : chr(0) 
#>  .. .. .. .. $ recipient_postal_code                                           : chr(0) 
#>  .. .. .. .. $ covered_recipient_primary_type_1                                : chr(0) 
#>  .. .. .. .. $ covered_recipient_primary_type_2                                : chr(0) 
#>  .. .. .. .. $ covered_recipient_primary_type_3                                : chr(0) 
#>  .. .. .. .. $ covered_recipient_primary_type_4                                : chr(0) 
#>  .. .. .. .. $ covered_recipient_primary_type_5                                : chr(0) 
#>  .. .. .. .. $ covered_recipient_primary_type_6                                : chr(0) 
#>  .. .. .. .. $ covered_recipient_specialty_1                                   : chr(0) 
#>  .. .. .. .. $ covered_recipient_specialty_2                                   : chr(0) 
#>  .. .. .. .. $ covered_recipient_specialty_3                                   : chr(0) 
#>  .. .. .. .. $ covered_recipient_specialty_4                                   : chr(0) 
#>  .. .. .. .. $ covered_recipient_specialty_5                                   : chr(0) 
#>  .. .. .. .. $ covered_recipient_specialty_6                                   : chr(0) 
#>  .. .. .. .. $ covered_recipient_license_state_code1                           : chr(0) 
#>  .. .. .. .. $ covered_recipient_license_state_code2                           : chr(0) 
#>  .. .. .. .. $ covered_recipient_license_state_code3                           : chr(0) 
#>  .. .. .. .. $ covered_recipient_license_state_code4                           : chr(0) 
#>  .. .. .. .. $ covered_recipient_license_state_code5                           : chr(0) 
#>  .. .. .. .. $ submitting_applicable_manufacturer_or_applicable_gpo_name       : chr(0) 
#>  .. .. .. .. $ applicable_manufacturer_or_applicable_gpo_making_payment_id     : chr(0) 
#>  .. .. .. .. $ applicable_manufacturer_or_applicable_gpo_making_payment_name   : chr(0) 
#>  .. .. .. .. $ applicable_manufacturer_or_applicable_gpo_making_payment_state  : chr(0) 
#>  .. .. .. .. $ applicable_manufacturer_or_applicable_gpo_making_payment_country: chr(0) 
#>  .. .. .. .. $ total_amount_of_payment_usdollars                               : chr(0) 
#>  .. .. .. .. $ date_of_payment                                                 : chr(0) 
#>  .. .. .. .. $ number_of_payments_included_in_total_amount                     : chr(0) 
#>  .. .. .. .. $ form_of_payment_or_transfer_of_value                            : chr(0) 
#>  .. .. .. .. $ nature_of_payment_or_transfer_of_value                          : chr(0) 
#>  .. .. .. .. $ city_of_travel                                                  : chr(0) 
#>  .. .. .. .. $ state_of_travel                                                 : chr(0) 
#>  .. .. .. .. $ country_of_travel                                               : chr(0) 
#>  .. .. .. .. $ physician_ownership_indicator                                   : chr(0) 
#>  .. .. .. .. $ third_party_payment_recipient_indicator                         : chr(0) 
#>  .. .. .. .. $ name_of_third_party_entity_receiving_payment_or_transfer_of_ccfc: chr(0) 
#>  .. .. .. .. $ charity_indicator                                               : chr(0) 
#>  .. .. .. .. $ third_party_equals_covered_recipient_indicator                  : chr(0) 
#>  .. .. .. .. $ contextual_information                                          : chr(0) 
#>  .. .. .. .. $ delay_in_publication_indicator                                  : chr(0) 
#>  .. .. .. .. $ record_id                                                       : chr(0) 
#>  .. .. .. .. $ dispute_status_for_publication                                  : chr(0) 
#>  .. .. .. .. $ related_product_indicator                                       : chr(0) 
#>  .. .. .. .. $ covered_or_noncovered_indicator_1                               : chr(0) 
#>  .. .. .. .. $ indicate_drug_or_biological_or_device_or_medical_supply_1       : chr(0) 
#>  .. .. .. .. $ product_category_or_therapeutic_area_1                          : chr(0) 
#>  .. .. .. .. $ name_of_drug_or_biological_or_device_or_medical_supply_1        : chr(0) 
#>  .. .. .. .. $ associated_drug_or_biological_ndc_1                             : chr(0) 
#>  .. .. .. .. $ associated_device_or_medical_supply_pdi_1                       : chr(0) 
#>  .. .. .. .. $ covered_or_noncovered_indicator_2                               : chr(0) 
#>  .. .. .. .. $ indicate_drug_or_biological_or_device_or_medical_supply_2       : chr(0) 
#>  .. .. .. .. $ product_category_or_therapeutic_area_2                          : chr(0) 
#>  .. .. .. .. $ name_of_drug_or_biological_or_device_or_medical_supply_2        : chr(0) 
#>  .. .. .. .. $ associated_drug_or_biological_ndc_2                             : chr(0) 
#>  .. .. .. .. $ associated_device_or_medical_supply_pdi_2                       : chr(0) 
#>  .. .. .. .. $ covered_or_noncovered_indicator_3                               : chr(0) 
#>  .. .. .. .. $ indicate_drug_or_biological_or_device_or_medical_supply_3       : chr(0) 
#>  .. .. .. .. $ product_category_or_therapeutic_area_3                          : chr(0) 
#>  .. .. .. .. $ name_of_drug_or_biological_or_device_or_medical_supply_3        : chr(0) 
#>  .. .. .. .. $ associated_drug_or_biological_ndc_3                             : chr(0) 
#>  .. .. .. .. $ associated_device_or_medical_supply_pdi_3                       : chr(0) 
#>  .. .. .. .. $ covered_or_noncovered_indicator_4                               : chr(0) 
#>  .. .. .. .. $ indicate_drug_or_biological_or_device_or_medical_supply_4       : chr(0) 
#>  .. .. .. .. $ product_category_or_therapeutic_area_4                          : chr(0) 
#>  .. .. .. .. $ name_of_drug_or_biological_or_device_or_medical_supply_4        : chr(0) 
#>  .. .. .. .. $ associated_drug_or_biological_ndc_4                             : chr(0) 
#>  .. .. .. .. $ associated_device_or_medical_supply_pdi_4                       : chr(0) 
#>  .. .. .. .. $ covered_or_noncovered_indicator_5                               : chr(0) 
#>  .. .. .. .. $ indicate_drug_or_biological_or_device_or_medical_supply_5       : chr(0) 
#>  .. .. .. .. $ product_category_or_therapeutic_area_5                          : chr(0) 
#>  .. .. .. .. $ name_of_drug_or_biological_or_device_or_medical_supply_5        : chr(0) 
#>  .. .. .. .. $ associated_drug_or_biological_ndc_5                             : chr(0) 
#>  .. .. .. .. $ associated_device_or_medical_supply_pdi_5                       : chr(0) 
#>  .. .. .. .. $ program_year                                                    : chr(0) 
#>  .. .. .. .. $ payment_publication_date                                        : chr(0) 
#>  ..  ..@ endpoints : tibble [7 × 3] (S3: tbl_df/tbl/data.frame)
#>  $ year      : int [1:7] 2023 2022 2021 2020 2019 2018 2017
#>  $ identifier: chr [1:7] "https://openpaymentsdata.cms.gov/api/1/datastore/query/fb3a65aa-c901-4a38-a813-b04b00dfa2a9/0" "https://openpaymentsdata.cms.gov/api/1/datastore/query/df01c2f8-dc1f-4e79-96cb-8208beaf143c/0" "https://openpaymentsdata.cms.gov/api/1/datastore/query/0380bbeb-aea1-58b6-b708-829f92a48202/0" "https://openpaymentsdata.cms.gov/api/1/datastore/query/a08c4b30-5cf3-4948-ad40-36f404619019/0" ...
#>  $ download  : chr [1:7] "https://download.cms.gov/openpayments/PGYR2023_P01302025_01212025/OP_DTL_GNRL_PGYR2023_P01302025_01212025.csv" "https://download.cms.gov/openpayments/PGYR2022_P01302025_01212025/OP_DTL_GNRL_PGYR2022_P01302025_01212025.csv" "https://download.cms.gov/openpayments/PGYR2021_P01302025_01212025/OP_DTL_GNRL_PGYR2021_P01302025_01212025.csv" "https://download.cms.gov/openpayments/PGYR2020_P01302025_01212025/OP_DTL_GNRL_PGYR2020_P01302025_01212025.csv" ...
#>  .. $ payment_ownership: <class_temporal>
#>  ..  ..@ metadata  :List of 4
#>  .. .. .. $ api        : chr "Open Payments [Temporal]"
#>  .. .. .. $ title      : chr "Ownership Payment Data"
#>  .. .. .. $ description: chr "All ownership and investment payments from the program year"
#>  .. .. .. $ modified   : Date[1:1], format: "2025-01-30"
#>  ..  ..@ dimensions: <class_dimensions>
#>  .. .. .. @ limit : int 500
#>  .. .. .. @ rows  : int 4013
#>  .. .. .. @ pages : int 9
#>  .. .. .. @ fields:List of 30
#>  .. .. .. .. $ change_type                                                     : chr(0) 
#>  .. .. .. .. $ physician_profile_id                                            : chr(0) 
#>  .. .. .. .. $ physician_npi                                                   : chr(0) 
#>  .. .. .. .. $ physician_first_name                                            : chr(0) 
#>  .. .. .. .. $ physician_middle_name                                           : chr(0) 
#>  .. .. .. .. $ physician_last_name                                             : chr(0) 
#>  .. .. .. .. $ physician_name_suffix                                           : chr(0) 
#>  .. .. .. .. $ recipient_primary_business_street_address_line1                 : chr(0) 
#>  .. .. .. .. $ recipient_primary_business_street_address_line2                 : chr(0) 
#>  .. .. .. .. $ recipient_city                                                  : chr(0) 
#>  .. .. .. .. $ recipient_state                                                 : chr(0) 
#>  .. .. .. .. $ recipient_zip_code                                              : chr(0) 
#>  .. .. .. .. $ recipient_country                                               : chr(0) 
#>  .. .. .. .. $ recipient_province                                              : chr(0) 
#>  .. .. .. .. $ recipient_postal_code                                           : chr(0) 
#>  .. .. .. .. $ physician_primary_type                                          : chr(0) 
#>  .. .. .. .. $ physician_specialty                                             : chr(0) 
#>  .. .. .. .. $ record_id                                                       : chr(0) 
#>  .. .. .. .. $ program_year                                                    : chr(0) 
#>  .. .. .. .. $ total_amount_invested_usdollars                                 : chr(0) 
#>  .. .. .. .. $ value_of_interest                                               : chr(0) 
#>  .. .. .. .. $ terms_of_interest                                               : chr(0) 
#>  .. .. .. .. $ submitting_applicable_manufacturer_or_applicable_gpo_name       : chr(0) 
#>  .. .. .. .. $ applicable_manufacturer_or_applicable_gpo_making_payment_id     : chr(0) 
#>  .. .. .. .. $ applicable_manufacturer_or_applicable_gpo_making_payment_name   : chr(0) 
#>  .. .. .. .. $ applicable_manufacturer_or_applicable_gpo_making_payment_state  : chr(0) 
#>  .. .. .. .. $ applicable_manufacturer_or_applicable_gpo_making_payment_country: chr(0) 
#>  .. .. .. .. $ dispute_status_for_publication                                  : chr(0) 
#>  .. .. .. .. $ interest_held_by_physician_or_an_immediate_family_member        : chr(0) 
#>  .. .. .. .. $ payment_publication_date                                        : chr(0) 
#>  ..  ..@ endpoints : tibble [7 × 3] (S3: tbl_df/tbl/data.frame)
#>  $ year      : int [1:7] 2023 2022 2021 2020 2019 2018 2017
#>  $ identifier: chr [1:7] "https://openpaymentsdata.cms.gov/api/1/datastore/query/ac0bc85c-02e3-45d9-89e8-2ff43da85df7/0" "https://openpaymentsdata.cms.gov/api/1/datastore/query/37792388-800f-427a-9e02-b11601454eeb/0" "https://openpaymentsdata.cms.gov/api/1/datastore/query/b0c03b8d-06df-58f2-8ce2-4daeffee147e/0" "https://openpaymentsdata.cms.gov/api/1/datastore/query/a9a0bf48-6b96-4589-b4c2-3c5dcfbeaca2/0" ...
#>  $ download  : chr [1:7] "https://download.cms.gov/openpayments/PGYR2023_P01302025_01212025/OP_DTL_OWNRSHP_PGYR2023_P01302025_01212025.csv" "https://download.cms.gov/openpayments/PGYR2022_P01302025_01212025/OP_DTL_OWNRSHP_PGYR2022_P01302025_01212025.csv" "https://download.cms.gov/openpayments/PGYR2021_P01302025_01212025/OP_DTL_OWNRSHP_PGYR2021_P01302025_01212025.csv" "https://download.cms.gov/openpayments/PGYR2020_P01302025_01212025/OP_DTL_OWNRSHP_PGYR2020_P01302025_01212025.csv" ...
#>  .. $ payment_research : <class_temporal>
#>  ..  ..@ metadata  :List of 4
#>  .. .. .. $ api        : chr "Open Payments [Temporal]"
#>  .. .. .. $ title      : chr "Research Payment Data"
#>  .. .. .. $ description: chr "All research-related payments from the program year"
#>  .. .. .. $ modified   : Date[1:1], format: "2025-01-30"
#>  ..  ..@ dimensions: <class_dimensions>
#>  .. .. .. @ limit : int 500
#>  .. .. .. @ rows  : int 1027926
#>  .. .. .. @ pages : int 2056
#>  .. .. .. @ fields:List of 252
#>  .. .. .. .. $ change_type                                                     : chr(0) 
#>  .. .. .. .. $ covered_recipient_type                                          : chr(0) 
#>  .. .. .. .. $ noncovered_recipient_entity_name                                : chr(0) 
#>  .. .. .. .. $ teaching_hospital_ccn                                           : chr(0) 
#>  .. .. .. .. $ teaching_hospital_id                                            : chr(0) 
#>  .. .. .. .. $ teaching_hospital_name                                          : chr(0) 
#>  .. .. .. .. $ covered_recipient_profile_id                                    : chr(0) 
#>  .. .. .. .. $ covered_recipient_npi                                           : chr(0) 
#>  .. .. .. .. $ covered_recipient_first_name                                    : chr(0) 
#>  .. .. .. .. $ covered_recipient_middle_name                                   : chr(0) 
#>  .. .. .. .. $ covered_recipient_last_name                                     : chr(0) 
#>  .. .. .. .. $ covered_recipient_name_suffix                                   : chr(0) 
#>  .. .. .. .. $ recipient_primary_business_street_address_line1                 : chr(0) 
#>  .. .. .. .. $ recipient_primary_business_street_address_line2                 : chr(0) 
#>  .. .. .. .. $ recipient_city                                                  : chr(0) 
#>  .. .. .. .. $ recipient_state                                                 : chr(0) 
#>  .. .. .. .. $ recipient_zip_code                                              : chr(0) 
#>  .. .. .. .. $ recipient_country                                               : chr(0) 
#>  .. .. .. .. $ recipient_province                                              : chr(0) 
#>  .. .. .. .. $ recipient_postal_code                                           : chr(0) 
#>  .. .. .. .. $ covered_recipient_primary_type_1                                : chr(0) 
#>  .. .. .. .. $ covered_recipient_primary_type_2                                : chr(0) 
#>  .. .. .. .. $ covered_recipient_primary_type_3                                : chr(0) 
#>  .. .. .. .. $ covered_recipient_primary_type_4                                : chr(0) 
#>  .. .. .. .. $ covered_recipient_primary_type_5                                : chr(0) 
#>  .. .. .. .. $ covered_recipient_primary_type_6                                : chr(0) 
#>  .. .. .. .. $ covered_recipient_specialty_1                                   : chr(0) 
#>  .. .. .. .. $ covered_recipient_specialty_2                                   : chr(0) 
#>  .. .. .. .. $ covered_recipient_specialty_3                                   : chr(0) 
#>  .. .. .. .. $ covered_recipient_specialty_4                                   : chr(0) 
#>  .. .. .. .. $ covered_recipient_specialty_5                                   : chr(0) 
#>  .. .. .. .. $ covered_recipient_specialty_6                                   : chr(0) 
#>  .. .. .. .. $ covered_recipient_license_state_code1                           : chr(0) 
#>  .. .. .. .. $ covered_recipient_license_state_code2                           : chr(0) 
#>  .. .. .. .. $ covered_recipient_license_state_code3                           : chr(0) 
#>  .. .. .. .. $ covered_recipient_license_state_code4                           : chr(0) 
#>  .. .. .. .. $ covered_recipient_license_state_code5                           : chr(0) 
#>  .. .. .. .. $ principal_investigator_1_covered_recipient_type                 : chr(0) 
#>  .. .. .. .. $ principal_investigator_1_profile_id                             : chr(0) 
#>  .. .. .. .. $ principal_investigator_1_npi                                    : chr(0) 
#>  .. .. .. .. $ principal_investigator_1_first_name                             : chr(0) 
#>  .. .. .. .. $ principal_investigator_1_middle_name                            : chr(0) 
#>  .. .. .. .. $ principal_investigator_1_last_name                              : chr(0) 
#>  .. .. .. .. $ principal_investigator_1_name_suffix                            : chr(0) 
#>  .. .. .. .. $ principal_investigator_1_business_street_address_line1          : chr(0) 
#>  .. .. .. .. $ principal_investigator_1_business_street_address_line2          : chr(0) 
#>  .. .. .. .. $ principal_investigator_1_city                                   : chr(0) 
#>  .. .. .. .. $ principal_investigator_1_state                                  : chr(0) 
#>  .. .. .. .. $ principal_investigator_1_zip_code                               : chr(0) 
#>  .. .. .. .. $ principal_investigator_1_country                                : chr(0) 
#>  .. .. .. .. $ principal_investigator_1_province                               : chr(0) 
#>  .. .. .. .. $ principal_investigator_1_postal_code                            : chr(0) 
#>  .. .. .. .. $ principal_investigator_1_primary_type_1                         : chr(0) 
#>  .. .. .. .. $ principal_investigator_1_primary_type_2                         : chr(0) 
#>  .. .. .. .. $ principal_investigator_1_primary_type_3                         : chr(0) 
#>  .. .. .. .. $ principal_investigator_1_primary_type_4                         : chr(0) 
#>  .. .. .. .. $ principal_investigator_1_primary_type_5                         : chr(0) 
#>  .. .. .. .. $ principal_investigator_1_primary_type_6                         : chr(0) 
#>  .. .. .. .. $ principal_investigator_1_specialty_1                            : chr(0) 
#>  .. .. .. .. $ principal_investigator_1_specialty_2                            : chr(0) 
#>  .. .. .. .. $ principal_investigator_1_specialty_3                            : chr(0) 
#>  .. .. .. .. $ principal_investigator_1_specialty_4                            : chr(0) 
#>  .. .. .. .. $ principal_investigator_1_specialty_5                            : chr(0) 
#>  .. .. .. .. $ principal_investigator_1_specialty_6                            : chr(0) 
#>  .. .. .. .. $ principal_investigator_1_license_state_code1                    : chr(0) 
#>  .. .. .. .. $ principal_investigator_1_license_state_code2                    : chr(0) 
#>  .. .. .. .. $ principal_investigator_1_license_state_code3                    : chr(0) 
#>  .. .. .. .. $ principal_investigator_1_license_state_code4                    : chr(0) 
#>  .. .. .. .. $ principal_investigator_1_license_state_code5                    : chr(0) 
#>  .. .. .. .. $ principal_investigator_2_covered_recipient_type                 : chr(0) 
#>  .. .. .. .. $ principal_investigator_2_profile_id                             : chr(0) 
#>  .. .. .. .. $ principal_investigator_2_npi                                    : chr(0) 
#>  .. .. .. .. $ principal_investigator_2_first_name                             : chr(0) 
#>  .. .. .. .. $ principal_investigator_2_middle_name                            : chr(0) 
#>  .. .. .. .. $ principal_investigator_2_last_name                              : chr(0) 
#>  .. .. .. .. $ principal_investigator_2_name_suffix                            : chr(0) 
#>  .. .. .. .. $ principal_investigator_2_business_street_address_line1          : chr(0) 
#>  .. .. .. .. $ principal_investigator_2_business_street_address_line2          : chr(0) 
#>  .. .. .. .. $ principal_investigator_2_city                                   : chr(0) 
#>  .. .. .. .. $ principal_investigator_2_state                                  : chr(0) 
#>  .. .. .. .. $ principal_investigator_2_zip_code                               : chr(0) 
#>  .. .. .. .. $ principal_investigator_2_country                                : chr(0) 
#>  .. .. .. .. $ principal_investigator_2_province                               : chr(0) 
#>  .. .. .. .. $ principal_investigator_2_postal_code                            : chr(0) 
#>  .. .. .. .. $ principal_investigator_2_primary_type_1                         : chr(0) 
#>  .. .. .. .. $ principal_investigator_2_primary_type_2                         : chr(0) 
#>  .. .. .. .. $ principal_investigator_2_primary_type_3                         : chr(0) 
#>  .. .. .. .. $ principal_investigator_2_primary_type_4                         : chr(0) 
#>  .. .. .. .. $ principal_investigator_2_primary_type_5                         : chr(0) 
#>  .. .. .. .. $ principal_investigator_2_primary_type_6                         : chr(0) 
#>  .. .. .. .. $ principal_investigator_2_specialty_1                            : chr(0) 
#>  .. .. .. .. $ principal_investigator_2_specialty_2                            : chr(0) 
#>  .. .. .. .. $ principal_investigator_2_specialty_3                            : chr(0) 
#>  .. .. .. .. $ principal_investigator_2_specialty_4                            : chr(0) 
#>  .. .. .. .. $ principal_investigator_2_specialty_5                            : chr(0) 
#>  .. .. .. .. $ principal_investigator_2_specialty_6                            : chr(0) 
#>  .. .. .. .. $ principal_investigator_2_license_state_code1                    : chr(0) 
#>  .. .. .. .. $ principal_investigator_2_license_state_code2                    : chr(0) 
#>  .. .. .. .. $ principal_investigator_2_license_state_code3                    : chr(0) 
#>  .. .. .. ..  [list output truncated]
#>  ..  ..@ endpoints : tibble [7 × 3] (S3: tbl_df/tbl/data.frame)
#>  $ year      : int [1:7] 2023 2022 2021 2020 2019 2018 2017
#>  $ identifier: chr [1:7] "https://openpaymentsdata.cms.gov/api/1/datastore/query/ec9521bf-9d97-4603-814c-f4132d34bc4f/0" "https://openpaymentsdata.cms.gov/api/1/datastore/query/fdc3c773-018a-412c-8a81-d7b8a13a037b/0" "https://openpaymentsdata.cms.gov/api/1/datastore/query/ce1d28dd-0094-5060-a036-580329439600/0" "https://openpaymentsdata.cms.gov/api/1/datastore/query/9c248e7e-7c7f-478b-ab84-ce0919d72c1c/0" ...
#>  $ download  : chr [1:7] "https://download.cms.gov/openpayments/PGYR2023_P01302025_01212025/OP_DTL_RSRCH_PGYR2023_P01302025_01212025.csv" "https://download.cms.gov/openpayments/PGYR2022_P01302025_01212025/OP_DTL_RSRCH_PGYR2022_P01302025_01212025.csv" "https://download.cms.gov/openpayments/PGYR2021_P01302025_01212025/OP_DTL_RSRCH_PGYR2021_P01302025_01212025.csv" "https://download.cms.gov/openpayments/PGYR2020_P01302025_01212025/OP_DTL_RSRCH_PGYR2020_P01302025_01212025.csv" ...