Provider API Endpoint Classes

Provider API Endpoint Classes

Usage

pro_endpoint(alias)

pro_group(alias)

Arguments

alias

<chr> endpoint alias

Value

An S7 <pro_endpoint> or<pro_group> object

Examples

pro_endpoint("asc_facility")
#> <pro_endpoint>
#>  @ identifier: chr "https://data.cms.gov/provider-data/api/1/datastore/query/4jcv-atw7/0"
#>  @ metadata  :List of 9
#>  .. $ title      : chr "Ambulatory Surgical Center Quality Measures - Facility"
#>  .. $ group      : chr "Hospitals"
#>  .. $ description: chr "A list of ambulatory surgical centers participating in the Ambulatory Surgical Center Quality Reporting (ASCQR)"| __truncated__
#>  .. $ issued     : Date[1:1], format: "2024-10-10"
#>  .. $ modified   : Date[1:1], format: "2025-04-08"
#>  .. $ released   : Date[1:1], format: "2025-04-30"
#>  .. $ site       : chr "https://data.cms.gov/provider-data/dataset/4jcv-atw7"
#>  .. $ dictionary : chr "https://data.cms.gov/provider-data/dataset/4jcv-atw7#data-dictionary"
#>  .. $ download   : chr "https://data.cms.gov/provider-data/sites/default/files/resources/dd03994fc93e296bb0297f1cd43cc987_1744668308/ASC_Facility.csv"
#>  @ dimensions: <class_dimensions>
#>  .. @ limit : int 2000
#>  .. @ rows  : int 11288
#>  .. @ pages : int 6
#>  .. @ fields:List of 50
#>  .. .. $ facility_name             : chr "facility_name"
#>  .. .. $ facility_id               : chr "facility_id"
#>  .. .. $ npi                       : chr "npi"
#>  .. .. $ citytown                  : chr "citytown"
#>  .. .. $ state                     : chr "state"
#>  .. .. $ zip_code                  : chr "zip_code"
#>  .. .. $ year                      : chr "year"
#>  .. .. $ asc1_rate                 : chr "asc1_rate"
#>  .. .. $ asc1_footnote             : chr "asc1_footnote"
#>  .. .. $ asc2_rate                 : chr "asc2_rate"
#>  .. .. $ asc2_footnote             : chr "asc2_footnote"
#>  .. .. $ asc3_rate                 : chr "asc3_rate"
#>  .. .. $ asc3_footnote             : chr "asc3_footnote"
#>  .. .. $ asc4_rate                 : chr "asc4_rate"
#>  .. .. $ asc4_footnote             : chr "asc4_footnote"
#>  .. .. $ asc9_rate                 : chr "asc9_rate"
#>  .. .. $ asc9_footnote             : chr "asc9_footnote"
#>  .. .. $ asc11_rate                : chr "asc11_rate"
#>  .. .. $ asc11_footnote            : chr "asc11_footnote"
#>  .. .. $ asc12_total_cases         : chr "asc12_total_cases"
#>  .. .. $ asc12_performance_category: chr "asc12_performance_category"
#>  .. .. $ asc12_rshv_rate           : chr "asc12_rshv_rate"
#>  .. .. $ asc12_interval_lower_limit: chr "asc12_interval_lower_limit"
#>  .. .. $ asc12_interval_upper_limit: chr "asc12_interval_upper_limit"
#>  .. .. $ asc12_footnote            : chr "asc12_footnote"
#>  .. .. $ asc13_rate                : chr "asc13_rate"
#>  .. .. $ asc13_footnote            : chr "asc13_footnote"
#>  .. .. $ asc14_rate                : chr "asc14_rate"
#>  .. .. $ asc14_footnote            : chr "asc14_footnote"
#>  .. .. $ asc17_total_cases         : chr "asc17_total_cases"
#>  .. .. $ asc17_performance_category: chr "asc17_performance_category"
#>  .. .. $ asc17_rshv_rate           : chr "asc17_rshv_rate"
#>  .. .. $ asc17_interval_lower_limit: chr "asc17_interval_lower_limit"
#>  .. .. $ asc17_interval_upper_limit: chr "asc17_interval_upper_limit"
#>  .. .. $ asc17_footnote            : chr "asc17_footnote"
#>  .. .. $ asc18_total_cases         : chr "asc18_total_cases"
#>  .. .. $ asc18_performance_category: chr "asc18_performance_category"
#>  .. .. $ asc18_rshv_rate           : chr "asc18_rshv_rate"
#>  .. .. $ asc18_interval_lower_limit: chr "asc18_interval_lower_limit"
#>  .. .. $ asc18_interval_upper_limit: chr "asc18_interval_upper_limit"
#>  .. .. $ asc18_footnote            : chr "asc18_footnote"
#>  .. .. $ asc19_total_cases         : chr "asc19_total_cases"
#>  .. .. $ asc19_performance_category: chr "asc19_performance_category"
#>  .. .. $ asc19_rshv_rate           : chr "asc19_rshv_rate"
#>  .. .. $ asc19_interval_lower_limit: chr "asc19_interval_lower_limit"
#>  .. .. $ asc19_interval_upper_limit: chr "asc19_interval_upper_limit"
#>  .. .. $ asc19_footnote            : chr "asc19_footnote"
#>  .. .. $ asc20_sample              : chr "asc20_sample"
#>  .. .. $ asc20_rate                : chr "asc20_rate"
#>  .. .. $ asc20_footnote            : chr "asc20_footnote"
pro_group("pro_dialysis")
#> <pro_group>
#>  @ group  : chr "Dialysis Facilities"
#>  @ members:List of 3
#>  .. $ dialysis_facility: <pro_endpoint>
#>  ..  ..@ identifier: chr "https://data.cms.gov/provider-data/api/1/datastore/query/23ew-n7w9/0"
#>  ..  ..@ metadata  :List of 9
#>  .. .. .. $ title      : chr "Dialysis Facility - Listing by Facility"
#>  .. .. .. $ group      : chr "Dialysis facilities"
#>  .. .. .. $ description: chr "A list of all dialysis facilities registered with Medicare that includes addresses and phone numbers, as well a"| __truncated__
#>  .. .. .. $ issued     : Date[1:1], format: "2020-03-14"
#>  .. .. .. $ modified   : Date[1:1], format: "2025-03-24"
#>  .. .. .. $ released   : Date[1:1], format: "2025-04-23"
#>  .. .. .. $ site       : chr "https://data.cms.gov/provider-data/dataset/23ew-n7w9"
#>  .. .. .. $ dictionary : chr "https://data.cms.gov/provider-data/dataset/23ew-n7w9#data-dictionary"
#>  .. .. .. $ download   : chr "https://data.cms.gov/provider-data/sites/default/files/resources/c04d84bc5c641284494bee4f20f17f9c_1742846704/DFC_FACILITY.csv"
#>  ..  ..@ dimensions: <class_dimensions>
#>  .. .. .. @ limit : int 2000
#>  .. .. .. @ rows  : int 7556
#>  .. .. .. @ pages : int 4
#>  .. .. .. @ fields:List of 142
#>  .. .. .. .. $ cms_certification_number_ccn                                    : chr "cms_certification_number_ccn"
#>  .. .. .. .. $ network                                                         : chr "network"
#>  .. .. .. .. $ facility_name                                                   : chr "facility_name"
#>  .. .. .. .. $ five_star_date                                                  : chr "five_star_date"
#>  .. .. .. .. $ five_star                                                       : chr "five_star"
#>  .. .. .. .. $ five_star_data_availability_code                                : chr "five_star_data_availability_code"
#>  .. .. .. .. $ address_line_1                                                  : chr "address_line_1"
#>  .. .. .. .. $ address_line_2                                                  : chr "address_line_2"
#>  .. .. .. .. $ citytown                                                        : chr "citytown"
#>  .. .. .. .. $ state                                                           : chr "state"
#>  .. .. .. .. $ zip_code                                                        : chr "zip_code"
#>  .. .. .. .. $ countyparish                                                    : chr "countyparish"
#>  .. .. .. .. $ telephone_number                                                : chr "telephone_number"
#>  .. .. .. .. $ profit_or_nonprofit                                             : chr "profit_or_nonprofit"
#>  .. .. .. .. $ chain_owned                                                     : chr "chain_owned"
#>  .. .. .. .. $ chain_organization                                              : chr "chain_organization"
#>  .. .. .. .. $ late_shift                                                      : chr "late_shift"
#>  .. .. .. .. $ of_dialysis_stations                                            : chr "of_dialysis_stations"
#>  .. .. .. .. $ offers_incenter_hemodialysis                                    : chr "offers_incenter_hemodialysis"
#>  .. .. .. .. $ offers_peritoneal_dialysis                                      : chr "offers_peritoneal_dialysis"
#>  .. .. .. .. $ offers_home_hemodialysis_training                               : chr "offers_home_hemodialysis_training"
#>  .. .. .. .. $ certification_date                                              : chr "certification_date"
#>  .. .. .. .. $ claims_date                                                     : chr "claims_date"
#>  .. .. .. .. $ eqrs_date                                                       : chr "eqrs_date"
#>  .. .. .. .. $ smr_date                                                        : chr "smr_date"
#>  .. .. .. .. $ patient_survival_category_text                                  : chr "patient_survival_category_text"
#>  .. .. .. .. $ patient_survival_data_availability_code                         : chr "patient_survival_data_availability_code"
#>  .. .. .. .. $ number_of_patients_included_in_survival_summary                 : chr "number_of_patients_included_in_survival_summary"
#>  .. .. .. .. $ mortality_rate_facility                                         : chr "mortality_rate_facility"
#>  .. .. .. .. $ mortality_rate_upper_confidence_limit_975                       : chr "mortality_rate_upper_confidence_limit_975"
#>  .. .. .. .. $ mortality_rate_lower_confidence_limit_25                        : chr "mortality_rate_lower_confidence_limit_25"
#>  .. .. .. .. $ shr_date                                                        : chr "shr_date"
#>  .. .. .. .. $ patient_hospitalization_category_text                           : chr "patient_hospitalization_category_text"
#>  .. .. .. .. $ patient_hospitalization_data_availability_code                  : chr "patient_hospitalization_data_availability_code"
#>  .. .. .. .. $ number_of_patients_included_in_hospitalization_summary          : chr "number_of_patients_included_in_hospitalization_summary"
#>  .. .. .. .. $ hospitalization_rate_facility                                   : chr "hospitalization_rate_facility"
#>  .. .. .. .. $ hospitalization_rate_upper_confidence_limit_975                 : chr "hospitalization_rate_upper_confidence_limit_975"
#>  .. .. .. .. $ hospitalization_rate_lower_confidence_limit_25                  : chr "hospitalization_rate_lower_confidence_limit_25"
#>  .. .. .. .. $ srr_date                                                        : chr "srr_date"
#>  .. .. .. .. $ patient_hospital_readmission_category                           : chr "patient_hospital_readmission_category"
#>  .. .. .. .. $ patient_hospital_readmission_data_availability_code             : chr "patient_hospital_readmission_data_availability_code"
#>  .. .. .. .. $ number_of_hospitalizations_included_in_hospital_readmission_fc2b: chr "number_of_hospitalizations_included_in_hospital_readmission_fc2b"
#>  .. .. .. .. $ readmission_rate_facility                                       : chr "readmission_rate_facility"
#>  .. .. .. .. $ readmission_rate_upper_confidence_limit_975                     : chr "readmission_rate_upper_confidence_limit_975"
#>  .. .. .. .. $ readmission_rate_lower_confidence_limit_25                      : chr "readmission_rate_lower_confidence_limit_25"
#>  .. .. .. .. $ strr_date                                                       : chr "strr_date"
#>  .. .. .. .. $ patient_transfusion_category_text                               : chr "patient_transfusion_category_text"
#>  .. .. .. .. $ patient_transfusion_data_availability_code                      : chr "patient_transfusion_data_availability_code"
#>  .. .. .. .. $ number_of_patients_included_in_the_transfusion_summary          : chr "number_of_patients_included_in_the_transfusion_summary"
#>  .. .. .. .. $ transfusion_rate_facility                                       : chr "transfusion_rate_facility"
#>  .. .. .. .. $ transfusion_rate_upper_confidence_limit_975                     : chr "transfusion_rate_upper_confidence_limit_975"
#>  .. .. .. .. $ transfusion_rate_lower_confidence_limit_25                      : chr "transfusion_rate_lower_confidence_limit_25"
#>  .. .. .. .. $ swr_date                                                        : chr "swr_date"
#>  .. .. .. .. $ swr_category_text                                               : chr "swr_category_text"
#>  .. .. .. .. $ patient_transplant_waitlist_data_availability_code              : chr "patient_transplant_waitlist_data_availability_code"
#>  .. .. .. .. $ number_of_patients_in_this_facility_for_swr                     : chr "number_of_patients_in_this_facility_for_swr"
#>  .. .. .. .. $ standardized_first_kidney_transplant_waitlist_ratio             : chr "standardized_first_kidney_transplant_waitlist_ratio"
#>  .. .. .. .. $ 95_ci_upper_limit_for_swr                                       : chr "95_ci_upper_limit_for_swr"
#>  .. .. .. .. $ 95_ci_lower_limit_for_swr                                       : chr "95_ci_lower_limit_for_swr"
#>  .. .. .. .. $ pppw_category_text                                              : chr "pppw_category_text"
#>  .. .. .. .. $ patient_prevalent_transplant_waitlist_data_availability_code    : chr "patient_prevalent_transplant_waitlist_data_availability_code"
#>  .. .. .. .. $ number_of_patients_for_pppw                                     : chr "number_of_patients_for_pppw"
#>  .. .. .. .. $ percentage_of_prevalent_patients_waitlisted                     : chr "percentage_of_prevalent_patients_waitlisted"
#>  .. .. .. .. $ 95_ci_upper_limit_for_pppw                                      : chr "95_ci_upper_limit_for_pppw"
#>  .. .. .. .. $ 95_ci_lower_limit_for_pppw                                      : chr "95_ci_lower_limit_for_pppw"
#>  .. .. .. .. $ sedr_date                                                       : chr "sedr_date"
#>  .. .. .. .. $ sedr_category_text                                              : chr "sedr_category_text"
#>  .. .. .. .. $ emergency_department_encounter_data_availability_code           : chr "emergency_department_encounter_data_availability_code"
#>  .. .. .. .. $ number_of_patients_included_in_sedr_summary                     : chr "number_of_patients_included_in_sedr_summary"
#>  .. .. .. .. $ standardized_ed_visits_ratio_facility                           : chr "standardized_ed_visits_ratio_facility"
#>  .. .. .. .. $ sedr_upper_confidence_limit_975                                 : chr "sedr_upper_confidence_limit_975"
#>  .. .. .. .. $ sedr_lower_confidence_limit_25                                  : chr "sedr_lower_confidence_limit_25"
#>  .. .. .. .. $ ed30_date                                                       : chr "ed30_date"
#>  .. .. .. .. $ ed30_category_text                                              : chr "ed30_category_text"
#>  .. .. .. .. $ emergency_department_encounter_ratio_occurring_within_30_da_f8e3: chr "emergency_department_encounter_ratio_occurring_within_30_da_f8e3"
#>  .. .. .. .. $ number_of_hospitalizations_included_in_ed30_summary             : chr "number_of_hospitalizations_included_in_ed30_summary"
#>  .. .. .. .. $ standardized_ed_visits_within_30_days_of_hospital_discharge_6307: chr "standardized_ed_visits_within_30_days_of_hospital_discharge_6307"
#>  .. .. .. .. $ ed30_upper_confidence_limit_975                                 : chr "ed30_upper_confidence_limit_975"
#>  .. .. .. .. $ ed30_lower_confidence_limit_25                                  : chr "ed30_lower_confidence_limit_25"
#>  .. .. .. .. $ years_modality_switch_based_upon                                : chr "years_modality_switch_based_upon"
#>  .. .. .. .. $ smosr_classification_category_facility                          : chr "smosr_classification_category_facility"
#>  .. .. .. .. $ modality_switch_data_availability_code                          : chr "modality_switch_data_availability_code"
#>  .. .. .. .. $ smosr_number_of_eligible_patients_facility                      : chr "smosr_number_of_eligible_patients_facility"
#>  .. .. .. .. $ smosr_standardized_modality_switch_ratio_facility               : chr "smosr_standardized_modality_switch_ratio_facility"
#>  .. .. .. .. $ smosr_upper_confidence_limit_facility                           : chr "smosr_upper_confidence_limit_facility"
#>  .. .. .. .. $ smosr_lower_confidence_limit_facility                           : chr "smosr_lower_confidence_limit_facility"
#>  .. .. .. .. $ sir_date                                                        : chr "sir_date"
#>  .. .. .. .. $ patient_infection_category_text                                 : chr "patient_infection_category_text"
#>  .. .. .. .. $ patient_infection_data_availability_code                        : chr "patient_infection_data_availability_code"
#>  .. .. .. .. $ standard_infection_ratio                                        : chr "standard_infection_ratio"
#>  .. .. .. .. $ sir_upper_confidence_limit_975                                  : chr "sir_upper_confidence_limit_975"
#>  .. .. .. .. $ sir_lower_confidence_limit_25                                   : chr "sir_lower_confidence_limit_25"
#>  .. .. .. .. $ fistula_category_text                                           : chr "fistula_category_text"
#>  .. .. .. .. $ fistula_data_availability_code                                  : chr "fistula_data_availability_code"
#>  .. .. .. .. $ number_of_patients_included_in_fistula_summary                  : chr "number_of_patients_included_in_fistula_summary"
#>  .. .. .. .. $ fistula_rate_facility                                           : chr "fistula_rate_facility"
#>  .. .. .. .. $ fistula_rate_upper_confidence_limit_975                         : chr "fistula_rate_upper_confidence_limit_975"
#>  .. .. .. .. $ fistula_rate_lower_confidence_limit_25                          : chr "fistula_rate_lower_confidence_limit_25"
#>  .. .. .. .. $ hcp_vaccination_data_collection_dates                           : chr "hcp_vaccination_data_collection_dates"
#>  .. .. .. ..  [list output truncated]
#>  .. $ dialysis_national: <pro_endpoint>
#>  ..  ..@ identifier: chr "https://data.cms.gov/provider-data/api/1/datastore/query/2rkq-ygai/0"
#>  ..  ..@ metadata  :List of 9
#>  .. .. .. $ title      : chr "Dialysis Facility - National Averages"
#>  .. .. .. $ group      : chr "Dialysis facilities"
#>  .. .. .. $ description: chr "National averages of common dialysis quality measures."
#>  .. .. .. $ issued     : Date[1:1], format: "2020-03-14"
#>  .. .. .. $ modified   : Date[1:1], format: "2025-03-24"
#>  .. .. .. $ released   : Date[1:1], format: "2025-04-23"
#>  .. .. .. $ site       : chr "https://data.cms.gov/provider-data/dataset/2rkq-ygai"
#>  .. .. .. $ dictionary : chr "https://data.cms.gov/provider-data/dataset/2rkq-ygai#data-dictionary"
#>  .. .. .. $ download   : chr "https://data.cms.gov/provider-data/sites/default/files/resources/ae019f56e76839ee1da1b51f502954cc_1742846706/DFC_NATIONAL.csv"
#>  ..  ..@ dimensions: <class_dimensions>
#>  .. .. .. @ limit : int 2000
#>  .. .. .. @ rows  : int 1
#>  .. .. .. @ pages : int 1
#>  .. .. .. @ fields:List of 59
#>  .. .. .. .. $ country                                                         : chr "country"
#>  .. .. .. .. $ transfusion_rate_us                                             : chr "transfusion_rate_us"
#>  .. .. .. .. $ transfusions__better_than_expected_us                           : chr "transfusions__better_than_expected_us"
#>  .. .. .. .. $ transfusions__as_expected_us                                    : chr "transfusions__as_expected_us"
#>  .. .. .. .. $ transfusions__worse_than_expected_us                            : chr "transfusions__worse_than_expected_us"
#>  .. .. .. .. $ percentage_of_adult_hd_patients_with_ktv12                      : chr "percentage_of_adult_hd_patients_with_ktv12"
#>  .. .. .. .. $ percentage_of_adult_pd_patients_with_ktv17                      : chr "percentage_of_adult_pd_patients_with_ktv17"
#>  .. .. .. .. $ percentage_of_pediatric_hd_patients_with_ktv12                  : chr "percentage_of_pediatric_hd_patients_with_ktv12"
#>  .. .. .. .. $ percentage_of_pediatric_pd_patients_with_ktv18                  : chr "percentage_of_pediatric_pd_patients_with_ktv18"
#>  .. .. .. .. $ percentage_of_pediatric_hd_patients_with_npcr                   : chr "percentage_of_pediatric_hd_patients_with_npcr"
#>  .. .. .. .. $ fistula_rate_us                                                 : chr "fistula_rate_us"
#>  .. .. .. .. $ fistula_rate__better_than_expected_us                           : chr "fistula_rate__better_than_expected_us"
#>  .. .. .. .. $ fistula_rate__as_expected_us                                    : chr "fistula_rate__as_expected_us"
#>  .. .. .. .. $ fistula_rate__worse_than_expected_us                            : chr "fistula_rate__worse_than_expected_us"
#>  .. .. .. .. $ percentage_of_adult_patients_with_long_term_catheter_in_use     : chr "percentage_of_adult_patients_with_long_term_catheter_in_use"
#>  .. .. .. .. $ percentage_of_adult_patients_with_hypercalcemia_serum_calci_044d: chr "percentage_of_adult_patients_with_hypercalcemia_serum_calci_044d"
#>  .. .. .. .. $ percentage_of_adult_patients_with_serum_phosphorus_less_tha_c222: chr "percentage_of_adult_patients_with_serum_phosphorus_less_tha_c222"
#>  .. .. .. .. $ percentage_of_adult_patients_with_serum_phosphorus_between__85e8: chr "percentage_of_adult_patients_with_serum_phosphorus_between__85e8"
#>  .. .. .. .. $ percentage_of_adult_patients_with_serum_phosphorus_between__fad7: chr "percentage_of_adult_patients_with_serum_phosphorus_between__fad7"
#>  .. .. .. .. $ percentage_of_adult_patients_with_serum_phosphorus_between__ff32: chr "percentage_of_adult_patients_with_serum_phosphorus_between__ff32"
#>  .. .. .. .. $ percentage_of_adult_patients_with_serum_phosphorus_greater__d8e3: chr "percentage_of_adult_patients_with_serum_phosphorus_greater__d8e3"
#>  .. .. .. .. $ hospitalization_rate_us                                         : chr "hospitalization_rate_us"
#>  .. .. .. .. $ hospitalizations__better_than_expected_us                       : chr "hospitalizations__better_than_expected_us"
#>  .. .. .. .. $ hospitalizations__as_expected_us                                : chr "hospitalizations__as_expected_us"
#>  .. .. .. .. $ hospitalizations__worse_than_expected_us                        : chr "hospitalizations__worse_than_expected_us"
#>  .. .. .. .. $ readmission_rate_us                                             : chr "readmission_rate_us"
#>  .. .. .. .. $ hospital_readmission__better_than_expected_us                   : chr "hospital_readmission__better_than_expected_us"
#>  .. .. .. .. $ hospital_readmission__as_expected_us                            : chr "hospital_readmission__as_expected_us"
#>  .. .. .. .. $ hospital__readmission__worse_than_expected_us                   : chr "hospital__readmission__worse_than_expected_us"
#>  .. .. .. .. $ mortality_rate_us                                               : chr "mortality_rate_us"
#>  .. .. .. .. $ survival__better_than_expected_us                               : chr "survival__better_than_expected_us"
#>  .. .. .. .. $ survival__as_expected_us                                        : chr "survival__as_expected_us"
#>  .. .. .. .. $ survival__worse_than_expected_us                                : chr "survival__worse_than_expected_us"
#>  .. .. .. .. $ standardized_first_kidney_transplant_waitlist_ratio_us          : chr "standardized_first_kidney_transplant_waitlist_ratio_us"
#>  .. .. .. .. $ incident_patients_transplant_waitlisting__better_than_expec_a24d: chr "incident_patients_transplant_waitlisting__better_than_expec_a24d"
#>  .. .. .. .. $ incident_patients_transplant_waitlisting__as_expected_us        : chr "incident_patients_transplant_waitlisting__as_expected_us"
#>  .. .. .. .. $ incident_patients_transplant_waitlisting__worse_than_expected_us: chr "incident_patients_transplant_waitlisting__worse_than_expected_us"
#>  .. .. .. .. $ percentage_of_prevalent_patients_waitlisted_us                  : chr "percentage_of_prevalent_patients_waitlisted_us"
#>  .. .. .. .. $ prevalent_patients_transplant_waitlisting_better_than_expec_c0cd: chr "prevalent_patients_transplant_waitlisting_better_than_expec_c0cd"
#>  .. .. .. .. $ prevalent_patients_transplant_waitlisting__as_expected_us       : chr "prevalent_patients_transplant_waitlisting__as_expected_us"
#>  .. .. .. .. $ prevalent_patients_transplant_waitlisting__worse_than_expec_e6e9: chr "prevalent_patients_transplant_waitlisting__worse_than_expec_e6e9"
#>  .. .. .. .. $ healthcare_worker_covid19_vaccination_adherence_percentage_us   : chr "healthcare_worker_covid19_vaccination_adherence_percentage_us"
#>  .. .. .. .. $ percentage_of_patients_with_hgb10_gdl_us                        : chr "percentage_of_patients_with_hgb10_gdl_us"
#>  .. .. .. .. $ percentage_of_patients_with_hgb12_gdl_us                        : chr "percentage_of_patients_with_hgb12_gdl_us"
#>  .. .. .. .. $ infection__better_than_expected_us                              : chr "infection__better_than_expected_us"
#>  .. .. .. .. $ infection__as_expected_us                                       : chr "infection__as_expected_us"
#>  .. .. .. .. $ infection__worse_than_expected_us                               : chr "infection__worse_than_expected_us"
#>  .. .. .. .. $ standardized_ed_visits_ratio_us                                 : chr "standardized_ed_visits_ratio_us"
#>  .. .. .. .. $ sedr__better_than_expected_us                                   : chr "sedr__better_than_expected_us"
#>  .. .. .. .. $ sedr__as_expected_us                                            : chr "sedr__as_expected_us"
#>  .. .. .. .. $ sedr__worse_than_expected_us                                    : chr "sedr__worse_than_expected_us"
#>  .. .. .. .. $ standardized_ed_visits_within_30_days_of_hospital_discharge_d0c6: chr "standardized_ed_visits_within_30_days_of_hospital_discharge_d0c6"
#>  .. .. .. .. $ ed30__better_than_expected_us                                   : chr "ed30__better_than_expected_us"
#>  .. .. .. .. $ ed30__as_expected_us                                            : chr "ed30__as_expected_us"
#>  .. .. .. .. $ ed30__worse_than_expected_us                                    : chr "ed30__worse_than_expected_us"
#>  .. .. .. .. $ standardized_modality_switch_ratio                              : chr "standardized_modality_switch_ratio"
#>  .. .. .. .. $ n_for_smsr_category_1_smsrbetter_than_expected_us               : chr "n_for_smsr_category_1_smsrbetter_than_expected_us"
#>  .. .. .. .. $ n_for_smsr_category_2_smsras_expected_us                        : chr "n_for_smsr_category_2_smsras_expected_us"
#>  .. .. .. .. $ n_for_smsr_category_3_smsrworse_than_expected_us                : chr "n_for_smsr_category_3_smsrworse_than_expected_us"
#>  .. $ dialysis_state   : <pro_endpoint>
#>  ..  ..@ identifier: chr "https://data.cms.gov/provider-data/api/1/datastore/query/2fpu-cgbb/0"
#>  ..  ..@ metadata  :List of 9
#>  .. .. .. $ title      : chr "Dialysis Facility - State Averages"
#>  .. .. .. $ group      : chr "Dialysis facilities"
#>  .. .. .. $ description: chr "State averages of common dialysis quality measures."
#>  .. .. .. $ issued     : Date[1:1], format: "2020-03-14"
#>  .. .. .. $ modified   : Date[1:1], format: "2025-03-24"
#>  .. .. .. $ released   : Date[1:1], format: "2025-04-23"
#>  .. .. .. $ site       : chr "https://data.cms.gov/provider-data/dataset/2fpu-cgbb"
#>  .. .. .. $ dictionary : chr "https://data.cms.gov/provider-data/dataset/2fpu-cgbb#data-dictionary"
#>  .. .. .. $ download   : chr "https://data.cms.gov/provider-data/sites/default/files/resources/3b5f0449f531ee07c467e49abed89954_1742846705/DFC_STATE.csv"
#>  ..  ..@ dimensions: <class_dimensions>
#>  .. .. .. @ limit : int 2000
#>  .. .. .. @ rows  : int 56
#>  .. .. .. @ pages : int 1
#>  .. .. .. @ fields:List of 49
#>  .. .. .. .. $ state                                                           : chr "state"
#>  .. .. .. .. $ survival_better_than_expected_state                             : chr "survival_better_than_expected_state"
#>  .. .. .. .. $ survival_as_expected_state                                      : chr "survival_as_expected_state"
#>  .. .. .. .. $ survival_worse_than_expected_state                              : chr "survival_worse_than_expected_state"
#>  .. .. .. .. $ hospitalizations_better_than_expected_state                     : chr "hospitalizations_better_than_expected_state"
#>  .. .. .. .. $ hospitalizations_as_expected_state                              : chr "hospitalizations_as_expected_state"
#>  .. .. .. .. $ hospitalizations_worse_than_expected_state                      : chr "hospitalizations_worse_than_expected_state"
#>  .. .. .. .. $ hospital_readmission__better_than_expected_state                : chr "hospital_readmission__better_than_expected_state"
#>  .. .. .. .. $ hospital_readmission__as_expected_state                         : chr "hospital_readmission__as_expected_state"
#>  .. .. .. .. $ hospital__readmission__worse_than_expected_state                : chr "hospital__readmission__worse_than_expected_state"
#>  .. .. .. .. $ transfusions_better_than_expected_state                         : chr "transfusions_better_than_expected_state"
#>  .. .. .. .. $ transfusions_as_expected_state                                  : chr "transfusions_as_expected_state"
#>  .. .. .. .. $ transfusions_worse_than_expected_state                          : chr "transfusions_worse_than_expected_state"
#>  .. .. .. .. $ infection_better_than_expected_state                            : chr "infection_better_than_expected_state"
#>  .. .. .. .. $ infection_as_expected_state                                     : chr "infection_as_expected_state"
#>  .. .. .. .. $ infection_worse_than_expected_state                             : chr "infection_worse_than_expected_state"
#>  .. .. .. .. $ percentage_of_adult_hd_patients_with_ktv12                      : chr "percentage_of_adult_hd_patients_with_ktv12"
#>  .. .. .. .. $ percentage_of_adult_pd_patients_with_ktv17                      : chr "percentage_of_adult_pd_patients_with_ktv17"
#>  .. .. .. .. $ percentage_of_pediatric_hd_patients_with_ktv12                  : chr "percentage_of_pediatric_hd_patients_with_ktv12"
#>  .. .. .. .. $ percentage_of_pediatric_pd_patients_with_ktv18                  : chr "percentage_of_pediatric_pd_patients_with_ktv18"
#>  .. .. .. .. $ percentage_of_pediatric_hd_patients_with_npcr_in_use            : chr "percentage_of_pediatric_hd_patients_with_npcr_in_use"
#>  .. .. .. .. $ fistula_rate__better_than_expected_state                        : chr "fistula_rate__better_than_expected_state"
#>  .. .. .. .. $ fistula_rate__as_expected_state                                 : chr "fistula_rate__as_expected_state"
#>  .. .. .. .. $ fistula_rate__worse_than_expected_state                         : chr "fistula_rate__worse_than_expected_state"
#>  .. .. .. .. $ percentage_of_adult_patients_with_long_term_catheter_in_use     : chr "percentage_of_adult_patients_with_long_term_catheter_in_use"
#>  .. .. .. .. $ percentage_of_adult_patients_with_hypercalcemia_serum_calci_044d: chr "percentage_of_adult_patients_with_hypercalcemia_serum_calci_044d"
#>  .. .. .. .. $ percentage_of_adult_patients_with_serum_phosphorus_less_tha_c222: chr "percentage_of_adult_patients_with_serum_phosphorus_less_tha_c222"
#>  .. .. .. .. $ percentage_of_adult_patients_with_serum_phosphorus_between__85e8: chr "percentage_of_adult_patients_with_serum_phosphorus_between__85e8"
#>  .. .. .. .. $ percentage_of_adult_patients_with_serum_phosphorus_between__fad7: chr "percentage_of_adult_patients_with_serum_phosphorus_between__fad7"
#>  .. .. .. .. $ percentage_of_adult_patients_with_serum_phosphorus_between__ff32: chr "percentage_of_adult_patients_with_serum_phosphorus_between__ff32"
#>  .. .. .. .. $ percentage_of_adult_patients_with_serum_phosphorus_greater__d8e3: chr "percentage_of_adult_patients_with_serum_phosphorus_greater__d8e3"
#>  .. .. .. .. $ incident_patients_transplant_waitlisting_better_than_expect_49e4: chr "incident_patients_transplant_waitlisting_better_than_expect_49e4"
#>  .. .. .. .. $ incident_patients_transplant_waitlisting__as_expected_state     : chr "incident_patients_transplant_waitlisting__as_expected_state"
#>  .. .. .. .. $ incident_patients_transplant_waitlisting__worse_than_expect_9b13: chr "incident_patients_transplant_waitlisting__worse_than_expect_9b13"
#>  .. .. .. .. $ prevalent_patients_transplant_waitlisting_better_than_expec_c046: chr "prevalent_patients_transplant_waitlisting_better_than_expec_c046"
#>  .. .. .. .. $ prevalent_patients_transplant_waitlisting__as_expected_state    : chr "prevalent_patients_transplant_waitlisting__as_expected_state"
#>  .. .. .. .. $ prevalent_patients_transplant_waitlisting__worse_than_expec_eeb0: chr "prevalent_patients_transplant_waitlisting__worse_than_expec_eeb0"
#>  .. .. .. .. $ healthcare_worker_covid19_vaccination_adherence_percentage_state: chr "healthcare_worker_covid19_vaccination_adherence_percentage_state"
#>  .. .. .. .. $ percentage_of_patients_with_hgb10_gdl_state                     : chr "percentage_of_patients_with_hgb10_gdl_state"
#>  .. .. .. .. $ percentage_of_patients_with_hgb12_gdl_state                     : chr "percentage_of_patients_with_hgb12_gdl_state"
#>  .. .. .. .. $ sedr__better_than_expected_state                                : chr "sedr__better_than_expected_state"
#>  .. .. .. .. $ sedr__as_expected_state                                         : chr "sedr__as_expected_state"
#>  .. .. .. .. $ sedr__worse_than_expected_state                                 : chr "sedr__worse_than_expected_state"
#>  .. .. .. .. $ ed30__better_than_expected_state                                : chr "ed30__better_than_expected_state"
#>  .. .. .. .. $ ed30__as_expected_state                                         : chr "ed30__as_expected_state"
#>  .. .. .. .. $ ed30__worse_than_expected_state                                 : chr "ed30__worse_than_expected_state"
#>  .. .. .. .. $ n_for_smsr_category_1_smsrbetter_than_expected_state            : chr "n_for_smsr_category_1_smsrbetter_than_expected_state"
#>  .. .. .. .. $ n_for_smsr_category_2_smsras_expected_state                     : chr "n_for_smsr_category_2_smsras_expected_state"
#>  .. .. .. .. $ n_for_smsr_category_3_smsrworse_than_expected_state             : chr "n_for_smsr_category_3_smsrworse_than_expected_state"