Provider API Endpoint Classes

Provider API Endpoint Classes

Usage

pro_endpoint(alias)

pro_group(alias)

Arguments

alias

<chr> endpoint alias

Value

An S7 <pro_endpoint> or<pro_group> object

Examples

pro_endpoint("asc_facility")
#> <pro_endpoint>
#>  @ identifier: chr "https://data.cms.gov/provider-data/api/1/datastore/query/4jcv-atw7/0"
#>  @ metadata  :List of 9
#>  .. $ title      : chr "Ambulatory Surgical Center Quality Measures - Facility"
#>  .. $ description: chr "A list of ambulatory surgical centers participating in the Ambulatory Surgical Center Quality Reporting (ASCQR)"| __truncated__
#>  .. $ modified   : Date[1:1], format: "2025-04-08"
#>  .. $ group      : chr "Hospitals"
#>  .. $ issued     : Date[1:1], format: "2024-10-10"
#>  .. $ released   : Date[1:1], format: "2025-04-30"
#>  .. $ download   : chr "https://data.cms.gov/provider-data/sites/default/files/resources/dd03994fc93e296bb0297f1cd43cc987_1744668308/ASC_Facility.csv"
#>  .. $ dictionary : chr "https://data.cms.gov/provider-data/dataset/4jcv-atw7#data-dictionary"
#>  .. $ site       : chr "https://data.cms.gov/provider-data/dataset/4jcv-atw7"
#>  @ dimensions: <class_dimensions>
#>  .. @ limit : int 2000
#>  .. @ rows  : int 11288
#>  .. @ pages : int 6
#>  .. @ fields:List of 50
#>  .. .. $ facility_name             : chr(0) 
#>  .. .. $ facility_id               : chr(0) 
#>  .. .. $ npi                       : chr(0) 
#>  .. .. $ citytown                  : chr(0) 
#>  .. .. $ state                     : chr(0) 
#>  .. .. $ zip_code                  : chr(0) 
#>  .. .. $ year                      : chr(0) 
#>  .. .. $ asc1_rate                 : chr(0) 
#>  .. .. $ asc1_footnote             : chr(0) 
#>  .. .. $ asc2_rate                 : chr(0) 
#>  .. .. $ asc2_footnote             : chr(0) 
#>  .. .. $ asc3_rate                 : chr(0) 
#>  .. .. $ asc3_footnote             : chr(0) 
#>  .. .. $ asc4_rate                 : chr(0) 
#>  .. .. $ asc4_footnote             : chr(0) 
#>  .. .. $ asc9_rate                 : chr(0) 
#>  .. .. $ asc9_footnote             : chr(0) 
#>  .. .. $ asc11_rate                : chr(0) 
#>  .. .. $ asc11_footnote            : chr(0) 
#>  .. .. $ asc12_total_cases         : chr(0) 
#>  .. .. $ asc12_performance_category: chr(0) 
#>  .. .. $ asc12_rshv_rate           : chr(0) 
#>  .. .. $ asc12_interval_lower_limit: chr(0) 
#>  .. .. $ asc12_interval_upper_limit: chr(0) 
#>  .. .. $ asc12_footnote            : chr(0) 
#>  .. .. $ asc13_rate                : chr(0) 
#>  .. .. $ asc13_footnote            : chr(0) 
#>  .. .. $ asc14_rate                : chr(0) 
#>  .. .. $ asc14_footnote            : chr(0) 
#>  .. .. $ asc17_total_cases         : chr(0) 
#>  .. .. $ asc17_performance_category: chr(0) 
#>  .. .. $ asc17_rshv_rate           : chr(0) 
#>  .. .. $ asc17_interval_lower_limit: chr(0) 
#>  .. .. $ asc17_interval_upper_limit: chr(0) 
#>  .. .. $ asc17_footnote            : chr(0) 
#>  .. .. $ asc18_total_cases         : chr(0) 
#>  .. .. $ asc18_performance_category: chr(0) 
#>  .. .. $ asc18_rshv_rate           : chr(0) 
#>  .. .. $ asc18_interval_lower_limit: chr(0) 
#>  .. .. $ asc18_interval_upper_limit: chr(0) 
#>  .. .. $ asc18_footnote            : chr(0) 
#>  .. .. $ asc19_total_cases         : chr(0) 
#>  .. .. $ asc19_performance_category: chr(0) 
#>  .. .. $ asc19_rshv_rate           : chr(0) 
#>  .. .. $ asc19_interval_lower_limit: chr(0) 
#>  .. .. $ asc19_interval_upper_limit: chr(0) 
#>  .. .. $ asc19_footnote            : chr(0) 
#>  .. .. $ asc20_sample              : chr(0) 
#>  .. .. $ asc20_rate                : chr(0) 
#>  .. .. $ asc20_footnote            : chr(0) 
pro_group("pro_dialysis")
#> <pro_group>
#>  @ group  : chr "Dialysis Facilities"
#>  @ members:List of 3
#>  .. $ dialysis_facility: <pro_endpoint>
#>  ..  ..@ identifier: chr "https://data.cms.gov/provider-data/api/1/datastore/query/23ew-n7w9/0"
#>  ..  ..@ metadata  :List of 9
#>  .. .. .. $ title      : chr "Dialysis Facility - Listing by Facility"
#>  .. .. .. $ description: chr "A list of all dialysis facilities registered with Medicare that includes addresses and phone numbers, as well a"| __truncated__
#>  .. .. .. $ modified   : Date[1:1], format: "2025-03-24"
#>  .. .. .. $ group      : chr "Dialysis facilities"
#>  .. .. .. $ issued     : Date[1:1], format: "2020-03-14"
#>  .. .. .. $ released   : Date[1:1], format: "2025-04-23"
#>  .. .. .. $ download   : chr "https://data.cms.gov/provider-data/sites/default/files/resources/c04d84bc5c641284494bee4f20f17f9c_1742846704/DFC_FACILITY.csv"
#>  .. .. .. $ dictionary : chr "https://data.cms.gov/provider-data/dataset/23ew-n7w9#data-dictionary"
#>  .. .. .. $ site       : chr "https://data.cms.gov/provider-data/dataset/23ew-n7w9"
#>  ..  ..@ dimensions: <class_dimensions>
#>  .. .. .. @ limit : int 2000
#>  .. .. .. @ rows  : int 7556
#>  .. .. .. @ pages : int 4
#>  .. .. .. @ fields:List of 142
#>  .. .. .. .. $ cms_certification_number_ccn                                    : chr(0) 
#>  .. .. .. .. $ network                                                         : chr(0) 
#>  .. .. .. .. $ facility_name                                                   : chr(0) 
#>  .. .. .. .. $ five_star_date                                                  : chr(0) 
#>  .. .. .. .. $ five_star                                                       : chr(0) 
#>  .. .. .. .. $ five_star_data_availability_code                                : chr(0) 
#>  .. .. .. .. $ address_line_1                                                  : chr(0) 
#>  .. .. .. .. $ address_line_2                                                  : chr(0) 
#>  .. .. .. .. $ citytown                                                        : chr(0) 
#>  .. .. .. .. $ state                                                           : chr(0) 
#>  .. .. .. .. $ zip_code                                                        : chr(0) 
#>  .. .. .. .. $ countyparish                                                    : chr(0) 
#>  .. .. .. .. $ telephone_number                                                : chr(0) 
#>  .. .. .. .. $ profit_or_nonprofit                                             : chr(0) 
#>  .. .. .. .. $ chain_owned                                                     : chr(0) 
#>  .. .. .. .. $ chain_organization                                              : chr(0) 
#>  .. .. .. .. $ late_shift                                                      : chr(0) 
#>  .. .. .. .. $ of_dialysis_stations                                            : chr(0) 
#>  .. .. .. .. $ offers_incenter_hemodialysis                                    : chr(0) 
#>  .. .. .. .. $ offers_peritoneal_dialysis                                      : chr(0) 
#>  .. .. .. .. $ offers_home_hemodialysis_training                               : chr(0) 
#>  .. .. .. .. $ certification_date                                              : chr(0) 
#>  .. .. .. .. $ claims_date                                                     : chr(0) 
#>  .. .. .. .. $ eqrs_date                                                       : chr(0) 
#>  .. .. .. .. $ smr_date                                                        : chr(0) 
#>  .. .. .. .. $ patient_survival_category_text                                  : chr(0) 
#>  .. .. .. .. $ patient_survival_data_availability_code                         : chr(0) 
#>  .. .. .. .. $ number_of_patients_included_in_survival_summary                 : chr(0) 
#>  .. .. .. .. $ mortality_rate_facility                                         : chr(0) 
#>  .. .. .. .. $ mortality_rate_upper_confidence_limit_975                       : chr(0) 
#>  .. .. .. .. $ mortality_rate_lower_confidence_limit_25                        : chr(0) 
#>  .. .. .. .. $ shr_date                                                        : chr(0) 
#>  .. .. .. .. $ patient_hospitalization_category_text                           : chr(0) 
#>  .. .. .. .. $ patient_hospitalization_data_availability_code                  : chr(0) 
#>  .. .. .. .. $ number_of_patients_included_in_hospitalization_summary          : chr(0) 
#>  .. .. .. .. $ hospitalization_rate_facility                                   : chr(0) 
#>  .. .. .. .. $ hospitalization_rate_upper_confidence_limit_975                 : chr(0) 
#>  .. .. .. .. $ hospitalization_rate_lower_confidence_limit_25                  : chr(0) 
#>  .. .. .. .. $ srr_date                                                        : chr(0) 
#>  .. .. .. .. $ patient_hospital_readmission_category                           : chr(0) 
#>  .. .. .. .. $ patient_hospital_readmission_data_availability_code             : chr(0) 
#>  .. .. .. .. $ number_of_hospitalizations_included_in_hospital_readmission_fc2b: chr(0) 
#>  .. .. .. .. $ readmission_rate_facility                                       : chr(0) 
#>  .. .. .. .. $ readmission_rate_upper_confidence_limit_975                     : chr(0) 
#>  .. .. .. .. $ readmission_rate_lower_confidence_limit_25                      : chr(0) 
#>  .. .. .. .. $ strr_date                                                       : chr(0) 
#>  .. .. .. .. $ patient_transfusion_category_text                               : chr(0) 
#>  .. .. .. .. $ patient_transfusion_data_availability_code                      : chr(0) 
#>  .. .. .. .. $ number_of_patients_included_in_the_transfusion_summary          : chr(0) 
#>  .. .. .. .. $ transfusion_rate_facility                                       : chr(0) 
#>  .. .. .. .. $ transfusion_rate_upper_confidence_limit_975                     : chr(0) 
#>  .. .. .. .. $ transfusion_rate_lower_confidence_limit_25                      : chr(0) 
#>  .. .. .. .. $ swr_date                                                        : chr(0) 
#>  .. .. .. .. $ swr_category_text                                               : chr(0) 
#>  .. .. .. .. $ patient_transplant_waitlist_data_availability_code              : chr(0) 
#>  .. .. .. .. $ number_of_patients_in_this_facility_for_swr                     : chr(0) 
#>  .. .. .. .. $ standardized_first_kidney_transplant_waitlist_ratio             : chr(0) 
#>  .. .. .. .. $ 95_ci_upper_limit_for_swr                                       : chr(0) 
#>  .. .. .. .. $ 95_ci_lower_limit_for_swr                                       : chr(0) 
#>  .. .. .. .. $ pppw_category_text                                              : chr(0) 
#>  .. .. .. .. $ patient_prevalent_transplant_waitlist_data_availability_code    : chr(0) 
#>  .. .. .. .. $ number_of_patients_for_pppw                                     : chr(0) 
#>  .. .. .. .. $ percentage_of_prevalent_patients_waitlisted                     : chr(0) 
#>  .. .. .. .. $ 95_ci_upper_limit_for_pppw                                      : chr(0) 
#>  .. .. .. .. $ 95_ci_lower_limit_for_pppw                                      : chr(0) 
#>  .. .. .. .. $ sedr_date                                                       : chr(0) 
#>  .. .. .. .. $ sedr_category_text                                              : chr(0) 
#>  .. .. .. .. $ emergency_department_encounter_data_availability_code           : chr(0) 
#>  .. .. .. .. $ number_of_patients_included_in_sedr_summary                     : chr(0) 
#>  .. .. .. .. $ standardized_ed_visits_ratio_facility                           : chr(0) 
#>  .. .. .. .. $ sedr_upper_confidence_limit_975                                 : chr(0) 
#>  .. .. .. .. $ sedr_lower_confidence_limit_25                                  : chr(0) 
#>  .. .. .. .. $ ed30_date                                                       : chr(0) 
#>  .. .. .. .. $ ed30_category_text                                              : chr(0) 
#>  .. .. .. .. $ emergency_department_encounter_ratio_occurring_within_30_da_f8e3: chr(0) 
#>  .. .. .. .. $ number_of_hospitalizations_included_in_ed30_summary             : chr(0) 
#>  .. .. .. .. $ standardized_ed_visits_within_30_days_of_hospital_discharge_6307: chr(0) 
#>  .. .. .. .. $ ed30_upper_confidence_limit_975                                 : chr(0) 
#>  .. .. .. .. $ ed30_lower_confidence_limit_25                                  : chr(0) 
#>  .. .. .. .. $ years_modality_switch_based_upon                                : chr(0) 
#>  .. .. .. .. $ smosr_classification_category_facility                          : chr(0) 
#>  .. .. .. .. $ modality_switch_data_availability_code                          : chr(0) 
#>  .. .. .. .. $ smosr_number_of_eligible_patients_facility                      : chr(0) 
#>  .. .. .. .. $ smosr_standardized_modality_switch_ratio_facility               : chr(0) 
#>  .. .. .. .. $ smosr_upper_confidence_limit_facility                           : chr(0) 
#>  .. .. .. .. $ smosr_lower_confidence_limit_facility                           : chr(0) 
#>  .. .. .. .. $ sir_date                                                        : chr(0) 
#>  .. .. .. .. $ patient_infection_category_text                                 : chr(0) 
#>  .. .. .. .. $ patient_infection_data_availability_code                        : chr(0) 
#>  .. .. .. .. $ standard_infection_ratio                                        : chr(0) 
#>  .. .. .. .. $ sir_upper_confidence_limit_975                                  : chr(0) 
#>  .. .. .. .. $ sir_lower_confidence_limit_25                                   : chr(0) 
#>  .. .. .. .. $ fistula_category_text                                           : chr(0) 
#>  .. .. .. .. $ fistula_data_availability_code                                  : chr(0) 
#>  .. .. .. .. $ number_of_patients_included_in_fistula_summary                  : chr(0) 
#>  .. .. .. .. $ fistula_rate_facility                                           : chr(0) 
#>  .. .. .. .. $ fistula_rate_upper_confidence_limit_975                         : chr(0) 
#>  .. .. .. .. $ fistula_rate_lower_confidence_limit_25                          : chr(0) 
#>  .. .. .. .. $ hcp_vaccination_data_collection_dates                           : chr(0) 
#>  .. .. .. ..  [list output truncated]
#>  .. $ dialysis_national: <pro_endpoint>
#>  ..  ..@ identifier: chr "https://data.cms.gov/provider-data/api/1/datastore/query/2rkq-ygai/0"
#>  ..  ..@ metadata  :List of 9
#>  .. .. .. $ title      : chr "Dialysis Facility - National Averages"
#>  .. .. .. $ description: chr "National averages of common dialysis quality measures."
#>  .. .. .. $ modified   : Date[1:1], format: "2025-03-24"
#>  .. .. .. $ group      : chr "Dialysis facilities"
#>  .. .. .. $ issued     : Date[1:1], format: "2020-03-14"
#>  .. .. .. $ released   : Date[1:1], format: "2025-04-23"
#>  .. .. .. $ download   : chr "https://data.cms.gov/provider-data/sites/default/files/resources/ae019f56e76839ee1da1b51f502954cc_1742846706/DFC_NATIONAL.csv"
#>  .. .. .. $ dictionary : chr "https://data.cms.gov/provider-data/dataset/2rkq-ygai#data-dictionary"
#>  .. .. .. $ site       : chr "https://data.cms.gov/provider-data/dataset/2rkq-ygai"
#>  ..  ..@ dimensions: <class_dimensions>
#>  .. .. .. @ limit : int 2000
#>  .. .. .. @ rows  : int 1
#>  .. .. .. @ pages : int 1
#>  .. .. .. @ fields:List of 59
#>  .. .. .. .. $ country                                                         : chr(0) 
#>  .. .. .. .. $ transfusion_rate_us                                             : chr(0) 
#>  .. .. .. .. $ transfusions__better_than_expected_us                           : chr(0) 
#>  .. .. .. .. $ transfusions__as_expected_us                                    : chr(0) 
#>  .. .. .. .. $ transfusions__worse_than_expected_us                            : chr(0) 
#>  .. .. .. .. $ percentage_of_adult_hd_patients_with_ktv12                      : chr(0) 
#>  .. .. .. .. $ percentage_of_adult_pd_patients_with_ktv17                      : chr(0) 
#>  .. .. .. .. $ percentage_of_pediatric_hd_patients_with_ktv12                  : chr(0) 
#>  .. .. .. .. $ percentage_of_pediatric_pd_patients_with_ktv18                  : chr(0) 
#>  .. .. .. .. $ percentage_of_pediatric_hd_patients_with_npcr                   : chr(0) 
#>  .. .. .. .. $ fistula_rate_us                                                 : chr(0) 
#>  .. .. .. .. $ fistula_rate__better_than_expected_us                           : chr(0) 
#>  .. .. .. .. $ fistula_rate__as_expected_us                                    : chr(0) 
#>  .. .. .. .. $ fistula_rate__worse_than_expected_us                            : chr(0) 
#>  .. .. .. .. $ percentage_of_adult_patients_with_long_term_catheter_in_use     : chr(0) 
#>  .. .. .. .. $ percentage_of_adult_patients_with_hypercalcemia_serum_calci_044d: chr(0) 
#>  .. .. .. .. $ percentage_of_adult_patients_with_serum_phosphorus_less_tha_c222: chr(0) 
#>  .. .. .. .. $ percentage_of_adult_patients_with_serum_phosphorus_between__85e8: chr(0) 
#>  .. .. .. .. $ percentage_of_adult_patients_with_serum_phosphorus_between__fad7: chr(0) 
#>  .. .. .. .. $ percentage_of_adult_patients_with_serum_phosphorus_between__ff32: chr(0) 
#>  .. .. .. .. $ percentage_of_adult_patients_with_serum_phosphorus_greater__d8e3: chr(0) 
#>  .. .. .. .. $ hospitalization_rate_us                                         : chr(0) 
#>  .. .. .. .. $ hospitalizations__better_than_expected_us                       : chr(0) 
#>  .. .. .. .. $ hospitalizations__as_expected_us                                : chr(0) 
#>  .. .. .. .. $ hospitalizations__worse_than_expected_us                        : chr(0) 
#>  .. .. .. .. $ readmission_rate_us                                             : chr(0) 
#>  .. .. .. .. $ hospital_readmission__better_than_expected_us                   : chr(0) 
#>  .. .. .. .. $ hospital_readmission__as_expected_us                            : chr(0) 
#>  .. .. .. .. $ hospital__readmission__worse_than_expected_us                   : chr(0) 
#>  .. .. .. .. $ mortality_rate_us                                               : chr(0) 
#>  .. .. .. .. $ survival__better_than_expected_us                               : chr(0) 
#>  .. .. .. .. $ survival__as_expected_us                                        : chr(0) 
#>  .. .. .. .. $ survival__worse_than_expected_us                                : chr(0) 
#>  .. .. .. .. $ standardized_first_kidney_transplant_waitlist_ratio_us          : chr(0) 
#>  .. .. .. .. $ incident_patients_transplant_waitlisting__better_than_expec_a24d: chr(0) 
#>  .. .. .. .. $ incident_patients_transplant_waitlisting__as_expected_us        : chr(0) 
#>  .. .. .. .. $ incident_patients_transplant_waitlisting__worse_than_expected_us: chr(0) 
#>  .. .. .. .. $ percentage_of_prevalent_patients_waitlisted_us                  : chr(0) 
#>  .. .. .. .. $ prevalent_patients_transplant_waitlisting_better_than_expec_c0cd: chr(0) 
#>  .. .. .. .. $ prevalent_patients_transplant_waitlisting__as_expected_us       : chr(0) 
#>  .. .. .. .. $ prevalent_patients_transplant_waitlisting__worse_than_expec_e6e9: chr(0) 
#>  .. .. .. .. $ healthcare_worker_covid19_vaccination_adherence_percentage_us   : chr(0) 
#>  .. .. .. .. $ percentage_of_patients_with_hgb10_gdl_us                        : chr(0) 
#>  .. .. .. .. $ percentage_of_patients_with_hgb12_gdl_us                        : chr(0) 
#>  .. .. .. .. $ infection__better_than_expected_us                              : chr(0) 
#>  .. .. .. .. $ infection__as_expected_us                                       : chr(0) 
#>  .. .. .. .. $ infection__worse_than_expected_us                               : chr(0) 
#>  .. .. .. .. $ standardized_ed_visits_ratio_us                                 : chr(0) 
#>  .. .. .. .. $ sedr__better_than_expected_us                                   : chr(0) 
#>  .. .. .. .. $ sedr__as_expected_us                                            : chr(0) 
#>  .. .. .. .. $ sedr__worse_than_expected_us                                    : chr(0) 
#>  .. .. .. .. $ standardized_ed_visits_within_30_days_of_hospital_discharge_d0c6: chr(0) 
#>  .. .. .. .. $ ed30__better_than_expected_us                                   : chr(0) 
#>  .. .. .. .. $ ed30__as_expected_us                                            : chr(0) 
#>  .. .. .. .. $ ed30__worse_than_expected_us                                    : chr(0) 
#>  .. .. .. .. $ standardized_modality_switch_ratio                              : chr(0) 
#>  .. .. .. .. $ n_for_smsr_category_1_smsrbetter_than_expected_us               : chr(0) 
#>  .. .. .. .. $ n_for_smsr_category_2_smsras_expected_us                        : chr(0) 
#>  .. .. .. .. $ n_for_smsr_category_3_smsrworse_than_expected_us                : chr(0) 
#>  .. $ dialysis_state   : <pro_endpoint>
#>  ..  ..@ identifier: chr "https://data.cms.gov/provider-data/api/1/datastore/query/2fpu-cgbb/0"
#>  ..  ..@ metadata  :List of 9
#>  .. .. .. $ title      : chr "Dialysis Facility - State Averages"
#>  .. .. .. $ description: chr "State averages of common dialysis quality measures."
#>  .. .. .. $ modified   : Date[1:1], format: "2025-03-24"
#>  .. .. .. $ group      : chr "Dialysis facilities"
#>  .. .. .. $ issued     : Date[1:1], format: "2020-03-14"
#>  .. .. .. $ released   : Date[1:1], format: "2025-04-23"
#>  .. .. .. $ download   : chr "https://data.cms.gov/provider-data/sites/default/files/resources/3b5f0449f531ee07c467e49abed89954_1742846705/DFC_STATE.csv"
#>  .. .. .. $ dictionary : chr "https://data.cms.gov/provider-data/dataset/2fpu-cgbb#data-dictionary"
#>  .. .. .. $ site       : chr "https://data.cms.gov/provider-data/dataset/2fpu-cgbb"
#>  ..  ..@ dimensions: <class_dimensions>
#>  .. .. .. @ limit : int 2000
#>  .. .. .. @ rows  : int 56
#>  .. .. .. @ pages : int 1
#>  .. .. .. @ fields:List of 49
#>  .. .. .. .. $ state                                                           : chr(0) 
#>  .. .. .. .. $ survival_better_than_expected_state                             : chr(0) 
#>  .. .. .. .. $ survival_as_expected_state                                      : chr(0) 
#>  .. .. .. .. $ survival_worse_than_expected_state                              : chr(0) 
#>  .. .. .. .. $ hospitalizations_better_than_expected_state                     : chr(0) 
#>  .. .. .. .. $ hospitalizations_as_expected_state                              : chr(0) 
#>  .. .. .. .. $ hospitalizations_worse_than_expected_state                      : chr(0) 
#>  .. .. .. .. $ hospital_readmission__better_than_expected_state                : chr(0) 
#>  .. .. .. .. $ hospital_readmission__as_expected_state                         : chr(0) 
#>  .. .. .. .. $ hospital__readmission__worse_than_expected_state                : chr(0) 
#>  .. .. .. .. $ transfusions_better_than_expected_state                         : chr(0) 
#>  .. .. .. .. $ transfusions_as_expected_state                                  : chr(0) 
#>  .. .. .. .. $ transfusions_worse_than_expected_state                          : chr(0) 
#>  .. .. .. .. $ infection_better_than_expected_state                            : chr(0) 
#>  .. .. .. .. $ infection_as_expected_state                                     : chr(0) 
#>  .. .. .. .. $ infection_worse_than_expected_state                             : chr(0) 
#>  .. .. .. .. $ percentage_of_adult_hd_patients_with_ktv12                      : chr(0) 
#>  .. .. .. .. $ percentage_of_adult_pd_patients_with_ktv17                      : chr(0) 
#>  .. .. .. .. $ percentage_of_pediatric_hd_patients_with_ktv12                  : chr(0) 
#>  .. .. .. .. $ percentage_of_pediatric_pd_patients_with_ktv18                  : chr(0) 
#>  .. .. .. .. $ percentage_of_pediatric_hd_patients_with_npcr_in_use            : chr(0) 
#>  .. .. .. .. $ fistula_rate__better_than_expected_state                        : chr(0) 
#>  .. .. .. .. $ fistula_rate__as_expected_state                                 : chr(0) 
#>  .. .. .. .. $ fistula_rate__worse_than_expected_state                         : chr(0) 
#>  .. .. .. .. $ percentage_of_adult_patients_with_long_term_catheter_in_use     : chr(0) 
#>  .. .. .. .. $ percentage_of_adult_patients_with_hypercalcemia_serum_calci_044d: chr(0) 
#>  .. .. .. .. $ percentage_of_adult_patients_with_serum_phosphorus_less_tha_c222: chr(0) 
#>  .. .. .. .. $ percentage_of_adult_patients_with_serum_phosphorus_between__85e8: chr(0) 
#>  .. .. .. .. $ percentage_of_adult_patients_with_serum_phosphorus_between__fad7: chr(0) 
#>  .. .. .. .. $ percentage_of_adult_patients_with_serum_phosphorus_between__ff32: chr(0) 
#>  .. .. .. .. $ percentage_of_adult_patients_with_serum_phosphorus_greater__d8e3: chr(0) 
#>  .. .. .. .. $ incident_patients_transplant_waitlisting_better_than_expect_49e4: chr(0) 
#>  .. .. .. .. $ incident_patients_transplant_waitlisting__as_expected_state     : chr(0) 
#>  .. .. .. .. $ incident_patients_transplant_waitlisting__worse_than_expect_9b13: chr(0) 
#>  .. .. .. .. $ prevalent_patients_transplant_waitlisting_better_than_expec_c046: chr(0) 
#>  .. .. .. .. $ prevalent_patients_transplant_waitlisting__as_expected_state    : chr(0) 
#>  .. .. .. .. $ prevalent_patients_transplant_waitlisting__worse_than_expec_eeb0: chr(0) 
#>  .. .. .. .. $ healthcare_worker_covid19_vaccination_adherence_percentage_state: chr(0) 
#>  .. .. .. .. $ percentage_of_patients_with_hgb10_gdl_state                     : chr(0) 
#>  .. .. .. .. $ percentage_of_patients_with_hgb12_gdl_state                     : chr(0) 
#>  .. .. .. .. $ sedr__better_than_expected_state                                : chr(0) 
#>  .. .. .. .. $ sedr__as_expected_state                                         : chr(0) 
#>  .. .. .. .. $ sedr__worse_than_expected_state                                 : chr(0) 
#>  .. .. .. .. $ ed30__better_than_expected_state                                : chr(0) 
#>  .. .. .. .. $ ed30__as_expected_state                                         : chr(0) 
#>  .. .. .. .. $ ed30__worse_than_expected_state                                 : chr(0) 
#>  .. .. .. .. $ n_for_smsr_category_1_smsrbetter_than_expected_state            : chr(0) 
#>  .. .. .. .. $ n_for_smsr_category_2_smsras_expected_state                     : chr(0) 
#>  .. .. .. .. $ n_for_smsr_category_3_smsrworse_than_expected_state             : chr(0)