Provider API Endpoint Classes

Provider API Endpoint Classes

Usage

pro_endpoint(alias)

pro_group(alias)

Arguments

alias

<chr> endpoint alias

Value

An S7 <pro_endpoint> or<pro_group> object

Examples

pro_endpoint("PSI90_6digit")
#> <pro_endpoint>
#>  @ identifier: chr "https://data.cms.gov/provider-data/api/1/datastore/query/muwa-iene/0"
#>  @ metadata  :List of 8
#>  .. $ title      : chr "CMS Medicare PSI-90 and component measures - six-digit estimate dataset"
#>  .. $ description: chr "This data set includes the Patient Safety and Adverse Events Composite measure (CMS Medicare PSI 90) and the in"| __truncated__
#>  .. $ issued     : Date[1:1], format: "2020-12-10"
#>  .. $ modified   : Date[1:1], format: "2025-04-08"
#>  .. $ released   : Date[1:1], format: "2025-04-30"
#>  .. $ site       : chr "https://data.cms.gov/provider-data/dataset/muwa-iene"
#>  .. $ dictionary : chr "https://data.cms.gov/provider-data/dataset/muwa-iene#data-dictionary"
#>  .. $ download   : chr "https://data.cms.gov/provider-data/sites/default/files/resources/84bd78c1b4e386185bcef2af963d8cf9_1744668313/CM"| __truncated__
#>  @ dimensions: <class_dimensions>
#>  .. @ limit : int 2000
#>  .. @ rows  : int 51975
#>  .. @ pages : int 26
#>  .. @ fields:List of 13
#>  .. .. $ facility_id  : chr "facility_id"
#>  .. .. $ facility_name: chr "facility_name"
#>  .. .. $ address      : chr "address"
#>  .. .. $ citytown     : chr "citytown"
#>  .. .. $ state        : chr "state"
#>  .. .. $ zip_code     : chr "zip_code"
#>  .. .. $ countyparish : chr "countyparish"
#>  .. .. $ measure_id   : chr "measure_id"
#>  .. .. $ measure_name : chr "measure_name"
#>  .. .. $ rate         : chr "rate"
#>  .. .. $ footnote     : chr "footnote"
#>  .. .. $ start_date   : chr "start_date"
#>  .. .. $ end_date     : chr "end_date"
pro_endpoint("joint_replace")
#> <pro_endpoint>
#>  @ identifier: chr "https://data.cms.gov/provider-data/api/1/datastore/query/tqkv-mgxq/0"
#>  @ metadata  :List of 8
#>  .. $ title      : chr "Comprehensive Care For Joint Replacement Model - Provider Data"
#>  .. $ description: chr "Comprehensive Care for Joint Replacement Model - provider data. This data set includes provider data for two qu"| __truncated__
#>  .. $ issued     : Date[1:1], format: "2020-12-10"
#>  .. $ modified   : Date[1:1], format: "2025-01-08"
#>  .. $ released   : Date[1:1], format: "2025-04-30"
#>  .. $ site       : chr "https://data.cms.gov/provider-data/dataset/tqkv-mgxq"
#>  .. $ dictionary : chr "https://data.cms.gov/provider-data/dataset/tqkv-mgxq#data-dictionary"
#>  .. $ download   : chr "https://data.cms.gov/provider-data/sites/default/files/resources/26896abd613a88df1d7d9637e9f85f19_1736791511/CJ"| __truncated__
#>  @ dimensions: <class_dimensions>
#>  .. @ limit : int 2000
#>  .. @ rows  : int 399
#>  .. @ pages : int 1
#>  .. @ fields:List of 18
#>  .. .. $ facility_id            : chr "facility_id"
#>  .. .. $ facility_name          : chr "facility_name"
#>  .. .. $ msa                    : chr "msa"
#>  .. .. $ msa_title              : chr "msa_title"
#>  .. .. $ hcahps_hlmr            : chr "hcahps_hlmr"
#>  .. .. $ hcahps_hlmr_percentile : chr "hcahps_hlmr_percentile"
#>  .. .. $ hcahps_start_date      : chr "hcahps_start_date"
#>  .. .. $ hcahps_end_date        : chr "hcahps_end_date"
#>  .. .. $ hcahps_footnote        : chr "hcahps_footnote"
#>  .. .. $ comphipknee            : chr "comphipknee"
#>  .. .. $ comphipknee_percentile : chr "comphipknee_percentile"
#>  .. .. $ comp_start_date        : chr "comp_start_date"
#>  .. .. $ comp_end_date          : chr "comp_end_date"
#>  .. .. $ comp_footnote          : chr "comp_footnote"
#>  .. .. $ pro                    : chr "pro"
#>  .. .. $ pro_start_date         : chr "pro_start_date"
#>  .. .. $ pro_end_date           : chr "pro_end_date"
#>  .. .. $ reconciliation_footnote: chr "reconciliation_footnote"
pro_endpoint("footnote_xwalk")
#> <pro_endpoint>
#>  @ identifier: chr "https://data.cms.gov/provider-data/api/1/datastore/query/y9us-9xdf/0"
#>  @ metadata  :List of 8
#>  .. $ title      : chr "Footnote Crosswalk"
#>  .. $ description: chr "The footnote crosswalk displays a list of footnotes and their summary."
#>  .. $ issued     : Date[1:1], format: "2020-12-10"
#>  .. $ modified   : Date[1:1], format: "2025-01-13"
#>  .. $ released   : Date[1:1], format: "2025-04-30"
#>  .. $ site       : chr "https://data.cms.gov/provider-data/dataset/y9us-9xdf"
#>  .. $ dictionary : chr "https://data.cms.gov/provider-data/dataset/y9us-9xdf#data-dictionary"
#>  .. $ download   : chr "https://data.cms.gov/provider-data/sites/default/files/resources/f29bb7c812e242f6edfef0a4b7d0eaca_1736805905/Fo"| __truncated__
#>  @ dimensions: <class_dimensions>
#>  .. @ limit : int 2000
#>  .. @ rows  : int 33
#>  .. @ pages : int 1
#>  .. @ fields:List of 2
#>  .. .. $ footnote     : chr "footnote"
#>  .. .. $ footnote_text: chr "footnote_text"
pro_group("PDC")
#> <pro_group>
#>  @ group  : chr "Provider Data Catalog"
#>  @ members:List of 3
#>  .. $ PDC_affiliations: <pro_endpoint>
#>  ..  ..@ identifier: chr "https://data.cms.gov/provider-data/api/1/datastore/query/27ea-46a8/0"
#>  ..  ..@ metadata  :List of 8
#>  .. .. .. $ title      : chr "Facility Affiliation Data"
#>  .. .. .. $ description: chr "This is the facility affiliations data publicly reported in the Provider Data Catalog."
#>  .. .. .. $ issued     : Date[1:1], format: "2023-08-17"
#>  .. .. .. $ modified   : Date[1:1], format: "2025-04-18"
#>  .. .. .. $ released   : Date[1:1], format: "2025-05-01"
#>  .. .. .. $ site       : chr "https://data.cms.gov/provider-data/dataset/27ea-46a8"
#>  .. .. .. $ dictionary : chr "https://data.cms.gov/provider-data/dataset/27ea-46a8#data-dictionary"
#>  .. .. .. $ download   : chr "https://data.cms.gov/provider-data/sites/default/files/resources/b7c4080ae144663e43353a9c35cd3f53_1744999531/Fa"| __truncated__
#>  ..  ..@ dimensions: <class_dimensions>
#>  .. .. .. @ limit : int 2000
#>  .. .. .. @ rows  : int 1557102
#>  .. .. .. @ pages : int 779
#>  .. .. .. @ fields:List of 9
#>  .. .. .. .. $ npi                                       : chr "npi"
#>  .. .. .. .. $ ind_pac_id                                : chr "ind_pac_id"
#>  .. .. .. .. $ provider_last_name                        : chr "provider_last_name"
#>  .. .. .. .. $ provider_first_name                       : chr "provider_first_name"
#>  .. .. .. .. $ provider_middle_name                      : chr "provider_middle_name"
#>  .. .. .. .. $ suff                                      : chr "suff"
#>  .. .. .. .. $ facility_type                             : chr "facility_type"
#>  .. .. .. .. $ facility_affiliations_certification_number: chr "facility_affiliations_certification_number"
#>  .. .. .. .. $ facility_type_certification_number        : chr "facility_type_certification_number"
#>  .. $ PDC_clinicians  : <pro_endpoint>
#>  ..  ..@ identifier: chr "https://data.cms.gov/provider-data/api/1/datastore/query/mj5m-pzi6/0"
#>  ..  ..@ metadata  :List of 8
#>  .. .. .. $ title      : chr "National Downloadable File"
#>  .. .. .. $ description: chr "The Doctors and Clinicians national downloadable file is organized such that each line is unique at the clinici"| __truncated__
#>  .. .. .. $ issued     : Date[1:1], format: "2023-08-17"
#>  .. .. .. $ modified   : Date[1:1], format: "2025-04-18"
#>  .. .. .. $ released   : Date[1:1], format: "2025-05-01"
#>  .. .. .. $ site       : chr "https://data.cms.gov/provider-data/dataset/mj5m-pzi6"
#>  .. .. .. $ dictionary : chr "https://data.cms.gov/provider-data/dataset/mj5m-pzi6#data-dictionary"
#>  .. .. .. $ download   : chr "https://data.cms.gov/provider-data/sites/default/files/resources/52c3f098d7e56028a298fd297cb0b38d_1744999530/DA"| __truncated__
#>  ..  ..@ dimensions: <class_dimensions>
#>  .. .. .. @ limit : int 2000
#>  .. .. .. @ rows  : int 2640783
#>  .. .. .. @ pages : int 1321
#>  .. .. .. @ fields:List of 31
#>  .. .. .. .. $ npi                 : chr "npi"
#>  .. .. .. .. $ ind_pac_id          : chr "ind_pac_id"
#>  .. .. .. .. $ ind_enrl_id         : chr "ind_enrl_id"
#>  .. .. .. .. $ provider_last_name  : chr "provider_last_name"
#>  .. .. .. .. $ provider_first_name : chr "provider_first_name"
#>  .. .. .. .. $ provider_middle_name: chr "provider_middle_name"
#>  .. .. .. .. $ suff                : chr "suff"
#>  .. .. .. .. $ gndr                : chr "gndr"
#>  .. .. .. .. $ cred                : chr "cred"
#>  .. .. .. .. $ med_sch             : chr "med_sch"
#>  .. .. .. .. $ grd_yr              : chr "grd_yr"
#>  .. .. .. .. $ pri_spec            : chr "pri_spec"
#>  .. .. .. .. $ sec_spec_1          : chr "sec_spec_1"
#>  .. .. .. .. $ sec_spec_2          : chr "sec_spec_2"
#>  .. .. .. .. $ sec_spec_3          : chr "sec_spec_3"
#>  .. .. .. .. $ sec_spec_4          : chr "sec_spec_4"
#>  .. .. .. .. $ sec_spec_all        : chr "sec_spec_all"
#>  .. .. .. .. $ telehlth            : chr "telehlth"
#>  .. .. .. .. $ facility_name       : chr "facility_name"
#>  .. .. .. .. $ org_pac_id          : chr "org_pac_id"
#>  .. .. .. .. $ num_org_mem         : chr "num_org_mem"
#>  .. .. .. .. $ adr_ln_1            : chr "adr_ln_1"
#>  .. .. .. .. $ adr_ln_2            : chr "adr_ln_2"
#>  .. .. .. .. $ ln_2_sprs           : chr "ln_2_sprs"
#>  .. .. .. .. $ citytown            : chr "citytown"
#>  .. .. .. .. $ state               : chr "state"
#>  .. .. .. .. $ zip_code            : chr "zip_code"
#>  .. .. .. .. $ telephone_number    : chr "telephone_number"
#>  .. .. .. .. $ ind_assgn           : chr "ind_assgn"
#>  .. .. .. .. $ grp_assgn           : chr "grp_assgn"
#>  .. .. .. .. $ adrs_id             : chr "adrs_id"
#>  .. $ PDC_utilization : <pro_endpoint>
#>  ..  ..@ identifier: chr "https://data.cms.gov/provider-data/api/1/datastore/query/n0yb-util/0"
#>  ..  ..@ metadata  :List of 8
#>  .. .. .. $ title      : chr "Utilization Data"
#>  .. .. .. $ description: chr "The Doctors and Clinicians utilization data file reports volume information for procedures of interest on clini"| __truncated__
#>  .. .. .. $ issued     : Date[1:1], format: "2024-01-04"
#>  .. .. .. $ modified   : Date[1:1], format: "2025-03-24"
#>  .. .. .. $ released   : Date[1:1], format: "2025-04-03"
#>  .. .. .. $ site       : chr "https://data.cms.gov/provider-data/dataset/n0yb-util"
#>  .. .. .. $ dictionary : chr "https://data.cms.gov/provider-data/dataset/n0yb-util#data-dictionary"
#>  .. .. .. $ download   : chr "https://data.cms.gov/provider-data/sites/default/files/resources/72eedb23fcbe1ef3666f89db01b85e1d_1742861112/Utilization.csv"
#>  ..  ..@ dimensions: <class_dimensions>
#>  .. .. .. @ limit : int 2000
#>  .. .. .. @ rows  : int 301894
#>  .. .. .. @ pages : int 151
#>  .. .. .. @ fields:List of 10
#>  .. .. .. .. $ npi                      : chr "npi"
#>  .. .. .. .. $ ind_pac_id               : chr "ind_pac_id"
#>  .. .. .. .. $ provider_last_name       : chr "provider_last_name"
#>  .. .. .. .. $ provider_first_name      : chr "provider_first_name"
#>  .. .. .. .. $ provider_middle_name     : chr "provider_middle_name"
#>  .. .. .. .. $ suff                     : chr "suff"
#>  .. .. .. .. $ procedure_category       : chr "procedure_category"
#>  .. .. .. .. $ count                    : chr "count"
#>  .. .. .. .. $ percentile               : chr "percentile"
#>  .. .. .. .. $ profile_display_indicator: chr "profile_display_indicator"
pro_group("MIPS")
#> <pro_group>
#>  @ group  : chr "PY 2022 MIPS Public Reporting"
#>  @ members:List of 5
#>  .. $ MIPS_performance: <pro_endpoint>
#>  ..  ..@ identifier: chr "https://data.cms.gov/provider-data/api/1/datastore/query/a174-a962/0"
#>  ..  ..@ metadata  :List of 8
#>  .. .. .. $ title      : chr "PY 2022 Clinician Public Reporting: Overall MIPS Performance"
#>  .. .. .. $ description: chr "This file contains Merit-Based Incentive Payment System (MIPS) Final Scores and performance category scores for"| __truncated__
#>  .. .. .. $ issued     : Date[1:1], format: "2023-08-17"
#>  .. .. .. $ modified   : Date[1:1], format: "2024-07-25"
#>  .. .. .. $ released   : Date[1:1], format: "2024-08-01"
#>  .. .. .. $ site       : chr "https://data.cms.gov/provider-data/dataset/a174-a962"
#>  .. .. .. $ dictionary : chr "https://data.cms.gov/provider-data/dataset/a174-a962#data-dictionary"
#>  .. .. .. $ download   : chr "https://data.cms.gov/provider-data/sites/default/files/resources/6b9e57db797c95853b034b329b1212b2_1721952317/ec_score_file.csv"
#>  ..  ..@ dimensions: <class_dimensions>
#>  .. .. .. @ limit : int 2000
#>  .. .. .. @ rows  : int 624209
#>  .. .. .. @ pages : int 313
#>  .. .. .. @ fields:List of 13
#>  .. .. .. .. $ npi                                       : chr "npi"
#>  .. .. .. .. $ org_pac_id                                : chr "org_pac_id"
#>  .. .. .. .. $ provider_last_name                        : chr "provider_last_name"
#>  .. .. .. .. $ provider_first_name                       : chr "provider_first_name"
#>  .. .. .. .. $ source                                    : chr "source"
#>  .. .. .. .. $ facilitybased_scoring_certification_number: chr "facilitybased_scoring_certification_number"
#>  .. .. .. .. $ facility_name                             : chr "facility_name"
#>  .. .. .. .. $ quality_category_score                    : chr "quality_category_score"
#>  .. .. .. .. $ pi_category_score                         : chr "pi_category_score"
#>  .. .. .. .. $ ia_category_score                         : chr "ia_category_score"
#>  .. .. .. .. $ cost_category_score                       : chr "cost_category_score"
#>  .. .. .. .. $ final_mips_score_without_cpb              : chr "final_mips_score_without_cpb"
#>  .. .. .. .. $ final_mips_score                          : chr "final_mips_score"
#>  .. $ MIPS_patient    : <pro_endpoint>
#>  ..  ..@ identifier: chr "https://data.cms.gov/provider-data/api/1/datastore/query/8c70-d353/0"
#>  ..  ..@ metadata  :List of 8
#>  .. .. .. $ title      : chr "PY 2022 Group Public Reporting: Patient Experience"
#>  .. .. .. $ description: chr "This file contains information for the Consumer Assessment of Healthcare Providers and Systems (CAHPS) for MIPS"| __truncated__
#>  .. .. .. $ issued     : Date[1:1], format: "2023-08-17"
#>  .. .. .. $ modified   : Date[1:1], format: "2024-05-16"
#>  .. .. .. $ released   : Date[1:1], format: "2024-05-30"
#>  .. .. .. $ site       : chr "https://data.cms.gov/provider-data/dataset/8c70-d353"
#>  .. .. .. $ dictionary : chr "https://data.cms.gov/provider-data/dataset/8c70-d353#data-dictionary"
#>  .. .. .. $ download   : chr "https://data.cms.gov/provider-data/sites/default/files/resources/6083bb351293e72eda42881d6a93d803_1715976323/gr"| __truncated__
#>  ..  ..@ dimensions: <class_dimensions>
#>  .. .. .. @ limit : int 2000
#>  .. .. .. @ rows  : int 432
#>  .. .. .. @ pages : int 1
#>  .. .. .. @ fields:List of 7
#>  .. .. .. .. $ facility_name: chr "facility_name"
#>  .. .. .. .. $ org_pac_id   : chr "org_pac_id"
#>  .. .. .. .. $ measure_cd   : chr "measure_cd"
#>  .. .. .. .. $ measure_title: chr "measure_title"
#>  .. .. .. .. $ prf_rate     : chr "prf_rate"
#>  .. .. .. .. $ patient_count: chr "patient_count"
#>  .. .. .. .. $ fn           : chr "fn"
#>  .. $ MIPS_clinician  : <pro_endpoint>
#>  ..  ..@ identifier: chr "https://data.cms.gov/provider-data/api/1/datastore/query/7d6a-e7a6/0"
#>  ..  ..@ metadata  :List of 8
#>  .. .. .. $ title      : chr "PY 2022 Clinician Public Reporting: MIPS Measures and Attestations"
#>  .. .. .. $ description: chr "This file contains performance information for Merit-Based Incentive Payment System (MIPS) Quality, Promoting I"| __truncated__
#>  .. .. .. $ issued     : Date[1:1], format: "2020-11-30"
#>  .. .. .. $ modified   : Date[1:1], format: "2024-05-16"
#>  .. .. .. $ released   : Date[1:1], format: "2024-05-30"
#>  .. .. .. $ site       : chr "https://data.cms.gov/provider-data/dataset/7d6a-e7a6"
#>  .. .. .. $ dictionary : chr "https://data.cms.gov/provider-data/dataset/7d6a-e7a6#data-dictionary"
#>  .. .. .. $ download   : chr "https://data.cms.gov/provider-data/sites/default/files/resources/b9c4d902c52ae898d3f422fffbdc59fe_1715976323/ec"| __truncated__
#>  ..  ..@ dimensions: <class_dimensions>
#>  .. .. .. @ limit : int 2000
#>  .. .. .. @ rows  : int 523282
#>  .. .. .. @ pages : int 262
#>  .. .. .. @ fields:List of 18
#>  .. .. .. .. $ npi                : chr "npi"
#>  .. .. .. .. $ ind_pac_id         : chr "ind_pac_id"
#>  .. .. .. .. $ provider_last_name : chr "provider_last_name"
#>  .. .. .. .. $ provider_first_name: chr "provider_first_name"
#>  .. .. .. .. $ apm_affl_1         : chr "apm_affl_1"
#>  .. .. .. .. $ apm_affl_2         : chr "apm_affl_2"
#>  .. .. .. .. $ apm_affl_3         : chr "apm_affl_3"
#>  .. .. .. .. $ apm_affl_4         : chr "apm_affl_4"
#>  .. .. .. .. $ measure_cd         : chr "measure_cd"
#>  .. .. .. .. $ measure_title      : chr "measure_title"
#>  .. .. .. .. $ invs_msr           : chr "invs_msr"
#>  .. .. .. .. $ attestation_value  : chr "attestation_value"
#>  .. .. .. .. $ prf_rate           : chr "prf_rate"
#>  .. .. .. .. $ patient_count      : chr "patient_count"
#>  .. .. .. .. $ star_value         : chr "star_value"
#>  .. .. .. .. $ five_star_benchmark: chr "five_star_benchmark"
#>  .. .. .. .. $ collection_type    : chr "collection_type"
#>  .. .. .. .. $ ccxp_ind           : chr "ccxp_ind"
#>  .. $ MIPS_group      : <pro_endpoint>
#>  ..  ..@ identifier: chr "https://data.cms.gov/provider-data/api/1/datastore/query/0ba7-2cb0/0"
#>  ..  ..@ metadata  :List of 8
#>  .. .. .. $ title      : chr "PY 2022 Group Public Reporting: MIPS Measures and Attestations"
#>  .. .. .. $ description: chr "This file contains performance information for Merit-Based Incentive Payment System (MIPS) Quality, Promoting I"| __truncated__
#>  .. .. .. $ issued     : Date[1:1], format: "2023-08-17"
#>  .. .. .. $ modified   : Date[1:1], format: "2024-05-16"
#>  .. .. .. $ released   : Date[1:1], format: "2024-05-30"
#>  .. .. .. $ site       : chr "https://data.cms.gov/provider-data/dataset/0ba7-2cb0"
#>  .. .. .. $ dictionary : chr "https://data.cms.gov/provider-data/dataset/0ba7-2cb0#data-dictionary"
#>  .. .. .. $ download   : chr "https://data.cms.gov/provider-data/sites/default/files/resources/410d73868de5a4ea585a61ca1365a256_1715976322/gr"| __truncated__
#>  ..  ..@ dimensions: <class_dimensions>
#>  .. .. .. @ limit : int 2000
#>  .. .. .. @ rows  : int 193488
#>  .. .. .. @ pages : int 97
#>  .. .. .. @ fields:List of 16
#>  .. .. .. .. $ facility_name      : chr "facility_name"
#>  .. .. .. .. $ org_pac_id         : chr "org_pac_id"
#>  .. .. .. .. $ aco_id_1           : chr "aco_id_1"
#>  .. .. .. .. $ aco_nm_1           : chr "aco_nm_1"
#>  .. .. .. .. $ aco_id_2           : chr "aco_id_2"
#>  .. .. .. .. $ aco_nm_2           : chr "aco_nm_2"
#>  .. .. .. .. $ measure_cd         : chr "measure_cd"
#>  .. .. .. .. $ measure_title      : chr "measure_title"
#>  .. .. .. .. $ invs_msr           : chr "invs_msr"
#>  .. .. .. .. $ attestation_value  : chr "attestation_value"
#>  .. .. .. .. $ prf_rate           : chr "prf_rate"
#>  .. .. .. .. $ patient_count      : chr "patient_count"
#>  .. .. .. .. $ star_value         : chr "star_value"
#>  .. .. .. .. $ five_star_benchmark: chr "five_star_benchmark"
#>  .. .. .. .. $ collection_type    : chr "collection_type"
#>  .. .. .. .. $ ccxp_ind           : chr "ccxp_ind"
#>  .. $ MIPS_virtual    : <pro_endpoint>
#>  ..  ..@ identifier: chr "https://data.cms.gov/provider-data/api/1/datastore/query/46dc-a66c/0"
#>  ..  ..@ metadata  :List of 8
#>  .. .. .. $ title      : chr "PY 2022 Virtual Group Public Reporting: MIPS Measures and Attestations"
#>  .. .. .. $ description: chr "This file contains performance information for Merit-Based Incentive Payment System (MIPS) Quality, Promoting I"| __truncated__
#>  .. .. .. $ issued     : Date[1:1], format: "2021-06-04"
#>  .. .. .. $ modified   : Date[1:1], format: "2024-05-16"
#>  .. .. .. $ released   : Date[1:1], format: "2024-05-30"
#>  .. .. .. $ site       : chr "https://data.cms.gov/provider-data/dataset/46dc-a66c"
#>  .. .. .. $ dictionary : chr "https://data.cms.gov/provider-data/dataset/46dc-a66c#data-dictionary"
#>  .. .. .. $ download   : chr "https://data.cms.gov/provider-data/sites/default/files/resources/635d18fe1428158bf5b033cd27da6cc3_1715976333/vg"| __truncated__
#>  ..  ..@ dimensions: <class_dimensions>
#>  .. .. .. @ limit : int 2000
#>  .. .. .. @ rows  : int 57
#>  .. .. .. @ pages : int 1
#>  .. .. .. @ fields:List of 8
#>  .. .. .. .. $ virtual_group_id : chr "virtual_group_id"
#>  .. .. .. .. $ measure_cd       : chr "measure_cd"
#>  .. .. .. .. $ measure_title    : chr "measure_title"
#>  .. .. .. .. $ invs_msr         : chr "invs_msr"
#>  .. .. .. .. $ attestation_value: chr "attestation_value"
#>  .. .. .. .. $ prf_rate         : chr "prf_rate"
#>  .. .. .. .. $ patient_count    : chr "patient_count"
#>  .. .. .. .. $ collection_type  : chr "collection_type"