Public Provider Enrollment Dataset Object

Public Provider Enrollment Dataset Object

Usage

public_Dataset(dataset)

Arguments

dataset

<chr> dataset title

Value

<S7_class> object

Examples

public_Dataset(dataset = "Public Provider Enrollment")
#> <providertwo::Dataset>
#>  @ type              : chr "dcat:Dataset"
#>  @ accessLevel       : chr "public"
#>  @ accrualPeriodicity: chr "Quarterly [R/P3M]"
#>  @ bureauCode        : chr "009:38"
#>  @ contactPoint      : <providertwo::dataset_contactPoint>
#>  .. @ type    : chr "vcard:Contact"
#>  .. @ fn      : chr "Provider Enrollment Data Requests - CPI"
#>  .. @ hasEmail: chr "mailto:ProviderEnrollmentDataRequests@cms.hhs.gov"
#>  .. @ hasURL  : NULL
#>  @ describedBy       : chr "https://data.cms.gov/resources/medicare-fee-for-service-public-provider-enrollment-data-dictionary"
#>  @ description       : chr "The Medicare Fee-For-Service Public Provider Enrollment dataset includes information on providers who are activ"| __truncated__
#>  @ identifier        : <providertwo::public_Identifier>
#>  .. @ url    : chr "https://data.cms.gov/data-api/v1/dataset/2457ea29-fc82-48b0-86ec-3b0755de7515/data-viewer"
#>  .. @ request:List of 7
#>  .. .. $ url     : chr "https://data.cms.gov/data-api/v1/dataset/2457ea29-fc82-48b0-86ec-3b0755de7515/data-viewer"
#>  .. .. $ method  : NULL
#>  .. .. $ headers : list()
#>  .. .. $ body    : NULL
#>  .. .. $ fields  : list()
#>  .. .. $ options : list()
#>  .. .. $ policies: list()
#>  .. .. - attr(*, "class")= chr "httr2_request"
#>  .. @ rows   : int 2945846
#>  .. @ fields : chr [1:11] "NPI" "PECOS_ASCT_CNTL_ID" "ENRLMT_ID" ...
#>  @ keyword           : chr "Medicare, Original Medicare, Provider Enrollment"
#>  @ landingPage       : chr "https://data.cms.gov/provider-characteristics/medicare-provider-supplier-enrollment/medicare-fee-for-service-pu"| __truncated__
#>  @ modified          : Date[1:1], format: "2025-01-15"
#>  @ programCode       : chr "009:000"
#>  @ publisher         : <providertwo::dataset_Publisher>
#>  .. @ type: chr "org:Organization"
#>  .. @ name: chr "Centers for Medicare & Medicaid Services"
#>  @ references        : chr "https://data.cms.gov/resources/fee-for-service-public-provider-enrollment-methodology"
#>  @ temporal          : chr "2023-01-01/2024-12-31"
#>  @ title             : chr "Medicare Fee-For-Service  Public Provider Enrollment"
#>  @ resourcesAPI      : <providertwo::dataset_Resources>
#>  .. @ url  : chr "https://data.cms.gov/data-api/v1/dataset-resources/7dcf9ea6-ee2f-4bf1-8b5d-39c18b0e8541"
#>  .. @ files: tibble [7 × 4] (S3: tbl_df/tbl/data.frame)
#>  $ name       : chr [1:7] "Medicare FFS Public Provider Enrollment Q4 2024" "Medicare FFS Public Provider Enrollment Data Dictionary" "Medicare FFS  Public Provider Enrollment Methodology" "Reassignment Sub-File 2024 Q4" ...
#>  $ fileSize   : chr [1:7] "316.55 MB" "144.35 kB" "61.36 kB" "111.75 MB" ...
#>  $ fileType   : chr [1:7] "csv" "pdf" "pdf" "csv" ...
#>  $ downloadURL: chr [1:7] "https://data.cms.gov/sites/default/files/2025-01/8ced283e-4017-4b90-b511-142aabb8531b/PPEF_Enrollment_Extract_2025.01.02.csv" "https://data.cms.gov/sites/default/files/2022-07/adb9dd34-73da-4488-855f-f091a648f87c/Medicare%20Fee-For-Servic"| __truncated__ "https://data.cms.gov/sites/default/files/2020-08/PPEF_Data_Guidance_07_31_2020.pdf" "https://data.cms.gov/sites/default/files/2025-01/PPEF_Reassignment_Extract_2025.01.02.csv" ...
public_Dataset(dataset = "Hospital Enrollments")
#> <providertwo::Dataset>
#>  @ type              : chr "dcat:Dataset"
#>  @ accessLevel       : chr "public"
#>  @ accrualPeriodicity: chr "Monthly [R/P1M]"
#>  @ bureauCode        : chr "009:38"
#>  @ contactPoint      : <providertwo::dataset_contactPoint>
#>  .. @ type    : chr "vcard:Contact"
#>  .. @ fn      : chr "Provider Enrollment Data Requests - CPI"
#>  .. @ hasEmail: chr "mailto:ProviderEnrollmentDataRequests@cms.hhs.gov"
#>  .. @ hasURL  : NULL
#>  @ describedBy       : chr "https://data.cms.gov/sites/default/files/2023-07/Hospital_Enrollments_Data_Dictionary.pdf"
#>  @ description       : chr "The Hospital Enrollments dataset provides enrollment information of all Hospitals currently enrolled in Medicar"| __truncated__
#>  @ identifier        : <providertwo::public_Identifier>
#>  .. @ url    : chr "https://data.cms.gov/data-api/v1/dataset/f6f6505c-e8b0-4d57-b258-e2b94133aaf2/data-viewer"
#>  .. @ request:List of 7
#>  .. .. $ url     : chr "https://data.cms.gov/data-api/v1/dataset/f6f6505c-e8b0-4d57-b258-e2b94133aaf2/data-viewer"
#>  .. .. $ method  : NULL
#>  .. .. $ headers : list()
#>  .. .. $ body    : NULL
#>  .. .. $ fields  : list()
#>  .. .. $ options : list()
#>  .. .. $ policies: list()
#>  .. .. - attr(*, "class")= chr "httr2_request"
#>  .. @ rows   : int 9272
#>  .. @ fields : chr [1:39] "ENROLLMENT ID" "ENROLLMENT STATE" ...
#>  @ keyword           : chr "Medicare, Original Medicare, Medicare Advantage, Hospitals & Facilities, Provider Enrollment"
#>  @ landingPage       : chr "https://data.cms.gov/provider-characteristics/hospitals-and-other-facilities/hospital-enrollments"
#>  @ modified          : Date[1:1], format: "2025-01-15"
#>  @ programCode       : chr "009:000"
#>  @ publisher         : <providertwo::dataset_Publisher>
#>  .. @ type: chr "org:Organization"
#>  .. @ name: chr "Centers for Medicare & Medicaid Services"
#>  @ references        : chr "https://data.cms.gov/sites/default/files/2024-10/Hospital_Data_Guidance.pdf"
#>  @ temporal          : chr "2022-11-01/2025-01-31"
#>  @ title             : chr "Hospital Enrollments"
#>  @ resourcesAPI      : <providertwo::dataset_Resources>
#>  .. @ url  : chr "https://data.cms.gov/data-api/v1/dataset-resources/4a4388c8-4a14-4a4d-b03a-f67d85d041f0"
#>  .. @ files: tibble [5 × 4] (S3: tbl_df/tbl/data.frame)
#>  $ name       : chr [1:5] "Hospital Enrollments Jan 2025" "Hospital Enrollments Data Dictionary (Current)" "Hospital All Owners and Enrollments Data Guidance-4" "Hospital Additional NPIs" ...
#>  $ fileSize   : chr [1:5] "2.59 MB" "364.20 kB" "509.26 kB" "71.81 kB" ...
#>  $ fileType   : chr [1:5] "csv" "pdf" "pdf" "csv" ...
#>  $ downloadURL: chr [1:5] "https://data.cms.gov/sites/default/files/2025-01/0ef8ef0d-bf93-4b00-a542-a4e2853947cc/Hospital_Enrollments_2025.01.02.csv" "https://data.cms.gov/sites/default/files/2023-07/Hospital_Enrollments_Data_Dictionary.pdf" "https://data.cms.gov/sites/default/files/2024-10/Hospital_Data_Guidance.pdf" "https://data.cms.gov/sites/default/files/2025-01/Hospital_Additional_NPIs_2025.01.02.csv" ...