Load Public API Dataset

Load Public API Dataset

Usage

public_Dataset(dataset, fname = TRUE)

Arguments

dataset

<chr> dataset title

fname

<lgl> Is dataset a function name?; default is TRUE

Value

<Dataset> object

Examples

public_Dataset("enrollees")
#> ══ Medicare Fee-For-Service  Public Provider Enrollment ════
#> 
#> ☰ Rows 2,945,846 ☰ Fields 11 ☰ Pages 590 ☰ Resources 7 ☰
#> 
#>   The Medicare Fee-For-Service Public Provider Enrollment
#>   dataset includes information on providers who are
#>   actively approved to bill Medicare or have completed the
#>   855O at the time the data was pulled from the Provider
#>   Enrollment, Chain, and Ownership System (PECOS). The
#>   release of this provider enrollment data is not related
#>   to other provider information releases such as Physician
#>   Compare or Data Transparency. Note: This full dataset
#>   contains more records than most spreadsheet programs can
#>   handle, which will result in an incomplete load of data.
#>   Use of a database or statistical software is required.
#> 
#> Timespan ❯ 2023-01-01 - 2024-12-31 Released ❯ Quarterly [R/P3M] Modified ❯
#> 2025-01-15
#> 
#> 
#> ℹ Data Dictionary | Landing Page | References

public_Dataset("hospitals")
#> ══ Hospital Enrollments ════════════════════════════════════
#> 
#> ☰ Rows 9,272 ☰ Fields 39 ☰ Pages 2 ☰ Resources 5 ☰
#> 
#>   The Hospital Enrollments dataset provides enrollment
#>   information of all Hospitals currently enrolled in
#>   Medicare. This data includes information on the
#>   Hospital's sub-group type, legal business name, doing
#>   business as name, organization type and address. 
#> 
#> Timespan ❯ 2022-11-01 - 2025-01-31 Released ❯ Monthly [R/P1M] Modified ❯
#> 2025-01-15
#> 
#> 
#> ℹ Data Dictionary | Landing Page | References

public_Dataset("reassignments")
#> ══ Revalidation Reassignment List ══════════════════════════
#> 
#> ☰ Rows 3,691,875 ☰ Fields 16 ☰ Pages 739 ☰ Resources 3 ☰
#> 
#>   The Revalidation Reassignment List dataset provides
#>   information on reassignments of providers who are due for
#>   revalidation. Note: This full dataset contains more
#>   records than most spreadsheet programs can handle, which
#>   will result in an incomplete load of data. Use of a
#>   database or statistical software is required.
#> 
#> Timespan ❯ 2022-02-01 - 2025-01-31 Released ❯ Monthly [R/P1M] Modified ❯
#> 2025-01-02
#> 
#> 
#> ℹ Data Dictionary | Landing Page | References

public_Dataset("opt_out")
#> ══ Opt Out Affidavits ══════════════════════════════════════
#> 
#> ☰ Rows 48,702 ☰ Fields 13 ☰ Pages 10 ☰ Resources 3 ☰
#> 
#>   ATTENTION USERSSome Providers' Opt-Out Status may end
#>   early due to COVID 19 waivers. Please contact your
#>   respective MAC for further information. For more
#>   information on the opt-out process, see Manage Your
#>   Enrollment or view the FAQ section below. The Opt Out
#>   Affidavits dataset provides information on providers who
#>   have decided not to participate in Medicare. It contains
#>   provider's NPI, specialty, address, and effective dates.
#> 
#> Timespan ❯ 2023-04-01 - 2024-12-31 Released ❯ Monthly [R/P1M] Modified ❯
#> 2025-01-22
#> 
#> 
#> ℹ Data Dictionary | Landing Page | References

public_Dataset("laboratories")
#> ══ Provider of Services File - Clinical Laboratories ═══════
#> 
#> ☰ Rows 653,223 ☰ Fields 103 ☰ Pages 131 ☰ Resources 3 ☰
#> 
#>   The Provider of Services (POS) Clinical Laboratories
#>   (CLIA) data provides information on CLIA demographics and
#>   types of testing services the facility provides. In this
#>   file you will find provider number, name, address and
#>   characteristics of the participating institution
#>   providers.
#> 
#> Timespan ❯ 2011-10-01 - 2024-12-31 Released ❯ Quarterly [R/P3M] Modified ❯
#> 2025-01-16
#> 
#> 
#> ℹ Data Dictionary | Landing Page | References