Quickly access CMS data

quick() is a convenience function to quickly access various CMS data endpoints. It provides a simple way to retrieve data from different CMS programs and categories.

Usage

quick(x, offset = 0L, limit = 1L, call = caller_env())

Arguments

x

A string representing the CMS data endpoint or category.

offset

An integer representing the offset for pagination. Default is 0.

limit

An integer representing the maximum number of records to retrieve. Default is 10,000.

call

The calling environment. Default is the current environment.

Value

A data frame or object containing the requested CMS data.

Examples

quick("MLR_summary") |> str()
#> tibble [1 × 14] (S3: tbl_df/tbl/data.frame)
#>  $ state                                            : chr "AL"
#>  $ program_name                                     : chr "Maternity Care Program"
#>  $ program_type                                     : chr "Other PIHP"
#>  $ eligibility_group                                : chr "All populations"
#>  $ if_other_describe_eligibility_group              : chr NA
#>  $ mco_pihp_or_pahp_name                            : chr "District 1"
#>  $ mlr_reporting_period_start_date                  : chr "2018-01-01"
#>  $ mlr_reporting_period_end_date                    : chr "2018-12-31"
#>  $ report_year                                      : chr "2018"
#>  $ mlr_numerator                                    : chr "$2,254,490"
#>  $ mlr_denominator                                  : chr "$2,329,022"
#>  $ statereported_member_months                      : chr "894"
#>  $ adjusted_mlr                                     : chr NA
#>  $ remittance_dollar_amount_for_mlr_reporting_period: chr NA