Clinicians Utilization

The Doctors and Clinicians utilization data file reports volume information for procedures of interest on clinician profile pages and in the provider data catalog (PDC) to inform patients and caregivers about clinicians' experience.

Usage

utilization(
  npi = NULL,
  pac = NULL,
  last_name = NULL,
  first_name = NULL,
  middle_name = NULL,
  suffix = NULL,
  procedure = NULL,
  count = NULL,
  percentile = NULL
)

Arguments

npi

<chr> 10-digit Individual NPI

pac

<chr> 10-digit PECOS Associate Control (PAC) ID

first_name, middle_name, last_name, suffix

<chr> Individual provider's name

procedure

<chr> Procedure category

count

<int> Number of procedures performed

percentile

<int> Percentile of procedures performed

Value

<tibble> of search results

Examples

utilization(last_name = "CURRY")
#> # A tibble: 44 × 10
#>    npi     pac   last_name first_name middle_name suff  procedure_category count
#>  * <chr>   <chr> <chr>     <chr>      <chr>       <chr> <chr>              <chr>
#>  1 103320… 8325… CURRY     ROBERT     EUGENE      NA    Hernia repair - g… 1-10 
#>  2 104342… 5890… CURRY     JOSEPH     M           NA    Upper gastrointes… 13   
#>  3 104342… 5890… CURRY     JOSEPH     M           NA    Melanoma (skin ca… 21   
#>  4 107371… 7517… CURRY     TRAVIS     NA          NA    Hip replacement    75   
#>  5 107371… 7517… CURRY     TRAVIS     NA          NA    Knee replacement   109  
#>  6 107371… 7517… CURRY     TRAVIS     NA          NA    Upper limb (arm) … 1-10 
#>  7 113475… 2860… CURRY     AMANDA     R           NA    Hip replacement    22   
#>  8 113475… 2860… CURRY     AMANDA     R           NA    Knee replacement   16   
#>  9 114431… 7214… CURRY     THOMAS     D           NA    Cataract surgery   15   
#> 10 114451… 8325… CURRY     MICHAEL    CODY        NA    Hip replacement    26   
#> # ℹ 34 more rows
#> # ℹ 2 more variables: percentile <chr>, display <chr>
utilization(npi = "1043245657")
#> NULL
utilization(npi = "1003001785")
#> # A tibble: 2 × 10
#>   npi      pac   last_name first_name middle_name suff  procedure_category count
#> * <chr>    <chr> <chr>     <chr>      <chr>       <chr> <chr>              <chr>
#> 1 1003001… 5799… JONES     JACLYN     C           NA    Lower limb (leg) … 17   
#> 2 1003001… 5799… JONES     JACLYN     C           NA    Upper limb (arm) … 27   
#> # ℹ 2 more variables: percentile <chr>, display <chr>